Struct xpcom::interfaces::imgIRequest

#[repr(C)]

pub struct imgIRequest { /* fields omitted */ }

interface imgIRequest : nsIRequest

/**
 * imgIRequest interface
 *
 * @author Stuart Parmenter <stuart@mozilla.com>
 * @version 0.1
 * @see imagelib2
 */

Methods

impl imgIRequest

pub fn coerce<T: imgIRequestCoerce>(&self) -> &T

Cast this imgIRequest to one of its base interfaces.

impl imgIRequest

pub const STATUS_NONE: i64

STATUS_NONE: i64 = 0

/**
   * Bits set in the return value from imageStatus
   * @name statusflags
   *
   * Meanings:
   *
   * STATUS_NONE: Nothing to report.
   *
   * STATUS_SIZE_AVAILABLE: We received enough image data
   * from the network or filesystem that we know the width
   * and height of the image, and have thus called SetSize()
   * on the container.
   *
   * STATUS_LOAD_COMPLETE: The data has been fully loaded
   * to memory, but not necessarily fully decoded.
   *
   * STATUS_ERROR: An error occurred loading the image.
   *
   * STATUS_FRAME_COMPLETE: The first frame has been
   * completely decoded.
   *
   * STATUS_DECODE_COMPLETE: The whole image has been decoded.
   *
   * STATUS_IS_ANIMATED: The image is animated.
   *
   * STATUS_HAS_TRANSPARENCY: The image is partially or completely transparent.
   */

pub const STATUS_SIZE_AVAILABLE: i64

STATUS_SIZE_AVAILABLE: i64 = 1

pub const STATUS_LOAD_COMPLETE: i64

STATUS_LOAD_COMPLETE: i64 = 2

pub const STATUS_ERROR: i64

STATUS_ERROR: i64 = 4

pub const STATUS_FRAME_COMPLETE: i64

STATUS_FRAME_COMPLETE: i64 = 8

pub const STATUS_DECODE_COMPLETE: i64

STATUS_DECODE_COMPLETE: i64 = 16

pub const STATUS_IS_ANIMATED: i64

STATUS_IS_ANIMATED: i64 = 32

pub const STATUS_HAS_TRANSPARENCY: i64

STATUS_HAS_TRANSPARENCY: i64 = 64

pub const CORS_NONE: i64

CORS_NONE: i64 = 1

/**
   * CORS modes images can be loaded with.
   *
   * By default, all images are loaded with CORS_NONE and cannot be used
   * cross-origin in context like WebGL.
   *
   * If an HTML img element has the crossorigin attribute set, the imgIRequest
   * will be validated for cross-origin usage with CORS, and, if successful,
   * will have its CORS mode set to the relevant type.
   */

pub const CORS_ANONYMOUS: i64

CORS_ANONYMOUS: i64 = 2

pub const CORS_USE_CREDENTIALS: i64

CORS_USE_CREDENTIALS: i64 = 3

pub const CATEGORY_FRAME_INIT: i64

CATEGORY_FRAME_INIT: i64 = 1

/**
   * Request loading priority boost to requested category, each category
   * of request increases priority only one time..
   *
   * CATEGORY_FRAME_INIT: increase priority when the imgRequest is associated
   * with an nsImageFrame.
   *
   * CATEGORY_SIZE_QUERY: increase priority when size decoding is necessary to
   * determine the layout size of the associated nsImageFrame.
   *
   * CATEGORY_DISPLAY: increase priority when the image is about to be displayed
   * in the viewport.
   */

pub const CATEGORY_SIZE_QUERY: i64

CATEGORY_SIZE_QUERY: i64 = 2

pub const CATEGORY_DISPLAY: i64

CATEGORY_DISPLAY: i64 = 4

pub unsafe fn GetImage(&self, aImage: *mut *const imgIContainer) -> nsresult

/**
   * the image container...
   * @return the image object associated with the request.
   * @attention NEED DOCS
   */

readonly attribute imgIContainer image;

pub unsafe fn GetImageStatus(&self, aImageStatus: *mut uint32_t) -> nsresult

/**
   * Status flags of the STATUS_* variety.
   */

readonly attribute unsigned long imageStatus;

pub unsafe fn GetImageErrorCode(
&self,
aImageErrorCode: *mut nsresult
) -> nsresult

[noscript] readonly attribute nsresult imageErrorCode;

pub unsafe fn GetURI(&self, aURI: *mut *const nsIURI) -> nsresult

/**
   * The URI the image load was started with.  Note that this might not be the
   * actual URI for the image (e.g. if HTTP redirects happened during the
   * load).
   */

readonly attribute nsIURI URI;

pub unsafe fn GetFinalURI(&self, aFinalURI: *mut *const nsIURI) -> nsresult

/**
   * The URI of the resource we ended up loading after all redirects, etc.
   */

readonly attribute nsIURI finalURI;

pub unsafe fn GetNotificationObserver(
&self,
aNotificationObserver: *mut *const imgINotificationObserver
) -> nsresult

readonly attribute imgINotificationObserver notificationObserver;

pub unsafe fn GetMimeType(&self, aMimeType: *mut *const c_char) -> nsresult

readonly attribute string mimeType;

pub unsafe fn Clone(
&self,
aObserver: *const imgINotificationObserver,
_retval: *mut *const imgIRequest
) -> nsresult

/**
   * Clone this request; the returned request will have aObserver as the
   * observer.  aObserver will be notified synchronously (before the clone()
   * call returns) with all the notifications that have already been dispatched
   * for this image load.
   */

imgIRequest clone (in imgINotificationObserver aObserver);

pub unsafe fn GetImagePrincipal(
&self,
aImagePrincipal: *mut *const nsIPrincipal
) -> nsresult

/**
   * The principal gotten from the channel the image was loaded from.
   */

readonly attribute nsIPrincipal imagePrincipal;

pub unsafe fn GetMultipart(&self, aMultipart: *mut bool) -> nsresult

/**
   * Whether the request is multipart (ie, multipart/x-mixed-replace)
   */

readonly attribute bool multipart;

pub unsafe fn GetCORSMode(&self, aCORSMode: *mut int32_t) -> nsresult

/**
   * The CORS mode that this image was loaded with.
   */

readonly attribute long CORSMode;

pub unsafe fn CancelAndForgetObserver(&self, aStatus: nsresult) -> nsresult

/**
   * Cancels this request as in nsIRequest::Cancel(); further, also nulls out
   * decoderObserver so it gets no further notifications from us.
   *
   * NOTE: You should not use this in any new code; instead, use cancel(). Note
   * that cancel() is asynchronous, which means that some time after you call
   * it, the listener/observer will get an OnStopRequest(). This means that, if
   * you're the observer, you can't call cancel() from your destructor.
   */

void cancelAndForgetObserver (in nsresult aStatus);

pub unsafe fn StartDecoding(&self, aFlags: uint32_t) -> nsresult

/**
   * Requests a synchronous decode for the image.
   *
   * imgIContainer has a startDecoding() method, but callers may want to request
   * a decode before the container has necessarily been instantiated. Calling
   * startDecoding() on the imgIRequest simply forwards along the request if the
   * container already exists, or calls it once the container becomes available
   * if it does not yet exist.
   */

void startDecoding (in uint32_t aFlags);

pub unsafe fn StartDecodingWithResult(&self, aFlags: uint32_t) -> bool

/**
   * Exactly like startDecoding above except returns whether the current frame
   * of the image is complete or not.
   */

[noscript,notxpcom] boolean startDecodingWithResult (in uint32_t aFlags);

pub unsafe fn LockImage(&self) -> nsresult

/**
   * Locks an image. If the image does not exist yet, locks it once it becomes
   * available. The lock persists for the lifetime of the imgIRequest (until
   * unlockImage is called) even if the underlying image changes.
   *
   * If you don't call unlockImage() by the time this imgIRequest goes away, it
   * will be called for you automatically.
   *
   * @see imgIContainer::lockImage for documentation of the underlying call.
   */

void lockImage ();

pub unsafe fn UnlockImage(&self) -> nsresult

/**
   * Unlocks an image.
   *
   * @see imgIContainer::unlockImage for documentation of the underlying call.
   */

void unlockImage ();

pub unsafe fn RequestDiscard(&self) -> nsresult

/**
   * If this image is unlocked, discard the image's decoded data.  If the image
   * is locked or is already discarded, do nothing.
   */

void requestDiscard ();

pub unsafe fn GetStaticRequest(
&self,
_retval: *mut *const imgIRequest
) -> nsresult

/**
   * If this request is for an animated image, the method creates a new
   * request which contains the current frame of the image.
   * Otherwise returns the same request.
   */

imgIRequest getStaticRequest ();

pub unsafe fn IncrementAnimationConsumers(&self) -> nsresult

/**
   * Requests that the image animate (if it has an animation).
   *
   * @see Image::IncrementAnimationConsumers for documentation of the
   * underlying call.
   */

void incrementAnimationConsumers ();

pub unsafe fn DecrementAnimationConsumers(&self) -> nsresult

/**
   * Tell the image it can forget about a request that the image animate.
   *
   * @see Image::DecrementAnimationConsumers for documentation of the
   * underlying call.
   */

void decrementAnimationConsumers ();

pub unsafe fn BoostPriority(&self, aCategory: uint32_t) -> nsresult

void boostPriority (in uint32_t aCategory);