Struct xpcom::interfaces::inIDOMUtils

#[repr(C)]

pub struct inIDOMUtils { /* fields omitted */ }

interface inIDOMUtils : nsISupports

Methods

impl inIDOMUtils

pub fn coerce<T: inIDOMUtilsCoerce>(&self) -> &T

Cast this inIDOMUtils to one of its base interfaces.

impl inIDOMUtils

pub const EXCLUDE_SHORTHANDS: i64

EXCLUDE_SHORTHANDS: i64 = 1

pub const INCLUDE_ALIASES: i64

INCLUDE_ALIASES: i64 = 2

pub const TYPE_LENGTH: i64

TYPE_LENGTH: i64 = 0

pub const TYPE_PERCENTAGE: i64

TYPE_PERCENTAGE: i64 = 1

pub const TYPE_COLOR: i64

TYPE_COLOR: i64 = 2

pub const TYPE_URL: i64

TYPE_URL: i64 = 3

pub const TYPE_ANGLE: i64

TYPE_ANGLE: i64 = 4

pub const TYPE_FREQUENCY: i64

TYPE_FREQUENCY: i64 = 5

pub const TYPE_TIME: i64

TYPE_TIME: i64 = 6

pub const TYPE_GRADIENT: i64

TYPE_GRADIENT: i64 = 7

pub const TYPE_TIMING_FUNCTION: i64

TYPE_TIMING_FUNCTION: i64 = 8

pub const TYPE_IMAGE_RECT: i64

TYPE_IMAGE_RECT: i64 = 9

pub const TYPE_NUMBER: i64

TYPE_NUMBER: i64 = 10

pub unsafe fn GetAllStyleSheets(
&self,
aDoc: *const nsIDOMDocument,
aLength: *mut uint32_t,
aSheets: *mut *mut *const nsISupports
) -> nsresult

void getAllStyleSheets (in nsIDOMDocument aDoc, [optional] out unsigned long aLength, [array, size_is (aLength), retval] out nsISupports aSheets);

pub unsafe fn GetCSSStyleRules(
&self,
aElement: *const nsIDOMElement,
aPseudo: &nsAString,
_retval: *mut *const nsIArrayExtensions
) -> nsresult

nsIArrayExtensions getCSSStyleRules (in nsIDOMElement aElement, [optional] in DOMString aPseudo);

pub unsafe fn GetRuleLine(
&self,
aRule: *const nsIDOMCSSRule,
_retval: *mut uint32_t
) -> nsresult

/**
   * Get the line number of a rule.
   *
   * @param nsIDOMCSSRule aRule The rule.
   * @return The rule's line number.  Line numbers are 1-based.
   */

unsigned long getRuleLine (in nsIDOMCSSRule aRule);

pub unsafe fn GetRuleColumn(
&self,
aRule: *const nsIDOMCSSRule,
_retval: *mut uint32_t
) -> nsresult

/**
   * Get the column number of a rule.
   *
   * @param nsIDOMCSSRule aRule The rule.
   * @return The rule's column number.  Column numbers are 1-based.
   */

unsigned long getRuleColumn (in nsIDOMCSSRule aRule);

pub unsafe fn GetRelativeRuleLine(
&self,
aRule: *const nsIDOMCSSRule,
_retval: *mut uint32_t
) -> nsresult

/**
   * Like getRuleLine, but if the rule is in a <style> element,
   * returns a line number relative to the start of the element.
   *
   * @param nsIDOMCSSRule aRule the rule to examine
   * @return the line number of the rule, possibly relative to the
   *         <style> element
   */

unsigned long getRelativeRuleLine (in nsIDOMCSSRule aRule);

pub unsafe fn GetSelectorCount(
&self,
aRule: *const nsIDOMCSSStyleRule,
_retval: *mut uint32_t
) -> nsresult

unsigned long getSelectorCount (in nsIDOMCSSStyleRule aRule);

pub unsafe fn GetSelectorText(
&self,
aRule: *const nsIDOMCSSStyleRule,
aSelectorIndex: uint32_t,
_retval: &mut nsAString
) -> nsresult

AString getSelectorText (in nsIDOMCSSStyleRule aRule, in unsigned long aSelectorIndex);

pub unsafe fn GetSpecificity(
&self,
aRule: *const nsIDOMCSSStyleRule,
aSelectorIndex: uint32_t,
_retval: *mut uint64_t
) -> nsresult

unsigned long long getSpecificity (in nsIDOMCSSStyleRule aRule, in unsigned long aSelectorIndex);

pub unsafe fn SelectorMatchesElement(
&self,
aElement: *const nsIDOMElement,
aRule: *const nsIDOMCSSStyleRule,
aSelectorIndex: uint32_t,
aPseudo: &nsAString,
_retval: *mut bool
) -> nsresult

bool selectorMatchesElement (in nsIDOMElement aElement, in nsIDOMCSSStyleRule aRule, in unsigned long aSelectorIndex, [optional] in DOMString aPseudo);

pub unsafe fn IsInheritedProperty(
&self,
aPropertyName: &nsAString,
_retval: *mut bool
) -> nsresult

bool isInheritedProperty (in AString aPropertyName);

pub unsafe fn GetCSSPropertyNames(
&self,
aFlags: uint32_t,
aCount: *mut uint32_t,
aProps: *mut *mut *const int16_t
) -> nsresult

void getCSSPropertyNames ([optional] in unsigned long aFlags, [optional] out unsigned long aCount, [array, size_is (aCount), retval] out wstring aProps);

pub unsafe fn GetCSSValuesForProperty(
&self,
aProperty: &nsAString,
aLength: *mut uint32_t,
aValues: *mut *mut *const int16_t
) -> nsresult

void getCSSValuesForProperty (in AString aProperty, [optional] out unsigned long aLength, [array, size_is (aLength), retval] out wstring aValues);

pub unsafe fn RgbToColorName(
&self,
aR: uint8_t,
aG: uint8_t,
aB: uint8_t,
_retval: &mut nsAString
) -> nsresult

AString rgbToColorName (in octet aR, in octet aG, in octet aB);

pub unsafe fn IsValidCSSColor(
&self,
aColorString: &nsAString,
_retval: *mut bool
) -> nsresult

bool isValidCSSColor (in AString aColorString);

pub unsafe fn GetSubpropertiesForCSSProperty(
&self,
aProperty: &nsAString,
aLength: *mut uint32_t,
aValues: *mut *mut *const int16_t
) -> nsresult

void getSubpropertiesForCSSProperty (in AString aProperty, [optional] out unsigned long aLength, [array, size_is (aLength), retval] out wstring aValues);

pub unsafe fn CssPropertyIsShorthand(
&self,
aProperty: &nsAString,
_retval: *mut bool
) -> nsresult

bool cssPropertyIsShorthand (in AString aProperty);

pub unsafe fn CssPropertySupportsType(
&self,
aProperty: &nsAString,
type_: uint32_t,
_retval: *mut bool
) -> nsresult

bool cssPropertySupportsType (in AString aProperty, in unsigned long type);

pub unsafe fn IsIgnorableWhitespace(
&self,
aDataNode: *const nsIDOMCharacterData,
_retval: *mut bool
) -> nsresult

boolean isIgnorableWhitespace (in nsIDOMCharacterData aDataNode);

pub unsafe fn GetParentForNode(
&self,
aNode: *const nsIDOMNode,
aShowingAnonymousContent: bool,
_retval: *mut *const nsIDOMNode
) -> nsresult

nsIDOMNode getParentForNode (in nsIDOMNode aNode, in boolean aShowingAnonymousContent);

pub unsafe fn GetChildrenForNode(
&self,
aNode: *const nsIDOMNode,
aShowingAnonymousContent: bool,
_retval: *mut *const nsIDOMNodeList
) -> nsresult

nsIDOMNodeList getChildrenForNode (in nsIDOMNode aNode, in boolean aShowingAnonymousContent);

pub unsafe fn GetBindingURLs(
&self,
aElement: *const nsIDOMElement,
_retval: *mut *const nsIArray
) -> nsresult

nsIArray getBindingURLs (in nsIDOMElement aElement);

pub unsafe fn GetContentState(
&self,
aElement: *const nsIDOMElement,
_retval: *mut uint64_t
) -> nsresult

unsigned long long getContentState (in nsIDOMElement aElement);

pub unsafe fn SetContentState(
&self,
aElement: *const nsIDOMElement,
aState: uint64_t,
_retval: *mut bool
) -> nsresult

/**
   * Setting and removing content state on an element. Both these functions
   * calling EventStateManager::SetContentState internally, the difference is
   * that for the remove case we simply pass in nullptr for the element.
   * Use them accordingly.
   *
   * When removing the active state, you may optionally also clear the active
   * document as well by setting aClearActiveDocument
   *
   * @return Returns true if the state was set successfully. See more details
   * in EventStateManager.h SetContentState.
   */

bool setContentState (in nsIDOMElement aElement, in unsigned long long aState);

pub unsafe fn RemoveContentState(
&self,
aElement: *const nsIDOMElement,
aState: uint64_t,
aClearActiveDocument: bool,
_retval: *mut bool
) -> nsresult

bool removeContentState (in nsIDOMElement aElement, in unsigned long long aState, [optional] in bool aClearActiveDocument);

pub unsafe fn GetUsedFontFaces(
&self,
aRange: *const nsIDOMRange,
_retval: *mut *const nsIDOMFontFaceList
) -> nsresult

nsIDOMFontFaceList getUsedFontFaces (in nsIDOMRange aRange);

pub unsafe fn GetCSSPseudoElementNames(
&self,
aCount: *mut uint32_t,
aNames: *mut *mut *const int16_t
) -> nsresult

/**
   * Get the names of all the supported pseudo-elements.
   * Pseudo-elements which are only accepted in UA style sheets are
   * not included.
   *
   * @param {unsigned long} aCount the number of items returned
   * @param {wstring[]} aNames the names
   */

void getCSSPseudoElementNames ([optional] out unsigned long aCount, [array, size_is (aCount), retval] out wstring aNames);

pub unsafe fn RemovePseudoClassLock(
&self,
aElement: *const nsIDOMElement,
aPseudoClass: &nsAString
) -> nsresult

void removePseudoClassLock (in nsIDOMElement aElement, in DOMString aPseudoClass);

pub unsafe fn HasPseudoClassLock(
&self,
aElement: *const nsIDOMElement,
aPseudoClass: &nsAString,
_retval: *mut bool
) -> nsresult

bool hasPseudoClassLock (in nsIDOMElement aElement, in DOMString aPseudoClass);

pub unsafe fn ClearPseudoClassLocks(
&self,
aElement: *const nsIDOMElement
) -> nsresult

void clearPseudoClassLocks (in nsIDOMElement aElement);

pub unsafe fn ParseStyleSheet(
&self,
aSheet: *const nsIDOMCSSStyleSheet,
aInput: &nsAString
) -> nsresult

/**
   * Parse CSS and update the style sheet in place.
   *
   * @param DOMCSSStyleSheet aSheet
   * @param DOMString aInput
   *        The new source string for the style sheet.
   */

void parseStyleSheet (in nsIDOMCSSStyleSheet aSheet, in DOMString aInput);

pub unsafe fn ScrollElementIntoView(
&self,
aElement: *const nsIDOMElement
) -> nsresult

/**
   * Scroll an element completely into view, if possible.
   * This is similar to ensureElementIsVisible but for all ancestors.
   *
   * @param DOMElement aElement
   */

void scrollElementIntoView (in nsIDOMElement aElement);