Struct xpcom::interfaces::mozIAnnotatedResult

#[repr(C)]

pub struct mozIAnnotatedResult { /* fields omitted */ }

interface mozIAnnotatedResult : nsISupports

Methods

impl mozIAnnotatedResult

pub fn coerce<T: mozIAnnotatedResultCoerce>(&self) -> &T

Cast this mozIAnnotatedResult to one of its base interfaces.

impl mozIAnnotatedResult

pub unsafe fn GetGuid(&self, aGuid: &mut nsACString) -> nsresult

/**
 * Represents a place annotated with a given annotation.  If a place has
 * multiple annotations, it can be represented by multiple
 * mozIAnnotatedResult(s).
 */
/**
     * The globally unique identifier of the place with this annotation.
     *
     * @note if itemId is valid this is the guid of the bookmark, otherwise
     *       of the page.
     */

readonly attribute AUTF8String guid;

pub unsafe fn GetUri(&self, aUri: *mut *const nsIURI) -> nsresult

/**
     * The URI of the place with this annotation, if available, null otherwise.
     */

readonly attribute nsIURI uri;

pub unsafe fn GetItemId(&self, aItemId: *mut int64_t) -> nsresult

/**
     * The bookmark id of the place with this annotation, if available,
     * -1 otherwise.
     *
     * @note if itemId is -1, it doesn't mean the page is not bookmarked, just
     *       that this annotation is relative to the page, not to the bookmark.
     */

readonly attribute long long itemId;

pub unsafe fn GetAnnotationName(
&self,
aAnnotationName: &mut nsACString
) -> nsresult

/**
     * Name of the annotation.
     */

readonly attribute AUTF8String annotationName;

pub unsafe fn GetAnnotationValue(
&self,
aAnnotationValue: *mut *const nsIVariant
) -> nsresult

/**
     * Value of the annotation.
     */

readonly attribute nsIVariant annotationValue;