Struct xpcom::interfaces::mozIGeckoMediaPluginChromeService

#[repr(C)]

pub struct mozIGeckoMediaPluginChromeService { /* fields omitted */ }

interface mozIGeckoMediaPluginChromeService : nsISupports

Methods

impl mozIGeckoMediaPluginChromeService

pub fn coerce<T: mozIGeckoMediaPluginChromeServiceCoerce>(&self) -> &T

Cast this mozIGeckoMediaPluginChromeService to one of its base interfaces.

impl mozIGeckoMediaPluginChromeService

pub unsafe fn AddPluginDirectory(&self, directory: &nsAString) -> nsresult

/**
   * Add a directory to scan for gecko media plugins.
   * @note Main-thread API.
   */

void addPluginDirectory (in AString directory);

pub unsafe fn RemovePluginDirectory(&self, directory: &nsAString) -> nsresult

/**
   * Remove a directory for gecko media plugins.
   * @note Main-thread API.
   */

void removePluginDirectory (in AString directory);

pub unsafe fn RemoveAndDeletePluginDirectory(
&self,
directory: &nsAString,
defer: bool
) -> nsresult

/**
   * Remove a directory for gecko media plugins and delete it from disk.
   * If |defer| is true, wait until the plugin is unused before removing.
   * @note Main-thread API.
   */

void removeAndDeletePluginDirectory (in AString directory, [optional] in bool defer);

pub unsafe fn ForgetThisSite(
&self,
site: &nsAString,
aPattern: &nsAString
) -> nsresult

/**
   * Clears storage data associated with the site and the originAttributes
   * pattern in JSON format.
   */

void forgetThisSite (in AString site, in DOMString aPattern);

pub unsafe fn IsPersistentStorageAllowed(
&self,
nodeId: &nsACString,
_retval: *mut bool
) -> nsresult

/**
   * Returns true if the given node id is allowed to store things
   * persistently on disk. Private Browsing and local content are not
   * allowed to store persistent data.
   */

bool isPersistentStorageAllowed (in ACString nodeId);

pub unsafe fn GetStorageDir(&self, _retval: *mut *const nsIFile) -> nsresult

/**
   * Returns the directory to use as the base for storing data about GMPs.
   */

nsIFile getStorageDir ();