Struct xpcom::interfaces::mozILivemark

#[repr(C)]

pub struct mozILivemark { /* fields omitted */ }

interface mozILivemark : mozILivemarkInfo

Methods

impl mozILivemark

pub fn coerce<T: mozILivemarkCoerce>(&self) -> &T

Cast this mozILivemark to one of its base interfaces.

impl mozILivemark

pub const STATUS_READY: i64

STATUS_READY: i64 = 0

pub const STATUS_LOADING: i64

STATUS_LOADING: i64 = 1

pub const STATUS_FAILED: i64

STATUS_FAILED: i64 = 2

pub unsafe fn GetStatus(&self, aStatus: *mut uint16_t) -> nsresult

/**
   * Status of this livemark.  One of the STATUS_* constants above.
   */

readonly attribute unsigned short status;

pub unsafe fn Reload(&self, aForceUpdate: bool) -> nsresult

/**
   * Reload livemark contents if they are expired or if forced to do so.
   *
   * @param [optional]aForceUpdate
   *        If set to true forces a reload even if contents are still valid.
   *
   * @note The update process is asynchronous, it's possible to register a
   *       result observer to be notified of updated contents through
   *       registerForUpdates.
   */

void reload ([optional] in boolean aForceUpdate);