Struct xpcom::interfaces::mozIPlacesAutoComplete

#[repr(C)]

pub struct mozIPlacesAutoComplete { /* fields omitted */ }

interface mozIPlacesAutoComplete : nsISupports

/**
 * This interface provides some constants used by the Places AutoComplete
 * search provider as well as methods to track opened pages for AutoComplete
 * purposes.
 */

Methods

impl mozIPlacesAutoComplete

pub fn coerce<T: mozIPlacesAutoCompleteCoerce>(&self) -> &T

Cast this mozIPlacesAutoComplete to one of its base interfaces.

impl mozIPlacesAutoComplete

pub const MATCH_ANYWHERE: i64

MATCH_ANYWHERE: i64 = 0

/**
   * Match anywhere in each searchable term.
   */

pub const MATCH_BOUNDARY_ANYWHERE: i64

MATCH_BOUNDARY_ANYWHERE: i64 = 1

/**
   * Match first on word boundaries, and if we do not get enough results, then
   * match anywhere in each searchable term.
   */

pub const MATCH_BOUNDARY: i64

MATCH_BOUNDARY: i64 = 2

/**
   * Match on word boundaries in each searchable term.
   */

pub const MATCH_BEGINNING: i64

MATCH_BEGINNING: i64 = 3

/**
   * Match only the beginning of each search term.
   */

pub const MATCH_ANYWHERE_UNMODIFIED: i64

MATCH_ANYWHERE_UNMODIFIED: i64 = 4

/**
   * Match anywhere in each searchable term without doing any transformation
   * or stripping on the underlying data.
   */

pub const MATCH_BEGINNING_CASE_SENSITIVE: i64

MATCH_BEGINNING_CASE_SENSITIVE: i64 = 5

/**
   * Match only the beginning of each search term using a case sensitive
   * comparator.
   */

pub const BEHAVIOR_HISTORY: i64

BEHAVIOR_HISTORY: i64 = 1

/**
   * Search through history.
   */

pub const BEHAVIOR_BOOKMARK: i64

BEHAVIOR_BOOKMARK: i64 = 2

/**
   * Search though bookmarks.
   */

pub const BEHAVIOR_TAG: i64

BEHAVIOR_TAG: i64 = 4

/**
   * Search through tags.
   */

pub const BEHAVIOR_TITLE: i64

BEHAVIOR_TITLE: i64 = 8

/**
   * Search the title of pages.
   */

pub const BEHAVIOR_URL: i64

BEHAVIOR_URL: i64 = 16

/**
   * Search the URL of pages.
   */

pub const BEHAVIOR_TYPED: i64

BEHAVIOR_TYPED: i64 = 32

/**
   * Search for typed pages.
   */

pub const BEHAVIOR_JAVASCRIPT: i64

BEHAVIOR_JAVASCRIPT: i64 = 64

/**
   * Search javascript: URLs.
   */

pub const BEHAVIOR_OPENPAGE: i64

BEHAVIOR_OPENPAGE: i64 = 128

/**
   * Search for pages that have been marked as being opened, such as a tab
   * in a tabbrowser.
   */

pub const BEHAVIOR_RESTRICT: i64

BEHAVIOR_RESTRICT: i64 = 256

/**
   * Use intersection between history, typed, bookmark, tag and openpage
   * instead of union, when the restrict bit is set.
   */

pub const BEHAVIOR_SEARCHES: i64

BEHAVIOR_SEARCHES: i64 = 512

/**
   * Include search suggestions from the currently selected search provider.
   */

pub unsafe fn RegisterOpenPage(
&self,
aURI: *const nsIURI,
aUserContextId: uint32_t
) -> nsresult

/**
   * Mark a page as being currently open.
   *
   * @note Pages will not be automatically unregistered when Private Browsing
   *       mode is entered or exited.  Therefore, consumers MUST unregister or
   *       register themselves.
   *
   * @param aURI
   *        The URI to register as an open page.
   * @param aUserContextId
   *        The Container Id of the tab.
   */

void registerOpenPage (in nsIURI aURI, in uint32_t aUserContextId);

pub unsafe fn UnregisterOpenPage(
&self,
aURI: *const nsIURI,
aUserContextId: uint32_t
) -> nsresult

/**
   * Mark a page as no longer being open (either by closing the window or tab,
   * or by navigating away from that page).
   *
   * @note Pages will not be automatically unregistered when Private Browsing
   *       mode is entered or exited.  Therefore, consumers MUST unregister or
   *       register themselves.
   *
   * @param aURI
   *        The URI to unregister as an open page.
   * @param aUserContextId
   *        The Container Id of the tab.
   */

void unregisterOpenPage (in nsIURI aURI, in uint32_t aUserContextId);