Struct xpcom::interfaces::mozIStorageRow

#[repr(C)]

pub struct mozIStorageRow { /* fields omitted */ }

interface mozIStorageRow : mozIStorageValueArray

Methods

impl mozIStorageRow

pub fn coerce<T: mozIStorageRowCoerce>(&self) -> &T

Cast this mozIStorageRow to one of its base interfaces.

impl mozIStorageRow

pub unsafe fn GetResultByIndex(
&self,
aIndex: uint32_t,
_retval: *mut *const nsIVariant
) -> nsresult

/**
   * Obtains the result of a given column specified by aIndex.
   *
   * @param aIndex
   *        Zero-based index of the result to get from the tuple.
   * @returns the result of the specified column.
   */

nsIVariant getResultByIndex (in unsigned long aIndex);

pub unsafe fn GetResultByName(
&self,
aName: &nsACString,
_retval: *mut *const nsIVariant
) -> nsresult

/**
   * Obtains the result of a given column specified by aName.
   *
   * @param aName
   *        Name of the result to get from the tuple.
   * @returns the result of the specified column.
   * @note The name of a result column is the value of the "AS" clause for that
   *       column.  If there is no AS clause then the name of the column is
   *       unspecified and may change from one release to the next.
   */

nsIVariant getResultByName (in AUTF8String aName);