Struct xpcom::interfaces::mozIStorageVacuumParticipant

#[repr(C)]

pub struct mozIStorageVacuumParticipant { /* fields omitted */ }

interface mozIStorageVacuumParticipant : nsISupports

/**
 * This interface contains the information that the Storage service needs to
 * vacuum a database.  This interface is created as a service through the
 * category manager with the category "vacuum-participant".
 * Please see https://developer.mozilla.org/en/mozIStorageVacuumParticipant for
 * more information.
 */

Methods

impl mozIStorageVacuumParticipant

pub fn coerce<T: mozIStorageVacuumParticipantCoerce>(&self) -> &T

Cast this mozIStorageVacuumParticipant to one of its base interfaces.

impl mozIStorageVacuumParticipant

pub unsafe fn GetExpectedDatabasePageSize(
&self,
aExpectedDatabasePageSize: *mut int32_t
) -> nsresult

/**
   * The expected page size in bytes for the database.  The vacuum manager will
   * try to correct the page size during idle based on this value.
   *
   * @note If the database is using the WAL journal mode, the page size won't
  *        be changed to the requested value.  See bug 634374.
   * @note Valid page size values are powers of 2 between 512 and 65536.
   *       The suggested value is mozIStorageConnection::defaultPageSize.
   */

readonly attribute long expectedDatabasePageSize;

pub unsafe fn GetDatabaseConnection(
&self,
aDatabaseConnection: *mut *const mozIStorageConnection
) -> nsresult

/**
   * Connection to the database file to be vacuumed.
   */

readonly attribute mozIStorageConnection databaseConnection;

pub unsafe fn OnBeginVacuum(&self, _retval: *mut bool) -> nsresult

/**
   * Notifies when a vacuum operation begins.  Listeners should avoid using the
   * database till onEndVacuum is received.
   *
   * @return true to proceed with the vacuum, false if the participant wants to
   *         opt-out for now, it will be retried later.  Useful when participant
   *         is running some other heavy operation that can't be interrupted.
   *
   * @note When a vacuum operation starts or ends it will also dispatch a global
   *       "heavy-io-task" notification through the observer service with the
   *       data argument being either "vacuum-begin" or "vacuum-end".
   */

boolean onBeginVacuum ();

pub unsafe fn OnEndVacuum(&self, aSucceeded: bool) -> nsresult

/**
   * Notifies when a vacuum operation ends.
   *
   * @param aSucceeded
   *        reports if the vacuum succeeded or failed.
   */

void onEndVacuum (in boolean aSucceeded);