Struct xpcom::interfaces::mozIStorageValueArray

#[repr(C)]

pub struct mozIStorageValueArray { /* fields omitted */ }

interface mozIStorageValueArray : nsISupports

/**
 * mozIStorageValueArray wraps an array of SQL values, such as a single database
 * row.
 */

Methods

impl mozIStorageValueArray

pub fn coerce<T: mozIStorageValueArrayCoerce>(&self) -> &T

Cast this mozIStorageValueArray to one of its base interfaces.

impl mozIStorageValueArray

pub const VALUE_TYPE_NULL: i64

VALUE_TYPE_NULL: i64 = 0

/**
   * These type values are returned by getTypeOfIndex
   * to indicate what type of value is present at
   * a given column.
   */

pub const VALUE_TYPE_INTEGER: i64

VALUE_TYPE_INTEGER: i64 = 1

pub const VALUE_TYPE_FLOAT: i64

VALUE_TYPE_FLOAT: i64 = 2

pub const VALUE_TYPE_TEXT: i64

VALUE_TYPE_TEXT: i64 = 3

pub const VALUE_TYPE_BLOB: i64

VALUE_TYPE_BLOB: i64 = 4

pub unsafe fn GetNumEntries(&self, aNumEntries: *mut uint32_t) -> nsresult

/**
   * numEntries
   *
   * number of entries in the array (each corresponding to a column
   * in the database row)
   */

readonly attribute unsigned long numEntries;

pub unsafe fn GetTypeOfIndex(
&self,
aIndex: uint32_t,
_retval: *mut int32_t
) -> nsresult

/**
   * Returns the type of the value at the given column index;
   * one of VALUE_TYPE_NULL, VALUE_TYPE_INTEGER, VALUE_TYPE_FLOAT,
   * VALUE_TYPE_TEXT, VALUE_TYPE_BLOB.
   */

long getTypeOfIndex (in unsigned long aIndex);

pub unsafe fn GetInt32(
&self,
aIndex: uint32_t,
_retval: *mut int32_t
) -> nsresult

/**
   * Obtain a value for the given entry (column) index.
   * Due to SQLite's type conversion rules, any of these are valid
   * for any column regardless of the column's data type.  However,
   * if the specific type matters, getTypeOfIndex should be used
   * first to identify the column type, and then the appropriate
   * get method should be called.
   *
   * If you ask for a string value for a NULL column, you will get an empty
   * string with IsVoid set to distinguish it from an explicitly set empty
   * string.
   */

long getInt32 (in unsigned long aIndex);

pub unsafe fn GetInt64(
&self,
aIndex: uint32_t,
_retval: *mut int64_t
) -> nsresult

long long getInt64 (in unsigned long aIndex);

pub unsafe fn GetDouble(
&self,
aIndex: uint32_t,
_retval: *mut c_double
) -> nsresult

double getDouble (in unsigned long aIndex);

pub unsafe fn GetUTF8String(
&self,
aIndex: uint32_t,
_retval: &mut nsACString
) -> nsresult

AUTF8String getUTF8String (in unsigned long aIndex);

pub unsafe fn GetString(
&self,
aIndex: uint32_t,
_retval: &mut nsAString
) -> nsresult

AString getString (in unsigned long aIndex);

pub unsafe fn GetBlob(
&self,
aIndex: uint32_t,
aDataSize: *mut uint32_t,
aData: *mut *mut uint8_t
) -> nsresult

void getBlob (in unsigned long aIndex, out unsigned long aDataSize, [array, size_is (aDataSize)] out octet aData);

pub unsafe fn GetBlobAsString(
&self,
aIndex: uint32_t,
_retval: &mut nsAString
) -> nsresult

AString getBlobAsString (in unsigned long aIndex);

pub unsafe fn GetBlobAsUTF8String(
&self,
aIndex: uint32_t,
_retval: &mut nsACString
) -> nsresult

AUTF8String getBlobAsUTF8String (in unsigned long aIndex);

pub unsafe fn GetIsNull(&self, aIndex: uint32_t, _retval: *mut bool) -> nsresult

boolean getIsNull (in unsigned long aIndex);

pub unsafe fn GetSharedUTF8String(
&self,
aIndex: uint32_t,
aLength: *mut uint32_t,
aResult: *mut *const c_char
) -> nsresult

/**
   * Returns a shared string pointer
   */

[noscript] void getSharedUTF8String (in unsigned long aIndex, out unsigned long aLength, [shared, retval] out string aResult);

pub unsafe fn GetSharedString(
&self,
aIndex: uint32_t,
aLength: *mut uint32_t,
aResult: *mut *const int16_t
) -> nsresult

[noscript] void getSharedString (in unsigned long aIndex, out unsigned long aLength, [shared, retval] out wstring aResult);