Struct xpcom::interfaces::mozIVisitInfoCallback

#[repr(C)]

pub struct mozIVisitInfoCallback { /* fields omitted */ }

interface mozIVisitInfoCallback : nsISupports

/**
 * Shared Callback interface for mozIAsyncHistory methods. The semantics
 * for each method are detailed in mozIAsyncHistory.
 */

Methods

impl mozIVisitInfoCallback

pub fn coerce<T: mozIVisitInfoCallbackCoerce>(&self) -> &T

Cast this mozIVisitInfoCallback to one of its base interfaces.

impl mozIVisitInfoCallback

pub unsafe fn HandleError(
&self,
aResultCode: nsresult,
aPlaceInfo: *const mozIPlaceInfo
) -> nsresult

/**
   * Called when the given place could not be processed.
   *
   * @param aResultCode
   *        nsresult indicating the failure reason.
   * @param aPlaceInfo
   *        The information that was given to the caller for the place.
   */

void handleError (in nsresult aResultCode, in mozIPlaceInfo aPlaceInfo);

pub unsafe fn HandleResult(&self, aPlaceInfo: *const mozIPlaceInfo) -> nsresult

/**
   * Called for each place processed successfully.
   *
   * @param aPlaceInfo
   *        The current info stored for the place.
   */

void handleResult (in mozIPlaceInfo aPlaceInfo);

pub unsafe fn HandleCompletion(&self, aUpdatedItems: uint32_t) -> nsresult

/**
   * Called when all records were processed.
   * @param aUpdatedItems
   *        How many items were successfully updated.
   */

void handleCompletion (in unsigned long aUpdatedItems);

pub unsafe fn GetIgnoreResults(&self, aIgnoreResults: *mut bool) -> nsresult

/**
   * These two attributes govern whether we attempt to call
   * handleResult and handleError, respectively, if/once
   * results/errors occur.
   */

readonly attribute bool ignoreResults;

pub unsafe fn GetIgnoreErrors(&self, aIgnoreErrors: *mut bool) -> nsresult

readonly attribute bool ignoreErrors;