Struct xpcom::interfaces::nsIASN1Object

#[repr(C)]

pub struct nsIASN1Object { /* fields omitted */ }

interface nsIASN1Object : nsISupports

/**
 * This represents an ASN.1 object,
 * where ASN.1 is "Abstract Syntax Notation number One".
 *
 * The additional state information carried in this interface
 * makes it fit for being used as the data structure
 * when working with visual reprenstation of ASN.1 objects
 * in a human user interface, like in a tree widget
 * where open/close state of nodes must be remembered.
 */

Methods

impl nsIASN1Object

pub fn coerce<T: nsIASN1ObjectCoerce>(&self) -> &T

Cast this nsIASN1Object to one of its base interfaces.

impl nsIASN1Object

pub const ASN1_END_CONTENTS: i64

ASN1_END_CONTENTS: i64 = 0

/**
   *  Identifiers for the possible types of object.
   */

pub const ASN1_BOOLEAN: i64

ASN1_BOOLEAN: i64 = 1

pub const ASN1_INTEGER: i64

ASN1_INTEGER: i64 = 2

pub const ASN1_BIT_STRING: i64

ASN1_BIT_STRING: i64 = 3

pub const ASN1_OCTET_STRING: i64

ASN1_OCTET_STRING: i64 = 4

pub const ASN1_NULL: i64

ASN1_NULL: i64 = 5

pub const ASN1_OBJECT_ID: i64

ASN1_OBJECT_ID: i64 = 6

pub const ASN1_ENUMERATED: i64

ASN1_ENUMERATED: i64 = 10

pub const ASN1_UTF8_STRING: i64

ASN1_UTF8_STRING: i64 = 12

pub const ASN1_SEQUENCE: i64

ASN1_SEQUENCE: i64 = 16

pub const ASN1_SET: i64

ASN1_SET: i64 = 17

pub const ASN1_PRINTABLE_STRING: i64

ASN1_PRINTABLE_STRING: i64 = 19

pub const ASN1_T61_STRING: i64

ASN1_T61_STRING: i64 = 20

pub const ASN1_IA5_STRING: i64

ASN1_IA5_STRING: i64 = 22

pub const ASN1_UTC_TIME: i64

ASN1_UTC_TIME: i64 = 23

pub const ASN1_GEN_TIME: i64

ASN1_GEN_TIME: i64 = 24

pub const ASN1_VISIBLE_STRING: i64

ASN1_VISIBLE_STRING: i64 = 26

pub const ASN1_UNIVERSAL_STRING: i64

ASN1_UNIVERSAL_STRING: i64 = 28

pub const ASN1_BMP_STRING: i64

ASN1_BMP_STRING: i64 = 30

pub const ASN1_HIGH_TAG_NUMBER: i64

ASN1_HIGH_TAG_NUMBER: i64 = 31

pub const ASN1_CONTEXT_SPECIFIC: i64

ASN1_CONTEXT_SPECIFIC: i64 = 32

pub const ASN1_APPLICATION: i64

ASN1_APPLICATION: i64 = 33

pub const ASN1_PRIVATE: i64

ASN1_PRIVATE: i64 = 34

pub unsafe fn GetType(&self, aType: *mut uint32_t) -> nsresult

/**
   *  "type" will be equal to one of the defined object identifiers.
   */

attribute unsigned long type;

pub unsafe fn SetType(&self, aType: uint32_t) -> nsresult

/**
   *  "type" will be equal to one of the defined object identifiers.
   */

attribute unsigned long type;

pub unsafe fn GetTag(&self, aTag: *mut uint32_t) -> nsresult

/**
   *  This contains a tag as explained in ASN.1 standards documents.
   */

attribute unsigned long tag;

pub unsafe fn SetTag(&self, aTag: uint32_t) -> nsresult

/**
   *  This contains a tag as explained in ASN.1 standards documents.
   */

attribute unsigned long tag;

pub unsafe fn GetDisplayName(&self, aDisplayName: &mut nsAString) -> nsresult

/**
   *  "displayName" contains a human readable explanatory label.
   */

attribute AString displayName;

pub unsafe fn SetDisplayName(&self, aDisplayName: &nsAString) -> nsresult

/**
   *  "displayName" contains a human readable explanatory label.
   */

attribute AString displayName;

pub unsafe fn GetDisplayValue(&self, aDisplayValue: &mut nsAString) -> nsresult

/**
   *  "displayValue" contains the human readable value.
   */

attribute AString displayValue;

pub unsafe fn SetDisplayValue(&self, aDisplayValue: &nsAString) -> nsresult

/**
   *  "displayValue" contains the human readable value.
   */

attribute AString displayValue;