Struct xpcom::interfaces::nsIASN1Sequence

#[repr(C)]

pub struct nsIASN1Sequence { /* fields omitted */ }

interface nsIASN1Sequence : nsIASN1Object

/**
 * This represents a sequence of ASN.1 objects,
 * where ASN.1 is "Abstract Syntax Notation number One".
 *
 * Overview of how this ASN1 interface is intended to
 * work.
 *
 * First off, the nsIASN1Sequence is any type in ASN1
 * that consists of sub-elements (ie SEQUENCE, SET)
 * nsIASN1Printable Items are all the other types that
 * can be viewed by themselves without interpreting further.
 * Examples would include INTEGER, UTF-8 STRING, OID.
 * These are not intended to directly reflect the numberous
 * types that exist in ASN1, but merely an interface to ease
 * producing a tree display the ASN1 structure of any DER
 * object.
 *
 * The additional state information carried in this interface
 * makes it fit for being used as the data structure
 * when working with visual reprenstation of ASN.1 objects
 * in a human user interface, like in a tree widget
 * where open/close state of nodes must be remembered.
 */

Methods

impl nsIASN1Sequence

pub fn coerce<T: nsIASN1SequenceCoerce>(&self) -> &T

Cast this nsIASN1Sequence to one of its base interfaces.

impl nsIASN1Sequence

pub unsafe fn GetASN1Objects(
&self,
aASN1Objects: *mut *const nsIMutableArray
) -> nsresult

/**
   *  The array of objects stored in the sequence.
   */

attribute nsIMutableArray ASN1Objects;

pub unsafe fn SetASN1Objects(
&self,
aASN1Objects: *const nsIMutableArray
) -> nsresult

/**
   *  The array of objects stored in the sequence.
   */

attribute nsIMutableArray ASN1Objects;

pub unsafe fn GetIsValidContainer(
&self,
aIsValidContainer: *mut bool
) -> nsresult

/**
   *  Whether the node at this position in the ASN.1 data structure
   *  sequence contains sub elements understood by the
   *  application.
   */

attribute boolean isValidContainer;

pub unsafe fn SetIsValidContainer(&self, aIsValidContainer: bool) -> nsresult

/**
   *  Whether the node at this position in the ASN.1 data structure
   *  sequence contains sub elements understood by the
   *  application.
   */

attribute boolean isValidContainer;

pub unsafe fn GetIsExpanded(&self, aIsExpanded: *mut bool) -> nsresult

/**
   *  Whether the contained objects should be shown or hidden.
   *  A UI implementation can use this flag to store the current
   *  expansion state when shown in a tree widget.
   */

attribute boolean isExpanded;

pub unsafe fn SetIsExpanded(&self, aIsExpanded: bool) -> nsresult

/**
   *  Whether the contained objects should be shown or hidden.
   *  A UI implementation can use this flag to store the current
   *  expansion state when shown in a tree widget.
   */

attribute boolean isExpanded;