Struct xpcom::interfaces::nsIAboutModule

#[repr(C)]

pub struct nsIAboutModule { /* fields omitted */ }

interface nsIAboutModule : nsISupports

Methods

impl nsIAboutModule

pub fn coerce<T: nsIAboutModuleCoerce>(&self) -> &T

Cast this nsIAboutModule to one of its base interfaces.

impl nsIAboutModule

pub const URI_SAFE_FOR_UNTRUSTED_CONTENT: i64

URI_SAFE_FOR_UNTRUSTED_CONTENT: i64 = 1

/**
     * A flag that indicates whether a URI should be run with content
     * privileges. If it is, the about: protocol handler will enforce that
     * the principal of channels created for it be based on their
     * originalURI or URI (depending on the channel flags), by setting
     * their "owner" to null.
     * If content needs to be able to link to this URI, specify
     * URI_CONTENT_LINKABLE as well.
     */

pub const ALLOW_SCRIPT: i64

ALLOW_SCRIPT: i64 = 2

/**
     * A flag that indicates whether script should be enabled for the
     * given about: URI even if it's disabled in general.
     */

pub const HIDE_FROM_ABOUTABOUT: i64

HIDE_FROM_ABOUTABOUT: i64 = 4

/**
     * A flag that indicates whether this about: URI doesn't want to be listed
     * in about:about, especially if it's not useful without a query string.
     */

pub const ENABLE_INDEXED_DB: i64

ENABLE_INDEXED_DB: i64 = 8

/**
     * A flag that indicates whether this about: URI wants Indexed DB enabled.
     */

pub const URI_CAN_LOAD_IN_CHILD: i64

URI_CAN_LOAD_IN_CHILD: i64 = 16

/**
     * A flag that indicates that this URI can be loaded in a child process
     */

pub const URI_MUST_LOAD_IN_CHILD: i64

URI_MUST_LOAD_IN_CHILD: i64 = 32

/**
     * A flag that indicates that this URI must be loaded in a child process
     */

pub const MAKE_UNLINKABLE: i64

MAKE_UNLINKABLE: i64 = 64

/**
     * Obsolete. This flag no longer has any effect and will be removed in future.
     */

pub const MAKE_LINKABLE: i64

MAKE_LINKABLE: i64 = 128

/**
     * A flag that indicates that this URI should be linkable from content.
     * Ignored unless URI_SAFE_FOR_UNTRUSTED_CONTENT is also specified.
     */

pub unsafe fn NewChannel(
&self,
aURI: *const nsIURI,
aLoadInfo: *const nsILoadInfo,
_retval: *mut *const nsIChannel
) -> nsresult

/**
     * Constructs a new channel for the about protocol module.
     *
     * @param aURI the uri of the new channel
     * @param aLoadInfo the loadinfo of the new channel
     */

nsIChannel newChannel (in nsIURI aURI, in nsILoadInfo aLoadInfo);

pub unsafe fn GetURIFlags(
&self,
aURI: *const nsIURI,
_retval: *mut uint32_t
) -> nsresult

/**
     * A method to get the flags that apply to a given about: URI.  The URI
     * passed in is guaranteed to be one of the URIs that this module
     * registered to deal with.
     */

unsigned long getURIFlags (in nsIURI aURI);