Struct xpcom::interfaces::nsIAboutNewTabService

#[repr(C)]

pub struct nsIAboutNewTabService { /* fields omitted */ }

interface nsIAboutNewTabService : nsISupports

/**
 * Allows to override about:newtab to point to a different location
 * than the one specified within AboutRedirector.cpp
 */

Methods

impl nsIAboutNewTabService

pub fn coerce<T: nsIAboutNewTabServiceCoerce>(&self) -> &T

Cast this nsIAboutNewTabService to one of its base interfaces.

impl nsIAboutNewTabService

pub unsafe fn GetNewTabURL(&self, aNewTabURL: &mut nsACString) -> nsresult

/**
   * Returns the url of the resource for the newtab page if not overridden,
   * otherwise a string represenation of the new URL.
   */

attribute ACString newTabURL;

pub unsafe fn SetNewTabURL(&self, aNewTabURL: &nsACString) -> nsresult

/**
   * Returns the url of the resource for the newtab page if not overridden,
   * otherwise a string represenation of the new URL.
   */

attribute ACString newTabURL;

pub unsafe fn GetDefaultURL(&self, aDefaultURL: &mut nsACString) -> nsresult

/**
   * Returns the default URL (local or activity stream depending on pref)
   */

attribute ACString defaultURL;

pub unsafe fn SetDefaultURL(&self, aDefaultURL: &nsACString) -> nsresult

/**
   * Returns the default URL (local or activity stream depending on pref)
   */

attribute ACString defaultURL;

pub unsafe fn GetWillNotifyUser(&self, aWillNotifyUser: *mut bool) -> nsresult

/**
   * Returns true if opening the New Tab page will notify the user of a change.
   */

attribute bool willNotifyUser;

pub unsafe fn SetWillNotifyUser(&self, aWillNotifyUser: bool) -> nsresult

/**
   * Returns true if opening the New Tab page will notify the user of a change.
   */

attribute bool willNotifyUser;

pub unsafe fn GetOverridden(&self, aOverridden: *mut bool) -> nsresult

/**
   * Returns true if the default resource got overridden.
   */

readonly attribute bool overridden;

pub unsafe fn GetActivityStreamEnabled(
&self,
aActivityStreamEnabled: *mut bool
) -> nsresult

/**
   * Returns true if the default resource is activity stream and isn't
   * overridden
   */

readonly attribute bool activityStreamEnabled;

pub unsafe fn GetActivityStreamPrerender(
&self,
aActivityStreamPrerender: *mut bool
) -> nsresult

/**
   * Returns true if the the prerendering pref for activity stream is true
   */

readonly attribute bool activityStreamPrerender;

pub unsafe fn GetActivityStreamDebug(
&self,
aActivityStreamDebug: *mut bool
) -> nsresult

/**
   * Returns true if the the debug pref for activity stream is true
   */

readonly attribute bool activityStreamDebug;

pub unsafe fn ResetNewTabURL(&self) -> nsresult

/**
   * Resets to the default resource and also resets the
   * overridden attribute to false.
   */

void resetNewTabURL ();