Struct xpcom::interfaces::nsIAccessibilityService

#[repr(C)]

pub struct nsIAccessibilityService { /* fields omitted */ }

interface nsIAccessibilityService : nsISupports

/**
 * An interface for in-process accessibility clients wishing to get an
 * nsIAccessible for a given DOM node.  More documentation at:
 *   http://www.mozilla.org/projects/ui/accessibility
 */

Methods

impl nsIAccessibilityService

pub fn coerce<T: nsIAccessibilityServiceCoerce>(&self) -> &T

Cast this nsIAccessibilityService to one of its base interfaces.

impl nsIAccessibilityService

pub unsafe fn GetApplicationAccessible(
&self,
_retval: *mut *const nsIAccessible
) -> nsresult

/**
   * Return application accessible.
   */

nsIAccessible getApplicationAccessible ();

pub unsafe fn GetAccessibleFor(
&self,
aNode: *const nsIDOMNode,
_retval: *mut *const nsIAccessible
) -> nsresult

/**
   * Return an nsIAccessible for a DOM node in pres shell 0.
   * Create a new accessible of the appropriate type if necessary,
   * or use one from the accessibility cache if it already exists.
   * @param aNode The DOM node to get an accessible for.
   * @return The nsIAccessible for the given DOM node.
   */

nsIAccessible getAccessibleFor (in nsIDOMNode aNode);

pub unsafe fn GetStringRole(
&self,
aRole: uint32_t,
_retval: &mut nsAString
) -> nsresult

/**
    * Returns accessible role as a string.
    *
    * @param aRole - the accessible role constants.
    */

AString getStringRole (in unsigned long aRole);

pub unsafe fn GetStringStates(
&self,
aStates: uint32_t,
aExtraStates: uint32_t,
_retval: *mut *const nsISupports
) -> nsresult

/**
    * Returns list which contains accessible states as a strings.
    *
    * @param aStates - accessible states.
    * @param aExtraStates - accessible extra states.
    */

nsISupports getStringStates (in unsigned long aStates, in unsigned long aExtraStates);

pub unsafe fn GetStringEventType(
&self,
aEventType: uint32_t,
_retval: &mut nsAString
) -> nsresult

/**
   * Get the type of accessible event as a string.
   *
   * @param aEventType - the accessible event type constant
   * @return - accessible event type presented as human readable string
   */

AString getStringEventType (in unsigned long aEventType);

pub unsafe fn GetStringRelationType(
&self,
aRelationType: uint32_t,
_retval: &mut nsAString
) -> nsresult

/**
   * Get the type of accessible relation as a string.
   *
   * @param aRelationType - the accessible relation type constant
   * @return - accessible relation type presented as human readable string
   */

AString getStringRelationType (in unsigned long aRelationType);

pub unsafe fn GetAccessibleFromCache(
&self,
aNode: *const nsIDOMNode,
_retval: *mut *const nsIAccessible
) -> nsresult

/**
   * Return an accessible for the given DOM node from the cache.
   * @note  the method is intended for testing purposes
   *
   * @param aNode  [in] the DOM node to get an accessible for
   *
   * @return       cached accessible for the given DOM node if any
   */

nsIAccessible getAccessibleFromCache (in nsIDOMNode aNode);

pub unsafe fn CreateAccessiblePivot(
&self,
aRoot: *const nsIAccessible,
_retval: *mut *const nsIAccessiblePivot
) -> nsresult

/**
   * Create a new pivot for tracking a position and traversing a subtree.
   *
   * @param aRoot [in] the accessible root for the pivot
   * @return a new pivot
   */

nsIAccessiblePivot createAccessiblePivot (in nsIAccessible aRoot);

pub unsafe fn SetLogging(&self, aModules: &nsACString) -> nsresult

/**
   * Enable logging for the given modules, all other modules aren't logged.
   *
   * @param aModules [in] list of modules, format is comma separated list
   *                      like 'docload,doccreate'.
   * @note Works on debug build only.
   * @see Logging.cpp for list of possible values.
   */

void setLogging (in ACString aModules);

pub unsafe fn IsLogged(
&self,
aModule: &nsAString,
_retval: *mut bool
) -> nsresult

/**
   * Return true if the given module is logged.
   */

boolean isLogged (in AString aModule);

pub unsafe fn GetConsumers(&self, _retval: &mut nsAString) -> nsresult

/**
   * Get the current accessibility service consumers.
   * @returns a JSON string representing the accessibility service consumers.
   */

AString getConsumers ();