Struct xpcom::interfaces::nsIAccessibleDocument

#[repr(C)]

pub struct nsIAccessibleDocument { /* fields omitted */ }

interface nsIAccessibleDocument : nsISupports

/**
 * An interface for in-process accessibility clients
 * that wish to retrieve information about a document.
 * When accessibility is turned on in Gecko,
 * there is an nsIAccessibleDocument for each document
 * whether it is XUL, HTML or whatever.
 * You can QueryInterface to nsIAccessibleDocument from the nsIAccessible for
 * the root node of a document or you can get one from
 * nsIAccessible::GetDocument().
 */

Methods

impl nsIAccessibleDocument

pub fn coerce<T: nsIAccessibleDocumentCoerce>(&self) -> &T

Cast this nsIAccessibleDocument to one of its base interfaces.

impl nsIAccessibleDocument

pub unsafe fn GetURL(&self, aURL: &mut nsAString) -> nsresult

/**
   * The URL of the document
   */

readonly attribute AString URL;

pub unsafe fn GetTitle(&self, aTitle: &mut nsAString) -> nsresult

/**
   * The title of the document, as specified in the document.
   */

readonly attribute AString title;

pub unsafe fn GetMimeType(&self, aMimeType: &mut nsAString) -> nsresult

/**
   * The mime type of the document
   */

readonly attribute AString mimeType;

pub unsafe fn GetDocType(&self, aDocType: &mut nsAString) -> nsresult

/**
   * The doc type of the document, as specified in the document.
   */

readonly attribute AString docType;

pub unsafe fn GetDOMDocument(
&self,
aDOMDocument: *mut *const nsIDOMDocument
) -> nsresult

/**
   * The nsIDOMDocument interface associated with this document.
   */

readonly attribute nsIDOMDocument DOMDocument;

pub unsafe fn GetWindow(
&self,
aWindow: *mut *const mozIDOMWindowProxy
) -> nsresult

/**
   * The nsIDOMWindow that the document resides in.
   */

readonly attribute mozIDOMWindowProxy window;

pub unsafe fn GetParentDocument(
&self,
aParentDocument: *mut *const nsIAccessibleDocument
) -> nsresult

/**
   * Return the parent document accessible.
   */

readonly attribute nsIAccessibleDocument parentDocument;

pub unsafe fn GetChildDocumentCount(
&self,
aChildDocumentCount: *mut uint32_t
) -> nsresult

/**
   * Return the count of child document accessibles.
   */

readonly attribute unsigned long childDocumentCount;

pub unsafe fn GetVirtualCursor(
&self,
aVirtualCursor: *mut *const nsIAccessiblePivot
) -> nsresult

/**
   * The virtual cursor pivot this document manages.
   */

readonly attribute nsIAccessiblePivot virtualCursor;

pub unsafe fn GetChildDocumentAt(
&self,
index: uint32_t,
_retval: *mut *const nsIAccessibleDocument
) -> nsresult

/**
   * Return the child document accessible at the given index.
   */

nsIAccessibleDocument getChildDocumentAt (in unsigned long index);