Struct xpcom::interfaces::nsIAccessibleEvent

#[repr(C)]

pub struct nsIAccessibleEvent { /* fields omitted */ }

interface nsIAccessibleEvent : nsISupports

/**
 * An interface for accessibility events listened to
 * by in-process accessibility clients, which can be used
 * to find out how to get accessibility and DOM interfaces for
 * the event and its target. To listen to in-process accessibility invents,
 * make your object an nsIObserver, and listen for accessible-event by
 * using code something like this:
 *   nsCOMPtr<nsIObserverService> observerService =
 *     do_GetService("@mozilla.org/observer-service;1", &rv);
 *   if (NS_SUCCEEDED(rv))
 *     rv = observerService->AddObserver(this, "accessible-event", PR_TRUE);
 */

Methods

impl nsIAccessibleEvent

pub fn coerce<T: nsIAccessibleEventCoerce>(&self) -> &T

Cast this nsIAccessibleEvent to one of its base interfaces.

impl nsIAccessibleEvent

pub const EVENT_SHOW: i64

EVENT_SHOW: i64 = 1

/**
   * An object has been created.
   */

pub const EVENT_HIDE: i64

EVENT_HIDE: i64 = 2

/**
   * An object has been destroyed.
   */

pub const EVENT_REORDER: i64

EVENT_REORDER: i64 = 3

/**
   * An object's children have changed
   */

pub const EVENT_ACTIVE_DECENDENT_CHANGED: i64

EVENT_ACTIVE_DECENDENT_CHANGED: i64 = 4

/**
   * The active descendant of a component has changed. The active descendant
   * is used in objects with transient children.
   */

pub const EVENT_FOCUS: i64

EVENT_FOCUS: i64 = 5

/**
   * An object has received the keyboard focus.
   */

pub const EVENT_STATE_CHANGE: i64

EVENT_STATE_CHANGE: i64 = 6

/**
   * An object's state has changed.
   */

pub const EVENT_LOCATION_CHANGE: i64

EVENT_LOCATION_CHANGE: i64 = 7

/**
   * An object has changed location, shape, or size.
   */

pub const EVENT_NAME_CHANGE: i64

EVENT_NAME_CHANGE: i64 = 8

/**
   * An object's Name property has changed.
   */

pub const EVENT_DESCRIPTION_CHANGE: i64

EVENT_DESCRIPTION_CHANGE: i64 = 9

/**
   * An object's Description property has changed.
   */

pub const EVENT_VALUE_CHANGE: i64

EVENT_VALUE_CHANGE: i64 = 10

/**
   * An object's numeric Value has changed.
   */

pub const EVENT_HELP_CHANGE: i64

EVENT_HELP_CHANGE: i64 = 11

/**
   * An object's help has changed.
   */

pub const EVENT_DEFACTION_CHANGE: i64

EVENT_DEFACTION_CHANGE: i64 = 12

/**
   * An object's default action has changed.
   */

pub const EVENT_ACTION_CHANGE: i64

EVENT_ACTION_CHANGE: i64 = 13

/**
   * An object's action has changed.
   */

pub const EVENT_ACCELERATOR_CHANGE: i64

EVENT_ACCELERATOR_CHANGE: i64 = 14

/**
   * An object's keyboard shortcut has changed.
   */

pub const EVENT_SELECTION: i64

EVENT_SELECTION: i64 = 15

/**
   * The selection within a container object has changed.
   */

pub const EVENT_SELECTION_ADD: i64

EVENT_SELECTION_ADD: i64 = 16

/**
   * An item within a container object has been added to the selection.
   */

pub const EVENT_SELECTION_REMOVE: i64

EVENT_SELECTION_REMOVE: i64 = 17

/**
   * An item within a container object has been removed from the selection.
   */

pub const EVENT_SELECTION_WITHIN: i64

EVENT_SELECTION_WITHIN: i64 = 18

/**
   * Numerous selection changes have occurred within a container object.
   */

pub const EVENT_ALERT: i64

EVENT_ALERT: i64 = 19

/**
   * An alert has been generated. Server applications send this event when a
   * user needs to know that a user interface element has changed.
   */

pub const EVENT_FOREGROUND: i64

EVENT_FOREGROUND: i64 = 20

/**
   * The foreground window has changed.
   */

pub const EVENT_MENU_START: i64

EVENT_MENU_START: i64 = 21

/**
   * A menu item on the menu bar has been selected.
   */

pub const EVENT_MENU_END: i64

EVENT_MENU_END: i64 = 22

/**
   * A menu from the menu bar has been closed.
   */

pub const EVENT_MENUPOPUP_START: i64

EVENT_MENUPOPUP_START: i64 = 23

/**
   * A pop-up menu has been displayed.
   */

pub const EVENT_MENUPOPUP_END: i64

EVENT_MENUPOPUP_END: i64 = 24

/**
   * A pop-up menu has been closed.
   */

pub const EVENT_CAPTURE_START: i64

EVENT_CAPTURE_START: i64 = 25

/**
   * A window has received mouse capture.
   */

pub const EVENT_CAPTURE_END: i64

EVENT_CAPTURE_END: i64 = 26

/**
   * A window has lost mouse capture.
   */

pub const EVENT_MOVESIZE_START: i64

EVENT_MOVESIZE_START: i64 = 27

/**
   * A window is being moved or resized.
   */

pub const EVENT_MOVESIZE_END: i64

EVENT_MOVESIZE_END: i64 = 28

/**
  * The movement or resizing of a window has finished
  */

pub const EVENT_CONTEXTHELP_START: i64

EVENT_CONTEXTHELP_START: i64 = 29

/**
   * A window has entered context-sensitive Help mode
   */

pub const EVENT_CONTEXTHELP_END: i64

EVENT_CONTEXTHELP_END: i64 = 30

/**
   * A window has exited context-sensitive Help mode
   */

pub const EVENT_DRAGDROP_START: i64

EVENT_DRAGDROP_START: i64 = 31

/**
   * An application is about to enter drag-and-drop mode
   */

pub const EVENT_DRAGDROP_END: i64

EVENT_DRAGDROP_END: i64 = 32

/**
   * An application is about to exit drag-and-drop mode
   */

pub const EVENT_DIALOG_START: i64

EVENT_DIALOG_START: i64 = 33

/**
   * A dialog box has been displayed
   */

pub const EVENT_DIALOG_END: i64

EVENT_DIALOG_END: i64 = 34

/**
   * A dialog box has been closed
   */

pub const EVENT_SCROLLING_START: i64

EVENT_SCROLLING_START: i64 = 35

/**
   * Scrolling has started on a scroll bar
   */

pub const EVENT_SCROLLING_END: i64

EVENT_SCROLLING_END: i64 = 36

/**
   * Scrolling has ended on a scroll bar
   */

pub const EVENT_MINIMIZE_START: i64

EVENT_MINIMIZE_START: i64 = 37

/**
   * A window object is about to be minimized or maximized
   */

pub const EVENT_MINIMIZE_END: i64

EVENT_MINIMIZE_END: i64 = 38

/**
   * A window object has been minimized or maximized
   */

pub const EVENT_DOCUMENT_LOAD_COMPLETE: i64

EVENT_DOCUMENT_LOAD_COMPLETE: i64 = 39

/**
   * The loading of the document has completed.
   */

pub const EVENT_DOCUMENT_RELOAD: i64

EVENT_DOCUMENT_RELOAD: i64 = 40

/**
   * The document contents are being reloaded.
   */

pub const EVENT_DOCUMENT_LOAD_STOPPED: i64

EVENT_DOCUMENT_LOAD_STOPPED: i64 = 41

/**
   * The loading of the document was interrupted.
   */

pub const EVENT_DOCUMENT_ATTRIBUTES_CHANGED: i64

EVENT_DOCUMENT_ATTRIBUTES_CHANGED: i64 = 42

/**
   * The document wide attributes of the document object have changed.
   */

pub const EVENT_DOCUMENT_CONTENT_CHANGED: i64

EVENT_DOCUMENT_CONTENT_CHANGED: i64 = 43

/**
   * The contents of the document have changed.
   */

pub const EVENT_PROPERTY_CHANGED: i64

EVENT_PROPERTY_CHANGED: i64 = 44

pub const EVENT_PAGE_CHANGED: i64

EVENT_PAGE_CHANGED: i64 = 45

/**
   * A slide changed in a presentation document or a page boundary was
   * crossed in a word processing document.
   */

pub const EVENT_TEXT_ATTRIBUTE_CHANGED: i64

EVENT_TEXT_ATTRIBUTE_CHANGED: i64 = 46

/**
   * A text object's attributes changed.
   * Also see EVENT_OBJECT_ATTRIBUTE_CHANGED.
   */

pub const EVENT_TEXT_CARET_MOVED: i64

EVENT_TEXT_CARET_MOVED: i64 = 47

/**
   * The caret has moved to a new position.
   */

pub const EVENT_TEXT_CHANGED: i64

EVENT_TEXT_CHANGED: i64 = 48

/**
   * This event indicates general text changes, i.e. changes to text that is
   * exposed through the IAccessibleText and IAccessibleEditableText interfaces.
   */

pub const EVENT_TEXT_INSERTED: i64

EVENT_TEXT_INSERTED: i64 = 49

/**
   * Text was inserted.
   */

pub const EVENT_TEXT_REMOVED: i64

EVENT_TEXT_REMOVED: i64 = 50

/**
   * Text was removed.
   */

pub const EVENT_TEXT_UPDATED: i64

EVENT_TEXT_UPDATED: i64 = 51

/**
   * Text was updated.
   */

pub const EVENT_TEXT_SELECTION_CHANGED: i64

EVENT_TEXT_SELECTION_CHANGED: i64 = 52

/**
   * The text selection changed.
   */

pub const EVENT_VISIBLE_DATA_CHANGED: i64

EVENT_VISIBLE_DATA_CHANGED: i64 = 53

/**
   * A visibile data event indicates the change of the visual appearance
   * of an accessible object.  This includes for example most of the
   * attributes available via the IAccessibleComponent interface.
   */

pub const EVENT_TEXT_COLUMN_CHANGED: i64

EVENT_TEXT_COLUMN_CHANGED: i64 = 54

/**
   * The caret moved from one column to the next.
   */

pub const EVENT_SECTION_CHANGED: i64

EVENT_SECTION_CHANGED: i64 = 55

/**
   * The caret moved from one section to the next.
   */

pub const EVENT_TABLE_CAPTION_CHANGED: i64

EVENT_TABLE_CAPTION_CHANGED: i64 = 56

/**
   * A table caption changed.
   */

pub const EVENT_TABLE_MODEL_CHANGED: i64

EVENT_TABLE_MODEL_CHANGED: i64 = 57

/**
   * A table's data changed.
   */

pub const EVENT_TABLE_SUMMARY_CHANGED: i64

EVENT_TABLE_SUMMARY_CHANGED: i64 = 58

/**
   * A table's summary changed.
   */

pub const EVENT_TABLE_ROW_DESCRIPTION_CHANGED: i64

EVENT_TABLE_ROW_DESCRIPTION_CHANGED: i64 = 59

/**
   * A table's row description changed.
   */

pub const EVENT_TABLE_ROW_HEADER_CHANGED: i64

EVENT_TABLE_ROW_HEADER_CHANGED: i64 = 60

/**
   * A table's row header changed.
   */

pub const EVENT_TABLE_ROW_INSERT: i64

EVENT_TABLE_ROW_INSERT: i64 = 61

pub const EVENT_TABLE_ROW_DELETE: i64

EVENT_TABLE_ROW_DELETE: i64 = 62

pub const EVENT_TABLE_ROW_REORDER: i64

EVENT_TABLE_ROW_REORDER: i64 = 63

pub const EVENT_TABLE_COLUMN_DESCRIPTION_CHANGED: i64

EVENT_TABLE_COLUMN_DESCRIPTION_CHANGED: i64 = 64

/**
   * A table's column description changed.
   */

pub const EVENT_TABLE_COLUMN_HEADER_CHANGED: i64

EVENT_TABLE_COLUMN_HEADER_CHANGED: i64 = 65

/**
   * A table's column header changed.
   */

pub const EVENT_TABLE_COLUMN_INSERT: i64

EVENT_TABLE_COLUMN_INSERT: i64 = 66

pub const EVENT_TABLE_COLUMN_DELETE: i64

EVENT_TABLE_COLUMN_DELETE: i64 = 67

pub const EVENT_TABLE_COLUMN_REORDER: i64

EVENT_TABLE_COLUMN_REORDER: i64 = 68

pub const EVENT_WINDOW_ACTIVATE: i64

EVENT_WINDOW_ACTIVATE: i64 = 69

pub const EVENT_WINDOW_CREATE: i64

EVENT_WINDOW_CREATE: i64 = 70

pub const EVENT_WINDOW_DEACTIVATE: i64

EVENT_WINDOW_DEACTIVATE: i64 = 71

pub const EVENT_WINDOW_DESTROY: i64

EVENT_WINDOW_DESTROY: i64 = 72

pub const EVENT_WINDOW_MAXIMIZE: i64

EVENT_WINDOW_MAXIMIZE: i64 = 73

pub const EVENT_WINDOW_MINIMIZE: i64

EVENT_WINDOW_MINIMIZE: i64 = 74

pub const EVENT_WINDOW_RESIZE: i64

EVENT_WINDOW_RESIZE: i64 = 75

pub const EVENT_WINDOW_RESTORE: i64

EVENT_WINDOW_RESTORE: i64 = 76

pub const EVENT_HYPERLINK_END_INDEX_CHANGED: i64

EVENT_HYPERLINK_END_INDEX_CHANGED: i64 = 77

/**
   * The ending index of this link within the containing string has changed.
   */

pub const EVENT_HYPERLINK_NUMBER_OF_ANCHORS_CHANGED: i64

EVENT_HYPERLINK_NUMBER_OF_ANCHORS_CHANGED: i64 = 78

/**
   * The number of anchors assoicated with this hyperlink object has changed.
   */

pub const EVENT_HYPERLINK_SELECTED_LINK_CHANGED: i64

EVENT_HYPERLINK_SELECTED_LINK_CHANGED: i64 = 79

/**
   * The hyperlink selected state changed from selected to unselected or
   * from unselected to selected.
   */

pub const EVENT_HYPERTEXT_LINK_ACTIVATED: i64

EVENT_HYPERTEXT_LINK_ACTIVATED: i64 = 80

/**
   * One of the links associated with the hypertext object has been activated.
   */

pub const EVENT_HYPERTEXT_LINK_SELECTED: i64

EVENT_HYPERTEXT_LINK_SELECTED: i64 = 81

/**
   * One of the links associated with the hypertext object has been selected.
   */

pub const EVENT_HYPERLINK_START_INDEX_CHANGED: i64

EVENT_HYPERLINK_START_INDEX_CHANGED: i64 = 82

/**
   * The starting index of this link within the containing string has changed.
   */

pub const EVENT_HYPERTEXT_CHANGED: i64

EVENT_HYPERTEXT_CHANGED: i64 = 83

/**
   * Focus has changed from one hypertext object to another, or focus moved
   * from a non-hypertext object to a hypertext object, or focus moved from a
   * hypertext object to a non-hypertext object.
   */

pub const EVENT_HYPERTEXT_NLINKS_CHANGED: i64

EVENT_HYPERTEXT_NLINKS_CHANGED: i64 = 84

/**
   * The number of hyperlinks associated with a hypertext object changed.
   */

pub const EVENT_OBJECT_ATTRIBUTE_CHANGED: i64

EVENT_OBJECT_ATTRIBUTE_CHANGED: i64 = 85

/**
   * An object's attributes changed. Also see EVENT_TEXT_ATTRIBUTE_CHANGED.
   */

pub const EVENT_VIRTUALCURSOR_CHANGED: i64

EVENT_VIRTUALCURSOR_CHANGED: i64 = 86

/**
   * A cursorable's virtual cursor has changed.
   */

pub const EVENT_TEXT_VALUE_CHANGE: i64

EVENT_TEXT_VALUE_CHANGE: i64 = 87

/**
   * An object's text Value has changed.
   */

pub const EVENT_LAST_ENTRY: i64

EVENT_LAST_ENTRY: i64 = 88

/**
   * Help make sure event map does not get out-of-line.
   */

pub unsafe fn GetEventType(&self, aEventType: *mut uint32_t) -> nsresult

/**
   * The type of event, based on the enumerated event values
   * defined in this interface.
   */

readonly attribute unsigned long eventType;

pub unsafe fn GetAccessible(
&self,
aAccessible: *mut *const nsIAccessible
) -> nsresult

/**
   * The nsIAccessible associated with the event.
   * May return null if no accessible is available
   */

readonly attribute nsIAccessible accessible;

pub unsafe fn GetAccessibleDocument(
&self,
aAccessibleDocument: *mut *const nsIAccessibleDocument
) -> nsresult

/**
   * The nsIAccessibleDocument that the event target nsIAccessible
   * resides in. This can be used to get the DOM window,
   * the DOM document and the window handler, among other things.
   */

readonly attribute nsIAccessibleDocument accessibleDocument;

pub unsafe fn GetDOMNode(&self, aDOMNode: *mut *const nsIDOMNode) -> nsresult

/**
   * The nsIDOMNode associated with the event
   * May return null if accessible for event has been shut down
   */

readonly attribute nsIDOMNode DOMNode;

pub unsafe fn GetIsFromUserInput(&self, aIsFromUserInput: *mut bool) -> nsresult

/**
   * Returns true if the event was caused by explicit user input,
   * as opposed to purely originating from a timer or mouse movement
   */

readonly attribute boolean isFromUserInput;