Struct xpcom::interfaces::nsIAccessibleHyperLink

#[repr(C)]

pub struct nsIAccessibleHyperLink { /* fields omitted */ }

interface nsIAccessibleHyperLink : nsISupports

/**
 * A cross-platform interface that supports hyperlink-specific properties and
 * methods.  Anchors, image maps, xul:labels with class="text-link" implement this interface.
 */

Methods

impl nsIAccessibleHyperLink

pub fn coerce<T: nsIAccessibleHyperLinkCoerce>(&self) -> &T

Cast this nsIAccessibleHyperLink to one of its base interfaces.

impl nsIAccessibleHyperLink

pub unsafe fn GetStartIndex(&self, aStartIndex: *mut int32_t) -> nsresult

/**
   * Returns the offset of the link within the parent accessible.
   */

readonly attribute long startIndex;

pub unsafe fn GetEndIndex(&self, aEndIndex: *mut int32_t) -> nsresult

/**
   * Returns the end index of the link within the parent accessible.
   *
   * @note  The link itself is represented by one embedded character within the
   * parent text, so the endIndex should be startIndex + 1.
   */

readonly attribute long endIndex;

pub unsafe fn GetValid(&self, aValid: *mut bool) -> nsresult

/**
   * Determines whether the link is valid (e. g. points to a valid URL).
   *
   * @note  XXX Currently only used with ARIA links, and the author has to
   * specify that the link is invalid via the aria-invalid="true" attribute.
   * In all other cases, TRUE is returned.
   */

readonly attribute boolean valid;

pub unsafe fn GetAnchorCount(&self, aAnchorCount: *mut int32_t) -> nsresult

/**
   * The numbber of anchors within this Hyperlink. Is normally 1 for anchors.
   * This anchor is, for example, the visible output of the html:a tag.
   * With an Image Map, reflects the actual areas within the map.
   */

readonly attribute long anchorCount;

pub unsafe fn GetURI(
&self,
index: int32_t,
_retval: *mut *const nsIURI
) -> nsresult

/**
   * Returns the URI at the given index.
   *
   * @note  ARIA hyperlinks do not have an URI to point to, since clicks are
   * processed via JavaScript. Therefore this property does not work on ARIA
   * links.
   *
   * @param index  The 0-based index of the URI to be returned.
   *
   * @return the nsIURI object containing the specifications for the URI.
   */

nsIURI getURI (in long index);

pub unsafe fn GetAnchor(
&self,
index: int32_t,
_retval: *mut *const nsIAccessible
) -> nsresult

/**
   * Returns a reference to the object at the given index.
   *
   * @param index  The 0-based index whose object is to be returned.
   *
   * @return the nsIAccessible object at the desired index.
   */

nsIAccessible getAnchor (in long index);