Struct xpcom::interfaces::nsIAccessiblePivot

#[repr(C)]

pub struct nsIAccessiblePivot { /* fields omitted */ }

interface nsIAccessiblePivot : nsISupports

/**
 * The pivot interface encapsulates a reference to a single place in an accessible
 * subtree. The pivot is a point or a range in the accessible tree. This interface
 * provides traversal methods to move the pivot to next/prev state that complies
 * to a given rule.
 */

Methods

impl nsIAccessiblePivot

pub fn coerce<T: nsIAccessiblePivotCoerce>(&self) -> &T

Cast this nsIAccessiblePivot to one of its base interfaces.

impl nsIAccessiblePivot

pub const CHAR_BOUNDARY: i64

CHAR_BOUNDARY: i64 = 0

pub const WORD_BOUNDARY: i64

WORD_BOUNDARY: i64 = 1

pub const LINE_BOUNDARY: i64

LINE_BOUNDARY: i64 = 2

pub const ATTRIBUTE_RANGE_BOUNDARY: i64

ATTRIBUTE_RANGE_BOUNDARY: i64 = 3

pub const REASON_NONE: i64

REASON_NONE: i64 = 0

pub const REASON_NEXT: i64

REASON_NEXT: i64 = 1

pub const REASON_PREV: i64

REASON_PREV: i64 = 2

pub const REASON_FIRST: i64

REASON_FIRST: i64 = 3

pub const REASON_LAST: i64

REASON_LAST: i64 = 4

pub const REASON_TEXT: i64

REASON_TEXT: i64 = 5

pub const REASON_POINT: i64

REASON_POINT: i64 = 6

pub unsafe fn GetPosition(
&self,
aPosition: *mut *const nsIAccessible
) -> nsresult

/**
   * The accessible the pivot is currently pointed at.
   */

attribute nsIAccessible position;

pub unsafe fn SetPosition(&self, aPosition: *const nsIAccessible) -> nsresult

/**
   * The accessible the pivot is currently pointed at.
   */

attribute nsIAccessible position;

pub unsafe fn GetRoot(&self, aRoot: *mut *const nsIAccessible) -> nsresult

/**
   * The root of the subtree in which the pivot traverses.
   */

readonly attribute nsIAccessible root;

pub unsafe fn GetModalRoot(
&self,
aModalRoot: *mut *const nsIAccessible
) -> nsresult

/**
   * The temporary modal root to which traversal is limited to.
   */

attribute nsIAccessible modalRoot;

pub unsafe fn SetModalRoot(&self, aModalRoot: *const nsIAccessible) -> nsresult

/**
   * The temporary modal root to which traversal is limited to.
   */

attribute nsIAccessible modalRoot;

pub unsafe fn GetStartOffset(&self, aStartOffset: *mut int32_t) -> nsresult

/**
   * The start offset of the text range the pivot points at, otherwise -1.
   */

readonly attribute long startOffset;

pub unsafe fn GetEndOffset(&self, aEndOffset: *mut int32_t) -> nsresult

/**
   * The end offset of the text range the pivot points at, otherwise -1.
   */

readonly attribute long endOffset;

pub unsafe fn AddObserver(
&self,
aObserver: *const nsIAccessiblePivotObserver
) -> nsresult

/**
   * Add an observer for pivot changes.
   *
   * @param aObserver [in] the observer object to be notified of pivot changes.
   */

void addObserver (in nsIAccessiblePivotObserver aObserver);

pub unsafe fn RemoveObserver(
&self,
aObserver: *const nsIAccessiblePivotObserver
) -> nsresult

/**
   * Remove an observer for pivot changes.
   *
   * @param aObserver [in] the observer object to remove from being notified.
   */

void removeObserver (in nsIAccessiblePivotObserver aObserver);