Struct xpcom::interfaces::nsIAccessibleRelation

#[repr(C)]

pub struct nsIAccessibleRelation { /* fields omitted */ }

interface nsIAccessibleRelation : nsISupports

/**
 * This interface gives access to an accessible's set of relations.
 */

Methods

impl nsIAccessibleRelation

pub fn coerce<T: nsIAccessibleRelationCoerce>(&self) -> &T

Cast this nsIAccessibleRelation to one of its base interfaces.

impl nsIAccessibleRelation

pub const RELATION_LABELLED_BY: i64

RELATION_LABELLED_BY: i64 = 0

/**
   * This object is labelled by a target object.
   */

pub const RELATION_LABEL_FOR: i64

RELATION_LABEL_FOR: i64 = 1

/**
   * This object is label for a target object.
   */

pub const RELATION_DESCRIBED_BY: i64

RELATION_DESCRIBED_BY: i64 = 2

/**
   * This object is described by the target object.
   */

pub const RELATION_DESCRIPTION_FOR: i64

RELATION_DESCRIPTION_FOR: i64 = 3

/**
   * This object is describes the target object.
   */

pub const RELATION_NODE_CHILD_OF: i64

RELATION_NODE_CHILD_OF: i64 = 4

/**
   * This object is a child of a target object.
   */

pub const RELATION_NODE_PARENT_OF: i64

RELATION_NODE_PARENT_OF: i64 = 5

/**
   * This object is a parent of a target object. A dual relation to
   * RELATION_NODE_CHILD_OF
   */

pub const RELATION_CONTROLLED_BY: i64

RELATION_CONTROLLED_BY: i64 = 6

/**
   * Some attribute of this object is affected by a target object.
   */

pub const RELATION_CONTROLLER_FOR: i64

RELATION_CONTROLLER_FOR: i64 = 7

/**
   * This object is interactive and controls some attribute of a target object.
   */

pub const RELATION_FLOWS_TO: i64

RELATION_FLOWS_TO: i64 = 8

/**
   * Content flows from this object to a target object, i.e. has content that
   * flows logically to another object in a sequential way, e.g. text flow.
   */

pub const RELATION_FLOWS_FROM: i64

RELATION_FLOWS_FROM: i64 = 9

/**
   * Content flows to this object from a target object, i.e. has content that
   * flows logically from another object in a sequential way, e.g. text flow.
   */

pub const RELATION_MEMBER_OF: i64

RELATION_MEMBER_OF: i64 = 10

/**
   * This object is a member of a group of one or more objects. When there is
   * more than one object in the group each member may have one and the same
   * target, e.g. a grouping object.  It is also possible that each member has
   * multiple additional targets, e.g. one for every other member in the group.
   */

pub const RELATION_SUBWINDOW_OF: i64

RELATION_SUBWINDOW_OF: i64 = 11

/**
   * This object is a sub window of a target object.
   */

pub const RELATION_EMBEDS: i64

RELATION_EMBEDS: i64 = 12

/**
   * This object embeds a target object. This relation can be used on the
   * OBJID_CLIENT accessible for a top level window to show where the content
   * areas are.
   */

pub const RELATION_EMBEDDED_BY: i64

RELATION_EMBEDDED_BY: i64 = 13

/**
   * This object is embedded by a target object.
   */

pub const RELATION_POPUP_FOR: i64

RELATION_POPUP_FOR: i64 = 14

/**
   * This object is a transient component related to the target object. When
   * this object is activated the target object doesn't lose focus.
   */

pub const RELATION_PARENT_WINDOW_OF: i64

RELATION_PARENT_WINDOW_OF: i64 = 15

/**
   * This object is a parent window of the target object.
   */

pub const RELATION_DEFAULT_BUTTON: i64

RELATION_DEFAULT_BUTTON: i64 = 16

/**
   * Part of a form/dialog with a related default button. It is used for
   * MSAA/XPCOM, it isn't for IA2 or ATK.
   */

pub const RELATION_CONTAINING_DOCUMENT: i64

RELATION_CONTAINING_DOCUMENT: i64 = 17

/**
   * The target object is the containing document object.
   */

pub const RELATION_CONTAINING_TAB_PANE: i64

RELATION_CONTAINING_TAB_PANE: i64 = 18

/**
   * The target object is the topmost containing document object in the tab pane.
   */

pub const RELATION_CONTAINING_WINDOW: i64

RELATION_CONTAINING_WINDOW: i64 = 19

/**
   * The target object is the containing window object.
   */

pub const RELATION_CONTAINING_APPLICATION: i64

RELATION_CONTAINING_APPLICATION: i64 = 20

/**
   * The target object is the containing application object.
   */

pub const RELATION_DETAILS: i64

RELATION_DETAILS: i64 = 21

/**
   * The target object provides the detailed, extended description for this
   * object. It provides more detailed information than would normally be
   * provided using the DESCRIBED_BY relation. A common use for this relation is
   * in digital publishing where an extended description needs to be conveyed in
   * a book that requires structural markup or the embedding of other technology
   * to provide illustrative content.
   */

pub const RELATION_DETAILS_FOR: i64

RELATION_DETAILS_FOR: i64 = 22

/**
   * This object provides the detailed, extended description for the target
   * object. See DETAILS relation.
   */

pub const RELATION_ERRORMSG: i64

RELATION_ERRORMSG: i64 = 23

/**
   * The target object is the error message for this object.
   */

pub const RELATION_ERRORMSG_FOR: i64

RELATION_ERRORMSG_FOR: i64 = 24

/**
   * This object is the error message for the target object.
   */

pub unsafe fn GetRelationType(&self, aRelationType: *mut uint32_t) -> nsresult

/**
   * Returns the type of the relation.
   */

readonly attribute unsigned long relationType;

pub unsafe fn GetTargetsCount(&self, aTargetsCount: *mut uint32_t) -> nsresult

/**
   * Returns the number of targets for this relation.
   */

readonly attribute unsigned long targetsCount;

pub unsafe fn GetTarget(
&self,
index: uint32_t,
_retval: *mut *const nsIAccessible
) -> nsresult

/**
   * Returns one accessible relation target.
   * @param index - 0 based index of relation target.
   */

nsIAccessible getTarget (in unsigned long index);

pub unsafe fn GetTargets(&self, _retval: *mut *const nsIArray) -> nsresult

/**
   * Returns multiple accessible relation targets.
   */

nsIArray getTargets ();