Struct xpcom::interfaces::nsIAccessibleTextRange

#[repr(C)]

pub struct nsIAccessibleTextRange { /* fields omitted */ }

interface nsIAccessibleTextRange : nsISupports

/**
 * A range representing a piece of text in the document.
 */

Methods

impl nsIAccessibleTextRange

pub fn coerce<T: nsIAccessibleTextRangeCoerce>(&self) -> &T

Cast this nsIAccessibleTextRange to one of its base interfaces.

impl nsIAccessibleTextRange

pub const EndPoint_Start: i64

EndPoint_Start: i64 = 1

/**
   * The two endpoints of the range (starting and ending).
   */

pub const EndPoint_End: i64

EndPoint_End: i64 = 2

pub const FormatUnit: i64

FormatUnit: i64 = 0

pub const WordUnit: i64

WordUnit: i64 = 1

pub const LineUnit: i64

LineUnit: i64 = 2

pub const ParagraphUnit: i64

ParagraphUnit: i64 = 3

pub const PageUnit: i64

PageUnit: i64 = 4

pub const DocumentUnit: i64

DocumentUnit: i64 = 5

pub const AnimationStyleAttr: i64

AnimationStyleAttr: i64 = 0

/**
   * Text attributes. Used in conjunction with findAttrs().
   */

pub const AnnotationObjectsAttr: i64

AnnotationObjectsAttr: i64 = 1

pub const AnnotationTypesAttr: i64

AnnotationTypesAttr: i64 = 2

pub const BackgroundColorAttr: i64

BackgroundColorAttr: i64 = 3

pub const BulletStyleAttr: i64

BulletStyleAttr: i64 = 4

pub const CapStyleAttr: i64

CapStyleAttr: i64 = 5

pub const CaretBidiModeAttr: i64

CaretBidiModeAttr: i64 = 6

pub const CaretPositionAttr: i64

CaretPositionAttr: i64 = 7

pub const CultureAttr: i64

CultureAttr: i64 = 8

pub const FontNameAttr: i64

FontNameAttr: i64 = 9

pub const FontSizeAttr: i64

FontSizeAttr: i64 = 10

pub const FontWeightAttr: i64

FontWeightAttr: i64 = 11

pub const ForegroundColorAttr: i64

ForegroundColorAttr: i64 = 12

pub const HorizontalTextAlignmentAttr: i64

HorizontalTextAlignmentAttr: i64 = 13

pub const IndentationFirstLineAttr: i64

IndentationFirstLineAttr: i64 = 14

pub const IndentationLeadingAttr: i64

IndentationLeadingAttr: i64 = 15

pub const IndentationTrailingAttr: i64

IndentationTrailingAttr: i64 = 16

pub const IsActiveAttr: i64

IsActiveAttr: i64 = 17

pub const IsHiddenAttr: i64

IsHiddenAttr: i64 = 18

pub const IsItalicAttr: i64

IsItalicAttr: i64 = 19

pub const IsReadOnlyAttr: i64

IsReadOnlyAttr: i64 = 20

pub const IsSubscriptAttr: i64

IsSubscriptAttr: i64 = 21

pub const IsSuperscriptAttr: i64

IsSuperscriptAttr: i64 = 22

pub const LinkAttr: i64

LinkAttr: i64 = 23

pub const MarginBottomAttr: i64

MarginBottomAttr: i64 = 24

pub const MarginLeadingAttr: i64

MarginLeadingAttr: i64 = 25

pub const MarginTopAttr: i64

MarginTopAttr: i64 = 26

pub const MarginTrailingAttr: i64

MarginTrailingAttr: i64 = 27

pub const OutlineStylesAttr: i64

OutlineStylesAttr: i64 = 28

pub const OverlineColorAttr: i64

OverlineColorAttr: i64 = 29

pub const OverlineStyleAttr: i64

OverlineStyleAttr: i64 = 30

pub const SelectionActiveEndAttr: i64

SelectionActiveEndAttr: i64 = 31

pub const StrikethroughColorAttr: i64

StrikethroughColorAttr: i64 = 32

pub const StrikethroughStyleAttr: i64

StrikethroughStyleAttr: i64 = 33

pub const StyleIdAttr: i64

StyleIdAttr: i64 = 34

pub const StyleNameAttr: i64

StyleNameAttr: i64 = 35

pub const TabsAttr: i64

TabsAttr: i64 = 36

pub const TextFlowDirectionsAttr: i64

TextFlowDirectionsAttr: i64 = 37

pub const UnderlineColorAttr: i64

UnderlineColorAttr: i64 = 38

pub const UnderlineStyleAttr: i64

UnderlineStyleAttr: i64 = 39

pub const AlignToTop: i64

AlignToTop: i64 = 0

pub const AlignToBottom: i64

AlignToBottom: i64 = 1

pub unsafe fn GetStartContainer(
&self,
aStartContainer: *mut *const nsIAccessibleText
) -> nsresult

readonly attribute nsIAccessibleText startContainer;

pub unsafe fn GetStartOffset(&self, aStartOffset: *mut int32_t) -> nsresult

readonly attribute long startOffset;

pub unsafe fn GetEndContainer(
&self,
aEndContainer: *mut *const nsIAccessibleText
) -> nsresult

readonly attribute nsIAccessibleText endContainer;

pub unsafe fn GetEndOffset(&self, aEndOffset: *mut int32_t) -> nsresult

readonly attribute long endOffset;

pub unsafe fn GetContainer(
&self,
aContainer: *mut *const nsIAccessible
) -> nsresult

/**
   * Return an accessible containing the whole range
   */

readonly attribute nsIAccessible container;

pub unsafe fn GetEmbeddedChildren(
&self,
aEmbeddedChildren: *mut *const nsIArray
) -> nsresult

/**
   * Return embedded children within the range.
   */

readonly attribute nsIArray embeddedChildren;

pub unsafe fn Compare(
&self,
aOtherRange: *const nsIAccessibleTextRange,
_retval: *mut bool
) -> nsresult

/**
   * Return true if this range has the same end points of the given range.
   */

boolean compare (in nsIAccessibleTextRange aOtherRange);

pub unsafe fn CompareEndPoints(
&self,
aEndPoint: uint32_t,
aOtherRange: *const nsIAccessibleTextRange,
aOtherRangeEndPoint: uint32_t,
_retval: *mut int32_t
) -> nsresult

/**
   * Compare this and given ranges end points.
   *
   * @return -1/0/1 if this range end point is before/equal/after the given
   *          range end point.
   */

long compareEndPoints (in unsigned long aEndPoint, in nsIAccessibleTextRange aOtherRange, in unsigned long aOtherRangeEndPoint);

pub unsafe fn GetText(&self, aText: &mut nsAString) -> nsresult

/**
   * Return text within the range.
   */

readonly attribute AString text;

pub unsafe fn GetBounds(&self, aBounds: *mut *const nsIArray) -> nsresult

/**
   * Return list of rects of the range.
   */

readonly attribute nsIArray bounds;

pub unsafe fn Move(&self, aUnit: uint32_t, aCount: int32_t) -> nsresult

/**
   *  Move the boundary(ies) by the given number of the unit.
   */

void move (in unsigned long aUnit, in long aCount);

pub unsafe fn MoveStart(&self, aUnit: uint32_t, aCount: int32_t) -> nsresult

void moveStart (in unsigned long aUnit, in long aCount);

pub unsafe fn MoveEnd(&self, aUnit: uint32_t, aCount: int32_t) -> nsresult

void moveEnd (in unsigned long aUnit, in long aCount);

pub unsafe fn Normalize(&self, aUnit: uint32_t) -> nsresult

/**
   * Normalize the range to the closest unit of the given type.
   */

void normalize (in unsigned long aUnit);

pub unsafe fn Crop(
&self,
aContainer: *const nsIAccessible,
_retval: *mut bool
) -> nsresult

/**
   * Crops the range by the given accessible element.
   */

boolean crop (in nsIAccessible aContainer);

pub unsafe fn FindText(
&self,
aText: &nsAString,
aIsBackward: bool,
aIsIgnoreCase: bool,
_retval: *mut *const nsIAccessibleTextRange
) -> nsresult

/**
   * Return range enclosing the found text.
   */

nsIAccessibleTextRange findText (in AString aText, in boolean aIsBackward, in boolean aIsIgnoreCase);

pub unsafe fn FindAttr(
&self,
aAttr: uint32_t,
aValue: *const nsIVariant,
aIsBackward: bool,
_retval: *mut *const nsIAccessibleTextRange
) -> nsresult

/**
   * Return range enslosing the text having requested attribute.
   */

nsIAccessibleTextRange findAttr (in unsigned long aAttr, in nsIVariant aValue, in boolean aIsBackward);

pub unsafe fn AddToSelection(&self) -> nsresult

/**
   * Add/remove the text range from selection.
   */

void addToSelection ();

pub unsafe fn RemoveFromSelection(&self) -> nsresult

void removeFromSelection ();

pub unsafe fn Select(&self) -> nsresult

void select ();

pub unsafe fn ScrollIntoView(&self, aHow: uint32_t) -> nsresult

/**
   * Scroll the range into view.
   */

void scrollIntoView (in unsigned long aHow);