Struct xpcom::interfaces::nsIAddonPolicyService

#[repr(C)]

pub struct nsIAddonPolicyService { /* fields omitted */ }

interface nsIAddonPolicyService : nsISupports

/**
 * This interface allows the security manager to query custom per-addon security
 * policy.
 */

Methods

impl nsIAddonPolicyService

pub fn coerce<T: nsIAddonPolicyServiceCoerce>(&self) -> &T

Cast this nsIAddonPolicyService to one of its base interfaces.

impl nsIAddonPolicyService

pub unsafe fn GetBaseCSP(&self, aBaseCSP: &mut nsAString) -> nsresult

/**
   * Returns the base content security policy, which is applied to all
   * extension documents, in addition to any custom policies.
   */

readonly attribute AString baseCSP;

pub unsafe fn GetDefaultCSP(&self, aDefaultCSP: &mut nsAString) -> nsresult

/**
   * Returns the default content security policy which applies to extension
   * documents which do not specify any custom policies.
   */

readonly attribute AString defaultCSP;

pub unsafe fn GetAddonCSP(
&self,
aAddonId: &nsAString,
_retval: &mut nsAString
) -> nsresult

/**
   * Returns the content security policy which applies to documents belonging
   * to the extension with the given ID. This may be either a custom policy,
   * if one was supplied, or the default policy if one was not.
   */

AString getAddonCSP (in AString aAddonId);

pub unsafe fn GetGeneratedBackgroundPageUrl(
&self,
aAddonId: &nsACString,
_retval: &mut nsACString
) -> nsresult

/**
   * Returns the generated background page as a data-URI, if any. If the addon
   * does not have an auto-generated background page, an empty string is
   * returned.
   */

ACString getGeneratedBackgroundPageUrl (in ACString aAddonId);

pub unsafe fn AddonHasPermission(
&self,
aAddonId: &nsAString,
aPerm: &nsAString,
_retval: *mut bool
) -> nsresult

/**
   * Returns true if the addon was granted the |aPerm| API permission.
   */

boolean addonHasPermission (in AString aAddonId, in AString aPerm);

pub unsafe fn AddonMayLoadURI(
&self,
aAddonId: &nsAString,
aURI: *const nsIURI,
aExplicit: bool,
_retval: *mut bool
) -> nsresult

/**
   * Returns true if unprivileged code associated with the given addon may load
   * data from |aURI|.  If |aExplicit| is true, the <all_urls> permission and
   * permissive host globs are ignored when checking for a match.
   */

boolean addonMayLoadURI (in AString aAddonId, in nsIURI aURI, [optional] in boolean aExplicit);

pub unsafe fn GetExtensionName(
&self,
aAddonId: &nsAString,
_retval: &mut nsAString
) -> nsresult

/**
   * Returns the name of the WebExtension with the given ID, or the ID string
   * if no matching add-on can be found.
   */

AString getExtensionName (in AString aAddonId);

pub unsafe fn ExtensionURILoadableByAnyone(
&self,
aURI: *const nsIURI,
_retval: *mut bool
) -> nsresult

/**
   * Returns true if a given extension:// URI is web-accessible.
   */

boolean extensionURILoadableByAnyone (in nsIURI aURI);

pub unsafe fn ExtensionURIToAddonId(
&self,
aURI: *const nsIURI,
_retval: &mut nsAString
) -> nsresult

/**
   * Maps an extension URI to the ID of the addon it belongs to.
   */

AString extensionURIToAddonId (in nsIURI aURI);