Struct xpcom::interfaces::nsIAppShellService

#[repr(C)]

pub struct nsIAppShellService { /* fields omitted */ }

interface nsIAppShellService : nsISupports

Methods

impl nsIAppShellService

pub fn coerce<T: nsIAppShellServiceCoerce>(&self) -> &T

Cast this nsIAppShellService to one of its base interfaces.

impl nsIAppShellService

pub const SIZE_TO_CONTENT: i64

SIZE_TO_CONTENT: i64 = -1

/**
   * Create a window, which will be initially invisible.
   * @param aParent the parent window.  Can be null.
   * @param aUrl the contents of the new window.
   * @param aChromeMask chrome flags affecting the kind of OS border
   *                    given to the window. see nsIBrowserWindow for
   *                    bit/flag definitions.
   * @param aCallbacks interface providing C++ hooks for window initialization
   *                   before the window is made visible.  Can be null.
   *                   Deprecated.
   * @param aInitialWidth width, in pixels, of the window.  Width of window
   *                      at creation.  Can be overridden by the "width"
   *                      tag in the XUL.  Set to NS_SIZETOCONTENT to force
   *                      the window to wrap to its contents.
   * @param aInitialHeight like aInitialWidth, but subtly different.
   * @param aOpeningTab The TabParent that requested that this window be opened.
   *                    Can be left null.
   * @param aOpenerWindow The Window Proxy which requested that this window be opened.
   *                      Can be left null.
   */

pub unsafe fn CreateTopLevelWindow(
&self,
aParent: *const nsIXULWindow,
aUrl: *const nsIURI,
aChromeMask: uint32_t,
aInitialWidth: int32_t,
aInitialHeight: int32_t,
aOpeningTab: *const nsITabParent,
aOpenerWindow: *const mozIDOMWindowProxy,
_retval: *mut *const nsIXULWindow
) -> nsresult

nsIXULWindow createTopLevelWindow (in nsIXULWindow aParent, in nsIURI aUrl, in uint32_t aChromeMask, in long aInitialWidth, in long aInitialHeight, in nsITabParent aOpeningTab, in mozIDOMWindowProxy aOpenerWindow);

pub unsafe fn CreateWindowlessBrowser(
&self,
aIsChrome: bool,
_retval: *mut *const nsIWindowlessBrowser
) -> nsresult

/**
   * This is the constructor for creating an invisible DocShell.
   * It is used to simulate DOM windows without an actual physical
   * representation.
   * @param aIsChrome Set true if you want to use it for chrome content.
   */

nsIWindowlessBrowser createWindowlessBrowser ([optional] in bool aIsChrome);

pub unsafe fn CreateHiddenWindow(&self) -> nsresult

[noscript] void createHiddenWindow ();

pub unsafe fn DestroyHiddenWindow(&self) -> nsresult

void destroyHiddenWindow ();

pub unsafe fn SetScreenId(&self, aScreenId: uint32_t) -> nsresult

/**
   * B2G multi-screen support. When open another top-level window on b2g,
   * a screen ID is needed for identifying which screen this window is
   * opened to.
   * @param aScreenId Differentiate screens of windows. It is platform-
   *                  specific due to the hardware limitation for now.
   */

[noscript] void setScreenId (in uint32_t aScreenId);

pub unsafe fn GetHiddenWindow(
&self,
aHiddenWindow: *mut *const nsIXULWindow
) -> nsresult

/**
   * Return the (singleton) application hidden window, automatically created
   * and maintained by this AppShellService.
   * @param aResult the hidden window.  Do not unhide hidden window.
   *                Do not taunt hidden window.
   */

readonly attribute nsIXULWindow hiddenWindow;

pub unsafe fn GetHiddenDOMWindow(
&self,
aHiddenDOMWindow: *mut *const mozIDOMWindowProxy
) -> nsresult

/**
   * Return the (singleton) application hidden window, automatically created
   * and maintained by this AppShellService.
   * @param aResult the hidden window.  Do not unhide hidden window.
   *                Do not taunt hidden window.
   */

readonly attribute mozIDOMWindowProxy hiddenDOMWindow;

pub unsafe fn GetHiddenPrivateWindow(
&self,
aHiddenPrivateWindow: *mut *const nsIXULWindow
) -> nsresult

/**
   * Return the (singleton) application hidden private window, automatically
   * created and maintained by this AppShellService.  This window is created
   * in private browsing mode.
   * @param aResult the hidden private window.  Do not unhide hidden window.
   *                Do not taunt hidden window.
   */

readonly attribute nsIXULWindow hiddenPrivateWindow;

pub unsafe fn GetHiddenPrivateDOMWindow(
&self,
aHiddenPrivateDOMWindow: *mut *const mozIDOMWindowProxy
) -> nsresult

/**
   * Return the (singleton) application hidden private window, automatically
   * created and maintained by this AppShellService.  This window is created
   * in private browsing mode.
   * @param aResult the hidden private window.  Do not unhide hidden window.
   *                Do not taunt hidden window.
   */

readonly attribute mozIDOMWindowProxy hiddenPrivateDOMWindow;

pub unsafe fn GetApplicationProvidedHiddenWindow(
&self,
aApplicationProvidedHiddenWindow: *mut bool
) -> nsresult

/**
   * Return true if the application hidden window was provided by the
   * application. If it wasn't, the default hidden window was used. This will
   * usually be false on all non-mac platforms.
   */

readonly attribute boolean applicationProvidedHiddenWindow;

pub unsafe fn RegisterTopLevelWindow(
&self,
aWindow: *const nsIXULWindow
) -> nsresult

/**
   * Add a window to the application's registry of windows.  These windows
   * are generally shown in the Windows taskbar, and the application
   * knows it can't quit until it's out of registered windows.
   * @param aWindow the window to register
   * @note When this method is successful, it fires the global notification
   *       "xul-window-registered"
   */

void registerTopLevelWindow (in nsIXULWindow aWindow);

pub unsafe fn UnregisterTopLevelWindow(
&self,
aWindow: *const nsIXULWindow
) -> nsresult

/**
   * Remove a window from the application's window registry. Note that
   * this method won't automatically attempt to quit the app when
   * the last window is unregistered. For that, see Quit().
   * @param aWindow you see the pattern
   */

void unregisterTopLevelWindow (in nsIXULWindow aWindow);

pub unsafe fn GetHasHiddenPrivateWindow(
&self,
aHasHiddenPrivateWindow: *mut bool
) -> nsresult

/**
   * Whether the hidden private window has been lazily created.
   */

[noscript] readonly attribute boolean hasHiddenPrivateWindow;

pub unsafe fn StartEventLoopLagTracking(&self, _retval: *mut bool) -> nsresult

/**
   * Start/stop tracking lags in the event loop.
   * If the event loop gets unresponsive, a "event-loop-lag" notification
   * is sent. Note that calling `startEventLoopLagTracking` when tracking
   * is already enabled has no effect.
   * @return true if tracking succeeded.
   */

bool startEventLoopLagTracking ();

pub unsafe fn StopEventLoopLagTracking(&self) -> nsresult

void stopEventLoopLagTracking ();