Struct xpcom::interfaces::nsIApplicationUpdateService

#[repr(C)]

pub struct nsIApplicationUpdateService { /* fields omitted */ }

interface nsIApplicationUpdateService : nsISupports

/**
 * An interface describing a global application service that handles performing
 * background update checks and provides utilities for selecting and
 * downloading update patches.
 */

Methods

impl nsIApplicationUpdateService

pub fn coerce<T: nsIApplicationUpdateServiceCoerce>(&self) -> &T

Cast this nsIApplicationUpdateService to one of its base interfaces.

impl nsIApplicationUpdateService

pub unsafe fn CheckForBackgroundUpdates(&self) -> nsresult

/**
   * Checks for available updates in the background using the listener provided
   * by the application update service for background checks.
   */

void checkForBackgroundUpdates ();

pub unsafe fn GetBackgroundChecker(
&self,
aBackgroundChecker: *mut *const nsIUpdateChecker
) -> nsresult

/**
   * The Update Checker used for background update checking.
   */

readonly attribute nsIUpdateChecker backgroundChecker;

pub unsafe fn SelectUpdate(
&self,
updates: *mut *const nsIUpdate,
updateCount: uint32_t,
_retval: *mut *const nsIUpdate
) -> nsresult

/**
   * Selects the best update to install from a list of available updates.
   * @param   updates
   *          An array of updates that are available
   * @param   updateCount
   *          The length of the |updates| array
   */

nsIUpdate selectUpdate ([array, size_is (updateCount)] in nsIUpdate updates, in unsigned long updateCount);

pub unsafe fn AddDownloadListener(
&self,
listener: *const nsIRequestObserver
) -> nsresult

/**
   * Adds a listener that receives progress and state information about the
   * update that is currently being downloaded, e.g. to update a user
   * interface.
   * @param   listener
   *          An object implementing nsIRequestObserver and optionally
   *          nsIProgressEventSink that is to be notified of state and
   *          progress information as the update is downloaded.
   */

void addDownloadListener (in nsIRequestObserver listener);

pub unsafe fn RemoveDownloadListener(
&self,
listener: *const nsIRequestObserver
) -> nsresult

/**
   * Removes a listener that is receiving progress and state information
   * about the update that is currently being downloaded.
   * @param   listener
   *          The listener object to remove.
   */

void removeDownloadListener (in nsIRequestObserver listener);

pub unsafe fn DownloadUpdate(
&self,
update: *const nsIUpdate,
background: bool,
_retval: &mut nsAString
) -> nsresult

/**
   *
   */

AString downloadUpdate (in nsIUpdate update, in boolean background);

pub unsafe fn GetUpdatesDirectory(
&self,
_retval: *mut *const nsIFile
) -> nsresult

/**
   * Get the Active Updates directory
   * @returns An nsIFile for the active updates directory.
   */

nsIFile getUpdatesDirectory ();

pub unsafe fn PauseDownload(&self) -> nsresult

/**
   * Pauses the active update download process
   */

void pauseDownload ();

pub unsafe fn GetIsDownloading(&self, aIsDownloading: *mut bool) -> nsresult

/**
   * Whether or not there is an download happening at the moment.
   */

readonly attribute boolean isDownloading;

pub unsafe fn GetCanCheckForUpdates(
&self,
aCanCheckForUpdates: *mut bool
) -> nsresult

/**
   * Whether or not the Update Service can check for updates. This is a function
   * of whether or not application update is disabled by the application and the
   * platform the application is running on.
   */

readonly attribute boolean canCheckForUpdates;

pub unsafe fn GetElevationRequired(
&self,
aElevationRequired: *mut bool
) -> nsresult

/**
   * Whether or not the installation requires elevation. Currently only
   * implemented on OSX, returns false on other platforms.
   */

readonly attribute boolean elevationRequired;

pub unsafe fn GetCanApplyUpdates(&self, aCanApplyUpdates: *mut bool) -> nsresult

/**
   * Whether or not the Update Service can download and install updates.
   * On Windows, this is a function of whether or not the maintenance service
   * is installed and enabled. On other systems, and as a fallback on Windows,
   * this depends on whether the current user has write access to the install
   * directory.
   */

readonly attribute boolean canApplyUpdates;

pub unsafe fn GetIsOtherInstanceHandlingUpdates(
&self,
aIsOtherInstanceHandlingUpdates: *mut bool
) -> nsresult

/**
   * Whether or not a different instance is handling updates of this
   * installation.  This currently only ever returns true on Windows
   * when 2 instances of an application are open. Only one of the instances
   * will actually handle updates for the installation.
   */

readonly attribute boolean isOtherInstanceHandlingUpdates;

pub unsafe fn GetCanStageUpdates(&self, aCanStageUpdates: *mut bool) -> nsresult

/**
   * Whether the Update Service is able to stage updates.
   */

readonly attribute boolean canStageUpdates;