Struct xpcom::interfaces::nsIArrayExtensions

#[repr(C)]

pub struct nsIArrayExtensions { /* fields omitted */ }

interface nsIArrayExtensions : nsIArray

/**
 * Helper interface for allowing scripts to treat nsIArray instances as if
 * they were nsISupportsArray instances while iterating.
 *
 * nsISupportsArray is convenient to iterate over in JavaScript:
 *
 * for (let i = 0; i < array.Count(); ++i) {
 *   let elem = array.GetElementAt(i);
 *   ...
 * }
 *
 * but doing the same with nsIArray is somewhat less convenient, since
 * queryElementAt is not nearly so nice to use from JavaScript.  So we provide
 * this extension interface so interfaces that currently return
 * nsISupportsArray can start returning nsIArrayExtensions and all JavaScript
 * should Just Work. Eventually we'll roll this interface into nsIArray
 * itself, possibly getting rid of the Count() method, as it duplicates
 * nsIArray functionality.
 */

Methods

impl nsIArrayExtensions

pub fn coerce<T: nsIArrayExtensionsCoerce>(&self) -> &T

Cast this nsIArrayExtensions to one of its base interfaces.

impl nsIArrayExtensions

pub unsafe fn Count(&self, _retval: *mut uint32_t) -> nsresult

/**
   * Count()
   *
   * Retrieves the length of the array. This is an alias for the
   * |nsIArray.length| attribute.
   */

uint32_t Count ();

pub unsafe fn GetElementAt(
&self,
index: uint32_t,
_retval: *mut *const nsISupports
) -> nsresult

/**
   * GetElementAt()
   *
   * Retrieve a specific element of the array. null is a valid result for
   * this method.
   *
   * Note: If the index is out of bounds null will be returned.
   *       This differs from the behavior of nsIArray.queryElementAt() which
   *       will throw if an invalid index is specified.
   *
   * @param index position of element
   */

nsISupports GetElementAt (in uint32_t index);