Struct xpcom::interfaces::nsIAsyncShutdownService

#[repr(C)]

pub struct nsIAsyncShutdownService { /* fields omitted */ }

interface nsIAsyncShutdownService : nsISupports

/**
 * A service that allows registering shutdown-time dependencies.
 */

Methods

impl nsIAsyncShutdownService

pub fn coerce<T: nsIAsyncShutdownServiceCoerce>(&self) -> &T

Cast this nsIAsyncShutdownService to one of its base interfaces.

impl nsIAsyncShutdownService

pub unsafe fn MakeBarrier(
&self,
aName: &nsAString,
_retval: *mut *const nsIAsyncShutdownBarrier
) -> nsresult

/**
   * Create a new barrier.
   *
   * By convention, the name should respect the following format:
   * "MyModuleName: Doing something while it's time"
   * e.g.
   * "OS.File: Waiting for clients to flush before shutting down"
   *
   * This attribute is uploaded as part of crash reports.
   */

nsIAsyncShutdownBarrier makeBarrier (in AString aName);

pub unsafe fn GetProfileBeforeChange(
&self,
aProfileBeforeChange: *mut *const nsIAsyncShutdownClient
) -> nsresult

/**
   * Barrier for notification profile-before-change.
   */

readonly attribute nsIAsyncShutdownClient profileBeforeChange;

pub unsafe fn GetProfileChangeTeardown(
&self,
aProfileChangeTeardown: *mut *const nsIAsyncShutdownClient
) -> nsresult

/**
   * Barrier for notification profile-change-teardown.
   */

readonly attribute nsIAsyncShutdownClient profileChangeTeardown;

pub unsafe fn GetQuitApplicationGranted(
&self,
aQuitApplicationGranted: *mut *const nsIAsyncShutdownClient
) -> nsresult

/**
   * Barrier for notification quit-application-granted.
   */

readonly attribute nsIAsyncShutdownClient quitApplicationGranted;

pub unsafe fn GetSendTelemetry(
&self,
aSendTelemetry: *mut *const nsIAsyncShutdownClient
) -> nsresult

/**
   * Barrier for notification profile-before-change-telemetry.
   */

readonly attribute nsIAsyncShutdownClient sendTelemetry;

pub unsafe fn GetWebWorkersShutdown(
&self,
aWebWorkersShutdown: *mut *const nsIAsyncShutdownClient
) -> nsresult

/**
   * Barrier for notification web-workers-shutdown.
   */

readonly attribute nsIAsyncShutdownClient webWorkersShutdown;

pub unsafe fn GetXpcomWillShutdown(
&self,
aXpcomWillShutdown: *mut *const nsIAsyncShutdownClient
) -> nsresult

/**
   * Barrier for notification xpcom-will-shutdown.
   */

readonly attribute nsIAsyncShutdownClient xpcomWillShutdown;