Struct xpcom::interfaces::nsIAutoCompleteSimpleResult

#[repr(C)]

pub struct nsIAutoCompleteSimpleResult { /* fields omitted */ }

interface nsIAutoCompleteSimpleResult : nsIAutoCompleteResult

/**
 * This class implements nsIAutoCompleteResult and provides simple methods
 * for setting the value and result items. It can be used whenever some basic
 * auto complete results are needed that can be pre-generated and filled into
 * an array.
 */

Methods

impl nsIAutoCompleteSimpleResult

pub fn coerce<T: nsIAutoCompleteSimpleResultCoerce>(&self) -> &T

Cast this nsIAutoCompleteSimpleResult to one of its base interfaces.

impl nsIAutoCompleteSimpleResult

pub unsafe fn SetSearchString(&self, aSearchString: &nsAString) -> nsresult

/**
   * A writer for the readonly attribute 'searchString' which should contain
   * the string that the user typed.
   */

void setSearchString (in AString aSearchString);

pub unsafe fn SetErrorDescription(
&self,
aErrorDescription: &nsAString
) -> nsresult

/**
   * A writer for the readonly attribute 'errorDescription'.
   */

void setErrorDescription (in AString aErrorDescription);

pub unsafe fn SetDefaultIndex(&self, aDefaultIndex: int32_t) -> nsresult

/**
   * A writer for the readonly attribute 'defaultIndex' which should contain
   * the index of the list that will be selected by default (normally 0).
   */

void setDefaultIndex (in long aDefaultIndex);

pub unsafe fn SetSearchResult(&self, aSearchResult: uint16_t) -> nsresult

/**
   * A writer for the readonly attribute 'searchResult' which should contain
   * one of the constants nsIAutoCompleteResult.RESULT_* indicating the success
   * of the search.
   */

void setSearchResult (in unsigned short aSearchResult);

pub unsafe fn InsertMatchAt(
&self,
aIndex: int32_t,
aValue: &nsAString,
aComment: &nsAString,
aImage: &nsAString,
aStyle: &nsAString,
aFinalCompleteValue: &nsAString,
aLabel: &nsAString
) -> nsresult

/**
   * Inserts a match consisting of the given value, comment, image, style and
   * the value to use for defaultIndex completion at a given position.
   * @param aIndex
   *        The index to insert at
   * @param aValue
   *        The value to autocomplete to
   * @param aComment
   *        Comment shown in the autocomplete widget to describe this match
   * @param aImage
   *        Image shown in the autocomplete widget for this match.
   * @param aStyle
   *        Describes how to style the match in the autocomplete widget
   * @param aFinalCompleteValue
   *        Value used when the user confirms selecting this match. If not
   *        provided, aValue will be used.
   */

void insertMatchAt (in long aIndex, in AString aValue, in AString aComment, [optional] in AString aImage, [optional] in AString aStyle, [optional] in AString aFinalCompleteValue, [optional] in AString aLabel);

pub unsafe fn AppendMatch(
&self,
aValue: &nsAString,
aComment: &nsAString,
aImage: &nsAString,
aStyle: &nsAString,
aFinalCompleteValue: &nsAString,
aLabel: &nsAString
) -> nsresult

/**
   * Appends a match consisting of the given value, comment, image, style and
   * the value to use for defaultIndex completion.
   * @param aValue
   *        The value to autocomplete to
   * @param aComment
   *        Comment shown in the autocomplete widget to describe this match
   * @param aImage
   *        Image shown in the autocomplete widget for this match.
   * @param aStyle
   *        Describes how to style the match in the autocomplete widget
   * @param aFinalCompleteValue
   *        Value used when the user confirms selecting this match. If not
   *        provided, aValue will be used.
   */

void appendMatch (in AString aValue, in AString aComment, [optional] in AString aImage, [optional] in AString aStyle, [optional] in AString aFinalCompleteValue, [optional] in AString aLabel);

pub unsafe fn RemoveMatchAt(&self, aIndex: int32_t) -> nsresult

/**
   * Removes an existing match.
   * @note this is different from removeValueAt, since it's not a consequence of
   * a user action, and as such it won't notify onValueRemoved.
   */

void removeMatchAt (in long aIndex);

pub unsafe fn GetListener(
&self,
_retval: *mut *const nsIAutoCompleteSimpleResultListener
) -> nsresult

/**
   * Gets the listener for changes in the result.
   */

nsIAutoCompleteSimpleResultListener getListener ();

pub unsafe fn SetListener(
&self,
aListener: *const nsIAutoCompleteSimpleResultListener
) -> nsresult

/**
   * Sets a listener for changes in the result.
   */

void setListener (in nsIAutoCompleteSimpleResultListener aListener);