Struct xpcom::interfaces::nsIBaseWindow

#[repr(C)]

pub struct nsIBaseWindow { /* fields omitted */ }

interface nsIBaseWindow : nsISupports

/**
 * The nsIBaseWindow describes a generic window and basic operations that
 * can be performed on it.  This is not to be a complete windowing interface
 * but rather a common set that nearly all windowed objects support.
 */

Methods

impl nsIBaseWindow

pub fn coerce<T: nsIBaseWindowCoerce>(&self) -> &T

Cast this nsIBaseWindow to one of its base interfaces.

impl nsIBaseWindow

pub const eRepaint: i64

eRepaint: i64 = 1

/**
     * The 'flags' argument to setPositionAndSize is a set of these bits.
     */

pub const eDelayResize: i64

eDelayResize: i64 = 2

pub unsafe fn Create(&self) -> nsresult

void create ();

pub unsafe fn Destroy(&self) -> nsresult

void destroy ();

pub unsafe fn SetPosition(&self, x: int32_t, y: int32_t) -> nsresult

void setPosition (in long x, in long y);

pub unsafe fn SetPositionDesktopPix(&self, x: int32_t, y: int32_t) -> nsresult

void setPositionDesktopPix (in long x, in long y);

pub unsafe fn GetPosition(&self, x: *mut int32_t, y: *mut int32_t) -> nsresult

void getPosition (out long x, out long y);

pub unsafe fn SetSize(
&self,
cx: int32_t,
cy: int32_t,
fRepaint: bool
) -> nsresult

void setSize (in long cx, in long cy, in boolean fRepaint);

pub unsafe fn GetSize(&self, cx: *mut int32_t, cy: *mut int32_t) -> nsresult

void getSize (out long cx, out long cy);

pub unsafe fn SetPositionAndSize(
&self,
x: int32_t,
y: int32_t,
cx: int32_t,
cy: int32_t,
flags: uint32_t
) -> nsresult

void setPositionAndSize (in long x, in long y, in long cx, in long cy, in unsigned long flags);

pub unsafe fn GetPositionAndSize(
&self,
x: *mut int32_t,
y: *mut int32_t,
cx: *mut int32_t,
cy: *mut int32_t
) -> nsresult

void getPositionAndSize (out long x, out long y, out long cx, out long cy);

pub unsafe fn Repaint(&self, force: bool) -> nsresult

/**
     * Tell the window to repaint itself
     * @param aForce - if true, repaint immediately
     *                 if false, the window may defer repainting as it sees fit.
     */

void repaint (in boolean force);

pub unsafe fn GetParentNativeWindow(
&self,
aParentNativeWindow: *mut nativeWindow
) -> nsresult

attribute nativeWindow parentNativeWindow;

pub unsafe fn SetParentNativeWindow(
&self,
aParentNativeWindow: nativeWindow
) -> nsresult

attribute nativeWindow parentNativeWindow;

pub unsafe fn GetNativeHandle(&self, aNativeHandle: &mut nsAString) -> nsresult

readonly attribute DOMString nativeHandle;

pub unsafe fn GetVisibility(&self, aVisibility: *mut bool) -> nsresult

attribute boolean visibility;

pub unsafe fn SetVisibility(&self, aVisibility: bool) -> nsresult

attribute boolean visibility;

pub unsafe fn GetEnabled(&self, aEnabled: *mut bool) -> nsresult

attribute boolean enabled;

pub unsafe fn SetEnabled(&self, aEnabled: bool) -> nsresult

attribute boolean enabled;

pub unsafe fn GetUnscaledDevicePixelsPerCSSPixel(
&self,
aUnscaledDevicePixelsPerCSSPixel: *mut c_double
) -> nsresult

readonly attribute double unscaledDevicePixelsPerCSSPixel;

pub unsafe fn GetDevicePixelsPerDesktopPixel(
&self,
aDevicePixelsPerDesktopPixel: *mut c_double
) -> nsresult

readonly attribute double devicePixelsPerDesktopPixel;

pub unsafe fn SetFocus(&self) -> nsresult

/**
    * Give the window focus.
    */

void setFocus ();

pub unsafe fn GetTitle(&self, aTitle: &mut nsAString) -> nsresult

attribute AString title;

pub unsafe fn SetTitle(&self, aTitle: &nsAString) -> nsresult

attribute AString title;