Struct xpcom::interfaces::nsIBinaryInputStream

#[repr(C)]

pub struct nsIBinaryInputStream { /* fields omitted */ }

interface nsIBinaryInputStream : nsIInputStream

/**
 * This interface allows consumption of primitive data types from a "binary
 * stream" containing untagged, big-endian binary data, i.e. as produced by an
 * implementation of nsIBinaryOutputStream.  This might be used, for example,
 * to implement network protocols or to read from architecture-neutral disk
 * files, i.e. ones that can be read and written by both big-endian and
 * little-endian platforms.
 *
 * @See nsIBinaryOutputStream
 */

Methods

impl nsIBinaryInputStream

pub fn coerce<T: nsIBinaryInputStreamCoerce>(&self) -> &T

Cast this nsIBinaryInputStream to one of its base interfaces.

impl nsIBinaryInputStream

pub unsafe fn SetInputStream(
&self,
aInputStream: *const nsIInputStream
) -> nsresult

void setInputStream (in nsIInputStream aInputStream);

pub unsafe fn ReadBoolean(&self, _retval: *mut bool) -> nsresult

/**
     * Read 8-bits from the stream.
     *
     * @return that byte to be treated as a boolean.
     */

boolean readBoolean ();

pub unsafe fn Read8(&self, _retval: *mut uint8_t) -> nsresult

uint8_t read8 ();

pub unsafe fn Read16(&self, _retval: *mut uint16_t) -> nsresult

uint16_t read16 ();

pub unsafe fn Read32(&self, _retval: *mut uint32_t) -> nsresult

uint32_t read32 ();

pub unsafe fn Read64(&self, _retval: *mut uint64_t) -> nsresult

uint64_t read64 ();

pub unsafe fn ReadFloat(&self, _retval: *mut c_float) -> nsresult

float readFloat ();

pub unsafe fn ReadDouble(&self, _retval: *mut c_double) -> nsresult

double readDouble ();

pub unsafe fn ReadCString(&self, _retval: &mut nsACString) -> nsresult

/**
     * Read an 8-bit pascal style string from the stream.
     * 32-bit length field, followed by length 8-bit chars.
     */

ACString readCString ();

pub unsafe fn ReadString(&self, _retval: &mut nsAString) -> nsresult

/**
     * Read an 16-bit pascal style string from the stream.
     * 32-bit length field, followed by length PRUnichars.
     */

AString readString ();

pub unsafe fn ReadBytes(
&self,
aLength: uint32_t,
aString: *mut *const c_char
) -> nsresult

/**
     * Read an opaque byte array from the stream.
     *
     * @param aLength the number of bytes that must be read.
     *
     * @throws NS_ERROR_FAILURE if it can't read aLength bytes
     */

void readBytes (in uint32_t aLength, [size_is (aLength), retval] out string aString);

pub unsafe fn ReadByteArray(
&self,
aLength: uint32_t,
aBytes: *mut *mut uint8_t
) -> nsresult

/**
     * Read an opaque byte array from the stream, storing the results
     * as an array of PRUint8s.
     *
     * @param aLength the number of bytes that must be read.
     *
     * @throws NS_ERROR_FAILURE if it can't read aLength bytes
     */

void readByteArray (in uint32_t aLength, [array, size_is (aLength), retval] out uint8_t aBytes);