Struct xpcom::interfaces::nsIBinaryOutputStream

#[repr(C)]

pub struct nsIBinaryOutputStream { /* fields omitted */ }

interface nsIBinaryOutputStream : nsIOutputStream

/**
 * This interface allows writing of primitive data types (integers,
 * floating-point values, booleans, etc.) to a stream in a binary, untagged,
 * fixed-endianness format.  This might be used, for example, to implement
 * network protocols or to produce architecture-neutral binary disk files,
 * i.e. ones that can be read and written by both big-endian and little-endian
 * platforms.  Output is written in big-endian order (high-order byte first),
 * as this is traditional network order.
 *
 * @See nsIBinaryInputStream
 */

Methods

impl nsIBinaryOutputStream

pub fn coerce<T: nsIBinaryOutputStreamCoerce>(&self) -> &T

Cast this nsIBinaryOutputStream to one of its base interfaces.

impl nsIBinaryOutputStream

pub unsafe fn SetOutputStream(
&self,
aOutputStream: *const nsIOutputStream
) -> nsresult

void setOutputStream (in nsIOutputStream aOutputStream);

pub unsafe fn WriteBoolean(&self, aBoolean: bool) -> nsresult

/**
     * Write a boolean as an 8-bit char to the stream.
     */

void writeBoolean (in boolean aBoolean);

pub unsafe fn Write8(&self, aByte: uint8_t) -> nsresult

void write8 (in uint8_t aByte);

pub unsafe fn Write16(&self, a16: uint16_t) -> nsresult

void write16 (in uint16_t a16);

pub unsafe fn Write32(&self, a32: uint32_t) -> nsresult

void write32 (in uint32_t a32);

pub unsafe fn Write64(&self, a64: uint64_t) -> nsresult

void write64 (in uint64_t a64);

pub unsafe fn WriteFloat(&self, aFloat: c_float) -> nsresult

void writeFloat (in float aFloat);

pub unsafe fn WriteDouble(&self, aDouble: c_double) -> nsresult

void writeDouble (in double aDouble);

pub unsafe fn WriteStringZ(&self, aString: *const c_char) -> nsresult

/**
     * Write an 8-bit pascal style string to the stream.
     * 32-bit length field, followed by length 8-bit chars.
     */

void writeStringZ (in string aString);

pub unsafe fn WriteWStringZ(&self, aString: *const int16_t) -> nsresult

/**
     * Write a 16-bit pascal style string to the stream.
     * 32-bit length field, followed by length PRUnichars.
     */

void writeWStringZ (in wstring aString);

pub unsafe fn WriteUtf8Z(&self, aString: *const int16_t) -> nsresult

/**
     * Write an 8-bit pascal style string (UTF8-encoded) to the stream.
     * 32-bit length field, followed by length 8-bit chars.
     */

void writeUtf8Z (in wstring aString);

pub unsafe fn WriteBytes(
&self,
aString: *const c_char,
aLength: uint32_t
) -> nsresult

/**
     * Write an opaque byte array to the stream.
     */

void writeBytes ([size_is (aLength)] in string aString, in uint32_t aLength);

pub unsafe fn WriteByteArray(
&self,
aBytes: *mut uint8_t,
aLength: uint32_t
) -> nsresult

/**
     * Write an opaque byte array to the stream.
     */

void writeByteArray ([array, size_is (aLength)] in uint8_t aBytes, in uint32_t aLength);