Struct xpcom::interfaces::nsIBufferedOutputStream

#[repr(C)]

pub struct nsIBufferedOutputStream { /* fields omitted */ }

interface nsIBufferedOutputStream : nsIOutputStream

/**
 * An output stream that stores up data to write out to another output stream
 * and does the entire write only when the buffer is full, so that fewer writes
 * to the underlying output stream are necessary.
 */

Methods

impl nsIBufferedOutputStream

pub fn coerce<T: nsIBufferedOutputStreamCoerce>(&self) -> &T

Cast this nsIBufferedOutputStream to one of its base interfaces.

impl nsIBufferedOutputStream

pub unsafe fn Init(
&self,
sinkToStream: *const nsIOutputStream,
bufferSize: uint32_t
) -> nsresult

/**
     * @param sinkToStream - add buffering to this stream
     * @param bufferSize   - specifies the maximum buffer size
     */

void init (in nsIOutputStream sinkToStream, in unsigned long bufferSize);

pub unsafe fn GetData(&self, aData: *mut *const nsIOutputStream) -> nsresult

/**
     * Get the wrapped data stream
     */

readonly attribute nsIOutputStream data;