Struct xpcom::interfaces::nsICacheInfoChannel

#[repr(C)]

pub struct nsICacheInfoChannel { /* fields omitted */ }

interface nsICacheInfoChannel : nsISupports

Methods

impl nsICacheInfoChannel

pub fn coerce<T: nsICacheInfoChannelCoerce>(&self) -> &T

Cast this nsICacheInfoChannel to one of its base interfaces.

impl nsICacheInfoChannel

pub unsafe fn GetCacheTokenFetchCount(
&self,
aCacheTokenFetchCount: *mut int32_t
) -> nsresult

/**
   * Get the number of times the cache entry has been opened. This attribute is
   * equivalent to nsICachingChannel.cacheToken.fetchCount.
   *
   * @throws NS_ERROR_NOT_AVAILABLE if the cache entry or the alternate data
   *         cache entry cannot be read.
   */

readonly attribute int32_t cacheTokenFetchCount;

pub unsafe fn GetCacheTokenExpirationTime(
&self,
aCacheTokenExpirationTime: *mut uint32_t
) -> nsresult

/**
   * Get expiration time from cache token. This attribute is equivalent to
   * nsICachingChannel.cacheToken.expirationTime.
   */

readonly attribute uint32_t cacheTokenExpirationTime;

pub unsafe fn GetCacheTokenCachedCharset(
&self,
aCacheTokenCachedCharset: &mut nsACString
) -> nsresult

/**
   * Set/get charset of cache entry. Accessing this attribute is equivalent to
   * calling nsICachingChannel.cacheToken.getMetaDataElement("charset") and
   * nsICachingChannel.cacheToken.setMetaDataElement("charset").
   */

attribute ACString cacheTokenCachedCharset;

pub unsafe fn SetCacheTokenCachedCharset(
&self,
aCacheTokenCachedCharset: &nsACString
) -> nsresult

/**
   * Set/get charset of cache entry. Accessing this attribute is equivalent to
   * calling nsICachingChannel.cacheToken.getMetaDataElement("charset") and
   * nsICachingChannel.cacheToken.setMetaDataElement("charset").
   */

attribute ACString cacheTokenCachedCharset;

pub unsafe fn IsFromCache(&self, _retval: *mut bool) -> nsresult

/**
   * TRUE if this channel's data is being loaded from the cache.  This value
   * is undefined before the channel fires its OnStartRequest notification
   * and after the channel fires its OnStopRequest notification.
   */

boolean isFromCache ();

pub unsafe fn GetCacheEntryId(&self, _retval: *mut uint64_t) -> nsresult

/**
   * The unique ID of the corresponding nsICacheEntry from which the response is
   * retrieved. By comparing the returned value, we can judge whether the data
   * of two distinct nsICacheInfoChannels is from the same nsICacheEntry. This
   * scenario could be useful when verifying whether the alternative data from
   * one nsICacheInfochannel matches the main data from another one.
   *
   * Note: NS_ERROR_NOT_AVAILABLE is thrown when a nsICacheInfoChannel has no
   * valid corresponding nsICacheEntry.
   */

uint64_t getCacheEntryId ();

pub unsafe fn GetCacheKey(&self, aCacheKey: *mut *const nsISupports) -> nsresult

/**
   * Set/get the cache key... uniquely identifies the data in the cache
   * for this channel.  Holding a reference to this key does NOT prevent
   * the cached data from being removed.
   *
   * A cache key retrieved from a particular instance of nsICacheInfoChannel
   * could be set on another instance of nsICacheInfoChannel provided the
   * underlying implementations are compatible and provided the new
   * channel instance was created with the same URI.  The implementation of
   * nsICacheInfoChannel would be expected to use the cache entry identified
   * by the cache token.  Depending on the value of nsIRequest::loadFlags,
   * the cache entry may be validated, overwritten, or simply read.
   *
   * The cache key may be NULL indicating that the URI of the channel is
   * sufficient to locate the same cache entry.  Setting a NULL cache key
   * is likewise valid.
   */

attribute nsISupports cacheKey;

pub unsafe fn SetCacheKey(&self, aCacheKey: *const nsISupports) -> nsresult

/**
   * Set/get the cache key... uniquely identifies the data in the cache
   * for this channel.  Holding a reference to this key does NOT prevent
   * the cached data from being removed.
   *
   * A cache key retrieved from a particular instance of nsICacheInfoChannel
   * could be set on another instance of nsICacheInfoChannel provided the
   * underlying implementations are compatible and provided the new
   * channel instance was created with the same URI.  The implementation of
   * nsICacheInfoChannel would be expected to use the cache entry identified
   * by the cache token.  Depending on the value of nsIRequest::loadFlags,
   * the cache entry may be validated, overwritten, or simply read.
   *
   * The cache key may be NULL indicating that the URI of the channel is
   * sufficient to locate the same cache entry.  Setting a NULL cache key
   * is likewise valid.
   */

attribute nsISupports cacheKey;

pub unsafe fn GetAllowStaleCacheContent(
&self,
aAllowStaleCacheContent: *mut bool
) -> nsresult

/**
   * Tells the channel to behave as if the LOAD_FROM_CACHE flag has been set,
   * but without affecting the loads for the entire loadGroup in case of this
   * channel being the default load group's channel.
   */

attribute boolean allowStaleCacheContent;

pub unsafe fn SetAllowStaleCacheContent(
&self,
aAllowStaleCacheContent: bool
) -> nsresult

/**
   * Tells the channel to behave as if the LOAD_FROM_CACHE flag has been set,
   * but without affecting the loads for the entire loadGroup in case of this
   * channel being the default load group's channel.
   */

attribute boolean allowStaleCacheContent;

pub unsafe fn PreferAlternativeDataType(&self, type_: &nsACString) -> nsresult

/**
   * Calling this method instructs the channel to serve the alternative data
   * if that was previously saved in the cache, otherwise it will serve the
   * real data.
   * Must be called before AsyncOpen.
   */

void preferAlternativeDataType (in ACString type);

pub unsafe fn GetPreferredAlternativeDataType(
&self,
aPreferredAlternativeDataType: &mut nsACString
) -> nsresult

/**
   * Get the preferred alternative data type set by preferAlternativeDataType().
   * This attribute stands for the desired data type instead of the type of the
   * information retrieved from the network stack.
   */

readonly attribute ACString preferredAlternativeDataType;

pub unsafe fn GetAlternativeDataType(
&self,
aAlternativeDataType: &mut nsACString
) -> nsresult

/**
   * Holds the type of the alternative data representation that the channel
   * is returning.
   * Is empty string if no alternative data representation was requested, or
   * if the requested representation wasn't found in the cache.
   * Can only be called during or after OnStartRequest.
   */

readonly attribute ACString alternativeDataType;

pub unsafe fn OpenAlternativeOutputStream(
&self,
type_: &nsACString,
_retval: *mut *const nsIOutputStream
) -> nsresult

/**
   * Opens and returns an output stream that a consumer may use to save an
   * alternate representation of the data.
   * Must be called after the OnStopRequest that delivered the real data.
   * The consumer may choose to replace the saved alt representation.
   * Opening the output stream will fail if there are any open input streams
   * reading the already saved alt representation.
   */

nsIOutputStream openAlternativeOutputStream (in ACString type);