Struct xpcom::interfaces::nsICachingChannel

#[repr(C)]

pub struct nsICachingChannel { /* fields omitted */ }

interface nsICachingChannel : nsICacheInfoChannel

/**
 * A channel may optionally implement this interface to allow clients
 * to affect its behavior with respect to how it uses the cache service.
 *
 * This interface provides:
 *   1) Support for "stream as file" semantics (for JAR and plugins).
 *   2) Support for "pinning" cached data in the cache (for printing and save-as).
 *   3) Support for uniquely identifying cached data in cases when the URL
 *      is insufficient (e.g., HTTP form submission).
 */

Methods

impl nsICachingChannel

pub fn coerce<T: nsICachingChannelCoerce>(&self) -> &T

Cast this nsICachingChannel to one of its base interfaces.

impl nsICachingChannel

pub const LOAD_NO_NETWORK_IO: i64

LOAD_NO_NETWORK_IO: i64 = 67108864

/**************************************************************************
     * Caching channel specific load flags:
     */
/**
     * This load flag inhibits fetching from the net.  An error of
     * NS_ERROR_DOCUMENT_NOT_CACHED will be sent to the listener's
     * onStopRequest if network IO is necessary to complete the request.
     *
     * This flag can be used to find out whether fetching this URL would
     * cause validation of the cache entry via the network.
     *
     * Combining this flag with LOAD_BYPASS_LOCAL_CACHE will cause all
     * loads to fail. This flag differs from LOAD_ONLY_FROM_CACHE in that
     * this flag fails the load if validation is required while
     * LOAD_ONLY_FROM_CACHE skips validation where possible.
     */

pub const LOAD_CHECK_OFFLINE_CACHE: i64

LOAD_CHECK_OFFLINE_CACHE: i64 = 134217728

/**
     * This load flag causes the offline cache to be checked when fetching
     * a request.  It will be set automatically if the browser is offline.
     *
     * This flag will not be transferred through a redirect.
     */

pub const LOAD_BYPASS_LOCAL_CACHE: i64

LOAD_BYPASS_LOCAL_CACHE: i64 = 268435456

/**
     * This load flag causes the local cache to be skipped when fetching a
     * request.  Unlike LOAD_BYPASS_CACHE, it does not force an end-to-end load
     * (i.e., it does not affect proxy caches).
     */

pub const LOAD_BYPASS_LOCAL_CACHE_IF_BUSY: i64

LOAD_BYPASS_LOCAL_CACHE_IF_BUSY: i64 = 536870912

/**
     * This load flag causes the local cache to be skipped if the request
     * would otherwise block waiting to access the cache.
     */

pub const LOAD_ONLY_FROM_CACHE: i64

LOAD_ONLY_FROM_CACHE: i64 = 1073741824

/**
     * This load flag inhibits fetching from the net if the data in the cache
     * has been evicted.  An error of NS_ERROR_DOCUMENT_NOT_CACHED will be sent
     * to the listener's onStopRequest in this case.  This flag is set
     * automatically when the application is offline.
     */

pub const LOAD_ONLY_IF_MODIFIED: i64

LOAD_ONLY_IF_MODIFIED: i64 = 2147483648

/**
     * This load flag controls what happens when a document would be loaded
     * from the cache to satisfy a call to AsyncOpen.  If this attribute is
     * set to TRUE, then the document will not be loaded from the cache.  A
     * stream listener can check nsICachingChannel::isFromCache to determine
     * if the AsyncOpen will actually result in data being streamed.
     *
     * If this flag has been set, and the request can be satisfied via the
     * cache, then the OnDataAvailable events will be skipped.  The listener
     * will only see OnStartRequest followed by OnStopRequest.
     */

pub unsafe fn GetCacheToken(
&self,
aCacheToken: *mut *const nsISupports
) -> nsresult

/**
     * Set/get the cache token... uniquely identifies the data in the cache.
     * Holding a reference to this token prevents the cached data from being
     * removed.
     *
     * A cache token retrieved from a particular instance of nsICachingChannel
     * could be set on another instance of nsICachingChannel provided the
     * underlying implementations are compatible.  The implementation of
     * nsICachingChannel would be expected to only read from the cache entry
     * identified by the cache token and not try to validate it.
     *
     * The cache token can be QI'd to a nsICacheEntryInfo if more detail
     * about the cache entry is needed (e.g., expiration time).
     */

attribute nsISupports cacheToken;

pub unsafe fn SetCacheToken(&self, aCacheToken: *const nsISupports) -> nsresult

/**
     * Set/get the cache token... uniquely identifies the data in the cache.
     * Holding a reference to this token prevents the cached data from being
     * removed.
     *
     * A cache token retrieved from a particular instance of nsICachingChannel
     * could be set on another instance of nsICachingChannel provided the
     * underlying implementations are compatible.  The implementation of
     * nsICachingChannel would be expected to only read from the cache entry
     * identified by the cache token and not try to validate it.
     *
     * The cache token can be QI'd to a nsICacheEntryInfo if more detail
     * about the cache entry is needed (e.g., expiration time).
     */

attribute nsISupports cacheToken;

pub unsafe fn GetOfflineCacheToken(
&self,
aOfflineCacheToken: *mut *const nsISupports
) -> nsresult

/**
     * The same as above but accessing the offline app cache token if there
     * is any.
     *
     * @throws
     *      NS_ERROR_NOT_AVAILABLE when there is not offline cache token
     */

attribute nsISupports offlineCacheToken;

pub unsafe fn SetOfflineCacheToken(
&self,
aOfflineCacheToken: *const nsISupports
) -> nsresult

/**
     * The same as above but accessing the offline app cache token if there
     * is any.
     *
     * @throws
     *      NS_ERROR_NOT_AVAILABLE when there is not offline cache token
     */

attribute nsISupports offlineCacheToken;

pub unsafe fn GetCacheOnlyMetadata(
&self,
aCacheOnlyMetadata: *mut bool
) -> nsresult

/**
     * Instructs the channel to only store the metadata of the entry, and not
     * the content. When reading an existing entry, this automatically sets
     * LOAD_ONLY_IF_MODIFIED flag.
     * Must be called before asyncOpen().
     */

attribute boolean cacheOnlyMetadata;

pub unsafe fn SetCacheOnlyMetadata(&self, aCacheOnlyMetadata: bool) -> nsresult

/**
     * Instructs the channel to only store the metadata of the entry, and not
     * the content. When reading an existing entry, this automatically sets
     * LOAD_ONLY_IF_MODIFIED flag.
     * Must be called before asyncOpen().
     */

attribute boolean cacheOnlyMetadata;

pub unsafe fn GetPin(&self, aPin: *mut bool) -> nsresult

/**
     * Tells the channel to use the pinning storage.
     */

attribute boolean pin;

pub unsafe fn SetPin(&self, aPin: bool) -> nsresult

/**
     * Tells the channel to use the pinning storage.
     */

attribute boolean pin;

pub unsafe fn ForceCacheEntryValidFor(
&self,
aSecondsToTheFuture: uint32_t
) -> nsresult

/**
     * Overrides cache validation for a time specified in seconds.
     *
     * @param aSecondsToTheFuture
     *
     */

void forceCacheEntryValidFor (in unsigned long aSecondsToTheFuture);