Struct xpcom::interfaces::nsIClientAuthDialogs

#[repr(C)]

pub struct nsIClientAuthDialogs { /* fields omitted */ }

interface nsIClientAuthDialogs : nsISupports

/**
 * Provides UI for SSL client-auth dialogs.
 */

Methods

impl nsIClientAuthDialogs

pub fn coerce<T: nsIClientAuthDialogsCoerce>(&self) -> &T

Cast this nsIClientAuthDialogs to one of its base interfaces.

impl nsIClientAuthDialogs

pub unsafe fn ChooseCertificate(
&self,
ctx: *const nsIInterfaceRequestor,
hostname: &nsACString,
port: int32_t,
organization: &nsACString,
issuerOrg: &nsACString,
certList: *const nsIArray,
selectedIndex: *mut uint32_t,
_retval: *mut bool
) -> nsresult

/**
   * Called when a user is asked to choose a certificate for client auth.
   *
   * @param ctx Context that allows at least nsIClientAuthUserDecision to be
   *            queried.
   * @param hostname Hostname of the server.
   * @param port Port of the server.
   * @param organization Organization field of the server cert.
   * @param issuerOrg Organization field of the issuer cert of the server cert.
   * @param certList List of certificates the user can choose from.
   * @param selectedIndex Index of the cert in |certList| that the user chose.
   *                      Ignored if the return value is false.
   * @return true if a certificate was chosen. false if the user canceled.
   */

[must_use] boolean chooseCertificate (in nsIInterfaceRequestor ctx, in AUTF8String hostname, in long port, in AUTF8String organization, in AUTF8String issuerOrg, in nsIArray certList, out unsigned long selectedIndex);