Struct xpcom::interfaces::nsIClipboardCommands

#[repr(C)]

pub struct nsIClipboardCommands { /* fields omitted */ }

interface nsIClipboardCommands : nsISupports

/**
 * An interface for embedding clients who wish to interact with
 * the system-wide OS clipboard. Mozilla does not use a private
 * clipboard, instead it places its data directly onto the system
 * clipboard. The webshell implements this interface.
 */

Methods

impl nsIClipboardCommands

pub fn coerce<T: nsIClipboardCommandsCoerce>(&self) -> &T

Cast this nsIClipboardCommands to one of its base interfaces.

impl nsIClipboardCommands

pub unsafe fn CanCutSelection(&self, _retval: *mut bool) -> nsresult

/**
   * Returns whether there is a selection and it is not read-only.
   *
   * @return <code>true</code> if the current selection can be cut,
   *          <code>false</code> otherwise.
   */

boolean canCutSelection ();

pub unsafe fn CanCopySelection(&self, _retval: *mut bool) -> nsresult

/**
   * Returns whether there is a selection and it is copyable.
   *
   * @return <code>true</code> if there is a selection,
   *          <code>false</code> otherwise.
   */

boolean canCopySelection ();

pub unsafe fn CanCopyLinkLocation(&self, _retval: *mut bool) -> nsresult

/**
   * Returns whether we can copy a link location.
   *
   * @return <code>true</code> if a link is selected,
   *           <code>false</code> otherwise.
   */

boolean canCopyLinkLocation ();

pub unsafe fn CanCopyImageLocation(&self, _retval: *mut bool) -> nsresult

/**
   * Returns whether we can copy an image location.
   *
   * @return <code>true</code> if an image is selected,
              <code>false</code> otherwise.
   */

boolean canCopyImageLocation ();

pub unsafe fn CanCopyImageContents(&self, _retval: *mut bool) -> nsresult

/**
   * Returns whether we can copy an image's contents.
   *
   * @return <code>true</code> if an image is selected,
   *          <code>false</code> otherwise
   */

boolean canCopyImageContents ();

pub unsafe fn CanPaste(&self, _retval: *mut bool) -> nsresult

/**
   * Returns whether the current contents of the clipboard can be
   * pasted and if the current selection is not read-only.
   *
   * @return <code>true</code> there is data to paste on the clipboard
   *          and the current selection is not read-only,
   *          <code>false</code> otherwise
   */

boolean canPaste ();

pub unsafe fn CutSelection(&self) -> nsresult

/**
   * Cut the current selection onto the clipboard.
   */

void cutSelection ();

pub unsafe fn CopySelection(&self) -> nsresult

/**
   * Copy the current selection onto the clipboard.
   */

void copySelection ();

pub unsafe fn CopyLinkLocation(&self) -> nsresult

/**
   * Copy the link location of the current selection (e.g.,
   * the |href| attribute of a selected |a| tag).
   */

void copyLinkLocation ();

pub unsafe fn CopyImageLocation(&self) -> nsresult

/**
   * Copy the location of the selected image.
   */

void copyImageLocation ();

pub unsafe fn CopyImageContents(&self) -> nsresult

/**
   * Copy the contents of the selected image.
   */

void copyImageContents ();

pub unsafe fn Paste(&self) -> nsresult

/**
   * Paste the contents of the clipboard into the current selection.
   */

void paste ();

pub unsafe fn SelectAll(&self) -> nsresult

/**
   * Select the entire contents.
   */

void selectAll ();

pub unsafe fn SelectNone(&self) -> nsresult

/**
   * Clear the current selection (if any). Insertion point ends up
   * at beginning of current selection.
   */

void selectNone ();