Struct xpcom::interfaces::nsIComponentManager

#[repr(C)]

pub struct nsIComponentManager { /* fields omitted */ }

interface nsIComponentManager : nsISupports

Methods

impl nsIComponentManager

pub fn coerce<T: nsIComponentManagerCoerce>(&self) -> &T

Cast this nsIComponentManager to one of its base interfaces.

impl nsIComponentManager

pub unsafe fn GetClassObject(
&self,
aClass: &nsCID,
aIID: &nsIID,
result: *mut *mut c_void
) -> nsresult

/**
     * getClassObject
     *
     * Returns the factory object that can be used to create instances of
     * CID aClass
     *
     * @param aClass The classid of the factory that is being requested
     */

void getClassObject (in nsCIDRef aClass, in nsIIDRef aIID, [iid_is (aIID), retval] out nsQIResult result);

pub unsafe fn GetClassObjectByContractID(
&self,
aContractID: *const c_char,
aIID: &nsIID,
result: *mut *mut c_void
) -> nsresult

/**
     * getClassObjectByContractID
     *
     * Returns the factory object that can be used to create instances of
     * CID aClass
     *
     * @param aClass The classid of the factory that is being requested
     */

void getClassObjectByContractID (in string aContractID, in nsIIDRef aIID, [iid_is (aIID), retval] out nsQIResult result);

pub unsafe fn CreateInstance(
&self,
aClass: &nsCID,
aDelegate: *const nsISupports,
aIID: &nsIID,
result: *mut *mut c_void
) -> nsresult

/**
     * createInstance
     *
     * Create an instance of the CID aClass and return the interface aIID.
     *
     * @param aClass : ClassID of object instance requested
     * @param aDelegate : Used for aggregation
     * @param aIID : IID of interface requested
     */

void createInstance (in nsCIDRef aClass, in nsISupports aDelegate, in nsIIDRef aIID, [iid_is (aIID), retval] out nsQIResult result);

pub unsafe fn CreateInstanceByContractID(
&self,
aContractID: *const c_char,
aDelegate: *const nsISupports,
aIID: &nsIID,
result: *mut *mut c_void
) -> nsresult

/**
     * createInstanceByContractID
     *
     * Create an instance of the CID that implements aContractID and return the
     * interface aIID.
     *
     * @param aContractID : aContractID of object instance requested
     * @param aDelegate : Used for aggregation
     * @param aIID : IID of interface requested
     */

void createInstanceByContractID (in string aContractID, in nsISupports aDelegate, in nsIIDRef aIID, [iid_is (aIID), retval] out nsQIResult result);

pub unsafe fn AddBootstrappedManifestLocation(
&self,
aLocation: *const nsIFile
) -> nsresult

/**
     * addBootstrappedManifestLocation
     *
     * Adds a bootstrapped manifest location on runtime.
     *
     * @param aLocation : A directory where chrome.manifest resides,
     *                    or an XPI with it on the root.
     */

void addBootstrappedManifestLocation (in nsIFile aLocation);

pub unsafe fn RemoveBootstrappedManifestLocation(
&self,
aLocation: *const nsIFile
) -> nsresult

/**
     * removeBootstrappedManifestLocation
     *
     * Removes a bootstrapped manifest location on runtime.
     *
     * @param aLocation : A directory where chrome.manifest resides,
     *                    or an XPI with it on the root.
     */

void removeBootstrappedManifestLocation (in nsIFile aLocation);

pub unsafe fn GetManifestLocations(
&self,
_retval: *mut *const nsIArray
) -> nsresult

/**
     * getManifestLocations
     *
     * Get an array of nsIURIs of all registered and builtin manifest locations.
     */

nsIArray getManifestLocations ();