Struct xpcom::interfaces::nsIContentDispatchChooser

#[repr(C)]

pub struct nsIContentDispatchChooser { /* fields omitted */ }

interface nsIContentDispatchChooser : nsISupports

/**
 * This is used to ask a user what they would like to do with a given piece of
 * content.
 */

Methods

impl nsIContentDispatchChooser

pub fn coerce<T: nsIContentDispatchChooserCoerce>(&self) -> &T

Cast this nsIContentDispatchChooser to one of its base interfaces.

impl nsIContentDispatchChooser

pub const REASON_CANNOT_HANDLE: i64

REASON_CANNOT_HANDLE: i64 = 0

/**
  * This request is passed to the helper app dialog because Gecko can not
  * handle content of this type.
  */

pub unsafe fn Ask(
&self,
aHandler: *const nsIHandlerInfo,
aWindowContext: *const nsIInterfaceRequestor,
aURI: *const nsIURI,
aReason: uint32_t
) -> nsresult

/**
  * Asks the user what to do with the content.
  *
  * @param aHander
  *        The interface describing the details of how this content should or
  *        can be handled.
  * @param aWindowContext
  *        The parent window context to show this chooser.  This can be null,
  *        and some implementations may not care about it.  Generally, you'll
  *        want to pass an nsIDOMWindow in so the chooser can be properly
  *        parented when opened.
  * @param aURI
  *        The URI of the resource that we are asking about.
  * @param aReason
  *        The reason why we are asking (see above).
  */

void ask (in nsIHandlerInfo aHandler, in nsIInterfaceRequestor aWindowContext, in nsIURI aURI, in unsigned long aReason);