Struct xpcom::interfaces::nsIContentPrefCallback2

#[repr(C)]

pub struct nsIContentPrefCallback2 { /* fields omitted */ }

interface nsIContentPrefCallback2 : nsISupports

Methods

impl nsIContentPrefCallback2

pub fn coerce<T: nsIContentPrefCallback2Coerce>(&self) -> &T

Cast this nsIContentPrefCallback2 to one of its base interfaces.

impl nsIContentPrefCallback2

pub const COMPLETE_OK: i64

COMPLETE_OK: i64 = 0

pub const COMPLETE_ERROR: i64

COMPLETE_ERROR: i64 = 1

pub unsafe fn HandleResult(&self, pref: *const nsIContentPref) -> nsresult

/**
 * The callback used by the above methods.
 */
/**
   * For the retrieval methods, this is called once for each retrieved
   * preference.  It is not called for other methods.
   *
   * @param pref  The retrieved preference.
   */

void handleResult (in nsIContentPref pref);

pub unsafe fn HandleError(&self, error: nsresult) -> nsresult

/**
   * Called when an error occurs.  This may be called multiple times before
   * handleCompletion is called.
   *
   * @param error  A number in Components.results describing the error.
   */

void handleError (in nsresult error);

pub unsafe fn HandleCompletion(&self, reason: uint16_t) -> nsresult

/**
   * Called when the method finishes.  This will be called exactly once for
   * each method invocation, and afterward no other callback methods will be
   * called.
   *
   * @param reason  One of the COMPLETE_* values indicating the manner in which
   *                the method completed.
   */

void handleCompletion (in unsigned short reason);