Struct xpcom::interfaces::nsIContentSniffer

#[repr(C)]

pub struct nsIContentSniffer { /* fields omitted */ }

interface nsIContentSniffer : nsISupports

/**
 * Content sniffer interface. Components implementing this interface can
 * determine a MIME type from a chunk of bytes.
 */

Methods

impl nsIContentSniffer

pub fn coerce<T: nsIContentSnifferCoerce>(&self) -> &T

Cast this nsIContentSniffer to one of its base interfaces.

impl nsIContentSniffer

pub unsafe fn GetMIMETypeFromContent(
&self,
aRequest: *const nsIRequest,
aData: *const uint8_t,
aLength: uint32_t,
_retval: &mut nsACString
) -> nsresult

/**
   * Given a chunk of data, determines a MIME type. Information from the given
   * request may be used in order to make a better decision.
   *
   * @param aRequest The request where this data came from. May be null.
   * @param aData Data to check
   * @param aLength Length of the data
   *
   * @return The content type
   *
   * @throw NS_ERROR_NOT_AVAILABLE if no MIME type could be determined.
   *
   * @note Implementations should consider the request read-only. Especially,
   * they should not attempt to set the content type property that subclasses of
   * nsIRequest might offer.
   */

ACString getMIMETypeFromContent (in nsIRequest aRequest, [array, size_is (aLength), const] in octet aData, in unsigned long aLength);