Struct xpcom::interfaces::nsIContentViewer

#[repr(C)]

pub struct nsIContentViewer { /* fields omitted */ }

interface nsIContentViewer : nsISupports

Methods

impl nsIContentViewer

pub fn coerce<T: nsIContentViewerCoerce>(&self) -> &T

Cast this nsIContentViewer to one of its base interfaces.

impl nsIContentViewer

pub const ePrompt: i64

ePrompt: i64 = 0

/**
   * aPermitUnloadFlags are passed to PermitUnload to indicate what action to take
   * if a beforeunload handler wants to prompt the user.  It is also used by
   * permitUnloadInternal to ensure we only prompt once.
   *
   * ePrompt: Prompt and return the user's choice (default).
   * eDontPromptAndDontUnload: Don't prompt and return false (unload not permitted)
   *                           if the document (or its children) asks us to prompt.
   * eDontPromptAndUnload: Don't prompt and return true (unload permitted) no matter what.
   */

pub const eDontPromptAndDontUnload: i64

eDontPromptAndDontUnload: i64 = 1

pub const eDontPromptAndUnload: i64

eDontPromptAndUnload: i64 = 2

pub const eDelayResize: i64

eDelayResize: i64 = 1

/**
   * The 'aFlags' argument to setBoundsWithFlags is a set of these bits.
   */

pub unsafe fn GetContainer(
&self,
aContainer: *mut *const nsIDocShell
) -> nsresult

attribute nsIDocShell container;

pub unsafe fn SetContainer(&self, aContainer: *const nsIDocShell) -> nsresult

attribute nsIDocShell container;

pub unsafe fn LoadComplete(&self, aStatus: nsresult) -> nsresult

void loadComplete (in nsresult aStatus);

pub unsafe fn GetLoadCompleted(&self, aLoadCompleted: *mut bool) -> nsresult

[noscript] readonly attribute boolean loadCompleted;

pub unsafe fn GetIsStopped(&self, aIsStopped: *mut bool) -> nsresult

[noscript] readonly attribute boolean isStopped;

pub unsafe fn PermitUnload(
&self,
aPermitUnloadFlags: uint32_t,
_retval: *mut bool
) -> nsresult

/**
   * Overload PermitUnload method for C++ consumers with no aPermitUnloadFlags
   * argument.
   */
/**
   * Checks if the document wants to prevent unloading by firing beforeunload on
   * the document, and if it does, takes action directed by aPermitUnloadFlags.
   * The result is returned.
   */

boolean permitUnload ([optional] in unsigned long aPermitUnloadFlags);

pub unsafe fn GetInPermitUnload(&self, aInPermitUnload: *mut bool) -> nsresult

/**
   * Exposes whether we're blocked in a call to permitUnload.
   */

readonly attribute boolean inPermitUnload;

pub unsafe fn GetBeforeUnloadFiring(
&self,
aBeforeUnloadFiring: *mut bool
) -> nsresult

/**
   * Exposes whether we're in the process of firing the beforeunload event.
   * In this case, the corresponding docshell will not allow navigation.
   */

readonly attribute boolean beforeUnloadFiring;

pub unsafe fn PageHide(&self, isUnload: bool) -> nsresult

void pageHide (in boolean isUnload);

pub unsafe fn Close(&self, historyEntry: *const nsISHEntry) -> nsresult

/**
   * All users of a content viewer are responsible for calling both
   * close() and destroy(), in that order.
   *
   * close() should be called when the load of a new page for the next
   * content viewer begins, and destroy() should be called when the next
   * content viewer replaces this one.
   *
   * |historyEntry| sets the session history entry for the content viewer.  If
   * this is null, then Destroy() will be called on the document by close().
   * If it is non-null, the document will not be destroyed, and the following
   * actions will happen when destroy() is called (*):
   *  - Sanitize() will be called on the viewer's document
   *  - The content viewer will set the contentViewer property on the
   *    history entry, and release its reference (ownership reversal).
   *  - hide() will be called, and no further destruction will happen.
   *
   *  (*) unless the document is currently being printed, in which case
   *      it will never be saved in session history.
   *
   */

void close (in nsISHEntry historyEntry);

pub unsafe fn Destroy(&self) -> nsresult

void destroy ();

pub unsafe fn Stop(&self) -> nsresult

void stop ();

pub unsafe fn GetDOMDocument(
&self,
aDOMDocument: *mut *const nsIDOMDocument
) -> nsresult

attribute nsIDOMDocument DOMDocument;

pub unsafe fn SetDOMDocument(
&self,
aDOMDocument: *const nsIDOMDocument
) -> nsresult

attribute nsIDOMDocument DOMDocument;

pub unsafe fn GetDocument(&self) -> *const nsIDocument

/**
   * Returns DOMDocument as nsIDocument and without addrefing.
   */

[noscript,notxpcom] nsIDocument getDocument ();

pub unsafe fn GetPreviousViewer(
&self,
aPreviousViewer: *mut *const nsIContentViewer
) -> nsresult

/**
   * The previous content viewer, which has been |close|d but not
   * |destroy|ed.
   */

[noscript] attribute nsIContentViewer previousViewer;

pub unsafe fn SetPreviousViewer(
&self,
aPreviousViewer: *const nsIContentViewer
) -> nsresult

/**
   * The previous content viewer, which has been |close|d but not
   * |destroy|ed.
   */

[noscript] attribute nsIContentViewer previousViewer;

pub unsafe fn Move(&self, aX: int32_t, aY: int32_t) -> nsresult

void move (in long aX, in long aY);

pub unsafe fn Show(&self) -> nsresult

void show ();

pub unsafe fn Hide(&self) -> nsresult

void hide ();

pub unsafe fn GetSticky(&self, aSticky: *mut bool) -> nsresult

attribute boolean sticky;

pub unsafe fn SetSticky(&self, aSticky: bool) -> nsresult

attribute boolean sticky;

pub unsafe fn RequestWindowClose(&self, _retval: *mut bool) -> nsresult

boolean requestWindowClose ();

pub unsafe fn Open(
&self,
aState: *const nsISupports,
aSHEntry: *const nsISHEntry
) -> nsresult

/**
   * Attach the content viewer to its DOM window and docshell.
   * @param aState A state object that might be useful in attaching the DOM
   *               window.
   * @param aSHEntry The history entry that the content viewer was stored in.
   *                 The entry must have the docshells for all of the child
   *                 documents stored in its child shell list.
   */

void open (in nsISupports aState, in nsISHEntry aSHEntry);

pub unsafe fn ClearHistoryEntry(&self) -> nsresult

/**
   * Clears the current history entry.  This is used if we need to clear out
   * the saved presentation state.
   */

void clearHistoryEntry ();

pub unsafe fn SetPageMode(
&self,
aPageMode: bool,
aPrintSettings: *const nsIPrintSettings
) -> nsresult

/**
   * Change the layout to view the document with page layout (like print preview), but
   * dynamic and editable (like Galley layout).
   */

void setPageMode (in boolean aPageMode, in nsIPrintSettings aPrintSettings);

pub unsafe fn GetHistoryEntry(
&self,
aHistoryEntry: *mut *const nsISHEntry
) -> nsresult

/**
   * Get the history entry that this viewer will save itself into when
   * destroyed.  Can return null
   */

readonly attribute nsISHEntry historyEntry;

pub unsafe fn GetIsTabModalPromptAllowed(
&self,
aIsTabModalPromptAllowed: *mut bool
) -> nsresult

/**
   * Indicates when we're in a state where content shouldn't be allowed to
   * trigger a tab-modal prompt (as opposed to a window-modal prompt) because
   * we're part way through some operation (eg beforeunload) that shouldn't be
   * rentrant if the user closes the tab while the prompt is showing.
   * See bug 613800.
   */

readonly attribute boolean isTabModalPromptAllowed;

pub unsafe fn GetIsHidden(&self, aIsHidden: *mut bool) -> nsresult

/**
   * Returns whether this content viewer is in a hidden state.
   *
   * @note Only Gecko internal code should set the attribute!
   */

attribute boolean isHidden;

pub unsafe fn SetIsHidden(&self, aIsHidden: bool) -> nsresult

/**
   * Returns whether this content viewer is in a hidden state.
   *
   * @note Only Gecko internal code should set the attribute!
   */

attribute boolean isHidden;

pub unsafe fn SetDocumentInternal(
&self,
aDocument: *const nsIDocument,
aForceReuseInnerWindow: bool
) -> nsresult

[noscript] void setDocumentInternal (in nsIDocument aDocument, in boolean aForceReuseInnerWindow);

pub unsafe fn ScrollToNode(&self, node: *const nsIDOMNode) -> nsresult

void scrollToNode (in nsIDOMNode node);

pub unsafe fn GetTextZoom(&self, aTextZoom: *mut c_float) -> nsresult

/** The amount by which to scale all text. Default is 1.0. */

attribute float textZoom;

pub unsafe fn SetTextZoom(&self, aTextZoom: c_float) -> nsresult

/** The amount by which to scale all text. Default is 1.0. */

attribute float textZoom;

pub unsafe fn GetEffectiveTextZoom(
&self,
aEffectiveTextZoom: *mut c_float
) -> nsresult

/** The actual text zoom in effect, as modified by the system font scale. */

readonly attribute float effectiveTextZoom;

pub unsafe fn GetFullZoom(&self, aFullZoom: *mut c_float) -> nsresult

/** The amount by which to scale all lengths. Default is 1.0. */

attribute float fullZoom;

pub unsafe fn SetFullZoom(&self, aFullZoom: c_float) -> nsresult

/** The amount by which to scale all lengths. Default is 1.0. */

attribute float fullZoom;

pub unsafe fn GetDeviceFullZoom(
&self,
aDeviceFullZoom: *mut c_float
) -> nsresult

/**
   * The actual full zoom in effect, as modified by the device context.
   * For a requested full zoom, the device context may choose a slightly
   * different effectiveFullZoom to accomodate integer rounding of app units
   * per dev pixel. This property returns the actual zoom amount in use,
   * though it may not be good user experience to report that a requested zoom
   * of 90% is actually 89.1%, for example. This value is provided primarily to
   * support media queries of dppx values, because those queries are matched
   * against the actual native device pixel ratio and the actual full zoom.
   */

readonly attribute float deviceFullZoom;

pub unsafe fn GetOverrideDPPX(&self, aOverrideDPPX: *mut c_float) -> nsresult

/**
   * The value used to override devicePixelRatio and media queries dppx.
   * Default is 0.0, that means no overriding is done (only a positive value
   * is applied).
   */

attribute float overrideDPPX;

pub unsafe fn SetOverrideDPPX(&self, aOverrideDPPX: c_float) -> nsresult

/**
   * The value used to override devicePixelRatio and media queries dppx.
   * Default is 0.0, that means no overriding is done (only a positive value
   * is applied).
   */

attribute float overrideDPPX;

pub unsafe fn GetAuthorStyleDisabled(
&self,
aAuthorStyleDisabled: *mut bool
) -> nsresult

/** Disable entire author style level (including HTML presentation hints) */

attribute boolean authorStyleDisabled;

pub unsafe fn SetAuthorStyleDisabled(
&self,
aAuthorStyleDisabled: bool
) -> nsresult

/** Disable entire author style level (including HTML presentation hints) */

attribute boolean authorStyleDisabled;

pub unsafe fn GetForceCharacterSet(
&self,
aForceCharacterSet: &mut nsACString
) -> nsresult

/**
   * XXX comm-central only: bug 829543. Not the Character Encoding menu in
   * browser!
   */

attribute ACString forceCharacterSet;

pub unsafe fn SetForceCharacterSet(
&self,
aForceCharacterSet: &nsACString
) -> nsresult

/**
   * XXX comm-central only: bug 829543. Not the Character Encoding menu in
   * browser!
   */

attribute ACString forceCharacterSet;

pub unsafe fn GetHintCharacterSet(
&self,
aHintCharacterSet: &mut nsACString
) -> nsresult

/**
   * XXX comm-central only: bug 829543.
   */

attribute ACString hintCharacterSet;

pub unsafe fn SetHintCharacterSet(
&self,
aHintCharacterSet: &nsACString
) -> nsresult

/**
   * XXX comm-central only: bug 829543.
   */

attribute ACString hintCharacterSet;

pub unsafe fn GetHintCharacterSetSource(
&self,
aHintCharacterSetSource: *mut int32_t
) -> nsresult

/**
   * XXX comm-central only: bug 829543.
   */

attribute int32_t hintCharacterSetSource;

pub unsafe fn SetHintCharacterSetSource(
&self,
aHintCharacterSetSource: int32_t
) -> nsresult

/**
   * XXX comm-central only: bug 829543.
   */

attribute int32_t hintCharacterSetSource;

pub unsafe fn GetContentSize(
&self,
width: *mut int32_t,
height: *mut int32_t
) -> nsresult

/**
   * Requests the size of the content to the container.
   */

void getContentSize (out long width, out long height);

pub unsafe fn GetContentSizeConstrained(
&self,
maxWidth: int32_t,
maxHeight: int32_t,
width: *mut int32_t,
height: *mut int32_t
) -> nsresult

/**
   * Returns the preferred width and height of the content, constrained to the
   * given maximum values. If either maxWidth or maxHeight is less than zero,
   * that dimension is not constrained.
   *
   * All input and output values are in device pixels, rather than CSS pixels.
   */

void getContentSizeConstrained (in long maxWidth, in long maxHeight, out long width, out long height);

pub unsafe fn GetMinFontSize(&self, aMinFontSize: *mut int32_t) -> nsresult

/** The minimum font size  */

attribute long minFontSize;

pub unsafe fn SetMinFontSize(&self, aMinFontSize: int32_t) -> nsresult

/** The minimum font size  */

attribute long minFontSize;

pub unsafe fn PausePainting(&self) -> nsresult

/**
   * Instruct the refresh driver to discontinue painting until further
   * notice.
   */

void pausePainting ();

pub unsafe fn ResumePainting(&self) -> nsresult

/**
   * Instruct the refresh driver to resume painting after a previous call to
   * pausePainting().
   */

void resumePainting ();

pub unsafe fn EmulateMedium(&self, aMediaType: &nsAString) -> nsresult

void emulateMedium (in AString aMediaType);

pub unsafe fn StopEmulatingMedium(&self) -> nsresult

void stopEmulatingMedium ();