Struct xpcom::interfaces::nsIControllerCommandTable

#[repr(C)]

pub struct nsIControllerCommandTable { /* fields omitted */ }

interface nsIControllerCommandTable : nsISupports

/**
 * nsIControllerCommandTable
 *
 * An interface via which a controller can maintain a series of commands,
 * and efficiently dispatch commands to their respective handlers.
 *
 * Controllers that use an nsIControllerCommandTable should support
 * nsIInterfaceRequestor, and be able to return an interface to their
 * controller command table via getInterface().
 *
 */

Methods

impl nsIControllerCommandTable

pub fn coerce<T: nsIControllerCommandTableCoerce>(&self) -> &T

Cast this nsIControllerCommandTable to one of its base interfaces.

impl nsIControllerCommandTable

pub unsafe fn MakeImmutable(&self) -> nsresult

/**
   * Make this command table immutable, so that commands cannot
   * be registered or unregistered. Some command tables are made
   * mutable after command registration so that they can be
   * used as singletons.
   */

void makeImmutable ();

pub unsafe fn RegisterCommand(
&self,
aCommandName: *const c_char,
aCommand: *const nsIControllerCommand
) -> nsresult

/**
   * Register and unregister commands with the command table.
   *
   * @param aCommandName  the name of the command under which to register or
   *                      unregister the given command handler.
   *
   * @param aCommand      the handler for this command.
   */

void registerCommand (in string aCommandName, in nsIControllerCommand aCommand);

pub unsafe fn UnregisterCommand(
&self,
aCommandName: *const c_char,
aCommand: *const nsIControllerCommand
) -> nsresult

void unregisterCommand (in string aCommandName, in nsIControllerCommand aCommand);

pub unsafe fn FindCommandHandler(
&self,
aCommandName: *const c_char,
_retval: *mut *const nsIControllerCommand
) -> nsresult

/**
   * Find the command handler which has been registered to handle the named command.
   *
   * @param aCommandName  the name of the command to find the handler for.
   */

nsIControllerCommand findCommandHandler (in string aCommandName);

pub unsafe fn IsCommandEnabled(
&self,
aCommandName: *const c_char,
aCommandRefCon: *const nsISupports,
_retval: *mut bool
) -> nsresult

/**
   * Get whether the named command is enabled.
   *
   * @param aCommandName    the name of the command to test
   * @param aCommandRefCon  the command context data
   */

boolean isCommandEnabled (in string aCommandName, in nsISupports aCommandRefCon);

pub unsafe fn UpdateCommandState(
&self,
aCommandName: *const c_char,
aCommandRefCon: *const nsISupports
) -> nsresult

/**
   * Tell the command to update its state (if it is a state updating command)
   *
   * @param aCommandName    the name of the command to update
   * @param aCommandRefCon  the command context data
   */

void updateCommandState (in string aCommandName, in nsISupports aCommandRefCon);

pub unsafe fn SupportsCommand(
&self,
aCommandName: *const c_char,
aCommandRefCon: *const nsISupports,
_retval: *mut bool
) -> nsresult

/**
   * Get whether the named command is supported.
   *
   * @param aCommandName    the name of the command to test
   * @param aCommandRefCon  the command context data
   */

boolean supportsCommand (in string aCommandName, in nsISupports aCommandRefCon);

pub unsafe fn DoCommand(
&self,
aCommandName: *const c_char,
aCommandRefCon: *const nsISupports
) -> nsresult

/**
   * Execute the named command.
   *
   * @param aCommandName    the name of the command to execute
   * @param aCommandRefCon  the command context data
   */

void doCommand (in string aCommandName, in nsISupports aCommandRefCon);

pub unsafe fn DoCommandParams(
&self,
aCommandName: *const c_char,
aParam: *const nsICommandParams,
aCommandRefCon: *const nsISupports
) -> nsresult

void doCommandParams (in string aCommandName, in nsICommandParams aParam, in nsISupports aCommandRefCon);

pub unsafe fn GetCommandState(
&self,
aCommandName: *const c_char,
aParam: *const nsICommandParams,
aCommandRefCon: *const nsISupports
) -> nsresult

void getCommandState (in string aCommandName, in nsICommandParams aParam, in nsISupports aCommandRefCon);

pub unsafe fn GetSupportedCommands(
&self,
count: *mut uint32_t,
commands: *mut *mut *const c_char
) -> nsresult

void getSupportedCommands (out unsigned long count, [array, size_is (count), retval] out string commands);