Struct xpcom::interfaces::nsIConverterOutputStream

#[repr(C)]

pub struct nsIConverterOutputStream { /* fields omitted */ }

interface nsIConverterOutputStream : nsIUnicharOutputStream

/**
 * This interface allows writing strings to a stream, doing automatic
 * character encoding conversion.
 */

Methods

impl nsIConverterOutputStream

pub fn coerce<T: nsIConverterOutputStreamCoerce>(&self) -> &T

Cast this nsIConverterOutputStream to one of its base interfaces.

impl nsIConverterOutputStream

pub unsafe fn Init(
&self,
aOutStream: *const nsIOutputStream,
aCharset: *const c_char
) -> nsresult

/**
     * Initialize this stream. Must be called before any other method on this
     * interface, or you will crash. The output stream passed to this method
     * must not be null, or you will crash.
     *
     * @param aOutStream
     *        The underlying output stream to which the converted strings will
     *        be written.
     * @param aCharset
     *        The character set to use for encoding the characters. A null
     *        charset will be interpreted as UTF-8.
     */

void init (in nsIOutputStream aOutStream, in string aCharset);