Struct xpcom::interfaces::nsICookie

#[repr(C)]

pub struct nsICookie { /* fields omitted */ }

interface nsICookie : nsISupports

Methods

impl nsICookie

pub fn coerce<T: nsICookieCoerce>(&self) -> &T

Cast this nsICookie to one of its base interfaces.

impl nsICookie

pub const STATUS_UNKNOWN: i64

STATUS_UNKNOWN: i64 = 0

/**
     * @DEPRECATED status implementation will return STATUS_UNKNOWN in all cases.
     */

pub const STATUS_ACCEPTED: i64

STATUS_ACCEPTED: i64 = 1

pub const STATUS_DOWNGRADED: i64

STATUS_DOWNGRADED: i64 = 2

pub const STATUS_FLAGGED: i64

STATUS_FLAGGED: i64 = 3

pub const STATUS_REJECTED: i64

STATUS_REJECTED: i64 = 4

pub const POLICY_UNKNOWN: i64

POLICY_UNKNOWN: i64 = 0

/**
     * @DEPRECATED policy implementation will return POLICY_UNKNOWN in all cases.
     */

pub const POLICY_NONE: i64

POLICY_NONE: i64 = 1

pub const POLICY_NO_CONSENT: i64

POLICY_NO_CONSENT: i64 = 2

pub const POLICY_IMPLICIT_CONSENT: i64

POLICY_IMPLICIT_CONSENT: i64 = 3

pub const POLICY_EXPLICIT_CONSENT: i64

POLICY_EXPLICIT_CONSENT: i64 = 4

pub const POLICY_NO_II: i64

POLICY_NO_II: i64 = 5

pub unsafe fn GetName(&self, aName: &mut nsACString) -> nsresult

/**
     * the name of the cookie
     */

readonly attribute ACString name;

pub unsafe fn GetValue(&self, aValue: &mut nsACString) -> nsresult

/**
     * the cookie value
     */

readonly attribute AUTF8String value;

pub unsafe fn GetIsDomain(&self, aIsDomain: *mut bool) -> nsresult

/**
     * true if the cookie is a domain cookie, false otherwise
     */

readonly attribute boolean isDomain;

pub unsafe fn GetHost(&self, aHost: &mut nsACString) -> nsresult

/**
     * the host (possibly fully qualified) of the cookie
     */

readonly attribute AUTF8String host;

pub unsafe fn GetPath(&self, aPath: &mut nsACString) -> nsresult

/**
     * the path pertaining to the cookie
     */

readonly attribute AUTF8String path;

pub unsafe fn GetIsSecure(&self, aIsSecure: *mut bool) -> nsresult

/**
     * true if the cookie was transmitted over ssl, false otherwise
     */

readonly attribute boolean isSecure;

pub unsafe fn GetExpires(&self, aExpires: *mut uint64_t) -> nsresult

/**
     * @DEPRECATED use nsICookie2.expiry and nsICookie2.isSession instead.
     *
     * expiration time in seconds since midnight (00:00:00), January 1, 1970 UTC.
     * expires = 0 represents a session cookie.
     * expires = 1 represents an expiration time earlier than Jan 1, 1970.
     */

readonly attribute uint64_t expires;

pub unsafe fn GetStatus(&self, aStatus: *mut nsCookieStatus) -> nsresult

readonly attribute nsCookieStatus status;

pub unsafe fn GetPolicy(&self, aPolicy: *mut nsCookiePolicy) -> nsresult

readonly attribute nsCookiePolicy policy;