Struct xpcom::interfaces::nsICookie2

#[repr(C)]

pub struct nsICookie2 { /* fields omitted */ }

interface nsICookie2 : nsICookie

/**
 * Main cookie object interface for use by consumers:
 * extends nsICookie, a frozen interface for external
 * access of cookie objects
 */

Methods

impl nsICookie2

pub fn coerce<T: nsICookie2Coerce>(&self) -> &T

Cast this nsICookie2 to one of its base interfaces.

impl nsICookie2

pub const SAMESITE_UNSET: i64

SAMESITE_UNSET: i64 = 0

pub const SAMESITE_LAX: i64

SAMESITE_LAX: i64 = 1

pub const SAMESITE_STRICT: i64

SAMESITE_STRICT: i64 = 2

pub unsafe fn GetRawHost(&self, aRawHost: &mut nsACString) -> nsresult

/**
     * the host (possibly fully qualified) of the cookie,
     * without a leading dot to represent if it is a
     * domain cookie.
     */

readonly attribute AUTF8String rawHost;

pub unsafe fn GetIsSession(&self, aIsSession: *mut bool) -> nsresult

/**
     * true if the cookie is a session cookie.
     * note that expiry time will also be honored
     * for session cookies (see below); thus, whichever is
     * the more restrictive of the two will take effect.
     */

readonly attribute boolean isSession;

pub unsafe fn GetExpiry(&self, aExpiry: *mut int64_t) -> nsresult

/**
     * the actual expiry time of the cookie, in seconds
     * since midnight (00:00:00), January 1, 1970 UTC.
     *
     * this is distinct from nsICookie::expires, which
     * has different and obsolete semantics.
     */

readonly attribute int64_t expiry;

pub unsafe fn GetIsHttpOnly(&self, aIsHttpOnly: *mut bool) -> nsresult

/**
     * true if the cookie is an http only cookie
     */

readonly attribute boolean isHttpOnly;

pub unsafe fn GetCreationTime(&self, aCreationTime: *mut int64_t) -> nsresult

/**
     * the creation time of the cookie, in microseconds
     * since midnight (00:00:00), January 1, 1970 UTC.
     */

readonly attribute int64_t creationTime;

pub unsafe fn GetLastAccessed(&self, aLastAccessed: *mut int64_t) -> nsresult

/**
     * the last time the cookie was accessed (i.e. created,
     * modified, or read by the server), in microseconds
     * since midnight (00:00:00), January 1, 1970 UTC.
     *
     * note that this time may be approximate.
     */

readonly attribute int64_t lastAccessed;

pub unsafe fn GetSameSite(&self, aSameSite: *mut int32_t) -> nsresult

/**
     * the sameSite attribute; this controls the cookie behavior for cross-site
     * requests as per
     * https://tools.ietf.org/html/draft-west-first-party-cookies-07
     *
     * This should be one of:
     * - SAMESITE_UNSET - the SameSite attribute is not present
     * - SAMESITE_LAX - the SameSite attribute is present, but not strict
     * - SAMESITE_STRICT - the SameSite attribute is present and strict
     */

readonly attribute int32_t sameSite;