Struct xpcom::interfaces::nsICookieManager

#[repr(C)]

pub struct nsICookieManager { /* fields omitted */ }

interface nsICookieManager : nsISupports

/**
 * An optional interface for accessing or removing the cookies
 * that are in the cookie list
 */

Methods

impl nsICookieManager

pub fn coerce<T: nsICookieManagerCoerce>(&self) -> &T

Cast this nsICookieManager to one of its base interfaces.

impl nsICookieManager

pub unsafe fn RemoveAll(&self) -> nsresult

/**
   * Called to remove all cookies from the cookie list
   */

void removeAll ();

pub unsafe fn GetEnumerator(
&self,
aEnumerator: *mut *const nsISimpleEnumerator
) -> nsresult

/**
   * Called to enumerate through each cookie in the cookie list.
   * The objects enumerated over are of type nsICookie
   * This enumerator should only be used for non-private browsing cookies.
   * To retrieve an enumerator for private browsing cookies, use
   * getCookiesWithOriginAttributes.
   */

readonly attribute nsISimpleEnumerator enumerator;

pub unsafe fn GetSessionEnumerator(
&self,
aSessionEnumerator: *mut *const nsISimpleEnumerator
) -> nsresult

/**
   * Called to enumerate through each session cookie in the cookie list.
   * The objects enumerated over are of type nsICookie
   * This enumerator should only be used for non-private browsing cookies.
   */

readonly attribute nsISimpleEnumerator sessionEnumerator;

pub unsafe fn CountCookiesFromHost(
&self,
aHost: &nsACString,
_retval: *mut uint32_t
) -> nsresult

/**
   * Count how many cookies exist within the base domain of 'aHost'.
   * Thus, for a host "weather.yahoo.com", the base domain would be "yahoo.com",
   * and any host or domain cookies for "yahoo.com" and its subdomains would be
   * counted.
   *
   * @param aHost
   *        the host string to search for, e.g. "google.com". this should consist
   *        of only the host portion of a URI. see @add for a description of
   *        acceptable host strings.
   *
   * @return the number of cookies found.
   */

unsigned long countCookiesFromHost (in AUTF8String aHost);

pub unsafe fn ImportCookies(&self, aCookieFile: *const nsIFile) -> nsresult

/**
   * Import an old-style cookie file. Imported cookies will be added to the
   * existing database. If the database contains any cookies the same as those
   * being imported (i.e. domain, name, and path match), they will be replaced.
   *
   * @param aCookieFile the file to import, usually cookies.txt
   */

void importCookies (in nsIFile aCookieFile);

pub unsafe fn GetCookiesWithOriginAttributes(
&self,
aPattern: &nsAString,
aHost: &nsACString,
_retval: *mut *const nsISimpleEnumerator
) -> nsresult

/**
   * Returns an enumerator of all cookies whose origin attributes matches aPattern
   *
   * @param aPattern origin attribute pattern in JSON format
   *
   * @param aHost
   *        the host string to search for, e.g. "google.com". this should consist
   *        of only the host portion of a URI. see @add for a description of
   *        acceptable host strings. This attribute is optional. It will search
   *        all hosts if this attribute is not given.
   */

nsISimpleEnumerator getCookiesWithOriginAttributes (in DOMString aPattern, [optional] in AUTF8String aHost);

pub unsafe fn RemoveCookiesWithOriginAttributes(
&self,
aPattern: &nsAString,
aHost: &nsACString
) -> nsresult

/**
   * Remove all the cookies whose origin attributes matches aPattern
   *
   * @param aPattern origin attribute pattern in JSON format
   */

void removeCookiesWithOriginAttributes (in DOMString aPattern, [optional] in AUTF8String aHost);