Struct xpcom::interfaces::nsICrashReporter

#[repr(C)]

pub struct nsICrashReporter { /* fields omitted */ }

interface nsICrashReporter : nsISupports

/**
 * Provides access to crash reporting functionality.
 *
 * @status UNSTABLE - This interface is not frozen and will probably change in
 *                    future releases.
 */

Methods

impl nsICrashReporter

pub fn coerce<T: nsICrashReporterCoerce>(&self) -> &T

Cast this nsICrashReporter to one of its base interfaces.

impl nsICrashReporter

pub unsafe fn GetEnabled(&self, aEnabled: *mut bool) -> nsresult

/**
   * Get the enabled status of the crash reporter.
   */

readonly attribute boolean enabled;

pub unsafe fn SetEnabled(&self, enabled: bool) -> nsresult

/**
   * Enable or disable crash reporting at runtime. Not available to script
   * because the JS engine relies on proper exception handler chaining.
   */

[noscript] void setEnabled (in bool enabled);

pub unsafe fn GetServerURL(&self, aServerURL: *mut *const nsIURL) -> nsresult

/**
   * Get or set the URL to which crash reports will be submitted.
   * Only https and http URLs are allowed, as the submission is handled
   * by OS-native networking libraries.
   *
   * @throw NS_ERROR_NOT_INITIALIZED if crash reporting is not initialized
   * @throw NS_ERROR_INVALID_ARG on set if a non-http(s) URL is assigned
   * @throw NS_ERROR_FAILURE on get if no URL is set
   */

attribute nsIURL serverURL;

pub unsafe fn SetServerURL(&self, aServerURL: *const nsIURL) -> nsresult

/**
   * Get or set the URL to which crash reports will be submitted.
   * Only https and http URLs are allowed, as the submission is handled
   * by OS-native networking libraries.
   *
   * @throw NS_ERROR_NOT_INITIALIZED if crash reporting is not initialized
   * @throw NS_ERROR_INVALID_ARG on set if a non-http(s) URL is assigned
   * @throw NS_ERROR_FAILURE on get if no URL is set
   */

attribute nsIURL serverURL;

pub unsafe fn GetMinidumpPath(
&self,
aMinidumpPath: *mut *const nsIFile
) -> nsresult

/**
   * Get or set the path on the local system to which minidumps will be
   * written when a crash happens.
   *
   * @throw NS_ERROR_NOT_INITIALIZED if crash reporting is not initialized
   */

attribute nsIFile minidumpPath;

pub unsafe fn SetMinidumpPath(&self, aMinidumpPath: *const nsIFile) -> nsresult

/**
   * Get or set the path on the local system to which minidumps will be
   * written when a crash happens.
   *
   * @throw NS_ERROR_NOT_INITIALIZED if crash reporting is not initialized
   */

attribute nsIFile minidumpPath;

pub unsafe fn GetMinidumpForID(
&self,
id: &nsAString,
_retval: *mut *const nsIFile
) -> nsresult

/**
   * Get the minidump file corresponding to the specified ID.
   *
   * @param id
   *        ID of the crash. Likely a UUID.
   *
   * @return The minidump file associated with the ID.
   *
   * @throw NS_ERROR_FILE_NOT_FOUND if the minidump could not be found
   */

nsIFile getMinidumpForID (in AString id);

pub unsafe fn GetExtraFileForID(
&self,
id: &nsAString,
_retval: *mut *const nsIFile
) -> nsresult

/**
   * Get the extra file corresponding to the specified ID.
   *
   * @param id
   *        ID of the crash. Likely a UUID.
   *
   * @return The extra file associated with the ID.
   *
   * @throw NS_ERROR_FILE_NOT_FOUND if the extra file could not be found
   */

nsIFile getExtraFileForID (in AString id);

pub unsafe fn AnnotateCrashReport(
&self,
key: &nsACString,
data: &nsACString
) -> nsresult

/**
   * Add some extra data to be submitted with a crash report.
   *
   * @param key
   *        Name of the data to be added.
   * @param data
   *        Data to be added.
   *
   * @throw NS_ERROR_NOT_INITIALIZED if crash reporting not initialized
   * @throw NS_ERROR_INVALID_ARG if key or data contain invalid characters.
   *                             Invalid characters for key are '=' and
   *                             '\n'.  Invalid character for data is '\0'.
   */

void annotateCrashReport (in AUTF8String key, in AUTF8String data);

pub unsafe fn AppendAppNotesToCrashReport(&self, data: &nsACString) -> nsresult

/**
   * Append some data to the "Notes" field, to be submitted with a crash report.
   * Unlike annotateCrashReport, this method will append to existing data.
   *
   * @param data
   *        Data to be added.
   *
   * @throw NS_ERROR_NOT_INITIALIZED if crash reporting not initialized
   * @throw NS_ERROR_INVALID_ARG if data contains invalid characters.
   *                             The only invalid character is '\0'.
   */

void appendAppNotesToCrashReport (in ACString data);

pub unsafe fn RegisterAppMemory(
&self,
ptr: uint64_t,
size: uint64_t
) -> nsresult

/**
   * Register a given memory range to be included in the crash report.
   *
   * @param ptr
   *        The starting address for the bytes.
   * @param size
   *        The number of bytes to include.
   *
   * @throw NS_ERROR_NOT_INITIALIZED if crash reporting not initialized
   * @throw NS_ERROR_NOT_IMPLEMENTED if unavailable on the current OS
   */

void registerAppMemory (in unsigned long long ptr, in unsigned long long size);

pub unsafe fn WriteMinidumpForException(
&self,
aExceptionInfo: *mut c_void
) -> nsresult

/**
   * Write a minidump immediately, with the user-supplied exception
   * information. This is implemented on Windows only, because
   * SEH (structured exception handling) exists on Windows only.
   *
   * @param aExceptionInfo  EXCEPTION_INFO* provided by Window's SEH
   */

[noscript] void writeMinidumpForException (in voidPtr aExceptionInfo);

pub unsafe fn AppendObjCExceptionInfoToAppNotes(
&self,
aException: *mut c_void
) -> nsresult

/**
   * Append note containing an Obj-C exception's info.
   *
   * @param aException  NSException object to append note for
   */

[noscript] void appendObjCExceptionInfoToAppNotes (in voidPtr aException);

pub unsafe fn GetSubmitReports(&self, aSubmitReports: *mut bool) -> nsresult

/**
   * User preference for submitting crash reports.
   */

attribute boolean submitReports;

pub unsafe fn SetSubmitReports(&self, aSubmitReports: bool) -> nsresult

/**
   * User preference for submitting crash reports.
   */

attribute boolean submitReports;

pub unsafe fn UpdateCrashEventsDir(&self) -> nsresult

/**
   * Cause the crash reporter to re-evaluate where crash events should go.
   *
   * This should be called during application startup and whenever profiles
   * change.
   */

void UpdateCrashEventsDir ();

pub unsafe fn SaveMemoryReport(&self) -> nsresult

/**
   * Save an anonymized memory report file for inclusion in a future crash
   * report in this session.
   *
   * @throws NS_ERROR_NOT_INITIALIZED if crash reporting is disabled.
   */

void saveMemoryReport ();

pub unsafe fn SetTelemetrySessionId(&self, id: &nsACString) -> nsresult

/**
   * Set the telemetry session ID which is recorded in crash metadata. This is
   * saved in the crash manager and telemetry but is not submitted as a
   * crash-stats annotation.
   */

void setTelemetrySessionId (in AUTF8String id);