Struct xpcom::interfaces::nsIDNSService

#[repr(C)]

pub struct nsIDNSService { /* fields omitted */ }

interface nsIDNSService : nsISupports

/**
 * nsIDNSService
 */

Methods

impl nsIDNSService

pub fn coerce<T: nsIDNSServiceCoerce>(&self) -> &T

Cast this nsIDNSService to one of its base interfaces.

impl nsIDNSService

pub const RESOLVE_BYPASS_CACHE: i64

RESOLVE_BYPASS_CACHE: i64 = 1

/*************************************************************************
     * Listed below are the various flags that may be OR'd together to form
     * the aFlags parameter passed to asyncResolve() and resolve().
     */
/**
     * if set, this flag suppresses the internal DNS lookup cache.
     */

pub const RESOLVE_CANONICAL_NAME: i64

RESOLVE_CANONICAL_NAME: i64 = 2

/**
     * if set, the canonical name of the specified host will be queried.
     */

pub const RESOLVE_PRIORITY_MEDIUM: i64

RESOLVE_PRIORITY_MEDIUM: i64 = 4

/**
     * if set, the query is given lower priority. Medium takes precedence
     * if both are used.
     */

pub const RESOLVE_PRIORITY_LOW: i64

RESOLVE_PRIORITY_LOW: i64 = 8

pub const RESOLVE_SPECULATE: i64

RESOLVE_SPECULATE: i64 = 16

/**
     * if set, indicates request is speculative. Speculative requests
     * return errors if prefetching is disabled by configuration.
     */

pub const RESOLVE_DISABLE_IPV6: i64

RESOLVE_DISABLE_IPV6: i64 = 32

/**
     * If set, only IPv4 addresses will be returned from resolve/asyncResolve.
     */

pub const RESOLVE_OFFLINE: i64

RESOLVE_OFFLINE: i64 = 64

/**
     * If set, only literals and cached entries will be returned from resolve/
     * asyncResolve.
     */

pub const RESOLVE_DISABLE_IPV4: i64

RESOLVE_DISABLE_IPV4: i64 = 128

/**
     * If set, only IPv6 addresses will be returned from resolve/asyncResolve.
     */

pub const RESOLVE_ALLOW_NAME_COLLISION: i64

RESOLVE_ALLOW_NAME_COLLISION: i64 = 256

/**
     * If set, allow name collision results (127.0.53.53) which are normally filtered.
     */

pub unsafe fn GetMyHostName(&self, aMyHostName: &mut nsACString) -> nsresult

/**
     * @return the hostname of the operating system.
     */

readonly attribute AUTF8String myHostName;