Struct xpcom::interfaces::nsIDOMEvent

#[repr(C)]

pub struct nsIDOMEvent { /* fields omitted */ }

interface nsIDOMEvent : nsISupports

Methods

impl nsIDOMEvent

pub fn coerce<T: nsIDOMEventCoerce>(&self) -> &T

Cast this nsIDOMEvent to one of its base interfaces.

impl nsIDOMEvent

pub const NONE: i64

NONE: i64 = 0

/**
 * The nsIDOMEvent interface is the primary datatype for all events in
 * the Document Object Model.
 *
 * For more information on this interface please see
 * http://dev.w3.org/2006/webapi/DOM-Level-3-Events/html/DOM3-Events.html and
 * http://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html
 */
/**
   * The event isn't being dispatched.
   */

pub const CAPTURING_PHASE: i64

CAPTURING_PHASE: i64 = 1

/**
   * The current event phase is the capturing phase.
   */

pub const AT_TARGET: i64

AT_TARGET: i64 = 2

/**
   * The event is currently being evaluated at the target EventTarget.
   */

pub const BUBBLING_PHASE: i64

BUBBLING_PHASE: i64 = 3

/**
   * The current event phase is the bubbling phase.
   */

pub const ALT_MASK: i64

ALT_MASK: i64 = 1

pub const CONTROL_MASK: i64

CONTROL_MASK: i64 = 2

pub const SHIFT_MASK: i64

SHIFT_MASK: i64 = 4

pub const META_MASK: i64

META_MASK: i64 = 8

pub unsafe fn GetType(&self, aType: &mut nsAString) -> nsresult

/**
   * The name of the event (case-insensitive). The name must be an XML
   * name.
   */

readonly attribute DOMString type;

pub unsafe fn GetTarget(
&self,
aTarget: *mut *const nsIDOMEventTarget
) -> nsresult

/**
   * Used to indicate the EventTarget to which the event was originally
   * dispatched.
   */

readonly attribute nsIDOMEventTarget target;

pub unsafe fn GetCurrentTarget(
&self,
aCurrentTarget: *mut *const nsIDOMEventTarget
) -> nsresult

/**
   * Used to indicate the EventTarget whose EventListeners are currently
   * being processed. This is particularly useful during capturing and
   * bubbling.
   */

readonly attribute nsIDOMEventTarget currentTarget;

pub unsafe fn GetEventPhase(&self, aEventPhase: *mut uint16_t) -> nsresult

/**
   * Used to indicate which phase of event flow is currently being
   * evaluated.
   */

readonly attribute unsigned short eventPhase;

pub unsafe fn GetBubbles(&self, aBubbles: *mut bool) -> nsresult

/**
   * Used to indicate whether or not an event is a bubbling event. If the
   * event can bubble the value is true, else the value is false.
   */

readonly attribute boolean bubbles;

pub unsafe fn GetCancelable(&self, aCancelable: *mut bool) -> nsresult

/**
   * Used to indicate whether or not an event can have its default action
   * prevented. If the default action can be prevented the value is true,
   * else the value is false.
   */

readonly attribute boolean cancelable;

pub unsafe fn GetTimeStamp(&self, aTimeStamp: *mut DOMTimeStamp) -> nsresult

/**
   * Used to specify the time (in milliseconds relative to the epoch) at
   * which the event was created. Due to the fact that some systems may
   * not provide this information the value of timeStamp may be not
   * available for all events. When not available, a value of 0 will be
   * returned. Examples of epoch time are the time of the system start or
   * 0:0:0 UTC 1st January 1970.
   */

readonly attribute DOMTimeStamp timeStamp;

pub unsafe fn StopPropagation(&self) -> nsresult

/**
   * The stopPropagation method is used prevent further propagation of an
   * event during event flow. If this method is called by any
   * EventListener the event will cease propagating through the tree. The
   * event will complete dispatch to all listeners on the current
   * EventTarget before event flow stops. This method may be used during
   * any stage of event flow.
   */

void stopPropagation ();

pub unsafe fn PreventDefault(&self) -> nsresult

/**
   * If an event is cancelable, the preventDefault method is used to
   * signify that the event is to be canceled, meaning any default action
   * normally taken by the implementation as a result of the event will
   * not occur. If, during any stage of event flow, the preventDefault
   * method is called the event is canceled. Any default action associated
   * with the event will not occur. Calling this method for a
   * non-cancelable event has no effect. Once preventDefault has been
   * called it will remain in effect throughout the remainder of the
   * event's propagation. This method may be used during any stage of
   * event flow.
   */

void preventDefault ();

pub unsafe fn GetDefaultPrevented(
&self,
aDefaultPrevented: *mut bool
) -> nsresult

/**
   * Used to indicate whether preventDefault() has been called for this event.
   */

readonly attribute boolean defaultPrevented;

pub unsafe fn StopImmediatePropagation(&self) -> nsresult

/**
   * Prevents other event listeners from being triggered and,
   * unlike Event.stopPropagation() its effect is immediate.
   */

void stopImmediatePropagation ();

pub unsafe fn GetOriginalTarget(
&self,
aOriginalTarget: *mut *const nsIDOMEventTarget
) -> nsresult

/** The original target of the event, before any retargetings. */

readonly attribute nsIDOMEventTarget originalTarget;

pub unsafe fn GetExplicitOriginalTarget(
&self,
aExplicitOriginalTarget: *mut *const nsIDOMEventTarget
) -> nsresult

/**
   * The explicit original target of the event.  If the event was retargeted
   * for some reason other than an anonymous boundary crossing, this will be set
   * to the target before the retargeting occurs.  For example, mouse events
   * are retargeted to their parent node when they happen over text nodes (bug
   * 185889), and in that case .target will show the parent and
   * .explicitOriginalTarget will show the text node.
   * .explicitOriginalTarget differs from .originalTarget in that it will never
   * contain anonymous content.
   */

readonly attribute nsIDOMEventTarget explicitOriginalTarget;

pub unsafe fn GetIsTrusted(&self, aIsTrusted: *mut bool) -> nsresult

readonly attribute boolean isTrusted;

pub unsafe fn GetCancelBubble(&self, aCancelBubble: *mut bool) -> nsresult

attribute boolean cancelBubble;

pub unsafe fn SetCancelBubble(&self, aCancelBubble: bool) -> nsresult

attribute boolean cancelBubble;

pub unsafe fn DuplicatePrivateData(&self) -> nsresult

[noscript] void duplicatePrivateData ();

pub unsafe fn SetTarget(&self, aTarget: *const nsIDOMEventTarget) -> nsresult

[noscript] void setTarget (in nsIDOMEventTarget aTarget);

pub unsafe fn IsDispatchStopped(&self) -> bool

[notxpcom] boolean IsDispatchStopped ();

pub unsafe fn SetTrusted(&self, aTrusted: bool) -> c_void

[noscript,notxpcom] void SetTrusted (in boolean aTrusted);

pub unsafe fn StopCrossProcessForwarding(&self) -> nsresult

[noscript] void stopCrossProcessForwarding ();