Struct xpcom::interfaces::nsIDOMOfflineResourceList

#[repr(C)]

pub struct nsIDOMOfflineResourceList { /* fields omitted */ }

interface nsIDOMOfflineResourceList : nsISupports

Methods

impl nsIDOMOfflineResourceList

pub fn coerce<T: nsIDOMOfflineResourceListCoerce>(&self) -> &T

Cast this nsIDOMOfflineResourceList to one of its base interfaces.

impl nsIDOMOfflineResourceList

pub const UNCACHED: i64

UNCACHED: i64 = 0

/**
   * State of the application cache this object is associated with.
   */

pub const IDLE: i64

IDLE: i64 = 1

pub const CHECKING: i64

CHECKING: i64 = 2

pub const DOWNLOADING: i64

DOWNLOADING: i64 = 3

pub const UPDATEREADY: i64

UPDATEREADY: i64 = 4

pub const OBSOLETE: i64

OBSOLETE: i64 = 5

pub unsafe fn GetMozItems(&self, aMozItems: *mut *const nsISupports) -> nsresult

/**
   * Get the list of dynamically-managed entries.
   */

readonly attribute nsISupports mozItems;

pub unsafe fn MozHasItem(&self, uri: &nsAString, _retval: *mut bool) -> nsresult

/**
   * Check that an entry exists in the list of dynamically-managed entries.
   *
   * @param uri
   *        The resource to check.
   */

boolean mozHasItem (in DOMString uri);

pub unsafe fn GetMozLength(&self, aMozLength: *mut uint32_t) -> nsresult

/**
   * Get the number of dynamically-managed entries.
   * @status DEPRECATED
   *         Clients should use the "items" attribute.
   */

readonly attribute unsigned long mozLength;

pub unsafe fn MozItem(
&self,
index: uint32_t,
_retval: &mut nsAString
) -> nsresult

/**
   * Get the URI of a dynamically-managed entry.
   * @status DEPRECATED
   *         Clients should use the "items" attribute.
   */

DOMString mozItem (in unsigned long index);

pub unsafe fn MozAdd(&self, uri: &nsAString) -> nsresult

/**
   * Add an item to the list of dynamically-managed entries.  The resource
   * will be fetched into the application cache.
   *
   * @param uri
   *        The resource to add.
   */

void mozAdd (in DOMString uri);

pub unsafe fn MozRemove(&self, uri: &nsAString) -> nsresult

/**
   * Remove an item from the list of dynamically-managed entries.  If this
   * was the last reference to a URI in the application cache, the cache
   * entry will be removed.
   *
   * @param uri
   *        The resource to remove.
   */

void mozRemove (in DOMString uri);

pub unsafe fn GetStatus(&self, aStatus: *mut uint16_t) -> nsresult

readonly attribute unsigned short status;

pub unsafe fn Update(&self) -> nsresult

/**
   * Begin the application update process on the associated application cache.
   */

void update ();

pub unsafe fn SwapCache(&self) -> nsresult

/**
   * Swap in the newest version of the application cache, or disassociate
   * from the cache if the cache group is obsolete.
   */

void swapCache ();