Struct xpcom::interfaces::nsIDOMXULContainerElement

#[repr(C)]

pub struct nsIDOMXULContainerElement { /* fields omitted */ }

interface nsIDOMXULContainerElement : nsIDOMXULContainerItemElement

Methods

impl nsIDOMXULContainerElement

pub fn coerce<T: nsIDOMXULContainerElementCoerce>(&self) -> &T

Cast this nsIDOMXULContainerElement to one of its base interfaces.

impl nsIDOMXULContainerElement

pub unsafe fn AppendItem(
&self,
aLabel: &nsAString,
aValue: &nsAString,
_retval: *mut *const nsIDOMXULElement
) -> nsresult

/**
   * Creates an item for the given label and value and appends it to the
   * container.
   *
   * @param aLabel - the label for the new item
   * @param aValue - the value of the new item
   */

nsIDOMXULElement appendItem (in DOMString aLabel, in DOMString aValue);

pub unsafe fn InsertItemAt(
&self,
aIndex: int32_t,
aLabel: &nsAString,
aValue: &nsAString,
_retval: *mut *const nsIDOMXULElement
) -> nsresult

/**
   * Creates an item for the given label and value and inserts it into the
   * container at the specified position.
   *
   * @param aIndex - the index where the new item will be inserted
   * @param aLabel - the label for the new item
   * @param aValue - the value of the new item
   */

nsIDOMXULElement insertItemAt (in long aIndex, in DOMString aLabel, in DOMString aValue);

pub unsafe fn RemoveItemAt(
&self,
aIndex: int32_t,
_retval: *mut *const nsIDOMXULElement
) -> nsresult

/**
   * Removes an item from the container.
   *
   * @param aIndex - index of the item to remove
   */

nsIDOMXULElement removeItemAt (in long aIndex);

pub unsafe fn GetItemCount(&self, aItemCount: *mut uint32_t) -> nsresult

/**
   * Returns a count of items in the container.
   */

readonly attribute unsigned long itemCount;

pub unsafe fn GetIndexOfItem(
&self,
aItem: *const nsIDOMXULElement,
_retval: *mut int32_t
) -> nsresult

/**
   * Returns the index of an item or -1 if the item is not in the container.
   *
   * @param aItem - the item to determine the index of
   */

long getIndexOfItem (in nsIDOMXULElement aItem);

pub unsafe fn GetItemAtIndex(
&self,
aIndex: int32_t,
_retval: *mut *const nsIDOMXULElement
) -> nsresult

/**
   * Returns the item at a given index or null if the item is not is the
   * container.
   *
   * @param aIndex - the index of the item to return
   */

nsIDOMXULElement getItemAtIndex (in long aIndex);