Struct xpcom::interfaces::nsIDOMXULDocument

#[repr(C)]

pub struct nsIDOMXULDocument { /* fields omitted */ }

interface nsIDOMXULDocument : nsIDOMDocument

Methods

impl nsIDOMXULDocument

pub fn coerce<T: nsIDOMXULDocumentCoerce>(&self) -> &T

Cast this nsIDOMXULDocument to one of its base interfaces.

impl nsIDOMXULDocument

pub unsafe fn GetPopupNode(
&self,
aPopupNode: *mut *const nsIDOMNode
) -> nsresult

attribute nsIDOMNode popupNode;

pub unsafe fn SetPopupNode(&self, aPopupNode: *const nsIDOMNode) -> nsresult

attribute nsIDOMNode popupNode;

pub unsafe fn GetPopupRangeParent(
&self,
aPopupRangeParent: *mut *const nsIDOMNode
) -> nsresult

/**
   * These attributes correspond to trustedGetPopupNode().rangeOffset and
   * rangeParent. They will help you find where in the DOM the popup is
   * happening. Can be accessed from chrome only, and only during a popup
   * event. Accessing any other time will be an error.
   */

readonly attribute nsIDOMNode popupRangeParent;

pub unsafe fn GetPopupRangeOffset(
&self,
aPopupRangeOffset: *mut int32_t
) -> nsresult

readonly attribute long popupRangeOffset;

pub unsafe fn GetTooltipNode(
&self,
aTooltipNode: *mut *const nsIDOMNode
) -> nsresult

attribute nsIDOMNode tooltipNode;

pub unsafe fn SetTooltipNode(&self, aTooltipNode: *const nsIDOMNode) -> nsresult

attribute nsIDOMNode tooltipNode;

pub unsafe fn GetCommandDispatcher(
&self,
aCommandDispatcher: *mut *const nsIDOMXULCommandDispatcher
) -> nsresult

readonly attribute nsIDOMXULCommandDispatcher commandDispatcher;

pub unsafe fn GetWidth(&self, aWidth: *mut int32_t) -> nsresult

readonly attribute long width;

pub unsafe fn GetHeight(&self, aHeight: *mut int32_t) -> nsresult

readonly attribute long height;

pub unsafe fn GetElementsByAttribute(
&self,
name: &nsAString,
value: &nsAString,
_retval: *mut *const nsIDOMNodeList
) -> nsresult

nsIDOMNodeList getElementsByAttribute (in DOMString name, in DOMString value);

pub unsafe fn GetElementsByAttributeNS(
&self,
namespaceURI: &nsAString,
name: &nsAString,
value: &nsAString,
_retval: *mut *const nsIDOMNodeList
) -> nsresult

nsIDOMNodeList getElementsByAttributeNS (in DOMString namespaceURI, in DOMString name, in DOMString value);

pub unsafe fn AddBroadcastListenerFor(
&self,
broadcaster: *const nsIDOMElement,
observer: *const nsIDOMElement,
attr: &nsAString
) -> nsresult

void addBroadcastListenerFor (in nsIDOMElement broadcaster, in nsIDOMElement observer, in DOMString attr);

pub unsafe fn RemoveBroadcastListenerFor(
&self,
broadcaster: *const nsIDOMElement,
observer: *const nsIDOMElement,
attr: &nsAString
) -> nsresult

void removeBroadcastListenerFor (in nsIDOMElement broadcaster, in nsIDOMElement observer, in DOMString attr);

pub unsafe fn Persist(&self, id: &nsAString, attr: &nsAString) -> nsresult

void persist (in DOMString id, in DOMString attr);

pub unsafe fn GetBoxObjectFor(
&self,
elt: *const nsIDOMElement,
_retval: *mut *const nsIBoxObject
) -> nsresult

nsIBoxObject getBoxObjectFor (in nsIDOMElement elt);

pub unsafe fn LoadOverlay(
&self,
url: &nsAString,
aObserver: *const nsIObserver
) -> nsresult

/**
   * Loads a XUL overlay and merges it with the current document, notifying an
   * observer when the merge is complete.
   * @param   url
   *          The URL of the overlay to load and merge
   * @param   observer
   *          An object implementing nsIObserver that will be notified with a
   *          message of topic "xul-overlay-merged" when the merge is complete.
   *          The subject parameter of |observe| will QI to a nsIURI - the URI
   *          of the merged overlay. This parameter is optional and may be null.
   *
   * NOTICE:  In the 2.0 timeframe this API will change such that the
   *          implementation will fire a DOMXULOverlayMerged event upon merge
   *          completion rather than notifying an observer. Do not rely on this
   *          API's behavior _not_ to change because it will!
   *          - Ben Goodger (8/23/2005)
   */

void loadOverlay (in DOMString url, in nsIObserver aObserver);