Struct xpcom::interfaces::nsIDateTimeInputArea

#[repr(C)]

pub struct nsIDateTimeInputArea { /* fields omitted */ }

interface nsIDateTimeInputArea : nsISupports

Methods

impl nsIDateTimeInputArea

pub fn coerce<T: nsIDateTimeInputAreaCoerce>(&self) -> &T

Cast this nsIDateTimeInputArea to one of its base interfaces.

impl nsIDateTimeInputArea

pub unsafe fn NotifyInputElementValueChanged(&self) -> nsresult

/**
   * Called from DOM/Layout when input element value has changed.
   */

void notifyInputElementValueChanged ();

pub unsafe fn NotifyMinMaxStepAttrChanged(&self) -> nsresult

/**
   * Called from DOM/Layout when input element min, max or step attribute has
   * changed.
   */

void notifyMinMaxStepAttrChanged ();

pub unsafe fn FocusInnerTextBox(&self) -> nsresult

/**
   * Called from DOM/Layout to set focus on inner text box.
   */

void focusInnerTextBox ();

pub unsafe fn BlurInnerTextBox(&self) -> nsresult

/**
   * Called from DOM/Layout to blur inner text box.
   */

void blurInnerTextBox ();

pub unsafe fn HasBadInput(&self, _retval: *mut bool) -> nsresult

/**
   * Called from DOM/Layout to know whether the current entered value is valid.
   */

boolean hasBadInput ();

pub unsafe fn SetPickerState(&self, isOpen: bool) -> nsresult

/**
   * Set the current state of the picker, true if it's opened, false otherwise.
   */

void setPickerState (in boolean isOpen);

pub unsafe fn SetEditAttribute(
&self,
name: &nsAString,
value: &nsAString
) -> nsresult

/**
   * Set the attribute of the inner text boxes. Only "tabindex", "readonly",
   * and "disabled" are allowed.
   */

void setEditAttribute (in DOMString name, in DOMString value);

pub unsafe fn RemoveEditAttribute(&self, name: &nsAString) -> nsresult

/**
   * Remove the attribute of the inner text boxes. Only "tabindex", "readonly",
   * and "disabled" are allowed.
   */

void removeEditAttribute (in DOMString name);