Struct xpcom::interfaces::nsIDirectoryServiceProvider2

#[repr(C)]

pub struct nsIDirectoryServiceProvider2 { /* fields omitted */ }

interface nsIDirectoryServiceProvider2 : nsIDirectoryServiceProvider

/**
 * nsIDirectoryServiceProvider2
 *
 * An extension of nsIDirectoryServiceProvider which allows
 * multiple files to be returned for the given key.
 */

Methods

impl nsIDirectoryServiceProvider2

pub fn coerce<T: nsIDirectoryServiceProvider2Coerce>(&self) -> &T

Cast this nsIDirectoryServiceProvider2 to one of its base interfaces.

impl nsIDirectoryServiceProvider2

pub unsafe fn GetFiles(
&self,
prop: *const c_char,
_retval: *mut *const nsISimpleEnumerator
) -> nsresult

/**
  * getFiles
  *
  * Directory Service calls this when it gets a request for
  * a prop and the requested type is nsISimpleEnumerator.
  *
  * @param prop         The symbolic name of the file list.
  *
  * @return             An enumerator for a list of file locations.
  *                     The elements in the enumeration are nsIFile
  * @returnCode         NS_SUCCESS_AGGREGATE_RESULT if this result should be
  *                     aggregated with other "lower" providers.
  */

nsISimpleEnumerator getFiles (in string prop);