Struct xpcom::interfaces::nsIDocShell

#[repr(C)]

pub struct nsIDocShell { /* fields omitted */ }

interface nsIDocShell : nsIDocShellTreeItem

Methods

impl nsIDocShell

pub fn coerce<T: nsIDocShellCoerce>(&self) -> &T

Cast this nsIDocShell to one of its base interfaces.

impl nsIDocShell

pub const INTERNAL_LOAD_FLAGS_NONE: i64

INTERNAL_LOAD_FLAGS_NONE: i64 = 0

pub const INTERNAL_LOAD_FLAGS_INHERIT_PRINCIPAL: i64

INTERNAL_LOAD_FLAGS_INHERIT_PRINCIPAL: i64 = 1

pub const INTERNAL_LOAD_FLAGS_DONT_SEND_REFERRER: i64

INTERNAL_LOAD_FLAGS_DONT_SEND_REFERRER: i64 = 2

pub const INTERNAL_LOAD_FLAGS_ALLOW_THIRD_PARTY_FIXUP: i64

INTERNAL_LOAD_FLAGS_ALLOW_THIRD_PARTY_FIXUP: i64 = 4

pub const INTERNAL_LOAD_FLAGS_FIRST_LOAD: i64

INTERNAL_LOAD_FLAGS_FIRST_LOAD: i64 = 8

pub const INTERNAL_LOAD_FLAGS_BYPASS_CLASSIFIER: i64

INTERNAL_LOAD_FLAGS_BYPASS_CLASSIFIER: i64 = 16

pub const INTERNAL_LOAD_FLAGS_FORCE_ALLOW_COOKIES: i64

INTERNAL_LOAD_FLAGS_FORCE_ALLOW_COOKIES: i64 = 32

pub const INTERNAL_LOAD_FLAGS_IS_SRCDOC: i64

INTERNAL_LOAD_FLAGS_IS_SRCDOC: i64 = 64

pub const INTERNAL_LOAD_FLAGS_NO_OPENER: i64

INTERNAL_LOAD_FLAGS_NO_OPENER: i64 = 256

pub const INTERNAL_LOAD_FLAGS_FORCE_ALLOW_DATA_URI: i64

INTERNAL_LOAD_FLAGS_FORCE_ALLOW_DATA_URI: i64 = 512

pub const ENUMERATE_FORWARDS: i64

ENUMERATE_FORWARDS: i64 = 0

/**
   * Get an enumerator over this docShell and its children.
   *
   * @param aItemType  - Only include docShells of this type, or if typeAll,
   *                     include all child shells.
   *                     Uses types from nsIDocShellTreeItem.
   * @param aDirection - Whether to enumerate forwards or backwards.
   */

pub const ENUMERATE_BACKWARDS: i64

ENUMERATE_BACKWARDS: i64 = 1

pub const APP_TYPE_UNKNOWN: i64

APP_TYPE_UNKNOWN: i64 = 0

/**
   * The type of application that created this window
   */

pub const APP_TYPE_MAIL: i64

APP_TYPE_MAIL: i64 = 1

pub const APP_TYPE_EDITOR: i64

APP_TYPE_EDITOR: i64 = 2

pub const BUSY_FLAGS_NONE: i64

BUSY_FLAGS_NONE: i64 = 0

/**
   * Current busy state for DocShell
   */

pub const BUSY_FLAGS_BUSY: i64

BUSY_FLAGS_BUSY: i64 = 1

pub const BUSY_FLAGS_BEFORE_PAGE_LOAD: i64

BUSY_FLAGS_BEFORE_PAGE_LOAD: i64 = 2

pub const BUSY_FLAGS_PAGE_LOADING: i64

BUSY_FLAGS_PAGE_LOADING: i64 = 4

pub const LOAD_CMD_NORMAL: i64

LOAD_CMD_NORMAL: i64 = 1

/**
   * Load commands for the document
   */

pub const LOAD_CMD_RELOAD: i64

LOAD_CMD_RELOAD: i64 = 2

pub const LOAD_CMD_HISTORY: i64

LOAD_CMD_HISTORY: i64 = 4

pub const LOAD_CMD_PUSHSTATE: i64

LOAD_CMD_PUSHSTATE: i64 = 8

pub const FRAME_TYPE_REGULAR: i64

FRAME_TYPE_REGULAR: i64 = 0

/**
   * The type of iframe that this docshell lives.
   */

pub const FRAME_TYPE_BROWSER: i64

FRAME_TYPE_BROWSER: i64 = 1

pub const TOUCHEVENTS_OVERRIDE_DISABLED: i64

TOUCHEVENTS_OVERRIDE_DISABLED: i64 = 0

/**
   * Override platform/pref default behaviour and force-disable touch events.
   */

pub const TOUCHEVENTS_OVERRIDE_ENABLED: i64

TOUCHEVENTS_OVERRIDE_ENABLED: i64 = 1

/**
   * Override platform/pref default behaviour and force-enable touch events.
   */

pub const TOUCHEVENTS_OVERRIDE_NONE: i64

TOUCHEVENTS_OVERRIDE_NONE: i64 = 2

/**
   * Don't override the platform/pref default behaviour for touch events.
   */

pub const DISPLAY_MODE_BROWSER: i64

DISPLAY_MODE_BROWSER: i64 = 0

/**
   * Allowed CSS display modes. This needs to be kept in
   * sync with similar values in nsStyleConsts.h
   */

pub const DISPLAY_MODE_MINIMAL_UI: i64

DISPLAY_MODE_MINIMAL_UI: i64 = 1

pub const DISPLAY_MODE_STANDALONE: i64

DISPLAY_MODE_STANDALONE: i64 = 2

pub const DISPLAY_MODE_FULLSCREEN: i64

DISPLAY_MODE_FULLSCREEN: i64 = 3

pub unsafe fn LoadURI(
&self,
uri: *const nsIURI,
loadInfo: *const nsIDocShellLoadInfo,
aLoadFlags: uint32_t,
firstParty: bool
) -> nsresult

/**
   * Loads a given URI.  This will give priority to loading the requested URI
   * in the object implementing this interface.  If it can't be loaded here
   * however, the URL dispatcher will go through its normal process of content
   * loading.
   *
   * @param uri        - The URI to load.
   * @param loadInfo   - This is the extended load info for this load.  This
   *                     most often will be null, but if you need to do
   *                     additional setup for this load you can get a loadInfo
   *                     object by calling createLoadInfo.  Once you have this
   *                     object you can set the needed properties on it and
   *                     then pass it to loadURI.
   * @param aLoadFlags - Flags to modify load behaviour. Flags are defined in
   *                     nsIWebNavigation.  Note that using flags outside
   *                     LOAD_FLAGS_MASK is only allowed if passing in a
   *                     non-null loadInfo.  And even some of those might not
   *                     be allowed.  Use at your own risk.
   */

[noscript] void loadURI (in nsIURI uri, in nsIDocShellLoadInfo loadInfo, in unsigned long aLoadFlags, in boolean firstParty);

pub unsafe fn LoadStream(
&self,
aStream: *const nsIInputStream,
aURI: *const nsIURI,
aContentType: &nsACString,
aContentCharset: &nsACString,
aLoadInfo: *const nsIDocShellLoadInfo
) -> nsresult

/**
   * Loads a given stream. This will give priority to loading the requested
   * stream in the object implementing this interface. If it can't be loaded
   * here however, the URL dispatched will go through its normal process of
   * content loading.
   *
   * @param aStream         - The input stream that provides access to the data
   *                          to be loaded.  This must be a blocking, threadsafe
   *                          stream implementation.
   * @param aURI            - The URI representing the stream, or null.
   * @param aContentType    - The type (MIME) of data being loaded (empty if unknown).
   * @param aContentCharset - The charset of the data being loaded (empty if unknown).
   * @param aLoadInfo       - This is the extended load info for this load.  This
   *                          most often will be null, but if you need to do
   *                          additional setup for this load you can get a
   *                          loadInfo object by calling createLoadInfo.  Once
   *                          you have this object you can set the needed
   *                          properties on it and then pass it to loadStream.
   */

[noscript] void loadStream (in nsIInputStream aStream, in nsIURI aURI, in ACString aContentType, in ACString aContentCharset, in nsIDocShellLoadInfo aLoadInfo);

pub unsafe fn CreateLoadInfo(
&self,
loadInfo: *mut *const nsIDocShellLoadInfo
) -> nsresult

/**
   * Creates a DocShellLoadInfo object that you can manipulate and then pass
   * to loadURI.
   */

void createLoadInfo (out nsIDocShellLoadInfo loadInfo);

pub unsafe fn PrepareForNewContentModel(&self) -> nsresult

/**
   * Reset state to a new content model within the current document and the document
   * viewer.  Called by the document before initiating an out of band document.write().
   */

void prepareForNewContentModel ();

pub unsafe fn SetCurrentURI(&self, aURI: *const nsIURI) -> nsresult

/**
   * For editors and suchlike who wish to change the URI associated with the
   * document. Note if you want to get the current URI, use the read-only
   * property on nsIWebNavigation.
   */

void setCurrentURI (in nsIURI aURI);

pub unsafe fn FirePageHideNotification(&self, isUnload: bool) -> nsresult

/**
   * Notify the associated content viewer and all child docshells that they are
   * about to be hidden.  If |isUnload| is true, then the document is being
   * unloaded and all dynamic subframe history entries are removed as well.
   *
   * @param isUnload
   *        True to fire the unload event in addition to the pagehide event,
   *        and remove all dynamic subframe history entries.
   */

[noscript] void firePageHideNotification (in boolean isUnload);

pub unsafe fn GetContentViewer(
&self,
aContentViewer: *mut *const nsIContentViewer
) -> nsresult

/**
   * Content Viewer that is currently loaded for this DocShell.  This may
   * change as the underlying content changes.
   */

readonly attribute nsIContentViewer contentViewer;

pub unsafe fn GetChromeEventHandler(
&self,
aChromeEventHandler: *mut *const nsIDOMEventTarget
) -> nsresult

/**
   * This attribute allows chrome to tie in to handle DOM events that may
   * be of interest to chrome.
   */

attribute nsIDOMEventTarget chromeEventHandler;

pub unsafe fn SetChromeEventHandler(
&self,
aChromeEventHandler: *const nsIDOMEventTarget
) -> nsresult

/**
   * This attribute allows chrome to tie in to handle DOM events that may
   * be of interest to chrome.
   */

attribute nsIDOMEventTarget chromeEventHandler;

pub unsafe fn GetCustomUserAgent(
&self,
aCustomUserAgent: &mut nsAString
) -> nsresult

/**
    * This allows chrome to set a custom User agent on a specific docshell
    */

attribute DOMString customUserAgent;

pub unsafe fn SetCustomUserAgent(
&self,
aCustomUserAgent: &nsAString
) -> nsresult

/**
    * This allows chrome to set a custom User agent on a specific docshell
    */

attribute DOMString customUserAgent;

pub unsafe fn GetAllowPlugins(&self, aAllowPlugins: *mut bool) -> nsresult

/**
   * Whether to allow plugin execution
   */

attribute boolean allowPlugins;

pub unsafe fn SetAllowPlugins(&self, aAllowPlugins: bool) -> nsresult

/**
   * Whether to allow plugin execution
   */

attribute boolean allowPlugins;

pub unsafe fn GetAllowJavascript(&self, aAllowJavascript: *mut bool) -> nsresult

/**
   * Whether to allow Javascript execution
   */

attribute boolean allowJavascript;

pub unsafe fn SetAllowJavascript(&self, aAllowJavascript: bool) -> nsresult

/**
   * Whether to allow Javascript execution
   */

attribute boolean allowJavascript;

pub unsafe fn GetAllowMetaRedirects(
&self,
aAllowMetaRedirects: *mut bool
) -> nsresult

/**
   * Attribute stating if refresh based redirects can be allowed
   */

attribute boolean allowMetaRedirects;

pub unsafe fn SetAllowMetaRedirects(
&self,
aAllowMetaRedirects: bool
) -> nsresult

/**
   * Attribute stating if refresh based redirects can be allowed
   */

attribute boolean allowMetaRedirects;

pub unsafe fn GetAllowSubframes(&self, aAllowSubframes: *mut bool) -> nsresult

/**
   * Attribute stating if it should allow subframes (framesets/iframes) or not
   */

attribute boolean allowSubframes;

pub unsafe fn SetAllowSubframes(&self, aAllowSubframes: bool) -> nsresult

/**
   * Attribute stating if it should allow subframes (framesets/iframes) or not
   */

attribute boolean allowSubframes;

pub unsafe fn GetAllowImages(&self, aAllowImages: *mut bool) -> nsresult

/**
   * Attribute stating whether or not images should be loaded.
   */

attribute boolean allowImages;

pub unsafe fn SetAllowImages(&self, aAllowImages: bool) -> nsresult

/**
   * Attribute stating whether or not images should be loaded.
   */

attribute boolean allowImages;

pub unsafe fn GetAllowMedia(&self) -> bool

/**
   * Attribute stating whether or not media (audio/video) should be loaded.
   */

[infallible] attribute boolean allowMedia;

pub unsafe fn SetAllowMedia(&self, aAllowMedia: bool) -> nsresult

/**
   * Attribute stating whether or not media (audio/video) should be loaded.
   */

[infallible] attribute boolean allowMedia;

pub unsafe fn GetAllowDNSPrefetch(
&self,
aAllowDNSPrefetch: *mut bool
) -> nsresult

/**
   * Attribute that determines whether DNS prefetch is allowed for this subtree
   * of the docshell tree.  Defaults to true.  Setting this will make it take
   * effect starting with the next document loaded in the docshell.
   */

attribute boolean allowDNSPrefetch;

pub unsafe fn SetAllowDNSPrefetch(&self, aAllowDNSPrefetch: bool) -> nsresult

/**
   * Attribute that determines whether DNS prefetch is allowed for this subtree
   * of the docshell tree.  Defaults to true.  Setting this will make it take
   * effect starting with the next document loaded in the docshell.
   */

attribute boolean allowDNSPrefetch;

pub unsafe fn GetAllowWindowControl(
&self,
aAllowWindowControl: *mut bool
) -> nsresult

/**
   * Attribute that determines whether window control (move/resize) is allowed.
   */

attribute boolean allowWindowControl;

pub unsafe fn SetAllowWindowControl(
&self,
aAllowWindowControl: bool
) -> nsresult

/**
   * Attribute that determines whether window control (move/resize) is allowed.
   */

attribute boolean allowWindowControl;

pub unsafe fn GetAllowContentRetargeting(&self) -> bool

/**
   * True if the docshell allows its content to be handled by a content listener
   * other than the docshell itself, including the external helper app service,
   * and false otherwise.  Defaults to true.
   */

[infallible] attribute boolean allowContentRetargeting;

pub unsafe fn SetAllowContentRetargeting(
&self,
aAllowContentRetargeting: bool
) -> nsresult

/**
   * True if the docshell allows its content to be handled by a content listener
   * other than the docshell itself, including the external helper app service,
   * and false otherwise.  Defaults to true.
   */

[infallible] attribute boolean allowContentRetargeting;

pub unsafe fn GetAllowContentRetargetingOnChildren(&self) -> bool

/**
   * True if new child docshells should allow content retargeting.
   * Setting allowContentRetargeting also overwrites this value.
   */

[infallible] attribute boolean allowContentRetargetingOnChildren;

pub unsafe fn SetAllowContentRetargetingOnChildren(
&self,
aAllowContentRetargetingOnChildren: bool
) -> nsresult

/**
   * True if new child docshells should allow content retargeting.
   * Setting allowContentRetargeting also overwrites this value.
   */

[infallible] attribute boolean allowContentRetargetingOnChildren;

pub unsafe fn GetInheritPrivateBrowsingId(&self) -> bool

/**
   * True if this docShell should inherit the private browsing ID from
   * its parent when reparented.
   *
   * NOTE: This should *not* be set false in new code, or for docShells
   * inserted anywhere other than as children of panels.
   */

[infallible] attribute boolean inheritPrivateBrowsingId;

pub unsafe fn SetInheritPrivateBrowsingId(
&self,
aInheritPrivateBrowsingId: bool
) -> nsresult

/**
   * True if this docShell should inherit the private browsing ID from
   * its parent when reparented.
   *
   * NOTE: This should *not* be set false in new code, or for docShells
   * inserted anywhere other than as children of panels.
   */

[infallible] attribute boolean inheritPrivateBrowsingId;

pub unsafe fn GetDocShellEnumerator(
&self,
aItemType: int32_t,
aDirection: int32_t,
_retval: *mut *const nsISimpleEnumerator
) -> nsresult

nsISimpleEnumerator getDocShellEnumerator (in long aItemType, in long aDirection);

pub unsafe fn GetAppType(&self, aAppType: *mut uint32_t) -> nsresult

attribute unsigned long appType;

pub unsafe fn SetAppType(&self, aAppType: uint32_t) -> nsresult

attribute unsigned long appType;

pub unsafe fn GetAllowAuth(&self, aAllowAuth: *mut bool) -> nsresult

/**
   * certain dochshells (like the message pane)
   * should not throw up auth dialogs
   * because it can act as a password trojan
   */

attribute boolean allowAuth;

pub unsafe fn SetAllowAuth(&self, aAllowAuth: bool) -> nsresult

/**
   * certain dochshells (like the message pane)
   * should not throw up auth dialogs
   * because it can act as a password trojan
   */

attribute boolean allowAuth;

pub unsafe fn GetZoom(&self, aZoom: *mut c_float) -> nsresult

/**
   * Set/Get the document scale factor.  When setting this attribute, a
   * NS_ERROR_NOT_IMPLEMENTED error may be returned by implementations
   * not supporting zoom.  Implementations not supporting zoom should return
   * 1.0 all the time for the Get operation.  1.0 by the way is the default
   * of zoom.  This means 100% of normal scaling or in other words normal size
   * no zoom.
   */

attribute float zoom;

pub unsafe fn SetZoom(&self, aZoom: c_float) -> nsresult

/**
   * Set/Get the document scale factor.  When setting this attribute, a
   * NS_ERROR_NOT_IMPLEMENTED error may be returned by implementations
   * not supporting zoom.  Implementations not supporting zoom should return
   * 1.0 all the time for the Get operation.  1.0 by the way is the default
   * of zoom.  This means 100% of normal scaling or in other words normal size
   * no zoom.
   */

attribute float zoom;

pub unsafe fn GetMarginWidth(&self, aMarginWidth: *mut int32_t) -> nsresult

attribute long marginWidth;

pub unsafe fn SetMarginWidth(&self, aMarginWidth: int32_t) -> nsresult

attribute long marginWidth;

pub unsafe fn GetMarginHeight(&self, aMarginHeight: *mut int32_t) -> nsresult

attribute long marginHeight;

pub unsafe fn SetMarginHeight(&self, aMarginHeight: int32_t) -> nsresult

attribute long marginHeight;

pub unsafe fn TabToTreeOwner(
&self,
forward: bool,
forDocumentNavigation: bool,
_retval: *mut bool
) -> nsresult

bool tabToTreeOwner (in boolean forward, in boolean forDocumentNavigation);

pub unsafe fn GetBusyFlags(&self, aBusyFlags: *mut uint32_t) -> nsresult

readonly attribute unsigned long busyFlags;

pub unsafe fn GetLoadType(&self, aLoadType: *mut uint32_t) -> nsresult

attribute unsigned long loadType;

pub unsafe fn SetLoadType(&self, aLoadType: uint32_t) -> nsresult

attribute unsigned long loadType;

pub unsafe fn GetDefaultLoadFlags(
&self,
aDefaultLoadFlags: *mut nsLoadFlags
) -> nsresult

attribute nsLoadFlags defaultLoadFlags;

pub unsafe fn SetDefaultLoadFlags(
&self,
aDefaultLoadFlags: nsLoadFlags
) -> nsresult

attribute nsLoadFlags defaultLoadFlags;

pub unsafe fn IsBeingDestroyed(&self, _retval: *mut bool) -> nsresult

boolean isBeingDestroyed ();

pub unsafe fn GetIsExecutingOnLoadHandler(
&self,
aIsExecutingOnLoadHandler: *mut bool
) -> nsresult

readonly attribute boolean isExecutingOnLoadHandler;

pub unsafe fn GetLayoutHistoryState(
&self,
aLayoutHistoryState: *mut *const nsILayoutHistoryState
) -> nsresult

attribute nsILayoutHistoryState layoutHistoryState;

pub unsafe fn SetLayoutHistoryState(
&self,
aLayoutHistoryState: *const nsILayoutHistoryState
) -> nsresult

attribute nsILayoutHistoryState layoutHistoryState;

pub unsafe fn GetShouldSaveLayoutState(
&self,
aShouldSaveLayoutState: *mut bool
) -> nsresult

readonly attribute boolean shouldSaveLayoutState;

pub unsafe fn GetSecurityUI(
&self,
aSecurityUI: *mut *const nsISecureBrowserUI
) -> nsresult

/**
   * The SecureBrowserUI object for this docshell.  This is set by XUL
   * <browser> or nsWebBrowser for their root docshell.
   */

attribute nsISecureBrowserUI securityUI;

pub unsafe fn SetSecurityUI(
&self,
aSecurityUI: *const nsISecureBrowserUI
) -> nsresult

/**
   * The SecureBrowserUI object for this docshell.  This is set by XUL
   * <browser> or nsWebBrowser for their root docshell.
   */

attribute nsISecureBrowserUI securityUI;

pub unsafe fn GetLoadURIDelegate(
&self,
aLoadURIDelegate: *mut *const nsILoadURIDelegate
) -> nsresult

/**
   * Object used to delegate URI loading to an upper context.
   * Currently only set for GeckoView to allow handling of load requests
   * at the application level.
   */

attribute nsILoadURIDelegate loadURIDelegate;

pub unsafe fn SetLoadURIDelegate(
&self,
aLoadURIDelegate: *const nsILoadURIDelegate
) -> nsresult

/**
   * Object used to delegate URI loading to an upper context.
   * Currently only set for GeckoView to allow handling of load requests
   * at the application level.
   */

attribute nsILoadURIDelegate loadURIDelegate;

pub unsafe fn SuspendRefreshURIs(&self) -> nsresult

/**
   * Cancel the XPCOM timers for each meta-refresh URI in this docshell,
   * and this docshell's children, recursively. The meta-refresh timers can be
   * restarted using resumeRefreshURIs().  If the timers are already suspended,
   * this has no effect.
   */

void suspendRefreshURIs ();

pub unsafe fn ResumeRefreshURIs(&self) -> nsresult

/**
   * Restart the XPCOM timers for each meta-refresh URI in this docshell,
   * and this docshell's children, recursively.  If the timers are already
   * running, this has no effect.
   */

void resumeRefreshURIs ();

pub unsafe fn BeginRestore(
&self,
viewer: *const nsIContentViewer,
top: bool
) -> nsresult

/**
   * Begin firing WebProgressListener notifications for restoring a page
   * presentation. |viewer| is the content viewer whose document we are
   * starting to load.  If null, it defaults to the docshell's current content
   * viewer, creating one if necessary.  |top| should be true for the toplevel
   * docshell that is being restored; it will be set to false when this method
   * is called for child docshells.  This method will post an event to
   * complete the simulated load after returning to the event loop.
   */

void beginRestore (in nsIContentViewer viewer, in boolean top);

pub unsafe fn FinishRestore(&self) -> nsresult

/**
   * Finish firing WebProgressListener notifications and DOM events for
   * restoring a page presentation.  This should only be called via
   * beginRestore().
   */

void finishRestore ();

pub unsafe fn GetRestoringDocument(
&self,
aRestoringDocument: *mut bool
) -> nsresult

readonly attribute boolean restoringDocument;

pub unsafe fn GetUseErrorPages(&self, aUseErrorPages: *mut bool) -> nsresult

attribute boolean useErrorPages;

pub unsafe fn SetUseErrorPages(&self, aUseErrorPages: bool) -> nsresult

attribute boolean useErrorPages;

pub unsafe fn DisplayLoadError(
&self,
aError: nsresult,
aURI: *const nsIURI,
aURL: *const int16_t,
aFailedChannel: *const nsIChannel,
_retval: *mut bool
) -> nsresult

/**
   * Display a load error in a frame while keeping that frame's currentURI
   * pointing correctly to the page where the error ocurred, rather than to
   * the error document page. You must provide either the aURI or aURL parameter.
   *
   * @param  aError         The error code to be displayed
   * @param  aURI           nsIURI of the page where the error happened
   * @param  aURL           wstring of the page where the error happened
   * @param  aFailedChannel The channel related to this error
   *
   * Returns whether or not we displayed an error page (note: will always
   * return false if in-content error pages are disabled!)
   */

boolean displayLoadError (in nsresult aError, in nsIURI aURI, in wstring aURL, [optional] in nsIChannel aFailedChannel);

pub unsafe fn GetFailedChannel(
&self,
aFailedChannel: *mut *const nsIChannel
) -> nsresult

/**
   * The channel that failed to load and resulted in an error page.
   * May be null. Relevant only to error pages.
   */

readonly attribute nsIChannel failedChannel;

pub unsafe fn GetPreviousTransIndex(
&self,
aPreviousTransIndex: *mut int32_t
) -> nsresult

/**
   * Keeps track of the previous SHTransaction index and the current
   * SHTransaction index at the time that the doc shell begins to load.
   * Used for ContentViewer eviction.
   */

readonly attribute long previousTransIndex;

pub unsafe fn GetLoadedTransIndex(
&self,
aLoadedTransIndex: *mut int32_t
) -> nsresult

readonly attribute long loadedTransIndex;

pub unsafe fn HistoryPurged(&self, numEntries: int32_t) -> nsresult

/**
   * Notification that entries have been removed from the beginning of a
   * nsSHistory which has this as its rootDocShell.
   *
   * @param numEntries - The number of entries removed
   */

void historyPurged (in long numEntries);

pub unsafe fn GetCurrentDocumentChannel(
&self,
aCurrentDocumentChannel: *mut *const nsIChannel
) -> nsresult

/**
   * Gets the channel for the currently loaded document, if any.
   * For a new document load, this will be the channel of the previous document
   * until after OnLocationChange fires.
   */

readonly attribute nsIChannel currentDocumentChannel;

pub unsafe fn GetChildOffset(&self, aChildOffset: *mut int32_t) -> nsresult

/**
   * The original offset of this child in its container. This property is -1 for
   * dynamically added docShells.
   */

[noscript] attribute long childOffset;

pub unsafe fn SetChildOffset(&self, aChildOffset: int32_t) -> nsresult

/**
   * The original offset of this child in its container. This property is -1 for
   * dynamically added docShells.
   */

[noscript] attribute long childOffset;

pub unsafe fn GetIsInUnload(&self, aIsInUnload: *mut bool) -> nsresult

/**
   * Find out whether the docshell is currently in the middle of a page
   * transition. This is set just before the pagehide/unload events fire.
   */

readonly attribute boolean isInUnload;

pub unsafe fn GetChannelIsUnsafe(&self, aChannelIsUnsafe: *mut bool) -> nsresult

/**
   * Find out if the currently loaded document came from a suspicious channel
   * (such as a JAR channel where the server-returned content type isn't a
   * known JAR type).
   */

readonly attribute boolean channelIsUnsafe;

pub unsafe fn GetHasMixedActiveContentLoaded(&self) -> bool

/**
   * This attribute determines whether Mixed Active Content is loaded on the
   * document. When it is true, mixed active content was not blocked and has
   * loaded (or is about to load) on the page. When it is false, mixed active content
   * has not loaded on the page, either because there was no mixed active content
   * requests on the page or such requests were blocked by nsMixedContentBlocker.
   * This boolean is set to true in nsMixedContentBlocker if Mixed Active Content
   * is allowed (either explicitly on the page by the user or when the about:config
   * setting security.mixed_content.block_active_content is set to false).
   */

[infallible] readonly attribute boolean hasMixedActiveContentLoaded;

pub unsafe fn GetHasMixedActiveContentBlocked(&self) -> bool

/**
   * This attribute determines whether a document has Mixed Active Content
   * that has been blocked from loading. When it is true, there is definitely
   * mixed active content on a page that has been blocked by
   * nsMixedContentBlocker.  When it is false, there may or may not be mixed
   * active content on a page, but if there is, it will load. Note that if the
   * about:config setting security.mixed_content.block_active_content is set
   * false, this boolean will be false, since blocking active content has been
   * disabled.
   */

[infallible] readonly attribute boolean hasMixedActiveContentBlocked;

pub unsafe fn GetHasMixedDisplayContentLoaded(&self) -> bool

/**
   * This attribute determines whether Mixed Display Content is loaded on the
   * document. When it is true, mixed display content was not blocked and has
   * loaded (or is about to load) on the page. Similar behavior to
   * hasMixedActiveContentLoaded.
   */

[infallible] readonly attribute boolean hasMixedDisplayContentLoaded;

pub unsafe fn GetHasMixedDisplayContentBlocked(&self) -> bool

/**
   * This attribute determines whether a document has Mixed Display Content
   * that has been blocked from loading. Similar behavior to
   * hasMixedActiveContentBlocked.
   */

[infallible] readonly attribute boolean hasMixedDisplayContentBlocked;

pub unsafe fn GetHasTrackingContentBlocked(&self) -> bool

/**
   * This attribute determines whether a document has Tracking Content
   * that has been blocked from loading.
   */

[infallible] readonly attribute boolean hasTrackingContentBlocked;

pub unsafe fn GetHasTrackingContentLoaded(&self) -> bool

/**
   * This attribute determines whether Tracking Content is loaded on the
   * document. When it is true, tracking content was not blocked and has
   * loaded (or is about to load) on the page.
   */

[infallible] readonly attribute boolean hasTrackingContentLoaded;

pub unsafe fn DetachEditorFromWindow(&self) -> c_void

/**
   * Disconnects this docshell's editor from its window, and stores the
   * editor data in the open document's session history entry.  This
   * should be called only during page transitions.
   */

[noscript,notxpcom] void DetachEditorFromWindow ();

pub unsafe fn GetIsOffScreenBrowser(
&self,
aIsOffScreenBrowser: *mut bool
) -> nsresult

/**
   * If true, this browser is not visible in the traditional sense, but
   * is actively being rendered to the screen (ex. painted on a canvas)
   * and should be treated accordingly.
   **/

attribute boolean isOffScreenBrowser;

pub unsafe fn SetIsOffScreenBrowser(
&self,
aIsOffScreenBrowser: bool
) -> nsresult

/**
   * If true, this browser is not visible in the traditional sense, but
   * is actively being rendered to the screen (ex. painted on a canvas)
   * and should be treated accordingly.
   **/

attribute boolean isOffScreenBrowser;

pub unsafe fn GetPrintPreview(
&self,
aPrintPreview: *mut *const nsIWebBrowserPrint
) -> nsresult

/**
   * If the current content viewer isn't initialized for print preview,
   * it is replaced with one which is and to which an about:blank document
   * is loaded.
   */

readonly attribute nsIWebBrowserPrint printPreview;

pub unsafe fn GetCanExecuteScripts(&self) -> bool

/**
   * Whether this docshell can execute scripts based on its hierarchy.
   * The rule of thumb here is that we disable js if this docshell or any
   * of its parents disallow scripting.
   */

[infallible] readonly attribute boolean canExecuteScripts;

pub unsafe fn GetIsActive(&self, aIsActive: *mut bool) -> nsresult

/**
   * Sets whether a docshell is active. An active docshell is one that is
   * visible, and thus is not a good candidate for certain optimizations
   * like image frame discarding. Docshells are active unless told otherwise.
   */

attribute boolean isActive;

pub unsafe fn SetIsActive(&self, aIsActive: bool) -> nsresult

/**
   * Sets whether a docshell is active. An active docshell is one that is
   * visible, and thus is not a good candidate for certain optimizations
   * like image frame discarding. Docshells are active unless told otherwise.
   */

attribute boolean isActive;

pub unsafe fn SetIsPrerendered(&self) -> nsresult

/**
   * Puts the docshell in prerendering mode. noscript because we want only
   * native code to be able to put a docshell in prerendering.
   */

[noscript] void SetIsPrerendered ();

pub unsafe fn GetIsPrerendered(&self) -> bool

/**
   * Whether this docshell is in prerender mode.
   */

[infallible] readonly attribute boolean isPrerendered;

pub unsafe fn GetHistoryID(&self, aHistoryID: *mut *mut nsID) -> nsresult

/**
   * The ID of the docshell in the session history.
   */

readonly attribute nsIDPtr historyID;

pub unsafe fn HistoryID(&self) -> nsID

/**
   * Helper method for accessing this value from C++
   */

[noscript,notxpcom] nsID HistoryID ();

pub unsafe fn GetIsAppTab(&self, aIsAppTab: *mut bool) -> nsresult

/**
   * Sets whether a docshell is an app tab. An app tab docshell may behave
   * differently than a non-app tab docshell in some cases, such as when
   * handling link clicks. Docshells are not app tabs unless told otherwise.
   */

attribute boolean isAppTab;

pub unsafe fn SetIsAppTab(&self, aIsAppTab: bool) -> nsresult

/**
   * Sets whether a docshell is an app tab. An app tab docshell may behave
   * differently than a non-app tab docshell in some cases, such as when
   * handling link clicks. Docshells are not app tabs unless told otherwise.
   */

attribute boolean isAppTab;

pub unsafe fn CreateAboutBlankContentViewer(
&self,
aPrincipal: *const nsIPrincipal
) -> nsresult

/**
   * Create a new about:blank document and content viewer.
   * @param aPrincipal the principal to use for the new document.
   */

void createAboutBlankContentViewer (in nsIPrincipal aPrincipal);

pub unsafe fn ForceCreateAboutBlankContentViewer(
&self,
aPrincipal: *const nsIPrincipal
) -> nsresult

/**
   * Like createAboutBlankContentViewer, but don't check for permit unload.
   * Only used by special session history operation.
   * @param aPrincipal the principal to use for the new document.
   */

[noscript] void forceCreateAboutBlankContentViewer (in nsIPrincipal aPrincipal);

pub unsafe fn GetCharset(&self, aCharset: &mut nsACString) -> nsresult

/**
   * Upon getting, returns the canonical encoding label of the document
   * currently loaded into this docshell.
   *
   * Upon setting, sets forcedCharset for compatibility with legacy callers.
   */

attribute ACString charset;

pub unsafe fn SetCharset(&self, aCharset: &nsACString) -> nsresult

/**
   * Upon getting, returns the canonical encoding label of the document
   * currently loaded into this docshell.
   *
   * Upon setting, sets forcedCharset for compatibility with legacy callers.
   */

attribute ACString charset;

pub unsafe fn GatherCharsetMenuTelemetry(&self) -> nsresult

/**
   * Called when the user chose an encoding override from the character
   * encoding menu. Separate from the setter for the charset property to avoid
   * extensions adding noise to the data.
   */

void gatherCharsetMenuTelemetry ();

pub unsafe fn GetForcedCharset(
&self,
aForcedCharset: &mut nsACString
) -> nsresult

/**
   * The charset forced by the user.
   */

attribute ACString forcedCharset;

pub unsafe fn SetForcedCharset(&self, aForcedCharset: &nsACString) -> nsresult

/**
   * The charset forced by the user.
   */

attribute ACString forcedCharset;

pub unsafe fn GetRecordProfileTimelineMarkers(&self) -> bool

/**
   * Whether the docShell records profile timeline markers at the moment
   */

[infallible] attribute boolean recordProfileTimelineMarkers;

pub unsafe fn SetRecordProfileTimelineMarkers(
&self,
aRecordProfileTimelineMarkers: bool
) -> nsresult

/**
   * Whether the docShell records profile timeline markers at the moment
   */

[infallible] attribute boolean recordProfileTimelineMarkers;

pub unsafe fn Now(&self, _retval: *mut DOMHighResTimeStamp) -> nsresult

/**
   * Return a DOMHighResTimeStamp representing the number of
   * milliseconds from an arbitrary point in time.  The reference
   * point is shared by all DocShells and is also used by timestamps
   * on markers.
   */

DOMHighResTimeStamp now ();

pub unsafe fn AddWeakPrivacyTransitionObserver(
&self,
obs: *const nsIPrivacyTransitionObserver
) -> nsresult

/**
   * Add an observer to the list of parties to be notified when this docshell's
   * private browsing status is changed. |obs| must support weak references.
   */

void addWeakPrivacyTransitionObserver (in nsIPrivacyTransitionObserver obs);

pub unsafe fn AddWeakReflowObserver(
&self,
obs: *const nsIReflowObserver
) -> nsresult

/**
   * Add an observer to the list of parties to be notified when reflows are
   * occurring. |obs| must support weak references.
   */

void addWeakReflowObserver (in nsIReflowObserver obs);

pub unsafe fn RemoveWeakReflowObserver(
&self,
obs: *const nsIReflowObserver
) -> nsresult

/**
   * Remove an observer from the list of parties to be notified about reflows.
   */

void removeWeakReflowObserver (in nsIReflowObserver obs);

pub unsafe fn NotifyReflowObservers(
&self,
interruptible: bool,
start: DOMHighResTimeStamp,
end: DOMHighResTimeStamp
) -> nsresult

/**
   * Notify all attached observers that a reflow has just occurred.
   *
   * @param interruptible if true, the reflow was interruptible.
   * @param start         timestamp when reflow started, in milliseconds since
   *                      navigationStart (accurate to 1/1000 of a ms)
   * @param end           timestamp when reflow ended, in milliseconds since
   *                      navigationStart (accurate to 1/1000 of a ms)
   */

[noscript] void notifyReflowObservers (in bool interruptible, in DOMHighResTimeStamp start, in DOMHighResTimeStamp end);

pub unsafe fn AddWeakScrollObserver(
&self,
obs: *const nsIScrollObserver
) -> nsresult

/**
   * Add an observer to the list of parties to be notified when scroll position
   * of some elements is changed.
   */

[noscript] void addWeakScrollObserver (in nsIScrollObserver obs);

pub unsafe fn RemoveWeakScrollObserver(
&self,
obs: *const nsIScrollObserver
) -> nsresult

/**
   * Add an observer to the list of parties to be notified when scroll position
   * of some elements is changed.
   */

[noscript] void removeWeakScrollObserver (in nsIScrollObserver obs);

pub unsafe fn NotifyScrollObservers(&self) -> nsresult

/**
   * Notify all attached observers that the scroll position of some element
   * has changed.
   */

[noscript] void notifyScrollObservers ();

pub unsafe fn GetFrameType(&self) -> uint32_t

[infallible] attribute unsigned long frameType;

pub unsafe fn SetFrameType(&self, aFrameType: uint32_t) -> nsresult

[infallible] attribute unsigned long frameType;

pub unsafe fn GetIsMozBrowser(&self) -> bool

/**
   * Returns true if this docshell corresponds to an <iframe mozbrowser>.
   * <xul:browser> returns false here.
   */

[infallible] readonly attribute boolean isMozBrowser;

pub unsafe fn GetIsIsolatedMozBrowserElement(&self) -> bool

/**
   * Returns true if this docshell corresponds to an isolated <iframe
   * mozbrowser>.
   *
   * <xul:browser> is not considered to be a mozbrowser element.
   * <iframe mozbrowser noisolation> does not count as isolated since
   * isolation is disabled.  Isolation can only be disabled if the
   * containing document is chrome.
   */

[infallible] readonly attribute boolean isIsolatedMozBrowserElement;

pub unsafe fn GetIsInIsolatedMozBrowserElement(&self) -> bool

/**
   * Returns true if this docshell corresponds to an isolated <iframe
   * mozbrowser> or if the docshell is contained in an isolated <iframe
   * mozbrowser>.
   *
   * <xul:browser> is not considered to be a mozbrowser element. <iframe
   * mozbrowser noisolation> does not count as isolated since isolation is
   * disabled.  Isolation can only be disabled if the containing document is
   * chrome.
   *
   * Our notion here of "contained in" means: Walk up the docshell hierarchy in
   * this process until we hit an <iframe mozbrowser> (or until the hierarchy
   * ends).  Return true iff the docshell we stopped on has
   * isIsolatedMozBrowserElement == true.
   */

[infallible] readonly attribute boolean isInIsolatedMozBrowserElement;

pub unsafe fn GetIsInMozBrowser(&self) -> bool

/**
   * Returns true if this docshell corresponds to an <iframe mozbrowser>, or
   * if this docshell is contained in an <iframe mozbrowser>. <xul:browser>
   * returns false here.
   *
   * To compute this value, we walk up the docshell hierarchy.  If we encounter
   * a docshell with isMozBrowser before we hit the end of the hierarchy,
   * we return true.  Otherwise, we return false.
   */

[infallible] readonly attribute boolean isInMozBrowser;

pub unsafe fn GetIsTopLevelContentDocShell(&self) -> bool

/**
   * Returns true if this docshell is the top level content docshell.
   */

[infallible] readonly attribute boolean isTopLevelContentDocShell;

pub unsafe fn GetSameTypeParentIgnoreBrowserBoundaries(
&self,
_retval: *mut *const nsIDocShell
) -> nsresult

/**
   * Like nsIDocShellTreeItem::GetSameTypeParent, except this ignores <iframe
   * mozbrowser> boundaries.
   */

nsIDocShell getSameTypeParentIgnoreBrowserBoundaries ();

pub unsafe fn GetSameTypeRootTreeItemIgnoreBrowserBoundaries(
&self,
_retval: *mut *const nsIDocShell
) -> nsresult

/**
   * Like nsIDocShellTreeItem::GetSameTypeRootTreeItem, except this ignores
   * <iframe mozbrowser> boundaries.
   */

nsIDocShell getSameTypeRootTreeItemIgnoreBrowserBoundaries ();

pub unsafe fn GetAsyncPanZoomEnabled(
&self,
aAsyncPanZoomEnabled: *mut bool
) -> nsresult

/**
   * True iff asynchronous panning and zooming is enabled for this
   * docshell.
   */

readonly attribute bool asyncPanZoomEnabled;

pub unsafe fn GetSandboxFlags(&self, aSandboxFlags: *mut uint32_t) -> nsresult

/**
   * The sandbox flags on the docshell. These reflect the value of the sandbox
   * attribute of the associated IFRAME or CSP-protectable content, if
   * existent. See the HTML5 spec for more details.
   * These flags on the docshell reflect the current state of the sandbox
   * attribute, which is modifiable. They are only used when loading new
   * content, sandbox flags are also immutably set on the document when it is
   * loaded.
   * The sandbox flags of a document depend on the sandbox flags on its
   * docshell and of its parent document, if any.
   * See nsSandboxFlags.h for the possible flags.
   */

attribute unsigned long sandboxFlags;

pub unsafe fn SetSandboxFlags(&self, aSandboxFlags: uint32_t) -> nsresult

/**
   * The sandbox flags on the docshell. These reflect the value of the sandbox
   * attribute of the associated IFRAME or CSP-protectable content, if
   * existent. See the HTML5 spec for more details.
   * These flags on the docshell reflect the current state of the sandbox
   * attribute, which is modifiable. They are only used when loading new
   * content, sandbox flags are also immutably set on the document when it is
   * loaded.
   * The sandbox flags of a document depend on the sandbox flags on its
   * docshell and of its parent document, if any.
   * See nsSandboxFlags.h for the possible flags.
   */

attribute unsigned long sandboxFlags;

pub unsafe fn GetOnePermittedSandboxedNavigator(
&self,
aOnePermittedSandboxedNavigator: *mut *const nsIDocShell
) -> nsresult

/**
   * When a new browsing context is opened by a sandboxed document, it needs to
   * keep track of the browsing context that opened it, so that it can be
   * navigated by it.  This is the "one permitted sandboxed navigator".
   */

attribute nsIDocShell onePermittedSandboxedNavigator;

pub unsafe fn SetOnePermittedSandboxedNavigator(
&self,
aOnePermittedSandboxedNavigator: *const nsIDocShell
) -> nsresult

/**
   * When a new browsing context is opened by a sandboxed document, it needs to
   * keep track of the browsing context that opened it, so that it can be
   * navigated by it.  This is the "one permitted sandboxed navigator".
   */

attribute nsIDocShell onePermittedSandboxedNavigator;

pub unsafe fn GetMixedContentChannel(
&self,
aMixedContentChannel: *mut *const nsIChannel
) -> nsresult

/**
   * This member variable determines whether a document has Mixed Active Content that
   * was initially blocked from loading, but the user has choosen to override the
   * block and allow the content to load. mMixedContentChannel is set to the document's
   * channel when the user allows mixed content. The nsMixedContentBlocker content policy
   * checks if the document's root channel matches the mMixedContentChannel.  If it matches,
   * then Mixed Content is loaded.  If it does match, mixed content is blocked.
   *
   * A match implies that there is definitely mixed active content on a page that was
   * initially blocked by nsMixedContentBlocker and then allowed and loaded by the user.
   * A miss imples that IF there is mixed active content on the page AND it was
   * blocked by nsMixedContentBlocker.cpp, the user has not choosen to override
   * the block. Note that if the about:config setting
   * security.mixed_content.block_active_content is set to false, this boolean
   * will be false, mMixedContentChannel will remain null since blocking active content has
   * been disabled and hence mMixedContentChannel will never be set.
   */

attribute nsIChannel mixedContentChannel;

pub unsafe fn SetMixedContentChannel(
&self,
aMixedContentChannel: *const nsIChannel
) -> nsresult

/**
   * This member variable determines whether a document has Mixed Active Content that
   * was initially blocked from loading, but the user has choosen to override the
   * block and allow the content to load. mMixedContentChannel is set to the document's
   * channel when the user allows mixed content. The nsMixedContentBlocker content policy
   * checks if the document's root channel matches the mMixedContentChannel.  If it matches,
   * then Mixed Content is loaded.  If it does match, mixed content is blocked.
   *
   * A match implies that there is definitely mixed active content on a page that was
   * initially blocked by nsMixedContentBlocker and then allowed and loaded by the user.
   * A miss imples that IF there is mixed active content on the page AND it was
   * blocked by nsMixedContentBlocker.cpp, the user has not choosen to override
   * the block. Note that if the about:config setting
   * security.mixed_content.block_active_content is set to false, this boolean
   * will be false, mMixedContentChannel will remain null since blocking active content has
   * been disabled and hence mMixedContentChannel will never be set.
   */

attribute nsIChannel mixedContentChannel;

pub unsafe fn GetAllowMixedContentAndConnectionData(
&self,
rootHasSecureConnection: *mut bool,
allowMixedContent: *mut bool,
isRootDocShell: *mut bool
) -> nsresult

/**
   * Checks whether the channel associated with the root docShell is equal to
   * mMixedContentChannel. If they are the same, allowMixedContent is set to true.
   * Checks if the root document has a secure connection. If it is, sets
   * rootHasSecureConnection to true. If the docShell is the root doc shell,
   * isRootDocShell is set to true.
   */

void GetAllowMixedContentAndConnectionData (out boolean rootHasSecureConnection, out boolean allowMixedContent, out boolean isRootDocShell);

pub unsafe fn PluginsAllowedInCurrentDoc(&self) -> bool

/**
   * Are plugins allowed in the current document loaded in this docshell ?
   * (if there is one). This depends on whether plugins are allowed by this
   * docshell itself or if the document is sandboxed and hence plugins should
   * not be allowed.
   */

[noscript,notxpcom] bool pluginsAllowedInCurrentDoc ();

pub unsafe fn GetFullscreenAllowed(&self) -> bool

/**
   * Attribute that determines whether fullscreen is allowed to be entered for
   * this subtree of the docshell tree. This is true when all iframes containing
   * this docshell have their "allowfullscreen" attribute set to "true".
   * fullscreenAllowed is only writable at content boundaries, where it is used
   * to propagate the value of the cross process parent's iframe's
   * "allowfullscreen" attribute to the child process. Setting
   * fullscreenAllowed on docshells which aren't content boundaries throws an
   * exception.
   */

[infallible] readonly attribute boolean fullscreenAllowed;

pub unsafe fn SetFullscreenAllowed(&self, allowed: bool) -> nsresult

void setFullscreenAllowed (in boolean allowed);

pub unsafe fn OrientationLock(&self) -> uint32_t

[notxpcom] uint32_t orientationLock ();

pub unsafe fn SetOrientationLock(&self, orientationLock: uint32_t) -> c_void

[notxpcom] void setOrientationLock (in uint32_t orientationLock);

pub unsafe fn GetAffectPrivateSessionLifetime(&self) -> bool

[infallible,noscript] attribute boolean affectPrivateSessionLifetime;

pub unsafe fn SetAffectPrivateSessionLifetime(
&self,
aAffectPrivateSessionLifetime: bool
) -> nsresult

[infallible,noscript] attribute boolean affectPrivateSessionLifetime;

pub unsafe fn GetMayEnableCharacterEncodingMenu(&self) -> bool

/**
   * Indicates whether the UI may enable the character encoding menu. The UI
   * must disable the menu when this property is false.
   */

[infallible] readonly attribute boolean mayEnableCharacterEncodingMenu;

pub unsafe fn GetEditor(&self, aEditor: *mut *const nsIEditor) -> nsresult

attribute nsIEditor editor;

pub unsafe fn SetEditor(&self, aEditor: *const nsIEditor) -> nsresult

attribute nsIEditor editor;

pub unsafe fn GetEditable(&self, aEditable: *mut bool) -> nsresult

readonly attribute boolean editable;

pub unsafe fn GetHasEditingSession(
&self,
aHasEditingSession: *mut bool
) -> nsresult

readonly attribute boolean hasEditingSession;

pub unsafe fn MakeEditable(&self, inWaitForUriLoad: bool) -> nsresult

/**
   * Make this docShell editable, setting a flag that causes
   * an editor to get created, either immediately, or after
   * a url has been loaded.
   *      @param  inWaitForUriLoad    true to wait for a URI before
   *                                  creating the editor.
   */

void makeEditable (in boolean inWaitForUriLoad);

pub unsafe fn GetChildSHEntry(
&self,
aChildOffset: int32_t,
_retval: *mut *const nsISHEntry
) -> nsresult

/**
   * Get the SHEntry associated with a child docshell
   */

nsISHEntry getChildSHEntry (in long aChildOffset);

pub unsafe fn AddChildSHEntry(
&self,
aCloneReference: *const nsISHEntry,
aHistoryEntry: *const nsISHEntry,
aChildOffset: int32_t,
aLoadType: uint32_t,
aCloneChilden: bool
) -> nsresult

/**
   * Add a Child SHEntry for a frameset page, given the child's loadtype.
   * If aCloneChildren is true, then aCloneReference's children will be
   * cloned onto aHistoryEntry.
   */

void addChildSHEntry (in nsISHEntry aCloneReference, in nsISHEntry aHistoryEntry, in long aChildOffset, in unsigned long aLoadType, in boolean aCloneChilden);

pub unsafe fn GetUseGlobalHistory(
&self,
aUseGlobalHistory: *mut bool
) -> nsresult

/**
   * Whether this docshell should save entries in global history.
   */

attribute boolean useGlobalHistory;

pub unsafe fn SetUseGlobalHistory(&self, aUseGlobalHistory: bool) -> nsresult

/**
   * Whether this docshell should save entries in global history.
   */

attribute boolean useGlobalHistory;

pub unsafe fn RemoveFromSessionHistory(&self) -> nsresult

/**
   * Removes nsISHEntry objects related to this docshell from session history.
   * Use this only with subdocuments, like iframes.
   */

void removeFromSessionHistory ();

pub unsafe fn GetCreatedDynamically(
&self,
aCreatedDynamically: *mut bool
) -> nsresult

/**
   * Set when an iframe/frame is added dynamically.
   */

attribute boolean createdDynamically;

pub unsafe fn SetCreatedDynamically(
&self,
aCreatedDynamically: bool
) -> nsresult

/**
   * Set when an iframe/frame is added dynamically.
   */

attribute boolean createdDynamically;

pub unsafe fn GetCurrentSHEntry(
&self,
aEntry: *mut *const nsISHEntry,
_retval: *mut bool
) -> nsresult

/**
   * Returns false for mLSHE, true for mOSHE
   */

boolean getCurrentSHEntry (out nsISHEntry aEntry);

pub unsafe fn IsCommandEnabled(
&self,
command: *const c_char,
_retval: *mut bool
) -> nsresult

/**
    * Cherry picked parts of nsIController.
    * They are here, because we want to call these functions
    * from JS.
    */

boolean isCommandEnabled (in string command);

pub unsafe fn DoCommand(&self, command: *const c_char) -> nsresult

void doCommand (in string command);

pub unsafe fn DoCommandWithParams(
&self,
command: *const c_char,
aParams: *const nsICommandParams
) -> nsresult

void doCommandWithParams (in string command, in nsICommandParams aParams);

pub unsafe fn IsInvisible(&self) -> bool

/**
   * Invisible DocShell are dummy construct to simulate DOM windows
   * without any actual visual representation. They have to be marked
   * at construction time, to avoid any painting activity.
   */

[noscript,notxpcom] bool IsInvisible ();

pub unsafe fn SetInvisible(&self, aIsInvisibleDochsell: bool) -> c_void

[noscript,notxpcom] void SetInvisible (in bool aIsInvisibleDochsell);

pub unsafe fn GetDeviceSizeIsPageSize(&self) -> bool

/**
   * If deviceSizeIsPageSize is set to true, device-width/height media queries
   * will be calculated from the page size, not the device size.
   *
   * Used by the Responsive Design Mode and B2G Simulator.
   *
   * Default is False.
   * Default value can be overriden with
   * docshell.device_size_is_page_size pref.
   */

[infallible] attribute boolean deviceSizeIsPageSize;

pub unsafe fn SetDeviceSizeIsPageSize(
&self,
aDeviceSizeIsPageSize: bool
) -> nsresult

/**
   * If deviceSizeIsPageSize is set to true, device-width/height media queries
   * will be calculated from the page size, not the device size.
   *
   * Used by the Responsive Design Mode and B2G Simulator.
   *
   * Default is False.
   * Default value can be overriden with
   * docshell.device_size_is_page_size pref.
   */

[infallible] attribute boolean deviceSizeIsPageSize;

pub unsafe fn GetHasLoadedNonBlankURI(&self) -> bool

/**
   * This attribute determines whether a document which is not about:blank has
   * already be loaded by this docShell.
   */

[infallible] readonly attribute boolean hasLoadedNonBlankURI;

pub unsafe fn GetWindowDraggingAllowed(
&self,
aWindowDraggingAllowed: *mut bool
) -> nsresult

/**
   * Allow usage of -moz-window-dragging:drag for content docshells.
   * True for top level chrome docshells. Throws if set to false with
   * top level chrome docshell.
   */

attribute boolean windowDraggingAllowed;

pub unsafe fn SetWindowDraggingAllowed(
&self,
aWindowDraggingAllowed: bool
) -> nsresult

/**
   * Allow usage of -moz-window-dragging:drag for content docshells.
   * True for top level chrome docshells. Throws if set to false with
   * top level chrome docshell.
   */

attribute boolean windowDraggingAllowed;

pub unsafe fn GetCurrentScrollRestorationIsManual(
&self,
aCurrentScrollRestorationIsManual: *mut bool
) -> nsresult

/**
   * Sets/gets the current scroll restoration mode.
   * @see https://html.spec.whatwg.org/#dom-history-scroll-restoration
   */

attribute boolean currentScrollRestorationIsManual;

pub unsafe fn SetCurrentScrollRestorationIsManual(
&self,
aCurrentScrollRestorationIsManual: bool
) -> nsresult

/**
   * Sets/gets the current scroll restoration mode.
   * @see https://html.spec.whatwg.org/#dom-history-scroll-restoration
   */

attribute boolean currentScrollRestorationIsManual;

pub unsafe fn GetEditingSession(
&self,
aEditingSession: *mut *const nsIEditingSession
) -> nsresult

/**
   * The editing session for this docshell.
   */

readonly attribute nsIEditingSession editingSession;

pub unsafe fn GetScriptableTabChild(
&self,
aTabChild: *mut *const nsITabChild
) -> nsresult

/**
   * The tab child for this docshell.
   */

[binaryname(ScriptableTabChild)] readonly attribute nsITabChild tabChild;

pub unsafe fn GetTouchEventsOverride(
&self,
aTouchEventsOverride: *mut uint32_t
) -> nsresult

/**
   * This allows chrome to override the default choice of whether touch events
   * are available on a specific docshell. Possible values are listed below.
   */

attribute unsigned long touchEventsOverride;

pub unsafe fn SetTouchEventsOverride(
&self,
aTouchEventsOverride: uint32_t
) -> nsresult

/**
   * This allows chrome to override the default choice of whether touch events
   * are available on a specific docshell. Possible values are listed below.
   */

attribute unsigned long touchEventsOverride;

pub unsafe fn GetIsOnlyToplevelInTabGroup(&self) -> bool

/**
   * This value is `true` if its corresponding unit of related browsing contexts
   * (TabGroup) contains only 1 toplevel window, and that window is the outer
   * window corresponding to this docshell.
   *
   * The value is `false` otherwise. This is the case if the docshell is an
   * iframe, has window.opener set, or another window with window.opener
   * referring to this window exists.
   *
   * If this value is `false`, it would be web content visible for a load
   * occuring in this docshell to be performed within a different docshell.
   */

[infallible] readonly attribute boolean isOnlyToplevelInTabGroup;

pub unsafe fn GetAwaitingLargeAlloc(&self) -> bool

/**
   * Returns `true` if this docshell was created due to a Large-Allocation
   * header, and has not seen the initiating load yet.
   */

[infallible] readonly attribute boolean awaitingLargeAlloc;

pub unsafe fn GetUseTrackingProtection(
&self,
aUseTrackingProtection: *mut bool
) -> nsresult

/**
   * Attribute that determines whether tracking protection is enabled.
   */

attribute boolean useTrackingProtection;

pub unsafe fn SetUseTrackingProtection(
&self,
aUseTrackingProtection: bool
) -> nsresult

/**
   * Attribute that determines whether tracking protection is enabled.
   */

attribute boolean useTrackingProtection;

pub unsafe fn DispatchLocationChangeEvent(&self) -> nsresult

/**
  * Fire a dummy location change event asynchronously.
  */

[noscript] void dispatchLocationChangeEvent ();

pub unsafe fn GetDisplayMode(&self, aDisplayMode: *mut uint32_t) -> nsresult

/**
   * Display mode for this docshell. Defaults to DISPLAY_MODE_BROWSER.
   * Media queries only look at the value in the top-most docshell.
   */

attribute unsigned long displayMode;

pub unsafe fn SetDisplayMode(&self, aDisplayMode: uint32_t) -> nsresult

/**
   * Display mode for this docshell. Defaults to DISPLAY_MODE_BROWSER.
   * Media queries only look at the value in the top-most docshell.
   */

attribute unsigned long displayMode;