Struct xpcom::interfaces::nsIDocShellLoadInfo

#[repr(C)]

pub struct nsIDocShellLoadInfo { /* fields omitted */ }

interface nsIDocShellLoadInfo : nsISupports

Methods

impl nsIDocShellLoadInfo

pub fn coerce<T: nsIDocShellLoadInfoCoerce>(&self) -> &T

Cast this nsIDocShellLoadInfo to one of its base interfaces.

impl nsIDocShellLoadInfo

pub const loadNormal: i64

loadNormal: i64 = 0

pub const loadNormalReplace: i64

loadNormalReplace: i64 = 1

pub const loadHistory: i64

loadHistory: i64 = 2

pub const loadReloadNormal: i64

loadReloadNormal: i64 = 3

pub const loadReloadBypassCache: i64

loadReloadBypassCache: i64 = 4

pub const loadReloadBypassProxy: i64

loadReloadBypassProxy: i64 = 5

pub const loadReloadBypassProxyAndCache: i64

loadReloadBypassProxyAndCache: i64 = 6

pub const loadLink: i64

loadLink: i64 = 7

pub const loadRefresh: i64

loadRefresh: i64 = 8

pub const loadReloadCharsetChange: i64

loadReloadCharsetChange: i64 = 9

pub const loadBypassHistory: i64

loadBypassHistory: i64 = 10

pub const loadStopContent: i64

loadStopContent: i64 = 11

pub const loadStopContentAndReplace: i64

loadStopContentAndReplace: i64 = 12

pub const loadNormalExternal: i64

loadNormalExternal: i64 = 13

pub const loadNormalBypassCache: i64

loadNormalBypassCache: i64 = 14

pub const loadNormalBypassProxy: i64

loadNormalBypassProxy: i64 = 15

pub const loadNormalBypassProxyAndCache: i64

loadNormalBypassProxyAndCache: i64 = 16

pub const loadPushState: i64

loadPushState: i64 = 17

pub const loadReplaceBypassCache: i64

loadReplaceBypassCache: i64 = 18

pub const loadReloadMixedContent: i64

loadReloadMixedContent: i64 = 19

pub const loadNormalAllowMixedContent: i64

loadNormalAllowMixedContent: i64 = 20

pub const loadReloadCharsetChangeBypassCache: i64

loadReloadCharsetChangeBypassCache: i64 = 21

pub const loadReloadCharsetChangeBypassProxyAndCache: i64

loadReloadCharsetChangeBypassProxyAndCache: i64 = 22

pub unsafe fn GetReferrer(&self, aReferrer: *mut *const nsIURI) -> nsresult

/** This is the referrer for the load. */

attribute nsIURI referrer;

pub unsafe fn SetReferrer(&self, aReferrer: *const nsIURI) -> nsresult

/** This is the referrer for the load. */

attribute nsIURI referrer;

pub unsafe fn GetOriginalURI(
&self,
aOriginalURI: *mut *const nsIURI
) -> nsresult

/**
     * The originalURI to be passed to nsIDocShell.internalLoad. May be null.
     */

attribute nsIURI originalURI;

pub unsafe fn SetOriginalURI(&self, aOriginalURI: *const nsIURI) -> nsresult

/**
     * The originalURI to be passed to nsIDocShell.internalLoad. May be null.
     */

attribute nsIURI originalURI;

pub unsafe fn GetResultPrincipalURI(
&self,
aResultPrincipalURI: *mut *const nsIURI
) -> nsresult

/**
     * Result principal URL from nsILoadInfo, may be null.  Valid only if
     * the "IsSome" part is true (has the same meaning as isSome()
     * on mozilla::Maybe.)
     *
     * In C++ please use Get/SetMaybeResultPrincipalURI helper functions.
     */

attribute nsIURI resultPrincipalURI;

pub unsafe fn SetResultPrincipalURI(
&self,
aResultPrincipalURI: *const nsIURI
) -> nsresult

/**
     * Result principal URL from nsILoadInfo, may be null.  Valid only if
     * the "IsSome" part is true (has the same meaning as isSome()
     * on mozilla::Maybe.)
     *
     * In C++ please use Get/SetMaybeResultPrincipalURI helper functions.
     */

attribute nsIURI resultPrincipalURI;

pub unsafe fn GetResultPrincipalURIIsSome(
&self,
aResultPrincipalURIIsSome: *mut bool
) -> nsresult

attribute boolean resultPrincipalURIIsSome;

pub unsafe fn SetResultPrincipalURIIsSome(
&self,
aResultPrincipalURIIsSome: bool
) -> nsresult

attribute boolean resultPrincipalURIIsSome;

pub unsafe fn GetLoadReplace(&self, aLoadReplace: *mut bool) -> nsresult

/**
     * loadReplace flag to be passed to nsIDocShell.internalLoad.
     */

attribute boolean loadReplace;

pub unsafe fn SetLoadReplace(&self, aLoadReplace: bool) -> nsresult

/**
     * loadReplace flag to be passed to nsIDocShell.internalLoad.
     */

attribute boolean loadReplace;

pub unsafe fn GetTriggeringPrincipal(
&self,
aTriggeringPrincipal: *mut *const nsIPrincipal
) -> nsresult

/** The principal of the load, that is, the entity responsible for
     *  causing the load to occur. In most cases the referrer and
     *  the triggeringPrincipal's URI will be identical.
     */

attribute nsIPrincipal triggeringPrincipal;

pub unsafe fn SetTriggeringPrincipal(
&self,
aTriggeringPrincipal: *const nsIPrincipal
) -> nsresult

/** The principal of the load, that is, the entity responsible for
     *  causing the load to occur. In most cases the referrer and
     *  the triggeringPrincipal's URI will be identical.
     */

attribute nsIPrincipal triggeringPrincipal;

pub unsafe fn GetInheritPrincipal(
&self,
aInheritPrincipal: *mut bool
) -> nsresult

/** If this attribute is true and no triggeringPrincipal is specified,
     * copy the principal from the referring document.
     */

attribute boolean inheritPrincipal;

pub unsafe fn SetInheritPrincipal(&self, aInheritPrincipal: bool) -> nsresult

/** If this attribute is true and no triggeringPrincipal is specified,
     * copy the principal from the referring document.
     */

attribute boolean inheritPrincipal;

pub unsafe fn GetPrincipalIsExplicit(
&self,
aPrincipalIsExplicit: *mut bool
) -> nsresult

/** If this attribute is true only ever use the principal specified
     *  by the triggeringPrincipal and inheritPrincipal attributes.
     *  If there are security reasons for why this is unsafe, such
     *  as trying to use a systemprincipal as the triggeringPrincipal
     *  for a content docshell the load fails.
     */

attribute boolean principalIsExplicit;

pub unsafe fn SetPrincipalIsExplicit(
&self,
aPrincipalIsExplicit: bool
) -> nsresult

/** If this attribute is true only ever use the principal specified
     *  by the triggeringPrincipal and inheritPrincipal attributes.
     *  If there are security reasons for why this is unsafe, such
     *  as trying to use a systemprincipal as the triggeringPrincipal
     *  for a content docshell the load fails.
     */

attribute boolean principalIsExplicit;

pub unsafe fn GetForceAllowDataURI(
&self,
aForceAllowDataURI: *mut bool
) -> nsresult

/**
     * If this attribute is true, then a top-level navigation
     * to a data URI will be allowed.
     */

attribute boolean forceAllowDataURI;

pub unsafe fn SetForceAllowDataURI(&self, aForceAllowDataURI: bool) -> nsresult

/**
     * If this attribute is true, then a top-level navigation
     * to a data URI will be allowed.
     */

attribute boolean forceAllowDataURI;

pub unsafe fn GetLoadType(
&self,
aLoadType: *mut nsDocShellInfoLoadType
) -> nsresult

/** Contains a load type as specified by the load* constants */

attribute nsDocShellInfoLoadType loadType;

pub unsafe fn SetLoadType(&self, aLoadType: nsDocShellInfoLoadType) -> nsresult

/** Contains a load type as specified by the load* constants */

attribute nsDocShellInfoLoadType loadType;

pub unsafe fn GetSHEntry(&self, aSHEntry: *mut *const nsISHEntry) -> nsresult

/** SHEntry for this page */

attribute nsISHEntry SHEntry;

pub unsafe fn SetSHEntry(&self, aSHEntry: *const nsISHEntry) -> nsresult

/** SHEntry for this page */

attribute nsISHEntry SHEntry;

pub unsafe fn GetTarget(&self, aTarget: *mut *const int16_t) -> nsresult

/** Target for load, like _content, _blank etc. */

attribute wstring target;

pub unsafe fn SetTarget(&self, aTarget: *const int16_t) -> nsresult

/** Target for load, like _content, _blank etc. */

attribute wstring target;

pub unsafe fn GetPostDataStream(
&self,
aPostDataStream: *mut *const nsIInputStream
) -> nsresult

/** Post data */

attribute nsIInputStream postDataStream;

pub unsafe fn SetPostDataStream(
&self,
aPostDataStream: *const nsIInputStream
) -> nsresult

/** Post data */

attribute nsIInputStream postDataStream;

pub unsafe fn GetHeadersStream(
&self,
aHeadersStream: *mut *const nsIInputStream
) -> nsresult

/** Additional headers */

attribute nsIInputStream headersStream;

pub unsafe fn SetHeadersStream(
&self,
aHeadersStream: *const nsIInputStream
) -> nsresult

/** Additional headers */

attribute nsIInputStream headersStream;

pub unsafe fn GetSendReferrer(&self, aSendReferrer: *mut bool) -> nsresult

/** True if the referrer should be sent, false if it shouldn't be
     *  sent, even if it's available. This attribute defaults to true.
     */

attribute boolean sendReferrer;

pub unsafe fn SetSendReferrer(&self, aSendReferrer: bool) -> nsresult

/** True if the referrer should be sent, false if it shouldn't be
     *  sent, even if it's available. This attribute defaults to true.
     */

attribute boolean sendReferrer;

pub unsafe fn GetReferrerPolicy(
&self,
aReferrerPolicy: *mut nsDocShellInfoReferrerPolicy
) -> nsresult

/** Referrer policy for the load. This attribute holds one of
     *  the values (REFERRER_POLICY_*) defined in nsIHttpChannel.
     */

attribute nsDocShellInfoReferrerPolicy referrerPolicy;

pub unsafe fn SetReferrerPolicy(
&self,
aReferrerPolicy: nsDocShellInfoReferrerPolicy
) -> nsresult

/** Referrer policy for the load. This attribute holds one of
     *  the values (REFERRER_POLICY_*) defined in nsIHttpChannel.
     */

attribute nsDocShellInfoReferrerPolicy referrerPolicy;

pub unsafe fn GetIsSrcdocLoad(&self, aIsSrcdocLoad: *mut bool) -> nsresult

/** True if the docshell has been created to load an iframe where the
     * srcdoc attribute has been set.  Set when srcdocData is specified.
     */

readonly attribute boolean isSrcdocLoad;

pub unsafe fn GetSrcdocData(&self, aSrcdocData: &mut nsAString) -> nsresult

/** When set, the load will be interpreted as a srcdoc load, where contents
     * of this string will be loaded instead of the URI.  Setting srcdocData
     * sets isSrcdocLoad to true
     */

attribute AString srcdocData;

pub unsafe fn SetSrcdocData(&self, aSrcdocData: &nsAString) -> nsresult

/** When set, the load will be interpreted as a srcdoc load, where contents
     * of this string will be loaded instead of the URI.  Setting srcdocData
     * sets isSrcdocLoad to true
     */

attribute AString srcdocData;

pub unsafe fn GetSourceDocShell(
&self,
aSourceDocShell: *mut *const nsIDocShell
) -> nsresult

/** When set, this is the Source Browsing Context for the navigation. */

attribute nsIDocShell sourceDocShell;

pub unsafe fn SetSourceDocShell(
&self,
aSourceDocShell: *const nsIDocShell
) -> nsresult

/** When set, this is the Source Browsing Context for the navigation. */

attribute nsIDocShell sourceDocShell;

pub unsafe fn GetBaseURI(&self, aBaseURI: *mut *const nsIURI) -> nsresult

/**
     * Used for srcdoc loads to give view-source knowledge of the load's base
     * URI as this information isn't embedded in the load's URI.
     */

attribute nsIURI baseURI;

pub unsafe fn SetBaseURI(&self, aBaseURI: *const nsIURI) -> nsresult

/**
     * Used for srcdoc loads to give view-source knowledge of the load's base
     * URI as this information isn't embedded in the load's URI.
     */

attribute nsIURI baseURI;