Struct xpcom::interfaces::nsIDocShellTreeItem

#[repr(C)]

pub struct nsIDocShellTreeItem { /* fields omitted */ }

interface nsIDocShellTreeItem : nsISupports

/**
 * The nsIDocShellTreeItem supplies the methods that are required of any item
 * that wishes to be able to live within the docshell tree either as a middle
 * node or a leaf.
 */

Methods

impl nsIDocShellTreeItem

pub fn coerce<T: nsIDocShellTreeItemCoerce>(&self) -> &T

Cast this nsIDocShellTreeItem to one of its base interfaces.

impl nsIDocShellTreeItem

pub const typeChrome: i64

typeChrome: i64 = 0

pub const typeContent: i64

typeContent: i64 = 1

pub const typeContentWrapper: i64

typeContentWrapper: i64 = 2

pub const typeChromeWrapper: i64

typeChromeWrapper: i64 = 3

pub const typeAll: i64

typeAll: i64 = 2147483647

pub unsafe fn GetName(&self, aName: &mut nsAString) -> nsresult

attribute AString name;

pub unsafe fn SetName(&self, aName: &nsAString) -> nsresult

attribute AString name;

pub unsafe fn NameEquals(
&self,
name: &nsAString,
_retval: *mut bool
) -> nsresult

/**
         * Compares the provided name against the item's name and
         * returns the appropriate result.
         *
         * @return <CODE>PR_TRUE</CODE> if names match;
         *         <CODE>PR_FALSE</CODE> otherwise.
         */

boolean nameEquals (in AString name);

pub unsafe fn GetItemType(&self, aItemType: *mut int32_t) -> nsresult

attribute long itemType;

pub unsafe fn SetItemType(&self, aItemType: int32_t) -> nsresult

attribute long itemType;

pub unsafe fn GetParent(
&self,
aParent: *mut *const nsIDocShellTreeItem
) -> nsresult

readonly attribute nsIDocShellTreeItem parent;

pub unsafe fn GetSameTypeParent(
&self,
aSameTypeParent: *mut *const nsIDocShellTreeItem
) -> nsresult

readonly attribute nsIDocShellTreeItem sameTypeParent;

pub unsafe fn GetRootTreeItem(
&self,
aRootTreeItem: *mut *const nsIDocShellTreeItem
) -> nsresult

readonly attribute nsIDocShellTreeItem rootTreeItem;

pub unsafe fn GetSameTypeRootTreeItem(
&self,
aSameTypeRootTreeItem: *mut *const nsIDocShellTreeItem
) -> nsresult

readonly attribute nsIDocShellTreeItem sameTypeRootTreeItem;

pub unsafe fn FindItemWithName(
&self,
name: &nsAString,
aRequestor: *const nsIDocShellTreeItem,
aOriginalRequestor: *const nsIDocShellTreeItem,
aSkipTabGroup: bool,
_retval: *mut *const nsIDocShellTreeItem
) -> nsresult

nsIDocShellTreeItem findItemWithName (in AString name, in nsIDocShellTreeItem aRequestor, in nsIDocShellTreeItem aOriginalRequestor, in bool aSkipTabGroup);

pub unsafe fn GetTreeOwner(
&self,
aTreeOwner: *mut *const nsIDocShellTreeOwner
) -> nsresult

readonly attribute nsIDocShellTreeOwner treeOwner;

pub unsafe fn SetTreeOwner(
&self,
treeOwner: *const nsIDocShellTreeOwner
) -> nsresult

[noscript] void setTreeOwner (in nsIDocShellTreeOwner treeOwner);

pub unsafe fn GetChildCount(&self, aChildCount: *mut int32_t) -> nsresult

readonly attribute long childCount;

pub unsafe fn AddChild(&self, child: *const nsIDocShellTreeItem) -> nsresult

void addChild (in nsIDocShellTreeItem child);

pub unsafe fn RemoveChild(&self, child: *const nsIDocShellTreeItem) -> nsresult

void removeChild (in nsIDocShellTreeItem child);

pub unsafe fn GetChildAt(
&self,
index: int32_t,
_retval: *mut *const nsIDocShellTreeItem
) -> nsresult

/**
     * Return the child at the index requested.  This is 0-based.
     *
     * @throws NS_ERROR_UNEXPECTED if the index is out of range
     */

nsIDocShellTreeItem getChildAt (in long index);

pub unsafe fn FindChildWithName(
&self,
aName: &nsAString,
aRecurse: bool,
aSameType: bool,
aRequestor: *const nsIDocShellTreeItem,
aOriginalRequestor: *const nsIDocShellTreeItem,
_retval: *mut *const nsIDocShellTreeItem
) -> nsresult

nsIDocShellTreeItem findChildWithName (in AString aName, in boolean aRecurse, in boolean aSameType, in nsIDocShellTreeItem aRequestor, in nsIDocShellTreeItem aOriginalRequestor);