Struct xpcom::interfaces::nsIDocShellTreeOwner

#[repr(C)]

pub struct nsIDocShellTreeOwner { /* fields omitted */ }

interface nsIDocShellTreeOwner : nsISupports

Methods

impl nsIDocShellTreeOwner

pub fn coerce<T: nsIDocShellTreeOwnerCoerce>(&self) -> &T

Cast this nsIDocShellTreeOwner to one of its base interfaces.

impl nsIDocShellTreeOwner

pub unsafe fn ContentShellAdded(
&self,
aContentShell: *const nsIDocShellTreeItem,
aPrimary: bool
) -> nsresult

/**
     * Called when a content shell is added to the docshell tree.  This is
     * _only_ called for "root" content shells (that is, ones whose parent is a
     * chrome shell).
     *
     * @param aContentShell the shell being added.
     * @param aPrimary whether the shell is primary.
     */

void contentShellAdded (in nsIDocShellTreeItem aContentShell, in boolean aPrimary);

pub unsafe fn ContentShellRemoved(
&self,
aContentShell: *const nsIDocShellTreeItem
) -> nsresult

/**
     * Called when a content shell is removed from the docshell tree.  This is
     * _only_ called for "root" content shells (that is, ones whose parent is a
     * chrome shell).  Note that if aContentShell was never added,
     * contentShellRemoved should just do nothing.
     *
     * @param aContentShell the shell being removed.
     */

void contentShellRemoved (in nsIDocShellTreeItem aContentShell);

pub unsafe fn GetPrimaryContentShell(
&self,
aPrimaryContentShell: *mut *const nsIDocShellTreeItem
) -> nsresult

readonly attribute nsIDocShellTreeItem primaryContentShell;

pub unsafe fn TabParentAdded(
&self,
aTab: *const nsITabParent,
aPrimary: bool
) -> nsresult

void tabParentAdded (in nsITabParent aTab, in boolean aPrimary);

pub unsafe fn TabParentRemoved(&self, aTab: *const nsITabParent) -> nsresult

void tabParentRemoved (in nsITabParent aTab);

pub unsafe fn GetPrimaryTabParent(
&self,
aPrimaryTabParent: *mut *const nsITabParent
) -> nsresult

readonly attribute nsITabParent primaryTabParent;

pub unsafe fn SizeShellTo(
&self,
shell: *const nsIDocShellTreeItem,
cx: int32_t,
cy: int32_t
) -> nsresult

void sizeShellTo (in nsIDocShellTreeItem shell, in long cx, in long cy);

pub unsafe fn GetPrimaryContentSize(
&self,
width: *mut int32_t,
height: *mut int32_t
) -> nsresult

void getPrimaryContentSize (out long width, out long height);

pub unsafe fn SetPrimaryContentSize(
&self,
width: int32_t,
height: int32_t
) -> nsresult

void setPrimaryContentSize (in long width, in long height);

pub unsafe fn GetRootShellSize(
&self,
width: *mut int32_t,
height: *mut int32_t
) -> nsresult

void getRootShellSize (out long width, out long height);

pub unsafe fn SetRootShellSize(
&self,
width: int32_t,
height: int32_t
) -> nsresult

void setRootShellSize (in long width, in long height);

pub unsafe fn SetPersistence(
&self,
aPersistPosition: bool,
aPersistSize: bool,
aPersistSizeMode: bool
) -> nsresult

void setPersistence (in boolean aPersistPosition, in boolean aPersistSize, in boolean aPersistSizeMode);

pub unsafe fn GetPersistence(
&self,
aPersistPosition: *mut bool,
aPersistSize: *mut bool,
aPersistSizeMode: *mut bool
) -> nsresult

void getPersistence (out boolean aPersistPosition, out boolean aPersistSize, out boolean aPersistSizeMode);

pub unsafe fn GetTabCount(&self, aTabCount: *mut uint32_t) -> nsresult

readonly attribute unsigned long tabCount;

pub unsafe fn GetHasPrimaryContent(
&self,
aHasPrimaryContent: *mut bool
) -> nsresult

readonly attribute bool hasPrimaryContent;