Struct xpcom::interfaces::nsIDragSession

#[repr(C)]

pub struct nsIDragSession { /* fields omitted */ }

interface nsIDragSession : nsISupports

Methods

impl nsIDragSession

pub fn coerce<T: nsIDragSessionCoerce>(&self) -> &T

Cast this nsIDragSession to one of its base interfaces.

impl nsIDragSession

pub unsafe fn GetCanDrop(&self, aCanDrop: *mut bool) -> nsresult

/**
    * Set the current state of the drag, whether it can be dropped or not.
    * usually the target "frame" sets this so the native system can render the correct feedback
    */

attribute boolean canDrop;

pub unsafe fn SetCanDrop(&self, aCanDrop: bool) -> nsresult

/**
    * Set the current state of the drag, whether it can be dropped or not.
    * usually the target "frame" sets this so the native system can render the correct feedback
    */

attribute boolean canDrop;

pub unsafe fn GetOnlyChromeDrop(&self, aOnlyChromeDrop: *mut bool) -> nsresult

/**
   * Indicates if the drop event should be dispatched only to chrome.
   */

attribute boolean onlyChromeDrop;

pub unsafe fn SetOnlyChromeDrop(&self, aOnlyChromeDrop: bool) -> nsresult

/**
   * Indicates if the drop event should be dispatched only to chrome.
   */

attribute boolean onlyChromeDrop;

pub unsafe fn GetDragAction(&self, aDragAction: *mut uint32_t) -> nsresult

/**
    * Sets the action (copy, move, link, et.c) for the current drag
    */

attribute unsigned long dragAction;

pub unsafe fn SetDragAction(&self, aDragAction: uint32_t) -> nsresult

/**
    * Sets the action (copy, move, link, et.c) for the current drag
    */

attribute unsigned long dragAction;

pub unsafe fn GetNumDropItems(&self, aNumDropItems: *mut uint32_t) -> nsresult

/**
    * Get the number of items that were dropped
    */

readonly attribute unsigned long numDropItems;

pub unsafe fn GetSourceDocument(
&self,
aSourceDocument: *mut *const nsIDOMDocument
) -> nsresult

/**
    * The document where the drag was started, which will be null if the
    * drag originated outside the application. Useful for determining if a drop
    * originated in the same document.
    */

readonly attribute nsIDOMDocument sourceDocument;

pub unsafe fn GetSourceNode(
&self,
aSourceNode: *mut *const nsIDOMNode
) -> nsresult

/**
    * The dom node that was originally dragged to start the session, which will be null if the
    * drag originated outside the application.
    */

readonly attribute nsIDOMNode sourceNode;

pub unsafe fn GetDataTransfer(
&self,
aDataTransfer: *mut *const nsIDOMDataTransfer
) -> nsresult

/**
   * The data transfer object for the current drag.
   */

attribute nsIDOMDataTransfer dataTransfer;

pub unsafe fn SetDataTransfer(
&self,
aDataTransfer: *const nsIDOMDataTransfer
) -> nsresult

/**
   * The data transfer object for the current drag.
   */

attribute nsIDOMDataTransfer dataTransfer;

pub unsafe fn GetData(
&self,
aTransferable: *const nsITransferable,
aItemIndex: uint32_t
) -> nsresult

/**
    * Get data from a Drag&Drop. Can be called while the drag is in process
    * or after the drop has completed.
    *
    * @param  aTransferable the transferable for the data to be put into
    * @param  aItemIndex which of multiple drag items, zero-based
    */

void getData (in nsITransferable aTransferable, in unsigned long aItemIndex);

pub unsafe fn IsDataFlavorSupported(
&self,
aDataFlavor: *const c_char,
_retval: *mut bool
) -> nsresult

/**
    * Check to set if any of the native data on the clipboard matches this data flavor
    */

boolean isDataFlavorSupported (in string aDataFlavor);

pub unsafe fn UserCancelled(&self) -> nsresult

void userCancelled ();

pub unsafe fn DragEventDispatchedToChildProcess(&self) -> nsresult

void dragEventDispatchedToChildProcess ();

pub unsafe fn UpdateDragEffect(&self) -> nsresult

void updateDragEffect ();

pub unsafe fn UpdateDragImage(
&self,
aImage: *const nsIDOMNode,
aImageX: int32_t,
aImageY: int32_t
) -> nsresult

void updateDragImage (in nsIDOMNode aImage, in long aImageX, in long aImageY);