Struct xpcom::interfaces::nsIEditingSession

#[repr(C)]

pub struct nsIEditingSession { /* fields omitted */ }

interface nsIEditingSession : nsISupports

Methods

impl nsIEditingSession

pub fn coerce<T: nsIEditingSessionCoerce>(&self) -> &T

Cast this nsIEditingSession to one of its base interfaces.

impl nsIEditingSession

pub const eEditorOK: i64

eEditorOK: i64 = 0

/**
   *  Error codes when we fail to create an editor
   *  is placed in attribute editorStatus
   */

pub const eEditorCreationInProgress: i64

eEditorCreationInProgress: i64 = 1

pub const eEditorErrorCantEditMimeType: i64

eEditorErrorCantEditMimeType: i64 = 2

pub const eEditorErrorFileNotFound: i64

eEditorErrorFileNotFound: i64 = 3

pub const eEditorErrorCantEditFramesets: i64

eEditorErrorCantEditFramesets: i64 = 8

pub const eEditorErrorUnknown: i64

eEditorErrorUnknown: i64 = 9

pub unsafe fn GetEditorStatus(&self, aEditorStatus: *mut uint32_t) -> nsresult

/**
   *  Status after editor creation and document loading
   *  Value is one of the above error codes
   */

readonly attribute unsigned long editorStatus;

pub unsafe fn MakeWindowEditable(
&self,
window: *const mozIDOMWindowProxy,
aEditorType: *const c_char,
doAfterUriLoad: bool,
aMakeWholeDocumentEditable: bool,
aInteractive: bool
) -> nsresult

/**
   *  Make this window editable
   *  @param aWindow nsIDOMWindow, the window the embedder needs to make editable
   *  @param aEditorType string, "html" "htmlsimple" "text" "textsimple"
   *  @param aMakeWholeDocumentEditable if PR_TRUE make the whole document in
   *                                    aWindow editable, otherwise it's the
   *                                    embedder who should make the document
   *                                    (or part of it) editable.
   *  @param aInteractive if PR_FALSE turn off scripting and plugins
   */

void makeWindowEditable (in mozIDOMWindowProxy window, in string aEditorType, in boolean doAfterUriLoad, in boolean aMakeWholeDocumentEditable, in boolean aInteractive);

pub unsafe fn WindowIsEditable(
&self,
window: *const mozIDOMWindowProxy,
_retval: *mut bool
) -> nsresult

/**
   *  Test whether a specific window has had its editable flag set; it may have an editor
   *  now, or will get one after the uri load.
   *
   *  Use this, passing the content root window, to test if we've set up editing
   *  for this content.
   */

boolean windowIsEditable (in mozIDOMWindowProxy window);

pub unsafe fn GetEditorForWindow(
&self,
window: *const mozIDOMWindowProxy,
_retval: *mut *const nsIEditor
) -> nsresult

/**
   *  Get the editor for this window. May return null
   */

nsIEditor getEditorForWindow (in mozIDOMWindowProxy window);

pub unsafe fn SetupEditorOnWindow(
&self,
window: *const mozIDOMWindowProxy
) -> nsresult

/**
   *  Setup editor and related support objects
   */

void setupEditorOnWindow (in mozIDOMWindowProxy window);

pub unsafe fn TearDownEditorOnWindow(
&self,
window: *const mozIDOMWindowProxy
) -> nsresult

/**
   *   Destroy editor and related support objects
   */

void tearDownEditorOnWindow (in mozIDOMWindowProxy window);

pub unsafe fn SetEditorOnControllers(
&self,
aWindow: *const mozIDOMWindowProxy,
aEditor: *const nsIEditor
) -> nsresult

void setEditorOnControllers (in mozIDOMWindowProxy aWindow, in nsIEditor aEditor);

pub unsafe fn DisableJSAndPlugins(
&self,
aWindow: *const mozIDOMWindowProxy
) -> nsresult

/**
   * Disable scripts and plugins in aWindow.
   */

void disableJSAndPlugins (in mozIDOMWindowProxy aWindow);

pub unsafe fn RestoreJSAndPlugins(
&self,
aWindow: *const mozIDOMWindowProxy
) -> nsresult

/**
   * Restore JS and plugins (enable/disable them) according to the state they
   * were before the last call to disableJSAndPlugins.
   */

void restoreJSAndPlugins (in mozIDOMWindowProxy aWindow);

pub unsafe fn DetachFromWindow(
&self,
aWindow: *const mozIDOMWindowProxy
) -> nsresult

/**
   * Removes all the editor's controllers/listeners etc and makes the window
   * uneditable.
   */

void detachFromWindow (in mozIDOMWindowProxy aWindow);

pub unsafe fn ReattachToWindow(
&self,
aWindow: *const mozIDOMWindowProxy
) -> nsresult

/**
   * Undos detachFromWindow(), reattaches this editing session/editor
   * to the window.
   */

void reattachToWindow (in mozIDOMWindowProxy aWindow);

pub unsafe fn GetJsAndPluginsDisabled(
&self,
aJsAndPluginsDisabled: *mut bool
) -> nsresult

/**
   * Whether this session has disabled JS and plugins.
   */

readonly attribute boolean jsAndPluginsDisabled;