Struct xpcom::interfaces::nsIEditorMailSupport

#[repr(C)]

pub struct nsIEditorMailSupport { /* fields omitted */ }

interface nsIEditorMailSupport : nsISupports

Methods

impl nsIEditorMailSupport

pub fn coerce<T: nsIEditorMailSupportCoerce>(&self) -> &T

Cast this nsIEditorMailSupport to one of its base interfaces.

impl nsIEditorMailSupport

pub unsafe fn PasteAsQuotation(&self, aSelectionType: int32_t) -> nsresult

/** Paste the text in the OS clipboard at the cursor position,
    * as a quotation (whose representation is dependant on the editor type),
    * replacing the selected text (if any).
    * @param aSelectionType Text or html?
    */

void pasteAsQuotation (in long aSelectionType);

pub unsafe fn InsertAsQuotation(
&self,
aQuotedText: &nsAString,
_retval: *mut *const nsIDOMNode
) -> nsresult

/** Insert a string as quoted text
    * (whose representation is dependant on the editor type),
    * replacing the selected text (if any).
    * @param aQuotedText  The actual text to be quoted
    * @return             The node which was inserted
    */

nsIDOMNode insertAsQuotation (in AString aQuotedText);

pub unsafe fn InsertTextWithQuotations(
&self,
aStringToInsert: &nsAString
) -> nsresult

/**
   * Inserts a plaintext string at the current location,
   * with special processing for lines beginning with ">",
   * which will be treated as mail quotes and inserted
   * as plaintext quoted blocks.
   * If the selection is not collapsed, the selection is deleted
   * and the insertion takes place at the resulting collapsed selection.
   *
   * @param aString   the string to be inserted
   */

void insertTextWithQuotations (in DOMString aStringToInsert);

pub unsafe fn PasteAsCitedQuotation(
&self,
aCitation: &nsAString,
aSelectionType: int32_t
) -> nsresult

/** Paste a string as quoted text,
    * whose representation is dependant on the editor type,
    * replacing the selected text (if any)
    * @param aCitation    The "mid" URL of the source message
    * @param aSelectionType Text or html?
    */

void pasteAsCitedQuotation (in AString aCitation, in long aSelectionType);

pub unsafe fn InsertAsCitedQuotation(
&self,
aQuotedText: &nsAString,
aCitation: &nsAString,
aInsertHTML: bool,
_retval: *mut *const nsIDOMNode
) -> nsresult

/** Insert a string as quoted text
    * (whose representation is dependant on the editor type),
    * replacing the selected text (if any),
    * including, if possible, a "cite" attribute.
    * @param aQuotedText  The actual text to be quoted
    * @param aCitation    The "mid" URL of the source message
    * @param aInsertHTML  Insert as html?  (vs plaintext)
    * @return             The node which was inserted
    */

nsIDOMNode insertAsCitedQuotation (in AString aQuotedText, in AString aCitation, in boolean aInsertHTML);

pub unsafe fn Rewrap(&self, aRespectNewlines: bool) -> nsresult

/**
   * Rewrap the selected part of the document, re-quoting if necessary.
   * @param aRespectNewlines  Try to maintain newlines in the original?
   */

void rewrap (in boolean aRespectNewlines);

pub unsafe fn StripCites(&self) -> nsresult

/**
   * Strip any citations in the selected part of the document.
   */

void stripCites ();

pub unsafe fn GetEmbeddedObjects(
&self,
_retval: *mut *const nsIArray
) -> nsresult

/**
   * Get a list of IMG and OBJECT tags in the current document.
   */

nsIArray getEmbeddedObjects ();