Struct xpcom::interfaces::nsIEditorStyleSheets

#[repr(C)]

pub struct nsIEditorStyleSheets { /* fields omitted */ }

interface nsIEditorStyleSheets : nsISupports

Methods

impl nsIEditorStyleSheets

pub fn coerce<T: nsIEditorStyleSheetsCoerce>(&self) -> &T

Cast this nsIEditorStyleSheets to one of its base interfaces.

impl nsIEditorStyleSheets

pub unsafe fn ReplaceStyleSheet(&self, aURL: &nsAString) -> nsresult

/** Load and apply the style sheet, specified by aURL, to the
    * editor's document, replacing the last style sheet added (if any).
    * This is always asynchronous, and may cause network I/O.
    *
    * @param aURL The style sheet to be loaded and applied.
    */

void replaceStyleSheet (in AString aURL);

pub unsafe fn AddStyleSheet(&self, aURL: &nsAString) -> nsresult

/** Add the given style sheet to the editor's document,
    * on top of any that are already there.
    * This is always asynchronous, and may cause network I/O.
    *
    * @param aURL The style sheet to be loaded and applied.
    */

void addStyleSheet (in AString aURL);

pub unsafe fn ReplaceOverrideStyleSheet(&self, aURL: &nsAString) -> nsresult

/** Load and apply the override style sheet, specified by aURL, to the
    * editor's document, replacing the last override style sheet added (if any).
    * This is always synchronous, so aURL should be a local file with only
    * local @imports. This action is not undoable. It is not intended for
    * "user" style sheets, only for editor developers to add sheets to change
    * display behavior for editing (like showing special cursors) that will
    * not be affected by loading "document" style sheets with addStyleSheet or
    * especially replaceStyleSheet.
    *
    * @param aURL The style sheet to be loaded and applied.
    */

void replaceOverrideStyleSheet (in AString aURL);

pub unsafe fn AddOverrideStyleSheet(&self, aURL: &nsAString) -> nsresult

/** Load and apply an override style sheet, specified by aURL, to
    * the editor's document, on top of any that are already there.
    * This is always synchronous, so the same caveats about local files and no
    * non-local @import as replaceOverrideStyleSheet apply here, too.
    *
    * @param aURL The style sheet to be loaded and applied.
    */

void addOverrideStyleSheet (in AString aURL);

pub unsafe fn RemoveStyleSheet(&self, aURL: &nsAString) -> nsresult

/** Remove the given style sheet from the editor's document
    * This is always synchronous
    *
    * @param aURL The style sheet to be removed
    */

void removeStyleSheet (in AString aURL);

pub unsafe fn RemoveOverrideStyleSheet(&self, aURL: &nsAString) -> nsresult

/** Remove the given override style sheet from the editor's document
    * This is always synchronous
    *
    * @param aURL The style sheet to be removed.
    */

void removeOverrideStyleSheet (in AString aURL);

pub unsafe fn EnableStyleSheet(
&self,
aURL: &nsAString,
aEnable: bool
) -> nsresult

/** Enable or disable the given style sheet from the editor's document
    * This is always synchronous
    *
    * @param aURL  The style sheet to be enabled or disabled
    * @param aEnable true to enable, or false to disable the style sheet
    */

void enableStyleSheet (in AString aURL, in boolean aEnable);