Struct xpcom::interfaces::nsIEmbeddingSiteWindow

#[repr(C)]

pub struct nsIEmbeddingSiteWindow { /* fields omitted */ }

interface nsIEmbeddingSiteWindow : nsISupports

/**
 * The nsIEmbeddingSiteWindow is implemented by the embedder to provide
 * Gecko with the means to call up to the host to resize the window,
 * hide or show it and set/get its title.
 */

Methods

impl nsIEmbeddingSiteWindow

pub fn coerce<T: nsIEmbeddingSiteWindowCoerce>(&self) -> &T

Cast this nsIEmbeddingSiteWindow to one of its base interfaces.

impl nsIEmbeddingSiteWindow

pub const DIM_FLAGS_POSITION: i64

DIM_FLAGS_POSITION: i64 = 1

/**
     * Flag indicates that position of the top left corner of the outer area
     * is required/specified.
     *
     * @see setDimensions
     * @see getDimensions
     */

pub const DIM_FLAGS_SIZE_INNER: i64

DIM_FLAGS_SIZE_INNER: i64 = 2

/**
     * Flag indicates that the size of the inner area is required/specified.
     *
     * @note The inner and outer flags are mutually exclusive and it is
     *       invalid to combine them.
     *
     * @see setDimensions
     * @see getDimensions
     * @see DIM_FLAGS_SIZE_OUTER
     */

pub const DIM_FLAGS_SIZE_OUTER: i64

DIM_FLAGS_SIZE_OUTER: i64 = 4

/**
     * Flag indicates that the size of the outer area is required/specified.
     *
     * @see setDimensions
     * @see getDimensions
     * @see DIM_FLAGS_SIZE_INNER
     */

pub const DIM_FLAGS_IGNORE_X: i64

DIM_FLAGS_IGNORE_X: i64 = 8

/**
     * Flag indicates that the x parameter should be ignored.
     *
     * @see setDimensions
     */

pub const DIM_FLAGS_IGNORE_Y: i64

DIM_FLAGS_IGNORE_Y: i64 = 16

/**
     * Flag indicates that the y parameter should be ignored.
     *
     * @see setDimensions
     */

pub const DIM_FLAGS_IGNORE_CX: i64

DIM_FLAGS_IGNORE_CX: i64 = 32

/**
     * Flag indicates that the cx parameter should be ignored.
     *
     * @see setDimensions
     */

pub const DIM_FLAGS_IGNORE_CY: i64

DIM_FLAGS_IGNORE_CY: i64 = 64

/**
     * Flag indicates that the cy parameter should be ignored.
     *
     * @see setDimensions
     */

pub unsafe fn SetDimensions(
&self,
flags: uint32_t,
x: int32_t,
y: int32_t,
cx: int32_t,
cy: int32_t
) -> nsresult

/**
     * Sets the dimensions for the window; the position & size. The
     * flags to indicate what the caller wants to set and whether the size
     * refers to the inner or outer area. The inner area refers to just
     * the embedded area, wheras the outer area can also include any
     * surrounding chrome, window frame, title bar, and so on.
     *
     * @param flags  Combination of position, inner and outer size flags.
     *               The ignore flags are telling the parent to use the
     *               current values for those dimensions and ignore the
     *               corresponding parameters the child sends.
     * @param x      Left hand corner of the outer area.
     * @param y      Top corner of the outer area.
     * @param cx     Width of the inner or outer area.
     * @param cy     Height of the inner or outer area.
     *
     * @return <code>NS_OK</code> if operation was performed correctly;
     *         <code>NS_ERROR_UNEXPECTED</code> if window could not be
     *           destroyed;
     *         <code>NS_ERROR_INVALID_ARG</code> for bad flag combination
     *           or illegal dimensions.
     *
     * @see getDimensions
     * @see DIM_FLAGS_POSITION
     * @see DIM_FLAGS_SIZE_OUTER
     * @see DIM_FLAGS_SIZE_INNER
     */

void setDimensions (in unsigned long flags, in long x, in long y, in long cx, in long cy);

pub unsafe fn GetDimensions(
&self,
flags: uint32_t,
x: *mut int32_t,
y: *mut int32_t,
cx: *mut int32_t,
cy: *mut int32_t
) -> nsresult

/**
     * Gets the dimensions of the window. The caller may pass
     * <CODE>nullptr</CODE> for any value it is uninterested in receiving.
     *
     * @param flags  Combination of position, inner and outer size flag .
     * @param x      Left hand corner of the outer area; or <CODE>nullptr</CODE>.
     * @param y      Top corner of the outer area; or <CODE>nullptr</CODE>.
     * @param cx     Width of the inner or outer area; or <CODE>nullptr</CODE>.
     * @param cy     Height of the inner or outer area; or <CODE>nullptr</CODE>.
     *
     * @see setDimensions
     * @see DIM_FLAGS_POSITION
     * @see DIM_FLAGS_SIZE_OUTER
     * @see DIM_FLAGS_SIZE_INNER
     */

void getDimensions (in unsigned long flags, out long x, out long y, out long cx, out long cy);

pub unsafe fn SetFocus(&self) -> nsresult

/**
     * Give the window focus.
     */

void setFocus ();

pub unsafe fn GetVisibility(&self, aVisibility: *mut bool) -> nsresult

/**
     * Visibility of the window.
     */

attribute boolean visibility;

pub unsafe fn SetVisibility(&self, aVisibility: bool) -> nsresult

/**
     * Visibility of the window.
     */

attribute boolean visibility;

pub unsafe fn GetTitle(&self, aTitle: &mut nsAString) -> nsresult

/**
     * Title of the window.
     */

attribute AString title;

pub unsafe fn SetTitle(&self, aTitle: &nsAString) -> nsresult

/**
     * Title of the window.
     */

attribute AString title;

pub unsafe fn GetSiteWindow(&self, aSiteWindow: *mut *mut c_void) -> nsresult

/**
     * Native window for the site's window. The implementor should copy the
     * native window object into the address supplied by the caller. The
     * type of the native window that the address refers to is  platform
     * and OS specific as follows:
     *
     * <ul>
     *   <li>On Win32 it is an <CODE>HWND</CODE>.</li>
     *   <li>On MacOS this is a <CODE>WindowPtr</CODE>.</li>
     *   <li>On GTK this is a <CODE>GtkWidget*</CODE>.</li>
     * </ul>
     */

[noscript] readonly attribute voidPtr siteWindow;

pub unsafe fn Blur(&self) -> nsresult

/**
     * Blur the window. This should unfocus the window and send an onblur event.
     */

void blur ();