Struct xpcom::interfaces::nsIEventListenerInfo

#[repr(C)]

pub struct nsIEventListenerInfo { /* fields omitted */ }

interface nsIEventListenerInfo : nsISupports

/**
 * An instance of this interface describes how an event listener
 * was added to an event target.
 */

Methods

impl nsIEventListenerInfo

pub fn coerce<T: nsIEventListenerInfoCoerce>(&self) -> &T

Cast this nsIEventListenerInfo to one of its base interfaces.

impl nsIEventListenerInfo

pub unsafe fn GetType(&self, aType: &mut nsAString) -> nsresult

/**
   * The type of the event for which the listener was added.
   * Null if the listener is for all the events.
   */

readonly attribute AString type;

pub unsafe fn GetCapturing(&self, aCapturing: *mut bool) -> nsresult

readonly attribute boolean capturing;

pub unsafe fn GetAllowsUntrusted(&self, aAllowsUntrusted: *mut bool) -> nsresult

readonly attribute boolean allowsUntrusted;

pub unsafe fn GetInSystemEventGroup(
&self,
aInSystemEventGroup: *mut bool
) -> nsresult

readonly attribute boolean inSystemEventGroup;

pub unsafe fn ToSource(&self, _retval: &mut nsAString) -> nsresult

/**
   * Tries to serialize event listener to a string.
   * Returns null if serialization isn't possible
   * (for example with C++ listeners).
   */

AString toSource ();