Struct xpcom::interfaces::nsIEventListenerService

#[repr(C)]

pub struct nsIEventListenerService { /* fields omitted */ }

interface nsIEventListenerService : nsISupports

Methods

impl nsIEventListenerService

pub fn coerce<T: nsIEventListenerServiceCoerce>(&self) -> &T

Cast this nsIEventListenerService to one of its base interfaces.

impl nsIEventListenerService

pub unsafe fn GetListenerInfoFor(
&self,
aEventTarget: *const nsIDOMEventTarget,
aCount: *mut uint32_t,
aOutArray: *mut *mut *const nsIEventListenerInfo
) -> nsresult

/**
   * Returns an array of nsIEventListenerInfo objects.
   * If aEventTarget doesn't have any listeners, this returns null.
   */

void getListenerInfoFor (in nsIDOMEventTarget aEventTarget, [optional] out unsigned long aCount, [array, size_is (aCount), retval] out nsIEventListenerInfo aOutArray);

pub unsafe fn GetEventTargetChainFor(
&self,
aEventTarget: *const nsIDOMEventTarget,
composed: bool,
aCount: *mut uint32_t,
aOutArray: *mut *mut *const nsIDOMEventTarget
) -> nsresult

/**
   * Returns an array of event targets.
   * aEventTarget will be at index 0.
   * The objects are the ones that would be used as DOMEvent.currentTarget while
   * dispatching an event to aEventTarget
   * @note Some events, especially 'load', may actually have a shorter
   *       event target chain than what this methods returns.
   */

void getEventTargetChainFor (in nsIDOMEventTarget aEventTarget, in boolean composed, [optional] out unsigned long aCount, [array, size_is (aCount), retval] out nsIDOMEventTarget aOutArray);

pub unsafe fn HasListenersFor(
&self,
aEventTarget: *const nsIDOMEventTarget,
aType: &nsAString,
_retval: *mut bool
) -> nsresult

/**
   * Returns true if a event target has any listener for the given type.
   */

boolean hasListenersFor (in nsIDOMEventTarget aEventTarget, in DOMString aType);

pub unsafe fn AddSystemEventListener(
&self,
target: *const nsIDOMEventTarget,
type_: &nsAString,
listener: *const nsIDOMEventListener,
useCapture: bool
) -> nsresult

/**
   * Add a system-group eventlistener to a event target.
   */

void addSystemEventListener (in nsIDOMEventTarget target, in DOMString type, in nsIDOMEventListener listener, in boolean useCapture);

pub unsafe fn RemoveSystemEventListener(
&self,
target: *const nsIDOMEventTarget,
type_: &nsAString,
listener: *const nsIDOMEventListener,
useCapture: bool
) -> nsresult

/**
   * Remove a system-group eventlistener from a event target.
   */

void removeSystemEventListener (in nsIDOMEventTarget target, in DOMString type, in nsIDOMEventListener listener, in boolean useCapture);

pub unsafe fn AddListenerForAllEvents(
&self,
target: *const nsIDOMEventTarget,
listener: *const nsIDOMEventListener,
aUseCapture: bool,
aWantsUntrusted: bool,
aSystemEventGroup: bool
) -> nsresult

void addListenerForAllEvents (in nsIDOMEventTarget target, in nsIDOMEventListener listener, [optional] in boolean aUseCapture, [optional] in boolean aWantsUntrusted, [optional] in boolean aSystemEventGroup);

pub unsafe fn RemoveListenerForAllEvents(
&self,
target: *const nsIDOMEventTarget,
listener: *const nsIDOMEventListener,
aUseCapture: bool,
aSystemEventGroup: bool
) -> nsresult

void removeListenerForAllEvents (in nsIDOMEventTarget target, in nsIDOMEventListener listener, [optional] in boolean aUseCapture, [optional] in boolean aSystemEventGroup);

pub unsafe fn AddListenerChangeListener(
&self,
aListener: *const nsIListenerChangeListener
) -> nsresult

void addListenerChangeListener (in nsIListenerChangeListener aListener);

pub unsafe fn RemoveListenerChangeListener(
&self,
aListener: *const nsIListenerChangeListener
) -> nsresult

void removeListenerChangeListener (in nsIListenerChangeListener aListener);