Struct xpcom::interfaces::nsIFakePluginTag

#[repr(C)]

pub struct nsIFakePluginTag { /* fields omitted */ }

interface nsIFakePluginTag : nsIPluginTag

/**
 * An interface representing a "fake" plugin: one implemented in JavaScript, not
 * as a NPAPI plug-in.  See nsIPluginHost.registerFakePlugin and the
 * documentation for the FakePluginTagInit dictionary.
 */

Methods

impl nsIFakePluginTag

pub fn coerce<T: nsIFakePluginTagCoerce>(&self) -> &T

Cast this nsIFakePluginTag to one of its base interfaces.

impl nsIFakePluginTag

pub unsafe fn GetHandlerURI(&self, aHandlerURI: *mut *const nsIURI) -> nsresult

/**
   * The URI that should be loaded into the tag (as a frame) to handle the
   * plugin. Note that the original data/src value for the plugin is not loaded
   * and will need to be requested by the handler via XHR or similar if desired.
   */

readonly attribute nsIURI handlerURI;

pub unsafe fn GetSandboxScript(
&self,
aSandboxScript: &mut nsAString
) -> nsresult

/**
   * Optional script to run in a sandbox when instantiating a plugin. If this
   * value is an empty string then no such script will be run.
   * The script runs in a sandbox with system principal in the process that
   * contains the element that instantiates the plugin (ie the EMBED or OBJECT
   * element). The sandbox global has a 'pluginElement' property that the script
   * can use to access the element that instantiates the plugin.
   */

readonly attribute AString sandboxScript;

pub unsafe fn GetId(&self, aId: *mut uint32_t) -> nsresult

/**
   * A unique id for this JS-implemented plugin. 0 is a valid id.
   */

readonly attribute unsigned long id;