Struct xpcom::interfaces::nsIFeed

#[repr(C)]

pub struct nsIFeed { /* fields omitted */ }

interface nsIFeed : nsIFeedContainer

/**
 * An nsIFeed represents a single Atom or RSS feed.
 */

Methods

impl nsIFeed

pub fn coerce<T: nsIFeedCoerce>(&self) -> &T

Cast this nsIFeed to one of its base interfaces.

impl nsIFeed

pub const TYPE_FEED: i64

TYPE_FEED: i64 = 0

pub const TYPE_AUDIO: i64

TYPE_AUDIO: i64 = 1

pub const TYPE_IMAGE: i64

TYPE_IMAGE: i64 = 2

pub const TYPE_VIDEO: i64

TYPE_VIDEO: i64 = 4

pub unsafe fn GetSubtitle(
&self,
aSubtitle: *mut *const nsIFeedTextConstruct
) -> nsresult

/**
  * Uses description, subtitle, and extensions
  * to generate a summary.
  */

attribute nsIFeedTextConstruct subtitle;

pub unsafe fn SetSubtitle(
&self,
aSubtitle: *const nsIFeedTextConstruct
) -> nsresult

/**
  * Uses description, subtitle, and extensions
  * to generate a summary.
  */

attribute nsIFeedTextConstruct subtitle;

pub unsafe fn GetType(&self, aType: *mut uint32_t) -> nsresult

/**
  * The type of feed. For example, a podcast would be TYPE_AUDIO.
  */

readonly attribute unsigned long type;

pub unsafe fn GetEnclosureCount(
&self,
aEnclosureCount: *mut int32_t
) -> nsresult

/**
  * The total number of enclosures found in the feed.
  */

attribute long enclosureCount;

pub unsafe fn SetEnclosureCount(&self, aEnclosureCount: int32_t) -> nsresult

/**
  * The total number of enclosures found in the feed.
  */

attribute long enclosureCount;

pub unsafe fn GetItems(&self, aItems: *mut *const nsIArray) -> nsresult

/**
  * The items or entries in feed.
  */

attribute nsIArray items;

pub unsafe fn SetItems(&self, aItems: *const nsIArray) -> nsresult

/**
  * The items or entries in feed.
  */

attribute nsIArray items;

pub unsafe fn GetCloud(
&self,
aCloud: *mut *const nsIWritablePropertyBag2
) -> nsresult

/**
  * No one really knows what cloud is for.
  *
  * It supposedly enables some sort of interaction with an XML-RPC or
  * SOAP service.
  */

attribute nsIWritablePropertyBag2 cloud;

pub unsafe fn SetCloud(
&self,
aCloud: *const nsIWritablePropertyBag2
) -> nsresult

/**
  * No one really knows what cloud is for.
  *
  * It supposedly enables some sort of interaction with an XML-RPC or
  * SOAP service.
  */

attribute nsIWritablePropertyBag2 cloud;

pub unsafe fn GetGenerator(
&self,
aGenerator: *mut *const nsIFeedGenerator
) -> nsresult

/**
  * Information about the software that produced the feed.
  */

attribute nsIFeedGenerator generator;

pub unsafe fn SetGenerator(
&self,
aGenerator: *const nsIFeedGenerator
) -> nsresult

/**
  * Information about the software that produced the feed.
  */

attribute nsIFeedGenerator generator;

pub unsafe fn GetImage(
&self,
aImage: *mut *const nsIWritablePropertyBag2
) -> nsresult

/**
  * An image url and some metadata (as defined by RSS2).
  *
  */

attribute nsIWritablePropertyBag2 image;

pub unsafe fn SetImage(
&self,
aImage: *const nsIWritablePropertyBag2
) -> nsresult

/**
  * An image url and some metadata (as defined by RSS2).
  *
  */

attribute nsIWritablePropertyBag2 image;

pub unsafe fn GetTextInput(
&self,
aTextInput: *mut *const nsIWritablePropertyBag2
) -> nsresult

/**
  * No one really knows what textInput is for.
  *
  * See
  * <http://www.cadenhead.org/workbench/news/2894/rss-joy-textinput>
  * for more details.
  */

attribute nsIWritablePropertyBag2 textInput;

pub unsafe fn SetTextInput(
&self,
aTextInput: *const nsIWritablePropertyBag2
) -> nsresult

/**
  * No one really knows what textInput is for.
  *
  * See
  * <http://www.cadenhead.org/workbench/news/2894/rss-joy-textinput>
  * for more details.
  */

attribute nsIWritablePropertyBag2 textInput;

pub unsafe fn GetSkipDays(&self, aSkipDays: *mut *const nsIArray) -> nsresult

/**
  * Days to skip fetching. This field was supposed to designate
  * intervals for feed fetching. It's not generally implemented. For
  * example, if this array contained "Monday", aggregators should not
  * fetch the feed on Mondays.
  */

attribute nsIArray skipDays;

pub unsafe fn SetSkipDays(&self, aSkipDays: *const nsIArray) -> nsresult

/**
  * Days to skip fetching. This field was supposed to designate
  * intervals for feed fetching. It's not generally implemented. For
  * example, if this array contained "Monday", aggregators should not
  * fetch the feed on Mondays.
  */

attribute nsIArray skipDays;

pub unsafe fn GetSkipHours(&self, aSkipHours: *mut *const nsIArray) -> nsresult

/**
  * Hours to skip fetching. This field was supposed to designate
  * intervals for feed fetching. It's not generally implemented. See
  * <http://blogs.law.harvard.edu/tech/rss> for more information.
  */

attribute nsIArray skipHours;

pub unsafe fn SetSkipHours(&self, aSkipHours: *const nsIArray) -> nsresult

/**
  * Hours to skip fetching. This field was supposed to designate
  * intervals for feed fetching. It's not generally implemented. See
  * <http://blogs.law.harvard.edu/tech/rss> for more information.
  */

attribute nsIArray skipHours;