Struct xpcom::interfaces::nsIFeedResultService

#[repr(C)]

pub struct nsIFeedResultService { /* fields omitted */ }

interface nsIFeedResultService : nsISupports

/**
 * nsIFeedResultService provides a globally-accessible object for retrieving
 * the results of feed processing.
 */

Methods

impl nsIFeedResultService

pub fn coerce<T: nsIFeedResultServiceCoerce>(&self) -> &T

Cast this nsIFeedResultService to one of its base interfaces.

impl nsIFeedResultService

pub unsafe fn GetForcePreviewPage(
&self,
aForcePreviewPage: *mut bool
) -> nsresult

/**
   * When set to true, forces the preview page to be displayed, regardless
   * of the user's preferences.
   */

attribute boolean forcePreviewPage;

pub unsafe fn SetForcePreviewPage(&self, aForcePreviewPage: bool) -> nsresult

/**
   * When set to true, forces the preview page to be displayed, regardless
   * of the user's preferences.
   */

attribute boolean forcePreviewPage;

pub unsafe fn AddToClientReader(
&self,
uri: &nsACString,
title: &nsAString,
subtitle: &nsAString,
feedType: uint32_t,
feedReader: &nsAString
) -> nsresult

/**
   * Adds a URI to the user's specified external feed handler, or live
   * bookmarks.
   * @param   uri
   *          The uri of the feed to add.
   * @param   title
   *          The title of the feed to add.
   * @param   subtitle
   *          The subtitle of the feed to add.
   * @param   feedType
   *          The nsIFeed type of the feed.  See nsIFeed.idl
   * @param   feedReader
   *          The type of feed reader we're using (client, bookmarks, default)
   *          If this parameter is null, the type is set to default
   */

void addToClientReader (in AUTF8String uri, in AString title, in AString subtitle, in unsigned long feedType, [optional] in AString feedReader);

pub unsafe fn AddFeedResult(&self, feedResult: *const nsIFeedResult) -> nsresult

/**
   * Registers a Feed Result object with a globally accessible service
   * so that it can be accessed by a singleton method outside the usual
   * flow of control in document loading.
   *
   * @param   feedResult
   *          An object implementing nsIFeedResult representing the feed.
   */

void addFeedResult (in nsIFeedResult feedResult);

pub unsafe fn GetFeedResult(
&self,
uri: *const nsIURI,
_retval: *mut *const nsIFeedResult
) -> nsresult

/**
   * Gets a Feed Handler object registered using addFeedResult.
   *
   * @param   uri
   *          The URI of the feed a handler is being requested for
   */

nsIFeedResult getFeedResult (in nsIURI uri);

pub unsafe fn RemoveFeedResult(&self, uri: *const nsIURI) -> nsresult

/**
   * Unregisters a Feed Handler object registered using addFeedResult.
   * @param   uri
   *          The feed URI the handler was registered under. This must be
   *          the same *instance* the feed was registered under.
   */

void removeFeedResult (in nsIURI uri);