Struct xpcom::interfaces::nsIFileOutputStream

#[repr(C)]

pub struct nsIFileOutputStream { /* fields omitted */ }

interface nsIFileOutputStream : nsIOutputStream

/**
 * An output stream that lets you stream to a file.
 */

Methods

impl nsIFileOutputStream

pub fn coerce<T: nsIFileOutputStreamCoerce>(&self) -> &T

Cast this nsIFileOutputStream to one of its base interfaces.

impl nsIFileOutputStream

pub const DEFER_OPEN: i64

DEFER_OPEN: i64 = 1

/**
     * See the same constant in nsIFileInputStream. The deferred open will
     * be performed when one of the following is called:
     *   - Seek
     *   - Tell
     *   - SetEOF
     *   - Write
     *   - Flush
     *
     * @note Using this flag results in the file not being opened
     *       during the call to Init.  This means that any errors that might
     *       happen when this flag is not set would happen during the
     *       first write, and if the file is to be created, then it will not
     *       appear on the disk until the first write.
     */

pub unsafe fn Init(
&self,
file: *const nsIFile,
ioFlags: int32_t,
perm: int32_t,
behaviorFlags: int32_t
) -> nsresult

/**
     * @param file          file to write to
     * @param ioFlags       file open flags listed in prio.h (see
     *                      PR_Open documentation) or -1 to open the
     *                      file in default mode (PR_WRONLY |
     *                      PR_CREATE_FILE | PR_TRUNCATE)
     * @param perm          file mode bits listed in prio.h or -1 to
     *                      use the default permissions (0664)
     * @param behaviorFlags flags specifying various behaviors of the class
     *        (currently none supported)
     */

void init (in nsIFile file, in long ioFlags, in long perm, in long behaviorFlags);

pub unsafe fn Preallocate(&self, length: int64_t) -> nsresult

/**
     * @param length        asks the operating system to allocate storage for
     *                      this file of at least |length| bytes long, and
     *                      set the file length to the corresponding size.
     * @throws NS_ERROR_FAILURE if the preallocation fails.
     * @throws NS_ERROR_NOT_INITIALIZED if the file is not opened.
     */

[noscript] void preallocate (in long long length);