Struct xpcom::interfaces::nsIFilePicker

#[repr(C)]

pub struct nsIFilePicker { /* fields omitted */ }

interface nsIFilePicker : nsISupports

Methods

impl nsIFilePicker

pub fn coerce<T: nsIFilePickerCoerce>(&self) -> &T

Cast this nsIFilePicker to one of its base interfaces.

impl nsIFilePicker

pub const modeOpen: i64

modeOpen: i64 = 0

pub const modeSave: i64

modeSave: i64 = 1

pub const modeGetFolder: i64

modeGetFolder: i64 = 2

pub const modeOpenMultiple: i64

modeOpenMultiple: i64 = 3

pub const returnOK: i64

returnOK: i64 = 0

pub const returnCancel: i64

returnCancel: i64 = 1

pub const returnReplace: i64

returnReplace: i64 = 2

pub const filterAll: i64

filterAll: i64 = 1

pub const filterHTML: i64

filterHTML: i64 = 2

pub const filterText: i64

filterText: i64 = 4

pub const filterImages: i64

filterImages: i64 = 8

pub const filterXML: i64

filterXML: i64 = 16

pub const filterXUL: i64

filterXUL: i64 = 32

pub const filterApps: i64

filterApps: i64 = 64

pub const filterAllowURLs: i64

filterAllowURLs: i64 = 128

pub const filterAudio: i64

filterAudio: i64 = 256

pub const filterVideo: i64

filterVideo: i64 = 512

pub unsafe fn Init(
&self,
parent: *const mozIDOMWindowProxy,
title: &nsAString,
mode: int16_t
) -> nsresult

/**
  * Initialize the file picker widget.  The file picker is not valid until this
  * method is called.
  *
  * @param      parent   mozIDOMWindow parent.  This dialog will be dependent
  *                      on this parent. parent must be non-null.
  * @param      title    The title for the file widget
  * @param      mode     load, save, or get folder
  *
  */

void init (in mozIDOMWindowProxy parent, in AString title, in short mode);

pub unsafe fn AppendFilters(&self, filterMask: int32_t) -> nsresult

/**
  * Append to the  filter list with things from the predefined list
  *
  * @param      filters  mask of filters i.e. (filterAll | filterHTML)
  *
  */

void appendFilters (in long filterMask);

pub unsafe fn AppendFilter(
&self,
title: &nsAString,
filter: &nsAString
) -> nsresult

/**
  * Add a filter
  *
  * @param      title    name of the filter
  * @param      filter   extensions to filter -- semicolon and space separated
  *
  */

void appendFilter (in AString title, in AString filter);

pub unsafe fn GetDefaultString(
&self,
aDefaultString: &mut nsAString
) -> nsresult

/**
  * The filename that should be suggested to the user as a default. This should
  * include the extension.
  *
  * @throws NS_ERROR_FAILURE on attempts to get
  */

attribute AString defaultString;

pub unsafe fn SetDefaultString(&self, aDefaultString: &nsAString) -> nsresult

/**
  * The filename that should be suggested to the user as a default. This should
  * include the extension.
  *
  * @throws NS_ERROR_FAILURE on attempts to get
  */

attribute AString defaultString;

pub unsafe fn GetDefaultExtension(
&self,
aDefaultExtension: &mut nsAString
) -> nsresult

/**
  * The extension that should be associated with files of the type we
  * want to work with.  On some platforms, this extension will be
  * automatically appended to filenames the user enters, if needed.
  */

attribute AString defaultExtension;

pub unsafe fn SetDefaultExtension(
&self,
aDefaultExtension: &nsAString
) -> nsresult

/**
  * The extension that should be associated with files of the type we
  * want to work with.  On some platforms, this extension will be
  * automatically appended to filenames the user enters, if needed.
  */

attribute AString defaultExtension;

pub unsafe fn GetFilterIndex(&self, aFilterIndex: *mut int32_t) -> nsresult

/**
  * The filter which is currently selected in the File Picker dialog
  *
  * @return Returns the index (0 based) of the selected filter in the filter list.
  */

attribute long filterIndex;

pub unsafe fn SetFilterIndex(&self, aFilterIndex: int32_t) -> nsresult

/**
  * The filter which is currently selected in the File Picker dialog
  *
  * @return Returns the index (0 based) of the selected filter in the filter list.
  */

attribute long filterIndex;

pub unsafe fn GetDisplayDirectory(
&self,
aDisplayDirectory: *mut *const nsIFile
) -> nsresult

/**
  * Set the directory that the file open/save dialog initially displays
  * Note that, if displaySpecialDirectory has been already set, this value will
  * be ignored.
  *
  * @param      displayDirectory  the name of the directory
  *
  */

attribute nsIFile displayDirectory;

pub unsafe fn SetDisplayDirectory(
&self,
aDisplayDirectory: *const nsIFile
) -> nsresult

/**
  * Set the directory that the file open/save dialog initially displays
  * Note that, if displaySpecialDirectory has been already set, this value will
  * be ignored.
  *
  * @param      displayDirectory  the name of the directory
  *
  */

attribute nsIFile displayDirectory;

pub unsafe fn GetDisplaySpecialDirectory(
&self,
aDisplaySpecialDirectory: &mut nsAString
) -> nsresult

/**
  * Set the directory that the file open/save dialog initially displays using
  * one of the special name as such as 'Desk', 'TmpD', and so on.
  * Note that, if displayDirectory has been already set, this value will be
  * ignored.
  *
  * @param      displaySpecialDirectory  the name of the special directory
  *
  */

attribute AString displaySpecialDirectory;

pub unsafe fn SetDisplaySpecialDirectory(
&self,
aDisplaySpecialDirectory: &nsAString
) -> nsresult

/**
  * Set the directory that the file open/save dialog initially displays using
  * one of the special name as such as 'Desk', 'TmpD', and so on.
  * Note that, if displayDirectory has been already set, this value will be
  * ignored.
  *
  * @param      displaySpecialDirectory  the name of the special directory
  *
  */

attribute AString displaySpecialDirectory;

pub unsafe fn GetFile(&self, aFile: *mut *const nsIFile) -> nsresult

/**
  * Get the nsIFile for the file or directory.
  *
  * @return Returns the file currently selected
  */

readonly attribute nsIFile file;

pub unsafe fn GetFileURL(&self, aFileURL: *mut *const nsIURI) -> nsresult

/**
  * Get the nsIURI for the file or directory.
  *
  * @return Returns the file currently selected
  */

readonly attribute nsIURI fileURL;

pub unsafe fn GetFiles(
&self,
aFiles: *mut *const nsISimpleEnumerator
) -> nsresult

/**
  * Get the enumerator for the selected files
  * only works in the modeOpenMultiple mode
  *
  * @return Returns the files currently selected
  */

readonly attribute nsISimpleEnumerator files;

pub unsafe fn GetDomFileOrDirectory(
&self,
aDomFileOrDirectory: *mut *const nsISupports
) -> nsresult

/**
  * Get the DOM File or the DOM Directory
  *
  * @return Returns the file or directory currently selected DOM object.
  */

readonly attribute nsISupports domFileOrDirectory;

pub unsafe fn GetDomFileOrDirectoryEnumerator(
&self,
aDomFileOrDirectoryEnumerator: *mut *const nsISimpleEnumerator
) -> nsresult

/**
  * Get the enumerator for the selected files or directories
  * only works in the modeOpenMultiple mode
  *
  * @return Returns the files/directories currently selected as DOM object.
  */

readonly attribute nsISimpleEnumerator domFileOrDirectoryEnumerator;

pub unsafe fn GetAddToRecentDocs(&self, aAddToRecentDocs: *mut bool) -> nsresult

/**
  * Controls whether the chosen file(s) should be added to the system's recent
  * documents list. This attribute will be ignored if the system has no "Recent
  * Docs" concept, or if the application is in private browsing mode (in which
  * case the file will not be added). Defaults to true.
  */

attribute boolean addToRecentDocs;

pub unsafe fn SetAddToRecentDocs(&self, aAddToRecentDocs: bool) -> nsresult

/**
  * Controls whether the chosen file(s) should be added to the system's recent
  * documents list. This attribute will be ignored if the system has no "Recent
  * Docs" concept, or if the application is in private browsing mode (in which
  * case the file will not be added). Defaults to true.
  */

attribute boolean addToRecentDocs;

pub unsafe fn Open(
&self,
aFilePickerShownCallback: *const nsIFilePickerShownCallback
) -> nsresult

/**
  * Opens the file dialog asynchrounously.
  * The passed in object's done method will be called upon completion.
  */

void open (in nsIFilePickerShownCallback aFilePickerShownCallback);

pub unsafe fn GetMode(&self, aMode: *mut int16_t) -> nsresult

/**
  * The picker's mode, as set by the 'mode' argument passed to init()
  * (one of the modeOpen et. al. constants specified above).
  */

readonly attribute short mode;

pub unsafe fn GetOkButtonLabel(
&self,
aOkButtonLabel: &mut nsAString
) -> nsresult

/**
   * If set to non-empty string, the nsIFilePicker implementation
   * may use okButtonLabel as the label for the button the user uses to accept
   * file selection.
   */

attribute AString okButtonLabel;

pub unsafe fn SetOkButtonLabel(&self, aOkButtonLabel: &nsAString) -> nsresult

/**
   * If set to non-empty string, the nsIFilePicker implementation
   * may use okButtonLabel as the label for the button the user uses to accept
   * file selection.
   */

attribute AString okButtonLabel;