Struct xpcom::interfaces::nsIFileURL

#[repr(C)]

pub struct nsIFileURL { /* fields omitted */ }

interface nsIFileURL : nsIURL

/**
 * nsIFileURL provides access to the underlying nsIFile object corresponding to
 * an URL.  The URL scheme need not be file:, since other local protocols may
 * map URLs to files (e.g., resource:).
 */

Methods

impl nsIFileURL

pub fn coerce<T: nsIFileURLCoerce>(&self) -> &T

Cast this nsIFileURL to one of its base interfaces.

impl nsIFileURL

pub unsafe fn GetFile(&self, aFile: *mut *const nsIFile) -> nsresult

/**
     * Get/Set nsIFile corresponding to this URL.
     *
     *  - Getter returns a reference to an immutable object.  Callers must clone
     *    before attempting to modify the returned nsIFile object.  NOTE: this
     *    constraint might not be enforced at runtime, so beware!!
     *
     *  - Setter clones the nsIFile object (allowing the caller to safely modify
     *    the nsIFile object after setting it on this interface).
     */

attribute nsIFile file;

pub unsafe fn SetFile(&self, aFile: *const nsIFile) -> nsresult

/**
     * Get/Set nsIFile corresponding to this URL.
     *
     *  - Getter returns a reference to an immutable object.  Callers must clone
     *    before attempting to modify the returned nsIFile object.  NOTE: this
     *    constraint might not be enforced at runtime, so beware!!
     *
     *  - Setter clones the nsIFile object (allowing the caller to safely modify
     *    the nsIFile object after setting it on this interface).
     */

attribute nsIFile file;