Struct xpcom::interfaces::nsIGNOMEShellService

#[repr(C)]

pub struct nsIGNOMEShellService { /* fields omitted */ }

interface nsIGNOMEShellService : nsIShellService

Methods

impl nsIGNOMEShellService

pub fn coerce<T: nsIGNOMEShellServiceCoerce>(&self) -> &T

Cast this nsIGNOMEShellService to one of its base interfaces.

impl nsIGNOMEShellService

pub unsafe fn GetCanSetDesktopBackground(
&self,
aCanSetDesktopBackground: *mut bool
) -> nsresult

/**
   * Used to determine whether or not to offer "Set as desktop background"
   * functionality. Even if shell service is available it is not
   * guaranteed that it is able to set the background for every desktop
   * which is especially true for Linux with its many different desktop
   * environments.
   */

readonly attribute boolean canSetDesktopBackground;