Struct xpcom::interfaces::nsIGZFileWriter

#[repr(C)]

pub struct nsIGZFileWriter { /* fields omitted */ }

interface nsIGZFileWriter : nsISupports

/**
 * A simple interface for writing to a .gz file.
 *
 * Note that the file that this interface produces has a different format than
 * what you'd get if you compressed your data as a gzip stream and dumped the
 * result to a file.
 *
 * The standard gunzip tool cannot decompress a raw gzip stream, but can handle
 * the files produced by this interface.
 */

Methods

impl nsIGZFileWriter

pub fn coerce<T: nsIGZFileWriterCoerce>(&self) -> &T

Cast this nsIGZFileWriter to one of its base interfaces.

impl nsIGZFileWriter

pub unsafe fn Init(&self, file: *const nsIFile) -> nsresult

/**
   * Initialize this object.  We'll write our gzip'ed data to the given file,
   * overwriting its contents if the file exists.
   *
   * init() will return an error if called twice.  It's an error to call any
   * other method on this interface without first calling init().
   */

[must_use] void init (in nsIFile file);

pub unsafe fn Write(&self, str: &nsACString) -> nsresult

/**
   * Write the given string to the file.
   */

[must_use] void write (in AUTF8String str);

pub unsafe fn Finish(&self) -> nsresult

/**
   * Close this nsIGZFileWriter.  Classes implementing nsIGZFileWriter will run
   * this method when the underlying object is destroyed, so it's not strictly
   * necessary to explicitly call it from your code.
   *
   * It's an error to call this method twice, and it's an error to call write()
   * after finish() has been called.
   */

void finish ();