Struct xpcom::interfaces::nsIGroupedSHistory

#[repr(C)]

pub struct nsIGroupedSHistory { /* fields omitted */ }

interface nsIGroupedSHistory : nsISupports

/**
 * nsIGroupedSHistory represent a combined session history across multiple
 * root docshells (usually browser tabs). The participating nsISHistory can
 * either be in chrome process or in content process, but nsIGroupedSHistory
 * itself lives in chrome process. The communication is proxyed through
 * nsIPartialSHistory.
 */

Methods

impl nsIGroupedSHistory

pub fn coerce<T: nsIGroupedSHistoryCoerce>(&self) -> &T

Cast this nsIGroupedSHistory to one of its base interfaces.

impl nsIGroupedSHistory

pub unsafe fn GetCount(&self) -> uint32_t

[infallible] readonly attribute unsigned long count;

pub unsafe fn GetActiveFrameLoader(
&self,
aActiveFrameLoader: *mut *const nsIFrameLoader
) -> nsresult

/**
   * The currently active frameloader controlled by this nsIGroupedSHistory.
   */

readonly attribute nsIFrameLoader activeFrameLoader;

pub unsafe fn AppendPartialSHistory(
&self,
aPartialHistory: *const nsIPartialSHistory
) -> nsresult

/**
   * Remove all partial histories after currently active one (if any) and then
   * append the given partial session history to the end of the list.
   */

void appendPartialSHistory (in nsIPartialSHistory aPartialHistory);

pub unsafe fn HandleSHistoryUpdate(
&self,
aPartialHistory: *const nsIPartialSHistory,
aTruncate: bool
) -> nsresult

/**
   * Notify the grouped session history that the active partial session history
   * has been modified.
   *
   * @param aPartialHistory The partial history which was updated
   * @param aTruncate If this parameter is true, all partial session histories
   *                  after this one will be removed.
   */

void handleSHistoryUpdate (in nsIPartialSHistory aPartialHistory, in boolean aTruncate);

pub unsafe fn GotoIndex(
&self,
aGlobalIndex: uint32_t,
_retval: *mut *const nsIFrameLoader
) -> nsresult

/**
   * Find the proper partial session history and navigate to the entry
   * corresponding to the given global index. Note it doesn't swap frameloaders,
   * but rather return the target loader for the caller to swap.
   *
   * This function may throw NS_ERROR_NOT_AVAILABLE if the frameloader to swap
   * to is dead.
   *
   * @param  aGlobalIndex
   *         The global index to navigate to.
   * @return The frameloader which needs to be swapped in, or null if no
   *         frameloader needs to be swapped.
   */

nsIFrameLoader gotoIndex (in unsigned long aGlobalIndex);

pub unsafe fn CloseInactiveFrameLoaderOwners(&self) -> nsresult

/**
   * Close the FrameLoaderOwners of the inactive PartialSHistories in this GlobalSHistory.
   * This does not remove the PartialSHistories from the GroupedSHistory.
   */

void closeInactiveFrameLoaderOwners ();

pub unsafe fn AddPrerenderingPartialSHistory(
&self,
aPrerendering: *const nsIPartialSHistory,
aId: int32_t
) -> nsresult

/**
   * Add a partialSHistory as a "prerendering" partialSHistory. This
   * partialSHistory's tab will have its lifetime managed by the
   * GroupedSHistory, and will be closed when closeInactiveFrameLoaderOwners is
   * called, or whenever a SHistory update is received.
   */

void addPrerenderingPartialSHistory (in nsIPartialSHistory aPrerendering, in long aId);

pub unsafe fn CancelPrerendering(&self, aId: int32_t) -> nsresult

/**
   * Cancel the prerendering with the given ID.
   */

void cancelPrerendering (in long aId);