Struct xpcom::interfaces::nsIHTMLAbsPosEditor

#[repr(C)]

pub struct nsIHTMLAbsPosEditor { /* fields omitted */ }

interface nsIHTMLAbsPosEditor : nsISupports

Methods

impl nsIHTMLAbsPosEditor

pub fn coerce<T: nsIHTMLAbsPosEditorCoerce>(&self) -> &T

Cast this nsIHTMLAbsPosEditor to one of its base interfaces.

impl nsIHTMLAbsPosEditor

pub unsafe fn GetSelectionContainerAbsolutelyPositioned(
&self,
aSelectionContainerAbsolutelyPositioned: *mut bool
) -> nsresult

/**
   * true if the selection container is absolutely positioned
   */

readonly attribute boolean selectionContainerAbsolutelyPositioned;

pub unsafe fn GetPositionedElement(
&self,
aPositionedElement: *mut *const nsIDOMElement
) -> nsresult

/**
   * this contains the absolutely positioned element currently edited
   * or null
   */

readonly attribute nsIDOMElement positionedElement;

pub unsafe fn GetAbsolutePositioningEnabled(
&self,
aAbsolutePositioningEnabled: *mut bool
) -> nsresult

/**
   * true if Absolute Positioning handling is enabled in the editor
   */

attribute boolean absolutePositioningEnabled;

pub unsafe fn SetAbsolutePositioningEnabled(
&self,
aAbsolutePositioningEnabled: bool
) -> nsresult

/**
   * true if Absolute Positioning handling is enabled in the editor
   */

attribute boolean absolutePositioningEnabled;

pub unsafe fn GetSnapToGridEnabled(
&self,
aSnapToGridEnabled: *mut bool
) -> nsresult

/**
   * true if Snap To Grid is enabled in the editor.
   */

attribute boolean snapToGridEnabled;

pub unsafe fn SetSnapToGridEnabled(&self, aSnapToGridEnabled: bool) -> nsresult

/**
   * true if Snap To Grid is enabled in the editor.
   */

attribute boolean snapToGridEnabled;

pub unsafe fn GetGridSize(&self, aGridSize: *mut uint32_t) -> nsresult

/**
   * sets the grid size in pixels.
   * @param aSizeInPixels [IN] the size of the grid in pixels
   */

attribute unsigned long gridSize;

pub unsafe fn SetGridSize(&self, aGridSize: uint32_t) -> nsresult

/**
   * sets the grid size in pixels.
   * @param aSizeInPixels [IN] the size of the grid in pixels
   */

attribute unsigned long gridSize;

pub unsafe fn GetAbsolutelyPositionedSelectionContainer(
&self,
aAbsolutelyPositionedSelectionContainer: *mut *const nsIDOMElement
) -> nsresult

/**
   * returns the deepest absolutely positioned container of the selection
   * if it exists or null.
   */

readonly attribute nsIDOMElement absolutelyPositionedSelectionContainer;

pub unsafe fn AbsolutePositionSelection(&self, aEnabled: bool) -> nsresult

/**
   * extracts the selection from the normal flow of the document and
   * positions it.
   * @param aEnabled [IN] true to absolutely position the selection,
   *                      false to put it back in the normal flow
   */

void absolutePositionSelection (in boolean aEnabled);

pub unsafe fn RelativeChangeZIndex(&self, aChange: int32_t) -> nsresult

/**
   * adds aChange to the z-index of the currently positioned element.
   * @param aChange [IN] relative change to apply to current z-index
   */

void relativeChangeZIndex (in long aChange);

pub unsafe fn AbsolutelyPositionElement(
&self,
aElement: *const nsIDOMElement,
aEnabled: bool
) -> nsresult

/**
   * extracts an element from the normal flow of the document and
   * positions it, and puts it back in the normal flow.
   * @param aElement [IN] the element
   * @param aEnabled [IN] true to absolutely position the element,
   *                      false to put it back in the normal flow
   */

void absolutelyPositionElement (in nsIDOMElement aElement, in boolean aEnabled);

pub unsafe fn SetElementPosition(
&self,
aElement: *const nsIDOMElement,
aX: int32_t,
aY: int32_t
) -> nsresult

/**
   * sets the position of an element; warning it does NOT check if the
   * element is already positioned or not and that's on purpose.
   * @param aElement [IN] the element
   * @param aX       [IN] the x position in pixels.
   * @param aY       [IN] the y position in pixels.
   */

void setElementPosition (in nsIDOMElement aElement, in long aX, in long aY);

pub unsafe fn GetElementZIndex(
&self,
aElement: *const nsIDOMElement,
_retval: *mut int32_t
) -> nsresult

/**
   * returns the absolute z-index of a positioned element. Never returns 'auto'.
   * @return         the z-index of the element
   * @param aElement [IN] the element.
   */

long getElementZIndex (in nsIDOMElement aElement);

pub unsafe fn SetElementZIndex(
&self,
aElement: *const nsIDOMElement,
aZorder: int32_t
) -> nsresult

/**
   * sets the z-index of an element.
   * @param aElement [IN] the element
   * @param aZorder  [IN] the z-index
   */

void setElementZIndex (in nsIDOMElement aElement, in long aZorder);

pub unsafe fn RelativeChangeElementZIndex(
&self,
aElement: *const nsIDOMElement,
aChange: int32_t,
_retval: *mut int32_t
) -> nsresult

/**
   * adds aChange to the z-index of an arbitrary element.
   * @return         the new z-index of the element
   * @param aElement [IN] the element
   * @param aChange  [IN] relative change to apply to current z-index of
   *                      the element
   */

long relativeChangeElementZIndex (in nsIDOMElement aElement, in long aChange);

pub unsafe fn ShowGrabberOnElement(
&self,
aElement: *const nsIDOMElement
) -> nsresult

/**
   * shows a grabber attached to an arbitrary element. The grabber is an image
   * positioned on the left hand side of the top border of the element. Dragging
   * and dropping it allows to change the element's absolute position in the
   * document. See chrome://editor/content/images/grabber.gif
   * @param aElement [IN] the element
   */

void showGrabberOnElement (in nsIDOMElement aElement);

pub unsafe fn HideGrabber(&self) -> nsresult

/**
   * hide the grabber if it shown.
   */

void hideGrabber ();

pub unsafe fn RefreshGrabber(&self) -> nsresult

/**
   * refreshes the grabber if it shown, possibly updating its position or
   * even hiding it.
   */

void refreshGrabber ();