Struct xpcom::interfaces::nsIHTMLInlineTableEditor

#[repr(C)]

pub struct nsIHTMLInlineTableEditor { /* fields omitted */ }

interface nsIHTMLInlineTableEditor : nsISupports

Methods

impl nsIHTMLInlineTableEditor

pub fn coerce<T: nsIHTMLInlineTableEditorCoerce>(&self) -> &T

Cast this nsIHTMLInlineTableEditor to one of its base interfaces.

impl nsIHTMLInlineTableEditor

pub unsafe fn GetInlineTableEditingEnabled(
&self,
aInlineTableEditingEnabled: *mut bool
) -> nsresult

/**
   * boolean indicating if inline table editing is enabled in the editor.
   * When inline table editing is enabled, and when the selection is
   * contained in a table cell, special buttons allowing to add/remove
   * a line/column are available on the cell's border.
   */

attribute boolean inlineTableEditingEnabled;

pub unsafe fn SetInlineTableEditingEnabled(
&self,
aInlineTableEditingEnabled: bool
) -> nsresult

/**
   * boolean indicating if inline table editing is enabled in the editor.
   * When inline table editing is enabled, and when the selection is
   * contained in a table cell, special buttons allowing to add/remove
   * a line/column are available on the cell's border.
   */

attribute boolean inlineTableEditingEnabled;

pub unsafe fn ShowInlineTableEditingUI(
&self,
aCell: *const nsIDOMElement
) -> nsresult

/**
   * Shows inline table editing UI around a table cell
   * @param aCell [IN] a DOM Element being a table cell, td or th
   */

void showInlineTableEditingUI (in nsIDOMElement aCell);

pub unsafe fn HideInlineTableEditingUI(&self) -> nsresult

/**
   * Hide all inline table editing UI
   */

void hideInlineTableEditingUI ();

pub unsafe fn DoInlineTableEditingAction(
&self,
aUIAnonymousElement: *const nsIDOMElement
) -> nsresult

/**
   * Modifies the table containing the selection according to the
   * activation of an inline table editing UI element
   * @param aUIAnonymousElement [IN] the inline table editing UI element
   */

void doInlineTableEditingAction (in nsIDOMElement aUIAnonymousElement);

pub unsafe fn RefreshInlineTableEditingUI(&self) -> nsresult

/**
   * Refresh already visible inline table editing UI
   */

void refreshInlineTableEditingUI ();