Struct xpcom::interfaces::nsIHelperAppLauncher

#[repr(C)]

pub struct nsIHelperAppLauncher { /* fields omitted */ }

interface nsIHelperAppLauncher : nsICancelable

/**
 * A helper app launcher is a small object created to handle the launching
 * of an external application.
 *
 * Note that cancelling the load via the nsICancelable interface will release
 * the reference to the launcher dialog.
 */

Methods

impl nsIHelperAppLauncher

pub fn coerce<T: nsIHelperAppLauncherCoerce>(&self) -> &T

Cast this nsIHelperAppLauncher to one of its base interfaces.

impl nsIHelperAppLauncher

pub unsafe fn GetMIMEInfo(&self, aMIMEInfo: *mut *const nsIMIMEInfo) -> nsresult

/**
   * The mime info object associated with the content type this helper app
   * launcher is currently attempting to load
   */

readonly attribute nsIMIMEInfo MIMEInfo;

pub unsafe fn GetSource(&self, aSource: *mut *const nsIURI) -> nsresult

/**
   * The source uri
   */

readonly attribute nsIURI source;

pub unsafe fn GetSuggestedFileName(
&self,
aSuggestedFileName: &mut nsAString
) -> nsresult

/**
   * The suggested name for this file
   */

readonly attribute AString suggestedFileName;

pub unsafe fn SaveToDisk(
&self,
aNewFileLocation: *const nsIFile,
aRememberThisPreference: bool
) -> nsresult

/**
   * Saves the final destination of the file. Does not actually perform the
   * save.
   * NOTE: This will release the reference to the
   * nsIHelperAppLauncherDialog.
   */

void saveToDisk (in nsIFile aNewFileLocation, in boolean aRememberThisPreference);

pub unsafe fn LaunchWithApplication(
&self,
aApplication: *const nsIFile,
aRememberThisPreference: bool
) -> nsresult

/**
   * Remembers that aApplication should be used to launch this content. Does
   * not actually launch the application.
   * NOTE: This will release the reference to the nsIHelperAppLauncherDialog.
   * @param aApplication nsIFile corresponding to the location of the application to use.
   * @param aRememberThisPreference TRUE if we should remember this choice.
   */

void launchWithApplication (in nsIFile aApplication, in boolean aRememberThisPreference);

pub unsafe fn SaveDestinationAvailable(&self, aFile: *const nsIFile) -> nsresult

/**
   * Callback invoked by nsIHelperAppLauncherDialog::promptForSaveToFileAsync
   * after the user has chosen a file through the File Picker (or dismissed it).
   * @param aFile The file that was chosen by the user (or null if dialog was dismissed).
   */

void saveDestinationAvailable (in nsIFile aFile);

pub unsafe fn SetWebProgressListener(
&self,
aWebProgressListener: *const nsIWebProgressListener2
) -> nsresult

/**
   * The following methods are used by the progress dialog to get or set
   * information on the current helper app launcher download.
   * This reference will be released when the download is finished (after the
   * listener receives the STATE_STOP notification).
   */

void setWebProgressListener (in nsIWebProgressListener2 aWebProgressListener);

pub unsafe fn GetTargetFile(&self, aTargetFile: *mut *const nsIFile) -> nsresult

/**
   * The file we are saving to
   */

readonly attribute nsIFile targetFile;

pub unsafe fn GetTargetFileIsExecutable(
&self,
aTargetFileIsExecutable: *mut bool
) -> nsresult

/**
   * The executable-ness of the target file
   */

readonly attribute boolean targetFileIsExecutable;

pub unsafe fn GetTimeDownloadStarted(
&self,
aTimeDownloadStarted: *mut PRTime
) -> nsresult

/**
   * Time when the download started
   */

readonly attribute PRTime timeDownloadStarted;

pub unsafe fn GetContentLength(&self, aContentLength: *mut int64_t) -> nsresult

/**
   * The download content length, or -1 if the length is not available.
   */

readonly attribute int64_t contentLength;