Struct xpcom::interfaces::nsIHttpActivityObserver

#[repr(C)]

pub struct nsIHttpActivityObserver { /* fields omitted */ }

interface nsIHttpActivityObserver : nsISupports

/**
 * nsIHttpActivityObserver
 *
 * This interface provides a way for http activities to be reported
 * to observers.
 */

Methods

impl nsIHttpActivityObserver

pub fn coerce<T: nsIHttpActivityObserverCoerce>(&self) -> &T

Cast this nsIHttpActivityObserver to one of its base interfaces.

impl nsIHttpActivityObserver

pub const ACTIVITY_TYPE_SOCKET_TRANSPORT: i64

ACTIVITY_TYPE_SOCKET_TRANSPORT: i64 = 1

pub const ACTIVITY_TYPE_HTTP_TRANSACTION: i64

ACTIVITY_TYPE_HTTP_TRANSACTION: i64 = 2

pub const ACTIVITY_SUBTYPE_REQUEST_HEADER: i64

ACTIVITY_SUBTYPE_REQUEST_HEADER: i64 = 20481

pub const ACTIVITY_SUBTYPE_REQUEST_BODY_SENT: i64

ACTIVITY_SUBTYPE_REQUEST_BODY_SENT: i64 = 20482

pub const ACTIVITY_SUBTYPE_RESPONSE_START: i64

ACTIVITY_SUBTYPE_RESPONSE_START: i64 = 20483

pub const ACTIVITY_SUBTYPE_RESPONSE_HEADER: i64

ACTIVITY_SUBTYPE_RESPONSE_HEADER: i64 = 20484

pub const ACTIVITY_SUBTYPE_RESPONSE_COMPLETE: i64

ACTIVITY_SUBTYPE_RESPONSE_COMPLETE: i64 = 20485

pub const ACTIVITY_SUBTYPE_TRANSACTION_CLOSE: i64

ACTIVITY_SUBTYPE_TRANSACTION_CLOSE: i64 = 20486

pub unsafe fn ObserveActivity(
&self,
aHttpChannel: *const nsISupports,
aActivityType: uint32_t,
aActivitySubtype: uint32_t,
aTimestamp: PRTime,
aExtraSizeData: uint64_t,
aExtraStringData: &nsACString
) -> nsresult

/**
     * observe activity from the http transport
     *
     * @param aHttpChannel
     *        nsISupports interface for the the http channel that
     *        generated this activity
     * @param aActivityType
     *        The value of this aActivityType will be one of
     *          ACTIVITY_TYPE_SOCKET_TRANSPORT or
     *          ACTIVITY_TYPE_HTTP_TRANSACTION
     * @param aActivitySubtype
     *        The value of this aActivitySubtype, will be depend
     *        on the value of aActivityType. When aActivityType
     *        is ACTIVITY_TYPE_SOCKET_TRANSPORT
     *          aActivitySubtype will be one of the
     *          nsISocketTransport::STATUS_???? values defined in
     *          nsISocketTransport.idl
     *        OR when aActivityType
     *        is ACTIVITY_TYPE_HTTP_TRANSACTION
     *          aActivitySubtype will be one of the
     *          nsIHttpActivityObserver::ACTIVITY_SUBTYPE_???? values
     *          defined below
     * @param aTimestamp
     *        microseconds past the epoch of Jan 1, 1970
     * @param aExtraSizeData
     *        Any extra size data optionally available with
     *        this activity
     * @param aExtraStringData
     *        Any extra string data optionally available with
     *        this activity
     */

[must_use] void observeActivity (in nsISupports aHttpChannel, in uint32_t aActivityType, in uint32_t aActivitySubtype, in PRTime aTimestamp, in uint64_t aExtraSizeData, in ACString aExtraStringData);

pub unsafe fn GetIsActive(&self, aIsActive: *mut bool) -> nsresult

/**
     * This attribute is true when this interface is active and should
     * observe http activities. When false, observeActivity() should not
     * be called. It is present for compatibility reasons and should be
     * implemented only by nsHttpActivityDistributor.
     */

[must_use] readonly attribute boolean isActive;