Struct xpcom::interfaces::nsIHttpChannel

#[repr(C)]

pub struct nsIHttpChannel { /* fields omitted */ }

interface nsIHttpChannel : nsIChannel

/**
 * nsIHttpChannel
 *
 * This interface allows for the modification of HTTP request parameters and
 * the inspection of the resulting HTTP response status and headers when they
 * become available.
 */

Methods

impl nsIHttpChannel

pub fn coerce<T: nsIHttpChannelCoerce>(&self) -> &T

Cast this nsIHttpChannel to one of its base interfaces.

impl nsIHttpChannel

pub const REFERRER_POLICY_UNSET: i64

REFERRER_POLICY_UNSET: i64 = 0

/**
     * Referrer policies. See ReferrerPolicy.h for more details.
     */

pub const REFERRER_POLICY_NO_REFERRER_WHEN_DOWNGRADE: i64

REFERRER_POLICY_NO_REFERRER_WHEN_DOWNGRADE: i64 = 1

pub const REFERRER_POLICY_NO_REFERRER: i64

REFERRER_POLICY_NO_REFERRER: i64 = 2

pub const REFERRER_POLICY_ORIGIN: i64

REFERRER_POLICY_ORIGIN: i64 = 3

pub const REFERRER_POLICY_ORIGIN_WHEN_XORIGIN: i64

REFERRER_POLICY_ORIGIN_WHEN_XORIGIN: i64 = 4

pub const REFERRER_POLICY_UNSAFE_URL: i64

REFERRER_POLICY_UNSAFE_URL: i64 = 5

pub const REFERRER_POLICY_SAME_ORIGIN: i64

REFERRER_POLICY_SAME_ORIGIN: i64 = 6

pub const REFERRER_POLICY_STRICT_ORIGIN: i64

REFERRER_POLICY_STRICT_ORIGIN: i64 = 7

pub const REFERRER_POLICY_STRICT_ORIGIN_WHEN_XORIGIN: i64

REFERRER_POLICY_STRICT_ORIGIN_WHEN_XORIGIN: i64 = 8

pub unsafe fn GetRequestMethod(
&self,
aRequestMethod: &mut nsACString
) -> nsresult

/**************************************************************************
     * REQUEST CONFIGURATION
     *
     * Modifying request parameters after asyncOpen has been called is an error.
     */
/**
     * Set/get the HTTP request method (default is "GET").  Both setter and
     * getter are case sensitive.
     *
     * This attribute may only be set before the channel is opened.
     *
     * NOTE: The data for a "POST" or "PUT" request can be configured via
     * nsIUploadChannel; however, after setting the upload data, it may be
     * necessary to set the request method explicitly.  The documentation
     * for nsIUploadChannel has further details.
     *
     * @throws NS_ERROR_IN_PROGRESS if set after the channel has been opened.
     */

[must_use] attribute ACString requestMethod;

pub unsafe fn SetRequestMethod(&self, aRequestMethod: &nsACString) -> nsresult

/**************************************************************************
     * REQUEST CONFIGURATION
     *
     * Modifying request parameters after asyncOpen has been called is an error.
     */
/**
     * Set/get the HTTP request method (default is "GET").  Both setter and
     * getter are case sensitive.
     *
     * This attribute may only be set before the channel is opened.
     *
     * NOTE: The data for a "POST" or "PUT" request can be configured via
     * nsIUploadChannel; however, after setting the upload data, it may be
     * necessary to set the request method explicitly.  The documentation
     * for nsIUploadChannel has further details.
     *
     * @throws NS_ERROR_IN_PROGRESS if set after the channel has been opened.
     */

[must_use] attribute ACString requestMethod;

pub unsafe fn GetReferrer(&self, aReferrer: *mut *const nsIURI) -> nsresult

/**
     * Get/set the HTTP referrer URI.  This is the address (URI) of the
     * resource from which this channel's URI was obtained (see RFC2616 section
     * 14.36).
     *
     * This attribute may only be set before the channel is opened.
     *
     * NOTE: The channel may silently refuse to set the Referer header if the
     * URI does not pass certain security checks (e.g., a "https://" URL will
     * never be sent as the referrer for a plaintext HTTP request).  The
     * implementation is not required to throw an exception when the referrer
     * URI is rejected.
     *
     * @throws NS_ERROR_IN_PROGRESS if set after the channel has been opened.
     * @throws NS_ERROR_FAILURE if used for setting referrer during
     *         visitRequestHeaders. Getting the value will not throw.
     */

[must_use] attribute nsIURI referrer;

pub unsafe fn SetReferrer(&self, aReferrer: *const nsIURI) -> nsresult

/**
     * Get/set the HTTP referrer URI.  This is the address (URI) of the
     * resource from which this channel's URI was obtained (see RFC2616 section
     * 14.36).
     *
     * This attribute may only be set before the channel is opened.
     *
     * NOTE: The channel may silently refuse to set the Referer header if the
     * URI does not pass certain security checks (e.g., a "https://" URL will
     * never be sent as the referrer for a plaintext HTTP request).  The
     * implementation is not required to throw an exception when the referrer
     * URI is rejected.
     *
     * @throws NS_ERROR_IN_PROGRESS if set after the channel has been opened.
     * @throws NS_ERROR_FAILURE if used for setting referrer during
     *         visitRequestHeaders. Getting the value will not throw.
     */

[must_use] attribute nsIURI referrer;

pub unsafe fn GetReferrerPolicy(
&self,
aReferrerPolicy: *mut uint32_t
) -> nsresult

/**
     * Get the HTTP referrer policy.  The policy is retrieved from the meta
     * referrer tag, which can be one of many values (see ReferrerPolicy.h for
     * more details).
     */

[must_use] readonly attribute unsigned long referrerPolicy;

pub unsafe fn SetReferrerWithPolicy(
&self,
referrer: *const nsIURI,
referrerPolicy: uint32_t
) -> nsresult

/**
     * Set the HTTP referrer URI with a referrer policy.
     * @throws NS_ERROR_FAILURE if called during visitRequestHeaders.
     */

[must_use] void setReferrerWithPolicy (in nsIURI referrer, in unsigned long referrerPolicy);

pub unsafe fn GetProtocolVersion(
&self,
aProtocolVersion: &mut nsACString
) -> nsresult

/**
     * Returns the network protocol used to fetch the resource as identified
     * by the ALPN Protocol ID.
     *
     * @throws NS_ERROR_NOT_AVAILABLE if called before the response
     *         has been received (before onStartRequest).
     */

[must_use] readonly attribute ACString protocolVersion;

pub unsafe fn GetTransferSize(&self, aTransferSize: *mut uint64_t) -> nsresult

/**
     * size consumed by the response header fields and the response payload body
     */

[must_use] readonly attribute uint64_t transferSize;

pub unsafe fn GetDecodedBodySize(
&self,
aDecodedBodySize: *mut uint64_t
) -> nsresult

/**
     * The size of the message body received by the client,
     * after removing any applied content-codings
     */

[must_use] readonly attribute uint64_t decodedBodySize;

pub unsafe fn GetEncodedBodySize(
&self,
aEncodedBodySize: *mut uint64_t
) -> nsresult

/**
     * The size in octets of the payload body, prior to removing content-codings
     */

[must_use] readonly attribute uint64_t encodedBodySize;

pub unsafe fn GetRequestHeader(
&self,
aHeader: &nsACString,
_retval: &mut nsACString
) -> nsresult

/**
     * Get the value of a particular request header.
     *
     * @param aHeader
     *        The case-insensitive name of the request header to query (e.g.,
     *        "Cache-Control").
     *
     * @return the value of the request header.
     * @throws NS_ERROR_NOT_AVAILABLE if the header is not set.
     */

[must_use] ACString getRequestHeader (in ACString aHeader);

pub unsafe fn SetRequestHeader(
&self,
aHeader: &nsACString,
aValue: &nsACString,
aMerge: bool
) -> nsresult

/**
     * Set the value of a particular request header.
     *
     * This method allows, for example, the cookies module to add "Cookie"
     * headers to the outgoing HTTP request.
     *
     * This method may only be called before the channel is opened.
     *
     * @param aHeader
     *        The case-insensitive name of the request header to set (e.g.,
     *        "Cookie").
     * @param aValue
     *        The request header value to set (e.g., "X=1").
     * @param aMerge
     *        If true, the new header value will be merged with any existing
     *        values for the specified header.  This flag is ignored if the
     *        specified header does not support merging (e.g., the "Content-
     *        Type" header can only have one value).  The list of headers for
     *        which this flag is ignored is an implementation detail.  If this
     *        flag is false, then the header value will be replaced with the
     *        contents of |aValue|.
     *
     * If aValue is empty and aMerge is false, the header will be cleared.
     *
     * @throws NS_ERROR_IN_PROGRESS if called after the channel has been
     *         opened.
     * @throws NS_ERROR_FAILURE if called during visitRequestHeaders.
     */

[must_use] void setRequestHeader (in ACString aHeader, in ACString aValue, in boolean aMerge);

pub unsafe fn SetEmptyRequestHeader(&self, aHeader: &nsACString) -> nsresult

/**
     * Set a request header with empty value.
     *
     * This should be used with caution in the cases where the behavior of
     * setRequestHeader ignoring empty header values is undesirable.
     *
     * This method may only be called before the channel is opened.
     *
     * @param aHeader
     *        The case-insensitive name of the request header to set (e.g.,
     *        "Cookie").
     *
     * @throws NS_ERROR_IN_PROGRESS if called after the channel has been
     *         opened.
     * @throws NS_ERROR_FAILURE if called during visitRequestHeaders.
     */

[must_use] void setEmptyRequestHeader (in ACString aHeader);

pub unsafe fn VisitRequestHeaders(
&self,
aVisitor: *const nsIHttpHeaderVisitor
) -> nsresult

/**
     * Call this method to visit all request headers.  Calling setRequestHeader
     * while visiting request headers has undefined behavior.  Don't do it!
     *
     * @param aVisitor
     *        the header visitor instance.
     */

[must_use] void visitRequestHeaders (in nsIHttpHeaderVisitor aVisitor);

pub unsafe fn VisitNonDefaultRequestHeaders(
&self,
aVisitor: *const nsIHttpHeaderVisitor
) -> nsresult

/**
     * Call this method to visit all non-default (UA-provided) request headers.
     * Calling setRequestHeader while visiting request headers has undefined
     * behavior. Don't do it!
     *
     * @param aVisitor
     *        the header visitor instance.
     */

[must_use] void visitNonDefaultRequestHeaders (in nsIHttpHeaderVisitor aVisitor);

pub unsafe fn GetAllowPipelining(&self, aAllowPipelining: *mut bool) -> nsresult

/**
     * This attribute no longer has any effect, it remains for backwards compat
     *
     * @throws NS_ERROR_FAILURE if set after the channel has been opened.
     */

[must_use] attribute boolean allowPipelining;

pub unsafe fn SetAllowPipelining(&self, aAllowPipelining: bool) -> nsresult

/**
     * This attribute no longer has any effect, it remains for backwards compat
     *
     * @throws NS_ERROR_FAILURE if set after the channel has been opened.
     */

[must_use] attribute boolean allowPipelining;

pub unsafe fn GetAllowSTS(&self, aAllowSTS: *mut bool) -> nsresult

/**
     * This attribute of the channel indicates whether or not
     * the underlying HTTP transaction should be honor stored Strict Transport
     * Security directives for its principal. It defaults to true. Using
     * OCSP to bootstrap the HTTPs is the likely use case for setting it to
     * false.
     *
     * This attribute may only be set before the channel is opened.
     *
     * @throws NS_ERROR_IN_PROGRESS or NS_ERROR_ALREADY_OPENED
     *         if called after the channel has been opened.
     */

[must_use] attribute boolean allowSTS;

pub unsafe fn SetAllowSTS(&self, aAllowSTS: bool) -> nsresult

/**
     * This attribute of the channel indicates whether or not
     * the underlying HTTP transaction should be honor stored Strict Transport
     * Security directives for its principal. It defaults to true. Using
     * OCSP to bootstrap the HTTPs is the likely use case for setting it to
     * false.
     *
     * This attribute may only be set before the channel is opened.
     *
     * @throws NS_ERROR_IN_PROGRESS or NS_ERROR_ALREADY_OPENED
     *         if called after the channel has been opened.
     */

[must_use] attribute boolean allowSTS;

pub unsafe fn GetRedirectionLimit(
&self,
aRedirectionLimit: *mut uint32_t
) -> nsresult

/**
     * This attribute specifies the number of redirects this channel is allowed
     * to make.  If zero, the channel will fail to redirect and will generate
     * a NS_ERROR_REDIRECT_LOOP failure status.
     *
     * NOTE: An HTTP redirect results in a new channel being created.  If the
     * new channel supports nsIHttpChannel, then it will be assigned a value
     * to its |redirectionLimit| attribute one less than the value of the
     * redirected channel's |redirectionLimit| attribute.  The initial value
     * for this attribute may be a configurable preference (depending on the
     * implementation).
     */

[must_use] attribute unsigned long redirectionLimit;

pub unsafe fn SetRedirectionLimit(
&self,
aRedirectionLimit: uint32_t
) -> nsresult

/**
     * This attribute specifies the number of redirects this channel is allowed
     * to make.  If zero, the channel will fail to redirect and will generate
     * a NS_ERROR_REDIRECT_LOOP failure status.
     *
     * NOTE: An HTTP redirect results in a new channel being created.  If the
     * new channel supports nsIHttpChannel, then it will be assigned a value
     * to its |redirectionLimit| attribute one less than the value of the
     * redirected channel's |redirectionLimit| attribute.  The initial value
     * for this attribute may be a configurable preference (depending on the
     * implementation).
     */

[must_use] attribute unsigned long redirectionLimit;

pub unsafe fn GetResponseStatus(
&self,
aResponseStatus: *mut uint32_t
) -> nsresult

/**************************************************************************
     * RESPONSE INFO
     *
     * Accessing response info before the onStartRequest event is an error.
     */
/**
     * Get the HTTP response code (e.g., 200).
     *
     * @throws NS_ERROR_NOT_AVAILABLE if called before the response
     *         has been received (before onStartRequest).
     */

[must_use] readonly attribute unsigned long responseStatus;

pub unsafe fn GetResponseStatusText(
&self,
aResponseStatusText: &mut nsACString
) -> nsresult

/**
     * Get the HTTP response status text (e.g., "OK").
     *
     * NOTE: This returns the raw (possibly 8-bit) text from the server.  There
     * are no assumptions made about the charset of the returned text.  You
     * have been warned!
     *
     * @throws NS_ERROR_NOT_AVAILABLE if called before the response
     *         has been received (before onStartRequest).
     */

[must_use] readonly attribute ACString responseStatusText;

pub unsafe fn GetRequestSucceeded(
&self,
aRequestSucceeded: *mut bool
) -> nsresult

/**
     * Returns true if the HTTP response code indicates success.  The value of
     * nsIRequest::status will be NS_OK even when processing a 404 response
     * because a 404 response may include a message body that (in some cases)
     * should be shown to the user.
     *
     * Use this attribute to distinguish server error pages from normal pages,
     * instead of comparing the response status manually against the set of
     * valid response codes, if that is required by your application.
     *
     * @throws NS_ERROR_NOT_AVAILABLE if called before the response
     *         has been received (before onStartRequest).
     */

[must_use] readonly attribute boolean requestSucceeded;

pub unsafe fn GetIsMainDocumentChannel(
&self,
aIsMainDocumentChannel: *mut bool
) -> nsresult

/** Indicates whether channel should be treated as the main one for the
    *  current document.  If manually set to true, will always remain true.  Otherwise,
    *  will be true if LOAD_DOCUMENT_URI is set in the channel's loadflags.
    */

[must_use] attribute boolean isMainDocumentChannel;

pub unsafe fn SetIsMainDocumentChannel(
&self,
aIsMainDocumentChannel: bool
) -> nsresult

/** Indicates whether channel should be treated as the main one for the
    *  current document.  If manually set to true, will always remain true.  Otherwise,
    *  will be true if LOAD_DOCUMENT_URI is set in the channel's loadflags.
    */

[must_use] attribute boolean isMainDocumentChannel;

pub unsafe fn GetResponseHeader(
&self,
header: &nsACString,
_retval: &mut nsACString
) -> nsresult

/**
     * Get the value of a particular response header.
     *
     * @param aHeader
     *        The case-insensitive name of the response header to query (e.g.,
     *        "Set-Cookie").
     *
     * @return the value of the response header.
     *
     * @throws NS_ERROR_NOT_AVAILABLE if called before the response
     *         has been received (before onStartRequest) or if the header is
     *         not set in the response.
     */

[must_use] ACString getResponseHeader (in ACString header);

pub unsafe fn SetResponseHeader(
&self,
header: &nsACString,
value: &nsACString,
merge: bool
) -> nsresult

/**
     * Set the value of a particular response header.
     *
     * This method allows, for example, the HTML content sink to inform the HTTP
     * channel about HTTP-EQUIV headers found in HTML <META> tags.
     *
     * @param aHeader
     *        The case-insensitive name of the response header to set (e.g.,
     *        "Cache-control").
     * @param aValue
     *        The response header value to set (e.g., "no-cache").
     * @param aMerge
     *        If true, the new header value will be merged with any existing
     *        values for the specified header.  This flag is ignored if the
     *        specified header does not support merging (e.g., the "Content-
     *        Type" header can only have one value).  The list of headers for
     *        which this flag is ignored is an implementation detail.  If this
     *        flag is false, then the header value will be replaced with the
     *        contents of |aValue|.
     *
     * If aValue is empty and aMerge is false, the header will be cleared.
     *
     * @throws NS_ERROR_NOT_AVAILABLE if called before the response
     *         has been received (before onStartRequest).
     * @throws NS_ERROR_ILLEGAL_VALUE if changing the value of this response
     *         header is not allowed.
     * @throws NS_ERROR_FAILURE if called during visitResponseHeaders,
     *         VisitOriginalResponseHeaders or getOriginalResponseHeader.
     */

[must_use] void setResponseHeader (in ACString header, in ACString value, in boolean merge);

pub unsafe fn VisitResponseHeaders(
&self,
aVisitor: *const nsIHttpHeaderVisitor
) -> nsresult

/**
     * Call this method to visit all response headers.  Calling
     * setResponseHeader while visiting response headers has undefined
     * behavior.  Don't do it!
     *
     * @param aVisitor
     *        the header visitor instance.
     *
     * @throws NS_ERROR_NOT_AVAILABLE if called before the response
     *         has been received (before onStartRequest).
     */

[must_use] void visitResponseHeaders (in nsIHttpHeaderVisitor aVisitor);

pub unsafe fn GetOriginalResponseHeader(
&self,
aHeader: &nsACString,
aVisitor: *const nsIHttpHeaderVisitor
) -> nsresult

/**
     * Get the value(s) of a particular response header in the form and order
     * it has been received from the remote peer. There can be multiple headers
     * with the same name.
     *
     * @param aHeader
     *        The case-insensitive name of the response header to query (e.g.,
     *        "Set-Cookie").
     *
     * @param aVisitor
     *        the header visitor instance.
     *
     * @throws NS_ERROR_NOT_AVAILABLE if called before the response
     *         has been received (before onStartRequest) or if the header is
     *         not set in the response.
     */

[must_use] void getOriginalResponseHeader (in ACString aHeader, in nsIHttpHeaderVisitor aVisitor);

pub unsafe fn VisitOriginalResponseHeaders(
&self,
aVisitor: *const nsIHttpHeaderVisitor
) -> nsresult

/**
     * Call this method to visit all response headers in the form and order as
     * they have been received from the remote peer.
     * Calling setResponseHeader while visiting response headers has undefined
     * behavior.  Don't do it!
     *
     * @param aVisitor
     *        the header visitor instance.
     *
     * @throws NS_ERROR_NOT_AVAILABLE if called before the response
     *         has been received (before onStartRequest).
     */

[must_use] void visitOriginalResponseHeaders (in nsIHttpHeaderVisitor aVisitor);

pub unsafe fn IsNoStoreResponse(&self, _retval: *mut bool) -> nsresult

/**
     * Returns true if the server sent a "Cache-Control: no-store" response
     * header.
     *
     * @throws NS_ERROR_NOT_AVAILABLE if called before the response
     *         has been received (before onStartRequest).
     */

[must_use] boolean isNoStoreResponse ();

pub unsafe fn IsNoCacheResponse(&self, _retval: *mut bool) -> nsresult

/**
     * Returns true if the server sent the equivalent of a "Cache-control:
     * no-cache" response header.  Equivalent response headers include:
     * "Pragma: no-cache", "Expires: 0", and "Expires" with a date value
     * in the past relative to the value of the "Date" header.
     *
     * @throws NS_ERROR_NOT_AVAILABLE if called before the response
     *         has been received (before onStartRequest).
     */

[must_use] boolean isNoCacheResponse ();

pub unsafe fn IsPrivateResponse(&self, _retval: *mut bool) -> nsresult

/**
     * Returns true if the server sent a "Cache-Control: private" response
     * header.
     *
     * @throws NS_ERROR_NOT_AVAILABLE if called before the response
     *         has been received (before onStartRequest).
     */

[must_use] boolean isPrivateResponse ();

pub unsafe fn RedirectTo(&self, aTargetURI: *const nsIURI) -> nsresult

/**
     * Instructs the channel to immediately redirect to a new destination.
     * Can only be called on channels that have not yet called their
     * listener's OnStartRequest(). Generally that means the latest time
     * this can be used is one of:
     *    "http-on-examine-response"
     *    "http-on-examine-merged-response"
     *    "http-on-examine-cached-response"
     *
     * When non-null URL is set before AsyncOpen:
     *  we attempt to redirect to the targetURI before we even start building
     *  and sending the request to the cache or the origin server.
     *  If the redirect is vetoed, we fail the channel.
     *
     * When set between AsyncOpen and first call to OnStartRequest being called:
     *  we attempt to redirect before we start delivery of network or cached
     *  response to the listener.  If vetoed, we continue with delivery of
     *  the original content to the channel listener.
     *
     * When passed aTargetURI is null the channel behaves normally (can be
     * rewritten).
     *
     * This method provides no explicit conflict resolution. The last
     * caller to call it wins.
     *
     * @throws NS_ERROR_NOT_AVAILABLE if called after the channel has already
     *         started to deliver the content to its listener.
     */

[must_use] void redirectTo (in nsIURI aTargetURI);

pub unsafe fn UpgradeToSecure(&self) -> nsresult

/**
     * Flags a channel to be upgraded to HTTPS.
     *
     * Upgrading to a secure channel must happen before or during
     * "http-on-modify-request". If redirectTo is called early as well, it
     * will win and upgradeToSecure will be a no-op.
     *
     * @throws NS_ERROR_NOT_AVAILABLE if called after the channel has already
     *         started to deliver the content to its listener.
     */

[must_use] void upgradeToSecure ();

pub unsafe fn GetRequestContextID(
&self,
aRequestContextID: *mut uint64_t
) -> nsresult

/**
     * Identifies the request context for this load.
     */

[must_use,noscript] attribute uint64_t requestContextID;

pub unsafe fn SetRequestContextID(
&self,
aRequestContextID: uint64_t
) -> nsresult

/**
     * Identifies the request context for this load.
     */

[must_use,noscript] attribute uint64_t requestContextID;

pub unsafe fn GetChannelId(&self, aChannelId: *mut uint64_t) -> nsresult

/**
     * Unique ID of the channel, shared between parent and child. Needed if
     * the channel activity needs to be monitored across process boundaries,
     * like in devtools net monitor. See bug 1274556.
     */

[must_use] attribute uint64_t channelId;

pub unsafe fn SetChannelId(&self, aChannelId: uint64_t) -> nsresult

/**
     * Unique ID of the channel, shared between parent and child. Needed if
     * the channel activity needs to be monitored across process boundaries,
     * like in devtools net monitor. See bug 1274556.
     */

[must_use] attribute uint64_t channelId;

pub unsafe fn GetTopLevelContentWindowId(
&self,
aTopLevelContentWindowId: *mut uint64_t
) -> nsresult

/**
     * ID of the top-level document's inner window.  Identifies the content
     * this channels is being load in.
     */

[must_use] attribute uint64_t topLevelContentWindowId;

pub unsafe fn SetTopLevelContentWindowId(
&self,
aTopLevelContentWindowId: uint64_t
) -> nsresult

/**
     * ID of the top-level document's inner window.  Identifies the content
     * this channels is being load in.
     */

[must_use] attribute uint64_t topLevelContentWindowId;

pub unsafe fn GetIsTrackingResource(&self) -> bool

/**
     * Returns true if the channel has loaded a resource that is on the tracking
     * protection list.  This is only available if the
     * privacy.trackingprotection.annotate_channels pref is set and its value
     * should only be relied on after the channel has established a connection.
     */

[infallible] readonly attribute boolean isTrackingResource;

pub unsafe fn GetTopLevelOuterContentWindowId(
&self,
aTopLevelOuterContentWindowId: *mut uint64_t
) -> nsresult

/**
     * ID of the top-level outer content window. Identifies this channel's
     * top-level window it comes from.
     *
     * NOTE: The setter of this attribute is currently for xpcshell test only.
     *       Don't alter it otherwise.
     */

[must_use] attribute uint64_t topLevelOuterContentWindowId;

pub unsafe fn SetTopLevelOuterContentWindowId(
&self,
aTopLevelOuterContentWindowId: uint64_t
) -> nsresult

/**
     * ID of the top-level outer content window. Identifies this channel's
     * top-level window it comes from.
     *
     * NOTE: The setter of this attribute is currently for xpcshell test only.
     *       Don't alter it otherwise.
     */

[must_use] attribute uint64_t topLevelOuterContentWindowId;

pub unsafe fn LogBlockedCORSRequest(&self, aMessage: &nsAString) -> nsresult

/**
     * In e10s, the information that the CORS response blocks the load is in the
     * parent, which doesn't know the true window id of the request, so we may
     * need to proxy the request to the child.
     *
     * @param aMessage
     *        The message to print in the console.
     */

void logBlockedCORSRequest (in AString aMessage);