Struct xpcom::interfaces::nsIHttpChannelAuthProvider

#[repr(C)]

pub struct nsIHttpChannelAuthProvider { /* fields omitted */ }

interface nsIHttpChannelAuthProvider : nsICancelable

/**
 * nsIHttpChannelAuthProvider
 *
 * This interface is intended for providing authentication for http-style
 * channels, like nsIHttpChannel and nsIWebSocket, which implement the
 * nsIHttpAuthenticableChannel interface.
 *
 * When requesting pages AddAuthorizationHeaders MUST be called
 * in order to get the http cached headers credentials. When the request is
 * unsuccessful because of receiving either a 401 or 407 http response code
 * ProcessAuthentication MUST be called and the page MUST be requested again
 * with the new credentials that the user has provided. After a successful
 * request, checkForSuperfluousAuth MAY be called, and disconnect MUST be
 * called.
 */

Methods

impl nsIHttpChannelAuthProvider

pub fn coerce<T: nsIHttpChannelAuthProviderCoerce>(&self) -> &T

Cast this nsIHttpChannelAuthProvider to one of its base interfaces.

impl nsIHttpChannelAuthProvider

pub unsafe fn Init(
&self,
channel: *const nsIHttpAuthenticableChannel
) -> nsresult

/**
   * Initializes the http authentication support for the channel.
   * Implementations must hold a weak reference of the channel.
   */

[must_use] void init (in nsIHttpAuthenticableChannel channel);

pub unsafe fn ProcessAuthentication(
&self,
httpStatus: uint32_t,
sslConnectFailed: bool
) -> nsresult

/**
   * Upon receipt of a server challenge, this function is called to determine
   * the credentials to send.
   *
   * @param httpStatus
   *        the http status received.
   * @param sslConnectFailed
   *        if the last ssl tunnel connection attempt was or not successful.
   * @param callback
   *        the callback to be called when it returns NS_ERROR_IN_PROGRESS.
   *        The implementation must hold a weak reference.
   *
   * @returns NS_OK if the credentials were got and set successfully.
   *          NS_ERROR_IN_PROGRESS if the credentials are going to be asked to
   *                               the user. The channel reference must be
   *                               alive until the feedback from
   *                               nsIHttpAuthenticableChannel's methods or
   *                               until disconnect be called.
   */

[must_use] void processAuthentication (in unsigned long httpStatus, in boolean sslConnectFailed);

pub unsafe fn AddAuthorizationHeaders(
&self,
dontUseCachedWWWCreds: bool
) -> nsresult

/**
   * Add credentials from the http auth cache.
   *
   * @param dontUseCachedWWWCreds
   *    When true, the method will not add any Authorization headers from
   *    the auth cache.
   */

[must_use] void addAuthorizationHeaders (in boolean dontUseCachedWWWCreds);

pub unsafe fn CheckForSuperfluousAuth(&self) -> nsresult

/**
   * Check if an unnecessary(and maybe malicious) url authentication has been
   * provided.
   */

[must_use] void checkForSuperfluousAuth ();

pub unsafe fn Disconnect(&self, status: nsresult) -> nsresult

/**
   * Cancel pending user auth prompts and release the callback and channel
   * weak references.
   */

[must_use] void disconnect (in nsresult status);