Struct xpcom::interfaces::nsIHttpServerStoppedCallback

#[repr(C)]

pub struct nsIHttpServerStoppedCallback { /* fields omitted */ }

interface nsIHttpServerStoppedCallback : nsISupports

Methods

impl nsIHttpServerStoppedCallback

pub fn coerce<T: nsIHttpServerStoppedCallbackCoerce>(&self) -> &T

Cast this nsIHttpServerStoppedCallback to one of its base interfaces.

impl nsIHttpServerStoppedCallback

pub unsafe fn OnStopped(&self) -> nsresult

/**
 * An interface through which a notification of the complete stopping (socket
 * closure, in-flight requests all fully served and responded to) of an HTTP
 * server may be received.
 */
/** Called when the corresponding server has been fully stopped. */

void onStopped ();