Struct xpcom::interfaces::nsIIDNService

#[repr(C)]

pub struct nsIIDNService { /* fields omitted */ }

interface nsIIDNService : nsISupports

/**
 * nsIIDNService interface.
 *
 * IDN (Internationalized Domain Name) support. Provides facilities
 * for manipulating IDN hostnames according to the specification set
 * forth by the IETF.
 *
 * IDN effort:
 * http://www.ietf.org/html.characters/idn-charter.html
 * http://www.i-dns.net
 *
 * IDNA specification:
 * http://search.ietf.org/internet-drafts/draft-ietf-idn-idna-06.txt
 */

Methods

impl nsIIDNService

pub fn coerce<T: nsIIDNServiceCoerce>(&self) -> &T

Cast this nsIIDNService to one of its base interfaces.

impl nsIIDNService

pub unsafe fn ConvertUTF8toACE(
&self,
input: &nsACString,
_retval: &mut nsACString
) -> nsresult

/**
     * Prepares the input hostname according to IDNA ToASCII operation,
     * the input hostname is assumed to be UTF8-encoded.
     */

ACString convertUTF8toACE (in AUTF8String input);

pub unsafe fn ConvertACEtoUTF8(
&self,
input: &nsACString,
_retval: &mut nsACString
) -> nsresult

/**
     * This is the ToUnicode operation as specified in the IDNA proposal,
     * with an additional step to encode the result in UTF-8.
     * It takes an ACE-encoded hostname and performs ToUnicode to it, then
     * encodes the resulting string into UTF8.
     */

AUTF8String convertACEtoUTF8 (in ACString input);

pub unsafe fn IsACE(&self, input: &nsACString, _retval: *mut bool) -> nsresult

/**
     * Checks if the input string is ACE encoded or not.
     */

boolean isACE (in ACString input);

pub unsafe fn Normalize(
&self,
input: &nsACString,
_retval: &mut nsACString
) -> nsresult

/**
     * Performs the unicode normalization needed for hostnames in IDN,
     * for callers that want early normalization.
     */

AUTF8String normalize (in AUTF8String input);

pub unsafe fn ConvertToDisplayIDN(
&self,
input: &nsACString,
isASCII: *mut bool,
_retval: &mut nsACString
) -> nsresult

/**
     * Normalizes and converts a host to UTF-8 if the host is in the IDN
     * whitelist, otherwise converts it to ACE. This is useful for display
     * purposes and to ensure an encoding consistent with nsIURI::GetHost().
     * If the result is ASCII or ACE encoded, |isASCII| will be true.
     */

AUTF8String convertToDisplayIDN (in AUTF8String input, out boolean isASCII);