Struct xpcom::interfaces::nsIInputIterator

#[repr(C)]

pub struct nsIInputIterator { /* fields omitted */ }

interface nsIInputIterator : nsISupports

/**
   * ...
   */

Methods

impl nsIInputIterator

pub fn coerce<T: nsIInputIteratorCoerce>(&self) -> &T

Cast this nsIInputIterator to one of its base interfaces.

impl nsIInputIterator

pub unsafe fn GetElement(&self, _retval: *mut *const nsISupports) -> nsresult

/**
       * Retrieve (and |AddRef()|) the element this iterator currently points to.
       *
       * The result is undefined if this iterator currently points outside the
       * useful range of the underlying container or sequence.
       *
       * @result a new reference to the element this iterator currently points to (if any)
       */

nsISupports getElement ();

pub unsafe fn StepForward(&self) -> nsresult

/**
       * Advance this iterator to the next position in the underlying container or sequence.
       */

void stepForward ();

pub unsafe fn IsEqualTo(
&self,
anotherIterator: *const nsISupports,
_retval: *mut bool
) -> nsresult

/**
       * Test if |anotherIterator| points to the same position in the underlying container or sequence.
       *
       * The result is undefined if |anotherIterator| was not created by or for the same underlying container or sequence.
       *
       * @param anotherIterator another iterator to compare against, created by or for the same underlying container or sequence
       * @result true if |anotherIterator| points to the same position in the underlying container or sequence
       */

boolean isEqualTo (in nsISupports anotherIterator);

pub unsafe fn Clone(&self, _retval: *mut *const nsISupports) -> nsresult

/**
       * Create a new iterator pointing to the same position in the underlying container or sequence to which this iterator currently points.
       * The returned iterator is suitable for use in a subsequent call to |isEqualTo()| against this iterator.
       *
       * @result a new iterator pointing at the same position in the same underlying container or sequence as this iterator
       */

nsISupports clone ();