Struct xpcom::interfaces::nsIInputStreamChannel

#[repr(C)]

pub struct nsIInputStreamChannel { /* fields omitted */ }

interface nsIInputStreamChannel : nsISupports

/**
 * nsIInputStreamChannel
 *
 * This interface provides methods to initialize an input stream channel.
 * The input stream channel serves as a data pump for an input stream.
 */

Methods

impl nsIInputStreamChannel

pub fn coerce<T: nsIInputStreamChannelCoerce>(&self) -> &T

Cast this nsIInputStreamChannel to one of its base interfaces.

impl nsIInputStreamChannel

pub unsafe fn SetURI(&self, aURI: *const nsIURI) -> nsresult

/**
     * Sets the URI for this channel.  This must be called before the
     * channel is opened, and it may only be called once.
     */

void setURI (in nsIURI aURI);

pub unsafe fn GetContentStream(
&self,
aContentStream: *mut *const nsIInputStream
) -> nsresult

/**
     * Get/set the content stream
     *
     * This stream contains the data that will be pushed to the channel's
     * stream listener.  If the stream is non-blocking and supports the
     * nsIAsyncInputStream interface, then the stream will be read directly.
     * Otherwise, the stream will be read on a background thread.
     *
     * This attribute must be set before the channel is opened, and it may
     * only be set once.
     *
     * @throws NS_ERROR_IN_PROGRESS if the setter is called after the channel
     * has been opened.
     */

attribute nsIInputStream contentStream;

pub unsafe fn SetContentStream(
&self,
aContentStream: *const nsIInputStream
) -> nsresult

/**
     * Get/set the content stream
     *
     * This stream contains the data that will be pushed to the channel's
     * stream listener.  If the stream is non-blocking and supports the
     * nsIAsyncInputStream interface, then the stream will be read directly.
     * Otherwise, the stream will be read on a background thread.
     *
     * This attribute must be set before the channel is opened, and it may
     * only be set once.
     *
     * @throws NS_ERROR_IN_PROGRESS if the setter is called after the channel
     * has been opened.
     */

attribute nsIInputStream contentStream;

pub unsafe fn GetSrcdocData(&self, aSrcdocData: &mut nsAString) -> nsresult

/**
     * Get/set the srcdoc data string.  When the input stream channel is
     * created to load a srcdoc iframe, this is set to hold the value of the
     * srcdoc attribute.
     *
     * This should be the same value used to create contentStream, but this is
     * not checked.
     *
     * Changing the value of this attribute will not otherwise affect the
     * functionality of the channel or input stream.
     */

attribute AString srcdocData;

pub unsafe fn SetSrcdocData(&self, aSrcdocData: &nsAString) -> nsresult

/**
     * Get/set the srcdoc data string.  When the input stream channel is
     * created to load a srcdoc iframe, this is set to hold the value of the
     * srcdoc attribute.
     *
     * This should be the same value used to create contentStream, but this is
     * not checked.
     *
     * Changing the value of this attribute will not otherwise affect the
     * functionality of the channel or input stream.
     */

attribute AString srcdocData;

pub unsafe fn GetIsSrcdocChannel(&self, aIsSrcdocChannel: *mut bool) -> nsresult

/**
     * Returns true if srcdocData has been set within the channel.
     */

readonly attribute boolean isSrcdocChannel;

pub unsafe fn GetBaseURI(&self, aBaseURI: *mut *const nsIURI) -> nsresult

/**
     * The base URI to be used for the channel.  Used when the base URI cannot
     * be inferred by other means, for example when this is a srcdoc channel.
     */

attribute nsIURI baseURI;

pub unsafe fn SetBaseURI(&self, aBaseURI: *const nsIURI) -> nsresult

/**
     * The base URI to be used for the channel.  Used when the base URI cannot
     * be inferred by other means, for example when this is a srcdoc channel.
     */

attribute nsIURI baseURI;