Struct xpcom::interfaces::nsILoadContext

#[repr(C)]

pub struct nsILoadContext { /* fields omitted */ }

interface nsILoadContext : nsISupports

/**
 * An nsILoadContext represents the context of a load.  This interface
 * can be queried for various information about where the load is
 * happening.
 */

Methods

impl nsILoadContext

pub fn coerce<T: nsILoadContextCoerce>(&self) -> &T

Cast this nsILoadContext to one of its base interfaces.

impl nsILoadContext

pub unsafe fn GetAssociatedWindow(
&self,
aAssociatedWindow: *mut *const mozIDOMWindowProxy
) -> nsresult

/**
   * associatedWindow is the window with which the load is associated, if any.
   * Note that the load may be triggered by a document which is different from
   * the document in associatedWindow, and in fact the source of the load need
   * not be same-origin with the document in associatedWindow.  This attribute
   * may be null if there is no associated window.
   */

readonly attribute mozIDOMWindowProxy associatedWindow;

pub unsafe fn GetTopWindow(
&self,
aTopWindow: *mut *const mozIDOMWindowProxy
) -> nsresult

/**
   * topWindow is the top window which is of same type as associatedWindow.
   * This is equivalent to associatedWindow.top, but is provided here as a
   * convenience.  All the same caveats as associatedWindow of apply, of
   * course.  This attribute may be null if there is no associated window.
   */

readonly attribute mozIDOMWindowProxy topWindow;

pub unsafe fn GetTopFrameElement(
&self,
aTopFrameElement: *mut *const nsIDOMElement
) -> nsresult

/**
   * topFrameElement is the <iframe>, <frame>, or <browser> element which
   * contains the topWindow with which the load is associated.
   *
   * Note that we may have a topFrameElement even when we don't have an
   * associatedWindow, if the topFrameElement's content lives out of process.
   * topFrameElement is available in single-process and multiprocess contexts.
   * Note that topFrameElement may be in chrome even when the nsILoadContext is
   * associated with content.
   */

readonly attribute nsIDOMElement topFrameElement;

pub unsafe fn GetNestedFrameId(&self, aNestedFrameId: *mut uint64_t) -> nsresult

/**
   * If this LoadContext corresponds to a nested remote iframe, we don't have
   * access to the topFrameElement.  Instead, we must use this id to send
   * messages. A return value of 0 signifies that this load context is not for
   * a nested frame.
   */

readonly attribute unsigned long long nestedFrameId;

pub unsafe fn GetIsContent(&self, aIsContent: *mut bool) -> nsresult

/**
   * True if the load context is content (as opposed to chrome).  This is
   * determined based on the type of window the load is performed in, NOT based
   * on any URIs that might be around.
   */

readonly attribute boolean isContent;

pub unsafe fn GetUsePrivateBrowsing(
&self,
aUsePrivateBrowsing: *mut bool
) -> nsresult

attribute boolean usePrivateBrowsing;

pub unsafe fn SetUsePrivateBrowsing(
&self,
aUsePrivateBrowsing: bool
) -> nsresult

attribute boolean usePrivateBrowsing;

pub unsafe fn GetUseRemoteTabs(&self, aUseRemoteTabs: *mut bool) -> nsresult

/**
   * Attribute that determines if remote (out-of-process) tabs should be used.
   */

readonly attribute boolean useRemoteTabs;

pub unsafe fn GetUseTrackingProtection(
&self,
aUseTrackingProtection: *mut bool
) -> nsresult

attribute boolean useTrackingProtection;

pub unsafe fn SetUseTrackingProtection(
&self,
aUseTrackingProtection: bool
) -> nsresult

attribute boolean useTrackingProtection;

pub unsafe fn SetPrivateBrowsing(&self, aInPrivateBrowsing: bool) -> nsresult

/**
   * Set the private browsing state of the load context, meant to be used internally.
   */

[noscript] void SetPrivateBrowsing (in boolean aInPrivateBrowsing);

pub unsafe fn SetRemoteTabs(&self, aUseRemoteTabs: bool) -> nsresult

/**
   * Set the remote tabs state of the load context, meant to be used internally.
   */

[noscript] void SetRemoteTabs (in boolean aUseRemoteTabs);

pub unsafe fn GetIsInIsolatedMozBrowserElement(
&self,
aIsInIsolatedMozBrowserElement: *mut bool
) -> nsresult

/**
   * Returns true iff the load is occurring inside an isolated mozbrowser
   * element. <xul:browser> is not considered to be a mozbrowser element.
   * <iframe mozbrowser noisolation> does not count as isolated since
   * isolation is disabled.  Isolation can only be disabled if the
   * containing document is chrome.
   */

readonly attribute boolean isInIsolatedMozBrowserElement;