Struct xpcom::interfaces::nsILoadGroupChild

#[repr(C)]

pub struct nsILoadGroupChild { /* fields omitted */ }

interface nsILoadGroupChild : nsISupports

/**
 * nsILoadGroupChild provides a hierarchy of load groups so that the
 * root load group can be used to conceptually tie a series of loading
 * operations into a logical whole while still leaving them separate
 * for the purposes of cancellation and status events.
 */

Methods

impl nsILoadGroupChild

pub fn coerce<T: nsILoadGroupChildCoerce>(&self) -> &T

Cast this nsILoadGroupChild to one of its base interfaces.

impl nsILoadGroupChild

pub unsafe fn GetParentLoadGroup(
&self,
aParentLoadGroup: *mut *const nsILoadGroup
) -> nsresult

/**
     * The parent of this load group. It is stored with
     * a nsIWeakReference/nsWeakPtr so there is no requirement for the
     * parentLoadGroup to out live the child, nor will the child keep a
     * reference count on the parent.
     */

attribute nsILoadGroup parentLoadGroup;

pub unsafe fn SetParentLoadGroup(
&self,
aParentLoadGroup: *const nsILoadGroup
) -> nsresult

/**
     * The parent of this load group. It is stored with
     * a nsIWeakReference/nsWeakPtr so there is no requirement for the
     * parentLoadGroup to out live the child, nor will the child keep a
     * reference count on the parent.
     */

attribute nsILoadGroup parentLoadGroup;

pub unsafe fn GetChildLoadGroup(
&self,
aChildLoadGroup: *mut *const nsILoadGroup
) -> nsresult

/**
     * The nsILoadGroup associated with this nsILoadGroupChild
     */

readonly attribute nsILoadGroup childLoadGroup;

pub unsafe fn GetRootLoadGroup(
&self,
aRootLoadGroup: *mut *const nsILoadGroup
) -> nsresult

/**
     * The rootLoadGroup is the recursive parent of this
     * load group where parent is defined as parentlLoadGroup if set
     * or childLoadGroup.loadGroup as a backup. (i.e. parentLoadGroup takes
     * precedence.) The nsILoadGroup child is the root if neither parent
     * nor loadgroup attribute is specified.
     */

readonly attribute nsILoadGroup rootLoadGroup;