Struct xpcom::interfaces::nsILocalHandlerApp

#[repr(C)]

pub struct nsILocalHandlerApp { /* fields omitted */ }

interface nsILocalHandlerApp : nsIHandlerApp

/**
 * nsILocalHandlerApp is a local OS-level executable
 */

Methods

impl nsILocalHandlerApp

pub fn coerce<T: nsILocalHandlerAppCoerce>(&self) -> &T

Cast this nsILocalHandlerApp to one of its base interfaces.

impl nsILocalHandlerApp

pub unsafe fn GetExecutable(&self, aExecutable: *mut *const nsIFile) -> nsresult

/**
     * Pointer to the executable file used to handle content
     */

attribute nsIFile executable;

pub unsafe fn SetExecutable(&self, aExecutable: *const nsIFile) -> nsresult

/**
     * Pointer to the executable file used to handle content
     */

attribute nsIFile executable;

pub unsafe fn GetParameterCount(
&self,
aParameterCount: *mut uint32_t
) -> nsresult

/**
     * Returns the current number of command line parameters.
     */

readonly attribute unsigned long parameterCount;

pub unsafe fn ClearParameters(&self) -> nsresult

/**
     * Clears the current list of command line parameters.
     */

void clearParameters ();

pub unsafe fn AppendParameter(&self, param: &nsAString) -> nsresult

/**
     * Appends a command line parameter to the command line
     * parameter list.
     *
     * @param param the parameter to add.
     */

void appendParameter (in AString param);

pub unsafe fn GetParameter(
&self,
parameterIndex: uint32_t,
_retval: &mut nsAString
) -> nsresult

/**
     * Retrieves a specific command line parameter.
     *
     * @param param the index of the parameter to return.
     *
     * @return the parameter string.
     *
     * @throw NS_ERROR_INVALID_ARG if the index is out of range.
     */

AString getParameter (in unsigned long parameterIndex);

pub unsafe fn ParameterExists(
&self,
param: &nsAString,
_retval: *mut bool
) -> nsresult

/**
     * Checks to see if a parameter exists in the command line
     * parameter list.
     *
     * @param param the parameter to search for.
     *
     * @return TRUE if the parameter exists in the current list.
     */

boolean parameterExists (in AString param);