Struct xpcom::interfaces::nsILoginInfo

#[repr(C)]

pub struct nsILoginInfo { /* fields omitted */ }

interface nsILoginInfo : nsISupports

/**
 * An object containing information for a login stored by the
 * password manager.
 */

Methods

impl nsILoginInfo

pub fn coerce<T: nsILoginInfoCoerce>(&self) -> &T

Cast this nsILoginInfo to one of its base interfaces.

impl nsILoginInfo

pub unsafe fn GetHostname(&self, aHostname: &mut nsAString) -> nsresult

/**
   * The hostname the login applies to.
   *
   * The hostname should be formatted as an URL. For example,
   * "https://site.com", "http://site.com:1234", "ftp://ftp.site.com".
   */

attribute AString hostname;

pub unsafe fn SetHostname(&self, aHostname: &nsAString) -> nsresult

/**
   * The hostname the login applies to.
   *
   * The hostname should be formatted as an URL. For example,
   * "https://site.com", "http://site.com:1234", "ftp://ftp.site.com".
   */

attribute AString hostname;

pub unsafe fn GetFormSubmitURL(
&self,
aFormSubmitURL: &mut nsAString
) -> nsresult

/**
   * The URL a form-based login was submitted to.
   *
   * For logins obtained from HTML forms, this field is the |action|
   * attribute from the |form| element, with the path removed. For
   * example "http://www.site.com". [Forms with no |action| attribute
   * default to submitting to their origin URL, so we store that.]
   *
   * For logins obtained from a HTTP or FTP protocol authentication,
   * this field is NULL.
   */

attribute AString formSubmitURL;

pub unsafe fn SetFormSubmitURL(&self, aFormSubmitURL: &nsAString) -> nsresult

/**
   * The URL a form-based login was submitted to.
   *
   * For logins obtained from HTML forms, this field is the |action|
   * attribute from the |form| element, with the path removed. For
   * example "http://www.site.com". [Forms with no |action| attribute
   * default to submitting to their origin URL, so we store that.]
   *
   * For logins obtained from a HTTP or FTP protocol authentication,
   * this field is NULL.
   */

attribute AString formSubmitURL;

pub unsafe fn GetHttpRealm(&self, aHttpRealm: &mut nsAString) -> nsresult

/**
   * The HTTP Realm a login was requested for.
   *
   * When an HTTP server sends a 401 result, the WWW-Authenticate
   * header includes a realm to identify the "protection space." See
   * RFC2617. If the response sent has a missing or blank realm, the
   * hostname is used instead.
   *
   * For logins obtained from HTML forms, this field is NULL.
   */

attribute AString httpRealm;

pub unsafe fn SetHttpRealm(&self, aHttpRealm: &nsAString) -> nsresult

/**
   * The HTTP Realm a login was requested for.
   *
   * When an HTTP server sends a 401 result, the WWW-Authenticate
   * header includes a realm to identify the "protection space." See
   * RFC2617. If the response sent has a missing or blank realm, the
   * hostname is used instead.
   *
   * For logins obtained from HTML forms, this field is NULL.
   */

attribute AString httpRealm;

pub unsafe fn GetUsername(&self, aUsername: &mut nsAString) -> nsresult

/**
   * The username for the login.
   */

attribute AString username;

pub unsafe fn SetUsername(&self, aUsername: &nsAString) -> nsresult

/**
   * The username for the login.
   */

attribute AString username;

pub unsafe fn GetUsernameField(
&self,
aUsernameField: &mut nsAString
) -> nsresult

/**
   * The |name| attribute for the username input field.
   *
   * For logins obtained from a HTTP or FTP protocol authentication,
   * this field is an empty string.
   */

attribute AString usernameField;

pub unsafe fn SetUsernameField(&self, aUsernameField: &nsAString) -> nsresult

/**
   * The |name| attribute for the username input field.
   *
   * For logins obtained from a HTTP or FTP protocol authentication,
   * this field is an empty string.
   */

attribute AString usernameField;

pub unsafe fn GetPassword(&self, aPassword: &mut nsAString) -> nsresult

/**
   * The password for the login.
   */

attribute AString password;

pub unsafe fn SetPassword(&self, aPassword: &nsAString) -> nsresult

/**
   * The password for the login.
   */

attribute AString password;

pub unsafe fn GetPasswordField(
&self,
aPasswordField: &mut nsAString
) -> nsresult

/**
   * The |name| attribute for the password input field.
   *
   * For logins obtained from a HTTP or FTP protocol authentication,
   * this field is an empty string.
   */

attribute AString passwordField;

pub unsafe fn SetPasswordField(&self, aPasswordField: &nsAString) -> nsresult

/**
   * The |name| attribute for the password input field.
   *
   * For logins obtained from a HTTP or FTP protocol authentication,
   * this field is an empty string.
   */

attribute AString passwordField;

pub unsafe fn Init(
&self,
aHostname: &nsAString,
aFormSubmitURL: &nsAString,
aHttpRealm: &nsAString,
aUsername: &nsAString,
aPassword: &nsAString,
aUsernameField: &nsAString,
aPasswordField: &nsAString
) -> nsresult

/**
   * Initialize a newly created nsLoginInfo object.
   *
   * The arguments are the fields for the new object.
   */

void init (in AString aHostname, in AString aFormSubmitURL, in AString aHttpRealm, in AString aUsername, in AString aPassword, in AString aUsernameField, in AString aPasswordField);

pub unsafe fn Equals(
&self,
aLoginInfo: *const nsILoginInfo,
_retval: *mut bool
) -> nsresult

/**
   * Test for strict equality with another nsILoginInfo object.
   *
   * @param aLoginInfo
   *        The other object to test.
   */

boolean equals (in nsILoginInfo aLoginInfo);

pub unsafe fn Matches(
&self,
aLoginInfo: *const nsILoginInfo,
ignorePassword: bool,
_retval: *mut bool
) -> nsresult

/**
   * Test for loose equivalency with another nsILoginInfo object. The
   * passwordField and usernameField values are ignored, and the password
   * values may be optionally ignored. If one login's formSubmitURL is an
   * empty string (but not null), it will be treated as a wildcard. [The
   * blank value indicates the login was stored before bug 360493 was fixed.]
   *
   * @param aLoginInfo
   *        The other object to test.
   * @param ignorePassword
   *        If true, ignore the password when checking for match.
   */

boolean matches (in nsILoginInfo aLoginInfo, in boolean ignorePassword);

pub unsafe fn Clone(&self, _retval: *mut *const nsILoginInfo) -> nsresult

/**
   * Create an identical copy of the login, duplicating all of the login's
   * nsILoginInfo and nsILoginMetaInfo properties.
   *
   * This allows code to be forwards-compatible, when additional properties
   * are added to nsILoginMetaInfo (or nsILoginInfo) in the future.
   */

nsILoginInfo clone ();