Struct xpcom::interfaces::nsILoginManagerStorage

#[repr(C)]

pub struct nsILoginManagerStorage { /* fields omitted */ }

interface nsILoginManagerStorage : nsISupports

Methods

impl nsILoginManagerStorage

pub fn coerce<T: nsILoginManagerStorageCoerce>(&self) -> &T

Cast this nsILoginManagerStorage to one of its base interfaces.

impl nsILoginManagerStorage

pub unsafe fn AddLogin(
&self,
aLogin: *const nsILoginInfo,
_retval: *mut *const nsILoginInfo
) -> nsresult

/**
   * Store a new login in the storage module.
   *
   * @param aLogin
   *        The login to be added.
   * @return a clone of the login info with the guid set (even if it was not provided).
   *
   * Default values for the login's nsILoginMetaInfo properties will be
   * created. However, if the caller specifies non-default values, they will
   * be used instead.
   */

nsILoginInfo addLogin (in nsILoginInfo aLogin);

pub unsafe fn RemoveLogin(&self, aLogin: *const nsILoginInfo) -> nsresult

/**
   * Remove a login from the storage module.
   *
   * @param aLogin
   *        The login to be removed.
   *
   * The specified login must exactly match a stored login. However, the
   * values of any nsILoginMetaInfo properties are ignored.
   */

void removeLogin (in nsILoginInfo aLogin);

pub unsafe fn ModifyLogin(
&self,
oldLogin: *const nsILoginInfo,
newLoginData: *const nsISupports
) -> nsresult

/**
   * Modify an existing login in the storage module.
   *
   * @param oldLogin
   *        The login to be modified.
   * @param newLoginData
   *        The new login values (either a nsILoginInfo or nsIProperyBag)
   *
   * If newLoginData is a nsILoginInfo, all of the old login's nsILoginInfo
   * properties are changed to the values from newLoginData (but the old
   * login's nsILoginMetaInfo properties are unmodified).
   *
   * If newLoginData is a nsIPropertyBag, only the specified properties
   * will be changed. The nsILoginMetaInfo properties of oldLogin can be
   * changed in this manner.
   *
   * If the propertybag contains an item named "timesUsedIncrement", the
   * login's timesUsed property will be incremented by the item's value.
   */

void modifyLogin (in nsILoginInfo oldLogin, in nsISupports newLoginData);

pub unsafe fn RemoveAllLogins(&self) -> nsresult

/**
   * Remove all stored logins.
   *
   * The browser sanitization feature allows the user to clear any stored
   * passwords. This interface allows that to be done without getting each
   * login first (which might require knowing the master password).
   *
   */

void removeAllLogins ();

pub unsafe fn GetAllLogins(
&self,
count: *mut uint32_t,
logins: *mut *mut *const nsILoginInfo
) -> nsresult

/**
   * Fetch all logins in the login manager. An array is always returned;
   * if there are no logins the array is empty.
   *
   * @param count
   *        The number of elements in the array. JS callers can simply use
   *        the array's .length property and omit this param.
   * @param logins
   *        An array of nsILoginInfo objects.
   *
   * NOTE: This can be called from JS as:
   *       var logins = pwmgr.getAllLogins();
   *       (|logins| is an array).
   */

void getAllLogins ([optional] out unsigned long count, [array, size_is (count), retval] out nsILoginInfo logins);

pub unsafe fn SearchLogins(
&self,
count: *mut uint32_t,
matchData: *const nsIPropertyBag,
logins: *mut *mut *const nsILoginInfo
) -> nsresult

/**
   * Search for logins in the login manager. An array is always returned;
   * if there are no logins the array is empty.
   *
   * @param count
   *        The number of elements in the array. JS callers can simply use
   *        the array's .length property, and supply an dummy object for
   *        this out param. For example: |searchLogins({}, matchData)|
   * @param matchData
   *        The data used to search. This does not follow the same
   *        requirements as findLogins for those fields. Wildcard matches are
   *        simply not specified.
   * @param logins
   *        An array of nsILoginInfo objects.
   *
   * NOTE: This can be called from JS as:
   *       var logins = pwmgr.searchLogins({}, matchData);
   *       (|logins| is an array).
   */

void searchLogins (out unsigned long count, in nsIPropertyBag matchData, [array, size_is (count), retval] out nsILoginInfo logins);

pub unsafe fn FindLogins(
&self,
count: *mut uint32_t,
aHostname: &nsAString,
aActionURL: &nsAString,
aHttpRealm: &nsAString,
logins: *mut *mut *const nsILoginInfo
) -> nsresult

/**
   * Search for logins matching the specified criteria. Called when looking
   * for logins that might be applicable to a form or authentication request.
   *
   * @param count
   *        The number of elements in the array. JS callers can simply use
   *        the array's .length property, and supply an dummy object for
   *        this out param. For example: |findLogins({}, hostname, ...)|
   * @param aHostname
   *        The hostname to restrict searches to, in URL format. For
   *        example: "http://www.site.com".
   * @param aActionURL
   *        For form logins, this argument should be the URL to which the
   *        form will be submitted. For protocol logins, specify null.
   * @param aHttpRealm
   *        For protocol logins, this argument should be the HTTP Realm
   *        for which the login applies. This is obtained from the
   *        WWW-Authenticate header. See RFC2617. For form logins,
   *        specify null.
   * @param logins
   *        An array of nsILoginInfo objects.
   *
   * NOTE: This can be called from JS as:
   *       var logins = pwmgr.findLogins({}, hostname, ...);
   *
   */

void findLogins (out unsigned long count, in AString aHostname, in AString aActionURL, in AString aHttpRealm, [array, size_is (count), retval] out nsILoginInfo logins);

pub unsafe fn CountLogins(
&self,
aHostname: &nsAString,
aActionURL: &nsAString,
aHttpRealm: &nsAString,
_retval: *mut uint32_t
) -> nsresult

/**
   * Search for logins matching the specified criteria, as with
   * findLogins(). This interface only returns the number of matching
   * logins (and not the logins themselves), which allows a caller to
   * check for logins without causing the user to be prompted for a master
   * password to decrypt the logins.
   *
   * @param aHostname
   *        The hostname to restrict searches to. Specify an empty string
   *        to match all hosts. A null value will not match any logins, and
   *        will thus always return a count of 0.
   * @param aActionURL
   *        The URL to which a form login will be submitted. To match any
   *        form login, specify an empty string. To not match any form
   *        login, specify null.
   * @param aHttpRealm
   *        The HTTP Realm for which the login applies. To match logins for
   *        any realm, specify an empty string. To not match logins for any
   *        realm, specify null.
   */

unsigned long countLogins (in AString aHostname, in AString aActionURL, in AString aHttpRealm);

pub unsafe fn GetUiBusy(&self, aUiBusy: *mut bool) -> nsresult

/**
   * True when a master password prompt is being shown.
   */

readonly attribute boolean uiBusy;

pub unsafe fn GetIsLoggedIn(&self, aIsLoggedIn: *mut bool) -> nsresult

/**
   * True when the master password has already been entered, and so a caller
   * can ask for decrypted logins without triggering a prompt.
   */

readonly attribute boolean isLoggedIn;