Struct xpcom::interfaces::nsILoginMetaInfo

#[repr(C)]

pub struct nsILoginMetaInfo { /* fields omitted */ }

interface nsILoginMetaInfo : nsISupports

/**
 * An object containing metainfo for a login stored by the login manager.
 *
 * Code using login manager can generally ignore this interface. When adding
 * logins, default value will be created. When modifying logins, these
 * properties will be unchanged unless a change is explicitly requested [by
 * using modifyLogin() with a nsIPropertyBag]. When deleting a login or
 * comparing logins, these properties are ignored.
 */

Methods

impl nsILoginMetaInfo

pub fn coerce<T: nsILoginMetaInfoCoerce>(&self) -> &T

Cast this nsILoginMetaInfo to one of its base interfaces.

impl nsILoginMetaInfo

pub unsafe fn GetGuid(&self, aGuid: &mut nsAString) -> nsresult

/**
   * The GUID to uniquely identify the login. This can be any arbitrary
   * string, but a format as created by nsIUUIDGenerator is recommended.
   * For example, "{d4e1a1f6-5ea0-40ee-bff5-da57982f21cf}"
   *
   * addLogin will generate a random value unless a value is provided.
   *
   * addLogin and modifyLogin will throw if the GUID already exists.
   */

attribute AString guid;

pub unsafe fn SetGuid(&self, aGuid: &nsAString) -> nsresult

/**
   * The GUID to uniquely identify the login. This can be any arbitrary
   * string, but a format as created by nsIUUIDGenerator is recommended.
   * For example, "{d4e1a1f6-5ea0-40ee-bff5-da57982f21cf}"
   *
   * addLogin will generate a random value unless a value is provided.
   *
   * addLogin and modifyLogin will throw if the GUID already exists.
   */

attribute AString guid;

pub unsafe fn GetTimeCreated(&self, aTimeCreated: *mut uint64_t) -> nsresult

/**
   * The time, in Unix Epoch milliseconds, when the login was first created.
   */

attribute unsigned long long timeCreated;

pub unsafe fn SetTimeCreated(&self, aTimeCreated: uint64_t) -> nsresult

/**
   * The time, in Unix Epoch milliseconds, when the login was first created.
   */

attribute unsigned long long timeCreated;

pub unsafe fn GetTimeLastUsed(&self, aTimeLastUsed: *mut uint64_t) -> nsresult

/**
   * The time, in Unix Epoch milliseconds, when the login was last submitted
   * in a form or used to begin an HTTP auth session.
   */

attribute unsigned long long timeLastUsed;

pub unsafe fn SetTimeLastUsed(&self, aTimeLastUsed: uint64_t) -> nsresult

/**
   * The time, in Unix Epoch milliseconds, when the login was last submitted
   * in a form or used to begin an HTTP auth session.
   */

attribute unsigned long long timeLastUsed;

pub unsafe fn GetTimePasswordChanged(
&self,
aTimePasswordChanged: *mut uint64_t
) -> nsresult

/**
   * The time, in Unix Epoch milliseconds, when the login was last modified.
   *
   * Contrary to what the name may suggest, this attribute takes into account
   * not only the password but also the username attribute.
   */

attribute unsigned long long timePasswordChanged;

pub unsafe fn SetTimePasswordChanged(
&self,
aTimePasswordChanged: uint64_t
) -> nsresult

/**
   * The time, in Unix Epoch milliseconds, when the login was last modified.
   *
   * Contrary to what the name may suggest, this attribute takes into account
   * not only the password but also the username attribute.
   */

attribute unsigned long long timePasswordChanged;

pub unsafe fn GetTimesUsed(&self, aTimesUsed: *mut uint32_t) -> nsresult

/**
   * The number of times the login was submitted in a form or used to begin
   * an HTTP auth session.
   */

attribute unsigned long timesUsed;

pub unsafe fn SetTimesUsed(&self, aTimesUsed: uint32_t) -> nsresult

/**
   * The number of times the login was submitted in a form or used to begin
   * an HTTP auth session.
   */

attribute unsigned long timesUsed;