Struct xpcom::interfaces::nsIMessageBroadcaster

#[repr(C)]

pub struct nsIMessageBroadcaster { /* fields omitted */ }

interface nsIMessageBroadcaster : nsIMessageListenerManager

/**
 * Message "broadcasters" don't have a single "other side" that they
 * send messages to, but rather a set of subordinate message managers.
 * For example, broadcasting a message through a window message
 * manager will broadcast the message to all frame message managers
 * within its window.
 */

Methods

impl nsIMessageBroadcaster

pub fn coerce<T: nsIMessageBroadcasterCoerce>(&self) -> &T

Cast this nsIMessageBroadcaster to one of its base interfaces.

impl nsIMessageBroadcaster

pub unsafe fn GetChildCount(&self, aChildCount: *mut uint32_t) -> nsresult

/**
   * Number of subordinate message managers.
   */

readonly attribute unsigned long childCount;

pub unsafe fn GetChildAt(
&self,
aIndex: uint32_t,
_retval: *mut *const nsIMessageListenerManager
) -> nsresult

/**
   * Return a single subordinate message manager.
   */

nsIMessageListenerManager getChildAt (in unsigned long aIndex);

pub unsafe fn ReleaseCachedProcesses(&self) -> nsresult

/**
   * Some processes are kept alive after their last tab/window are closed for testing
   * (see dom.ipc.keepProcessesAlive). This function releases those.
   */

void releaseCachedProcesses ();