Struct xpcom::interfaces::nsIMessageSender

#[repr(C)]

pub struct nsIMessageSender { /* fields omitted */ }

interface nsIMessageSender : nsIMessageListenerManager

/**
 * Message "senders" have a single "other side" to which messages are
 * sent.  For example, a child-process message manager will send
 * messages that are only delivered to its one parent-process message
 * manager.
 */

Methods

impl nsIMessageSender

pub fn coerce<T: nsIMessageSenderCoerce>(&self) -> &T

Cast this nsIMessageSender to one of its base interfaces.

impl nsIMessageSender

pub unsafe fn GetProcessMessageManager(
&self,
aProcessMessageManager: *mut *const nsIMessageSender
) -> nsresult

/**
  * For remote browsers there is always a corresponding process message
  * manager. The intention of this attribute is to link leaf level frame
  * message managers on the parent side with the corresponding process
  * message managers (if there is one). For any other cases this property
  * is null.
  */

readonly attribute nsIMessageSender processMessageManager;

pub unsafe fn GetRemoteType(&self, aRemoteType: &mut nsAString) -> nsresult

/**
  * For remote browsers, this contains the remoteType of the content child.
  * Otherwise, it is empty.
  */

readonly attribute AString remoteType;