Struct xpcom::interfaces::nsIMozIconURI

#[repr(C)]

pub struct nsIMozIconURI { /* fields omitted */ }

interface nsIMozIconURI : nsIURI

/**
   * nsIIconURI
   *
   * This interface derives from nsIURI, to provide additional information
   * about moz-icon URIs.
   *
   * What *is* a moz-icon URI you ask?  Well, it has the following syntax:
   *
   * moz-icon:[<valid-url> | //<file-with-extension> | //stock/<stock-icon>]?
   *            ['?'[<parameter-value-pairs>]]
   *
   * <valid-url> is a valid URL spec.
   *
   * <file-with-extension> is any filename with an extension, e.g. "dummy.html".
   * If the file you want an icon for isn't known to exist, you can use this
   * instead of a URL and just place a dummy file name with the extension or
   * content type you want.
   *
   * <stock-icon> is the name of a platform-dependant stock icon.
   *
   * Legal parameter value pairs are listed below:
   *
   *   Parameter:   size
   *   Values:      [<integer> | button | toolbar | toolbarsmall | menu |
   *                 dialog]
   *   Description: If integer, this is the desired size in square pixels of
   *                the icon
   *                Else, use the OS default for the specified keyword context.
   *
   *   Parameter:   state
   *   Values:      [normal | disabled]
   *   Description: The state of the icon.
   *
   *   Parameter:   contentType
   *   Values:      <mime-type>
   *   Description: The mime type we want an icon for. This is ignored by
   *                stock images.
   */

Methods

impl nsIMozIconURI

pub fn coerce<T: nsIMozIconURICoerce>(&self) -> &T

Cast this nsIMozIconURI to one of its base interfaces.

impl nsIMozIconURI

pub unsafe fn GetIconURL(&self, aIconURL: *mut *const nsIURL) -> nsresult

attribute nsIURL iconURL;

pub unsafe fn SetIconURL(&self, aIconURL: *const nsIURL) -> nsresult

attribute nsIURL iconURL;

pub unsafe fn GetImageSize(&self, aImageSize: *mut uint32_t) -> nsresult

attribute unsigned long imageSize;

pub unsafe fn SetImageSize(&self, aImageSize: uint32_t) -> nsresult

attribute unsigned long imageSize;

pub unsafe fn GetStockIcon(&self, aStockIcon: &mut nsACString) -> nsresult

readonly attribute ACString stockIcon;

pub unsafe fn GetIconSize(&self, aIconSize: &mut nsACString) -> nsresult

readonly attribute ACString iconSize;

pub unsafe fn GetIconState(&self, aIconState: &mut nsACString) -> nsresult

readonly attribute ACString iconState;

pub unsafe fn GetContentType(&self, aContentType: &mut nsACString) -> nsresult

attribute ACString contentType;

pub unsafe fn SetContentType(&self, aContentType: &nsACString) -> nsresult

attribute ACString contentType;

pub unsafe fn GetFileExtension(
&self,
aFileExtension: &mut nsACString
) -> nsresult

readonly attribute ACString fileExtension;