Struct xpcom::interfaces::nsINSSErrorsService

#[repr(C)]

pub struct nsINSSErrorsService { /* fields omitted */ }

interface nsINSSErrorsService : nsISupports

Methods

impl nsINSSErrorsService

pub fn coerce<T: nsINSSErrorsServiceCoerce>(&self) -> &T

Cast this nsINSSErrorsService to one of its base interfaces.

impl nsINSSErrorsService

pub const ERROR_CLASS_SSL_PROTOCOL: i64

ERROR_CLASS_SSL_PROTOCOL: i64 = 1

pub const ERROR_CLASS_BAD_CERT: i64

ERROR_CLASS_BAD_CERT: i64 = 2

pub const NSS_SEC_ERROR_BASE: i64

NSS_SEC_ERROR_BASE: i64 = -8192

/**
     *  The following values define the range of NSPR error codes used by NSS.
     *  NSS remains the authorative source for these numbers, as a result,
     *  the values might change in the future.
     *  The security module will perform a runtime check and assertion
     *  to ensure the values are in synch with NSS.
     */

pub const NSS_SEC_ERROR_LIMIT: i64

NSS_SEC_ERROR_LIMIT: i64 = -7192

pub const NSS_SSL_ERROR_BASE: i64

NSS_SSL_ERROR_BASE: i64 = -12288

pub const NSS_SSL_ERROR_LIMIT: i64

NSS_SSL_ERROR_LIMIT: i64 = -11288

pub const MOZILLA_PKIX_ERROR_BASE: i64

MOZILLA_PKIX_ERROR_BASE: i64 = -16384

/**
     * The error codes within each module must fit in 16 bits. We want these
     * errors to fit in the same module as the NSS errors but not overlap with
     * any of them. Converting an NSS SEC, NSS SSL, or mozilla::pkix error to
     * an NS error involves negating the value of the error and then
     * synthesizing an error in the NS_ERROR_MODULE_SECURITY module. Hence,
     * mozilla::pkix errors will start at a negative value that both doesn't
     * overlap with the current value ranges for NSS errors and that will fit
     * in 16 bits when negated.
     *
     * Keep these in sync with pkixnss.h.
     */

pub const MOZILLA_PKIX_ERROR_LIMIT: i64

MOZILLA_PKIX_ERROR_LIMIT: i64 = -15384

pub unsafe fn IsNSSErrorCode(
&self,
aNSPRCode: int32_t,
_retval: *mut bool
) -> nsresult

/**
     *  @param aNSPRCode An error code obtained using PR_GetError()
     *  @return True if it is error code defined by the NSS library
     */

[must_use] boolean isNSSErrorCode (in int32_t aNSPRCode);

pub unsafe fn GetXPCOMFromNSSError(
&self,
aNSPRCode: int32_t,
_retval: *mut nsresult
) -> nsresult

/**
     *  Function will fail if aNSPRCode is not an NSS error code.
     *  @param aNSPRCode An error code obtained using PR_GetError()
     *  @return The result of the conversion, an XPCOM error code
     */

[must_use] nsresult getXPCOMFromNSSError (in int32_t aNSPRCode);

pub unsafe fn GetErrorMessage(
&self,
aXPCOMErrorCode: nsresult,
_retval: &mut nsAString
) -> nsresult

/**
     *  Function will fail if aXPCOMErrorCode is not an NSS error code.
     *  @param aXPCOMErrorCode An error code obtained using getXPCOMFromNSSError
     *  return A localized human readable error explanation.
     */

AString getErrorMessage (in nsresult aXPCOMErrorCode);

pub unsafe fn GetErrorClass(
&self,
aXPCOMErrorCode: nsresult,
_retval: *mut uint32_t
) -> nsresult

/**
     *  Function will fail if aXPCOMErrorCode is not an NSS error code.
     *  @param aXPCOMErrorCode An error code obtained using getXPCOMFromNSSError
     *  return the error class of the code, either ERROR_CLASS_BAD_CERT
     *         or ERROR_CLASS_SSL_PROTOCOL
     */

[must_use] uint32_t getErrorClass (in nsresult aXPCOMErrorCode);