Struct xpcom::interfaces::nsINavHistoryContainerResultNode

#[repr(C)]

pub struct nsINavHistoryContainerResultNode { /* fields omitted */ }

interface nsINavHistoryContainerResultNode : nsINavHistoryResultNode

Methods

impl nsINavHistoryContainerResultNode

pub fn coerce<T: nsINavHistoryContainerResultNodeCoerce>(&self) -> &T

Cast this nsINavHistoryContainerResultNode to one of its base interfaces.

impl nsINavHistoryContainerResultNode

pub const STATE_CLOSED: i64

STATE_CLOSED: i64 = 0

pub const STATE_LOADING: i64

STATE_LOADING: i64 = 1

pub const STATE_OPENED: i64

STATE_OPENED: i64 = 2

pub unsafe fn GetContainerOpen(&self, aContainerOpen: *mut bool) -> nsresult

/**
 * Base class for container results. This includes all types of groupings.
 * Bookmark folders and places queries will be QueryResultNodes which extends
 * these items.
 */
/**
   * Set this to allow descent into the container. When closed, attempting
   * to call getChildren or childCount will result in an error. You should
   * set this to false when you are done reading.
   *
   * For HOST and DAY groupings, doing this is free since the children have
   * been precomputed. For queries and bookmark folders, being open means they
   * will keep themselves up-to-date by listening for updates and re-querying
   * as needed.
   */

attribute boolean containerOpen;

pub unsafe fn SetContainerOpen(&self, aContainerOpen: bool) -> nsresult

/**
 * Base class for container results. This includes all types of groupings.
 * Bookmark folders and places queries will be QueryResultNodes which extends
 * these items.
 */
/**
   * Set this to allow descent into the container. When closed, attempting
   * to call getChildren or childCount will result in an error. You should
   * set this to false when you are done reading.
   *
   * For HOST and DAY groupings, doing this is free since the children have
   * been precomputed. For queries and bookmark folders, being open means they
   * will keep themselves up-to-date by listening for updates and re-querying
   * as needed.
   */

attribute boolean containerOpen;

pub unsafe fn GetState(&self, aState: *mut uint16_t) -> nsresult

/**
   * Indicates whether the container is closed, loading, or opened.  Loading
   * implies that the container has been opened asynchronously and has not yet
   * fully opened.
   */

readonly attribute unsigned short state;

pub unsafe fn GetHasChildren(&self, aHasChildren: *mut bool) -> nsresult

/**
   * This indicates whether this node "may" have children, and can be used
   * when the container is open or closed. When the container is closed, it
   * will give you an exact answer if the node can easily be populated (for
   * example, a bookmark folder). If not (for example, a complex history query),
   * it will return true. When the container is open, it will always be
   * accurate. It is intended to be used to see if we should draw the "+" next
   * to a tree item.
   */

readonly attribute boolean hasChildren;

pub unsafe fn GetChildCount(&self, aChildCount: *mut uint32_t) -> nsresult

/**
   * This gives you the children of the nodes. It is preferrable to use this
   * interface over the array one, since it avoids creating an nsIArray object
   * and the interface is already the correct type.
   *
   * @throws NS_ERROR_NOT_AVAILABLE if containerOpen is false.
   */

readonly attribute unsigned long childCount;

pub unsafe fn GetChild(
&self,
aIndex: uint32_t,
_retval: *mut *const nsINavHistoryResultNode
) -> nsresult

nsINavHistoryResultNode getChild (in unsigned long aIndex);

pub unsafe fn GetChildIndex(
&self,
aNode: *const nsINavHistoryResultNode,
_retval: *mut uint32_t
) -> nsresult

/**
   * Get the index of a direct child in this container.
   *
   * @param aNode
   *        a result node.
   *
   * @return aNode's index in this container.
   * @throws NS_ERROR_NOT_AVAILABLE if containerOpen is false.
   * @throws NS_ERROR_INVALID_ARG if aNode isn't a direct child of this
   * container.
   */

unsigned long getChildIndex (in nsINavHistoryResultNode aNode);