Struct xpcom::interfaces::nsINavHistoryQueryOptions

#[repr(C)]

pub struct nsINavHistoryQueryOptions { /* fields omitted */ }

interface nsINavHistoryQueryOptions : nsISupports

Methods

impl nsINavHistoryQueryOptions

pub fn coerce<T: nsINavHistoryQueryOptionsCoerce>(&self) -> &T

Cast this nsINavHistoryQueryOptions to one of its base interfaces.

impl nsINavHistoryQueryOptions

pub const SORT_BY_NONE: i64

SORT_BY_NONE: i64 = 0

/**
 * This object represents the global options for executing a query.
 */
/**
   * You can ask for the results to be pre-sorted. Since the DB has indices
   * of many items, it can produce sorted results almost for free. These should
   * be self-explanatory.
   *
   * Note: re-sorting is slower, as is sorting by title or when you have a
   * host name.
   *
   * For bookmark items, SORT_BY_NONE means sort by the natural bookmark order.
   */

pub const SORT_BY_TITLE_ASCENDING: i64

SORT_BY_TITLE_ASCENDING: i64 = 1

pub const SORT_BY_TITLE_DESCENDING: i64

SORT_BY_TITLE_DESCENDING: i64 = 2

pub const SORT_BY_DATE_ASCENDING: i64

SORT_BY_DATE_ASCENDING: i64 = 3

pub const SORT_BY_DATE_DESCENDING: i64

SORT_BY_DATE_DESCENDING: i64 = 4

pub const SORT_BY_URI_ASCENDING: i64

SORT_BY_URI_ASCENDING: i64 = 5

pub const SORT_BY_URI_DESCENDING: i64

SORT_BY_URI_DESCENDING: i64 = 6

pub const SORT_BY_VISITCOUNT_ASCENDING: i64

SORT_BY_VISITCOUNT_ASCENDING: i64 = 7

pub const SORT_BY_VISITCOUNT_DESCENDING: i64

SORT_BY_VISITCOUNT_DESCENDING: i64 = 8

pub const SORT_BY_KEYWORD_ASCENDING: i64

SORT_BY_KEYWORD_ASCENDING: i64 = 9

pub const SORT_BY_KEYWORD_DESCENDING: i64

SORT_BY_KEYWORD_DESCENDING: i64 = 10

pub const SORT_BY_DATEADDED_ASCENDING: i64

SORT_BY_DATEADDED_ASCENDING: i64 = 11

pub const SORT_BY_DATEADDED_DESCENDING: i64

SORT_BY_DATEADDED_DESCENDING: i64 = 12

pub const SORT_BY_LASTMODIFIED_ASCENDING: i64

SORT_BY_LASTMODIFIED_ASCENDING: i64 = 13

pub const SORT_BY_LASTMODIFIED_DESCENDING: i64

SORT_BY_LASTMODIFIED_DESCENDING: i64 = 14

pub const SORT_BY_TAGS_ASCENDING: i64

SORT_BY_TAGS_ASCENDING: i64 = 17

pub const SORT_BY_TAGS_DESCENDING: i64

SORT_BY_TAGS_DESCENDING: i64 = 18

pub const SORT_BY_ANNOTATION_ASCENDING: i64

SORT_BY_ANNOTATION_ASCENDING: i64 = 19

pub const SORT_BY_ANNOTATION_DESCENDING: i64

SORT_BY_ANNOTATION_DESCENDING: i64 = 20

pub const SORT_BY_FRECENCY_ASCENDING: i64

SORT_BY_FRECENCY_ASCENDING: i64 = 21

pub const SORT_BY_FRECENCY_DESCENDING: i64

SORT_BY_FRECENCY_DESCENDING: i64 = 22

pub const RESULTS_AS_URI: i64

RESULTS_AS_URI: i64 = 0

/**
   * "URI" results, one for each URI visited in the range. Individual result
   * nodes will be of type "URI".
   */

pub const RESULTS_AS_VISIT: i64

RESULTS_AS_VISIT: i64 = 1

/**
   * "Visit" results, with one for each time a page was visited (this will
   * often give you multiple results for one URI). Individual result nodes will
   * have type "Visit"
   *
   * @note This result type is only supported by QUERY_TYPE_HISTORY.
   */

pub const RESULTS_AS_DATE_QUERY: i64

RESULTS_AS_DATE_QUERY: i64 = 3

/**
   * This returns query nodes for each predefined date range where we
   * had visits. The node contains information how to load its content:
   * - visits for the given date range will be loaded.
   *
   * @note This result type is only supported by QUERY_TYPE_HISTORY.
   */

pub const RESULTS_AS_SITE_QUERY: i64

RESULTS_AS_SITE_QUERY: i64 = 4

/**
   * This returns nsINavHistoryQueryResultNode nodes for each site where we
   * have visits. The node contains information how to load its content:
   * - last visit for each url in the given host will be loaded.
   *
   * @note This result type is only supported by QUERY_TYPE_HISTORY.
   */

pub const RESULTS_AS_DATE_SITE_QUERY: i64

RESULTS_AS_DATE_SITE_QUERY: i64 = 5

/**
   * This returns nsINavHistoryQueryResultNode nodes for each day where we
   * have visits. The node contains information how to load its content:
   * - list of hosts visited in the given period will be loaded.
   *
   * @note This result type is only supported by QUERY_TYPE_HISTORY.
   */

pub const RESULTS_AS_TAG_QUERY: i64

RESULTS_AS_TAG_QUERY: i64 = 6

/**
   * This returns nsINavHistoryQueryResultNode nodes for each tag.
   * The node contains information how to load its content:
   * - list of bookmarks with the given tag will be loaded.
   *
   * @note Setting this resultType will force queryType to QUERY_TYPE_BOOKMARKS.
   */

pub const RESULTS_AS_TAG_CONTENTS: i64

RESULTS_AS_TAG_CONTENTS: i64 = 7

/**
   * This is a container with an URI result type that contains the last
   * modified bookmarks for the given tag.
   * Tag folder id must be defined in the query.
   *
   * @note Setting this resultType will force queryType to QUERY_TYPE_BOOKMARKS.
   */

pub const QUERY_TYPE_HISTORY: i64

QUERY_TYPE_HISTORY: i64 = 0

pub const QUERY_TYPE_BOOKMARKS: i64

QUERY_TYPE_BOOKMARKS: i64 = 1

pub const QUERY_TYPE_UNIFIED: i64

QUERY_TYPE_UNIFIED: i64 = 2

pub unsafe fn GetSortingMode(&self, aSortingMode: *mut uint16_t) -> nsresult

/**
   * The sorting mode to be used for this query.
   * mode is one of SORT_BY_*
   */

attribute unsigned short sortingMode;

pub unsafe fn SetSortingMode(&self, aSortingMode: uint16_t) -> nsresult

/**
   * The sorting mode to be used for this query.
   * mode is one of SORT_BY_*
   */

attribute unsigned short sortingMode;

pub unsafe fn GetSortingAnnotation(
&self,
aSortingAnnotation: &mut nsACString
) -> nsresult

/**
   * The annotation to use in SORT_BY_ANNOTATION_* sorting modes.
   */

attribute AUTF8String sortingAnnotation;

pub unsafe fn SetSortingAnnotation(
&self,
aSortingAnnotation: &nsACString
) -> nsresult

/**
   * The annotation to use in SORT_BY_ANNOTATION_* sorting modes.
   */

attribute AUTF8String sortingAnnotation;

pub unsafe fn GetResultType(&self, aResultType: *mut uint16_t) -> nsresult

/**
   * Sets the result type. One of RESULT_TYPE_* which includes how URIs are
   * represented.
   */

attribute unsigned short resultType;

pub unsafe fn SetResultType(&self, aResultType: uint16_t) -> nsresult

/**
   * Sets the result type. One of RESULT_TYPE_* which includes how URIs are
   * represented.
   */

attribute unsigned short resultType;

pub unsafe fn GetExcludeItems(&self, aExcludeItems: *mut bool) -> nsresult

/**
   * This option excludes all URIs and separators from a bookmarks query.
   * This would be used if you just wanted a list of bookmark folders and
   * queries (such as the left pane of the places page).
   * Defaults to false.
   */

attribute boolean excludeItems;

pub unsafe fn SetExcludeItems(&self, aExcludeItems: bool) -> nsresult

/**
   * This option excludes all URIs and separators from a bookmarks query.
   * This would be used if you just wanted a list of bookmark folders and
   * queries (such as the left pane of the places page).
   * Defaults to false.
   */

attribute boolean excludeItems;

pub unsafe fn GetExcludeQueries(&self, aExcludeQueries: *mut bool) -> nsresult

/**
   * Set to true to exclude queries ("place:" URIs) from the query results.
   * Simple folder queries (bookmark folder symlinks) will still be included.
   * Defaults to false.
   */

attribute boolean excludeQueries;

pub unsafe fn SetExcludeQueries(&self, aExcludeQueries: bool) -> nsresult

/**
   * Set to true to exclude queries ("place:" URIs) from the query results.
   * Simple folder queries (bookmark folder symlinks) will still be included.
   * Defaults to false.
   */

attribute boolean excludeQueries;

pub unsafe fn GetExcludeReadOnlyFolders(
&self,
aExcludeReadOnlyFolders: *mut bool
) -> nsresult

/**
   * DO NOT USE THIS API. IT'LL BE REMOVED IN BUG 1072833.
   *
   * Set to true to exclude live bookmarks from the query results.
   */

attribute boolean excludeReadOnlyFolders;

pub unsafe fn SetExcludeReadOnlyFolders(
&self,
aExcludeReadOnlyFolders: bool
) -> nsresult

/**
   * DO NOT USE THIS API. IT'LL BE REMOVED IN BUG 1072833.
   *
   * Set to true to exclude live bookmarks from the query results.
   */

attribute boolean excludeReadOnlyFolders;

pub unsafe fn GetExpandQueries(&self, aExpandQueries: *mut bool) -> nsresult

/**
   * When set, allows items with "place:" URIs to appear as containers,
   * with the container's contents filled in from the stored query.
   * If not set, these will appear as normal items. Doesn't do anything if
   * excludeQueries is set. Defaults to false.
   *
   * Note that this has no effect on folder links, which are place: URIs
   * returned by nsINavBookmarkService.GetFolderURI. These are always expanded
   * and will appear as bookmark folders.
   */

attribute boolean expandQueries;

pub unsafe fn SetExpandQueries(&self, aExpandQueries: bool) -> nsresult

/**
   * When set, allows items with "place:" URIs to appear as containers,
   * with the container's contents filled in from the stored query.
   * If not set, these will appear as normal items. Doesn't do anything if
   * excludeQueries is set. Defaults to false.
   *
   * Note that this has no effect on folder links, which are place: URIs
   * returned by nsINavBookmarkService.GetFolderURI. These are always expanded
   * and will appear as bookmark folders.
   */

attribute boolean expandQueries;

pub unsafe fn GetIncludeHidden(&self, aIncludeHidden: *mut bool) -> nsresult

/**
   * Some pages in history are marked "hidden" and thus don't appear by default
   * in queries.  These include automatic framed visits and redirects.  Setting
   * this attribute will return all pages, even hidden ones.  Does nothing for
   * bookmark queries. Defaults to false.
   */

attribute boolean includeHidden;

pub unsafe fn SetIncludeHidden(&self, aIncludeHidden: bool) -> nsresult

/**
   * Some pages in history are marked "hidden" and thus don't appear by default
   * in queries.  These include automatic framed visits and redirects.  Setting
   * this attribute will return all pages, even hidden ones.  Does nothing for
   * bookmark queries. Defaults to false.
   */

attribute boolean includeHidden;

pub unsafe fn GetMaxResults(&self, aMaxResults: *mut uint32_t) -> nsresult

/**
   * This is the maximum number of results that you want. The query is exeucted,
   * the results are sorted, and then the top 'maxResults' results are taken
   * and returned. Set to 0 (the default) to get all results.
   *
   * THIS DOES NOT WORK IN CONJUNCTION WITH SORTING BY TITLE. This is because
   * sorting by title requires us to sort after using locale-sensetive sorting
   * (as opposed to letting the database do it for us).
   *
   * Instead, we get the result ordered by date, pick the maxResult most recent
   * ones, and THEN sort by title.
   */

attribute unsigned long maxResults;

pub unsafe fn SetMaxResults(&self, aMaxResults: uint32_t) -> nsresult

/**
   * This is the maximum number of results that you want. The query is exeucted,
   * the results are sorted, and then the top 'maxResults' results are taken
   * and returned. Set to 0 (the default) to get all results.
   *
   * THIS DOES NOT WORK IN CONJUNCTION WITH SORTING BY TITLE. This is because
   * sorting by title requires us to sort after using locale-sensetive sorting
   * (as opposed to letting the database do it for us).
   *
   * Instead, we get the result ordered by date, pick the maxResult most recent
   * ones, and THEN sort by title.
   */

attribute unsigned long maxResults;

pub unsafe fn GetQueryType(&self, aQueryType: *mut uint16_t) -> nsresult

/**
   * The type of search to use when querying the DB; This attribute is only
   * honored by query nodes. It is silently ignored for simple folder queries.
   */

attribute unsigned short queryType;

pub unsafe fn SetQueryType(&self, aQueryType: uint16_t) -> nsresult

/**
   * The type of search to use when querying the DB; This attribute is only
   * honored by query nodes. It is silently ignored for simple folder queries.
   */

attribute unsigned short queryType;

pub unsafe fn GetAsyncEnabled(&self, aAsyncEnabled: *mut bool) -> nsresult

/**
   * When this is true, the root container node generated by these options and
   * its descendant containers will be opened asynchronously if they support it.
   * This is false by default.
   *
   * @note Currently only bookmark folder containers support being opened
   *       asynchronously.
   */

attribute boolean asyncEnabled;

pub unsafe fn SetAsyncEnabled(&self, aAsyncEnabled: bool) -> nsresult

/**
   * When this is true, the root container node generated by these options and
   * its descendant containers will be opened asynchronously if they support it.
   * This is false by default.
   *
   * @note Currently only bookmark folder containers support being opened
   *       asynchronously.
   */

attribute boolean asyncEnabled;

pub unsafe fn Clone(
&self,
_retval: *mut *const nsINavHistoryQueryOptions
) -> nsresult

/**
   * Creates a new options item with the same parameters of this one.
   */

nsINavHistoryQueryOptions clone ();