Struct xpcom::interfaces::nsINavHistoryQueryResultNode

#[repr(C)]

pub struct nsINavHistoryQueryResultNode { /* fields omitted */ }

interface nsINavHistoryQueryResultNode : nsINavHistoryContainerResultNode

Methods

impl nsINavHistoryQueryResultNode

pub fn coerce<T: nsINavHistoryQueryResultNodeCoerce>(&self) -> &T

Cast this nsINavHistoryQueryResultNode to one of its base interfaces.

impl nsINavHistoryQueryResultNode

pub unsafe fn GetQueries(
&self,
queryCount: *mut uint32_t,
queries: *mut *mut *const nsINavHistoryQuery
) -> nsresult

/**
 * Used for places queries and as a base for bookmark folders.
 *
 * Note that if you request places to *not* be expanded in the options that
 * generated this node, this item will report it has no children and never try
 * to populate itself.
 */
/**
   * Get the queries which build this node's children.
   * Only valid for RESULT_TYPE_QUERY nodes.
   */

void getQueries ([optional] out unsigned long queryCount, [array, size_is (queryCount), retval] out nsINavHistoryQuery queries);

pub unsafe fn GetQueryOptions(
&self,
aQueryOptions: *mut *const nsINavHistoryQueryOptions
) -> nsresult

/**
   * Get the options which group this node's children.
   * Only valid for RESULT_TYPE_QUERY nodes.
   */

readonly attribute nsINavHistoryQueryOptions queryOptions;

pub unsafe fn GetFolderItemId(&self, aFolderItemId: *mut int64_t) -> nsresult

/**
   * For both simple folder queries and folder shortcut queries, this is set to
   * the concrete itemId of the folder (i.e. for folder shortcuts it's the
   * target folder id).  Otherwise, this is set to -1.
   */

readonly attribute long long folderItemId;

pub unsafe fn GetTargetFolderGuid(
&self,
aTargetFolderGuid: &mut nsACString
) -> nsresult

/**
   * For both simple folder queries and folder shortcut queries, this is set to
   * the concrete guid of the folder (i.e. for folder shortcuts it's the target
   * folder guid). Otherwise, this is set to an empty string.
   */

readonly attribute ACString targetFolderGuid;