Struct xpcom::interfaces::nsIObjectInputStream

#[repr(C)]

pub struct nsIObjectInputStream { /* fields omitted */ }

interface nsIObjectInputStream : nsIBinaryInputStream

/**
 * @see nsIObjectOutputStream
 * @see nsIBinaryInputStream
 */

Methods

impl nsIObjectInputStream

pub fn coerce<T: nsIObjectInputStreamCoerce>(&self) -> &T

Cast this nsIObjectInputStream to one of its base interfaces.

impl nsIObjectInputStream

pub unsafe fn ReadObject(
&self,
aIsStrongRef: bool,
_retval: *mut *const nsISupports
) -> nsresult

/**
     * Read an object from this stream to satisfy a strong or weak reference
     * to one of its interfaces.  If the interface was not along the primary
     * inheritance chain ending in the "root" or XPCOM-identity nsISupports,
     * readObject will QueryInterface from the deserialized object root to the
     * correct interface, which was specified when the object was serialized.
     *
     * @see nsIObjectOutputStream
     */

nsISupports readObject (in boolean aIsStrongRef);

pub unsafe fn ReadID(&self, aID: *mut nsID) -> nsresult

[notxpcom] nsresult readID (out nsID aID);

pub unsafe fn GetBuffer(
&self,
aLength: uint32_t,
aAlignMask: uint32_t
) -> *mut u8

/**
     * Optimized deserialization support -- see nsIStreamBufferAccess.idl.
     */

[notxpcom] charPtr getBuffer (in uint32_t aLength, in uint32_t aAlignMask);

pub unsafe fn PutBuffer(&self, aBuffer: *mut u8, aLength: uint32_t) -> c_void

[notxpcom] void putBuffer (in charPtr aBuffer, in uint32_t aLength);