Struct xpcom::interfaces::nsIObjectLoadingContent

#[repr(C)]

pub struct nsIObjectLoadingContent { /* fields omitted */ }

interface nsIObjectLoadingContent : nsISupports

/**
 * This interface represents a content node that loads objects.
 *
 * Please make sure to update the MozObjectLoadingContent WebIDL
 * interface to mirror this interface when changing it.
 */

Methods

impl nsIObjectLoadingContent

pub fn coerce<T: nsIObjectLoadingContentCoerce>(&self) -> &T

Cast this nsIObjectLoadingContent to one of its base interfaces.

impl nsIObjectLoadingContent

pub const TYPE_LOADING: i64

TYPE_LOADING: i64 = 0

/**
   * See notes in nsObjectLoadingContent.h
   */

pub const TYPE_IMAGE: i64

TYPE_IMAGE: i64 = 1

pub const TYPE_PLUGIN: i64

TYPE_PLUGIN: i64 = 2

pub const TYPE_FAKE_PLUGIN: i64

TYPE_FAKE_PLUGIN: i64 = 3

pub const TYPE_DOCUMENT: i64

TYPE_DOCUMENT: i64 = 4

pub const TYPE_NULL: i64

TYPE_NULL: i64 = 5

pub const PLUGIN_ACTIVE: i64

PLUGIN_ACTIVE: i64 = 255

pub const PLUGIN_UNSUPPORTED: i64

PLUGIN_UNSUPPORTED: i64 = 0

pub const PLUGIN_ALTERNATE: i64

PLUGIN_ALTERNATE: i64 = 1

pub const PLUGIN_DISABLED: i64

PLUGIN_DISABLED: i64 = 2

pub const PLUGIN_BLOCKLISTED: i64

PLUGIN_BLOCKLISTED: i64 = 3

pub const PLUGIN_OUTDATED: i64

PLUGIN_OUTDATED: i64 = 4

pub const PLUGIN_CRASHED: i64

PLUGIN_CRASHED: i64 = 5

pub const PLUGIN_SUPPRESSED: i64

PLUGIN_SUPPRESSED: i64 = 6

pub const PLUGIN_USER_DISABLED: i64

PLUGIN_USER_DISABLED: i64 = 7

pub const PLUGIN_CLICK_TO_PLAY: i64

PLUGIN_CLICK_TO_PLAY: i64 = 8

pub const PLUGIN_VULNERABLE_UPDATABLE: i64

PLUGIN_VULNERABLE_UPDATABLE: i64 = 9

pub const PLUGIN_VULNERABLE_NO_UPDATE: i64

PLUGIN_VULNERABLE_NO_UPDATE: i64 = 10

pub const PLUGIN_CLICK_TO_PLAY_QUIET: i64

PLUGIN_CLICK_TO_PLAY_QUIET: i64 = 11

pub const PLUGIN_PERMISSION_PROMPT_ACTION_QUIET: i64

PLUGIN_PERMISSION_PROMPT_ACTION_QUIET: i64 = 8

pub unsafe fn GetActualType(&self, aActualType: &mut nsACString) -> nsresult

/**
   * The actual mime type (the one we got back from the network
   * request) for the element.
   */

readonly attribute ACString actualType;

pub unsafe fn GetDisplayedType(&self, aDisplayedType: *mut uint32_t) -> nsresult

/**
   * Gets the type of the content that's currently loaded. See
   * the constants above for the list of possible values.
   */

readonly attribute unsigned long displayedType;

pub unsafe fn GetContentTypeForMIMEType(
&self,
aMimeType: &nsACString,
_retval: *mut uint32_t
) -> nsresult

/**
   * Gets the content type that corresponds to the give MIME type.  See the
   * constants above for the list of possible values.  If nothing else fits,
   * TYPE_NULL will be returned.
   */

unsigned long getContentTypeForMIMEType (in AUTF8String aMimeType);

pub unsafe fn GetBaseURI(&self, aBaseURI: *mut *const nsIURI) -> nsresult

/**
   * Returns the base URI of the object as seen by plugins. This differs from
   * the normal codebase in that it takes <param> tags and plugin-specific
   * quirks into account.
   */

[noscript] readonly attribute nsIURI baseURI;

pub unsafe fn PluginDestroyed(&self) -> nsresult

[noscript] void pluginDestroyed ();

pub unsafe fn PluginCrashed(
&self,
pluginTag: *const nsIPluginTag,
pluginDumpID: &nsAString,
browserDumpID: &nsAString,
submittedCrashReport: bool
) -> nsresult

[noscript] void pluginCrashed (in nsIPluginTag pluginTag, in AString pluginDumpID, in AString browserDumpID, in boolean submittedCrashReport);

pub unsafe fn Reload(&self, aClearActivation: bool) -> nsresult

/**
   * Forces a re-evaluation and reload of the tag, optionally invalidating its
   * click-to-play state.  This can be used when the MIME type that provides a
   * type has changed, for instance, to force the tag to re-evalulate the
   * handler to use.
   */

void reload (in boolean aClearActivation);

pub unsafe fn GetActivated(&self, aActivated: *mut bool) -> nsresult

/**
   * This attribute will return true if the current content type has been
   * activated, either explicitly or by passing checks that would have it be
   * click-to-play.
   */

readonly attribute boolean activated;

pub unsafe fn StopPluginInstance(&self) -> nsresult

[noscript] void stopPluginInstance ();

pub unsafe fn SyncStartPluginInstance(&self) -> nsresult

[noscript] void syncStartPluginInstance ();

pub unsafe fn AsyncStartPluginInstance(&self) -> nsresult

[noscript] void asyncStartPluginInstance ();

pub unsafe fn InitializeFromChannel(
&self,
request: *const nsIRequest
) -> nsresult

/**
   * Puts the tag in the "waiting on a channel" state and adopts this
   * channel. This does not override the normal logic of examining attributes
   * and the channel type, so the load may cancel this channel if it decides not
   * to use one.
   *
   * This assumes:
   *  - This tag has not begun loading yet
   *  - This channel has not yet hit OnStartRequest
   *  - The caller will continue to pass channel events to us as a listener
   */

[noscript] void initializeFromChannel (in nsIRequest request);

pub unsafe fn GetSrcURI(&self, aSrcURI: *mut *const nsIURI) -> nsresult

/**
   * The URL of the data/src loaded in the object. This may be null (i.e.
   * an <embed> with no src).
   */

readonly attribute nsIURI srcURI;

pub unsafe fn SkipFakePlugins(&self) -> nsresult

/**
   * Disable the use of fake plugins and reload the tag if necessary.
   */

void skipFakePlugins ();