Struct xpcom::interfaces::nsIObjectOutputStream

#[repr(C)]

pub struct nsIObjectOutputStream { /* fields omitted */ }

interface nsIObjectOutputStream : nsIBinaryOutputStream

/**
 * @See nsIObjectInputStream
 * @See nsIBinaryOutputStream
 */

Methods

impl nsIObjectOutputStream

pub fn coerce<T: nsIObjectOutputStreamCoerce>(&self) -> &T

Cast this nsIObjectOutputStream to one of its base interfaces.

impl nsIObjectOutputStream

pub unsafe fn WriteObject(
&self,
aObject: *const nsISupports,
aIsStrongRef: bool
) -> nsresult

/**
     * Write the object whose "root" or XPCOM-identity nsISupports is aObject.
     * The cause for writing this object is a strong or weak reference, so the
     * aIsStrongRef argument must tell which kind of pointer is being followed
     * here during serialization.
     *
     * If the object has only one strong reference in the serialization and no
     * weak refs, use writeSingleRefObject.  This is a valuable optimization:
     * it saves space in the stream, and cycles on both ends of the process.
     *
     * If the reference being serialized is a pointer to an interface not on
     * the primary inheritance chain ending in the root nsISupports, you must
     * call writeCompoundObject instead of this method.
     */

void writeObject (in nsISupports aObject, in boolean aIsStrongRef);

pub unsafe fn WriteSingleRefObject(
&self,
aObject: *const nsISupports
) -> nsresult

/**
     * Write an object referenced singly and strongly via its root nsISupports
     * or a subclass of its root nsISupports.  There must not be other refs to
     * aObject in memory, or in the serialization.
     */

void writeSingleRefObject (in nsISupports aObject);

pub unsafe fn WriteCompoundObject(
&self,
aObject: *const nsISupports,
aIID: &nsIID,
aIsStrongRef: bool
) -> nsresult

/**
     * Write the object referenced by an interface pointer at aObject that
     * inherits from a non-primary nsISupports, i.e., a reference to one of
     * the multiply inherited interfaces derived from an nsISupports other
     * than the root or XPCOM-identity nsISupports; or a reference to an
     * inner object in the case of true XPCOM aggregation.  aIID identifies
     * this interface.
     */

void writeCompoundObject (in nsISupports aObject, in nsIIDRef aIID, in boolean aIsStrongRef);

pub unsafe fn WriteID(&self, aID: &nsID) -> nsresult

void writeID (in nsIDRef aID);

pub unsafe fn GetBuffer(
&self,
aLength: uint32_t,
aAlignMask: uint32_t
) -> *mut u8

/**
     * Optimized serialization support -- see nsIStreamBufferAccess.idl.
     */

[notxpcom] charPtr getBuffer (in uint32_t aLength, in uint32_t aAlignMask);

pub unsafe fn PutBuffer(&self, aBuffer: *mut u8, aLength: uint32_t) -> c_void

[notxpcom] void putBuffer (in charPtr aBuffer, in uint32_t aLength);