Struct xpcom::interfaces::nsIOfflineCacheUpdateObserver

#[repr(C)]

pub struct nsIOfflineCacheUpdateObserver { /* fields omitted */ }

interface nsIOfflineCacheUpdateObserver : nsISupports

Methods

impl nsIOfflineCacheUpdateObserver

pub fn coerce<T: nsIOfflineCacheUpdateObserverCoerce>(&self) -> &T

Cast this nsIOfflineCacheUpdateObserver to one of its base interfaces.

impl nsIOfflineCacheUpdateObserver

pub const STATE_ERROR: i64

STATE_ERROR: i64 = 1

pub const STATE_CHECKING: i64

STATE_CHECKING: i64 = 2

pub const STATE_NOUPDATE: i64

STATE_NOUPDATE: i64 = 3

pub const STATE_OBSOLETE: i64

STATE_OBSOLETE: i64 = 4

pub const STATE_DOWNLOADING: i64

STATE_DOWNLOADING: i64 = 5

pub const STATE_ITEMSTARTED: i64

STATE_ITEMSTARTED: i64 = 6

pub const STATE_ITEMCOMPLETED: i64

STATE_ITEMCOMPLETED: i64 = 7

pub const STATE_ITEMPROGRESS: i64

STATE_ITEMPROGRESS: i64 = 8

pub const STATE_FINISHED: i64

STATE_FINISHED: i64 = 10

pub unsafe fn UpdateStateChanged(
&self,
aUpdate: *const nsIOfflineCacheUpdate,
state: uint32_t
) -> nsresult

/**
   * aUpdate has changed its state.
   *
   * @param aUpdate
   *        The nsIOfflineCacheUpdate being processed.
   * @param event
   *        See enumeration above
   */

void updateStateChanged (in nsIOfflineCacheUpdate aUpdate, in uint32_t state);

pub unsafe fn ApplicationCacheAvailable(
&self,
applicationCache: *const nsIApplicationCache
) -> nsresult

/**
   * Informs the observer about an application being available to associate.
   *
   * @param applicationCache
   *        The application cache instance that has been created or found by the
   *        update to associate with
   */

void applicationCacheAvailable (in nsIApplicationCache applicationCache);