Struct xpcom::interfaces::nsIParentRedirectingChannel

#[repr(C)]

pub struct nsIParentRedirectingChannel { /* fields omitted */ }

interface nsIParentRedirectingChannel : nsIParentChannel

/**
 * Implemented by chrome side of IPC protocols that support redirect responses.
 */

Methods

impl nsIParentRedirectingChannel

pub fn coerce<T: nsIParentRedirectingChannelCoerce>(&self) -> &T

Cast this nsIParentRedirectingChannel to one of its base interfaces.

impl nsIParentRedirectingChannel

pub unsafe fn StartRedirect(
&self,
newChannelId: uint32_t,
newChannel: *const nsIChannel,
redirectFlags: uint32_t,
callback: *const nsIAsyncVerifyRedirectCallback
) -> nsresult

/**
   * Called when the channel got a response that redirects it to a different
   * URI.  The implementation is responsible for calling the redirect observers
   * on the child process and provide the decision result to the callback.
   *
   * @param newChannelId
   *    id of the redirect channel obtained from nsIRedirectChannelRegistrar.
   * @param newURI
   *    the URI we redirect to
   * @param callback
   *    redirect result callback, usage is compatible with how
   *    nsIChannelEventSink defines it
   */

void startRedirect (in uint32_t newChannelId, in nsIChannel newChannel, in uint32_t redirectFlags, in nsIAsyncVerifyRedirectCallback callback);

pub unsafe fn ContinueVerification(
&self,
callback: *const nsIAsyncVerifyRedirectReadyCallback
) -> nsresult

/**
   * Called to new channel when the original channel got Redirect2Verify
   * response from child. Callback will be invoked when the new channel
   * finishes the preparation for Redirect2Verify and can be called immediately.
   *
   * @param callback
   *    redirect ready callback, will be called when redirect verification
   *    procedure can proceed.
   */

void continueVerification (in nsIAsyncVerifyRedirectReadyCallback callback);

pub unsafe fn CompleteRedirect(&self, succeeded: bool) -> nsresult

/**
   * Called after we are done with redirecting process and we know if to
   * redirect or not.  Forward the redirect result to the child process.  From
   * that moment the nsIParentChannel implementation expects it will be
   * forwarded all notifications from the 'real' channel.
   *
   * Primarilly used by HttpChannelParentListener::OnRedirectResult and kept
   * as mActiveChannel and mRedirectChannel in that class.
   */

void completeRedirect (in boolean succeeded);