Struct xpcom::interfaces::nsIParentalControlsService

#[repr(C)]

pub struct nsIParentalControlsService { /* fields omitted */ }

interface nsIParentalControlsService : nsISupports

Methods

impl nsIParentalControlsService

pub fn coerce<T: nsIParentalControlsServiceCoerce>(&self) -> &T

Cast this nsIParentalControlsService to one of its base interfaces.

impl nsIParentalControlsService

pub const DOWNLOAD: i64

DOWNLOAD: i64 = 1

/**
   * Action types that can be blocked for users.
   */

pub const INSTALL_EXTENSION: i64

INSTALL_EXTENSION: i64 = 2

pub const INSTALL_APP: i64

INSTALL_APP: i64 = 3

pub const BROWSE: i64

BROWSE: i64 = 4

pub const SHARE: i64

SHARE: i64 = 5

pub const BOOKMARK: i64

BOOKMARK: i64 = 6

pub const ADD_CONTACT: i64

ADD_CONTACT: i64 = 7

pub const SET_IMAGE: i64

SET_IMAGE: i64 = 8

pub const MODIFY_ACCOUNTS: i64

MODIFY_ACCOUNTS: i64 = 9

pub const REMOTE_DEBUGGING: i64

REMOTE_DEBUGGING: i64 = 10

pub const IMPORT_SETTINGS: i64

IMPORT_SETTINGS: i64 = 11

pub const PRIVATE_BROWSING: i64

PRIVATE_BROWSING: i64 = 12

pub const DATA_CHOICES: i64

DATA_CHOICES: i64 = 13

pub const CLEAR_HISTORY: i64

CLEAR_HISTORY: i64 = 14

pub const MASTER_PASSWORD: i64

MASTER_PASSWORD: i64 = 15

pub const GUEST_BROWSING: i64

GUEST_BROWSING: i64 = 16

pub const ADVANCED_SETTINGS: i64

ADVANCED_SETTINGS: i64 = 17

pub const CAMERA_MICROPHONE: i64

CAMERA_MICROPHONE: i64 = 18

pub const BLOCK_LIST: i64

BLOCK_LIST: i64 = 19

pub const TELEMETRY: i64

TELEMETRY: i64 = 20

pub const HEALTH_REPORT: i64

HEALTH_REPORT: i64 = 21

pub const DEFAULT_THEME: i64

DEFAULT_THEME: i64 = 22

pub const ePCLog_URIVisit: i64

ePCLog_URIVisit: i64 = 1

/**
   * Log entry types. Additional types can be defined and implemented
   * as needed. Other possible event types might include email events,
   * media related events, and IM events.
   */

pub const ePCLog_FileDownload: i64

ePCLog_FileDownload: i64 = 2

pub unsafe fn GetParentalControlsEnabled(
&self,
aParentalControlsEnabled: *mut bool
) -> nsresult

/**
   * @returns true if the current user account has parental controls
   * restrictions enabled.
   */

readonly attribute boolean parentalControlsEnabled;

pub unsafe fn GetBlockFileDownloadsEnabled(
&self,
aBlockFileDownloadsEnabled: *mut bool
) -> nsresult

/**
   * @returns true if the current user account parental controls
   * restrictions include the blocking of all file downloads.
   */

readonly attribute boolean blockFileDownloadsEnabled;

pub unsafe fn IsAllowed(
&self,
aAction: int16_t,
aUri: *const nsIURI,
_retval: *mut bool
) -> nsresult

/**
   * Check if the user can do the prescibed action for this uri.
   *
   * @param aAction             Action being performed
   * @param aUri                The uri requesting this action
   * @param aWindow             The window generating this event.
   */

boolean isAllowed (in short aAction, [optional] in nsIURI aUri);

pub unsafe fn RequestURIOverride(
&self,
aTarget: *const nsIURI,
aWindowContext: *const nsIInterfaceRequestor,
_retval: *mut bool
) -> nsresult

/**
   * Request that blocked URI(s) be allowed through parental
   * control filters. Returns true if the URI was successfully
   * overriden. Note, may block while native UI is shown.
   *
   * @param aTarget(s)          URI to be overridden. In the case of
   *                            multiple URI, the first URI in the array
   *                            should be the root URI of the site.
   * @param window              Window that generates the event.
   */

boolean requestURIOverride (in nsIURI aTarget, [optional] in nsIInterfaceRequestor aWindowContext);

pub unsafe fn RequestURIOverrides(
&self,
aTargets: *const nsIArray,
aWindowContext: *const nsIInterfaceRequestor,
_retval: *mut bool
) -> nsresult

boolean requestURIOverrides (in nsIArray aTargets, [optional] in nsIInterfaceRequestor aWindowContext);

pub unsafe fn GetLoggingEnabled(&self, aLoggingEnabled: *mut bool) -> nsresult

/**
   * @returns true if the current user account has parental controls
   * logging enabled. If true, applications should log relevent events
   * using 'log'.
   */

readonly attribute boolean loggingEnabled;

pub unsafe fn Log(
&self,
aEntryType: int16_t,
aFlag: bool,
aSource: *const nsIURI,
aTarget: *const nsIFile
) -> nsresult

/**
   * Log an application specific parental controls
   * event.
   *
   * @param aEntryType       Constant defining the type of event.
   * @param aFlag            A flag indicating if the subject content
   *                         was blocked.
   * @param aSource          The URI source of the subject content.
   * @param aTarget          The location the content was saved to if
   *                         no blocking occurred.
   */

void log (in short aEntryType, in boolean aFlag, in nsIURI aSource, [optional] in nsIFile aTarget);