Struct xpcom::interfaces::nsIPartialSHistory

#[repr(C)]

pub struct nsIPartialSHistory { /* fields omitted */ }

interface nsIPartialSHistory : nsISupports

/**
 * nsIPartialSHistory represents a part of nsIGroupedSHistory. It associates to
 * a "partial" nsISHistory in either local or remote process.
 */

Methods

impl nsIPartialSHistory

pub fn coerce<T: nsIPartialSHistoryCoerce>(&self) -> &T

Cast this nsIPartialSHistory to one of its base interfaces.

impl nsIPartialSHistory

pub const STATE_INACTIVE: i64

STATE_INACTIVE: i64 = 0

pub const STATE_ACTIVE: i64

STATE_ACTIVE: i64 = 1

pub const STATE_PRERENDER: i64

STATE_PRERENDER: i64 = 2

pub unsafe fn GetCount(&self) -> uint32_t

[infallible] readonly attribute unsigned long count;

pub unsafe fn GetGlobalIndex(&self) -> int32_t

[infallible] readonly attribute long globalIndex;

pub unsafe fn GetGlobalIndexOffset(&self) -> uint32_t

[infallible] readonly attribute unsigned long globalIndexOffset;

pub unsafe fn GetOwnerFrameLoader(
&self,
aOwnerFrameLoader: *mut *const nsIFrameLoader
) -> nsresult

readonly attribute nsIFrameLoader ownerFrameLoader;

pub unsafe fn GetGroupedSHistory(
&self,
aGroupedSHistory: *mut *const nsIGroupedSHistory
) -> nsresult

readonly attribute nsIGroupedSHistory groupedSHistory;

pub unsafe fn GetActiveState(&self) -> int32_t

[infallible] attribute long activeState;

pub unsafe fn SetActiveState(&self, aActiveState: int32_t) -> nsresult

[infallible] attribute long activeState;

pub unsafe fn OnAttachGroupedSHistory(
&self,
aGroup: *const nsIGroupedSHistory,
aOffset: uint32_t
) -> nsresult

/**
   * Notify that it's been added to a grouped session history. It also implies
   * it's becoming the active partial history of the group.
   *
   * @param aGroup                 The GroupedSHistory which this partialSHistory
   *                               is joining.
   *
   * @param aOffset                The number of entries in preceding partial
   *                               session histories.
   */

void onAttachGroupedSHistory (in nsIGroupedSHistory aGroup, in unsigned long aOffset);

pub unsafe fn HandleSHistoryUpdate(
&self,
aCount: uint32_t,
aLocalIndex: uint32_t,
aTruncate: bool
) -> nsresult

/**
   * This method is used by the TabParent to notify the PartialSHistory
   * that the state of its corresponding nsISHistory in the content process
   * has been updated. It is unused in the in-process case.
   *
   * @param aCount      The number of entries in the associated session history.
   * @param aLocalIndex The local index of the currently active entry in the
   *                    associated session history
   */

void handleSHistoryUpdate (in unsigned long aCount, in unsigned long aLocalIndex, in boolean aTruncate);

pub unsafe fn OnActive(
&self,
aGlobalLength: uint32_t,
aTargetLocalIndex: uint32_t
) -> nsresult

/**
   * Notify that the partial session history has been swapped in as the active
   * session history. Only an active session history can possibly add / remove /
   * replace its history entries.
   *
   * @param aGlobalLength      The up-to-date global length.
   * @param aTargetLocalIndex  The local index to navigate to.
   */

void onActive (in unsigned long aGlobalLength, in unsigned long aTargetLocalIndex);

pub unsafe fn OnDeactive(&self) -> nsresult

/**
   * Notify that the partial session history has been swapped out and is no
   * longer active.
   */

void onDeactive ();