Struct xpcom::interfaces::nsIPaymentRequestService

#[repr(C)]

pub struct nsIPaymentRequestService { /* fields omitted */ }

interface nsIPaymentRequestService : nsISupports

/**
 *  nsPaymentRequestService is used to manage the created PaymentRequest in the
 *  chrome process. It is also the IPC agent for payment UI to communicate with
 *  merchant side.
 */

Methods

impl nsIPaymentRequestService

pub fn coerce<T: nsIPaymentRequestServiceCoerce>(&self) -> &T

Cast this nsIPaymentRequestService to one of its base interfaces.

impl nsIPaymentRequestService

pub unsafe fn GetPaymentRequestById(
&self,
aRequestId: &nsAString,
_retval: *mut *const nsIPaymentRequest
) -> nsresult

/**
   *  Get the nsIPaymentRequest through the given payment request identifier.
   *  @param aRequestId - the payment request identifier.
   *                      This is an internal id generated by Gecko.
   *  @return           - the requested payment request. null if there is no
   *                      coressponding nsIPaymentRequest for aRequestId.
   */

nsIPaymentRequest getPaymentRequestById (in AString aRequestId);

pub unsafe fn Enumerate(
&self,
_retval: *mut *const nsISimpleEnumerator
) -> nsresult

/**
   *  Get the enumerator for all managed nsIPaymentRequests.
   *  @return - an enumerator for all managed nsIPaymentRequests.
   */

nsISimpleEnumerator enumerate ();

pub unsafe fn RespondPayment(
&self,
aResponse: *const nsIPaymentActionResponse
) -> nsresult

/**
   *  Send the user's response to the merchant.
   *  @param aResponse - the user's response.
   */

void respondPayment (in nsIPaymentActionResponse aResponse);

pub unsafe fn ChangeShippingAddress(
&self,
requestId: &nsAString,
aAddress: *const nsIPaymentAddress
) -> nsresult

/**
   *  Inform the merchant the shipping addres has changed.
   *  @param requestId - the request identifier of the payment request.
   *  @param aAddress - the new payment address.
   */

void changeShippingAddress (in AString requestId, in nsIPaymentAddress aAddress);

pub unsafe fn ChangeShippingOption(
&self,
requestId: &nsAString,
option: &nsAString
) -> nsresult

/**
   *  Inform the merchant the shipping option has changed.
   *  @param requestId - the request identifier of the payment request.
   *  @param option - the shipping option ID string.
   */

void changeShippingOption (in AString requestId, in AString option);

pub unsafe fn Cleanup(&self) -> nsresult

/**
   *  Following APIs are for testing or platform code only. UI implementation
   *  should not use them.
   */
/**
   *  Clean up the all managed payment requests.
   *  This API is for testing only.
   */

void cleanup ();

pub unsafe fn SetTestingUIService(
&self,
aUIService: *const nsIPaymentUIService
) -> nsresult

/**
   *  Setup the customized nsIPaymentUIService.
   *  This API is for testing only.
   */

void setTestingUIService (in nsIPaymentUIService aUIService);

pub unsafe fn RequestPayment(
&self,
aRequest: *const nsIPaymentActionRequest
) -> nsresult

/**
   *  Request a specified action on the specified PaymentRequest.
   *  @param aRequest - the requested action.
   */

void requestPayment (in nsIPaymentActionRequest aRequest);

pub unsafe fn RemoveActionCallback(
&self,
aCallback: *const nsIPaymentActionCallback
) -> nsresult

/**
   *  This is a cleanup function to break the association between
   *  nsIPaymentRequestService and nsIPaymentActionCallback.
   *  nsIPaymentActionCallback is an interface that registered to
   *  nsIPaymentRequestService when the merchant asks to perform actions, and it
   *  will be called when user's response send back to nsIPaymentRequestService.
   *  @param aCallback - the specified nsIPaymentActionCallback.
   */

void removeActionCallback (in nsIPaymentActionCallback aCallback);