Struct xpcom::interfaces::nsIPerformanceStatsService

#[repr(C)]

pub struct nsIPerformanceStatsService { /* fields omitted */ }

interface nsIPerformanceStatsService : nsISupports

Methods

impl nsIPerformanceStatsService

pub fn coerce<T: nsIPerformanceStatsServiceCoerce>(&self) -> &T

Cast this nsIPerformanceStatsService to one of its base interfaces.

impl nsIPerformanceStatsService

pub unsafe fn GetJankAlertThreshold(
&self,
aJankAlertThreshold: *mut uint64_t
) -> nsresult

/**
   * The threshold, in microseconds, above which a performance group is
   * considered "slow" and should raise performance alerts.
   */

attribute unsigned long long jankAlertThreshold;

pub unsafe fn SetJankAlertThreshold(
&self,
aJankAlertThreshold: uint64_t
) -> nsresult

/**
   * The threshold, in microseconds, above which a performance group is
   * considered "slow" and should raise performance alerts.
   */

attribute unsigned long long jankAlertThreshold;

pub unsafe fn GetAnimationJankLevelThreshold(
&self,
aAnimationJankLevelThreshold: *mut int16_t
) -> nsresult

/**
   * If a user is seeing an animation and we spend too long executing
   * JS code while blocking refresh, this will be visible to the user.
   *
   * We assume that any jank during an animation and lasting more than
   * 2^animationJankLevelThreshold ms will be visible.
   */

attribute short animationJankLevelThreshold;

pub unsafe fn SetAnimationJankLevelThreshold(
&self,
aAnimationJankLevelThreshold: int16_t
) -> nsresult

/**
   * If a user is seeing an animation and we spend too long executing
   * JS code while blocking refresh, this will be visible to the user.
   *
   * We assume that any jank during an animation and lasting more than
   * 2^animationJankLevelThreshold ms will be visible.
   */

attribute short animationJankLevelThreshold;

pub unsafe fn GetUserInputDelayThreshold(
&self,
aUserInputDelayThreshold: *mut uint64_t
) -> nsresult

/**
   * If a user performs an input (e.g. clicking, pressing a key, but
   * *NOT* moving the mouse), and we spend too long executing JS code
   * before displaying feedback, this will be visible to the user even
   * if there is no ongoing animation.
   *
   * We assume that any jank during `userInputDelayThreshold` us after
   * the user input will be visible.
   */

attribute unsigned long long userInputDelayThreshold;

pub unsafe fn SetUserInputDelayThreshold(
&self,
aUserInputDelayThreshold: uint64_t
) -> nsresult

/**
   * If a user performs an input (e.g. clicking, pressing a key, but
   * *NOT* moving the mouse), and we spend too long executing JS code
   * before displaying feedback, this will be visible to the user even
   * if there is no ongoing animation.
   *
   * We assume that any jank during `userInputDelayThreshold` us after
   * the user input will be visible.
   */

attribute unsigned long long userInputDelayThreshold;

pub unsafe fn GetJankAlertBufferingDelay(
&self,
aJankAlertBufferingDelay: *mut uint32_t
) -> nsresult

/**
   * A buffering delay, in milliseconds, used by the service to
   * regroup performance alerts, before observers are actually
   * noticed. Higher delays let the system avoid redundant
   * notifications for the same group, and are generally better for
   * performance.
   */

attribute unsigned long jankAlertBufferingDelay;

pub unsafe fn SetJankAlertBufferingDelay(
&self,
aJankAlertBufferingDelay: uint32_t
) -> nsresult

/**
   * A buffering delay, in milliseconds, used by the service to
   * regroup performance alerts, before observers are actually
   * noticed. Higher delays let the system avoid redundant
   * notifications for the same group, and are generally better for
   * performance.
   */

attribute unsigned long jankAlertBufferingDelay;

pub unsafe fn GetObservableWindow(
&self,
windowId: uint64_t,
_retval: *mut *const nsIPerformanceObservable
) -> nsresult

/**
   * Get a nsIPerformanceObservable representing a DOM window. This
   * observable may then be used to (un)register for watching
   * performance alerts for this window.
   *
   * Note that this covers only the current process.
   *
   * Use special window id 0 to get an observable that may be used to
   * (un)register for watching performance alerts of all windows at
   * once.
   */

nsIPerformanceObservable getObservableWindow (in unsigned long long windowId);