Struct xpcom::interfaces::nsIPermissionManager

#[repr(C)]

pub struct nsIPermissionManager { /* fields omitted */ }

interface nsIPermissionManager : nsISupports

Methods

impl nsIPermissionManager

pub fn coerce<T: nsIPermissionManagerCoerce>(&self) -> &T

Cast this nsIPermissionManager to one of its base interfaces.

impl nsIPermissionManager

pub const UNKNOWN_ACTION: i64

UNKNOWN_ACTION: i64 = 0

/**
   * Predefined return values for the testPermission method and for
   * the permission param of the add method
   * NOTE: UNKNOWN_ACTION (0) is reserved to represent the
   * default permission when no entry is found for a host, and
   * should not be used by consumers to indicate otherwise.
   */

pub const ALLOW_ACTION: i64

ALLOW_ACTION: i64 = 1

pub const DENY_ACTION: i64

DENY_ACTION: i64 = 2

pub const PROMPT_ACTION: i64

PROMPT_ACTION: i64 = 3

pub const EXPIRE_NEVER: i64

EXPIRE_NEVER: i64 = 0

/**
   * Predefined expiration types for permissions.  Permissions can be permanent
   * (never expire), expire at the end of the session, or expire at a specified
   * time. Permissions that expire at the end of a session may also have a
   * specified expiration time.
   *
   * EXPIRE_POLICY is a special expiration status. It is set when the permission
   * is set by reading an enterprise policy. These permissions cannot be overridden.
   */

pub const EXPIRE_SESSION: i64

EXPIRE_SESSION: i64 = 1

pub const EXPIRE_TIME: i64

EXPIRE_TIME: i64 = 2

pub const EXPIRE_POLICY: i64

EXPIRE_POLICY: i64 = 3

pub unsafe fn Add(
&self,
uri: *const nsIURI,
type_: *const c_char,
permission: uint32_t,
expireType: uint32_t,
expireTime: int64_t
) -> nsresult

/**
   * Add permission information for a given URI and permission type. This
   * operation will cause the type string to be registered if it does not
   * currently exist. If a permission already exists for a given type, it
   * will be modified.
   *
   * @param uri         the uri to add the permission for
   * @param type        a case-sensitive ASCII string, identifying the consumer.
   *                    Consumers should choose this string to be unique, with
   *                    respect to other consumers.
   * @param permission  an integer representing the desired action (e.g. allow
   *                    or deny). The interpretation of this number is up to the
   *                    consumer, and may represent different actions for different
   *                    types. Consumers may use one of the enumerated permission
   *                    actions defined above, for convenience.
   *                    NOTE: UNKNOWN_ACTION (0) is reserved to represent the
   *                    default permission when no entry is found for a host, and
   *                    should not be used by consumers to indicate otherwise.
   * @param expiretype  a constant defining whether this permission should
   *                    never expire (EXPIRE_NEVER), expire at the end of the
   *                    session (EXPIRE_SESSION), or expire at a specified time
   *                    (EXPIRE_TIME).
   * @param expiretime  an integer representation of when this permission
   *                    should be forgotten (milliseconds since Jan 1 1970 0:00:00).
   */

void add (in nsIURI uri, in string type, in uint32_t permission, [optional] in uint32_t expireType, [optional] in int64_t expireTime);

pub unsafe fn GetAllForURI(
&self,
uri: *const nsIURI,
_retval: *mut *const nsISimpleEnumerator
) -> nsresult

/**
   * Get all custom permissions for a given URI. This will return
   * an enumerator of all permissions which are not set to default
   * and which belong to the matching prinicpal of the given URI.
   *
   * @param uri  the URI to get all permissions for
   */

nsISimpleEnumerator getAllForURI (in nsIURI uri);

pub unsafe fn AddFromPrincipal(
&self,
principal: *const nsIPrincipal,
typed: *const c_char,
permission: uint32_t,
expireType: uint32_t,
expireTime: int64_t
) -> nsresult

/**
   * Add permission information for a given principal.
   * It is internally calling the other add() method using the nsIURI from the
   * principal.
   * Passing a system principal will be a no-op because they will always be
   * granted permissions.
   */

void addFromPrincipal (in nsIPrincipal principal, in string typed, in uint32_t permission, [optional] in uint32_t expireType, [optional] in int64_t expireTime);

pub unsafe fn Remove(
&self,
uri: *const nsIURI,
type_: *const c_char
) -> nsresult

/**
   * Remove permission information for a given URI and permission type. This will
   * remove the permission for the entire host described by the uri, acting as the
   * opposite operation to the add() method.
   *
   * @param uri    the uri to remove the permission for
   * @param type   a case-sensitive ASCII string, identifying the consumer.
   *               The type must have been previously registered using the
   *               add() method.
   */

void remove (in nsIURI uri, in string type);

pub unsafe fn RemoveFromPrincipal(
&self,
principal: *const nsIPrincipal,
type_: *const c_char
) -> nsresult

/**
   * Remove permission information for a given principal.
   * This is internally calling remove() with the host from the principal's URI.
   * Passing system principal will be a no-op because we never add them to the
   * database.
   */

void removeFromPrincipal (in nsIPrincipal principal, in string type);

pub unsafe fn RemovePermission(&self, perm: *const nsIPermission) -> nsresult

/**
   * Remove the given permission from the permission manager.
   *
   * @param perm   a permission obtained from the permission manager.
   */

void removePermission (in nsIPermission perm);

pub unsafe fn RemoveAll(&self) -> nsresult

/**
   * Clear permission information for all websites.
   */

void removeAll ();

pub unsafe fn RemoveAllSince(&self, since: int64_t) -> nsresult

/**
   * Clear all permission information added since the specified time.
   */

void removeAllSince (in int64_t since);

pub unsafe fn TestPermission(
&self,
uri: *const nsIURI,
type_: *const c_char,
_retval: *mut uint32_t
) -> nsresult

/**
   * Test whether a website has permission to perform the given action.
   * This function will perform a pref lookup to permissions.default.<type>
   * if the specific permission type is part of the whitelist for that functionality.
   * @param uri     the uri to be tested
   * @param type    a case-sensitive ASCII string, identifying the consumer
   * @param return  see add(), param permission. returns UNKNOWN_ACTION when
   *                there is no stored permission for this uri and / or type.
   */

uint32_t testPermission (in nsIURI uri, in string type);

pub unsafe fn TestPermissionFromPrincipal(
&self,
principal: *const nsIPrincipal,
type_: *const c_char,
_retval: *mut uint32_t
) -> nsresult

/**
   * Test whether the principal has the permission to perform a given action.
   * System principals will always have permissions granted.
   * This function will perform a pref lookup to permissions.default.<type>
   * if the specific permission type is part of the whitelist for that functionality.
   */

uint32_t testPermissionFromPrincipal (in nsIPrincipal principal, in string type);

pub unsafe fn TestPermissionFromWindow(
&self,
window: *const mozIDOMWindow,
type_: *const c_char,
_retval: *mut uint32_t
) -> nsresult

/**
   * Test whether the principal associated with the window's document has the
   * permission to perform a given action.  System principals will always
   * have permissions granted.
   * This function will perform a pref lookup to permissions.default.<type>
   * if the specific permission type is part of the whitelist for that functionality.
   */

uint32_t testPermissionFromWindow (in mozIDOMWindow window, in string type);

pub unsafe fn TestExactPermission(
&self,
uri: *const nsIURI,
type_: *const c_char,
_retval: *mut uint32_t
) -> nsresult

/**
   * Test whether a website has permission to perform the given action.
   * This requires an exact hostname match, subdomains are not a match.
   * This function will perform a pref lookup to permissions.default.<type>
   * if the specific permission type is part of the whitelist for that functionality.
   * @param uri     the uri to be tested
   * @param type    a case-sensitive ASCII string, identifying the consumer
   * @param return  see add(), param permission. returns UNKNOWN_ACTION when
   *                there is no stored permission for this uri and / or type.
   */

uint32_t testExactPermission (in nsIURI uri, in string type);

pub unsafe fn TestExactPermissionFromPrincipal(
&self,
principal: *const nsIPrincipal,
type_: *const c_char,
_retval: *mut uint32_t
) -> nsresult

/**
   * See testExactPermission() above.
   * System principals will always have permissions granted.
   * This function will perform a pref lookup to permissions.default.<type>
   * if the specific permission type is part of the whitelist for that functionality.
   */

uint32_t testExactPermissionFromPrincipal (in nsIPrincipal principal, in string type);

pub unsafe fn TestExactPermanentPermission(
&self,
principal: *const nsIPrincipal,
type_: *const c_char,
_retval: *mut uint32_t
) -> nsresult

/**
   * Test whether a website has permission to perform the given action
   * ignoring active sessions.
   * System principals will always have permissions granted.
   * This function will perform a pref lookup to permissions.default.<type>
   * if the specific permission type is part of the whitelist for that functionality.
   *
   * @param principal the principal
   * @param type      a case-sensitive ASCII string, identifying the consumer
   * @param return    see add(), param permission. returns UNKNOWN_ACTION when
   *                  there is no stored permission for this uri and / or type.
   */

uint32_t testExactPermanentPermission (in nsIPrincipal principal, in string type);

pub unsafe fn GetPermissionObjectForURI(
&self,
uri: *const nsIURI,
type_: *const c_char,
exactHost: bool,
_retval: *mut *const nsIPermission
) -> nsresult

/**
   * Get the permission object associated with the given URI and action.
   * @param uri The URI
   * @param type      A case-sensitive ASCII string identifying the consumer
   * @param exactHost If true, only the specific host will be matched,
   *                  @see testExactPermission. If false, subdomains will
   *                  also be searched, @see testPermission.
   * @returns The matching permission object, or null if no matching object
   *          was found. No matching object is equivalent to UNKNOWN_ACTION.
   * @note Clients in general should prefer the test* methods unless they
   *       need to know the specific stored details.
   * @note This method will always return null for the system principal.
   */

nsIPermission getPermissionObjectForURI (in nsIURI uri, in string type, in boolean exactHost);

pub unsafe fn GetPermissionObject(
&self,
principal: *const nsIPrincipal,
type_: *const c_char,
exactHost: bool,
_retval: *mut *const nsIPermission
) -> nsresult

/**
   * Get the permission object associated with the given principal and action.
   * @param principal The principal
   * @param type      A case-sensitive ASCII string identifying the consumer
   * @param exactHost If true, only the specific host will be matched,
   *                  @see testExactPermission. If false, subdomains will
   *                  also be searched, @see testPermission.
   * @returns The matching permission object, or null if no matching object
   *          was found. No matching object is equivalent to UNKNOWN_ACTION.
   * @note Clients in general should prefer the test* methods unless they
   *       need to know the specific stored details.
   * @note This method will always return null for the system principal.
   */

nsIPermission getPermissionObject (in nsIPrincipal principal, in string type, in boolean exactHost);

pub unsafe fn GetEnumerator(
&self,
aEnumerator: *mut *const nsISimpleEnumerator
) -> nsresult

/**
   * Allows enumeration of all stored permissions
   * @return an nsISimpleEnumerator interface that allows access to
   *         nsIPermission objects
   */

readonly attribute nsISimpleEnumerator enumerator;

pub unsafe fn RemovePermissionsWithAttributes(
&self,
patternAsJSON: &nsAString
) -> nsresult

/**
   * Remove all permissions that will match the origin pattern.
   */

void removePermissionsWithAttributes (in DOMString patternAsJSON);

pub unsafe fn UpdateExpireTime(
&self,
principal: *const nsIPrincipal,
type_: *const c_char,
exactHost: bool,
sessionExpireTime: uint64_t,
persistentExpireTime: uint64_t
) -> nsresult

/**
   * If the current permission is set to expire, reset the expiration time. If
   * there is no permission or the current permission does not expire, this
   * method will silently return.
   *
   * @param sessionExpiretime  an integer representation of when this permission
   *                           should be forgotten (milliseconds since
   *                           Jan 1 1970 0:00:00), if it is currently
   *                           EXPIRE_SESSION.
   * @param sessionExpiretime  an integer representation of when this permission
   *                           should be forgotten (milliseconds since
   *                           Jan 1 1970 0:00:00), if it is currently
   *                           EXPIRE_TIME.
   */

void updateExpireTime (in nsIPrincipal principal, in string type, in boolean exactHost, in uint64_t sessionExpireTime, in uint64_t persistentExpireTime);

pub unsafe fn BroadcastPermissionsForPrincipalToAllContentProcesses(
&self,
aPrincipal: *const nsIPrincipal
) -> nsresult

/**
   * Broadcasts permissions for the given principal to all content processes.
   *
   * DO NOT USE THIS METHOD if you can avoid it. It was added in bug XXX to
   * handle the current temporary implementation of ServiceWorker debugging. It
   * will be removed when service worker debugging is fixed.
   *
   * @param aPrincipal The principal to broadcast permissions for.
   */

void broadcastPermissionsForPrincipalToAllContentProcesses (in nsIPrincipal aPrincipal);

pub unsafe fn WhenPermissionsAvailable(
&self,
aPrincipal: *const nsIPrincipal,
aRunnable: *const nsIRunnable
) -> nsresult

/**
   * Add a callback which should be run when all permissions are available for
   * the given nsIPrincipal. This method invokes the callback runnable
   * synchronously when the permissions are already available. Otherwise the
   * callback will be run asynchronously in SystemGroup when all permissions
   * are available in the future.
   *
   * NOTE: This method will not request the permissions be sent by the parent
   * process. This should only be used to wait for permissions which may not
   * have arrived yet in order to ensure they are present.
   *
   * @param aPrincipal The principal to wait for permissions to be available for.
   * @param aRunnable  The runnable to run when permissions are available for the
   *                   given principal.
   */

void whenPermissionsAvailable (in nsIPrincipal aPrincipal, in nsIRunnable aRunnable);

pub unsafe fn GetHasPreloadPermissions(&self) -> bool

/**
   * True if any "preload" permissions are present. This is used to avoid making
   * potentially expensive permissions checks in nsContentBlocker.
   */

[infallible] readonly attribute boolean hasPreloadPermissions;