Struct xpcom::interfaces::nsIPlaintextEditor

#[repr(C)]

pub struct nsIPlaintextEditor { /* fields omitted */ }

interface nsIPlaintextEditor : nsISupports

Methods

impl nsIPlaintextEditor

pub fn coerce<T: nsIPlaintextEditorCoerce>(&self) -> &T

Cast this nsIPlaintextEditor to one of its base interfaces.

impl nsIPlaintextEditor

pub const eEditorPlaintextMask: i64

eEditorPlaintextMask: i64 = 1

pub const eEditorSingleLineMask: i64

eEditorSingleLineMask: i64 = 2

pub const eEditorPasswordMask: i64

eEditorPasswordMask: i64 = 4

pub const eEditorReadonlyMask: i64

eEditorReadonlyMask: i64 = 8

pub const eEditorDisabledMask: i64

eEditorDisabledMask: i64 = 16

pub const eEditorFilterInputMask: i64

eEditorFilterInputMask: i64 = 32

pub const eEditorMailMask: i64

eEditorMailMask: i64 = 64

pub const eEditorEnableWrapHackMask: i64

eEditorEnableWrapHackMask: i64 = 128

pub const eEditorWidgetMask: i64

eEditorWidgetMask: i64 = 256

pub const eEditorNoCSSMask: i64

eEditorNoCSSMask: i64 = 512

pub const eEditorAllowInteraction: i64

eEditorAllowInteraction: i64 = 1024

pub const eEditorDontEchoPassword: i64

eEditorDontEchoPassword: i64 = 2048

pub const eEditorRightToLeft: i64

eEditorRightToLeft: i64 = 4096

pub const eEditorLeftToRight: i64

eEditorLeftToRight: i64 = 8192

pub const eEditorSkipSpellCheck: i64

eEditorSkipSpellCheck: i64 = 16384

pub const eNewlinesPasteIntact: i64

eNewlinesPasteIntact: i64 = 0

pub const eNewlinesPasteToFirst: i64

eNewlinesPasteToFirst: i64 = 1

pub const eNewlinesReplaceWithSpaces: i64

eNewlinesReplaceWithSpaces: i64 = 2

pub const eNewlinesStrip: i64

eNewlinesStrip: i64 = 3

pub const eNewlinesReplaceWithCommas: i64

eNewlinesReplaceWithCommas: i64 = 4

pub const eNewlinesStripSurroundingWhitespace: i64

eNewlinesStripSurroundingWhitespace: i64 = 5

pub unsafe fn GetTextLength(&self, aTextLength: *mut int32_t) -> nsresult

/**
    * The length of the contents in characters.
    * XXX change this type to 'unsigned long'
    */

readonly attribute long textLength;

pub unsafe fn GetMaxTextLength(&self, aMaxTextLength: *mut int32_t) -> nsresult

/**
    * The maximum number of characters allowed.
    *   default: -1 (unlimited).
    */

attribute long maxTextLength;

pub unsafe fn SetMaxTextLength(&self, aMaxTextLength: int32_t) -> nsresult

/**
    * The maximum number of characters allowed.
    *   default: -1 (unlimited).
    */

attribute long maxTextLength;

pub unsafe fn GetWrapWidth(&self, aWrapWidth: *mut int32_t) -> nsresult

/** Get and set the body wrap width.
    *
    * Special values:
    *    0 = wrap to window width
    *   -1 = no wrap at all
    */

attribute long wrapWidth;

pub unsafe fn SetWrapWidth(&self, aWrapWidth: int32_t) -> nsresult

/** Get and set the body wrap width.
    *
    * Special values:
    *    0 = wrap to window width
    *   -1 = no wrap at all
    */

attribute long wrapWidth;

pub unsafe fn SetWrapColumn(&self, aWrapColumn: int32_t) -> nsresult

/**
   * Similar to the setter for wrapWidth, but just sets the editor
   * internal state without actually changing the content being edited
   * to wrap at that column.  This should only be used by callers who
   * are sure that their content is already set up correctly.
   */

void setWrapColumn (in long aWrapColumn);

pub unsafe fn GetNewlineHandling(
&self,
aNewlineHandling: *mut int32_t
) -> nsresult

/** Get and set newline handling.
   *
   *  Values are the constants defined above.
   */

attribute long newlineHandling;

pub unsafe fn SetNewlineHandling(&self, aNewlineHandling: int32_t) -> nsresult

/** Get and set newline handling.
   *
   *  Values are the constants defined above.
   */

attribute long newlineHandling;

pub unsafe fn InsertText(&self, aStringToInsert: &nsAString) -> nsresult

/**
   * Inserts a string at the current location,
   * given by the selection.
   * If the selection is not collapsed, the selection is deleted
   * and the insertion takes place at the resulting collapsed selection.
   *
   * @param aString   the string to be inserted
   */

void insertText (in DOMString aStringToInsert);

pub unsafe fn SetText(&self, aString: &nsAString) -> nsresult

/**
   * Replace existed string with a string.
   * This is fast path to replace all string when using single line control.
   *
   * @ param aString   the string to be set
   */

[noscript] void setText (in DOMString aString);

pub unsafe fn InsertLineBreak(&self) -> nsresult

/**
   * Insert a line break into the content model.
   * The interpretation of a break is up to the implementation:
   * it may enter a character, split a node in the tree, etc.
   * This may be more efficient than calling InsertText with a newline.
   */

void insertLineBreak ();