Struct xpcom::interfaces::nsIPluginHost

#[repr(C)]

pub struct nsIPluginHost { /* fields omitted */ }

interface nsIPluginHost : nsISupports

Methods

impl nsIPluginHost

pub fn coerce<T: nsIPluginHostCoerce>(&self) -> &T

Cast this nsIPluginHost to one of its base interfaces.

impl nsIPluginHost

pub const FLAG_CLEAR_ALL: i64

FLAG_CLEAR_ALL: i64 = 0

pub const FLAG_CLEAR_CACHE: i64

FLAG_CLEAR_CACHE: i64 = 1

pub const EXCLUDE_NONE: i64

EXCLUDE_NONE: i64 = 0

pub const EXCLUDE_DISABLED: i64

EXCLUDE_DISABLED: i64 = 1

pub const EXCLUDE_FAKE: i64

EXCLUDE_FAKE: i64 = 2

pub unsafe fn ReloadPlugins(&self) -> nsresult

/**
   * Causes the plugins directory to be searched again for new plugin
   * libraries.
   */

void reloadPlugins ();

pub unsafe fn GetPluginTags(
&self,
aPluginCount: *mut uint32_t,
aResults: *mut *mut *const nsIPluginTag
) -> nsresult

void getPluginTags ([optional] out unsigned long aPluginCount, [array, size_is (aPluginCount), retval] out nsIPluginTag aResults);

pub unsafe fn ClearSiteData(
&self,
plugin: *const nsIPluginTag,
domain: &nsACString,
flags: uint64_t,
maxAge: int64_t,
callback: *const nsIClearSiteDataCallback
) -> nsresult

void clearSiteData (in nsIPluginTag plugin, in AUTF8String domain, in uint64_t flags, in int64_t maxAge, in nsIClearSiteDataCallback callback);

pub unsafe fn SiteHasData(
&self,
plugin: *const nsIPluginTag,
domain: &nsACString,
_retval: *mut bool
) -> nsresult

boolean siteHasData (in nsIPluginTag plugin, in AUTF8String domain);

pub unsafe fn GetPermissionStringForType(
&self,
mimeType: &nsACString,
excludeFlags: uint32_t,
_retval: &mut nsACString
) -> nsresult

/**
   * Get the "permission string" for the plugin.  This is a string that can be
   * passed to the permission manager to see whether the plugin is allowed to
   * run, for example.  This will typically be based on the plugin's "nice name"
   * and its blocklist state.
   *
   * @mimeType The MIME type we're interested in.
   * @excludeFlags Set of the EXCLUDE_* flags above, defaulting to EXCLUDE_NONE.
   */

ACString getPermissionStringForType (in AUTF8String mimeType, [optional] in uint32_t excludeFlags);

pub unsafe fn GetPermissionStringForTag(
&self,
tag: *const nsIPluginTag,
excludeFlags: uint32_t,
_retval: &mut nsACString
) -> nsresult

/**
   * Get the "permission string" for the plugin.  This is a string that can be
   * passed to the permission manager to see whether the plugin is allowed to
   * run, for example.  This will typically be based on the plugin's "nice name"
   * and its blocklist state.
   *
   * @tag The tage we're interested in
   * @excludeFlags Set of the EXCLUDE_* flags above, defaulting to EXCLUDE_NONE.
   */

ACString getPermissionStringForTag (in nsIPluginTag tag, [optional] in uint32_t excludeFlags);

pub unsafe fn GetPluginTagForType(
&self,
mimeType: &nsACString,
excludeFlags: uint32_t,
_retval: *mut *const nsIPluginTag
) -> nsresult

/**
   * Get the nsIPluginTag for this MIME type. This method works with both
   * enabled and disabled/blocklisted plugins, but an enabled plugin will
   * always be returned if available.
   *
   * A fake plugin tag, if one exists and is available, will be returned in
   * preference to NPAPI plugin tags unless excluded by the excludeFlags.
   *
   * @mimeType The MIME type we're interested in.
   * @excludeFlags Set of the EXCLUDE_* flags above, defaulting to EXCLUDE_NONE.
   *
   * @throws NS_ERROR_NOT_AVAILABLE if no plugin is available for this MIME
   *         type.
   */

nsIPluginTag getPluginTagForType (in AUTF8String mimeType, [optional] in uint32_t excludeFlags);

pub unsafe fn GetStateForType(
&self,
mimeType: &nsACString,
excludeFlags: uint32_t,
_retval: *mut uint32_t
) -> nsresult

/**
   * Get the nsIPluginTag enabled state for this MIME type.  See
   * nsIPluginTag.enabledState.
   *
   * @mimeType The MIME type we're interested in.
   * @excludeFlags Set of the EXCLUDE_* flags above, defaulting to EXCLUDE_NONE.
   */

unsigned long getStateForType (in AUTF8String mimeType, [optional] in uint32_t excludeFlags);

pub unsafe fn GetBlocklistStateForType(
&self,
aMimeType: &nsACString,
excludeFlags: uint32_t,
_retval: *mut uint32_t
) -> nsresult

/**
   * Get the blocklist state for a MIME type.  See nsIPluginTag.blocklistState.
   *
   * @mimeType The MIME type we're interested in.
   * @excludeFlags Set of the EXCLUDE_* flags above, defaulting to EXCLUDE_NONE.
   */

uint32_t getBlocklistStateForType (in AUTF8String aMimeType, [optional] in uint32_t excludeFlags);

pub unsafe fn GetFakePlugin(
&self,
mimeType: &nsACString,
_retval: *mut *const nsIFakePluginTag
) -> nsresult

/**
   * Get a reference to an existing fake plugin tag for the given MIME type, if
   * any.  Can return null.
   */

nsIFakePluginTag getFakePlugin (in AUTF8String mimeType);

pub unsafe fn UnregisterFakePlugin(&self, handlerURI: &nsACString) -> nsresult

/**
   * Unregister a fake plugin.  The argument can be the .handlerURI.spec of an
   * existing nsIFakePluginTag, or just a known handler URI string that was
   * passed in the FakePluginTagInit when registering.
   */

void unregisterFakePlugin (in AUTF8String handlerURI);