Struct xpcom::interfaces::nsIPresentationControlService

#[repr(C)]

pub struct nsIPresentationControlService { /* fields omitted */ }

interface nsIPresentationControlService : nsISupports

/**
 * Presentation control service which can be used for both presentation
 * control client and server.
 */

Methods

impl nsIPresentationControlService

pub fn coerce<T: nsIPresentationControlServiceCoerce>(&self) -> &T

Cast this nsIPresentationControlService to one of its base interfaces.

impl nsIPresentationControlService

pub unsafe fn StartServer(&self, aEncrypted: bool, aPort: uint16_t) -> nsresult

/**
   * This method initializes server socket. Caller should set listener and
   * monitor onServerReady event to get the correct server info.
   * @param   aEncrypted
   *          True for using TLS control channel.
   * @param   aPort
   *          The port of the server socket.  Pass 0 or opt-out to indicate no
   *          preference, and a port will be selected automatically.
   * @throws  NS_ERROR_FAILURE if the server socket has been inited or the
   *          server socket can not be inited.
   */

void startServer (in boolean aEncrypted, [optional] in uint16_t aPort);

pub unsafe fn Connect(
&self,
aDeviceInfo: *const nsITCPDeviceInfo,
_retval: *mut *const nsIPresentationControlChannel
) -> nsresult

/**
   * Request connection to designated remote presentation control receiver.
   * @param   aDeviceInfo
   *          The remtoe device info for establish connection.
   * @returns The control channel for this session.
   * @throws  NS_ERROR_FAILURE if the Id hasn't been inited.
   */

nsIPresentationControlChannel connect (in nsITCPDeviceInfo aDeviceInfo);

pub unsafe fn IsCompatibleServer(
&self,
aVersion: uint32_t,
_retval: *mut bool
) -> nsresult

/**
   * Check the compatibility to remote presentation control server.
   * @param  aVersion
   *         The version of remote server.
   */

boolean isCompatibleServer (in uint32_t aVersion);

pub unsafe fn Close(&self) -> nsresult

/**
   * Close server socket and call |listener.onClose(NS_OK)|
   */

void close ();

pub unsafe fn GetPort(&self, aPort: *mut uint16_t) -> nsresult

/**
   * Get the listen port of the TCP socket, valid after the server is ready.
   * 0 indicates the server socket is not ready or is closed.
   */

readonly attribute uint16_t port;

pub unsafe fn GetVersion(&self, aVersion: *mut uint32_t) -> nsresult

/**
   * The protocol version implemented by this server.
   */

readonly attribute uint32_t version;

pub unsafe fn GetId(&self, aId: &mut nsACString) -> nsresult

/**
   * The id of the TCP presentation server. |requestSession| won't
   * work until the |id| is set.
   */

attribute AUTF8String id;

pub unsafe fn SetId(&self, aId: &nsACString) -> nsresult

/**
   * The id of the TCP presentation server. |requestSession| won't
   * work until the |id| is set.
   */

attribute AUTF8String id;

pub unsafe fn GetCertFingerprint(
&self,
aCertFingerprint: &mut nsACString
) -> nsresult

/**
   * The fingerprint of the TLS server certificate.
   * Empty string indicates the server is not ready or not encrypted.
   */

attribute AUTF8String certFingerprint;

pub unsafe fn SetCertFingerprint(
&self,
aCertFingerprint: &nsACString
) -> nsresult

/**
   * The fingerprint of the TLS server certificate.
   * Empty string indicates the server is not ready or not encrypted.
   */

attribute AUTF8String certFingerprint;

pub unsafe fn GetListener(
&self,
aListener: *mut *const nsIPresentationControlServerListener
) -> nsresult

/**
   * The listener for handling events of this presentation control server.
   * Listener must be provided before invoke |startServer| and |close|.
   */

attribute nsIPresentationControlServerListener listener;

pub unsafe fn SetListener(
&self,
aListener: *const nsIPresentationControlServerListener
) -> nsresult

/**
   * The listener for handling events of this presentation control server.
   * Listener must be provided before invoke |startServer| and |close|.
   */

attribute nsIPresentationControlServerListener listener;