Struct xpcom::interfaces::nsIPrintSettings

#[repr(C)]

pub struct nsIPrintSettings { /* fields omitted */ }

interface nsIPrintSettings : nsISupports

/**
 * Simplified graphics interface for JS rendering.
 */

Methods

impl nsIPrintSettings

pub fn coerce<T: nsIPrintSettingsCoerce>(&self) -> &T

Cast this nsIPrintSettings to one of its base interfaces.

impl nsIPrintSettings

pub const kInitSaveOddEvenPages: i64

kInitSaveOddEvenPages: i64 = 1

/**
   * PrintSettings to be Saved Navigation Constants
   */

pub const kInitSaveHeaderLeft: i64

kInitSaveHeaderLeft: i64 = 2

pub const kInitSaveHeaderCenter: i64

kInitSaveHeaderCenter: i64 = 4

pub const kInitSaveHeaderRight: i64

kInitSaveHeaderRight: i64 = 8

pub const kInitSaveFooterLeft: i64

kInitSaveFooterLeft: i64 = 16

pub const kInitSaveFooterCenter: i64

kInitSaveFooterCenter: i64 = 32

pub const kInitSaveFooterRight: i64

kInitSaveFooterRight: i64 = 64

pub const kInitSaveBGColors: i64

kInitSaveBGColors: i64 = 128

pub const kInitSaveBGImages: i64

kInitSaveBGImages: i64 = 256

pub const kInitSavePaperSize: i64

kInitSavePaperSize: i64 = 512

pub const kInitSaveResolution: i64

kInitSaveResolution: i64 = 1024

pub const kInitSaveDuplex: i64

kInitSaveDuplex: i64 = 2048

pub const kInitSavePaperData: i64

kInitSavePaperData: i64 = 8192

pub const kInitSaveUnwriteableMargins: i64

kInitSaveUnwriteableMargins: i64 = 16384

pub const kInitSaveEdges: i64

kInitSaveEdges: i64 = 32768

pub const kInitSaveReversed: i64

kInitSaveReversed: i64 = 65536

pub const kInitSaveInColor: i64

kInitSaveInColor: i64 = 131072

pub const kInitSaveOrientation: i64

kInitSaveOrientation: i64 = 262144

pub const kInitSavePrinterName: i64

kInitSavePrinterName: i64 = 1048576

pub const kInitSavePrintToFile: i64

kInitSavePrintToFile: i64 = 2097152

pub const kInitSaveToFileName: i64

kInitSaveToFileName: i64 = 4194304

pub const kInitSavePageDelay: i64

kInitSavePageDelay: i64 = 8388608

pub const kInitSaveMargins: i64

kInitSaveMargins: i64 = 16777216

pub const kInitSaveNativeData: i64

kInitSaveNativeData: i64 = 33554432

pub const kInitSaveShrinkToFit: i64

kInitSaveShrinkToFit: i64 = 134217728

pub const kInitSaveScaling: i64

kInitSaveScaling: i64 = 268435456

pub const kInitSaveAll: i64

kInitSaveAll: i64 = 4294967295

pub const kPrintOddPages: i64

kPrintOddPages: i64 = 1

pub const kPrintEvenPages: i64

kPrintEvenPages: i64 = 2

pub const kEnableSelectionRB: i64

kEnableSelectionRB: i64 = 4

pub const kRangeAllPages: i64

kRangeAllPages: i64 = 0

pub const kRangeSpecifiedPageRange: i64

kRangeSpecifiedPageRange: i64 = 1

pub const kRangeSelection: i64

kRangeSelection: i64 = 2

pub const kRangeFocusFrame: i64

kRangeFocusFrame: i64 = 3

pub const kJustLeft: i64

kJustLeft: i64 = 0

pub const kJustCenter: i64

kJustCenter: i64 = 1

pub const kJustRight: i64

kJustRight: i64 = 2

pub const kUseInternalDefault: i64

kUseInternalDefault: i64 = 0

/**
   * FrameSet Default Type Constants
   */

pub const kUseSettingWhenPossible: i64

kUseSettingWhenPossible: i64 = 1

pub const kPaperSizeNativeData: i64

kPaperSizeNativeData: i64 = 0

/**
   * Page Size Type Constants
   */

pub const kPaperSizeDefined: i64

kPaperSizeDefined: i64 = 1

pub const kPaperSizeInches: i64

kPaperSizeInches: i64 = 0

/**
   * Page Size Unit Constants
   */

pub const kPaperSizeMillimeters: i64

kPaperSizeMillimeters: i64 = 1

pub const kPortraitOrientation: i64

kPortraitOrientation: i64 = 0

/**
   * Orientation Constants
   */

pub const kLandscapeOrientation: i64

kLandscapeOrientation: i64 = 1

pub const kNoFrames: i64

kNoFrames: i64 = 0

/**
   * Print Frame Constants
   */

pub const kFramesAsIs: i64

kFramesAsIs: i64 = 1

pub const kSelectedFrame: i64

kSelectedFrame: i64 = 2

pub const kEachFrameSep: i64

kEachFrameSep: i64 = 3

pub const kFrameEnableNone: i64

kFrameEnableNone: i64 = 0

/**
   * How to Enable Frame Set Printing Constants
   */

pub const kFrameEnableAll: i64

kFrameEnableAll: i64 = 1

pub const kFrameEnableAsIsAndEach: i64

kFrameEnableAsIsAndEach: i64 = 2

pub const kOutputFormatNative: i64

kOutputFormatNative: i64 = 0

/**
   * Output file format
   */

pub const kOutputFormatPS: i64

kOutputFormatPS: i64 = 1

pub const kOutputFormatPDF: i64

kOutputFormatPDF: i64 = 2

pub unsafe fn SetPrintOptions(
&self,
aType: int32_t,
aTurnOnOff: bool
) -> nsresult

/**
   * Set PrintOptions
   */

void SetPrintOptions (in int32_t aType, in boolean aTurnOnOff);

pub unsafe fn GetPrintOptions(
&self,
aType: int32_t,
_retval: *mut bool
) -> nsresult

/**
   * Get PrintOptions
   */

boolean GetPrintOptions (in int32_t aType);

pub unsafe fn GetPrintOptionsBits(&self, _retval: *mut int32_t) -> nsresult

/**
   * Get PrintOptions Bit field
   */

int32_t GetPrintOptionsBits ();

pub unsafe fn SetPrintOptionsBits(&self, bits: int32_t) -> nsresult

/**
   * Set PrintOptions Bit field
   */

void SetPrintOptionsBits (in int32_t bits);

pub unsafe fn GetEffectivePageSize(
&self,
aWidth: *mut c_double,
aHeight: *mut c_double
) -> nsresult

/**
   * Get the page size in twips, considering the
   * orientation (portrait or landscape).
   */

void GetEffectivePageSize (out double aWidth, out double aHeight);

pub unsafe fn Clone(&self, _retval: *mut *const nsIPrintSettings) -> nsresult

/**
   * Makes a new copy
   */

nsIPrintSettings clone ();

pub unsafe fn Assign(&self, aPS: *const nsIPrintSettings) -> nsresult

/**
   * Assigns the internal values from the "in" arg to the current object
   */

void assign (in nsIPrintSettings aPS);

pub unsafe fn GetPrintSession(
&self,
aPrintSession: *mut *const nsIPrintSession
) -> nsresult

/**
   * Data Members
   */

[noscript] attribute nsIPrintSession printSession;

pub unsafe fn SetPrintSession(
&self,
aPrintSession: *const nsIPrintSession
) -> nsresult

/**
   * Data Members
   */

[noscript] attribute nsIPrintSession printSession;

pub unsafe fn GetStartPageRange(
&self,
aStartPageRange: *mut int32_t
) -> nsresult

attribute long startPageRange;

pub unsafe fn SetStartPageRange(&self, aStartPageRange: int32_t) -> nsresult

attribute long startPageRange;

pub unsafe fn GetEndPageRange(&self, aEndPageRange: *mut int32_t) -> nsresult

attribute long endPageRange;

pub unsafe fn SetEndPageRange(&self, aEndPageRange: int32_t) -> nsresult

attribute long endPageRange;

pub unsafe fn GetEdgeTop(&self, aEdgeTop: *mut c_double) -> nsresult

/**
   * The edge measurements define the positioning of the headers
   * and footers on the page. They're measured as an offset from
   * the "unwriteable margin" (described below).
   */

attribute double edgeTop;

pub unsafe fn SetEdgeTop(&self, aEdgeTop: c_double) -> nsresult

/**
   * The edge measurements define the positioning of the headers
   * and footers on the page. They're measured as an offset from
   * the "unwriteable margin" (described below).
   */

attribute double edgeTop;

pub unsafe fn GetEdgeLeft(&self, aEdgeLeft: *mut c_double) -> nsresult

attribute double edgeLeft;

pub unsafe fn SetEdgeLeft(&self, aEdgeLeft: c_double) -> nsresult

attribute double edgeLeft;

pub unsafe fn GetEdgeBottom(&self, aEdgeBottom: *mut c_double) -> nsresult

attribute double edgeBottom;

pub unsafe fn SetEdgeBottom(&self, aEdgeBottom: c_double) -> nsresult

attribute double edgeBottom;

pub unsafe fn GetEdgeRight(&self, aEdgeRight: *mut c_double) -> nsresult

attribute double edgeRight;

pub unsafe fn SetEdgeRight(&self, aEdgeRight: c_double) -> nsresult

attribute double edgeRight;

pub unsafe fn GetMarginTop(&self, aMarginTop: *mut c_double) -> nsresult

/**
   * The margins define the positioning of the content on the page.
   * They're treated as an offset from the "unwriteable margin"
   * (described below).
   */

attribute double marginTop;

pub unsafe fn SetMarginTop(&self, aMarginTop: c_double) -> nsresult

/**
   * The margins define the positioning of the content on the page.
   * They're treated as an offset from the "unwriteable margin"
   * (described below).
   */

attribute double marginTop;

pub unsafe fn GetMarginLeft(&self, aMarginLeft: *mut c_double) -> nsresult

attribute double marginLeft;

pub unsafe fn SetMarginLeft(&self, aMarginLeft: c_double) -> nsresult

attribute double marginLeft;

pub unsafe fn GetMarginBottom(&self, aMarginBottom: *mut c_double) -> nsresult

attribute double marginBottom;

pub unsafe fn SetMarginBottom(&self, aMarginBottom: c_double) -> nsresult

attribute double marginBottom;

pub unsafe fn GetMarginRight(&self, aMarginRight: *mut c_double) -> nsresult

attribute double marginRight;

pub unsafe fn SetMarginRight(&self, aMarginRight: c_double) -> nsresult

attribute double marginRight;

pub unsafe fn GetUnwriteableMarginTop(
&self,
aUnwriteableMarginTop: *mut c_double
) -> nsresult

/**
   * The unwriteable margin defines the printable region of the paper, creating
   * an invisible border from which the edge and margin attributes are measured.
   */

attribute double unwriteableMarginTop;

pub unsafe fn SetUnwriteableMarginTop(
&self,
aUnwriteableMarginTop: c_double
) -> nsresult

/**
   * The unwriteable margin defines the printable region of the paper, creating
   * an invisible border from which the edge and margin attributes are measured.
   */

attribute double unwriteableMarginTop;

pub unsafe fn GetUnwriteableMarginLeft(
&self,
aUnwriteableMarginLeft: *mut c_double
) -> nsresult

attribute double unwriteableMarginLeft;

pub unsafe fn SetUnwriteableMarginLeft(
&self,
aUnwriteableMarginLeft: c_double
) -> nsresult

attribute double unwriteableMarginLeft;

pub unsafe fn GetUnwriteableMarginBottom(
&self,
aUnwriteableMarginBottom: *mut c_double
) -> nsresult

attribute double unwriteableMarginBottom;

pub unsafe fn SetUnwriteableMarginBottom(
&self,
aUnwriteableMarginBottom: c_double
) -> nsresult

attribute double unwriteableMarginBottom;

pub unsafe fn GetUnwriteableMarginRight(
&self,
aUnwriteableMarginRight: *mut c_double
) -> nsresult

attribute double unwriteableMarginRight;

pub unsafe fn SetUnwriteableMarginRight(
&self,
aUnwriteableMarginRight: c_double
) -> nsresult

attribute double unwriteableMarginRight;

pub unsafe fn GetScaling(&self, aScaling: *mut c_double) -> nsresult

attribute double scaling;

pub unsafe fn SetScaling(&self, aScaling: c_double) -> nsresult

attribute double scaling;

pub unsafe fn GetPrintBGColors(&self, aPrintBGColors: *mut bool) -> nsresult

attribute boolean printBGColors;

pub unsafe fn SetPrintBGColors(&self, aPrintBGColors: bool) -> nsresult

attribute boolean printBGColors;

pub unsafe fn GetPrintBGImages(&self, aPrintBGImages: *mut bool) -> nsresult

attribute boolean printBGImages;

pub unsafe fn SetPrintBGImages(&self, aPrintBGImages: bool) -> nsresult

attribute boolean printBGImages;

pub unsafe fn GetPrintRange(&self, aPrintRange: *mut int16_t) -> nsresult

attribute short printRange;

pub unsafe fn SetPrintRange(&self, aPrintRange: int16_t) -> nsresult

attribute short printRange;

pub unsafe fn GetTitle(&self, aTitle: &mut nsAString) -> nsresult

attribute AString title;

pub unsafe fn SetTitle(&self, aTitle: &nsAString) -> nsresult

attribute AString title;

pub unsafe fn GetDocURL(&self, aDocURL: &mut nsAString) -> nsresult

attribute AString docURL;

pub unsafe fn SetDocURL(&self, aDocURL: &nsAString) -> nsresult

attribute AString docURL;

pub unsafe fn GetHeaderStrLeft(
&self,
aHeaderStrLeft: &mut nsAString
) -> nsresult

attribute AString headerStrLeft;

pub unsafe fn SetHeaderStrLeft(&self, aHeaderStrLeft: &nsAString) -> nsresult

attribute AString headerStrLeft;

pub unsafe fn GetHeaderStrCenter(
&self,
aHeaderStrCenter: &mut nsAString
) -> nsresult

attribute AString headerStrCenter;

pub unsafe fn SetHeaderStrCenter(
&self,
aHeaderStrCenter: &nsAString
) -> nsresult

attribute AString headerStrCenter;

pub unsafe fn GetHeaderStrRight(
&self,
aHeaderStrRight: &mut nsAString
) -> nsresult

attribute AString headerStrRight;

pub unsafe fn SetHeaderStrRight(&self, aHeaderStrRight: &nsAString) -> nsresult

attribute AString headerStrRight;

pub unsafe fn GetFooterStrLeft(
&self,
aFooterStrLeft: &mut nsAString
) -> nsresult

attribute AString footerStrLeft;

pub unsafe fn SetFooterStrLeft(&self, aFooterStrLeft: &nsAString) -> nsresult

attribute AString footerStrLeft;

pub unsafe fn GetFooterStrCenter(
&self,
aFooterStrCenter: &mut nsAString
) -> nsresult

attribute AString footerStrCenter;

pub unsafe fn SetFooterStrCenter(
&self,
aFooterStrCenter: &nsAString
) -> nsresult

attribute AString footerStrCenter;

pub unsafe fn GetFooterStrRight(
&self,
aFooterStrRight: &mut nsAString
) -> nsresult

attribute AString footerStrRight;

pub unsafe fn SetFooterStrRight(&self, aFooterStrRight: &nsAString) -> nsresult

attribute AString footerStrRight;

pub unsafe fn GetHowToEnableFrameUI(
&self,
aHowToEnableFrameUI: *mut int16_t
) -> nsresult

attribute short howToEnableFrameUI;

pub unsafe fn SetHowToEnableFrameUI(
&self,
aHowToEnableFrameUI: int16_t
) -> nsresult

attribute short howToEnableFrameUI;

pub unsafe fn GetIsCancelled(&self, aIsCancelled: *mut bool) -> nsresult

attribute boolean isCancelled;

pub unsafe fn SetIsCancelled(&self, aIsCancelled: bool) -> nsresult

attribute boolean isCancelled;

pub unsafe fn GetPrintFrameTypeUsage(
&self,
aPrintFrameTypeUsage: *mut int16_t
) -> nsresult

attribute short printFrameTypeUsage;

pub unsafe fn SetPrintFrameTypeUsage(
&self,
aPrintFrameTypeUsage: int16_t
) -> nsresult

attribute short printFrameTypeUsage;

pub unsafe fn GetPrintFrameType(
&self,
aPrintFrameType: *mut int16_t
) -> nsresult

attribute short printFrameType;

pub unsafe fn SetPrintFrameType(&self, aPrintFrameType: int16_t) -> nsresult

attribute short printFrameType;

pub unsafe fn GetPrintSilent(&self, aPrintSilent: *mut bool) -> nsresult

attribute boolean printSilent;

pub unsafe fn SetPrintSilent(&self, aPrintSilent: bool) -> nsresult

attribute boolean printSilent;

pub unsafe fn GetShrinkToFit(&self, aShrinkToFit: *mut bool) -> nsresult

attribute boolean shrinkToFit;

pub unsafe fn SetShrinkToFit(&self, aShrinkToFit: bool) -> nsresult

attribute boolean shrinkToFit;

pub unsafe fn GetShowPrintProgress(
&self,
aShowPrintProgress: *mut bool
) -> nsresult

attribute boolean showPrintProgress;

pub unsafe fn SetShowPrintProgress(&self, aShowPrintProgress: bool) -> nsresult

attribute boolean showPrintProgress;

pub unsafe fn GetPaperName(&self, aPaperName: &mut nsAString) -> nsresult

attribute AString paperName;

pub unsafe fn SetPaperName(&self, aPaperName: &nsAString) -> nsresult

attribute AString paperName;

pub unsafe fn GetPaperData(&self, aPaperData: *mut int16_t) -> nsresult

attribute short paperData;

pub unsafe fn SetPaperData(&self, aPaperData: int16_t) -> nsresult

attribute short paperData;

pub unsafe fn GetPaperWidth(&self, aPaperWidth: *mut c_double) -> nsresult

attribute double paperWidth;

pub unsafe fn SetPaperWidth(&self, aPaperWidth: c_double) -> nsresult

attribute double paperWidth;

pub unsafe fn GetPaperHeight(&self, aPaperHeight: *mut c_double) -> nsresult

attribute double paperHeight;

pub unsafe fn SetPaperHeight(&self, aPaperHeight: c_double) -> nsresult

attribute double paperHeight;

pub unsafe fn GetPaperSizeUnit(&self, aPaperSizeUnit: *mut int16_t) -> nsresult

attribute short paperSizeUnit;

pub unsafe fn SetPaperSizeUnit(&self, aPaperSizeUnit: int16_t) -> nsresult

attribute short paperSizeUnit;

pub unsafe fn GetPrintReversed(&self, aPrintReversed: *mut bool) -> nsresult

attribute boolean printReversed;

pub unsafe fn SetPrintReversed(&self, aPrintReversed: bool) -> nsresult

attribute boolean printReversed;

pub unsafe fn GetPrintInColor(&self, aPrintInColor: *mut bool) -> nsresult

attribute boolean printInColor;

pub unsafe fn SetPrintInColor(&self, aPrintInColor: bool) -> nsresult

attribute boolean printInColor;

pub unsafe fn GetOrientation(&self, aOrientation: *mut int32_t) -> nsresult

attribute long orientation;

pub unsafe fn SetOrientation(&self, aOrientation: int32_t) -> nsresult

attribute long orientation;

pub unsafe fn GetNumCopies(&self, aNumCopies: *mut int32_t) -> nsresult

attribute long numCopies;

pub unsafe fn SetNumCopies(&self, aNumCopies: int32_t) -> nsresult

attribute long numCopies;

pub unsafe fn GetPrinterName(&self, aPrinterName: &mut nsAString) -> nsresult

attribute AString printerName;

pub unsafe fn SetPrinterName(&self, aPrinterName: &nsAString) -> nsresult

attribute AString printerName;

pub unsafe fn GetPrintToFile(&self, aPrintToFile: *mut bool) -> nsresult

attribute boolean printToFile;

pub unsafe fn SetPrintToFile(&self, aPrintToFile: bool) -> nsresult

attribute boolean printToFile;

pub unsafe fn GetToFileName(&self, aToFileName: &mut nsAString) -> nsresult

attribute AString toFileName;

pub unsafe fn SetToFileName(&self, aToFileName: &nsAString) -> nsresult

attribute AString toFileName;

pub unsafe fn GetOutputFormat(&self, aOutputFormat: *mut int16_t) -> nsresult

attribute short outputFormat;

pub unsafe fn SetOutputFormat(&self, aOutputFormat: int16_t) -> nsresult

attribute short outputFormat;

pub unsafe fn GetPrintPageDelay(
&self,
aPrintPageDelay: *mut int32_t
) -> nsresult

attribute long printPageDelay;

pub unsafe fn SetPrintPageDelay(&self, aPrintPageDelay: int32_t) -> nsresult

attribute long printPageDelay;

pub unsafe fn GetResolution(&self, aResolution: *mut int32_t) -> nsresult

attribute long resolution;

pub unsafe fn SetResolution(&self, aResolution: int32_t) -> nsresult

attribute long resolution;

pub unsafe fn GetDuplex(&self, aDuplex: *mut int32_t) -> nsresult

attribute long duplex;

pub unsafe fn SetDuplex(&self, aDuplex: int32_t) -> nsresult

attribute long duplex;

pub unsafe fn GetIsInitializedFromPrinter(
&self,
aIsInitializedFromPrinter: *mut bool
) -> nsresult

/**
   * This attribute tracks whether the PS has been initialized
   * from a printer specified by the "printerName" attr.
   * If a different name is set into the "printerName"
   * attribute than the one it was initialized with the PS
   * will then get intialized from that printer.
   */

attribute boolean isInitializedFromPrinter;

pub unsafe fn SetIsInitializedFromPrinter(
&self,
aIsInitializedFromPrinter: bool
) -> nsresult

/**
   * This attribute tracks whether the PS has been initialized
   * from a printer specified by the "printerName" attr.
   * If a different name is set into the "printerName"
   * attribute than the one it was initialized with the PS
   * will then get intialized from that printer.
   */

attribute boolean isInitializedFromPrinter;

pub unsafe fn GetIsInitializedFromPrefs(
&self,
aIsInitializedFromPrefs: *mut bool
) -> nsresult

/**
   * This attribute tracks whether the PS has been initialized
   * from prefs. If a different name is set into the "printerName"
   * attribute than the one it was initialized with the PS
   * will then get intialized from prefs again.
   */

attribute boolean isInitializedFromPrefs;

pub unsafe fn SetIsInitializedFromPrefs(
&self,
aIsInitializedFromPrefs: bool
) -> nsresult

/**
   * This attribute tracks whether the PS has been initialized
   * from prefs. If a different name is set into the "printerName"
   * attribute than the one it was initialized with the PS
   * will then get intialized from prefs again.
   */

attribute boolean isInitializedFromPrefs;

pub unsafe fn SetupSilentPrinting(&self) -> nsresult

/**
   * We call this function so that anything that requires a run of the event loop
   * can do so safely. The print dialog runs the event loop but in silent printing
   * that doesn't happen.
   *
   * Either this or ShowPrintDialog (but not both) MUST be called by the print engine
   * before printing, otherwise printing can fail on some platforms.
   */

[noscript] void SetupSilentPrinting ();