Struct xpcom::interfaces::nsIProcess

#[repr(C)]

pub struct nsIProcess { /* fields omitted */ }

interface nsIProcess : nsISupports

Methods

impl nsIProcess

pub fn coerce<T: nsIProcessCoerce>(&self) -> &T

Cast this nsIProcess to one of its base interfaces.

impl nsIProcess

pub unsafe fn Init(&self, executable: *const nsIFile) -> nsresult

/**
   * Initialises the process with an executable to be run. Call the run method
   * to run the executable.
   * @param executable The executable to run.
   */

void init (in nsIFile executable);

pub unsafe fn Kill(&self) -> nsresult

/**
   * Kills the running process. After exiting the process will either have
   * been killed or a failure will have been returned.
   */

void kill ();

pub unsafe fn Run(
&self,
blocking: bool,
args: *mut *const c_char,
count: uint32_t
) -> nsresult

/**
   * Executes the file this object was initialized with
   * @param blocking   Whether to wait until the process terminates before
                       returning or not.
   * @param args       An array of arguments to pass to the process in the
   *                   native character set.
   * @param count      The length of the args array.
   */

void run (in boolean blocking, [array, size_is (count)] in string args, in unsigned long count);

pub unsafe fn RunAsync(
&self,
args: *mut *const c_char,
count: uint32_t,
observer: *const nsIObserver,
holdWeak: bool
) -> nsresult

/**
   * Executes the file this object was initialized with optionally calling
   * an observer after the process has finished running.
   * @param args       An array of arguments to pass to the process in the
   *                   native character set.
   * @param count      The length of the args array.
   * @param observer   An observer to notify when the process has completed. It
   *                   will receive this process instance as the subject and
   *                   "process-finished" or "process-failed" as the topic. The
   *                   observer will be notified on the main thread.
   * @param holdWeak   Whether to use a weak reference to hold the observer.
   */

void runAsync ([array, size_is (count)] in string args, in unsigned long count, [optional] in nsIObserver observer, [optional] in boolean holdWeak);

pub unsafe fn Runw(
&self,
blocking: bool,
args: *mut *const int16_t,
count: uint32_t
) -> nsresult

/**
   * Executes the file this object was initialized with
   * @param blocking   Whether to wait until the process terminates before
                       returning or not.
   * @param args       An array of arguments to pass to the process in UTF-16
   * @param count      The length of the args array.
   */

void runw (in boolean blocking, [array, size_is (count)] in wstring args, in unsigned long count);

pub unsafe fn RunwAsync(
&self,
args: *mut *const int16_t,
count: uint32_t,
observer: *const nsIObserver,
holdWeak: bool
) -> nsresult

/**
   * Executes the file this object was initialized with optionally calling
   * an observer after the process has finished running.
   * @param args       An array of arguments to pass to the process in UTF-16
   * @param count      The length of the args array.
   * @param observer   An observer to notify when the process has completed. It
   *                   will receive this process instance as the subject and
   *                   "process-finished" or "process-failed" as the topic. The
   *                   observer will be notified on the main thread.
   * @param holdWeak   Whether to use a weak reference to hold the observer.
   */

void runwAsync ([array, size_is (count)] in wstring args, in unsigned long count, [optional] in nsIObserver observer, [optional] in boolean holdWeak);

pub unsafe fn GetStartHidden(&self, aStartHidden: *mut bool) -> nsresult

/**
   * When set to true the process will not open a new window when started and
   * will run hidden from the user. This currently affects only the Windows
   * platform.
   */

attribute boolean startHidden;

pub unsafe fn SetStartHidden(&self, aStartHidden: bool) -> nsresult

/**
   * When set to true the process will not open a new window when started and
   * will run hidden from the user. This currently affects only the Windows
   * platform.
   */

attribute boolean startHidden;

pub unsafe fn GetNoShell(&self, aNoShell: *mut bool) -> nsresult

/**
   * When set to true the process will be launched directly without using the
   * shell. This currently affects only the Windows platform.
   */

attribute boolean noShell;

pub unsafe fn SetNoShell(&self, aNoShell: bool) -> nsresult

/**
   * When set to true the process will be launched directly without using the
   * shell. This currently affects only the Windows platform.
   */

attribute boolean noShell;

pub unsafe fn GetPid(&self, aPid: *mut uint32_t) -> nsresult

/**
   * The process identifier of the currently running process. This will only
   * be available after the process has started and may not be available on
   * some platforms.
   */

readonly attribute unsigned long pid;

pub unsafe fn GetExitValue(&self, aExitValue: *mut int32_t) -> nsresult

/**
   * The exit value of the process. This is only valid after the process has
   * exited.
   */

readonly attribute long exitValue;

pub unsafe fn GetIsRunning(&self, aIsRunning: *mut bool) -> nsresult

/**
   * Returns whether the process is currently running or not.
   */

readonly attribute boolean isRunning;