Struct xpcom::interfaces::nsIProfiler

#[repr(C)]

pub struct nsIProfiler { /* fields omitted */ }

interface nsIProfiler : nsISupports

Methods

impl nsIProfiler

pub fn coerce<T: nsIProfilerCoerce>(&self) -> &T

Cast this nsIProfiler to one of its base interfaces.

impl nsIProfiler

pub unsafe fn CanProfile(&self, _retval: *mut bool) -> nsresult

boolean CanProfile ();

pub unsafe fn StartProfiler(
&self,
aEntries: uint32_t,
aInterval: c_double,
aFeatures: *mut *const c_char,
aFeatureCount: uint32_t,
aFilters: *mut *const c_char,
aFilterCount: uint32_t
) -> nsresult

void StartProfiler (in uint32_t aEntries, in double aInterval, [array, size_is (aFeatureCount)] in string aFeatures, in uint32_t aFeatureCount, [array, size_is (aFilterCount), optional] in string aFilters, [optional] in uint32_t aFilterCount);

pub unsafe fn StopProfiler(&self) -> nsresult

void StopProfiler ();

pub unsafe fn IsPaused(&self, _retval: *mut bool) -> nsresult

boolean IsPaused ();

pub unsafe fn PauseSampling(&self) -> nsresult

void PauseSampling ();

pub unsafe fn ResumeSampling(&self) -> nsresult

void ResumeSampling ();

pub unsafe fn AddMarker(&self, aMarker: *const c_char) -> nsresult

void AddMarker (in string aMarker);

pub unsafe fn GetProfile(
&self,
aSinceTime: c_double,
_retval: *mut *const c_char
) -> nsresult

string GetProfile ([optional] in double aSinceTime);

pub unsafe fn IsActive(&self, _retval: *mut bool) -> nsresult

boolean IsActive ();

pub unsafe fn GetFeatures(
&self,
aCount: *mut uint32_t,
aFeatures: *mut *mut *const c_char
) -> nsresult

/**
   * Returns an array of the features that are supported in this build.
   * Features may vary depending on platform and build flags.
   */

void GetFeatures (out uint32_t aCount, [array, size_is (aCount), retval] out string aFeatures);

pub unsafe fn GetAllFeatures(
&self,
aCount: *mut uint32_t,
aFeatures: *mut *mut *const c_char
) -> nsresult

/**
   * Returns an array of all features that are supported by the profiler.
   * The array may contain features that are not supported in this build.
   */

void GetAllFeatures (out uint32_t aCount, [array, size_is (aCount), retval] out string aFeatures);

pub unsafe fn GetStartParams(
&self,
aStartParams: *mut *const nsIProfilerStartParams
) -> nsresult

/**
   * The starting parameters that were sent to the profiler for sampling.
   * If the profiler is not currently sampling, this will return null.
   */

[noscript] readonly attribute nsIProfilerStartParams startParams;

pub unsafe fn GetBufferInfo(
&self,
aCurrentPosition: *mut uint32_t,
aTotalSize: *mut uint32_t,
aGeneration: *mut uint32_t
) -> nsresult

void GetBufferInfo (out uint32_t aCurrentPosition, out uint32_t aTotalSize, out uint32_t aGeneration);

pub unsafe fn GetElapsedTime(&self, _retval: *mut c_double) -> nsresult

/**
   * Returns the elapsed time, in milliseconds, since the profiler's epoch.
   * The epoch is guaranteed to be constant for the duration of the
   * process, but is otherwise arbitrary.
   */

double getElapsedTime ();

pub unsafe fn DumpProfileToFile(&self, aFilename: *const c_char) -> nsresult

/**
   * Dump the collected profile to a file.
   */

void dumpProfileToFile (in string aFilename);