Struct xpcom::interfaces::nsIPrompt

#[repr(C)]

pub struct nsIPrompt { /* fields omitted */ }

interface nsIPrompt : nsISupports

Methods

impl nsIPrompt

pub fn coerce<T: nsIPromptCoerce>(&self) -> &T

Cast this nsIPrompt to one of its base interfaces.

impl nsIPrompt

pub const BUTTON_POS_0: i64

BUTTON_POS_0: i64 = 1

pub const BUTTON_POS_1: i64

BUTTON_POS_1: i64 = 256

pub const BUTTON_POS_2: i64

BUTTON_POS_2: i64 = 65536

pub const BUTTON_TITLE_OK: i64

BUTTON_TITLE_OK: i64 = 1

pub const BUTTON_TITLE_CANCEL: i64

BUTTON_TITLE_CANCEL: i64 = 2

pub const BUTTON_TITLE_YES: i64

BUTTON_TITLE_YES: i64 = 3

pub const BUTTON_TITLE_NO: i64

BUTTON_TITLE_NO: i64 = 4

pub const BUTTON_TITLE_SAVE: i64

BUTTON_TITLE_SAVE: i64 = 5

pub const BUTTON_TITLE_DONT_SAVE: i64

BUTTON_TITLE_DONT_SAVE: i64 = 6

pub const BUTTON_TITLE_REVERT: i64

BUTTON_TITLE_REVERT: i64 = 7

pub const BUTTON_TITLE_IS_STRING: i64

BUTTON_TITLE_IS_STRING: i64 = 127

pub const BUTTON_POS_0_DEFAULT: i64

BUTTON_POS_0_DEFAULT: i64 = 0

pub const BUTTON_POS_1_DEFAULT: i64

BUTTON_POS_1_DEFAULT: i64 = 16777216

pub const BUTTON_POS_2_DEFAULT: i64

BUTTON_POS_2_DEFAULT: i64 = 33554432

pub const BUTTON_DELAY_ENABLE: i64

BUTTON_DELAY_ENABLE: i64 = 67108864

pub const STD_OK_CANCEL_BUTTONS: i64

STD_OK_CANCEL_BUTTONS: i64 = 513

pub const STD_YES_NO_BUTTONS: i64

STD_YES_NO_BUTTONS: i64 = 1027

pub unsafe fn Alert(
&self,
dialogTitle: *const int16_t,
text: *const int16_t
) -> nsresult

void alert (in wstring dialogTitle, in wstring text);

pub unsafe fn AlertCheck(
&self,
dialogTitle: *const int16_t,
text: *const int16_t,
checkMsg: *const int16_t,
checkValue: *mut bool
) -> nsresult

void alertCheck (in wstring dialogTitle, in wstring text, in wstring checkMsg, inout boolean checkValue);

pub unsafe fn Confirm(
&self,
dialogTitle: *const int16_t,
text: *const int16_t,
_retval: *mut bool
) -> nsresult

boolean confirm (in wstring dialogTitle, in wstring text);

pub unsafe fn ConfirmCheck(
&self,
dialogTitle: *const int16_t,
text: *const int16_t,
checkMsg: *const int16_t,
checkValue: *mut bool,
_retval: *mut bool
) -> nsresult

boolean confirmCheck (in wstring dialogTitle, in wstring text, in wstring checkMsg, inout boolean checkValue);

pub unsafe fn ConfirmEx(
&self,
dialogTitle: *const int16_t,
text: *const int16_t,
buttonFlags: uint32_t,
button0Title: *const int16_t,
button1Title: *const int16_t,
button2Title: *const int16_t,
checkMsg: *const int16_t,
checkValue: *mut bool,
_retval: *mut int32_t
) -> nsresult

int32_t confirmEx (in wstring dialogTitle, in wstring text, in unsigned long buttonFlags, in wstring button0Title, in wstring button1Title, in wstring button2Title, in wstring checkMsg, inout boolean checkValue);

pub unsafe fn Prompt(
&self,
dialogTitle: *const int16_t,
text: *const int16_t,
value: *mut *const int16_t,
checkMsg: *const int16_t,
checkValue: *mut bool,
_retval: *mut bool
) -> nsresult

boolean prompt (in wstring dialogTitle, in wstring text, inout wstring value, in wstring checkMsg, inout boolean checkValue);

pub unsafe fn PromptPassword(
&self,
dialogTitle: *const int16_t,
text: *const int16_t,
password: *mut *const int16_t,
checkMsg: *const int16_t,
checkValue: *mut bool,
_retval: *mut bool
) -> nsresult

boolean promptPassword (in wstring dialogTitle, in wstring text, inout wstring password, in wstring checkMsg, inout boolean checkValue);

pub unsafe fn PromptUsernameAndPassword(
&self,
dialogTitle: *const int16_t,
text: *const int16_t,
username: *mut *const int16_t,
password: *mut *const int16_t,
checkMsg: *const int16_t,
checkValue: *mut bool,
_retval: *mut bool
) -> nsresult

boolean promptUsernameAndPassword (in wstring dialogTitle, in wstring text, inout wstring username, inout wstring password, in wstring checkMsg, inout boolean checkValue);

pub unsafe fn Select(
&self,
dialogTitle: *const int16_t,
text: *const int16_t,
count: uint32_t,
selectList: *mut *const int16_t,
outSelection: *mut int32_t,
_retval: *mut bool
) -> nsresult

boolean select (in wstring dialogTitle, in wstring text, in uint32_t count, [array, size_is (count)] in wstring selectList, out long outSelection);