Struct xpcom::interfaces::nsIProxyInfo

#[repr(C)]

pub struct nsIProxyInfo { /* fields omitted */ }

interface nsIProxyInfo : nsISupports

/**
 * This interface identifies a proxy server.
 */

Methods

impl nsIProxyInfo

pub fn coerce<T: nsIProxyInfoCoerce>(&self) -> &T

Cast this nsIProxyInfo to one of its base interfaces.

impl nsIProxyInfo

pub const TRANSPARENT_PROXY_RESOLVES_HOST: i64

TRANSPARENT_PROXY_RESOLVES_HOST: i64 = 1

/****************************************************************************
   * The following "Proxy Flags" may be bit-wise combined to construct the
   * flags attribute defined on this interface.  All unspecified bits are
   * reserved for future use.
   */
/**
   * This flag is set if the proxy is to perform name resolution itself.  If
   * this is the case, the hostname is used in some fashion, and we shouldn't
   * do any form of DNS lookup ourselves.
   */

pub unsafe fn GetHost(&self, aHost: &mut nsACString) -> nsresult

/**
   * This attribute specifies the hostname of the proxy server.
   */

readonly attribute AUTF8String host;

pub unsafe fn GetPort(&self, aPort: *mut int32_t) -> nsresult

/**
   * This attribute specifies the port number of the proxy server.
   */

readonly attribute long port;

pub unsafe fn GetType(&self, aType: &mut nsACString) -> nsresult

/**
   * This attribute specifies the type of the proxy server as an ASCII string.
   *
   * Some special values for this attribute include (but are not limited to)
   * the following:
   *   "http"     HTTP proxy (or SSL CONNECT for HTTPS)
   *   "https"    HTTP proxying over TLS connection to proxy
   *   "socks"    SOCKS v5 proxy
   *   "socks4"   SOCKS v4 proxy
   *   "direct"   no proxy
   *   "unknown"  unknown proxy (see nsIProtocolProxyService::resolve)
   *
   * A future version of this interface may define additional types.
   */

readonly attribute ACString type;

pub unsafe fn GetFlags(&self, aFlags: *mut uint32_t) -> nsresult

/**
   * This attribute specifies flags that modify the proxy type.  The value of
   * this attribute is the bit-wise combination of the Proxy Flags defined
   * below.  Any undefined bits are reserved for future use.
   */

readonly attribute unsigned long flags;

pub unsafe fn GetResolveFlags(&self, aResolveFlags: *mut uint32_t) -> nsresult

/**
   * This attribute specifies flags that were used by nsIProxyProtocolService when
   * creating this ProxyInfo element.
   */

readonly attribute unsigned long resolveFlags;

pub unsafe fn GetUsername(&self, aUsername: &mut nsACString) -> nsresult

/**
   * Specifies a proxy username.
   */

readonly attribute ACString username;

pub unsafe fn GetPassword(&self, aPassword: &mut nsACString) -> nsresult

/**
   * Specifies a proxy password.
   */

readonly attribute ACString password;

pub unsafe fn GetFailoverTimeout(
&self,
aFailoverTimeout: *mut uint32_t
) -> nsresult

/**
   * This attribute specifies the failover timeout in seconds for this proxy.
   * If a nsIProxyInfo is reported as failed via nsIProtocolProxyService::
   * getFailoverForProxy, then the failed proxy will not be used again for this
   * many seconds.
   */

readonly attribute unsigned long failoverTimeout;

pub unsafe fn GetFailoverProxy(
&self,
aFailoverProxy: *mut *const nsIProxyInfo
) -> nsresult

/**
   * This attribute specifies the proxy to failover to when this proxy fails.
   */

attribute nsIProxyInfo failoverProxy;

pub unsafe fn SetFailoverProxy(
&self,
aFailoverProxy: *const nsIProxyInfo
) -> nsresult

/**
   * This attribute specifies the proxy to failover to when this proxy fails.
   */

attribute nsIProxyInfo failoverProxy;