Struct xpcom::interfaces::nsIPushNotifier

#[repr(C)]

pub struct nsIPushNotifier { /* fields omitted */ }

interface nsIPushNotifier : nsISupports

/**
 * Fires XPCOM observer notifications and service worker events for
 * messages sent to push subscriptions.
 */

Methods

impl nsIPushNotifier

pub fn coerce<T: nsIPushNotifierCoerce>(&self) -> &T

Cast this nsIPushNotifier to one of its base interfaces.

impl nsIPushNotifier

pub unsafe fn NotifyPush(
&self,
scope: &nsACString,
principal: *const nsIPrincipal,
messageId: &nsAString
) -> nsresult

/**
   * Fires a `push-message` observer notification, and sends a `push` event to
   * the service worker registered for the |scope|. |messageId| is an opaque ID
   * used to report errors if the worker fails to handle the message.
   */

void notifyPush (in ACString scope, in nsIPrincipal principal, in DOMString messageId);

pub unsafe fn NotifyPushWithData(
&self,
scope: &nsACString,
principal: *const nsIPrincipal,
messageId: &nsAString,
dataLen: uint32_t,
data: *mut uint8_t
) -> nsresult

/**
   * Same as `notifyPush`, except the subject of the observer notification
   * receives an `nsIPushMessage` instance containing the |data|. Service
   * workers can access the |data| via the `PushMessageData` WebIDL interface.
   */

void notifyPushWithData (in ACString scope, in nsIPrincipal principal, in DOMString messageId, [optional] in uint32_t dataLen, [array, size_is (dataLen)] in uint8_t data);

pub unsafe fn NotifySubscriptionChange(
&self,
scope: &nsACString,
principal: *const nsIPrincipal
) -> nsresult

/**
   * Fires a `push-subscription-change` observer notification, and sends a
   * `pushsubscriptionchange` event to the service worker registered for the
   * |scope|.
   */

void notifySubscriptionChange (in ACString scope, in nsIPrincipal principal);

pub unsafe fn NotifySubscriptionModified(
&self,
scope: &nsACString,
principal: *const nsIPrincipal
) -> nsresult

/**
   * Fires a `push-subscription-modified` observer notification. Chrome code
   * can listen for this notification to see when a subscription is added,
   * updated, removed, or expired for any |scope|.
   *
   * This is useful for Dev Tools and debugging add-ons that passively observe
   * when subscriptions are created or dropped. Other callers should listen for
   * `push-subscription-change` and resubscribe instead.
   */

void notifySubscriptionModified (in ACString scope, in nsIPrincipal principal);

pub unsafe fn NotifyError(
&self,
scope: &nsACString,
principal: *const nsIPrincipal,
message: &nsAString,
flags: uint32_t
) -> nsresult

void notifyError (in ACString scope, in nsIPrincipal principal, in DOMString message, in uint32_t flags);