Struct xpcom::interfaces::nsIQueryContentEventResult

#[repr(C)]

pub struct nsIQueryContentEventResult { /* fields omitted */ }

interface nsIQueryContentEventResult : nsISupports

/**
 * The result of query content events.  succeeded propery can be used always.
 * Whether other properties can be used or not depends on the event.
 * See nsIDOMWindowUtils.idl, which properites can be used was documented.
 */

Methods

impl nsIQueryContentEventResult

pub fn coerce<T: nsIQueryContentEventResultCoerce>(&self) -> &T

Cast this nsIQueryContentEventResult to one of its base interfaces.

impl nsIQueryContentEventResult

pub unsafe fn GetOffset(&self, aOffset: *mut uint32_t) -> nsresult

readonly attribute unsigned long offset;

pub unsafe fn GetTentativeCaretOffset(
&self,
aTentativeCaretOffset: *mut uint32_t
) -> nsresult

readonly attribute unsigned long tentativeCaretOffset;

pub unsafe fn GetReversed(&self, aReversed: *mut bool) -> nsresult

readonly attribute boolean reversed;

pub unsafe fn GetLeft(&self, aLeft: *mut int32_t) -> nsresult

readonly attribute long left;

pub unsafe fn GetTop(&self, aTop: *mut int32_t) -> nsresult

readonly attribute long top;

pub unsafe fn GetWidth(&self, aWidth: *mut int32_t) -> nsresult

readonly attribute long width;

pub unsafe fn GetHeight(&self, aHeight: *mut int32_t) -> nsresult

readonly attribute long height;

pub unsafe fn GetText(&self, aText: &mut nsAString) -> nsresult

readonly attribute AString text;

pub unsafe fn GetCharacterRect(
&self,
offset: int32_t,
left: *mut int32_t,
top: *mut int32_t,
width: *mut int32_t,
height: *mut int32_t
) -> nsresult

void getCharacterRect (in long offset, out long left, out long top, out long width, out long height);

pub unsafe fn GetSucceeded(&self, aSucceeded: *mut bool) -> nsresult

readonly attribute boolean succeeded;

pub unsafe fn GetNotFound(&self, aNotFound: *mut bool) -> nsresult

readonly attribute boolean notFound;

pub unsafe fn GetTentativeCaretOffsetNotFound(
&self,
aTentativeCaretOffsetNotFound: *mut bool
) -> nsresult

readonly attribute boolean tentativeCaretOffsetNotFound;