Struct xpcom::interfaces::nsIRDFCompositeDataSource

#[repr(C)]

pub struct nsIRDFCompositeDataSource { /* fields omitted */ }

interface nsIRDFCompositeDataSource : nsIRDFDataSource

/**
 * An nsIRDFCompositeDataSource composes individual data sources, providing
 * the illusion of a single, coherent RDF graph.
 */

Methods

impl nsIRDFCompositeDataSource

pub fn coerce<T: nsIRDFCompositeDataSourceCoerce>(&self) -> &T

Cast this nsIRDFCompositeDataSource to one of its base interfaces.

impl nsIRDFCompositeDataSource

pub unsafe fn GetAllowNegativeAssertions(
&self,
aAllowNegativeAssertions: *mut bool
) -> nsresult

/**
     *
     * Set this value to <code>true</code> if the composite datasource
     * may contains at least one datasource that has <em>negative</em>
     * assertions. (This is the default.)
     *
     * Set this value to <code>false</code> if none of the datasources
     * being composed contains a negative assertion. This allows the
     * composite datasource to perform some query optimizations.
     *
     * By default, this value is <code>true</true>.
     */

attribute boolean allowNegativeAssertions;

pub unsafe fn SetAllowNegativeAssertions(
&self,
aAllowNegativeAssertions: bool
) -> nsresult

/**
     *
     * Set this value to <code>true</code> if the composite datasource
     * may contains at least one datasource that has <em>negative</em>
     * assertions. (This is the default.)
     *
     * Set this value to <code>false</code> if none of the datasources
     * being composed contains a negative assertion. This allows the
     * composite datasource to perform some query optimizations.
     *
     * By default, this value is <code>true</true>.
     */

attribute boolean allowNegativeAssertions;

pub unsafe fn GetCoalesceDuplicateArcs(
&self,
aCoalesceDuplicateArcs: *mut bool
) -> nsresult

/**
     * Set to <code>true</code> if the composite datasource should
     * take care to coalesce duplicate arcs when returning values from
     * queries. (This is the default.)
     *
     * Set to <code>false</code> if the composite datasource shouldn't
     * bother to check for duplicates. This allows the composite
     * datasource to more efficiently answer queries.
     *
     * By default, this value is <code>true</code>.
     */

attribute boolean coalesceDuplicateArcs;

pub unsafe fn SetCoalesceDuplicateArcs(
&self,
aCoalesceDuplicateArcs: bool
) -> nsresult

/**
     * Set to <code>true</code> if the composite datasource should
     * take care to coalesce duplicate arcs when returning values from
     * queries. (This is the default.)
     *
     * Set to <code>false</code> if the composite datasource shouldn't
     * bother to check for duplicates. This allows the composite
     * datasource to more efficiently answer queries.
     *
     * By default, this value is <code>true</code>.
     */

attribute boolean coalesceDuplicateArcs;

pub unsafe fn AddDataSource(
&self,
aDataSource: *const nsIRDFDataSource
) -> nsresult

/**
     * Add a datasource the the composite data source.
     * @param aDataSource the datasource to add to composite
     */

void AddDataSource (in nsIRDFDataSource aDataSource);

pub unsafe fn RemoveDataSource(
&self,
aDataSource: *const nsIRDFDataSource
) -> nsresult

/**
     * Remove a datasource from the composite data source.
     * @param aDataSource the datasource to remove from the composite
     */

void RemoveDataSource (in nsIRDFDataSource aDataSource);

pub unsafe fn GetDataSources(
&self,
_retval: *mut *const nsISimpleEnumerator
) -> nsresult

/**
     * Retrieve the datasources in the composite data source.
     * @return an nsISimpleEnumerator that will enumerate each
     * of the datasources in the composite
     */

nsISimpleEnumerator GetDataSources ();