Struct xpcom::interfaces::nsIRDFContainerUtils

#[repr(C)]

pub struct nsIRDFContainerUtils { /* fields omitted */ }

interface nsIRDFContainerUtils : nsISupports

Methods

impl nsIRDFContainerUtils

pub fn coerce<T: nsIRDFContainerUtilsCoerce>(&self) -> &T

Cast this nsIRDFContainerUtils to one of its base interfaces.

impl nsIRDFContainerUtils

pub unsafe fn IsOrdinalProperty(
&self,
aProperty: *const nsIRDFResource,
_retval: *mut bool
) -> nsresult

/**
     * Returns 'true' if the property is an RDF ordinal property.
     */

boolean IsOrdinalProperty (in nsIRDFResource aProperty);

pub unsafe fn IndexToOrdinalResource(
&self,
aIndex: int32_t,
_retval: *mut *const nsIRDFResource
) -> nsresult

/**
     * Convert the specified index to an ordinal property.
     */

nsIRDFResource IndexToOrdinalResource (in long aIndex);

pub unsafe fn OrdinalResourceToIndex(
&self,
aOrdinal: *const nsIRDFResource,
_retval: *mut int32_t
) -> nsresult

/**
     * Convert the specified ordinal property into an index
     */

long OrdinalResourceToIndex (in nsIRDFResource aOrdinal);

pub unsafe fn IsContainer(
&self,
aDataSource: *const nsIRDFDataSource,
aResource: *const nsIRDFResource,
_retval: *mut bool
) -> nsresult

/**
     * Return 'true' if the specified resource is a container
     */

boolean IsContainer (in nsIRDFDataSource aDataSource, in nsIRDFResource aResource);

pub unsafe fn IsEmpty(
&self,
aDataSource: *const nsIRDFDataSource,
aResource: *const nsIRDFResource,
_retval: *mut bool
) -> nsresult

/**
     * Return 'true' if the specified resource is a container and it is empty
     */

boolean IsEmpty (in nsIRDFDataSource aDataSource, in nsIRDFResource aResource);

pub unsafe fn IsBag(
&self,
aDataSource: *const nsIRDFDataSource,
aResource: *const nsIRDFResource,
_retval: *mut bool
) -> nsresult

/**
     * Return 'true' if the specified resource is a bag
     */

boolean IsBag (in nsIRDFDataSource aDataSource, in nsIRDFResource aResource);

pub unsafe fn IsSeq(
&self,
aDataSource: *const nsIRDFDataSource,
aResource: *const nsIRDFResource,
_retval: *mut bool
) -> nsresult

/**
     * Return 'true' if the specified resource is a sequence
     */

boolean IsSeq (in nsIRDFDataSource aDataSource, in nsIRDFResource aResource);

pub unsafe fn IsAlt(
&self,
aDataSource: *const nsIRDFDataSource,
aResource: *const nsIRDFResource,
_retval: *mut bool
) -> nsresult

/**
     * Return 'true' if the specified resource is an alternation
     */

boolean IsAlt (in nsIRDFDataSource aDataSource, in nsIRDFResource aResource);

pub unsafe fn MakeBag(
&self,
aDataSource: *const nsIRDFDataSource,
aResource: *const nsIRDFResource,
_retval: *mut *const nsIRDFContainer
) -> nsresult

/**
     * Decorates the specified resource appropriately to make it
     * usable as an empty bag in the specified data source.
     */

nsIRDFContainer MakeBag (in nsIRDFDataSource aDataSource, in nsIRDFResource aResource);

pub unsafe fn MakeSeq(
&self,
aDataSource: *const nsIRDFDataSource,
aResource: *const nsIRDFResource,
_retval: *mut *const nsIRDFContainer
) -> nsresult

/**
     * Decorates the specified resource appropriately to make it
     * usable as an empty sequence in the specified data source.
     */

nsIRDFContainer MakeSeq (in nsIRDFDataSource aDataSource, in nsIRDFResource aResource);

pub unsafe fn MakeAlt(
&self,
aDataSource: *const nsIRDFDataSource,
aResource: *const nsIRDFResource,
_retval: *mut *const nsIRDFContainer
) -> nsresult

/**
     * Decorates the specified resource appropriately to make it
     * usable as an empty alternation in the specified data source.
     */

nsIRDFContainer MakeAlt (in nsIRDFDataSource aDataSource, in nsIRDFResource aResource);

pub unsafe fn IndexOf(
&self,
aDataSource: *const nsIRDFDataSource,
aContainer: *const nsIRDFResource,
aElement: *const nsIRDFNode,
_retval: *mut int32_t
) -> nsresult

/**
     * Retrieve the index of element in the container. Returns -1 if
     * the element is not in the container.
     */

long indexOf (in nsIRDFDataSource aDataSource, in nsIRDFResource aContainer, in nsIRDFNode aElement);