Struct xpcom::interfaces::nsIRDFDataSource

#[repr(C)]

pub struct nsIRDFDataSource { /* fields omitted */ }

interface nsIRDFDataSource : nsISupports

Methods

impl nsIRDFDataSource

pub fn coerce<T: nsIRDFDataSourceCoerce>(&self) -> &T

Cast this nsIRDFDataSource to one of its base interfaces.

impl nsIRDFDataSource

pub unsafe fn GetURI(&self, aURI: &mut nsACString) -> nsresult

/** The "URI" of the data source. This used by the RDF service's
     * |GetDataSource()| method to cache datasources.
     */

readonly attribute ACString URI;

pub unsafe fn GetSource(
&self,
aProperty: *const nsIRDFResource,
aTarget: *const nsIRDFNode,
aTruthValue: bool,
_retval: *mut *const nsIRDFResource
) -> nsresult

/** Find an RDF resource that points to a given node over the
     * specified arc & truth value
     *
     * @throws NS_RDF_NO_VALUE if there is no source that leads
     * to the target with the specified property.
     */

nsIRDFResource GetSource (in nsIRDFResource aProperty, in nsIRDFNode aTarget, in boolean aTruthValue);

pub unsafe fn GetSources(
&self,
aProperty: *const nsIRDFResource,
aTarget: *const nsIRDFNode,
aTruthValue: bool,
_retval: *mut *const nsISimpleEnumerator
) -> nsresult

/**
     * Find all RDF resources that point to a given node over the
     * specified arc & truth value
     */

nsISimpleEnumerator GetSources (in nsIRDFResource aProperty, in nsIRDFNode aTarget, in boolean aTruthValue);

pub unsafe fn GetTarget(
&self,
aSource: *const nsIRDFResource,
aProperty: *const nsIRDFResource,
aTruthValue: bool,
_retval: *mut *const nsIRDFNode
) -> nsresult

/**
     * Find a child of that is related to the source by the given arc
     * arc and truth value
     *
     * @throws NS_RDF_NO_VALUE if there is no target accessible from the
     * source via the specified property.
     */

nsIRDFNode GetTarget (in nsIRDFResource aSource, in nsIRDFResource aProperty, in boolean aTruthValue);

pub unsafe fn GetTargets(
&self,
aSource: *const nsIRDFResource,
aProperty: *const nsIRDFResource,
aTruthValue: bool,
_retval: *mut *const nsISimpleEnumerator
) -> nsresult

/**
     * Find all children of that are related to the source by the given arc
     * arc and truth value.
     */

nsISimpleEnumerator GetTargets (in nsIRDFResource aSource, in nsIRDFResource aProperty, in boolean aTruthValue);

pub unsafe fn Assert(
&self,
aSource: *const nsIRDFResource,
aProperty: *const nsIRDFResource,
aTarget: *const nsIRDFNode,
aTruthValue: bool
) -> nsresult

/**
     * Add an assertion to the graph.
     */

void Assert (in nsIRDFResource aSource, in nsIRDFResource aProperty, in nsIRDFNode aTarget, in boolean aTruthValue);

pub unsafe fn Unassert(
&self,
aSource: *const nsIRDFResource,
aProperty: *const nsIRDFResource,
aTarget: *const nsIRDFNode
) -> nsresult

/**
     * Remove an assertion from the graph.
     */

void Unassert (in nsIRDFResource aSource, in nsIRDFResource aProperty, in nsIRDFNode aTarget);

pub unsafe fn Change(
&self,
aSource: *const nsIRDFResource,
aProperty: *const nsIRDFResource,
aOldTarget: *const nsIRDFNode,
aNewTarget: *const nsIRDFNode
) -> nsresult

/**
     * Change an assertion from
     *
     *   [aSource]--[aProperty]-->[aOldTarget]
     *
     * to
     *
     *   [aSource]--[aProperty]-->[aNewTarget]
     */

void Change (in nsIRDFResource aSource, in nsIRDFResource aProperty, in nsIRDFNode aOldTarget, in nsIRDFNode aNewTarget);

pub unsafe fn Move(
&self,
aOldSource: *const nsIRDFResource,
aNewSource: *const nsIRDFResource,
aProperty: *const nsIRDFResource,
aTarget: *const nsIRDFNode
) -> nsresult

/**
     * 'Move' an assertion from
     *
     *   [aOldSource]--[aProperty]-->[aTarget]
     *
     * to
     *
     *   [aNewSource]--[aProperty]-->[aTarget]
     */

void Move (in nsIRDFResource aOldSource, in nsIRDFResource aNewSource, in nsIRDFResource aProperty, in nsIRDFNode aTarget);

pub unsafe fn HasAssertion(
&self,
aSource: *const nsIRDFResource,
aProperty: *const nsIRDFResource,
aTarget: *const nsIRDFNode,
aTruthValue: bool,
_retval: *mut bool
) -> nsresult

/**
     * Query whether an assertion exists in this graph.
     */

boolean HasAssertion (in nsIRDFResource aSource, in nsIRDFResource aProperty, in nsIRDFNode aTarget, in boolean aTruthValue);

pub unsafe fn AddObserver(&self, aObserver: *const nsIRDFObserver) -> nsresult

/**
     * Add an observer to this data source. If the datasource
     * supports observers, the datasource source should hold a strong
     * reference to the observer.
     */

void AddObserver (in nsIRDFObserver aObserver);

pub unsafe fn RemoveObserver(
&self,
aObserver: *const nsIRDFObserver
) -> nsresult

/**
     * Remove an observer from this data source.
     */

void RemoveObserver (in nsIRDFObserver aObserver);

pub unsafe fn ArcLabelsIn(
&self,
aNode: *const nsIRDFNode,
_retval: *mut *const nsISimpleEnumerator
) -> nsresult

/**
     * Get a cursor to iterate over all the arcs that point into a node.
     */

nsISimpleEnumerator ArcLabelsIn (in nsIRDFNode aNode);

pub unsafe fn ArcLabelsOut(
&self,
aSource: *const nsIRDFResource,
_retval: *mut *const nsISimpleEnumerator
) -> nsresult

/**
     * Get a cursor to iterate over all the arcs that originate in
     * a resource.
     */

nsISimpleEnumerator ArcLabelsOut (in nsIRDFResource aSource);

pub unsafe fn GetAllResources(
&self,
_retval: *mut *const nsISimpleEnumerator
) -> nsresult

/**
     * Retrieve all of the resources that the data source currently
     * refers to.
     */

nsISimpleEnumerator GetAllResources ();

pub unsafe fn IsCommandEnabled(
&self,
aSources: *const nsISupports,
aCommand: *const nsIRDFResource,
aArguments: *const nsISupports,
_retval: *mut bool
) -> nsresult

/**
     * Returns whether a given command is enabled for a set of sources.
     */

boolean IsCommandEnabled (in nsISupports aSources, in nsIRDFResource aCommand, in nsISupports aArguments);

pub unsafe fn DoCommand(
&self,
aSources: *const nsISupports,
aCommand: *const nsIRDFResource,
aArguments: *const nsISupports
) -> nsresult

/**
     * Perform the specified command on set of sources.
     */

void DoCommand (in nsISupports aSources, in nsIRDFResource aCommand, in nsISupports aArguments);

pub unsafe fn GetAllCmds(
&self,
aSource: *const nsIRDFResource,
_retval: *mut *const nsISimpleEnumerator
) -> nsresult

/**
     * Returns the set of all commands defined for a given source.
     */

nsISimpleEnumerator GetAllCmds (in nsIRDFResource aSource);

pub unsafe fn HasArcIn(
&self,
aNode: *const nsIRDFNode,
aArc: *const nsIRDFResource,
_retval: *mut bool
) -> nsresult

/**
     * Returns true if the specified node is pointed to by the specified arc.
     * Equivalent to enumerating ArcLabelsIn and comparing for the specified arc.
     */

boolean hasArcIn (in nsIRDFNode aNode, in nsIRDFResource aArc);

pub unsafe fn HasArcOut(
&self,
aSource: *const nsIRDFResource,
aArc: *const nsIRDFResource,
_retval: *mut bool
) -> nsresult

/**
     * Returns true if the specified node has the specified outward arc.
     * Equivalent to enumerating ArcLabelsOut and comparing for the specified arc.
     */

boolean hasArcOut (in nsIRDFResource aSource, in nsIRDFResource aArc);

pub unsafe fn BeginUpdateBatch(&self) -> nsresult

/**
     * Notify observers that the datasource is about to send several
     * notifications at once.
     * This must be followed by calling endUpdateBatch(), otherwise
     * viewers will get out of sync.
     */

void beginUpdateBatch ();

pub unsafe fn EndUpdateBatch(&self) -> nsresult

/**
     * Notify observers that the datasource has completed issuing
     * a notification group.
     */

void endUpdateBatch ();