Struct xpcom::interfaces::nsIRDFLiteral

#[repr(C)]

pub struct nsIRDFLiteral { /* fields omitted */ }

interface nsIRDFLiteral : nsIRDFNode

/**
 * A literal node in the graph, whose value is a string.
 */

Methods

impl nsIRDFLiteral

pub fn coerce<T: nsIRDFLiteralCoerce>(&self) -> &T

Cast this nsIRDFLiteral to one of its base interfaces.

impl nsIRDFLiteral

pub unsafe fn GetValue(&self, aValue: *mut *const int16_t) -> nsresult

/**
     * The Unicode string value of the literal.
     */

readonly attribute wstring Value;

pub unsafe fn GetValueConst(&self, aConstValue: *mut *const int16_t) -> nsresult

/**
     * An unscriptable version used to avoid a string copy. Meant
     * for use as a performance optimization.
     */

[noscript] void GetValueConst ([shared] out wstring aConstValue);