Struct xpcom::interfaces::nsIRequestContext

#[repr(C)]

pub struct nsIRequestContext { /* fields omitted */ }

interface nsIRequestContext : nsISupports

/**
 * The nsIRequestContext is used to maintain state about connections
 * that are in some way associated with each other (often by being part
 * of the same load group) and how they interact with blocking items like
 * HEAD css/js loads.
 *
 * This used to be known as nsILoadGroupConnectionInfo and nsISchedulingContext.
 */

Methods

impl nsIRequestContext

pub fn coerce<T: nsIRequestContextCoerce>(&self) -> &T

Cast this nsIRequestContext to one of its base interfaces.

impl nsIRequestContext

pub unsafe fn GetID(&self, aID: *mut uint64_t) -> nsresult

/**
   * A unique identifier for this request context
   */

[noscript] readonly attribute unsigned long long ID;

pub unsafe fn BeginLoad(&self) -> nsresult

/**
   * Called by the associated document when its load starts.  This resets
   * context's internal states.
   */

void beginLoad ();

pub unsafe fn DOMContentLoaded(&self) -> nsresult

/**
  * Called when the associated document notified the DOMContentLoaded event.
  */

void DOMContentLoaded ();

pub unsafe fn GetBlockingTransactionCount(
&self,
aBlockingTransactionCount: *mut uint32_t
) -> nsresult

/**
   * Number of active blocking transactions associated with this context
   */

readonly attribute unsigned long blockingTransactionCount;

pub unsafe fn AddBlockingTransaction(&self) -> nsresult

/**
   * Increase the number of active blocking transactions associated
   * with this context by one.
   */

void addBlockingTransaction ();

pub unsafe fn RemoveBlockingTransaction(
&self,
_retval: *mut uint32_t
) -> nsresult

/**
   * Decrease the number of active blocking transactions associated
   * with this context by one. The return value is the number of remaining
   * blockers.
   */

unsigned long removeBlockingTransaction ();

pub unsafe fn GetUserAgentOverride(
&self,
aUserAgentOverride: &mut nsACString
) -> nsresult

/**
   * This holds a cached value of the user agent override.
   */

[noscript] attribute ACString userAgentOverride;

pub unsafe fn SetUserAgentOverride(
&self,
aUserAgentOverride: &nsACString
) -> nsresult

/**
   * This holds a cached value of the user agent override.
   */

[noscript] attribute ACString userAgentOverride;

pub unsafe fn AddNonTailRequest(&self) -> nsresult

/**
   * Increases/decrease the number of non-tailed requests in this context.
   * If the count drops to zero, all tail-blocked callbacks are notified
   * shortly after that to be unblocked.
   */

void addNonTailRequest ();

pub unsafe fn RemoveNonTailRequest(&self) -> nsresult

void removeNonTailRequest ();

pub unsafe fn IsContextTailBlocked(
&self,
callback: *const nsIRequestTailUnblockCallback,
_retval: *mut bool
) -> nsresult

/**
   * If the request context is in tail-blocked state, the callback
   * is queued and result is true.  The callback will be notified
   * about tail-unblocking or when the request context is canceled.
   */

[must_use] boolean isContextTailBlocked (in nsIRequestTailUnblockCallback callback);

pub unsafe fn CancelTailedRequest(
&self,
request: *const nsIRequestTailUnblockCallback
) -> nsresult

/**
   * Called when the request is sitting in the tail queue but has been
   * canceled before untailing.  This just removes the request from the
   * queue so that it is not notified on untail and not referenced.
   */

void cancelTailedRequest (in nsIRequestTailUnblockCallback request);

pub unsafe fn CancelTailPendingRequests(&self, aResult: nsresult) -> nsresult

/**
   * This notifies all queued tail-blocked requests, they will be notified
   * aResult and released afterwards.  Called by the load group when
   * it's canceled.
   */

void cancelTailPendingRequests (in nsresult aResult);