Struct xpcom::interfaces::nsISAXXMLReader

#[repr(C)]

pub struct nsISAXXMLReader { /* fields omitted */ }

interface nsISAXXMLReader : nsIStreamListener

/**
 * Interface for reading an XML document using callbacks.
 *
 * NOTE: This file (and related ones) used to implement a full-featured SAX XML
 * parser. But we have few uses of this interface and the ones we have only use
 * a fraction of the full SAX functionality. So in bug 1416038 we removed a lot
 * of it. See the changes associated with that bug if you need to reintroduce
 * any of that functionality.
 *
 * nsISAXXMLReader is the interface that an XML parser's SAX2
 * driver must implement.  This interface allows an application to set
 * and query features and properties in the parser, to register event
 * handlers for document processing, and to initiate a document
 * parse.
 */

Methods

impl nsISAXXMLReader

pub fn coerce<T: nsISAXXMLReaderCoerce>(&self) -> &T

Cast this nsISAXXMLReader to one of its base interfaces.

impl nsISAXXMLReader

pub unsafe fn GetBaseURI(&self, aBaseURI: *mut *const nsIURI) -> nsresult

/**
   * The base URI.
   */

attribute nsIURI baseURI;

pub unsafe fn SetBaseURI(&self, aBaseURI: *const nsIURI) -> nsresult

/**
   * The base URI.
   */

attribute nsIURI baseURI;

pub unsafe fn GetContentHandler(
&self,
aContentHandler: *mut *const nsISAXContentHandler
) -> nsresult

/**
   * If the application does not register a content handler, all
   * content events reported by the SAX parser will be silently
   * ignored.
   *
   * Applications may register a new or different handler in the
   * middle of a parse, and the SAX parser must begin using the new
   * handler immediately.
   */

attribute nsISAXContentHandler contentHandler;

pub unsafe fn SetContentHandler(
&self,
aContentHandler: *const nsISAXContentHandler
) -> nsresult

/**
   * If the application does not register a content handler, all
   * content events reported by the SAX parser will be silently
   * ignored.
   *
   * Applications may register a new or different handler in the
   * middle of a parse, and the SAX parser must begin using the new
   * handler immediately.
   */

attribute nsISAXContentHandler contentHandler;

pub unsafe fn GetErrorHandler(
&self,
aErrorHandler: *mut *const nsISAXErrorHandler
) -> nsresult

/**
   * If the application does not register an error handler, all
   * error events reported by the SAX parser will be silently ignored;
   * however, normal processing may not continue.  It is highly
   * recommended that all SAX applications implement an error handler
   * to avoid unexpected bugs.
   *
   * Applications may register a new or different handler in the
   * middle of a parse, and the SAX parser must begin using the new
   * handler immediately.
   */

attribute nsISAXErrorHandler errorHandler;

pub unsafe fn SetErrorHandler(
&self,
aErrorHandler: *const nsISAXErrorHandler
) -> nsresult

/**
   * If the application does not register an error handler, all
   * error events reported by the SAX parser will be silently ignored;
   * however, normal processing may not continue.  It is highly
   * recommended that all SAX applications implement an error handler
   * to avoid unexpected bugs.
   *
   * Applications may register a new or different handler in the
   * middle of a parse, and the SAX parser must begin using the new
   * handler immediately.
   */

attribute nsISAXErrorHandler errorHandler;

pub unsafe fn ParseFromString(
&self,
str: &nsAString,
contentType: *const c_char
) -> nsresult

/**
   * @param str The UTF16 string to be parsed
   * @param contentType The content type of the string (see parseFromStream)
   */

void parseFromString (in AString str, in string contentType);

pub unsafe fn ParseFromStream(
&self,
stream: *const nsIInputStream,
charset: *const c_char,
contentType: *const c_char
) -> nsresult

/**
   * @param stream The byte stream whose contents are parsed
   * @param charset The character set that was used to encode the byte
   *                stream. NULL if not specified.
   * @param contentType The content type of the string - either text/xml,
   *                    application/xml, or application/xhtml+xml.
   *                    Must not be NULL.
   */

void parseFromStream (in nsIInputStream stream, in string charset, in string contentType);

pub unsafe fn ParseAsync(&self, observer: *const nsIRequestObserver) -> nsresult

/**
   * Begin an asynchronous parse. This method initializes the parser,
   * and must be called before any nsIStreamListener methods. It is
   * then the caller's duty to call nsIStreamListener methods to drive
   * the parser. Once this method is called, the caller must not call
   * one of the other parse methods.
   *
   * @param observer The nsIRequestObserver to notify upon start or stop.
   *                 Can be NULL.
   */

void parseAsync (in nsIRequestObserver observer);