Struct xpcom::interfaces::nsISHEntry

#[repr(C)]

pub struct nsISHEntry { /* fields omitted */ }

interface nsISHEntry : nsISupports

Methods

impl nsISHEntry

pub fn coerce<T: nsISHEntryCoerce>(&self) -> &T

Cast this nsISHEntry to one of its base interfaces.

impl nsISHEntry

pub unsafe fn GetURI(&self, aURI: *mut *const nsIURI) -> nsresult

/**
     * A readonly property that returns the URI
     * of the current entry. The object returned is
     * of type nsIURI
     */

readonly attribute nsIURI URI;

pub unsafe fn GetOriginalURI(
&self,
aOriginalURI: *mut *const nsIURI
) -> nsresult

/**
     * A readonly property that returns the original URI of the current entry.
     * If an entry is the result of a redirect this attribute holds original
     * URI. The object returned is of type nsIURI
     */

attribute nsIURI originalURI;

pub unsafe fn SetOriginalURI(&self, aOriginalURI: *const nsIURI) -> nsresult

/**
     * A readonly property that returns the original URI of the current entry.
     * If an entry is the result of a redirect this attribute holds original
     * URI. The object returned is of type nsIURI
     */

attribute nsIURI originalURI;

pub unsafe fn GetResultPrincipalURI(
&self,
aResultPrincipalURI: *mut *const nsIURI
) -> nsresult

/**
     * URL as stored from nsILoadInfo.resultPrincipalURI.  See nsILoadInfo
     * for more details.
     */

attribute nsIURI resultPrincipalURI;

pub unsafe fn SetResultPrincipalURI(
&self,
aResultPrincipalURI: *const nsIURI
) -> nsresult

/**
     * URL as stored from nsILoadInfo.resultPrincipalURI.  See nsILoadInfo
     * for more details.
     */

attribute nsIURI resultPrincipalURI;

pub unsafe fn GetLoadReplace(&self, aLoadReplace: *mut bool) -> nsresult

/**
     *  This flag remembers whether channel has LOAD_REPLACE set.
     */

attribute boolean loadReplace;

pub unsafe fn SetLoadReplace(&self, aLoadReplace: bool) -> nsresult

/**
     *  This flag remembers whether channel has LOAD_REPLACE set.
     */

attribute boolean loadReplace;

pub unsafe fn GetTitle(&self, aTitle: *mut *const int16_t) -> nsresult

/**
     * A readonly property that returns the title
     * of the current entry.  The object returned
     * is a encoded string
     */

readonly attribute wstring title;

pub unsafe fn GetIsSubFrame(&self, aIsSubFrame: *mut bool) -> nsresult

/**
     * A readonly property that returns a boolean
     * flag which indicates if the entry was created as a
     * result of a subframe navigation. This flag will be
     * 'false' when a frameset page is visited for
     * the first time. This flag will be 'true' for all
     * history entries created as a result of a subframe
     * navigation.
     */

readonly attribute boolean isSubFrame;

pub unsafe fn SetURI(&self, aURI: *const nsIURI) -> nsresult

/** URI for the document */

void setURI (in nsIURI aURI);

pub unsafe fn GetReferrerURI(
&self,
aReferrerURI: *mut *const nsIURI
) -> nsresult

/** Referrer URI */

attribute nsIURI referrerURI;

pub unsafe fn SetReferrerURI(&self, aReferrerURI: *const nsIURI) -> nsresult

/** Referrer URI */

attribute nsIURI referrerURI;

pub unsafe fn GetReferrerPolicy(
&self,
aReferrerPolicy: *mut uint32_t
) -> nsresult

/** Referrer policy, holding one of the values (REFERRER_POLICY_*)
     *  defined in nsIHttpChannel.
     */

attribute unsigned long referrerPolicy;

pub unsafe fn SetReferrerPolicy(&self, aReferrerPolicy: uint32_t) -> nsresult

/** Referrer policy, holding one of the values (REFERRER_POLICY_*)
     *  defined in nsIHttpChannel.
     */

attribute unsigned long referrerPolicy;

pub unsafe fn GetContentViewer(
&self,
aContentViewer: *mut *const nsIContentViewer
) -> nsresult

/** Content viewer, for fast restoration of presentation */

attribute nsIContentViewer contentViewer;

pub unsafe fn SetContentViewer(
&self,
aContentViewer: *const nsIContentViewer
) -> nsresult

/** Content viewer, for fast restoration of presentation */

attribute nsIContentViewer contentViewer;

pub unsafe fn GetSticky(&self, aSticky: *mut bool) -> nsresult

/** Whether the content viewer is marked "sticky" */

attribute boolean sticky;

pub unsafe fn SetSticky(&self, aSticky: bool) -> nsresult

/** Whether the content viewer is marked "sticky" */

attribute boolean sticky;

pub unsafe fn GetWindowState(
&self,
aWindowState: *mut *const nsISupports
) -> nsresult

/** Saved state of the global window object */

attribute nsISupports windowState;

pub unsafe fn SetWindowState(
&self,
aWindowState: *const nsISupports
) -> nsresult

/** Saved state of the global window object */

attribute nsISupports windowState;

pub unsafe fn AddChildShell(
&self,
shell: *const nsIDocShellTreeItem
) -> nsresult

/**
     * Saved child docshells corresponding to contentViewer.  The child shells
     * are restored as children of the parent docshell, in this order, when the
     * parent docshell restores a saved presentation.
     */
/** Append a child shell to the end of our list. */

void addChildShell (in nsIDocShellTreeItem shell);

pub unsafe fn ChildShellAt(
&self,
index: int32_t,
_retval: *mut *const nsIDocShellTreeItem
) -> nsresult

/**
     * Get the child shell at |index|; returns null if |index| is out of bounds.
     */

nsIDocShellTreeItem childShellAt (in long index);

pub unsafe fn ClearChildShells(&self) -> nsresult

/**
     * Clear the child shell list.
     */

void clearChildShells ();

pub unsafe fn GetRefreshURIList(
&self,
aRefreshURIList: *mut *const nsIMutableArray
) -> nsresult

/** Saved refresh URI list for the content viewer */

attribute nsIMutableArray refreshURIList;

pub unsafe fn SetRefreshURIList(
&self,
aRefreshURIList: *const nsIMutableArray
) -> nsresult

/** Saved refresh URI list for the content viewer */

attribute nsIMutableArray refreshURIList;

pub unsafe fn SyncPresentationState(&self) -> nsresult

/**
     * Ensure that the cached presentation members are self-consistent.
     * If either |contentViewer| or |windowState| are null, then all of the
     * following members are cleared/reset:
     *  contentViewer, sticky, windowState, viewerBounds, childShells,
     *  refreshURIList.
     */

void syncPresentationState ();

pub unsafe fn SetTitle(&self, aTitle: &nsAString) -> nsresult

/** Title for the document */

void setTitle (in AString aTitle);

pub unsafe fn GetPostData(
&self,
aPostData: *mut *const nsIInputStream
) -> nsresult

/** Post Data for the document */

attribute nsIInputStream postData;

pub unsafe fn SetPostData(&self, aPostData: *const nsIInputStream) -> nsresult

/** Post Data for the document */

attribute nsIInputStream postData;

pub unsafe fn GetLayoutHistoryState(
&self,
aLayoutHistoryState: *mut *const nsILayoutHistoryState
) -> nsresult

/** LayoutHistoryState for scroll position and form values */

attribute nsILayoutHistoryState layoutHistoryState;

pub unsafe fn SetLayoutHistoryState(
&self,
aLayoutHistoryState: *const nsILayoutHistoryState
) -> nsresult

/** LayoutHistoryState for scroll position and form values */

attribute nsILayoutHistoryState layoutHistoryState;

pub unsafe fn InitLayoutHistoryState(
&self,
_retval: *mut *const nsILayoutHistoryState
) -> nsresult

/**
     * Initialises the LayoutHistoryState if it doesn't already exist
     * and returns a reference to it.
     */

nsILayoutHistoryState initLayoutHistoryState ();

pub unsafe fn GetParent(&self, aParent: *mut *const nsISHEntry) -> nsresult

/** parent of this entry */

attribute nsISHEntry parent;

pub unsafe fn SetParent(&self, aParent: *const nsISHEntry) -> nsresult

/** parent of this entry */

attribute nsISHEntry parent;

pub unsafe fn GetLoadType(&self, aLoadType: *mut uint32_t) -> nsresult

/**
     * The loadType for this entry. This is typically loadHistory except
     * when reload is pressed, it has the appropriate reload flag
     */

attribute unsigned long loadType;

pub unsafe fn SetLoadType(&self, aLoadType: uint32_t) -> nsresult

/**
     * The loadType for this entry. This is typically loadHistory except
     * when reload is pressed, it has the appropriate reload flag
     */

attribute unsigned long loadType;

pub unsafe fn GetID(&self, aID: *mut uint32_t) -> nsresult

/**
     * An ID to help identify this entry from others during
     * subframe navigation
     */

attribute unsigned long ID;

pub unsafe fn SetID(&self, aID: uint32_t) -> nsresult

/**
     * An ID to help identify this entry from others during
     * subframe navigation
     */

attribute unsigned long ID;

pub unsafe fn GetCacheKey(&self, aCacheKey: *mut *const nsISupports) -> nsresult

/** attribute to set and get the cache key for the entry */

attribute nsISupports cacheKey;

pub unsafe fn SetCacheKey(&self, aCacheKey: *const nsISupports) -> nsresult

/** attribute to set and get the cache key for the entry */

attribute nsISupports cacheKey;

pub unsafe fn GetSaveLayoutStateFlag(
&self,
aSaveLayoutStateFlag: *mut bool
) -> nsresult

/** attribute to indicate whether layoutHistoryState should be saved */

attribute boolean saveLayoutStateFlag;

pub unsafe fn SetSaveLayoutStateFlag(
&self,
aSaveLayoutStateFlag: bool
) -> nsresult

/** attribute to indicate whether layoutHistoryState should be saved */

attribute boolean saveLayoutStateFlag;

pub unsafe fn GetExpirationStatus(
&self,
aExpirationStatus: *mut bool
) -> nsresult

/** attribute to indicate whether the page is already expired in cache */

attribute boolean expirationStatus;

pub unsafe fn SetExpirationStatus(&self, aExpirationStatus: bool) -> nsresult

/** attribute to indicate whether the page is already expired in cache */

attribute boolean expirationStatus;

pub unsafe fn GetContentType(&self, aContentType: &mut nsACString) -> nsresult

/**
     * attribute to indicate the content-type of the document that this
     * is a session history entry for
     */

attribute ACString contentType;

pub unsafe fn SetContentType(&self, aContentType: &nsACString) -> nsresult

/**
     * attribute to indicate the content-type of the document that this
     * is a session history entry for
     */

attribute ACString contentType;

pub unsafe fn GetURIWasModified(&self, aURIWasModified: *mut bool) -> nsresult

/**
     * If we created this SHEntry via history.pushState or modified it via
     * history.replaceState, and if we changed the SHEntry's URI via the
     * push/replaceState call, and if the SHEntry's new URI differs from its
     * old URI by more than just the hash, then we set this field to true.
     *
     * Additionally, if this SHEntry was created by calling pushState from a
     * SHEntry whose URI was modified, this SHEntry's URIWasModified field is
     * true.
     *
     */

attribute boolean URIWasModified;

pub unsafe fn SetURIWasModified(&self, aURIWasModified: bool) -> nsresult

/**
     * If we created this SHEntry via history.pushState or modified it via
     * history.replaceState, and if we changed the SHEntry's URI via the
     * push/replaceState call, and if the SHEntry's new URI differs from its
     * old URI by more than just the hash, then we set this field to true.
     *
     * Additionally, if this SHEntry was created by calling pushState from a
     * SHEntry whose URI was modified, this SHEntry's URIWasModified field is
     * true.
     *
     */

attribute boolean URIWasModified;

pub unsafe fn SetScrollPosition(&self, x: int32_t, y: int32_t) -> nsresult

/** Set/Get scrollers' positon in anchored pages */

void setScrollPosition (in long x, in long y);

pub unsafe fn GetScrollPosition(
&self,
x: *mut int32_t,
y: *mut int32_t
) -> nsresult

void getScrollPosition (out long x, out long y);

pub unsafe fn Create(
&self,
URI: *const nsIURI,
title: &nsAString,
inputStream: *const nsIInputStream,
layoutHistoryState: *const nsILayoutHistoryState,
cacheKey: *const nsISupports,
contentType: &nsACString,
triggeringPrincipal: *const nsIPrincipal,
principalToInherit: *const nsIPrincipal,
docshellID: &nsID,
dynamicCreation: bool
) -> nsresult

/** Additional ways to create an entry */

[noscript] void create (in nsIURI URI, in AString title, in nsIInputStream inputStream, in nsILayoutHistoryState layoutHistoryState, in nsISupports cacheKey, in ACString contentType, in nsIPrincipal triggeringPrincipal, in nsIPrincipal principalToInherit, in nsIDRef docshellID, in boolean dynamicCreation);

pub unsafe fn Clone(&self, _retval: *mut *const nsISHEntry) -> nsresult

nsISHEntry clone ();

pub unsafe fn SetIsSubFrame(&self, aFlag: bool) -> nsresult

/** Attribute that indicates if this entry is for a subframe navigation */

void setIsSubFrame (in boolean aFlag);

pub unsafe fn GetAnyContentViewer(
&self,
ownerEntry: *mut *const nsISHEntry,
_retval: *mut *const nsIContentViewer
) -> nsresult

/** Return any content viewer present in or below this node in the
        nsSHEntry tree.  This will differ from contentViewer in the case
        where a child nsSHEntry has the content viewer for this tree. */

nsIContentViewer getAnyContentViewer (out nsISHEntry ownerEntry);

pub unsafe fn GetTriggeringPrincipal(
&self,
aTriggeringPrincipal: *mut *const nsIPrincipal
) -> nsresult

/**
     * Get the principal, if any, that was associated with the channel
     * that the document that was loaded to create this history entry
     * came from.
     */

attribute nsIPrincipal triggeringPrincipal;

pub unsafe fn SetTriggeringPrincipal(
&self,
aTriggeringPrincipal: *const nsIPrincipal
) -> nsresult

/**
     * Get the principal, if any, that was associated with the channel
     * that the document that was loaded to create this history entry
     * came from.
     */

attribute nsIPrincipal triggeringPrincipal;

pub unsafe fn GetPrincipalToInherit(
&self,
aPrincipalToInherit: *mut *const nsIPrincipal
) -> nsresult

/**
     * Get the principal, if any, that is used when the inherit flag
     * is set.
     */

attribute nsIPrincipal principalToInherit;

pub unsafe fn SetPrincipalToInherit(
&self,
aPrincipalToInherit: *const nsIPrincipal
) -> nsresult

/**
     * Get the principal, if any, that is used when the inherit flag
     * is set.
     */

attribute nsIPrincipal principalToInherit;

pub unsafe fn GetStateData(
&self,
aStateData: *mut *const nsIStructuredCloneContainer
) -> nsresult

/**
     * Get/set data associated with this history state via a pushState() call,
     * serialized using structured clone.
     **/

attribute nsIStructuredCloneContainer stateData;

pub unsafe fn SetStateData(
&self,
aStateData: *const nsIStructuredCloneContainer
) -> nsresult

/**
     * Get/set data associated with this history state via a pushState() call,
     * serialized using structured clone.
     **/

attribute nsIStructuredCloneContainer stateData;

pub unsafe fn HasDetachedEditor(&self) -> bool

/** Returns true if this shistory entry is storing a detached editor. */

[noscript,notxpcom] boolean hasDetachedEditor ();

pub unsafe fn IsDynamicallyAdded(&self, _retval: *mut bool) -> nsresult

/**
     * Returns true if the related docshell was added because of
     * dynamic addition of an iframe/frame.
     */

boolean isDynamicallyAdded ();

pub unsafe fn HasDynamicallyAddedChild(&self, _retval: *mut bool) -> nsresult

/**
     * Returns true if any of the child entries returns true
     * when isDynamicallyAdded is called on it.
     */

boolean hasDynamicallyAddedChild ();

pub unsafe fn GetDocshellID(&self, aDocshellID: *mut *mut nsID) -> nsresult

/**
     * The history ID of the docshell.
     */

attribute nsIDPtr docshellID;

pub unsafe fn SetDocshellID(&self, aDocshellID: *const nsID) -> nsresult

/**
     * The history ID of the docshell.
     */

attribute nsIDPtr docshellID;

pub unsafe fn DocshellID(&self) -> nsID

/**
     * Helper method for accessing this value from C++
     */

[noscript,notxpcom] nsID DocshellID ();

pub unsafe fn GetBFCacheEntry(
&self,
aBFCacheEntry: *mut *const nsIBFCacheEntry
) -> nsresult

readonly attribute nsIBFCacheEntry BFCacheEntry;

pub unsafe fn HasBFCacheEntry(&self, aEntry: *const nsIBFCacheEntry) -> bool

/**
     * Does this SHEntry point to the given BFCache entry?  If so, evicting
     * the BFCache entry will evict the SHEntry, since the two entries
     * correspond to the same document.
     */

[noscript,notxpcom] boolean hasBFCacheEntry (in nsIBFCacheEntry aEntry);

pub unsafe fn AdoptBFCacheEntry(&self, aEntry: *const nsISHEntry) -> nsresult

/**
     * Adopt aEntry's BFCacheEntry, so now both this and aEntry point to
     * aEntry's BFCacheEntry.
     */

void adoptBFCacheEntry (in nsISHEntry aEntry);

pub unsafe fn AbandonBFCacheEntry(&self) -> nsresult

/**
     * Create a new BFCache entry and drop our reference to our old one.  This
     * call unlinks this SHEntry from any other SHEntries for its document.
     */

void abandonBFCacheEntry ();

pub unsafe fn SharesDocumentWith(
&self,
aEntry: *const nsISHEntry,
_retval: *mut bool
) -> nsresult

/**
     * Does this SHEntry correspond to the same document as aEntry?  This is
     * true iff the two SHEntries have the same BFCacheEntry.  So in
     * particular, sharesDocumentWith(aEntry) is guaranteed to return true if
     * it's preceeded by a call to adoptBFCacheEntry(aEntry).
     */

boolean sharesDocumentWith (in nsISHEntry aEntry);

pub unsafe fn GetIsSrcdocEntry(&self, aIsSrcdocEntry: *mut bool) -> nsresult

/**
     * True if this SHEntry corresponds to a document created by a srcdoc iframe.
     * Set when a value is assigned to  srcdocData.
     */

readonly attribute boolean isSrcdocEntry;

pub unsafe fn GetSrcdocData(&self, aSrcdocData: &mut nsAString) -> nsresult

/**
     * Contents of the srcdoc attribute in a srcdoc iframe to be loaded instead
     * of the URI.  Similar to a Data URI, this information is needed to
     * recreate the document at a later stage.
     * Setting this sets isSrcdocEntry to true
     */

attribute AString srcdocData;

pub unsafe fn SetSrcdocData(&self, aSrcdocData: &nsAString) -> nsresult

/**
     * Contents of the srcdoc attribute in a srcdoc iframe to be loaded instead
     * of the URI.  Similar to a Data URI, this information is needed to
     * recreate the document at a later stage.
     * Setting this sets isSrcdocEntry to true
     */

attribute AString srcdocData;

pub unsafe fn GetBaseURI(&self, aBaseURI: *mut *const nsIURI) -> nsresult

/**
     * When isSrcdocEntry is true, this contains the baseURI of the srcdoc
     * document for use in situations where it cannot otherwise be determined,
     * for example with view-source.
     */

attribute nsIURI baseURI;

pub unsafe fn SetBaseURI(&self, aBaseURI: *const nsIURI) -> nsresult

/**
     * When isSrcdocEntry is true, this contains the baseURI of the srcdoc
     * document for use in situations where it cannot otherwise be determined,
     * for example with view-source.
     */

attribute nsIURI baseURI;

pub unsafe fn GetScrollRestorationIsManual(
&self,
aScrollRestorationIsManual: *mut bool
) -> nsresult

/**
     * Sets/gets the current scroll restoration state,
     * if true == "manual", false == "auto".
     */

attribute boolean scrollRestorationIsManual;

pub unsafe fn SetScrollRestorationIsManual(
&self,
aScrollRestorationIsManual: bool
) -> nsresult

/**
     * Sets/gets the current scroll restoration state,
     * if true == "manual", false == "auto".
     */

attribute boolean scrollRestorationIsManual;

pub unsafe fn GetLoadedInThisProcess(
&self,
aLoadedInThisProcess: *mut bool
) -> nsresult

/**
     * Flag to indicate that the history entry was originally loaded in the
     * current process. This flag does not survive a browser process switch.
     */

readonly attribute boolean loadedInThisProcess;

pub unsafe fn GetSHistory(&self, aSHistory: *mut *const nsISHistory) -> nsresult

/**
     * The session history it belongs to. It's usually only set on root entries.
     * SHEntry is strictly bound to the SHistory it belongs to; it should not be
     * changed once set to a non-null value.
     */

[noscript] attribute nsISHistory SHistory;

pub unsafe fn SetSHistory(&self, aSHistory: *const nsISHistory) -> nsresult

/**
     * The session history it belongs to. It's usually only set on root entries.
     * SHEntry is strictly bound to the SHistory it belongs to; it should not be
     * changed once set to a non-null value.
     */

[noscript] attribute nsISHistory SHistory;