Struct xpcom::interfaces::nsISHistoryInternal

#[repr(C)]

pub struct nsISHistoryInternal { /* fields omitted */ }

interface nsISHistoryInternal : nsISupports

Methods

impl nsISHistoryInternal

pub fn coerce<T: nsISHistoryInternalCoerce>(&self) -> &T

Cast this nsISHistoryInternal to one of its base interfaces.

impl nsISHistoryInternal

pub unsafe fn AddEntry(
&self,
aEntry: *const nsISHEntry,
aPersist: bool
) -> nsresult

/**
   * Add a new Entry to the History List
   * @param aEntry - The entry to add
   * @param aPersist - If true this specifies that the entry should persist
   * in the list.  If false, this means that when new entries are added
   * this element will not appear in the session history list.
   */

void addEntry (in nsISHEntry aEntry, in boolean aPersist);

pub unsafe fn GetRootTransaction(
&self,
aRootTransaction: *mut *const nsISHTransaction
) -> nsresult

/**
   * Get the root transaction
   */

readonly attribute nsISHTransaction rootTransaction;

pub unsafe fn SetRootDocShell(
&self,
rootDocShell: *const nsIDocShell
) -> nsresult

/**
   * Sets the toplevel docshell object to which this SHistory object belongs to.
   */

void setRootDocShell (in nsIDocShell rootDocShell);

pub unsafe fn UpdateIndex(&self) -> nsresult

/**
   * Update the index maintained by sessionHistory
   */

void updateIndex ();

pub unsafe fn ReplaceEntry(
&self,
aIndex: int32_t,
aReplaceEntry: *const nsISHEntry
) -> nsresult

/**
   * Replace the nsISHEntry at a particular index
   * @param aIndex - The index at which the entry should be replaced
   * @param aReplaceEntry - The replacement entry for the index.
   */

void replaceEntry (in long aIndex, in nsISHEntry aReplaceEntry);

pub unsafe fn NotifyOnHistoryReload(
&self,
aReloadURI: *const nsIURI,
aReloadFlags: uint32_t,
_retval: *mut bool
) -> nsresult

/**
   * Notifies all registered session history listeners about an impending
   * reload.
   *
   * @param aReloadURI    The URI of the document to be reloaded.
   * @param aReloadFlags  Flags that indicate how the document is to be
   *                      refreshed. See constants on the nsIWebNavigation
   *                      interface.
   * @return              Whether the operation can proceed.
   */

boolean notifyOnHistoryReload (in nsIURI aReloadURI, in unsigned long aReloadFlags);

pub unsafe fn EvictOutOfRangeContentViewers(&self, aIndex: int32_t) -> nsresult

/**
    * Evict content viewers which don't lie in the "safe" range around aIndex.
    * In practice, this should leave us with no more than gHistoryMaxViewers
    * viewers associated with this SHistory object.
    *
    * Also make sure that the total number of content viewers in all windows is
    * not greater than our global max; if it is, evict viewers as appropriate.
    *
    * @param aIndex - The index around which the "safe" range is centered.  In
    *   general, if you just navigated the history, aIndex should be the index
    *   history was navigated to.
    */

void evictOutOfRangeContentViewers (in long aIndex);

pub unsafe fn EvictExpiredContentViewerForEntry(
&self,
aEntry: *const nsIBFCacheEntry
) -> nsresult

/**
    * Evict the content viewer associated with a bfcache entry
    * that has timed out.
    */

void evictExpiredContentViewerForEntry (in nsIBFCacheEntry aEntry);

pub unsafe fn EvictAllContentViewers(&self) -> nsresult

/**
    * Evict all the content viewers in this session history
    */

void evictAllContentViewers ();

pub unsafe fn AddToExpirationTracker(
&self,
aEntry: *const nsIBFCacheEntry
) -> nsresult

/**
    * Add a BFCache entry to expiration tracker so it gets evicted on expiration.
    */

void addToExpirationTracker (in nsIBFCacheEntry aEntry);

pub unsafe fn RemoveFromExpirationTracker(
&self,
aEntry: *const nsIBFCacheEntry
) -> nsresult

/**
    * Remove a BFCache entry from expiration tracker.
    */

void removeFromExpirationTracker (in nsIBFCacheEntry aEntry);

pub unsafe fn RemoveDynEntries(
&self,
aIndex: int32_t,
aContainer: *const nsISHContainer
) -> c_void

/**
    * Remove dynamic entries found at given index.
    *
    * @param aIndex
    *        Index to remove dynamic entries from. It will be passed to
    *        RemoveEntries as aStartIndex.
    * @param aContainer (optional)
    *        The container to start looking for dynamic entries. Only the
    *        dynamic descendants of the container will be removed. If not given,
    *        all dynamic entries at the index will be removed.
    */

[noscript,notxpcom] void RemoveDynEntries (in long aIndex, in nsISHContainer aContainer);

pub unsafe fn RemoveDynEntriesForBFCacheEntry(
&self,
aEntry: *const nsIBFCacheEntry
) -> c_void

/**
    * Similar to RemoveDynEntries, but instead of specifying an index, use the
    * given BFCacheEntry to find the index and remove dynamic entries from the
    * index.
    *
    * The method takes no effect if the bfcache entry is not or no longer hold
    * by the SHistory instance.
    *
    * @param aEntry
    *        The bfcache entry to look up for index to remove dynamic entries
    *        from.
    */

[noscript,notxpcom] void RemoveDynEntriesForBFCacheEntry (in nsIBFCacheEntry aEntry);