Struct xpcom::interfaces::nsISSLSocketControl

#[repr(C)]

pub struct nsISSLSocketControl { /* fields omitted */ }

interface nsISSLSocketControl : nsISupports

Methods

impl nsISSLSocketControl

pub fn coerce<T: nsISSLSocketControlCoerce>(&self) -> &T

Cast this nsISSLSocketControl to one of its base interfaces.

impl nsISSLSocketControl

pub const KEY_EXCHANGE_UNKNOWN: i64

KEY_EXCHANGE_UNKNOWN: i64 = -1

pub const SSL_VERSION_3: i64

SSL_VERSION_3: i64 = 768

pub const TLS_VERSION_1: i64

TLS_VERSION_1: i64 = 769

pub const TLS_VERSION_1_1: i64

TLS_VERSION_1_1: i64 = 770

pub const TLS_VERSION_1_2: i64

TLS_VERSION_1_2: i64 = 771

pub const TLS_VERSION_1_3: i64

TLS_VERSION_1_3: i64 = 772

pub const SSL_VERSION_UNKNOWN: i64

SSL_VERSION_UNKNOWN: i64 = -1

pub const SSL_MAC_UNKNOWN: i64

SSL_MAC_UNKNOWN: i64 = -1

pub const SSL_MAC_NULL: i64

SSL_MAC_NULL: i64 = 0

pub const SSL_MAC_MD5: i64

SSL_MAC_MD5: i64 = 1

pub const SSL_MAC_SHA: i64

SSL_MAC_SHA: i64 = 2

pub const SSL_HMAC_MD5: i64

SSL_HMAC_MD5: i64 = 3

pub const SSL_HMAC_SHA: i64

SSL_HMAC_SHA: i64 = 4

pub const SSL_HMAC_SHA256: i64

SSL_HMAC_SHA256: i64 = 5

pub const SSL_MAC_AEAD: i64

SSL_MAC_AEAD: i64 = 6

pub unsafe fn GetNotificationCallbacks(
&self,
aNotificationCallbacks: *mut *const nsIInterfaceRequestor
) -> nsresult

attribute nsIInterfaceRequestor notificationCallbacks;

pub unsafe fn SetNotificationCallbacks(
&self,
aNotificationCallbacks: *const nsIInterfaceRequestor
) -> nsresult

attribute nsIInterfaceRequestor notificationCallbacks;

pub unsafe fn ProxyStartSSL(&self) -> nsresult

void proxyStartSSL ();

pub unsafe fn StartTLS(&self) -> nsresult

void StartTLS ();

pub unsafe fn GetNegotiatedNPN(
&self,
aNegotiatedNPN: &mut nsACString
) -> nsresult

readonly attribute ACString negotiatedNPN;

pub unsafe fn GetAlpnEarlySelection(&self, _retval: &mut nsACString) -> nsresult

ACString getAlpnEarlySelection ();

pub unsafe fn GetEarlyDataAccepted(
&self,
aEarlyDataAccepted: *mut bool
) -> nsresult

readonly attribute bool earlyDataAccepted;

pub unsafe fn DriveHandshake(&self) -> nsresult

void driveHandshake ();

pub unsafe fn JoinConnection(
&self,
npnProtocol: &nsACString,
hostname: &nsACString,
port: int32_t,
_retval: *mut bool
) -> nsresult

boolean joinConnection (in ACString npnProtocol, in ACString hostname, in long port);

pub unsafe fn TestJoinConnection(
&self,
npnProtocol: &nsACString,
hostname: &nsACString,
port: int32_t,
_retval: *mut bool
) -> nsresult

boolean testJoinConnection (in ACString npnProtocol, in ACString hostname, in long port);

pub unsafe fn IsAcceptableForHost(
&self,
hostname: &nsACString,
_retval: *mut bool
) -> nsresult

boolean isAcceptableForHost (in ACString hostname);

pub unsafe fn GetKEAUsed(&self) -> int16_t

[infallible] readonly attribute short KEAUsed;

pub unsafe fn GetKEAKeyBits(&self) -> uint32_t

[infallible] readonly attribute unsigned long KEAKeyBits;

pub unsafe fn GetProviderFlags(&self, aProviderFlags: *mut uint32_t) -> nsresult

readonly attribute uint32_t providerFlags;

pub unsafe fn GetProviderTlsFlags(
&self,
aProviderTlsFlags: *mut uint32_t
) -> nsresult

readonly attribute uint32_t providerTlsFlags;

pub unsafe fn GetSSLVersionUsed(&self) -> int16_t

[infallible] readonly attribute short SSLVersionUsed;

pub unsafe fn GetSSLVersionOffered(&self) -> int16_t

[infallible] readonly attribute short SSLVersionOffered;

pub unsafe fn GetMACAlgorithmUsed(&self) -> int16_t

[infallible] readonly attribute short MACAlgorithmUsed;

pub unsafe fn GetClientCert(
&self,
aClientCert: *mut *const nsIX509Cert
) -> nsresult

/**
     * If set before the server requests a client cert (assuming it does so at
     * all), then this cert will be presented to the server, instead of asking
     * the user or searching the set of rememebered user cert decisions.
     */

attribute nsIX509Cert clientCert;

pub unsafe fn SetClientCert(&self, aClientCert: *const nsIX509Cert) -> nsresult

/**
     * If set before the server requests a client cert (assuming it does so at
     * all), then this cert will be presented to the server, instead of asking
     * the user or searching the set of rememebered user cert decisions.
     */

attribute nsIX509Cert clientCert;

pub unsafe fn GetBypassAuthentication(&self) -> bool

/**
     * bypassAuthentication is true if the server certificate checks are
     * not be enforced. This is to enable non-secure transport over TLS.
     */

[infallible] readonly attribute boolean bypassAuthentication;

pub unsafe fn GetFailedVerification(&self) -> bool

[infallible] readonly attribute boolean failedVerification;