Struct xpcom::interfaces::nsIScriptError

#[repr(C)]

pub struct nsIScriptError { /* fields omitted */ }

interface nsIScriptError : nsIConsoleMessage

Methods

impl nsIScriptError

pub fn coerce<T: nsIScriptErrorCoerce>(&self) -> &T

Cast this nsIScriptError to one of its base interfaces.

impl nsIScriptError

pub const errorFlag: i64

errorFlag: i64 = 0

/** pseudo-flag for default case */

pub const warningFlag: i64

warningFlag: i64 = 1

/** message is warning */

pub const exceptionFlag: i64

exceptionFlag: i64 = 2

/** exception was thrown for this case - exception-aware hosts can ignore */

pub const strictFlag: i64

strictFlag: i64 = 4

/** error or warning is due to strict option */

pub const infoFlag: i64

infoFlag: i64 = 8

/** just a log message */

pub unsafe fn GetErrorMessage(&self, aErrorMessage: &mut nsAString) -> nsresult

/**
     * The error message without any context/line number information.
     *
     * @note nsIConsoleMessage.message will return the error formatted
     *       with file/line information.
     */

readonly attribute AString errorMessage;

pub unsafe fn GetSourceName(&self, aSourceName: &mut nsAString) -> nsresult

readonly attribute AString sourceName;

pub unsafe fn GetSourceLine(&self, aSourceLine: &mut nsAString) -> nsresult

readonly attribute AString sourceLine;

pub unsafe fn GetLineNumber(&self, aLineNumber: *mut uint32_t) -> nsresult

readonly attribute uint32_t lineNumber;

pub unsafe fn GetColumnNumber(&self, aColumnNumber: *mut uint32_t) -> nsresult

readonly attribute uint32_t columnNumber;

pub unsafe fn GetFlags(&self, aFlags: *mut uint32_t) -> nsresult

readonly attribute uint32_t flags;

pub unsafe fn GetCategory(&self, aCategory: *mut *const c_char) -> nsresult

/**
     * Categories I know about -
     * XUL javascript
     * content javascript (both of these from nsDocShell, currently)
     * system javascript (errors in JS components and other system JS)
     */

readonly attribute string category;

pub unsafe fn GetOuterWindowID(&self, aOuterWindowID: *mut uint64_t) -> nsresult

readonly attribute unsigned long long outerWindowID;

pub unsafe fn GetInnerWindowID(&self, aInnerWindowID: *mut uint64_t) -> nsresult

readonly attribute unsigned long long innerWindowID;

pub unsafe fn GetIsFromPrivateWindow(
&self,
aIsFromPrivateWindow: *mut bool
) -> nsresult

readonly attribute boolean isFromPrivateWindow;

pub unsafe fn GetErrorMessageName(
&self,
aErrorMessageName: &mut nsAString
) -> nsresult

/**
     * The name of a template string, as found in js.msg, associated with the
     * error message.
     */

attribute AString errorMessageName;

pub unsafe fn SetErrorMessageName(
&self,
aErrorMessageName: &nsAString
) -> nsresult

/**
     * The name of a template string, as found in js.msg, associated with the
     * error message.
     */

attribute AString errorMessageName;

pub unsafe fn GetNotes(&self, aNotes: *mut *const nsIArray) -> nsresult

readonly attribute nsIArray notes;

pub unsafe fn Init(
&self,
message: &nsAString,
sourceName: &nsAString,
sourceLine: &nsAString,
lineNumber: uint32_t,
columnNumber: uint32_t,
flags: uint32_t,
category: *const c_char
) -> nsresult

void init (in AString message, in AString sourceName, in AString sourceLine, in uint32_t lineNumber, in uint32_t columnNumber, in uint32_t flags, in string category);

pub unsafe fn InitWithWindowID(
&self,
message: &nsAString,
sourceName: &nsAString,
sourceLine: &nsAString,
lineNumber: uint32_t,
columnNumber: uint32_t,
flags: uint32_t,
category: &nsACString,
innerWindowID: uint64_t
) -> nsresult

void initWithWindowID (in AString message, in AString sourceName, in AString sourceLine, in uint32_t lineNumber, in uint32_t columnNumber, in uint32_t flags, in ACString category, in unsigned long long innerWindowID);

pub unsafe fn InitWithSanitizedSource(
&self,
message: &nsAString,
sourceName: &nsAString,
sourceLine: &nsAString,
lineNumber: uint32_t,
columnNumber: uint32_t,
flags: uint32_t,
category: &nsACString,
innerWindowID: uint64_t
) -> nsresult

void initWithSanitizedSource (in AString message, in AString sourceName, in AString sourceLine, in uint32_t lineNumber, in uint32_t columnNumber, in uint32_t flags, in ACString category, in unsigned long long innerWindowID);

pub unsafe fn InitWithSourceURI(
&self,
message: &nsAString,
sourceURI: *const nsIURI,
sourceLine: &nsAString,
lineNumber: uint32_t,
columnNumber: uint32_t,
flags: uint32_t,
category: &nsACString,
innerWindowID: uint64_t
) -> nsresult

void initWithSourceURI (in AString message, in nsIURI sourceURI, in AString sourceLine, in uint32_t lineNumber, in uint32_t columnNumber, in uint32_t flags, in ACString category, in unsigned long long innerWindowID);