Struct xpcom::interfaces::nsIScriptLoaderObserver

#[repr(C)]

pub struct nsIScriptLoaderObserver { /* fields omitted */ }

interface nsIScriptLoaderObserver : nsISupports

Methods

impl nsIScriptLoaderObserver

pub fn coerce<T: nsIScriptLoaderObserverCoerce>(&self) -> &T

Cast this nsIScriptLoaderObserver to one of its base interfaces.

impl nsIScriptLoaderObserver

pub unsafe fn ScriptAvailable(
&self,
aResult: nsresult,
aElement: *const nsIScriptElement,
aIsInlineClassicScript: bool,
aURI: *const nsIURI,
aLineNo: int32_t
) -> nsresult

/**
   * The script is available for evaluation. For inline scripts, this
   * method will be called synchronously. For externally loaded scripts,
   * this method will be called when the load completes.
   *
   * @param aResult A result code representing the result of loading
   *        a script. If this is a failure code, script evaluation
   *        will not occur.
   * @param aElement The element being processed.
   * @param aIsInline Is this an inline classic script (as opposed to an
   *        externally loaded classic script or module script)?
   * @param aURI What is the URI of the script (the document URI if
   *        it is inline).
   * @param aLineNo At what line does the script appear (generally 1
   *        if it is a loaded script).
   */

void scriptAvailable (in nsresult aResult, in nsIScriptElement aElement, in boolean aIsInlineClassicScript, in nsIURI aURI, in int32_t aLineNo);

pub unsafe fn ScriptEvaluated(
&self,
aResult: nsresult,
aElement: *const nsIScriptElement,
aIsInline: bool
) -> nsresult

/**
   * The script has been evaluated.
   *
   * @param aResult A result code representing the success or failure of
   *        the script evaluation.
   * @param aElement The element being processed.
   * @param aIsInline Is this an inline script or externally loaded?
   */

void scriptEvaluated (in nsresult aResult, in nsIScriptElement aElement, in boolean aIsInline);