Struct xpcom::interfaces::nsIScriptableRegion

#[repr(C)]

pub struct nsIScriptableRegion { /* fields omitted */ }

interface nsIScriptableRegion : nsISupports

Methods

impl nsIScriptableRegion

pub fn coerce<T: nsIScriptableRegionCoerce>(&self) -> &T

Cast this nsIScriptableRegion to one of its base interfaces.

impl nsIScriptableRegion

pub unsafe fn Init(&self) -> nsresult

void init ();

pub unsafe fn SetToRegion(
&self,
aRegion: *const nsIScriptableRegion
) -> nsresult

/**
  * copy operator equivalent that takes another region
  *
  * @param      region to copy
  * @return     void
  *
  **/

void setToRegion (in nsIScriptableRegion aRegion);

pub unsafe fn SetToRect(
&self,
aX: int32_t,
aY: int32_t,
aWidth: int32_t,
aHeight: int32_t
) -> nsresult

/**
  * copy operator equivalent that takes a rect
  *
  * @param      aX xoffset of rect to set region to
  * @param      aY yoffset of rect to set region to
  * @param      aWidth width of rect to set region to
  * @param      aHeight height of rect to set region to
  * @return     void
  *
  **/

void setToRect (in long aX, in long aY, in long aWidth, in long aHeight);

pub unsafe fn IntersectRegion(
&self,
aRegion: *const nsIScriptableRegion
) -> nsresult

/**
  * destructively intersect another region with this one
  *
  * @param      region to intersect
  * @return     void
  *
  **/

void intersectRegion (in nsIScriptableRegion aRegion);

pub unsafe fn IntersectRect(
&self,
aX: int32_t,
aY: int32_t,
aWidth: int32_t,
aHeight: int32_t
) -> nsresult

/**
  * destructively intersect a rect with this region
  *
  * @param      aX xoffset of rect to intersect with region
  * @param      aY yoffset of rect to intersect with region
  * @param      aWidth width of rect to intersect with region
  * @param      aHeight height of rect to intersect with region
  * @return     void
  *
  **/

void intersectRect (in long aX, in long aY, in long aWidth, in long aHeight);

pub unsafe fn UnionRegion(
&self,
aRegion: *const nsIScriptableRegion
) -> nsresult

/**
  * destructively union another region with this one
  *
  * @param      region to union
  * @return     void
  *
  **/

void unionRegion (in nsIScriptableRegion aRegion);

pub unsafe fn UnionRect(
&self,
aX: int32_t,
aY: int32_t,
aWidth: int32_t,
aHeight: int32_t
) -> nsresult

/**
  * destructively union a rect with this region
  *
  * @param      aX xoffset of rect to union with region
  * @param      aY yoffset of rect to union with region
  * @param      aWidth width of rect to union with region
  * @param      aHeight height of rect to union with region
  * @return     void
  *
  **/

void unionRect (in long aX, in long aY, in long aWidth, in long aHeight);

pub unsafe fn SubtractRegion(
&self,
aRegion: *const nsIScriptableRegion
) -> nsresult

/**
  * destructively subtract another region with this one
  *
  * @param      region to subtract
  * @return     void
  *
  **/

void subtractRegion (in nsIScriptableRegion aRegion);

pub unsafe fn SubtractRect(
&self,
aX: int32_t,
aY: int32_t,
aWidth: int32_t,
aHeight: int32_t
) -> nsresult

/**
  * destructively subtract a rect from this region
  *
  * @param      aX xoffset of rect to subtract with region
  * @param      aY yoffset of rect to subtract with region
  * @param      aWidth width of rect to subtract with region
  * @param      aHeight height of rect to subtract with region
  * @return     void
  *
  **/

void subtractRect (in long aX, in long aY, in long aWidth, in long aHeight);

pub unsafe fn IsEmpty(&self, _retval: *mut bool) -> nsresult

/**
  * is this region empty? i.e. does it contain any pixels
  *
  * @param      none
  * @return     returns whether the region is empty
  *
  **/

boolean isEmpty ();

pub unsafe fn IsEqualRegion(
&self,
aRegion: *const nsIScriptableRegion,
_retval: *mut bool
) -> nsresult

/**
  * == operator equivalent i.e. do the regions contain exactly
  * the same pixels
  *
  * @param      region to compare
  * @return     whether the regions are identical
  *
  **/

boolean isEqualRegion (in nsIScriptableRegion aRegion);

pub unsafe fn GetBoundingBox(
&self,
aX: *mut int32_t,
aY: *mut int32_t,
aWidth: *mut int32_t,
aHeight: *mut int32_t
) -> nsresult

/**
  * returns the bounding box of the region i.e. the smallest
  * rectangle that completely contains the region.
  *
  * @param      aX out parameter for xoffset of bounding rect for region
  * @param      aY out parameter for yoffset of bounding rect for region
  * @param      aWidth out parameter for width of bounding rect for region
  * @param      aHeight out parameter for height of bounding rect for region
  * @return     void
  *
  **/

void getBoundingBox (out long aX, out long aY, out long aWidth, out long aHeight);

pub unsafe fn Offset(&self, aXOffset: int32_t, aYOffset: int32_t) -> nsresult

/**
  * offsets the region in x and y
  *
  * @param  xoffset  pixel offset in x
  * @param  yoffset  pixel offset in y
  * @return          void
  *
  **/

void offset (in long aXOffset, in long aYOffset);

pub unsafe fn ContainsRect(
&self,
aX: int32_t,
aY: int32_t,
aWidth: int32_t,
aHeight: int32_t,
_retval: *mut bool
) -> nsresult

/**
  * does the region intersect the rectangle?
  *
  * @param      rect to check for containment
  * @return     true if the region intersects the rect
  *
  **/

boolean containsRect (in long aX, in long aY, in long aWidth, in long aHeight);