Struct xpcom::interfaces::nsIScriptableUnicodeConverter

#[repr(C)]

pub struct nsIScriptableUnicodeConverter { /* fields omitted */ }

interface nsIScriptableUnicodeConverter : nsISupports

/**
 * This interface is a unicode encoder for use by scripts
 *
 * @created         8/Jun/2000
 * @author          Makoto Kato [m_kato@ga2.so-net.ne.jp]
 */

Methods

impl nsIScriptableUnicodeConverter

pub fn coerce<T: nsIScriptableUnicodeConverterCoerce>(&self) -> &T

Cast this nsIScriptableUnicodeConverter to one of its base interfaces.

impl nsIScriptableUnicodeConverter

pub unsafe fn ConvertFromUnicode(
&self,
aSrc: &nsAString,
_retval: &mut nsACString
) -> nsresult

/**
   * Converts the data from Unicode to one Charset.
   * Returns the converted string. After converting, Finish should be called
   * and its return value appended to this return value.
   */

ACString ConvertFromUnicode (in AString aSrc);

pub unsafe fn Finish(&self, _retval: &mut nsACString) -> nsresult

/**
   * Returns the terminator string.
   * Should be called after ConvertFromUnicode() and appended to that
   * function's return value.
   */

ACString Finish ();

pub unsafe fn ConvertToUnicode(
&self,
aSrc: &nsACString,
_retval: &mut nsAString
) -> nsresult

/**
   * Converts the data from one Charset to Unicode.
   */

AString ConvertToUnicode (in ACString aSrc);

pub unsafe fn ConvertFromByteArray(
&self,
aData: *const uint8_t,
aCount: uint32_t,
_retval: &mut nsAString
) -> nsresult

/**
   * Converts an array of bytes to a unicode string.
   */

AString convertFromByteArray ([array, size_is (aCount), const] in octet aData, in unsigned long aCount);

pub unsafe fn ConvertToByteArray(
&self,
aString: &nsAString,
aLen: *mut uint32_t,
aData: *mut *mut uint8_t
) -> nsresult

/**
   * Convert a unicode string to an array of bytes. Finish does not need to be
   * called.
   */

void convertToByteArray (in AString aString, [optional] out unsigned long aLen, [array, size_is (aLen), retval] out octet aData);

pub unsafe fn ConvertToInputStream(
&self,
aString: &nsAString,
_retval: *mut *const nsIInputStream
) -> nsresult

/**
   * Converts a unicode string to an input stream. The bytes in the stream are
   * encoded according to the charset attribute.
   * The returned stream will be nonblocking.
   */

nsIInputStream convertToInputStream (in AString aString);

pub unsafe fn GetCharset(&self, aCharset: &mut nsACString) -> nsresult

/**
   * Current character set.
   *
   * @throw NS_ERROR_UCONV_NOCONV
   *        The requested charset is not supported.
   */

attribute ACString charset;

pub unsafe fn SetCharset(&self, aCharset: &nsACString) -> nsresult

/**
   * Current character set.
   *
   * @throw NS_ERROR_UCONV_NOCONV
   *        The requested charset is not supported.
   */

attribute ACString charset;

pub unsafe fn GetIsInternal(&self, aIsInternal: *mut bool) -> nsresult

/**
   * Meaningless
   */

attribute boolean isInternal;

pub unsafe fn SetIsInternal(&self, aIsInternal: bool) -> nsresult

/**
   * Meaningless
   */

attribute boolean isInternal;