Struct xpcom::interfaces::nsIScrollable

#[repr(C)]

pub struct nsIScrollable { /* fields omitted */ }

interface nsIScrollable : nsISupports

/**
 * The nsIScrollable is an interface that can be implemented by a control that
 * supports scrolling.  This is a generic interface without concern for the
 * type of content that may be inside.
 */

Methods

impl nsIScrollable

pub fn coerce<T: nsIScrollableCoerce>(&self) -> &T

Cast this nsIScrollable to one of its base interfaces.

impl nsIScrollable

pub const ScrollOrientation_X: i64

ScrollOrientation_X: i64 = 1

/**
   * Constants declaring the two scroll orientations a scrollbar can be in.
   * ScrollOrientation_X - Horizontal scrolling.  When passing this
   *         in to a method you are requesting or setting data for the
   *         horizontal scrollbar.
   * ScrollOrientation_Y - Vertical scrolling.  When passing this
   *         in to a method you are requesting or setting data for the
   *         vertical scrollbar.
   */

pub const ScrollOrientation_Y: i64

ScrollOrientation_Y: i64 = 2

pub const Scrollbar_Auto: i64

Scrollbar_Auto: i64 = 1

/**
   * Constants declaring the states of the scrollbars.
   * ScrollPref_Auto - bars visible only when needed.
   * ScrollPref_Never - bars never visible, even when scrolling still possible.
   * ScrollPref_Always - bars always visible, even when scrolling is not possible
   */

pub const Scrollbar_Never: i64

Scrollbar_Never: i64 = 2

pub const Scrollbar_Always: i64

Scrollbar_Always: i64 = 3

pub unsafe fn GetDefaultScrollbarPreferences(
&self,
scrollOrientation: int32_t,
_retval: *mut int32_t
) -> nsresult

/**
   * Get or set the default scrollbar state for all documents in
   * this shell.
   */

long getDefaultScrollbarPreferences (in long scrollOrientation);

pub unsafe fn SetDefaultScrollbarPreferences(
&self,
scrollOrientation: int32_t,
scrollbarPref: int32_t
) -> nsresult

void setDefaultScrollbarPreferences (in long scrollOrientation, in long scrollbarPref);

pub unsafe fn GetScrollbarVisibility(
&self,
verticalVisible: *mut bool,
horizontalVisible: *mut bool
) -> nsresult

/**
   * Get information about whether the vertical and horizontal scrollbars are
   * currently visible.  If you are only interested in one of the visibility
   * settings pass nullptr in for the one you aren't interested in.
   */

void getScrollbarVisibility (out boolean verticalVisible, out boolean horizontalVisible);