Struct xpcom::interfaces::nsISelectionPrivate

#[repr(C)]

pub struct nsISelectionPrivate { /* fields omitted */ }

interface nsISelectionPrivate : nsISupports

Methods

impl nsISelectionPrivate

pub fn coerce<T: nsISelectionPrivateCoerce>(&self) -> &T

Cast this nsISelectionPrivate to one of its base interfaces.

impl nsISelectionPrivate

pub const ENDOFPRECEDINGLINE: i64

ENDOFPRECEDINGLINE: i64 = 0

pub const STARTOFNEXTLINE: i64

STARTOFNEXTLINE: i64 = 1

pub const TABLESELECTION_NONE: i64

TABLESELECTION_NONE: i64 = 0

pub const TABLESELECTION_CELL: i64

TABLESELECTION_CELL: i64 = 1

pub const TABLESELECTION_ROW: i64

TABLESELECTION_ROW: i64 = 2

pub const TABLESELECTION_COLUMN: i64

TABLESELECTION_COLUMN: i64 = 3

pub const TABLESELECTION_TABLE: i64

TABLESELECTION_TABLE: i64 = 4

pub const TABLESELECTION_ALLCELLS: i64

TABLESELECTION_ALLCELLS: i64 = 5

pub unsafe fn GetInterlinePosition(
&self,
aInterlinePosition: *mut bool
) -> nsresult

attribute boolean interlinePosition;

pub unsafe fn SetInterlinePosition(&self, aInterlinePosition: bool) -> nsresult

attribute boolean interlinePosition;

pub unsafe fn GetAncestorLimiter(
&self,
aAncestorLimiter: *mut *const nsIContent
) -> nsresult

[noscript] attribute nsIContent ancestorLimiter;

pub unsafe fn SetAncestorLimiter(
&self,
aAncestorLimiter: *const nsIContent
) -> nsresult

[noscript] attribute nsIContent ancestorLimiter;

pub unsafe fn ToStringWithFormat(
&self,
formatType: *const c_char,
flags: uint32_t,
wrapColumn: int32_t,
_retval: &mut nsAString
) -> nsresult

DOMString toStringWithFormat (in string formatType, in unsigned long flags, in int32_t wrapColumn);

pub unsafe fn AddSelectionListener(
&self,
newListener: *const nsISelectionListener
) -> nsresult

void addSelectionListener (in nsISelectionListener newListener);

pub unsafe fn RemoveSelectionListener(
&self,
listenerToRemove: *const nsISelectionListener
) -> nsresult

void removeSelectionListener (in nsISelectionListener listenerToRemove);

pub unsafe fn GetTableSelectionType(
&self,
range: *const nsIDOMRange,
_retval: *mut int32_t
) -> nsresult

/** Test if supplied range points to a single table element:
      *    Result is one of above constants. "None" means
      *    a table element isn't selected.
      */

[noscript] long getTableSelectionType (in nsIDOMRange range);

pub unsafe fn GetCanCacheFrameOffset(
&self,
aCanCacheFrameOffset: *mut bool
) -> nsresult

[noscript] attribute boolean canCacheFrameOffset;

pub unsafe fn SetCanCacheFrameOffset(
&self,
aCanCacheFrameOffset: bool
) -> nsresult

[noscript] attribute boolean canCacheFrameOffset;

pub unsafe fn GetType(&self, aType: *mut int16_t) -> nsresult

/**
     * Returns the type of the selection (see nsISelectionController for
     * available constants).
     */

readonly attribute short type;

pub unsafe fn GetRangesForInterval(
&self,
beginNode: *const nsIDOMNode,
beginOffset: int32_t,
endNode: *const nsIDOMNode,
endOffset: int32_t,
allowAdjacent: bool,
resultCount: *mut uint32_t,
results: *mut *mut *const nsIDOMRange
) -> nsresult

/**
     * Return array of ranges intersecting with the given DOM interval.
     */

void GetRangesForInterval (in nsIDOMNode beginNode, in int32_t beginOffset, in nsIDOMNode endNode, in int32_t endOffset, in boolean allowAdjacent, out uint32_t resultCount, [array, size_is (resultCount), retval] out nsIDOMRange results);

pub unsafe fn ScrollIntoView(
&self,
aRegion: int16_t,
aIsSynchronous: bool,
aVPercent: int16_t,
aHPercent: int16_t
) -> nsresult

/**
     * Scrolls a region of the selection, so that it is visible in
     * the scrolled view.
     *
     * @param aRegion - the region inside the selection to scroll into view
     *                  (see selection region constants defined in
     *                   nsISelectionController).
     * @param aIsSynchronous - when true, scrolls the selection into view
     *                         before returning. If false, posts a request which
     *                         is processed at some point after the method returns.
     * @param aVPercent - how to align the frame vertically.
     * @param aHPercent - how to align the frame horizontally.
     */

void scrollIntoView (in short aRegion, in boolean aIsSynchronous, in int16_t aVPercent, in int16_t aHPercent);

pub unsafe fn SelectionLanguageChange(&self, langRTL: bool) -> nsresult

/**
     * Modifies the cursor Bidi level after a change in keyboard direction
     * @param langRTL is PR_TRUE if the new language is right-to-left or
     *                PR_FALSE if the new language is left-to-right.
     */

[noscript] void selectionLanguageChange (in boolean langRTL);

pub unsafe fn SetColors(
&self,
aForegroundColor: &nsAString,
aBackgroundColor: &nsAString,
aAltForegroundColor: &nsAString,
aAltBackgroundColor: &nsAString
) -> nsresult

/**
     * setColors() sets custom colors for the selection.
     * Currently, this is supported only when the selection type is SELECTION_FIND.
     * Otherwise, throws an exception.
     *
     * @param aForegroundColor     The foreground color of the selection.
     *                             If this is "currentColor", foreground color
     *                             isn't changed by this selection.
     * @param aBackgroundColor     The background color of the selection.
     *                             If this is "transparent", background color is
     *                             never painted.
     * @param aAltForegroundColor  The alternative foreground color of the
     *                             selection.
     *                             If aBackgroundColor doesn't have sufficient
     *                             contrast with its around or foreground color
     *                             if "currentColor" is specified, alternative
     *                             colors are used if it have higher contrast.
     * @param aAltBackgroundColor  The alternative background color of the
     *                             selection.
     */

void setColors (in DOMString aForegroundColor, in DOMString aBackgroundColor, in DOMString aAltForegroundColor, in DOMString aAltBackgroundColor);

pub unsafe fn ResetColors(&self) -> nsresult

/**
     * resetColors() forget the customized colors which were set by setColors().
     */

void resetColors ();