Struct xpcom::interfaces::nsISerialEventTarget

#[repr(C)]

pub struct nsISerialEventTarget { /* fields omitted */ }

interface nsISerialEventTarget : nsIEventTarget

/**
 * A serial event target is an event dispatching interface like
 * nsIEventTarget. Runnables dispatched to an nsISerialEventTarget are required
 * to execute serially. That is, two different runnables dispatched to the
 * target should never be allowed to execute simultaneously. One exception to
 * this rule is nested event loops. If a runnable spins a nested event loop,
 * causing another runnable dispatched to the target to run, the target may
 * still be considered "serial".
 *
 * Examples:
 * - nsIThread is a serial event target.
 * - Thread pools are not serial event targets.
 * - However, one can "convert" a thread pool into an nsISerialEventTarget
 *   by putting a TaskQueue in front of it.
 */

Methods

impl nsISerialEventTarget

pub fn coerce<T: nsISerialEventTargetCoerce>(&self) -> &T

Cast this nsISerialEventTarget to one of its base interfaces.