Struct xpcom::interfaces::nsISessionStore

#[repr(C)]

pub struct nsISessionStore { /* fields omitted */ }

interface nsISessionStore : nsISupports

/**
 * nsISessionStore keeps track of the current browsing state - i.e.
 * tab history, cookies, scroll state, form data, and window features
 * - and allows to restore everything into one browser window.
 *
 * The nsISessionStore API operates mostly on browser windows and the tabbrowser
 * tabs contained in them:
 *
 * * "Browser windows" are those DOM windows having loaded
 * chrome://browser/content/browser.xul . From overlays you can just pass the
 * global |window| object to the API, though (or |top| from a sidebar).
 * From elsewhere you can get browser windows through the nsIWindowMediator
 * by looking for "navigator:browser" windows.
 *
 * * "Tabbrowser tabs" are all the child nodes of a browser window's
 * |gBrowser.tabContainer| such as e.g. |gBrowser.selectedTab|.
 */

Methods

impl nsISessionStore

pub fn coerce<T: nsISessionStoreCoerce>(&self) -> &T

Cast this nsISessionStore to one of its base interfaces.

impl nsISessionStore

pub unsafe fn GetCanRestoreLastSession(
&self,
aCanRestoreLastSession: *mut bool
) -> nsresult

/**
   * Is it possible to restore the previous session. Will always be false when
   * in Private Browsing mode.
   */

attribute boolean canRestoreLastSession;

pub unsafe fn SetCanRestoreLastSession(
&self,
aCanRestoreLastSession: bool
) -> nsresult

/**
   * Is it possible to restore the previous session. Will always be false when
   * in Private Browsing mode.
   */

attribute boolean canRestoreLastSession;

pub unsafe fn RestoreLastSession(&self) -> nsresult

/**
   * Restore the previous session if possible. This will not overwrite the
   * current session. Instead the previous session will be merged into the
   * current session. Current windows will be reused if they were windows that
   * pinned tabs were previously restored into. New windows will be opened as
   * needed.
   *
   * Note: This will throw if there is no previous state to restore. Check with
   * canRestoreLastSession first to avoid thrown errors.
   */

void restoreLastSession ();

pub unsafe fn GetBrowserState(&self, _retval: &mut nsAString) -> nsresult

/**
   * Get the current browsing state.
   * @returns a JSON string representing the session state.
   */

AString getBrowserState ();

pub unsafe fn SetBrowserState(&self, aState: &nsAString) -> nsresult

/**
   * Set the browsing state.
   * This will immediately restore the state of the whole application to the state
   * passed in, *replacing* the current session.
   *
   * @param aState is a JSON string representing the session state.
   */

void setBrowserState (in AString aState);

pub unsafe fn GetWindowState(
&self,
aWindow: *const nsIDOMWindow,
_retval: &mut nsAString
) -> nsresult

/**
   * @param aWindow is the browser window whose state is to be returned.
   *
   * @returns a JSON string representing a session state with only one window.
   */

AString getWindowState (in nsIDOMWindow aWindow);

pub unsafe fn SetWindowState(
&self,
aWindow: *const nsIDOMWindow,
aState: &nsAString,
aOverwrite: bool
) -> nsresult

/**
   * @param aWindow    is the browser window whose state is to be set.
   * @param aState     is a JSON string representing a session state.
   * @param aOverwrite boolean overwrite existing tabs
   */

void setWindowState (in nsIDOMWindow aWindow, in AString aState, in boolean aOverwrite);

pub unsafe fn GetTabState(
&self,
aTab: *const nsIDOMNode,
_retval: &mut nsAString
) -> nsresult

/**
   * @param aTab is the tabbrowser tab whose state is to be returned.
   *
   * @returns a JSON string representing the state of the tab
   *         (note: doesn't contain cookies - if you need them, use getWindowState instead).
   */

AString getTabState (in nsIDOMNode aTab);

pub unsafe fn SetTabState(
&self,
aTab: *const nsIDOMNode,
aState: &nsAString
) -> nsresult

/**
   * @param aTab   is the tabbrowser tab whose state is to be set.
   * @param aState is a JSON string representing a session state.
   */

void setTabState (in nsIDOMNode aTab, in AString aState);

pub unsafe fn DuplicateTab(
&self,
aWindow: *const nsIDOMWindow,
aTab: *const nsIDOMNode,
aDelta: int32_t,
_retval: *mut *const nsIDOMNode
) -> nsresult

/**
   * Duplicates a given tab as thoroughly as possible.
   *
   * @param aWindow is the browser window into which the tab will be duplicated.
   * @param aTab    is the tabbrowser tab to duplicate (can be from a different window).
   * @param aDelta  is the offset to the history entry to load in the duplicated tab.
   * @returns a reference to the newly created tab.
   */

nsIDOMNode duplicateTab (in nsIDOMWindow aWindow, in nsIDOMNode aTab, [optional] in long aDelta);

pub unsafe fn GetClosedTabCount(
&self,
aWindow: *const nsIDOMWindow,
_retval: *mut uint32_t
) -> nsresult

/**
   * Get the number of restore-able tabs for a browser window
   */

unsigned long getClosedTabCount (in nsIDOMWindow aWindow);

pub unsafe fn GetClosedTabData(
&self,
aWindow: *const nsIDOMWindow,
_retval: &mut nsAString
) -> nsresult

/**
   * Get closed tab data
   *
   * @param aWindow is the browser window for which to get closed tab data
   * @returns a JSON string representing the list of closed tabs.
   */

AString getClosedTabData (in nsIDOMWindow aWindow);

pub unsafe fn UndoCloseTab(
&self,
aWindow: *const nsIDOMWindow,
aIndex: uint32_t,
_retval: *mut *const nsIDOMNode
) -> nsresult

/**
   * @param aWindow is the browser window to reopen a closed tab in.
   * @param aIndex  is the index of the tab to be restored (FIFO ordered).
   * @returns a reference to the reopened tab.
   */

nsIDOMNode undoCloseTab (in nsIDOMWindow aWindow, in unsigned long aIndex);

pub unsafe fn ForgetClosedTab(
&self,
aWindow: *const nsIDOMWindow,
aIndex: uint32_t,
_retval: *mut *const nsIDOMNode
) -> nsresult

/**
   * @param aWindow is the browser window associated with the closed tab.
   * @param aIndex  is the index of the closed tab to be removed (FIFO ordered).
   */

nsIDOMNode forgetClosedTab (in nsIDOMWindow aWindow, in unsigned long aIndex);

pub unsafe fn GetClosedWindowCount(&self, _retval: *mut uint32_t) -> nsresult

/**
   * Get the number of restore-able windows
   */

unsigned long getClosedWindowCount ();

pub unsafe fn GetClosedWindowData(&self, _retval: &mut nsAString) -> nsresult

/**
   * Get closed windows data
   *
   * @returns a JSON string representing the list of closed windows.
   */

AString getClosedWindowData ();

pub unsafe fn UndoCloseWindow(
&self,
aIndex: uint32_t,
_retval: *mut *const nsIDOMWindow
) -> nsresult

/**
   * @param aIndex is the index of the windows to be restored (FIFO ordered).
   * @returns the nsIDOMWindow object of the reopened window
   */

nsIDOMWindow undoCloseWindow (in unsigned long aIndex);

pub unsafe fn ForgetClosedWindow(
&self,
aIndex: uint32_t,
_retval: *mut *const nsIDOMNode
) -> nsresult

/**
   * @param aIndex  is the index of the closed window to be removed (FIFO ordered).
   *
   * @throws NS_ERROR_INVALID_ARG
   *   when aIndex does not map to a closed window
   */

nsIDOMNode forgetClosedWindow (in unsigned long aIndex);

pub unsafe fn GetWindowValue(
&self,
aWindow: *const nsIDOMWindow,
aKey: &nsAString,
_retval: &mut nsAString
) -> nsresult

/**
   * @param aWindow is the window to get the value for.
   * @param aKey    is the value's name.
   *
   * @returns A string value or an empty string if none is set.
   */

AString getWindowValue (in nsIDOMWindow aWindow, in AString aKey);

pub unsafe fn DeleteWindowValue(
&self,
aWindow: *const nsIDOMWindow,
aKey: &nsAString
) -> nsresult

/**
   * @param aWindow is the browser window to get the value for.
   * @param aKey    is the value's name.
   */

void deleteWindowValue (in nsIDOMWindow aWindow, in AString aKey);

pub unsafe fn GetTabValue(
&self,
aTab: *const nsIDOMNode,
aKey: &nsAString,
_retval: &mut nsAString
) -> nsresult

/**
   * @param aTab is the tabbrowser tab to get the value for.
   * @param aKey is the value's name.
   *
   * @returns A string value or an empty string if none is set.
   */

AString getTabValue (in nsIDOMNode aTab, in AString aKey);

pub unsafe fn DeleteTabValue(
&self,
aTab: *const nsIDOMNode,
aKey: &nsAString
) -> nsresult

/**
   * @param aTab is the tabbrowser tab to get the value for.
   * @param aKey is the value's name.
   */

void deleteTabValue (in nsIDOMNode aTab, in AString aKey);

pub unsafe fn GetGlobalValue(
&self,
aKey: &nsAString,
_retval: &mut nsAString
) -> nsresult

/**
   * @param aKey is the value's name.
   *
   * @returns A string value or an empty string if none is set.
   */

AString getGlobalValue (in AString aKey);

pub unsafe fn DeleteGlobalValue(&self, aKey: &nsAString) -> nsresult

/**
   * @param aTab is the browser tab to get the value for.
   * @param aKey is the value's name.
   */

void deleteGlobalValue (in AString aKey);

pub unsafe fn PersistTabAttribute(&self, aName: &nsAString) -> nsresult

/**
   * @param aName is the name of the attribute to save/restore for all tabbrowser tabs.
   */

void persistTabAttribute (in AString aName);