Struct xpcom::interfaces::nsISimpleStreamListener

#[repr(C)]

pub struct nsISimpleStreamListener { /* fields omitted */ }

interface nsISimpleStreamListener : nsIStreamListener

/**
 * A simple stream listener can be used with AsyncRead to supply data to
 * a output stream.
 */

Methods

impl nsISimpleStreamListener

pub fn coerce<T: nsISimpleStreamListenerCoerce>(&self) -> &T

Cast this nsISimpleStreamListener to one of its base interfaces.

impl nsISimpleStreamListener

pub unsafe fn Init(
&self,
aSink: *const nsIOutputStream,
aObserver: *const nsIRequestObserver
) -> nsresult

/**
     * Initialize the simple stream listener.
     *
     * @param aSink data will be read from the channel to this output stream.
     *              Must implement writeFrom.
     * @param aObserver optional stream observer (can be NULL)
     */

void init (in nsIOutputStream aSink, in nsIRequestObserver aObserver);