Struct xpcom::interfaces::nsISiteSecurityService

#[repr(C)]

pub struct nsISiteSecurityService { /* fields omitted */ }

interface nsISiteSecurityService : nsISupports

Methods

impl nsISiteSecurityService

pub fn coerce<T: nsISiteSecurityServiceCoerce>(&self) -> &T

Cast this nsISiteSecurityService to one of its base interfaces.

impl nsISiteSecurityService

pub const HEADER_HSTS: i64

HEADER_HSTS: i64 = 0

pub const HEADER_HPKP: i64

HEADER_HPKP: i64 = 1

pub const HEADER_OMS: i64

HEADER_OMS: i64 = 2

pub const Success: i64

Success: i64 = 0

pub const ERROR_UNKNOWN: i64

ERROR_UNKNOWN: i64 = 1

pub const ERROR_UNTRUSTWORTHY_CONNECTION: i64

ERROR_UNTRUSTWORTHY_CONNECTION: i64 = 2

pub const ERROR_COULD_NOT_PARSE_HEADER: i64

ERROR_COULD_NOT_PARSE_HEADER: i64 = 3

pub const ERROR_NO_MAX_AGE: i64

ERROR_NO_MAX_AGE: i64 = 4

pub const ERROR_MULTIPLE_MAX_AGES: i64

ERROR_MULTIPLE_MAX_AGES: i64 = 5

pub const ERROR_INVALID_MAX_AGE: i64

ERROR_INVALID_MAX_AGE: i64 = 6

pub const ERROR_MULTIPLE_INCLUDE_SUBDOMAINS: i64

ERROR_MULTIPLE_INCLUDE_SUBDOMAINS: i64 = 7

pub const ERROR_INVALID_INCLUDE_SUBDOMAINS: i64

ERROR_INVALID_INCLUDE_SUBDOMAINS: i64 = 8

pub const ERROR_INVALID_PIN: i64

ERROR_INVALID_PIN: i64 = 9

pub const ERROR_MULTIPLE_REPORT_URIS: i64

ERROR_MULTIPLE_REPORT_URIS: i64 = 10

pub const ERROR_PINSET_DOES_NOT_MATCH_CHAIN: i64

ERROR_PINSET_DOES_NOT_MATCH_CHAIN: i64 = 11

pub const ERROR_NO_BACKUP_PIN: i64

ERROR_NO_BACKUP_PIN: i64 = 12

pub const ERROR_COULD_NOT_SAVE_STATE: i64

ERROR_COULD_NOT_SAVE_STATE: i64 = 13

pub const ERROR_ROOT_NOT_BUILT_IN: i64

ERROR_ROOT_NOT_BUILT_IN: i64 = 14

pub const SOURCE_UNKNOWN: i64

SOURCE_UNKNOWN: i64 = 0

/**
     * nsISiteSecurityService::IsSecureURI can optionally return a flag
     * indicating the source of the HSTS cache entry, if it comes from the
     * preload list, was seen naturally, or is a result of HSTS priming.
     */

pub const SOURCE_PRELOAD_LIST: i64

SOURCE_PRELOAD_LIST: i64 = 1

pub const SOURCE_ORGANIC_REQUEST: i64

SOURCE_ORGANIC_REQUEST: i64 = 2

pub const SOURCE_HSTS_PRIMING: i64

SOURCE_HSTS_PRIMING: i64 = 3

pub unsafe fn ClearAll(&self) -> nsresult

/**
     * Removes all non-preloaded security state by resetting to factory-original
     * settings.
     */

[must_use] void clearAll ();

pub unsafe fn ClearPreloads(&self) -> nsresult

/**
     * Removes all preloaded security state.
     */

[must_use] void clearPreloads ();

pub unsafe fn SetHSTSPreload(
&self,
aHost: &nsACString,
aIncludesSubdomains: bool,
aExpires: int64_t,
_retval: *mut bool
) -> nsresult

/**
     * Set an HSTS preload entry for a host. The resulting entries will be
     * permanent and visible from private and non-private contexts. These
     * entries replace any already set by this mechanism or those built-in to
     * Gecko.
     *
     * @param aHost the hostname (punycode) that the entry applies to
     * @param aIncludeSubdomains whether this entry also applies to subdomains
     * @param aExpires the time this entry should expire (millis since epoch)
     */

[must_use] boolean setHSTSPreload (in ACString aHost, in boolean aIncludesSubdomains, in int64_t aExpires);

pub unsafe fn Enumerate(
&self,
aType: uint32_t,
_retval: *mut *const nsISimpleEnumerator
) -> nsresult

/**
     * Returns an enumerator of the nsISiteSecurityService storage. Each item in
     * the enumeration is a nsISiteSecurityState that can be QueryInterfaced to
     * the appropriate nsISiteHSTSState or nsISiteHPKPState, depending on the
     * provided type. Doesn't include preloaded entries (either the hard-coded
     * ones or the preloaded-delivered-by-kinto ones).
     *
     * @param aType the type of security state in question.
     */

[must_use] nsISimpleEnumerator enumerate (in uint32_t aType);