Struct xpcom::interfaces::nsISliderListener

#[repr(C)]

pub struct nsISliderListener { /* fields omitted */ }

interface nsISliderListener : nsISupports

/**
 * Used for <scale> to listen to slider changes to avoid mutation listeners
 */

Methods

impl nsISliderListener

pub fn coerce<T: nsISliderListenerCoerce>(&self) -> &T

Cast this nsISliderListener to one of its base interfaces.

impl nsISliderListener

pub unsafe fn ValueChanged(
&self,
which: &nsAString,
newValue: int32_t,
userChanged: bool
) -> nsresult

/**
   * Called when the current, minimum or maximum value has been changed to
   * newValue. The which parameter will either be 'curpos', 'minpos' or 'maxpos'.
   * If userChanged is true, then the user changed ths slider, otherwise it
   * was changed via some other means.
   */

void valueChanged (in AString which, in long newValue, in boolean userChanged);

pub unsafe fn DragStateChanged(&self, isDragging: bool) -> nsresult

/**
   * Called when the user begins or ends dragging the thumb.
   */

void dragStateChanged (in boolean isDragging);