Struct xpcom::interfaces::nsISocketProvider

#[repr(C)]

pub struct nsISocketProvider { /* fields omitted */ }

interface nsISocketProvider : nsISupports

/**
 * nsISocketProvider
 */

Methods

impl nsISocketProvider

pub fn coerce<T: nsISocketProviderCoerce>(&self) -> &T

Cast this nsISocketProvider to one of its base interfaces.

impl nsISocketProvider

pub const PROXY_RESOLVES_HOST: i64

PROXY_RESOLVES_HOST: i64 = 1

/**
     * PROXY_RESOLVES_HOST
     *
     * This flag is set if the proxy is to perform hostname resolution instead
     * of the client.  When set, the hostname parameter passed when in this
     * interface will be used instead of the address structure passed for a
     * later connect et al. request.
     */

pub const ANONYMOUS_CONNECT: i64

ANONYMOUS_CONNECT: i64 = 2

/**
     * When setting this flag, the socket will not apply any
     * credentials when establishing a connection. For example,
     * an SSL connection would not send any client-certificates
     * if this flag is set.
     */

pub const NO_PERMANENT_STORAGE: i64

NO_PERMANENT_STORAGE: i64 = 4

/**
     * If set, indicates that the connection was initiated from a source
     * defined as being private in the sense of Private Browsing. Generally,
     * there should be no state shared between connections that are private
     * and those that are not; it is OK for multiple private connections
     * to share state with each other, and it is OK for multiple non-private
     * connections to share state with each other.
     */

pub const MITM_OK: i64

MITM_OK: i64 = 8

/**
     * This flag is an explicit opt-in that allows a normally secure socket
     * provider to use, at its discretion, an insecure algorithm. e.g.
     * a TLS socket without authentication.
     */

pub const BE_CONSERVATIVE: i64

BE_CONSERVATIVE: i64 = 16

/**
     * If set, do not use newer protocol features that might have interop problems
     * on the Internet. Intended only for use with critical infra like the updater.
     * default is false.
     */