Struct xpcom::interfaces::nsIStorageStream

#[repr(C)]

pub struct nsIStorageStream { /* fields omitted */ }

interface nsIStorageStream : nsISupports

/**
 * The nsIStorageStream interface maintains an internal data buffer that can be
 * filled using a single output stream.  One or more independent input streams
 * can be created to read the data from the buffer non-destructively.
 */

Methods

impl nsIStorageStream

pub fn coerce<T: nsIStorageStreamCoerce>(&self) -> &T

Cast this nsIStorageStream to one of its base interfaces.

impl nsIStorageStream

pub unsafe fn Init(&self, segmentSize: uint32_t, maxSize: uint32_t) -> nsresult

/**
     *
     * Initialize the stream, setting up the amount of space that will be
     * allocated for the stream's backing-store.
     *
     * @param segmentSize
     *        Size of each segment. Must be a power of two.
     * @param maxSize
     *        Maximum total size of this stream. length will always be less
     *        than or equal to this value. Passing UINT32_MAX is safe.
     */

void init (in uint32_t segmentSize, in uint32_t maxSize);

pub unsafe fn GetOutputStream(
&self,
startPosition: int32_t,
_retval: *mut *const nsIOutputStream
) -> nsresult

/**
     * Get a reference to the one and only output stream for this instance.
     * The zero-based startPosition argument is used is used to set the initial
     * write cursor position.  The startPosition cannot be set larger than the
     * current buffer length.  Calling this method has the side-effect of
     * truncating the internal buffer to startPosition bytes.
     */

nsIOutputStream getOutputStream (in int32_t startPosition);

pub unsafe fn NewInputStream(
&self,
startPosition: int32_t,
_retval: *mut *const nsIInputStream
) -> nsresult

/**
     * Create a new input stream to read data (written by the singleton output
     * stream) from the internal buffer.  Multiple, independent input streams
     * can be created.
     */

nsIInputStream newInputStream (in int32_t startPosition);

pub unsafe fn GetLength(&self, aLength: *mut uint32_t) -> nsresult

/**
     * The length attribute indicates the total number of bytes stored in the
     * nsIStorageStream internal buffer, regardless of any consumption by input
     * streams.  Assigning to the length field can be used to truncate the
     * buffer data, but can not be used when either the instance's output
     * stream is in use.
     *
     * @See #writeInProgress */

attribute uint32_t length;

pub unsafe fn SetLength(&self, aLength: uint32_t) -> nsresult

/**
     * The length attribute indicates the total number of bytes stored in the
     * nsIStorageStream internal buffer, regardless of any consumption by input
     * streams.  Assigning to the length field can be used to truncate the
     * buffer data, but can not be used when either the instance's output
     * stream is in use.
     *
     * @See #writeInProgress */

attribute uint32_t length;

pub unsafe fn GetWriteInProgress(&self, aWriteInProgress: *mut bool) -> nsresult

/**
     * True, when output stream has not yet been Close'ed
     */

readonly attribute boolean writeInProgress;