Struct xpcom::interfaces::nsIStreamConverterService

#[repr(C)]

pub struct nsIStreamConverterService { /* fields omitted */ }

interface nsIStreamConverterService : nsISupports

/**
 * The nsIStreamConverterService is a higher level stream converter factory
 * responsible for locating and creating stream converters
 * (nsIStreamConverter).
 *
 * This service retrieves an interface that can convert data from a particular
 * MIME type, to a particular MIME type. It is responsible for any intermediary
 * conversion required in order to get from X to Z, assuming direct conversion
 * is not possible.
 *
 * @author Jud Valeski
 * @see nsIStreamConverter
 */

Methods

impl nsIStreamConverterService

pub fn coerce<T: nsIStreamConverterServiceCoerce>(&self) -> &T

Cast this nsIStreamConverterService to one of its base interfaces.

impl nsIStreamConverterService

pub unsafe fn CanConvert(
&self,
aFromType: *const c_char,
aToType: *const c_char,
_retval: *mut bool
) -> nsresult

/**
     * Tests whether conversion between the two specified types is possible.
     * This is cheaper than calling convert()/asyncConvertData(); it is not
     * necessary to call this function before calling one of those, though.
     */

boolean canConvert (in string aFromType, in string aToType);

pub unsafe fn Convert(
&self,
aFromStream: *const nsIInputStream,
aFromType: *const c_char,
aToType: *const c_char,
aContext: *const nsISupports,
_retval: *mut *const nsIInputStream
) -> nsresult

/**
     * <b>SYNCHRONOUS VERSION</b>
     * Converts a stream of one type, to a stream of another type.
     *
     * Use this method when you have a stream you want to convert.
     *
     * @param aFromStream   The stream representing the original/raw data.
     * @param aFromType     The MIME type of aFromStream.
     * @param aToType       The MIME type of the returned stream.
     * @param aContext      Either an opaque context, or a converter specific
     *                      context (implementation specific).
     * @return              The converted stream. NOTE: The returned stream
     *                      may not already be converted. An efficient stream
     *                      converter implementation will convert data on
     *                      demand rather than buffering the converted data
     *                      until it is used.
     */

nsIInputStream convert (in nsIInputStream aFromStream, in string aFromType, in string aToType, in nsISupports aContext);

pub unsafe fn AsyncConvertData(
&self,
aFromType: *const c_char,
aToType: *const c_char,
aListener: *const nsIStreamListener,
aContext: *const nsISupports,
_retval: *mut *const nsIStreamListener
) -> nsresult

/**
     * <b>ASYNCHRONOUS VERSION</b>
     * Retrieves a nsIStreamListener that receives the original/raw data via its
     * nsIStreamListener::OnDataAvailable() callback, then converts and pushes
     * the data to aListener.
     *
     * Use this method when you want to proxy (and convert) nsIStreamListener
     * callbacks asynchronously.
     *
     * @param aFromType     The MIME type of the original/raw data.
     * @param aToType       The MIME type of the converted data.
     * @param aListener     The listener that receives the converted data.
     * @param aCtxt         Either an opaque context, or a converter specific
     *                      context (implementation specific).
     * @return              A nsIStreamListener that receives data via its
     *                      OnDataAvailable() method.
     */

nsIStreamListener asyncConvertData (in string aFromType, in string aToType, in nsIStreamListener aListener, in nsISupports aContext);