Struct xpcom::interfaces::nsIStreamLoader

#[repr(C)]

pub struct nsIStreamLoader { /* fields omitted */ }

interface nsIStreamLoader : nsIStreamListener

Methods

impl nsIStreamLoader

pub fn coerce<T: nsIStreamLoaderCoerce>(&self) -> &T

Cast this nsIStreamLoader to one of its base interfaces.

impl nsIStreamLoader

pub unsafe fn Init(
&self,
aStreamObserver: *const nsIStreamLoaderObserver,
aRequestObserver: *const nsIRequestObserver
) -> nsresult

/**
 * Asynchronously loads a channel into a memory buffer.
 *
 * To use this interface, first call init() with a nsIStreamLoaderObserver
 * that will be notified when the data has been loaded. Then call asyncOpen()
 * on the channel with the nsIStreamLoader as the listener. The context
 * argument in the asyncOpen() call will be passed to the onStreamComplete()
 * callback.
 *
 * XXX define behaviour for sizes >4 GB
 */
/**
     * Initialize this stream loader, and start loading the data.
     *
     * @param aStreamObserver
     *        An observer that will be notified when the data is complete.
     * @param aRequestObserver
     *        An optional observer that will be notified when the request
     *        has started or stopped.
     */

void init (in nsIStreamLoaderObserver aStreamObserver, [optional] in nsIRequestObserver aRequestObserver);

pub unsafe fn GetNumBytesRead(&self, aNumBytesRead: *mut uint32_t) -> nsresult

/**
     * Gets the number of bytes read so far.
     */

readonly attribute unsigned long numBytesRead;

pub unsafe fn GetRequest(&self, aRequest: *mut *const nsIRequest) -> nsresult

/**
     * Gets the request that loaded this file.
     * null after the request has finished loading.
     */

readonly attribute nsIRequest request;