Struct xpcom::interfaces::nsITLSClientStatus

#[repr(C)]

pub struct nsITLSClientStatus { /* fields omitted */ }

interface nsITLSClientStatus : nsISupports

/**
 * Security summary for a given TLS client connection being handled by a
 * |nsITLSServerSocket| server.
 *
 * This is accessible through the security info object on the transport, which
 * will be an instance of |nsITLSServerConnectionInfo| (see below).
 *
 * The values of these attributes are available once the |onHandshakeDone|
 * method of the security observer has been called (see
 * |nsITLSServerSecurityObserver| below).
 */

Methods

impl nsITLSClientStatus

pub fn coerce<T: nsITLSClientStatusCoerce>(&self) -> &T

Cast this nsITLSClientStatus to one of its base interfaces.

impl nsITLSClientStatus

pub const SSL_VERSION_3: i64

SSL_VERSION_3: i64 = 768

/**
   * Values for tlsVersionUsed, as defined by TLS
   */

pub const TLS_VERSION_1: i64

TLS_VERSION_1: i64 = 769

pub const TLS_VERSION_1_1: i64

TLS_VERSION_1_1: i64 = 770

pub const TLS_VERSION_1_2: i64

TLS_VERSION_1_2: i64 = 771

pub const TLS_VERSION_1_3: i64

TLS_VERSION_1_3: i64 = 772

pub const TLS_VERSION_UNKNOWN: i64

TLS_VERSION_UNKNOWN: i64 = -1

pub unsafe fn GetPeerCert(&self, aPeerCert: *mut *const nsIX509Cert) -> nsresult

/**
   * peerCert
   *
   * The client's certificate, if one was requested via |requestCertificate|
   * above and supplied by the client.
   */

readonly attribute nsIX509Cert peerCert;

pub unsafe fn GetTlsVersionUsed(
&self,
aTlsVersionUsed: *mut int16_t
) -> nsresult

/**
   * tlsVersionUsed
   *
   * The version of TLS used by the connection.  See values above.
   */

readonly attribute short tlsVersionUsed;

pub unsafe fn GetCipherName(&self, aCipherName: &mut nsACString) -> nsresult

/**
   * cipherName
   *
   * Name of the cipher suite used, such as
   * "TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256".
   * See security/nss/lib/ssl/sslinfo.c for the possible values.
   */

readonly attribute ACString cipherName;

pub unsafe fn GetKeyLength(&self, aKeyLength: *mut uint32_t) -> nsresult

/**
   * keyLength
   *
   * The "effective" key size of the symmetric key in bits.
   */

readonly attribute unsigned long keyLength;

pub unsafe fn GetMacLength(&self, aMacLength: *mut uint32_t) -> nsresult

/**
   * macLength
   *
   * The size of the MAC in bits.
   */

readonly attribute unsigned long macLength;