Struct xpcom::interfaces::nsITLSServerConnectionInfo

#[repr(C)]

pub struct nsITLSServerConnectionInfo { /* fields omitted */ }

interface nsITLSServerConnectionInfo : nsISupports

/**
 * Connection info for a given TLS client connection being handled by a
 * |nsITLSServerSocket| server.  This object is thread-safe.
 *
 * This is exposed as the security info object on the transport, so it can be
 * accessed via |transport.securityInfo|.
 *
 * This interface is available by the time the |onSocketAttached| is called,
 * which is the first time the TLS server consumer is notified of a new client.
 */

Methods

impl nsITLSServerConnectionInfo

pub fn coerce<T: nsITLSServerConnectionInfoCoerce>(&self) -> &T

Cast this nsITLSServerConnectionInfo to one of its base interfaces.

impl nsITLSServerConnectionInfo

pub unsafe fn SetSecurityObserver(
&self,
observer: *const nsITLSServerSecurityObserver
) -> nsresult

/**
   * setSecurityObserver
   *
   * Set the security observer to be notified when the TLS handshake has
   * completed.
   */

void setSecurityObserver (in nsITLSServerSecurityObserver observer);

pub unsafe fn GetServerSocket(
&self,
aServerSocket: *mut *const nsITLSServerSocket
) -> nsresult

/**
   * serverSocket
   *
   * The nsITLSServerSocket instance that accepted this client connection.
   */

readonly attribute nsITLSServerSocket serverSocket;

pub unsafe fn GetStatus(
&self,
aStatus: *mut *const nsITLSClientStatus
) -> nsresult

/**
   * status
   *
   * Security summary for this TLS client connection.  Note that the values of
   * this interface are not available until the TLS handshake has completed.
   * See |nsITLSClientStatus| above for more details.
   */

readonly attribute nsITLSClientStatus status;