Struct xpcom::interfaces::nsITLSServerSocket

#[repr(C)]

pub struct nsITLSServerSocket { /* fields omitted */ }

interface nsITLSServerSocket : nsIServerSocket

Methods

impl nsITLSServerSocket

pub fn coerce<T: nsITLSServerSocketCoerce>(&self) -> &T

Cast this nsITLSServerSocket to one of its base interfaces.

impl nsITLSServerSocket

pub const REQUEST_NEVER: i64

REQUEST_NEVER: i64 = 0

/**
   * Values for setRequestClientCertificate
   */

pub const REQUEST_FIRST_HANDSHAKE: i64

REQUEST_FIRST_HANDSHAKE: i64 = 1

pub const REQUEST_ALWAYS: i64

REQUEST_ALWAYS: i64 = 2

pub const REQUIRE_FIRST_HANDSHAKE: i64

REQUIRE_FIRST_HANDSHAKE: i64 = 3

pub const REQUIRE_ALWAYS: i64

REQUIRE_ALWAYS: i64 = 4

pub unsafe fn GetServerCert(
&self,
aServerCert: *mut *const nsIX509Cert
) -> nsresult

/**
   * serverCert
   *
   * The server's certificate that is presented to the client during the TLS
   * handshake.  This is required to be set before calling |asyncListen|.
   */

attribute nsIX509Cert serverCert;

pub unsafe fn SetServerCert(&self, aServerCert: *const nsIX509Cert) -> nsresult

/**
   * serverCert
   *
   * The server's certificate that is presented to the client during the TLS
   * handshake.  This is required to be set before calling |asyncListen|.
   */

attribute nsIX509Cert serverCert;

pub unsafe fn SetSessionCache(&self, aSessionCache: bool) -> nsresult

/**
   * setSessionCache
   *
   * Whether the server should use a session cache.  Defaults to true.  This
   * should be set before calling |asyncListen| if you wish to change the
   * default.
   */

void setSessionCache (in boolean aSessionCache);

pub unsafe fn SetSessionTickets(&self, aSessionTickets: bool) -> nsresult

/**
   * setSessionTickets
   *
   * Whether the server should support session tickets.  Defaults to true.  This
   * should be set before calling |asyncListen| if you wish to change the
   * default.
   */

void setSessionTickets (in boolean aSessionTickets);

pub unsafe fn SetRequestClientCertificate(
&self,
aRequestClientCert: uint32_t
) -> nsresult

/**
   * setRequestClientCertificate
   *
   * Whether the server should request and/or require a client auth certificate
   * from the client.  Defaults to REQUEST_NEVER.  See the possible options
   * above.  This should be set before calling |asyncListen| if you wish to
   * change the default.
   */

void setRequestClientCertificate (in unsigned long aRequestClientCert);

pub unsafe fn SetCipherSuites(
&self,
aCipherSuites: *mut uint16_t,
aLength: uint32_t
) -> nsresult

/**
   * setCipherSuites
   *
   * The server's cipher suites that is used by the TLS handshake.
   * This is required to be set before calling |asyncListen|.
   */

void setCipherSuites ([array, size_is (aLength)] in unsigned short aCipherSuites, in unsigned long aLength);

pub unsafe fn SetVersionRange(
&self,
aMinVersion: uint16_t,
aMaxVersion: uint16_t
) -> nsresult

/**
   * setVersionRange
   *
   * The server's TLS versions that is used by the TLS handshake.
   * This is required to be set before calling |asyncListen|.
   *
   * aMinVersion and aMaxVersion is a TLS version value from
   * |nsITLSClientStatus| constants.
   */

void setVersionRange (in unsigned short aMinVersion, in unsigned short aMaxVersion);