Struct xpcom::interfaces::nsITabParent

#[repr(C)]

pub struct nsITabParent { /* fields omitted */ }

interface nsITabParent : nsISupports

Methods

impl nsITabParent

pub fn coerce<T: nsITabParentCoerce>(&self) -> &T

Cast this nsITabParent to one of its base interfaces.

impl nsITabParent

pub unsafe fn GetChildProcessOffset(
&self,
aCssX: *mut int32_t,
aCssY: *mut int32_t
) -> nsresult

void getChildProcessOffset (out int32_t aCssX, out int32_t aCssY);

pub unsafe fn GetUseAsyncPanZoom(&self, aUseAsyncPanZoom: *mut bool) -> nsresult

readonly attribute boolean useAsyncPanZoom;

pub unsafe fn GetDocShellIsActive(
&self,
aDocShellIsActive: *mut bool
) -> nsresult

/**
    * Manages the docshell active state of the remote browser.
    */

attribute boolean docShellIsActive;

pub unsafe fn SetDocShellIsActive(&self, aDocShellIsActive: bool) -> nsresult

/**
    * Manages the docshell active state of the remote browser.
    */

attribute boolean docShellIsActive;

pub unsafe fn GetIsPrerendered(&self) -> bool

/**
   * Whether this tabParent is in prerender mode.
   */

[infallible] readonly attribute boolean isPrerendered;

pub unsafe fn PreserveLayers(&self, aPreserveLayers: bool) -> nsresult

/**
    * As an optimisation, setting the docshell's active state to
    * inactive also triggers a layer invalidation to free up some
    * potentially unhelpful memory usage. Calling preserveLayers
    * will cause the layers to be preserved even for inactive
    * docshells.
    */

void preserveLayers (in boolean aPreserveLayers);

pub unsafe fn SuppressDisplayport(&self, aEnabled: bool) -> nsresult

/**
   * During interactions where painting performance
   * is more important than scrolling, we may temporarily
   * suppress the displayport. Each enable called must be matched
   * with a disable call.
   */

void suppressDisplayport (in bool aEnabled);

pub unsafe fn GetTabId(&self, aTabId: *mut uint64_t) -> nsresult

readonly attribute uint64_t tabId;

pub unsafe fn GetOsPid(&self, aOsPid: *mut int32_t) -> nsresult

/**
   * The OS level process Id of the related child process.
   */

readonly attribute int32_t osPid;

pub unsafe fn NavigateByKey(
&self,
aForward: bool,
aForDocumentNavigation: bool
) -> nsresult

/**
   * Navigate by key. If aForDocumentNavigation is true, navigate by document.
   * If aForDocumentNavigation is false, navigate by element.
   *
   * If aForward is true, navigate to the first focusable element or document.
   * If aForward is false, navigate to the last focusable element or document.
   */

void navigateByKey (in bool aForward, in bool aForDocumentNavigation);

pub unsafe fn GetHasContentOpener(
&self,
aHasContentOpener: *mut bool
) -> nsresult

readonly attribute boolean hasContentOpener;

pub unsafe fn GetHasPresented(&self, aHasPresented: *mut bool) -> nsresult

/**
   * True if we've previously received layers for this tab when switching to
   * it.
   */

readonly attribute boolean hasPresented;

pub unsafe fn TransmitPermissionsForPrincipal(
&self,
aPrincipal: *const nsIPrincipal
) -> nsresult

/**
   * Ensures that the content process which has this tab parent has all of the
   * permissions required to load a document with the given principal.
   */

void transmitPermissionsForPrincipal (in nsIPrincipal aPrincipal);

pub unsafe fn GetHasBeforeUnload(&self, aHasBeforeUnload: *mut bool) -> nsresult

/**
   * True if any of the frames loaded in the TabChild have registered
   * an onbeforeunload event handler.
   */

readonly attribute boolean hasBeforeUnload;

pub unsafe fn StartApzAutoscroll(
&self,
aAnchorX: c_float,
aAnchorY: c_float,
aScrollId: nsViewID,
aPresShellId: uint32_t,
_retval: *mut bool
) -> nsresult

/**
   * Notify APZ to start autoscrolling.
   * (aAnchorX, aAnchorY) are the coordinates of the autoscroll anchor,
   * in CSS coordinates relative to the screen. aScrollId and
   * aPresShellId identify the scroll frame that content chose to scroll.
   * Returns whether we were successfully able to notify APZ.
   * If this function returns true, APZ (which may live in another process)
   * may still reject the autoscroll, but it's then APZ's reponsibility
   * to notify content via an "autoscroll-rejected-by-apz" message.
   */

boolean startApzAutoscroll (in float aAnchorX, in float aAnchorY, in nsViewID aScrollId, in uint32_t aPresShellId);

pub unsafe fn StopApzAutoscroll(
&self,
aScrollId: nsViewID,
aPresShellId: uint32_t
) -> nsresult

/**
   * Notify APZ to stop autoscrolling.
   * aScrollId and aPresShellId identify the scroll frame that is being
   * autoscrolled.
   */

void stopApzAutoscroll (in nsViewID aScrollId, in uint32_t aPresShellId);