Struct xpcom::interfaces::nsITaggingService

#[repr(C)]

pub struct nsITaggingService { /* fields omitted */ }

interface nsITaggingService : nsISupports

Methods

impl nsITaggingService

pub fn coerce<T: nsITaggingServiceCoerce>(&self) -> &T

Cast this nsITaggingService to one of its base interfaces.

impl nsITaggingService

pub const MAX_TAG_LENGTH: i64

MAX_TAG_LENGTH: i64 = 100

/**
   * Defines the maximal length of a tag. Related to the bug 407821
   * (https://bugzilla.mozilla.org/show_bug.cgi?id=407821)
   */

pub unsafe fn TagURI(
&self,
aURI: *const nsIURI,
aTags: *const nsIVariant,
aSource: uint16_t
) -> nsresult

/**
   * Tags a URL with the given set of tags. Current tags set for the URL
   * persist. Tags in aTags which are already set for the given URL are
   * ignored.
   *
   * @param aURI
   *        the URL to tag.
   * @param aTags
   *        Array of tags to set for the given URL.  Each element within the
   *        array can be either a tag name (non-empty string) or a concrete
   *        itemId of a tag container.
   * @param [optional] aSource
   *        A change source constant from nsINavBookmarksService::SOURCE_*.
   *        Defaults to SOURCE_DEFAULT if omitted.
   */

void tagURI (in nsIURI aURI, in nsIVariant aTags, [optional] in unsigned short aSource);

pub unsafe fn UntagURI(
&self,
aURI: *const nsIURI,
aTags: *const nsIVariant,
aSource: uint16_t
) -> nsresult

/**
   * Removes tags from a URL. Tags from aTags which are not set for the
   * given URL are ignored.
   *
   * @param aURI
   *        the URL to un-tag.
   * @param aTags
   *        Array of tags to unset.  Pass null to remove all tags from the given
   *        url.  Each element within the array can be either a tag name
   *        (non-empty string) or a concrete itemId of a tag container.
   * @param [optional] aSource
   *        A change source constant from nsINavBookmarksService::SOURCE_*.
   *        Defaults to SOURCE_DEFAULT if omitted.
   */

void untagURI (in nsIURI aURI, in nsIVariant aTags, [optional] in unsigned short aSource);

pub unsafe fn GetURIsForTag(
&self,
aTag: &nsAString,
_retval: *mut *const nsIVariant
) -> nsresult

/**
   * Retrieves all URLs tagged with the given tag.
   *
   * @param aTag
   *        tag name
   * @returns Array of uris tagged with aTag.
   */

nsIVariant getURIsForTag (in AString aTag);

pub unsafe fn GetTagsForURI(
&self,
aURI: *const nsIURI,
length: *mut uint32_t,
aTags: *mut *mut *const int16_t
) -> nsresult

/**
   * Retrieves all tags set for the given URL.
   *
   * @param aURI
   *        a URL.
   * @returns array of tags (sorted by name).
   */

void getTagsForURI (in nsIURI aURI, [optional] out unsigned long length, [array, size_is (length), retval] out wstring aTags);

pub unsafe fn GetAllTags(&self, aAllTags: *mut *const nsIVariant) -> nsresult

/**
   * Retrieves all tags used to tag URIs in the data-base (sorted by name).
   */

readonly attribute nsIVariant allTags;

pub unsafe fn GetHasTags(&self, aHasTags: *mut bool) -> nsresult

/**
   * Whether any tags exist.
   *
   * @note This is faster than allTags.length, since doesn't need to sort tags.
   */

readonly attribute boolean hasTags;