Struct xpcom::interfaces::nsIThreadInternal

#[repr(C)]

pub struct nsIThreadInternal { /* fields omitted */ }

interface nsIThreadInternal : nsIThread

/**
 * The XPCOM thread object implements this interface, which allows a consumer
 * to observe dispatch activity on the thread.
 */

Methods

impl nsIThreadInternal

pub fn coerce<T: nsIThreadInternalCoerce>(&self) -> &T

Cast this nsIThreadInternal to one of its base interfaces.

impl nsIThreadInternal

pub unsafe fn GetObserver(
&self,
aObserver: *mut *const nsIThreadObserver
) -> nsresult

/**
   * Get/set the current thread observer (may be null).  This attribute may be
   * read from any thread, but must only be set on the thread corresponding to
   * this thread object.  The observer will be released on the thread
   * corresponding to this thread object after all other events have been
   * processed during a call to Shutdown.
   */

attribute nsIThreadObserver observer;

pub unsafe fn SetObserver(
&self,
aObserver: *const nsIThreadObserver
) -> nsresult

/**
   * Get/set the current thread observer (may be null).  This attribute may be
   * read from any thread, but must only be set on the thread corresponding to
   * this thread object.  The observer will be released on the thread
   * corresponding to this thread object after all other events have been
   * processed during a call to Shutdown.
   */

attribute nsIThreadObserver observer;

pub unsafe fn AddObserver(&self, observer: *const nsIThreadObserver) -> nsresult

/**
   * Add an observer that will *only* receive onProcessNextEvent,
   * beforeProcessNextEvent. and afterProcessNextEvent callbacks. Always called
   * on the target thread, and the implementation does not have to be
   * threadsafe. Order of callbacks is not guaranteed (i.e.
   * afterProcessNextEvent may be called first depending on whether or not the
   * observer is added in a nested loop). Holds a strong ref.
   */

void addObserver (in nsIThreadObserver observer);

pub unsafe fn RemoveObserver(
&self,
observer: *const nsIThreadObserver
) -> nsresult

/**
   * Remove an observer added via the addObserver call. Once removed the
   * observer will never be called again by the thread.
   */

void removeObserver (in nsIThreadObserver observer);