Struct xpcom::interfaces::nsIToolkitChromeRegistry

#[repr(C)]

pub struct nsIToolkitChromeRegistry { /* fields omitted */ }

interface nsIToolkitChromeRegistry : nsIXULChromeRegistry

Methods

impl nsIToolkitChromeRegistry

pub fn coerce<T: nsIToolkitChromeRegistryCoerce>(&self) -> &T

Cast this nsIToolkitChromeRegistry to one of its base interfaces.

impl nsIToolkitChromeRegistry

pub unsafe fn CheckForOSAccessibility(&self) -> nsresult

/**
   * If the OS has a "high-visibility" or "disabled-friendly" theme set,
   * we want to force mozilla into the classic theme, which (for the most part
   * obeys the system color/font settings. We cannot do this at initialization,
   * because it depends on the toolkit (GTK2) being initialized, which is
   * not the case in some embedding situations. Embedders have to manually
   * call this method during the startup process.
   */

void checkForOSAccessibility ();

pub unsafe fn GetLocalesForPackage(
&self,
aPackage: &nsACString,
_retval: *mut *const nsIUTF8StringEnumerator
) -> nsresult

/**
   * Get a list of locales available for the specified package.
   */

nsIUTF8StringEnumerator getLocalesForPackage (in AUTF8String aPackage);