Struct xpcom::interfaces::nsIToolkitProfileService

#[repr(C)]

pub struct nsIToolkitProfileService { /* fields omitted */ }

interface nsIToolkitProfileService : nsISupports

Methods

impl nsIToolkitProfileService

pub fn coerce<T: nsIToolkitProfileServiceCoerce>(&self) -> &T

Cast this nsIToolkitProfileService to one of its base interfaces.

impl nsIToolkitProfileService

pub unsafe fn GetStartWithLastProfile(
&self,
aStartWithLastProfile: *mut bool
) -> nsresult

attribute boolean startWithLastProfile;

pub unsafe fn SetStartWithLastProfile(
&self,
aStartWithLastProfile: bool
) -> nsresult

attribute boolean startWithLastProfile;

pub unsafe fn GetStartOffline(&self, aStartOffline: *mut bool) -> nsresult

attribute boolean startOffline;

pub unsafe fn SetStartOffline(&self, aStartOffline: bool) -> nsresult

attribute boolean startOffline;

pub unsafe fn GetProfiles(
&self,
aProfiles: *mut *const nsISimpleEnumerator
) -> nsresult

readonly attribute nsISimpleEnumerator profiles;

pub unsafe fn GetSelectedProfile(
&self,
aSelectedProfile: *mut *const nsIToolkitProfile
) -> nsresult

/**
     * The currently selected profile (the one used or about to be used by the
     * browser).
     */

attribute nsIToolkitProfile selectedProfile;

pub unsafe fn SetSelectedProfile(
&self,
aSelectedProfile: *const nsIToolkitProfile
) -> nsresult

/**
     * The currently selected profile (the one used or about to be used by the
     * browser).
     */

attribute nsIToolkitProfile selectedProfile;

pub unsafe fn GetDefaultProfile(
&self,
aDefaultProfile: *mut *const nsIToolkitProfile
) -> nsresult

/**
     * The default profile (the one used or about to be used by the
     * browser if no other profile is specified at runtime). This is the profile
     * marked with Default=1 in profiles.ini and is usually the same as
     * selectedProfile, except on Developer Edition.
     *
     * Developer Edition uses a profile named "dev-edition-default" as the
     * default profile (which it creates if it doesn't exist), unless a special
     * empty file named "ignore-dev-edition-profile" is present next to
     * profiles.ini. In that case Developer Edition behaves the same as any
     * other build of Firefox.
     */

attribute nsIToolkitProfile defaultProfile;

pub unsafe fn SetDefaultProfile(
&self,
aDefaultProfile: *const nsIToolkitProfile
) -> nsresult

/**
     * The default profile (the one used or about to be used by the
     * browser if no other profile is specified at runtime). This is the profile
     * marked with Default=1 in profiles.ini and is usually the same as
     * selectedProfile, except on Developer Edition.
     *
     * Developer Edition uses a profile named "dev-edition-default" as the
     * default profile (which it creates if it doesn't exist), unless a special
     * empty file named "ignore-dev-edition-profile" is present next to
     * profiles.ini. In that case Developer Edition behaves the same as any
     * other build of Firefox.
     */

attribute nsIToolkitProfile defaultProfile;

pub unsafe fn GetProfileByName(
&self,
aName: &nsACString,
_retval: *mut *const nsIToolkitProfile
) -> nsresult

/**
     * Get a profile by name. This is mainly for use by the -P
     * commandline flag.
     *
     * @param aName The profile name to find.
     */

nsIToolkitProfile getProfileByName (in AUTF8String aName);

pub unsafe fn LockProfilePath(
&self,
aDirectory: *const nsIFile,
aTempDirectory: *const nsIFile,
_retval: *mut *const nsIProfileLock
) -> nsresult

/**
     * Lock an arbitrary path as a profile. If the path does not exist, it
     * will be created and the defaults copied from the application directory.
     */

nsIProfileLock lockProfilePath (in nsIFile aDirectory, in nsIFile aTempDirectory);

pub unsafe fn CreateProfile(
&self,
aRootDir: *const nsIFile,
aName: &nsACString,
_retval: *mut *const nsIToolkitProfile
) -> nsresult

/**
     * Create a new profile.
     *
     * The profile temporary directory will be chosen based on where the
     * profile directory is located.
     *
     * @param aRootDir
     *        The profile directory. May be null, in which case a suitable
     *        default will be chosen based on the profile name.
     * @param aName
     *        The profile name.
     */

nsIToolkitProfile createProfile (in nsIFile aRootDir, in AUTF8String aName);

pub unsafe fn CreateDefaultProfileForApp(
&self,
aProfileName: &nsACString,
aAppName: &nsACString,
aVendorName: &nsACString,
_retval: *mut *const nsIToolkitProfile
) -> nsresult

/**
     * Create the default profile for an application.
     *
     * The profile will be typically in
     * {Application Data}/.profilename/{salt}.default or
     * {Application Data}/.appname/{salt}.default
     * or if aVendorName is provided
     * {Application Data}/.vendor/appname/{salt}.default
     *
     * @note Either aProfileName or aAppName must be non-empty
     *
     * @param  aProfileName
     *         The name of the profile
     * @param  aAppName
     *         The name of the application
     * @param  aVendorName
     *         The name of the vendor
     * @return The created profile.
     */

nsIToolkitProfile createDefaultProfileForApp (in AUTF8String aProfileName, in AUTF8String aAppName, in AUTF8String aVendorName);

pub unsafe fn GetProfileCount(&self, aProfileCount: *mut uint32_t) -> nsresult

/**
     * Returns the number of profiles.
     * @return 0, 1, or 2. More than 2 profiles will always return 2.
     */

readonly attribute unsigned long profileCount;

pub unsafe fn Flush(&self) -> nsresult

/**
     * Flush the profiles list file.
     */

void flush ();