Struct xpcom::interfaces::nsITreeBoxObject

#[repr(C)]

pub struct nsITreeBoxObject { /* fields omitted */ }

interface nsITreeBoxObject : nsISupports

Methods

impl nsITreeBoxObject

pub fn coerce<T: nsITreeBoxObjectCoerce>(&self) -> &T

Cast this nsITreeBoxObject to one of its base interfaces.

impl nsITreeBoxObject

pub unsafe fn GetColumns(
&self,
aColumns: *mut *const nsITreeColumns
) -> nsresult

/**
   * Obtain the columns.
   */

readonly attribute nsITreeColumns columns;

pub unsafe fn GetView(&self, aView: *mut *const nsITreeView) -> nsresult

/**
   * The view that backs the tree and that supplies it with its data.
   * It is dynamically settable, either using a view attribute on the
   * tree tag or by setting this attribute to a new value.
   */

attribute nsITreeView view;

pub unsafe fn SetView(&self, aView: *const nsITreeView) -> nsresult

/**
   * The view that backs the tree and that supplies it with its data.
   * It is dynamically settable, either using a view attribute on the
   * tree tag or by setting this attribute to a new value.
   */

attribute nsITreeView view;

pub unsafe fn GetFocused(&self, aFocused: *mut bool) -> nsresult

/**
   * Whether or not we are currently focused.
   */

attribute boolean focused;

pub unsafe fn SetFocused(&self, aFocused: bool) -> nsresult

/**
   * Whether or not we are currently focused.
   */

attribute boolean focused;

pub unsafe fn GetTreeBody(
&self,
aTreeBody: *mut *const nsIDOMElement
) -> nsresult

/**
   * Obtain the treebody content node
   */

readonly attribute nsIDOMElement treeBody;

pub unsafe fn GetRowHeight(&self, aRowHeight: *mut int32_t) -> nsresult

/**
   * Obtain the height of a row.
   */

readonly attribute long rowHeight;

pub unsafe fn GetRowWidth(&self, aRowWidth: *mut int32_t) -> nsresult

/**
   * Obtain the width of a row.
   */

readonly attribute long rowWidth;

pub unsafe fn GetHorizontalPosition(
&self,
aHorizontalPosition: *mut int32_t
) -> nsresult

/**
   * Get the pixel position of the horizontal scrollbar.
   */

readonly attribute long horizontalPosition;

pub unsafe fn GetSelectionRegion(
&self,
aSelectionRegion: *mut *const nsIScriptableRegion
) -> nsresult

/**
   * Return the region for the visible parts of the selection, in device pixels.
   */

readonly attribute nsIScriptableRegion selectionRegion;

pub unsafe fn GetFirstVisibleRow(&self, _retval: *mut int32_t) -> nsresult

/**
   * Get the index of the first visible row.
   */

long getFirstVisibleRow ();

pub unsafe fn GetLastVisibleRow(&self, _retval: *mut int32_t) -> nsresult

/**
   * Get the index of the last visible row.
   */

long getLastVisibleRow ();

pub unsafe fn GetPageLength(&self, _retval: *mut int32_t) -> nsresult

/**
   * Gets the number of possible visible rows.
   */

long getPageLength ();

pub unsafe fn EnsureRowIsVisible(&self, index: int32_t) -> nsresult

/**
   * Ensures that a row at a given index is visible.
   */

void ensureRowIsVisible (in long index);

pub unsafe fn EnsureCellIsVisible(
&self,
row: int32_t,
col: *const nsITreeColumn
) -> nsresult

/**
   * Ensures that a given cell in the tree is visible.
   */

void ensureCellIsVisible (in long row, in nsITreeColumn col);

pub unsafe fn ScrollToRow(&self, index: int32_t) -> nsresult

/**
   * Scrolls such that the row at index is at the top of the visible view.
   */

void scrollToRow (in long index);

pub unsafe fn ScrollByLines(&self, numLines: int32_t) -> nsresult

/**
   * Scroll the tree up or down by numLines lines. Positive
   * values move down in the tree. Prevents scrolling off the
   * end of the tree.
   */

void scrollByLines (in long numLines);

pub unsafe fn ScrollByPages(&self, numPages: int32_t) -> nsresult

/**
   * Scroll the tree up or down by numPages pages. A page
   * is considered to be the amount displayed by the tree.
   * Positive values move down in the tree. Prevents scrolling
   * off the end of the tree.
   */

void scrollByPages (in long numPages);

pub unsafe fn ScrollToCell(
&self,
row: int32_t,
col: *const nsITreeColumn
) -> nsresult

/**
   * Scrolls such that a given cell is visible (if possible)
   * at the top left corner of the visible view.
   */

void scrollToCell (in long row, in nsITreeColumn col);

pub unsafe fn ScrollToColumn(&self, col: *const nsITreeColumn) -> nsresult

/**
   * Scrolls horizontally so that the specified column is
   * at the left of the view (if possible).
   */

void scrollToColumn (in nsITreeColumn col);

pub unsafe fn ScrollToHorizontalPosition(
&self,
horizontalPosition: int32_t
) -> nsresult

/**
   * Scroll to a specific horizontal pixel position.
   */

void scrollToHorizontalPosition (in long horizontalPosition);

pub unsafe fn Invalidate(&self) -> nsresult

/**
   * Invalidation methods for fine-grained painting control.
   */

void invalidate ();

pub unsafe fn InvalidateColumn(&self, col: *const nsITreeColumn) -> nsresult

void invalidateColumn (in nsITreeColumn col);

pub unsafe fn InvalidateRow(&self, index: int32_t) -> nsresult

void invalidateRow (in long index);

pub unsafe fn InvalidateCell(
&self,
row: int32_t,
col: *const nsITreeColumn
) -> nsresult

void invalidateCell (in long row, in nsITreeColumn col);

pub unsafe fn InvalidateRange(
&self,
startIndex: int32_t,
endIndex: int32_t
) -> nsresult

void invalidateRange (in long startIndex, in long endIndex);

pub unsafe fn InvalidateColumnRange(
&self,
startIndex: int32_t,
endIndex: int32_t,
col: *const nsITreeColumn
) -> nsresult

void invalidateColumnRange (in long startIndex, in long endIndex, in nsITreeColumn col);

pub unsafe fn GetRowAt(
&self,
x: int32_t,
y: int32_t,
_retval: *mut int32_t
) -> nsresult

/**
   * A hit test that can tell you what row the mouse is over.
   * returns -1 for invalid mouse coordinates.
   *
   * The coordinate system is the client coordinate system for the
   * document this boxObject lives in, and the units are CSS pixels.
   */

long getRowAt (in long x, in long y);

pub unsafe fn GetCellAt(
&self,
x: int32_t,
y: int32_t,
row: *mut int32_t,
col: *mut *const nsITreeColumn,
childElt: &mut nsAString
) -> nsresult

/**
   * A hit test that can tell you what cell the mouse is over.  Row is the row index
   * hit,  returns -1 for invalid mouse coordinates.  ColID is the column hit.
   * ChildElt is the pseudoelement hit: this can have values of
   * "cell", "twisty", "image", and "text".
   *
   * The coordinate system is the client coordinate system for the
   * document this boxObject lives in, and the units are CSS pixels.
   */

void getCellAt (in long x, in long y, out long row, out nsITreeColumn col, out AString childElt);

pub unsafe fn GetCoordsForCellItem(
&self,
row: int32_t,
col: *const nsITreeColumn,
element: &nsAString,
x: *mut int32_t,
y: *mut int32_t,
width: *mut int32_t,
height: *mut int32_t
) -> nsresult

/**
   * Find the coordinates of an element within a specific cell.
   */

void getCoordsForCellItem (in long row, in nsITreeColumn col, in AString element, out long x, out long y, out long width, out long height);

pub unsafe fn IsCellCropped(
&self,
row: int32_t,
col: *const nsITreeColumn,
_retval: *mut bool
) -> nsresult

/**
   * Determine if the text of a cell is being cropped or not.
   */

boolean isCellCropped (in long row, in nsITreeColumn col);

pub unsafe fn RowCountChanged(&self, index: int32_t, count: int32_t) -> nsresult

/**
   * The view is responsible for calling these notification methods when
   * rows are added or removed.  Index is the position at which the new
   * rows were added or at which rows were removed.  For
   * non-contiguous additions/removals, this method should be called multiple times.
   */

void rowCountChanged (in long index, in long count);

pub unsafe fn BeginUpdateBatch(&self) -> nsresult

/**
   * Notify the tree that the view is about to perform a batch
   * update, that is, add, remove or invalidate several rows at once.
   * This must be followed by calling endUpdateBatch(), otherwise the tree
   * will get out of sync.
   */

void beginUpdateBatch ();

pub unsafe fn EndUpdateBatch(&self) -> nsresult

/**
   * Notify the tree that the view has completed a batch update.
   */

void endUpdateBatch ();

pub unsafe fn ClearStyleAndImageCaches(&self) -> nsresult

/**
   * Called on a theme switch to flush out the tree's style and image caches.
   */

void clearStyleAndImageCaches ();

pub unsafe fn RemoveImageCacheEntry(
&self,
row: int32_t,
col: *const nsITreeColumn
) -> nsresult

/**
   * Remove an image source from the image cache to allow its invalidation.
   *
   * @note This only affects images supplied by the view, not the ones supplied
   *       through the styling context, like twisties or checkboxes.
   */

void removeImageCacheEntry (in long row, in nsITreeColumn col);