Struct xpcom::interfaces::nsITreeColumns

#[repr(C)]

pub struct nsITreeColumns { /* fields omitted */ }

interface nsITreeColumns : nsISupports

Methods

impl nsITreeColumns

pub fn coerce<T: nsITreeColumnsCoerce>(&self) -> &T

Cast this nsITreeColumns to one of its base interfaces.

impl nsITreeColumns

pub unsafe fn GetTree(&self, aTree: *mut *const nsITreeBoxObject) -> nsresult

/**
   * The tree widget for these columns.
   */

readonly attribute nsITreeBoxObject tree;

pub unsafe fn GetCount(&self, aCount: *mut int32_t) -> nsresult

/**
   * The number of columns.
   */

readonly attribute long count;

pub unsafe fn GetLength(&self, aLength: *mut int32_t) -> nsresult

/**
   * An alias for count (for the benefit of scripts which treat this as an
   * array).
   */

readonly attribute long length;

pub unsafe fn GetFirstColumn(
&self,
_retval: *mut *const nsITreeColumn
) -> nsresult

/**
   * Get the first/last column.
   */

nsITreeColumn getFirstColumn ();

pub unsafe fn GetLastColumn(
&self,
_retval: *mut *const nsITreeColumn
) -> nsresult

nsITreeColumn getLastColumn ();

pub unsafe fn GetPrimaryColumn(
&self,
_retval: *mut *const nsITreeColumn
) -> nsresult

/**
   * Attribute based column getters.
   */

nsITreeColumn getPrimaryColumn ();

pub unsafe fn GetSortedColumn(
&self,
_retval: *mut *const nsITreeColumn
) -> nsresult

nsITreeColumn getSortedColumn ();

pub unsafe fn GetKeyColumn(
&self,
_retval: *mut *const nsITreeColumn
) -> nsresult

nsITreeColumn getKeyColumn ();

pub unsafe fn GetColumnFor(
&self,
element: *const nsIDOMElement,
_retval: *mut *const nsITreeColumn
) -> nsresult

/**
   * Get the column for the given element.
   */

nsITreeColumn getColumnFor (in nsIDOMElement element);

pub unsafe fn GetNamedColumn(
&self,
id: &nsAString,
_retval: *mut *const nsITreeColumn
) -> nsresult

/**
   * Parametric column getters.
   */

nsITreeColumn getNamedColumn (in AString id);

pub unsafe fn GetColumnAt(
&self,
index: int32_t,
_retval: *mut *const nsITreeColumn
) -> nsresult

nsITreeColumn getColumnAt (in long index);

pub unsafe fn InvalidateColumns(&self) -> nsresult

/**
   * This method is called whenever a treecol is added or removed and
   * the column cache needs to be rebuilt.
   */

void invalidateColumns ();

pub unsafe fn RestoreNaturalOrder(&self) -> nsresult

void restoreNaturalOrder ();