Struct xpcom::interfaces::nsITreeSelection

#[repr(C)]

pub struct nsITreeSelection { /* fields omitted */ }

interface nsITreeSelection : nsISupports

Methods

impl nsITreeSelection

pub fn coerce<T: nsITreeSelectionCoerce>(&self) -> &T

Cast this nsITreeSelection to one of its base interfaces.

impl nsITreeSelection

pub unsafe fn GetTree(&self, aTree: *mut *const nsITreeBoxObject) -> nsresult

/**
   * The tree widget for this selection.
   */

attribute nsITreeBoxObject tree;

pub unsafe fn SetTree(&self, aTree: *const nsITreeBoxObject) -> nsresult

/**
   * The tree widget for this selection.
   */

attribute nsITreeBoxObject tree;

pub unsafe fn GetSingle(&self, aSingle: *mut bool) -> nsresult

/**
   * This attribute is a boolean indicating single selection.
   */

readonly attribute boolean single;

pub unsafe fn GetCount(&self, aCount: *mut int32_t) -> nsresult

/**
   * The number of rows currently selected in this tree.
   */

readonly attribute long count;

pub unsafe fn IsSelected(&self, index: int32_t, _retval: *mut bool) -> nsresult

/**
   * Indicates whether or not the row at the specified index is
   * part of the selection.
   */

boolean isSelected (in long index);

pub unsafe fn Select(&self, index: int32_t) -> nsresult

/**
   * Deselect all rows and select the row at the specified index.
   */

void select (in long index);

pub unsafe fn TimedSelect(&self, index: int32_t, delay: int32_t) -> nsresult

/**
   * Perform a timed select.
   */

void timedSelect (in long index, in long delay);

pub unsafe fn ToggleSelect(&self, index: int32_t) -> nsresult

/**
   * Toggle the selection state of the row at the specified index.
   */

void toggleSelect (in long index);

pub unsafe fn RangedSelect(
&self,
startIndex: int32_t,
endIndex: int32_t,
augment: bool
) -> nsresult

/**
   * Select the range specified by the indices.  If augment is true,
   * then we add the range to the selection without clearing out anything
   * else.  If augment is false, everything is cleared except for the specified range.
   */

void rangedSelect (in long startIndex, in long endIndex, in boolean augment);

pub unsafe fn ClearRange(
&self,
startIndex: int32_t,
endIndex: int32_t
) -> nsresult

/**
   * Clears the range.
   */

void clearRange (in long startIndex, in long endIndex);

pub unsafe fn ClearSelection(&self) -> nsresult

/**
   * Clears the selection.
   */

void clearSelection ();

pub unsafe fn InvertSelection(&self) -> nsresult

/**
   * Inverts the selection.
   */

void invertSelection ();

pub unsafe fn SelectAll(&self) -> nsresult

/**
   * Selects all rows.
   */

void selectAll ();

pub unsafe fn GetRangeCount(&self, _retval: *mut int32_t) -> nsresult

/**
   * Iterate the selection using these methods.
   */

long getRangeCount ();

pub unsafe fn GetRangeAt(
&self,
i: int32_t,
min: *mut int32_t,
max: *mut int32_t
) -> nsresult

void getRangeAt (in long i, out long min, out long max);

pub unsafe fn InvalidateSelection(&self) -> nsresult

/**
   * Can be used to invalidate the selection.
   */

void invalidateSelection ();

pub unsafe fn AdjustSelection(&self, index: int32_t, count: int32_t) -> nsresult

/**
   * Called when the row count changes to adjust selection indices.
   */

void adjustSelection (in long index, in long count);

pub unsafe fn GetSelectEventsSuppressed(
&self,
aSelectEventsSuppressed: *mut bool
) -> nsresult

/**
   * This attribute is a boolean indicating whether or not the
   * "select" event should fire when the selection is changed using
   * one of our methods.  A view can use this to temporarily suppress
   * the selection while manipulating all of the indices, e.g., on
   * a sort.
   * Note: setting this attribute to false will fire a select event.
   */

attribute boolean selectEventsSuppressed;

pub unsafe fn SetSelectEventsSuppressed(
&self,
aSelectEventsSuppressed: bool
) -> nsresult

/**
   * This attribute is a boolean indicating whether or not the
   * "select" event should fire when the selection is changed using
   * one of our methods.  A view can use this to temporarily suppress
   * the selection while manipulating all of the indices, e.g., on
   * a sort.
   * Note: setting this attribute to false will fire a select event.
   */

attribute boolean selectEventsSuppressed;

pub unsafe fn GetCurrentIndex(&self, aCurrentIndex: *mut int32_t) -> nsresult

/**
   * The current item (the one that gets a focus rect in addition to being
   * selected).
   */

attribute long currentIndex;

pub unsafe fn SetCurrentIndex(&self, aCurrentIndex: int32_t) -> nsresult

/**
   * The current item (the one that gets a focus rect in addition to being
   * selected).
   */

attribute long currentIndex;

pub unsafe fn GetCurrentColumn(
&self,
aCurrentColumn: *mut *const nsITreeColumn
) -> nsresult

/**
   * The current column.
   */

attribute nsITreeColumn currentColumn;

pub unsafe fn SetCurrentColumn(
&self,
aCurrentColumn: *const nsITreeColumn
) -> nsresult

/**
   * The current column.
   */

attribute nsITreeColumn currentColumn;

pub unsafe fn GetShiftSelectPivot(
&self,
aShiftSelectPivot: *mut int32_t
) -> nsresult

/**
   * The selection "pivot".  This is the first item the user selected as
   * part of a ranged select.
   */

readonly attribute long shiftSelectPivot;