Struct xpcom::interfaces::nsIUDPSocket

#[repr(C)]

pub struct nsIUDPSocket { /* fields omitted */ }

interface nsIUDPSocket : nsISupports

/**
 * nsIUDPSocket
 *
 * An interface to a UDP socket that can accept incoming connections.
 */

Methods

impl nsIUDPSocket

pub fn coerce<T: nsIUDPSocketCoerce>(&self) -> &T

Cast this nsIUDPSocket to one of its base interfaces.

impl nsIUDPSocket

pub unsafe fn Close(&self) -> nsresult

/**
     * close
     *
     * This method closes a UDP socket.  This does not affect already
     * connected client sockets (i.e., the nsISocketTransport instances
     * created from this UDP socket).  This will cause the onStopListening
     * event to asynchronously fire with a status of NS_BINDING_ABORTED.
     */

void close ();

pub unsafe fn AsyncListen(
&self,
aListener: *const nsIUDPSocketListener
) -> nsresult

/**
     * asyncListen
     *
     * This method puts the UDP socket in the listening state.  It will
     * asynchronously listen for and accept client connections.  The listener
     * will be notified once for each client connection that is accepted.  The
     * listener's onSocketAccepted method will be called on the same thread
     * that called asyncListen (the calling thread must have a nsIEventTarget).
     *
     * The listener will be passed a reference to an already connected socket
     * transport (nsISocketTransport).  See below for more details.
     *
     * @param aListener
     *        The listener to be notified when client connections are accepted.
     */

void asyncListen (in nsIUDPSocketListener aListener);

pub unsafe fn GetLocalAddr(
&self,
aLocalAddr: *mut *const nsINetAddr
) -> nsresult

/**
     * Returns the local address of this UDP socket
     */

readonly attribute nsINetAddr localAddr;

pub unsafe fn GetPort(&self, aPort: *mut int32_t) -> nsresult

/**
     * Returns the port of this UDP socket.
     */

readonly attribute long port;

pub unsafe fn Send(
&self,
host: &nsACString,
port: uint16_t,
data: *const uint8_t,
dataLength: uint32_t,
_retval: *mut uint32_t
) -> nsresult

/**
     * send
     *
     * Send out the datagram to specified remote host and port.
     * DNS lookup will be triggered.
     *
     * @param host The remote host name.
     * @param port The remote port.
     * @param data The buffer containing the data to be written.
     * @param dataLength The maximum number of bytes to be written.
     * @return number of bytes written. (0 or dataLength)
     */

unsigned long send (in AUTF8String host, in unsigned short port, [array, size_is (dataLength), const] in uint8_t data, in unsigned long dataLength);

pub unsafe fn SendWithAddr(
&self,
addr: *const nsINetAddr,
data: *const uint8_t,
dataLength: uint32_t,
_retval: *mut uint32_t
) -> nsresult

/**
     * sendWithAddr
     *
     * Send out the datagram to specified remote host and port.
     *
     * @param addr The remote host address.
     * @param data The buffer containing the data to be written.
     * @param dataLength The maximum number of bytes to be written.
     * @return number of bytes written. (0 or dataLength)
     */

unsigned long sendWithAddr (in nsINetAddr addr, [array, size_is (dataLength), const] in uint8_t data, in unsigned long dataLength);

pub unsafe fn SendBinaryStream(
&self,
host: &nsACString,
port: uint16_t,
stream: *const nsIInputStream
) -> nsresult

/**
     * sendBinaryStream
     *
     * Send out the datagram to specified remote address and port.
     *
     * @param host The remote host name.
     * @param port The remote port.
     * @param stream The input stream to be sent. This must be a buffered stream implementation.
     */

void sendBinaryStream (in AUTF8String host, in unsigned short port, in nsIInputStream stream);

pub unsafe fn JoinMulticast(
&self,
addr: &nsACString,
iface: &nsACString
) -> nsresult

/**
     * joinMulticast
     *
     * Join the multicast group specified by |addr|.  You are then able to
     * receive future datagrams addressed to the group.
     *
     * @param addr
     *        The multicast group address.
     * @param iface
     *        The local address of the interface on which to join the group.  If
     *        this is not specified, the OS may join the group on all interfaces
     *        or only the primary interface.
     */

void joinMulticast (in AUTF8String addr, [optional] in AUTF8String iface);

pub unsafe fn LeaveMulticast(
&self,
addr: &nsACString,
iface: &nsACString
) -> nsresult

/**
     * leaveMulticast
     *
     * Leave the multicast group specified by |addr|.  You will no longer
     * receive future datagrams addressed to the group.
     *
     * @param addr
     *        The multicast group address.
     * @param iface
     *        The local address of the interface on which to leave the group.
     *        If this is not specified, the OS may leave the group on all
     *        interfaces or only the primary interface.
     */

void leaveMulticast (in AUTF8String addr, [optional] in AUTF8String iface);

pub unsafe fn GetMulticastLoopback(
&self,
aMulticastLoopback: *mut bool
) -> nsresult

/**
     * multicastLoopback
     *
     * Whether multicast datagrams sent via this socket should be looped back to
     * this host (assuming this host has joined the relevant group).  Defaults
     * to true.
     * Note: This is currently write-only.
     */

attribute boolean multicastLoopback;

pub unsafe fn SetMulticastLoopback(&self, aMulticastLoopback: bool) -> nsresult

/**
     * multicastLoopback
     *
     * Whether multicast datagrams sent via this socket should be looped back to
     * this host (assuming this host has joined the relevant group).  Defaults
     * to true.
     * Note: This is currently write-only.
     */

attribute boolean multicastLoopback;

pub unsafe fn GetMulticastInterface(
&self,
aMulticastInterface: &mut nsACString
) -> nsresult

/**
     * multicastInterface
     *
     * The interface that should be used for sending future multicast datagrams.
     * Note: This is currently write-only.
     */

attribute AUTF8String multicastInterface;

pub unsafe fn SetMulticastInterface(
&self,
aMulticastInterface: &nsACString
) -> nsresult

/**
     * multicastInterface
     *
     * The interface that should be used for sending future multicast datagrams.
     * Note: This is currently write-only.
     */

attribute AUTF8String multicastInterface;

pub unsafe fn GetRecvBufferSize(
&self,
aRecvBufferSize: *mut int32_t
) -> nsresult

/**
     * recvBufferSize
     *
     * The size of the receive buffer. Default depends on the OS.
     */

[noscript] attribute long recvBufferSize;

pub unsafe fn SetRecvBufferSize(&self, aRecvBufferSize: int32_t) -> nsresult

/**
     * recvBufferSize
     *
     * The size of the receive buffer. Default depends on the OS.
     */

[noscript] attribute long recvBufferSize;

pub unsafe fn GetSendBufferSize(
&self,
aSendBufferSize: *mut int32_t
) -> nsresult

/**
     * sendBufferSize
     *
     * The size of the send buffer. Default depends on the OS.
     */

[noscript] attribute long sendBufferSize;

pub unsafe fn SetSendBufferSize(&self, aSendBufferSize: int32_t) -> nsresult

/**
     * sendBufferSize
     *
     * The size of the send buffer. Default depends on the OS.
     */

[noscript] attribute long sendBufferSize;