Struct xpcom::interfaces::nsIURIFixupInfo

#[repr(C)]

pub struct nsIURIFixupInfo { /* fields omitted */ }

interface nsIURIFixupInfo : nsISupports

/**
 * Interface indicating what we found/corrected when fixing up a URI
 */

Methods

impl nsIURIFixupInfo

pub fn coerce<T: nsIURIFixupInfoCoerce>(&self) -> &T

Cast this nsIURIFixupInfo to one of its base interfaces.

impl nsIURIFixupInfo

pub unsafe fn GetConsumer(&self, aConsumer: *mut *const nsISupports) -> nsresult

/**
   * Consumer that asked for fixed up URI.
   */

attribute nsISupports consumer;

pub unsafe fn SetConsumer(&self, aConsumer: *const nsISupports) -> nsresult

/**
   * Consumer that asked for fixed up URI.
   */

attribute nsISupports consumer;

pub unsafe fn GetPreferredURI(
&self,
aPreferredURI: *mut *const nsIURI
) -> nsresult

/**
   * Our best guess as to what URI the consumer will want. Might
   * be null if we couldn't salvage anything (for instance, because
   * the input was invalid as a URI and FIXUP_FLAG_ALLOW_KEYWORD_LOOKUP
   * was not passed)
   */

readonly attribute nsIURI preferredURI;

pub unsafe fn GetFixedURI(&self, aFixedURI: *mut *const nsIURI) -> nsresult

/**
   * The fixed-up original input, *never* using a keyword search.
   * (might be null if the original input was not recoverable as
   * a URL, e.g. "foo bar"!)
   */

readonly attribute nsIURI fixedURI;

pub unsafe fn GetKeywordProviderName(
&self,
aKeywordProviderName: &mut nsAString
) -> nsresult

/**
   * The name of the keyword search provider used to provide a keyword search;
   * empty string if no keyword search was done.
   */

readonly attribute AString keywordProviderName;

pub unsafe fn GetKeywordAsSent(
&self,
aKeywordAsSent: &mut nsAString
) -> nsresult

/**
   * The keyword as used for the search (post trimming etc.)
   * empty string if no keyword search was done.
   */

readonly attribute AString keywordAsSent;

pub unsafe fn GetFixupChangedProtocol(
&self,
aFixupChangedProtocol: *mut bool
) -> nsresult

/**
   * Whether we changed the protocol instead of using one from the input as-is.
   */

readonly attribute boolean fixupChangedProtocol;

pub unsafe fn GetFixupCreatedAlternateURI(
&self,
aFixupCreatedAlternateURI: *mut bool
) -> nsresult

/**
   * Whether we created an alternative URI. We might have added a prefix and/or
   * suffix, the contents of which are controlled by the
   * browser.fixup.alternate.prefix and .suffix prefs, with the defaults being
   * "www." and ".com", respectively.
   */

readonly attribute boolean fixupCreatedAlternateURI;

pub unsafe fn GetOriginalInput(
&self,
aOriginalInput: &mut nsACString
) -> nsresult

/**
   * The original input
   */

readonly attribute AUTF8String originalInput;