Struct xpcom::interfaces::nsIUpdateManager

#[repr(C)]

pub struct nsIUpdateManager { /* fields omitted */ }

interface nsIUpdateManager : nsISupports

/**
 * An interface describing a global application service that maintains a list
 * of updates previously performed as well as the current active update.
 */

Methods

impl nsIUpdateManager

pub fn coerce<T: nsIUpdateManagerCoerce>(&self) -> &T

Cast this nsIUpdateManager to one of its base interfaces.

impl nsIUpdateManager

pub unsafe fn GetUpdateAt(
&self,
index: int32_t,
_retval: *mut *const nsIUpdate
) -> nsresult

/**
   * Gets the update at the specified index
   * @param   index
   *          The index within the updates array
   * @returns The nsIUpdate object at the specified index
   */

nsIUpdate getUpdateAt (in long index);

pub unsafe fn GetUpdateCount(&self, aUpdateCount: *mut int32_t) -> nsresult

/**
   * Gets the total number of updates in the history list.
   */

readonly attribute long updateCount;

pub unsafe fn GetActiveUpdate(
&self,
aActiveUpdate: *mut *const nsIUpdate
) -> nsresult

/**
   * The active (current) update. The active update is not in the history list.
   */

attribute nsIUpdate activeUpdate;

pub unsafe fn SetActiveUpdate(
&self,
aActiveUpdate: *const nsIUpdate
) -> nsresult

/**
   * The active (current) update. The active update is not in the history list.
   */

attribute nsIUpdate activeUpdate;

pub unsafe fn SaveUpdates(&self) -> nsresult

/**
   * Saves all updates to disk.
   */

void saveUpdates ();

pub unsafe fn RefreshUpdateStatus(&self) -> nsresult

/**
   * Refresh the update status based on the information in update.status.
   */

void refreshUpdateStatus ();

pub unsafe fn ElevationOptedIn(&self) -> nsresult

/**
   * The user agreed to proceed with an elevated update and we are now
   * permitted to show an elevation prompt.
   */

void elevationOptedIn ();

pub unsafe fn CleanupActiveUpdate(&self) -> nsresult

/**
   * Clean up and remove the active update without applying it.
   */

void cleanupActiveUpdate ();