Struct xpcom::interfaces::nsIUpdatePatch

#[repr(C)]

pub struct nsIUpdatePatch { /* fields omitted */ }

interface nsIUpdatePatch : nsISupports

/**
 * An interface that describes an object representing a patch file that can
 * be downloaded and applied to a version of this application so that it
 * can be updated.
 */

Methods

impl nsIUpdatePatch

pub fn coerce<T: nsIUpdatePatchCoerce>(&self) -> &T

Cast this nsIUpdatePatch to one of its base interfaces.

impl nsIUpdatePatch

pub unsafe fn GetType(&self, aType: &mut nsAString) -> nsresult

/**
   * The type of this patch:
   * "partial"      A binary difference between two application versions
   * "complete"     A complete patch containing all of the replacement files
   *                to update to the new version
   */

attribute AString type;

pub unsafe fn SetType(&self, aType: &nsAString) -> nsresult

/**
   * The type of this patch:
   * "partial"      A binary difference between two application versions
   * "complete"     A complete patch containing all of the replacement files
   *                to update to the new version
   */

attribute AString type;

pub unsafe fn GetURL(&self, aURL: &mut nsAString) -> nsresult

/**
   * The URL this patch was being downloaded from
   */

attribute AString URL;

pub unsafe fn SetURL(&self, aURL: &nsAString) -> nsresult

/**
   * The URL this patch was being downloaded from
   */

attribute AString URL;

pub unsafe fn GetFinalURL(&self, aFinalURL: &mut nsAString) -> nsresult

/**
   * The final URL this patch was being downloaded from
   */

attribute AString finalURL;

pub unsafe fn SetFinalURL(&self, aFinalURL: &nsAString) -> nsresult

/**
   * The final URL this patch was being downloaded from
   */

attribute AString finalURL;

pub unsafe fn GetSize(&self, aSize: *mut uint32_t) -> nsresult

/**
   * The size of this file, in bytes.
   */

attribute unsigned long size;

pub unsafe fn SetSize(&self, aSize: uint32_t) -> nsresult

/**
   * The size of this file, in bytes.
   */

attribute unsigned long size;

pub unsafe fn GetState(&self, aState: &mut nsAString) -> nsresult

/**
   * The state of this patch
   */

attribute AString state;

pub unsafe fn SetState(&self, aState: &nsAString) -> nsresult

/**
   * The state of this patch
   */

attribute AString state;

pub unsafe fn GetErrorCode(&self, aErrorCode: *mut int32_t) -> nsresult

/**
   * A numeric error code that conveys additional information about the state of
   * a failed update. If the update is not in the "failed" state the value is
   * zero. The possible values are located in common/errors.h and values between
   * 80 and 99 are in nsUpdateService.js.
   */

attribute long errorCode;

pub unsafe fn SetErrorCode(&self, aErrorCode: int32_t) -> nsresult

/**
   * A numeric error code that conveys additional information about the state of
   * a failed update. If the update is not in the "failed" state the value is
   * zero. The possible values are located in common/errors.h and values between
   * 80 and 99 are in nsUpdateService.js.
   */

attribute long errorCode;

pub unsafe fn GetSelected(&self, aSelected: *mut bool) -> nsresult

/**
   * true if this patch is currently selected as the patch to be downloaded and
   * installed for this update transaction, false if another patch from this
   * update has been selected.
   */

attribute boolean selected;

pub unsafe fn SetSelected(&self, aSelected: bool) -> nsresult

/**
   * true if this patch is currently selected as the patch to be downloaded and
   * installed for this update transaction, false if another patch from this
   * update has been selected.
   */

attribute boolean selected;

pub unsafe fn Serialize(
&self,
updates: *const nsIDOMDocument,
_retval: *mut *const nsIDOMElement
) -> nsresult

/**
   * Serializes this patch object into a DOM Element
   * @param   updates
   *          The document to serialize into
   * @returns The DOM Element created by the serialization process
   */

nsIDOMElement serialize (in nsIDOMDocument updates);