Struct xpcom::interfaces::nsIUrlClassifierDBService

#[repr(C)]

pub struct nsIUrlClassifierDBService { /* fields omitted */ }

interface nsIUrlClassifierDBService : nsISupports

/**
 * This is a proxy class that is instantiated and called from the JS thread.
 * It provides async methods for querying and updating the database.  As the
 * methods complete, they call the callback function.
 */

Methods

impl nsIUrlClassifierDBService

pub fn coerce<T: nsIUrlClassifierDBServiceCoerce>(&self) -> &T

Cast this nsIUrlClassifierDBService to one of its base interfaces.

impl nsIUrlClassifierDBService

pub unsafe fn Lookup(
&self,
principal: *const nsIPrincipal,
tables: &nsACString,
c: *const nsIUrlClassifierCallback
) -> nsresult

/**
   * Looks up a URI in the specified tables.
   *
   * @param principal: The principal containing the URI to search.
   * @param c: The callback will be called with a comma-separated list
   *        of tables to which the key belongs.
   */

void lookup (in nsIPrincipal principal, in ACString tables, in nsIUrlClassifierCallback c);

pub unsafe fn GetTables(&self, c: *const nsIUrlClassifierCallback) -> nsresult

/**
   * Lists the tables along with their meta info in the following format:
   *
   *   tablename;[metadata]\n
   *   tablename2;[metadata]\n
   *
   * For v2 tables, the metadata is the chunks info such as
   *
   *   goog-phish-shavar;a:10,14,30-40s:56,67
   *   goog-unwanted-shavar;a:1-3,5
   *
   * For v4 tables, base64 encoded state is currently the only info in the
   * metadata (can be extended whenever necessary). For exmaple,
   *
   *   goog-phish-proto;Cg0IARAGGAEiAzAwMTABEKqTARoCGAjT1gDD:oCGAjT1gDD\n
   *   goog-malware-proto;Cg0IAhAGGAEiAzAwMTABENCQARoCGAjx5Yty:BENCQARoCGAj\n
   *
   * Note that the metadata is colon-separated.
   *
   */

void getTables (in nsIUrlClassifierCallback c);

pub unsafe fn SetHashCompleter(
&self,
tableName: &nsACString,
completer: *const nsIUrlClassifierHashCompleter
) -> nsresult

/**
   * Set the nsIUrlClassifierCompleter object for a given table.  This
   * object will be used to request complete versions of partial
   * hashes.
   */

void setHashCompleter (in ACString tableName, in nsIUrlClassifierHashCompleter completer);

pub unsafe fn ClearLastResults(&self) -> nsresult

/**
   * Forget the results that were used in the last DB update.
   */

void clearLastResults ();

pub unsafe fn BeginUpdate(
&self,
updater: *const nsIUrlClassifierUpdateObserver,
tables: &nsACString
) -> nsresult

/**
   * Begin an update process.  Will throw NS_ERROR_NOT_AVAILABLE if there
   * is already an update in progress.
   *
   * @param updater The update observer tied to this update.
   * @param tables A comma-separated list of tables included in this update.
   */

void beginUpdate (in nsIUrlClassifierUpdateObserver updater, in ACString tables);

pub unsafe fn BeginStream(&self, table: &nsACString) -> nsresult

/**
   * Begin a stream update.  This should be called once per url being
   * fetched.
   *
   * @param table The table the contents of this stream will be associated
   *              with, or empty for the initial stream.
   */

void beginStream (in ACString table);

pub unsafe fn UpdateStream(&self, updateChunk: &nsACString) -> nsresult

/**
   * Update the table incrementally.
   */

void updateStream (in ACString updateChunk);

pub unsafe fn FinishStream(&self) -> nsresult

/**
   * Finish an individual stream update.  Must be called for every
   * beginStream() call, before the next beginStream() or finishUpdate().
   *
   * The update observer's streamFinished will be called once the
   * stream has been processed.
   */

void finishStream ();

pub unsafe fn FinishUpdate(&self) -> nsresult

/**
   * Finish an incremental update.  This will attempt to commit any
   * pending changes and resets the update interface.
   *
   * The update observer's updateSucceeded or updateError methods
   * will be called when the update has been processed.
   */

void finishUpdate ();

pub unsafe fn CancelUpdate(&self) -> nsresult

/**
   * Cancel an incremental update.  This rolls back any pending changes.
   * and resets the update interface.
   *
   * The update observer's updateError method will be called when the
   * update has been rolled back.
   */

void cancelUpdate ();

pub unsafe fn ResetDatabase(&self) -> nsresult

/**
   * Reset the url-classifier database.  This call will delete the existing
   * database, emptying all tables.  Mostly intended for use in unit tests.
   */

void resetDatabase ();

pub unsafe fn ReloadDatabase(&self) -> nsresult

/**
   * Reload he url-classifier database. This will empty all cache for
   * completions from gethash, and reload it from database. Mostly intended
   * for use in tests.
   */

void reloadDatabase ();

pub unsafe fn ClearCache(&self) -> nsresult

/**
   * Empty all the caches.
   */

void clearCache ();