Struct xpcom::interfaces::nsIUrlClassifierUtils

#[repr(C)]

pub struct nsIUrlClassifierUtils { /* fields omitted */ }

interface nsIUrlClassifierUtils : nsISupports

Methods

impl nsIUrlClassifierUtils

pub fn coerce<T: nsIUrlClassifierUtilsCoerce>(&self) -> &T

Cast this nsIUrlClassifierUtils to one of its base interfaces.

impl nsIUrlClassifierUtils

pub unsafe fn GetKeyForURI(
&self,
uri: *const nsIURI,
_retval: &mut nsACString
) -> nsresult

/**
   * Get the lookup string for a given URI.  This normalizes the hostname,
   * url-decodes the string, and strips off the protocol.
   *
   * @param uri URI to get the lookup key for.
   *
   * @returns String containing the canonicalized URI.
   */

ACString getKeyForURI (in nsIURI uri);

pub unsafe fn GetProvider(
&self,
tableName: &nsACString,
_retval: &mut nsACString
) -> nsresult

/**
   * Get the provider by table name.
   *
   * @param tableName The table name that we want to lookup
   *
   * @returns the provider name that the given table belongs.
   */

ACString getProvider (in ACString tableName);

pub unsafe fn GetTelemetryProvider(
&self,
tableName: &nsACString,
_retval: &mut nsACString
) -> nsresult

/**
   * Get the provider used for Telemetry.
   * Because recording Telemetry will leak user-controlled strings,
   * only built-in providers should be recorded.
   *
   * @param tableName The table name that we want to lookup
   *
   * @returns the filtered provider for telemetry.
   *
   */

ACString getTelemetryProvider (in ACString tableName);

pub unsafe fn GetProtocolVersion(
&self,
provider: &nsACString,
_retval: &mut nsACString
) -> nsresult

/**
   * Get the protocol version for the given provider.
   *
   * @param provider String the provider name. e.g. "google"
   *
   * @returns String to indicate the protocol version. e.g. "2.2"
   */

ACString getProtocolVersion (in ACString provider);

pub unsafe fn ConvertThreatTypeToListNames(
&self,
threatType: uint32_t,
_retval: &mut nsACString
) -> nsresult

/**
   * Convert threat type to list name.
   *
   * @param Integer to indicate threat type.
   *
   * @returns The list names separated by ','. For example,
   *          'goog-phish-proto,test-phish-proto'.
   */

ACString convertThreatTypeToListNames (in uint32_t threatType);

pub unsafe fn ConvertListNameToThreatType(
&self,
listName: &nsACString,
_retval: *mut uint32_t
) -> nsresult

/**
   * Convert list name to threat type.
   *
   * @param The list name.
   *
   * @returns The threat type in integer.
   */

uint32_t convertListNameToThreatType (in ACString listName);

pub unsafe fn MakeUpdateRequestV4(
&self,
aListNames: *mut *const c_char,
aStatesBase64: *mut *const c_char,
aCount: uint32_t,
_retval: &mut nsACString
) -> nsresult

/**
   * Make update request for given lists and their states.
   *
   * @param aListNames An array of list name represented in string.
   * @param aState An array of states (encoded in base64 format) for each list.
   * @param aCount The array length of aList and aState.
   *
   * @returns A base64url encoded string.
   */

ACString makeUpdateRequestV4 ([array, size_is (aCount)] in string aListNames, [array, size_is (aCount)] in string aStatesBase64, in uint32_t aCount);

pub unsafe fn MakeFindFullHashRequestV4(
&self,
aListNames: *mut *const c_char,
aListStatesBase64: *mut *const c_char,
aPrefixes: *mut *const c_char,
aListCount: uint32_t,
aPrefixCount: uint32_t,
_retval: &mut nsACString
) -> nsresult

/**
   * Make "find full hash" request by for the given prefixes.
   *
   * @param aListNames An array of list names represented in string.
   * @param aListStatesBase64 An array of list states represented in base64.
   * @param aPrefixes An array of prefixes for which we'd like to find full hashes..
   * @param aListCount The array length of aListNames
   * @param aPrefixCount The array length of aPrefixes
   *
   * @returns A base64url encoded string.
   */

ACString makeFindFullHashRequestV4 ([array, size_is (aListCount)] in string aListNames, [array, size_is (aListCount)] in string aListStatesBase64, [array, size_is (aPrefixCount)] in string aPrefixes, in uint32_t aListCount, in uint32_t aPrefixCount);

pub unsafe fn MakeThreatHitReport(
&self,
aChannel: *const nsIChannel,
aListName: &nsACString,
aHashBase64: &nsACString,
_retval: &mut nsACString
) -> nsresult

/**
   * Make ThreatHit report request body.
   *
   * @param aChannel channel which encountered the threat.
   * @param aListName listname represented in string.
   * @param aHashBase64 hash-based hit represented in base64.
   *
   * @returns A base64 encoded string.
   */

ACString makeThreatHitReport (in nsIChannel aChannel, in ACString aListName, in ACString aHashBase64);

pub unsafe fn ParseFindFullHashResponseV4(
&self,
aResponse: &nsACString,
aCallback: *const nsIUrlClassifierParseFindFullHashCallback
) -> nsresult

/**
   * Parse V4 FindFullHash response.
   *
   * @param aResponse Byte stream from the server.
   * @param aCallback The callback function on each complete hash parsed.
   *                  Can be called multiple times in one parsing.
   */

void parseFindFullHashResponseV4 (in ACString aResponse, in nsIUrlClassifierParseFindFullHashCallback aCallback);