Struct xpcom::interfaces::nsIUrlListManager

#[repr(C)]

pub struct nsIUrlListManager { /* fields omitted */ }

interface nsIUrlListManager : nsISupports

/**
 * Interface for a class that manages updates of the url classifier database.
 */

Methods

impl nsIUrlListManager

pub fn coerce<T: nsIUrlListManagerCoerce>(&self) -> &T

Cast this nsIUrlListManager to one of its base interfaces.

impl nsIUrlListManager

pub unsafe fn GetGethashUrl(
&self,
tableName: &nsACString,
_retval: &mut nsACString
) -> nsresult

/**
     * Get the gethash url for this table
     */

ACString getGethashUrl (in ACString tableName);

pub unsafe fn GetUpdateUrl(
&self,
tableName: &nsACString,
_retval: &mut nsACString
) -> nsresult

/**
     * Get the update url for this table
     */

ACString getUpdateUrl (in ACString tableName);

pub unsafe fn RegisterTable(
&self,
tableName: &nsACString,
providerName: &nsACString,
updateUrl: &nsACString,
gethashUrl: &nsACString,
_retval: *mut bool
) -> nsresult

/**
     * Add a table to the list of tables we are managing. The name is a
     * string of the format provider_name-semantic_type-table_type.  For
     * @param tableName A string of the format
     *        provider_name-semantic_type-table_type.  For example,
     *        goog-white-enchash or goog-black-url.
     * @param providerName The name of the entity providing the list.
     * @param updateUrl The URL from which to fetch updates.
     * @param gethashUrl The URL from which to fetch hash completions.
     */

boolean registerTable (in ACString tableName, in ACString providerName, in ACString updateUrl, in ACString gethashUrl);

pub unsafe fn UnregisterTable(&self, tableName: &nsACString) -> nsresult

/**
     * Unregister table from the list
     */

void unregisterTable (in ACString tableName);

pub unsafe fn EnableUpdate(&self, tableName: &nsACString) -> nsresult

/**
     * Turn on update checking for a table. I.e., during the next server
     * check, download updates for this table.
     */

void enableUpdate (in ACString tableName);

pub unsafe fn DisableUpdate(&self, tableName: &nsACString) -> nsresult

/**
     * Turn off update checking for a table.
     */

void disableUpdate (in ACString tableName);

pub unsafe fn MaybeToggleUpdateChecking(&self) -> nsresult

/**
     * Toggle update checking, if necessary.
     */

void maybeToggleUpdateChecking ();

pub unsafe fn CheckForUpdates(
&self,
updateUrl: &nsACString,
_retval: *mut bool
) -> nsresult

/**
     * This is currently used by about:url-classifier to force an update
     * for the update url. Update may still fail because of backoff algorithm.
     */

boolean checkForUpdates (in ACString updateUrl);

pub unsafe fn ForceUpdates(
&self,
tableNames: &nsACString,
_retval: *mut bool
) -> nsresult

/**
     * Force updates for the given tables, updates are still restricted to
     * backoff algorithm.
     * @param tables  A string lists all the tables that we want to trigger updates.
     *                table names are separated with ','.
     */

boolean forceUpdates (in ACString tableNames);

pub unsafe fn GetBackOffTime(
&self,
provider: &nsACString,
_retval: *mut uint64_t
) -> nsresult

/**
     * This is currently used by about:url-classifier to get back-off time
     * (in millisecond since epoch) for the given provider. Return 0 if we
     * are not in back-off mode.
     */

uint64_t getBackOffTime (in ACString provider);