Struct xpcom::interfaces::nsIWebBrowserChrome

#[repr(C)]

pub struct nsIWebBrowserChrome { /* fields omitted */ }

interface nsIWebBrowserChrome : nsISupports

/**
 * nsIWebBrowserChrome corresponds to the top-level, outermost window
 * containing an embedded Gecko web browser.
 */

Methods

impl nsIWebBrowserChrome

pub fn coerce<T: nsIWebBrowserChromeCoerce>(&self) -> &T

Cast this nsIWebBrowserChrome to one of its base interfaces.

impl nsIWebBrowserChrome

pub const STATUS_SCRIPT: i64

STATUS_SCRIPT: i64 = 1

pub const STATUS_LINK: i64

STATUS_LINK: i64 = 3

pub const CHROME_DEFAULT: i64

CHROME_DEFAULT: i64 = 1

/**
     * Definitions for the chrome flags
     */

pub const CHROME_WINDOW_BORDERS: i64

CHROME_WINDOW_BORDERS: i64 = 2

pub const CHROME_WINDOW_CLOSE: i64

CHROME_WINDOW_CLOSE: i64 = 4

pub const CHROME_WINDOW_RESIZE: i64

CHROME_WINDOW_RESIZE: i64 = 8

pub const CHROME_MENUBAR: i64

CHROME_MENUBAR: i64 = 16

pub const CHROME_TOOLBAR: i64

CHROME_TOOLBAR: i64 = 32

pub const CHROME_LOCATIONBAR: i64

CHROME_LOCATIONBAR: i64 = 64

pub const CHROME_STATUSBAR: i64

CHROME_STATUSBAR: i64 = 128

pub const CHROME_PERSONAL_TOOLBAR: i64

CHROME_PERSONAL_TOOLBAR: i64 = 256

pub const CHROME_SCROLLBARS: i64

CHROME_SCROLLBARS: i64 = 512

pub const CHROME_TITLEBAR: i64

CHROME_TITLEBAR: i64 = 1024

pub const CHROME_EXTRA: i64

CHROME_EXTRA: i64 = 2048

pub const CHROME_WITH_SIZE: i64

CHROME_WITH_SIZE: i64 = 4096

pub const CHROME_WITH_POSITION: i64

CHROME_WITH_POSITION: i64 = 8192

pub const CHROME_WINDOW_MIN: i64

CHROME_WINDOW_MIN: i64 = 16384

pub const CHROME_WINDOW_POPUP: i64

CHROME_WINDOW_POPUP: i64 = 32768

pub const CHROME_PRIVATE_WINDOW: i64

CHROME_PRIVATE_WINDOW: i64 = 65536

pub const CHROME_NON_PRIVATE_WINDOW: i64

CHROME_NON_PRIVATE_WINDOW: i64 = 131072

pub const CHROME_PRIVATE_LIFETIME: i64

CHROME_PRIVATE_LIFETIME: i64 = 262144

pub const CHROME_REMOTE_WINDOW: i64

CHROME_REMOTE_WINDOW: i64 = 1048576

pub const CHROME_SUPPRESS_ANIMATION: i64

CHROME_SUPPRESS_ANIMATION: i64 = 16777216

pub const CHROME_WINDOW_RAISED: i64

CHROME_WINDOW_RAISED: i64 = 33554432

pub const CHROME_WINDOW_LOWERED: i64

CHROME_WINDOW_LOWERED: i64 = 67108864

pub const CHROME_CENTER_SCREEN: i64

CHROME_CENTER_SCREEN: i64 = 134217728

pub const CHROME_DEPENDENT: i64

CHROME_DEPENDENT: i64 = 268435456

pub const CHROME_MODAL: i64

CHROME_MODAL: i64 = 536870912

pub const CHROME_OPENAS_DIALOG: i64

CHROME_OPENAS_DIALOG: i64 = 1073741824

pub const CHROME_OPENAS_CHROME: i64

CHROME_OPENAS_CHROME: i64 = 2147483648

pub const CHROME_ALL: i64

CHROME_ALL: i64 = 4094

pub unsafe fn SetStatus(
&self,
statusType: uint32_t,
status: *const int16_t
) -> nsresult

/**
     * Called when the status text in the chrome needs to be updated.
     * @param statusType indicates what is setting the text
     * @param status status string. null is an acceptable value meaning
     *               no status.
     */

void setStatus (in unsigned long statusType, in wstring status);

pub unsafe fn GetWebBrowser(
&self,
aWebBrowser: *mut *const nsIWebBrowser
) -> nsresult

/**
     * The currently loaded WebBrowser.  The browser chrome may be
     * told to set the WebBrowser object to a new object by setting this
     * attribute.  In this case the implementer is responsible for taking the
     * new WebBrowser object and doing any necessary initialization or setup
     * as if it had created the WebBrowser itself.  This includes positioning
     * setting up listeners etc.
     */

attribute nsIWebBrowser webBrowser;

pub unsafe fn SetWebBrowser(
&self,
aWebBrowser: *const nsIWebBrowser
) -> nsresult

/**
     * The currently loaded WebBrowser.  The browser chrome may be
     * told to set the WebBrowser object to a new object by setting this
     * attribute.  In this case the implementer is responsible for taking the
     * new WebBrowser object and doing any necessary initialization or setup
     * as if it had created the WebBrowser itself.  This includes positioning
     * setting up listeners etc.
     */

attribute nsIWebBrowser webBrowser;

pub unsafe fn GetChromeFlags(&self, aChromeFlags: *mut uint32_t) -> nsresult

/**
     * The chrome flags for this browser chrome. The implementation should
     * reflect the value of this attribute by hiding or showing its chrome
     * appropriately.
     */

attribute unsigned long chromeFlags;

pub unsafe fn SetChromeFlags(&self, aChromeFlags: uint32_t) -> nsresult

/**
     * The chrome flags for this browser chrome. The implementation should
     * reflect the value of this attribute by hiding or showing its chrome
     * appropriately.
     */

attribute unsigned long chromeFlags;

pub unsafe fn DestroyBrowserWindow(&self) -> nsresult

/**
     * Asks the implementer to destroy the window associated with this
     * WebBrowser object.
     */

void destroyBrowserWindow ();

pub unsafe fn SizeBrowserTo(&self, aCX: int32_t, aCY: int32_t) -> nsresult

/**
     * Tells the chrome to size itself such that the browser will be the
     * specified size.
     * @param aCX new width of the browser
     * @param aCY new height of the browser
     */

void sizeBrowserTo (in long aCX, in long aCY);

pub unsafe fn ShowAsModal(&self) -> nsresult

/**
     * Shows the window as a modal window.
     * @return (the function error code) the status value specified by
     *         in exitModalEventLoop.
     */

void showAsModal ();

pub unsafe fn IsWindowModal(&self, _retval: *mut bool) -> nsresult

/**
     * Is the window modal (that is, currently executing a modal loop)?
     * @return true if it's a modal window
     */

boolean isWindowModal ();

pub unsafe fn ExitModalEventLoop(&self, aStatus: nsresult) -> nsresult

/**
     * Exit a modal event loop if we're in one. The implementation
     * should also exit out of the loop if the window is destroyed.
     * @param aStatus - the result code to return from showAsModal
     */

void exitModalEventLoop (in nsresult aStatus);