Struct xpcom::interfaces::nsIWebBrowserChrome3

#[repr(C)]

pub struct nsIWebBrowserChrome3 { /* fields omitted */ }

interface nsIWebBrowserChrome3 : nsIWebBrowserChrome2

/**
 * nsIWebBrowserChrome3 is an extension to nsIWebBrowserChrome2.
 */

Methods

impl nsIWebBrowserChrome3

pub fn coerce<T: nsIWebBrowserChrome3Coerce>(&self) -> &T

Cast this nsIWebBrowserChrome3 to one of its base interfaces.

impl nsIWebBrowserChrome3

pub unsafe fn OnBeforeLinkTraversal(
&self,
originalTarget: &nsAString,
linkURI: *const nsIURI,
linkNode: *const nsIDOMNode,
isAppTab: bool,
_retval: &mut nsAString
) -> nsresult

/**
   * Determines the appropriate target for a link.
   *
   * @param originalTarget
   *        The original link target.
   * @param linkURI
   *        Link destination URI.
   * @param aDOMNode
   *        Link DOM node.
   * @param isAppTab
   *        Whether or not the link is in an app tab.
   * @returns A new link target, if appropriate.
   *          Otherwise returns originalTarget.
   */

AString onBeforeLinkTraversal (in AString originalTarget, in nsIURI linkURI, in nsIDOMNode linkNode, in boolean isAppTab);

pub unsafe fn ShouldLoadURI(
&self,
aDocShell: *const nsIDocShell,
aURI: *const nsIURI,
aReferrer: *const nsIURI,
aHasPostData: bool,
aTriggeringPrincipal: *const nsIPrincipal,
_retval: *mut bool
) -> nsresult

/**
   * Determines whether a load should continue.
   *
   * @param aDocShell
   *        The docshell performing the load.
   * @param aURI
   *        The URI being loaded.
   * @param aReferrer
   *        The referrer of the load.
   * @param aHasPostData
   *        True if the load which is being asked about has associated post data
   *        which would be discarded if the load was redirected across process
   *        boundaries.
   * @param aTriggeringPrincipal
   *        The principal that initiated the load of aURI.
   */

bool shouldLoadURI (in nsIDocShell aDocShell, in nsIURI aURI, in nsIURI aReferrer, in boolean aHasPostData, in nsIPrincipal aTriggeringPrincipal);

pub unsafe fn ShouldLoadURIInThisProcess(
&self,
aURI: *const nsIURI,
_retval: *mut bool
) -> nsresult

bool shouldLoadURIInThisProcess (in nsIURI aURI);

pub unsafe fn ReloadInFreshProcess(
&self,
aDocShell: *const nsIDocShell,
aURI: *const nsIURI,
aReferrer: *const nsIURI,
aTriggeringPrincipal: *const nsIPrincipal,
aLoadFlags: uint32_t,
_retval: *mut bool
) -> nsresult

/**
   * Attempts to load the currently loaded page into a fresh process to increase
   * available memory.
   *
   * @param aDocShell
   *        The docshell performing the load.
   */

bool reloadInFreshProcess (in nsIDocShell aDocShell, in nsIURI aURI, in nsIURI aReferrer, in nsIPrincipal aTriggeringPrincipal, in uint32_t aLoadFlags);

pub unsafe fn StartPrerenderingDocument(
&self,
aHref: *const nsIURI,
aReferrer: *const nsIURI,
aTriggeringPrincipal: *const nsIPrincipal
) -> nsresult

/**
   * Tell the browser to start prerendering the given document. This prerendering
   * _must_ be for the toplevel document.
   *
   * @param aHref The URI to begin prerendering
   * @param aReferrer The URI of the document requesting the prerender.
   * @param aTriggeringPrincipal The Principal that initiated to load aHref
   */

void startPrerenderingDocument (in nsIURI aHref, in nsIURI aReferrer, in nsIPrincipal aTriggeringPrincipal);

pub unsafe fn ShouldSwitchToPrerenderedDocument(
&self,
aHref: *const nsIURI,
aReferrer: *const nsIURI,
aSuccess: *const nsIRunnable,
aFailure: *const nsIRunnable,
_retval: *mut bool
) -> nsresult

/**
   * Check if there's a prerendered document which matches given URI /
   * referrer, and try to switch to the prerendered document immediately if
   * there is.
   *
   * @param aHref
   *        The URI which is being loaded.
   * @param aReferrer
   *        The referrer for the current load.
   * @param aSuccess
   *        (Optional) a runnable which will be run if the swap is successful.
   * @param aFailure
   *        (Optional) a runnable which will be run if the swap is not
   *        successful. Note it's not invoked if the function returns false.
   *
   * @return True if there is a matched prerendered document to swap with,
   *         false otherwise.
   */

bool shouldSwitchToPrerenderedDocument (in nsIURI aHref, in nsIURI aReferrer, in nsIRunnable aSuccess, in nsIRunnable aFailure);