Struct xpcom::interfaces::nsIWebBrowserFind

#[repr(C)]

pub struct nsIWebBrowserFind { /* fields omitted */ }

interface nsIWebBrowserFind : nsISupports

/**
 * nsIWebBrowserFind
 *
 * Searches for text in a web browser.
 *
 * Get one by doing a GetInterface on an nsIWebBrowser.
 *
 * By default, the implementation will search the focussed frame, or
 * if there is no focussed frame, the web browser content area. It
 * does not by default search subframes or iframes. To change this
 * behaviour, and to explicitly set the frame to search,
 * QueryInterface to nsIWebBrowserFindInFrames.
 */

Methods

impl nsIWebBrowserFind

pub fn coerce<T: nsIWebBrowserFindCoerce>(&self) -> &T

Cast this nsIWebBrowserFind to one of its base interfaces.

impl nsIWebBrowserFind

pub unsafe fn FindNext(&self, _retval: *mut bool) -> nsresult

/**
     * findNext
     *
     * Finds, highlights, and scrolls into view the next occurrence of the
     * search string, using the current search settings. Fails if the
     * search string is empty.
     *
     * @return  Whether an occurrence was found
     */

boolean findNext ();

pub unsafe fn GetSearchString(
&self,
aSearchString: *mut *const int16_t
) -> nsresult

/**
     * searchString
     *
     * The string to search for. This must be non-empty to search.
     */

attribute wstring searchString;

pub unsafe fn SetSearchString(&self, aSearchString: *const int16_t) -> nsresult

/**
     * searchString
     *
     * The string to search for. This must be non-empty to search.
     */

attribute wstring searchString;

pub unsafe fn GetFindBackwards(&self, aFindBackwards: *mut bool) -> nsresult

/**
     * findBackwards
     *
     * Whether to find backwards (towards the beginning of the document).
     * Default is false (search forward).
     */

attribute boolean findBackwards;

pub unsafe fn SetFindBackwards(&self, aFindBackwards: bool) -> nsresult

/**
     * findBackwards
     *
     * Whether to find backwards (towards the beginning of the document).
     * Default is false (search forward).
     */

attribute boolean findBackwards;

pub unsafe fn GetWrapFind(&self, aWrapFind: *mut bool) -> nsresult

/**
     * wrapFind
     *
     * Whether the search wraps around to the start (or end) of the document
     * if no match was found between the current position and the end (or
     * beginning). Works correctly when searching backwards. Default is
     * false.
     */

attribute boolean wrapFind;

pub unsafe fn SetWrapFind(&self, aWrapFind: bool) -> nsresult

/**
     * wrapFind
     *
     * Whether the search wraps around to the start (or end) of the document
     * if no match was found between the current position and the end (or
     * beginning). Works correctly when searching backwards. Default is
     * false.
     */

attribute boolean wrapFind;

pub unsafe fn GetEntireWord(&self, aEntireWord: *mut bool) -> nsresult

/**
     * entireWord
     *
     * Whether to match entire words only. Default is false.
     */

attribute boolean entireWord;

pub unsafe fn SetEntireWord(&self, aEntireWord: bool) -> nsresult

/**
     * entireWord
     *
     * Whether to match entire words only. Default is false.
     */

attribute boolean entireWord;

pub unsafe fn GetMatchCase(&self, aMatchCase: *mut bool) -> nsresult

/**
     * matchCase
     *
     * Whether to match case (case sensitive) when searching. Default is false.
     */

attribute boolean matchCase;

pub unsafe fn SetMatchCase(&self, aMatchCase: bool) -> nsresult

/**
     * matchCase
     *
     * Whether to match case (case sensitive) when searching. Default is false.
     */

attribute boolean matchCase;

pub unsafe fn GetSearchFrames(&self, aSearchFrames: *mut bool) -> nsresult

/**
     * searchFrames
     *
     * Whether to search through all frames in the content area. Default is true.
     *
     * Note that you can control whether the search propagates into child or
     * parent frames explicitly using nsIWebBrowserFindInFrames, but if one,
     * but not both, of searchSubframes and searchParentFrames are set, this
     * returns false.
     */

attribute boolean searchFrames;

pub unsafe fn SetSearchFrames(&self, aSearchFrames: bool) -> nsresult

/**
     * searchFrames
     *
     * Whether to search through all frames in the content area. Default is true.
     *
     * Note that you can control whether the search propagates into child or
     * parent frames explicitly using nsIWebBrowserFindInFrames, but if one,
     * but not both, of searchSubframes and searchParentFrames are set, this
     * returns false.
     */

attribute boolean searchFrames;