Struct xpcom::interfaces::nsIWebBrowserFindInFrames

#[repr(C)]

pub struct nsIWebBrowserFindInFrames { /* fields omitted */ }

interface nsIWebBrowserFindInFrames : nsISupports

/**
 * nsIWebBrowserFindInFrames
 *
 * Controls how find behaves when multiple frames or iframes are present.
 *
 * Get by doing a QueryInterface from nsIWebBrowserFind.
 */

Methods

impl nsIWebBrowserFindInFrames

pub fn coerce<T: nsIWebBrowserFindInFramesCoerce>(&self) -> &T

Cast this nsIWebBrowserFindInFrames to one of its base interfaces.

impl nsIWebBrowserFindInFrames

pub unsafe fn GetCurrentSearchFrame(
&self,
aCurrentSearchFrame: *mut *const mozIDOMWindowProxy
) -> nsresult

/**
     * currentSearchFrame
     *
     * Frame at which to start the search. Once the search is done, this will
     * be set to be the last frame searched, whether or not a result was found.
     * Has to be equal to or contained within the rootSearchFrame.
     */

attribute mozIDOMWindowProxy currentSearchFrame;

pub unsafe fn SetCurrentSearchFrame(
&self,
aCurrentSearchFrame: *const mozIDOMWindowProxy
) -> nsresult

/**
     * currentSearchFrame
     *
     * Frame at which to start the search. Once the search is done, this will
     * be set to be the last frame searched, whether or not a result was found.
     * Has to be equal to or contained within the rootSearchFrame.
     */

attribute mozIDOMWindowProxy currentSearchFrame;

pub unsafe fn GetRootSearchFrame(
&self,
aRootSearchFrame: *mut *const mozIDOMWindowProxy
) -> nsresult

/**
     * rootSearchFrame
     *
     * Frame within which to confine the search (normally the content area frame).
     * Set this to only search a subtree of the frame hierarchy.
     */

attribute mozIDOMWindowProxy rootSearchFrame;

pub unsafe fn SetRootSearchFrame(
&self,
aRootSearchFrame: *const mozIDOMWindowProxy
) -> nsresult

/**
     * rootSearchFrame
     *
     * Frame within which to confine the search (normally the content area frame).
     * Set this to only search a subtree of the frame hierarchy.
     */

attribute mozIDOMWindowProxy rootSearchFrame;

pub unsafe fn GetSearchSubframes(&self, aSearchSubframes: *mut bool) -> nsresult

/**
     * searchSubframes
     *
     * Whether to recurse down into subframes while searching. Default is true.
     *
     * Setting nsIWebBrowserfind.searchFrames to true sets this to true.
     */

attribute boolean searchSubframes;

pub unsafe fn SetSearchSubframes(&self, aSearchSubframes: bool) -> nsresult

/**
     * searchSubframes
     *
     * Whether to recurse down into subframes while searching. Default is true.
     *
     * Setting nsIWebBrowserfind.searchFrames to true sets this to true.
     */

attribute boolean searchSubframes;

pub unsafe fn GetSearchParentFrames(
&self,
aSearchParentFrames: *mut bool
) -> nsresult

/**
     * searchParentFrames
     *
     * Whether to allow the search to propagate out of the currentSearchFrame into its
     * parent frame(s). Search is always confined within the rootSearchFrame. Default
     * is true.
     *
     * Setting nsIWebBrowserfind.searchFrames to true sets this to true.
     */

attribute boolean searchParentFrames;

pub unsafe fn SetSearchParentFrames(
&self,
aSearchParentFrames: bool
) -> nsresult

/**
     * searchParentFrames
     *
     * Whether to allow the search to propagate out of the currentSearchFrame into its
     * parent frame(s). Search is always confined within the rootSearchFrame. Default
     * is true.
     *
     * Setting nsIWebBrowserfind.searchFrames to true sets this to true.
     */

attribute boolean searchParentFrames;