Struct xpcom::interfaces::nsIWebBrowserFocus

#[repr(C)]

pub struct nsIWebBrowserFocus { /* fields omitted */ }

interface nsIWebBrowserFocus : nsISupports

/**
 * nsIWebBrowserFocus
 * Interface that embedders use for controlling and interacting
 * with the browser focus management. The embedded browser can be focused by
 * clicking in it or tabbing into it. If the browser is currently focused and
 * the embedding application's top level window is disabled, deactivate() must
 * be called, and activate() called again when the top level window is
 * reactivated for the browser's focus memory to work correctly.
 */

Methods

impl nsIWebBrowserFocus

pub fn coerce<T: nsIWebBrowserFocusCoerce>(&self) -> &T

Cast this nsIWebBrowserFocus to one of its base interfaces.

impl nsIWebBrowserFocus

pub unsafe fn Activate(&self) -> nsresult

/**
   * MANDATORY
   * activate() is a mandatory call that must be made to the browser
   * when the embedding application's window is activated *and* the
   * browser area was the last thing in focus.  This method can also be called
   * if the embedding application wishes to give the browser area focus,
   * without affecting the currently focused element within the browser.
   *
   * @note
   * If you fail to make this call, mozilla focus memory will not work
   * correctly.
   */

void activate ();

pub unsafe fn Deactivate(&self) -> nsresult

/**
   * MANDATORY
   * deactivate() is a mandatory call that must be made to the browser
   * when the embedding application's window is deactivated *and* the
   * browser area was the last thing in focus.  On non-windows platforms,
   * deactivate() should also be called when focus moves from the browser
   * to the embedding chrome.
   *
   * @note
   * If you fail to make this call, mozilla focus memory will not work
   * correctly.
   */

void deactivate ();

pub unsafe fn SetFocusAtFirstElement(&self) -> nsresult

/**
   * Give the first element focus within mozilla
   * (i.e. TAB was pressed and focus should enter mozilla)
   */

void setFocusAtFirstElement ();

pub unsafe fn SetFocusAtLastElement(&self) -> nsresult

/**
   * Give the last element focus within mozilla
   * (i.e. SHIFT-TAB was pressed and focus should enter mozilla)
   */

void setFocusAtLastElement ();

pub unsafe fn GetFocusedWindow(
&self,
aFocusedWindow: *mut *const mozIDOMWindowProxy
) -> nsresult

/**
   * The currently focused nsDOMWindow when the browser is active,
   * or the last focused nsDOMWindow when the browser is inactive.
   */

attribute mozIDOMWindowProxy focusedWindow;

pub unsafe fn SetFocusedWindow(
&self,
aFocusedWindow: *const mozIDOMWindowProxy
) -> nsresult

/**
   * The currently focused nsDOMWindow when the browser is active,
   * or the last focused nsDOMWindow when the browser is inactive.
   */

attribute mozIDOMWindowProxy focusedWindow;

pub unsafe fn GetFocusedElement(
&self,
aFocusedElement: *mut *const nsIDOMElement
) -> nsresult

/**
   * The currently focused nsDOMElement when the browser is active,
   * or the last focused nsDOMElement when the browser is inactive.
   */

attribute nsIDOMElement focusedElement;

pub unsafe fn SetFocusedElement(
&self,
aFocusedElement: *const nsIDOMElement
) -> nsresult

/**
   * The currently focused nsDOMElement when the browser is active,
   * or the last focused nsDOMElement when the browser is inactive.
   */

attribute nsIDOMElement focusedElement;