Struct xpcom::interfaces::nsIWebBrowserPersistDocument

#[repr(C)]

pub struct nsIWebBrowserPersistDocument { /* fields omitted */ }

interface nsIWebBrowserPersistDocument : nsISupports

/**
 * Interface representing a document that can be serialized with
 * nsIWebBrowserPersist; it may or may not be in this process.  Some
 * information is exposed as attributes, which may or may not reflect
 * changes made to the underlying document; most of these are
 * self-explanatory from their names and types.
 */

Methods

impl nsIWebBrowserPersistDocument

pub fn coerce<T: nsIWebBrowserPersistDocumentCoerce>(&self) -> &T

Cast this nsIWebBrowserPersistDocument to one of its base interfaces.

impl nsIWebBrowserPersistDocument

pub unsafe fn GetIsPrivate(&self, aIsPrivate: *mut bool) -> nsresult

readonly attribute boolean isPrivate;

pub unsafe fn GetDocumentURI(&self, aDocumentURI: &mut nsACString) -> nsresult

readonly attribute AUTF8String documentURI;

pub unsafe fn GetBaseURI(&self, aBaseURI: &mut nsACString) -> nsresult

readonly attribute AUTF8String baseURI;

pub unsafe fn GetContentType(&self, aContentType: &mut nsACString) -> nsresult

readonly attribute ACString contentType;

pub unsafe fn GetCharacterSet(&self, aCharacterSet: &mut nsACString) -> nsresult

readonly attribute ACString characterSet;

pub unsafe fn GetTitle(&self, aTitle: &mut nsAString) -> nsresult

readonly attribute AString title;

pub unsafe fn GetReferrer(&self, aReferrer: &mut nsAString) -> nsresult

readonly attribute AString referrer;

pub unsafe fn GetContentDisposition(
&self,
aContentDisposition: &mut nsAString
) -> nsresult

readonly attribute AString contentDisposition;

pub unsafe fn GetPostData(
&self,
aPostData: *mut *const nsIInputStream
) -> nsresult

readonly attribute nsIInputStream postData;

pub unsafe fn GetCacheKey(&self, aCacheKey: *mut uint32_t) -> nsresult

/**
   * The cache key.  Unlike in nsISHEntry, where it's wrapped in an
   * nsISupportsPRUint32, this is just the integer.
   */

readonly attribute unsigned long cacheKey;

pub unsafe fn GetPersistFlags(&self, aPersistFlags: *mut uint32_t) -> nsresult

/**
   * This attribute is set by nsIWebBrowserPersist implementations to
   * propagate persist flags that apply to the DOM traversal and
   * serialization (rather than to managing file I/O).
   */

attribute unsigned long persistFlags;

pub unsafe fn SetPersistFlags(&self, aPersistFlags: uint32_t) -> nsresult

/**
   * This attribute is set by nsIWebBrowserPersist implementations to
   * propagate persist flags that apply to the DOM traversal and
   * serialization (rather than to managing file I/O).
   */

attribute unsigned long persistFlags;

pub unsafe fn ReadResources(
&self,
aVisitor: *const nsIWebBrowserPersistResourceVisitor
) -> nsresult

/**
   * Walk the DOM searching for external resources needed to render it.
   * The visitor callbacks may be called either before or after
   * readResources returns.
   *
   * @see nsIWebBrowserPersistResourceVisitor
   */

void readResources (in nsIWebBrowserPersistResourceVisitor aVisitor);

pub unsafe fn WriteContent(
&self,
aStream: *const nsIOutputStream,
aURIMap: *const nsIWebBrowserPersistURIMap,
aRequestedContentType: &nsACString,
aEncoderFlags: uint32_t,
aWrapColumn: uint32_t,
aCompletion: *const nsIWebBrowserPersistWriteCompletion
) -> nsresult

/**
   * Serialize the document's DOM.
   *
   * @param aStream       The output stream to write the document to.
   *
   * @param aURIMap       Optional; specifies URI rewriting to perform on
   *                      external references (as read by readResources).
   *                      If given, also causes relative hyperlinks to be
   *                      converted to absolute in the written text.
   *
   * @param aRequestedContentType
   *                      The desired MIME type to save the document as;
   *                      optional and defaults to the document's type.
   *                      (If no encoder exists for that type, "text/html"
   *                      is used instead.)
   *
   * @param aEncoderFlags Flags to pass to the encoder.
   *
   * @param aWrapColumn   Desired text width, ignored if wrapping is not
   *                      specified by the encoding flags, or if 0.
   *
   * @param aCompletion   Callback invoked when writing is complete.
   *                      It may be called either before or after writeContent
   *                      returns.
   *
   * @see nsIDocumentEncoder
   */

void writeContent (in nsIOutputStream aStream, in nsIWebBrowserPersistURIMap aURIMap, in ACString aRequestedContentType, in unsigned long aEncoderFlags, in unsigned long aWrapColumn, in nsIWebBrowserPersistWriteCompletion aCompletion);