Struct xpcom::interfaces::nsIWebBrowserPrint

#[repr(C)]

pub struct nsIWebBrowserPrint { /* fields omitted */ }

interface nsIWebBrowserPrint : nsISupports

/**
 * nsIWebBrowserPrint corresponds to the main interface
 * for printing an embedded Gecko web browser window/document
 */

Methods

impl nsIWebBrowserPrint

pub fn coerce<T: nsIWebBrowserPrintCoerce>(&self) -> &T

Cast this nsIWebBrowserPrint to one of its base interfaces.

impl nsIWebBrowserPrint

pub const PRINTPREVIEW_GOTO_PAGENUM: i64

PRINTPREVIEW_GOTO_PAGENUM: i64 = 0

/**
   * PrintPreview Navigation Constants
   */

pub const PRINTPREVIEW_PREV_PAGE: i64

PRINTPREVIEW_PREV_PAGE: i64 = 1

pub const PRINTPREVIEW_NEXT_PAGE: i64

PRINTPREVIEW_NEXT_PAGE: i64 = 2

pub const PRINTPREVIEW_HOME: i64

PRINTPREVIEW_HOME: i64 = 3

pub const PRINTPREVIEW_END: i64

PRINTPREVIEW_END: i64 = 4

pub unsafe fn GetGlobalPrintSettings(
&self,
aGlobalPrintSettings: *mut *const nsIPrintSettings
) -> nsresult

/**
   * Returns a "global" PrintSettings object
   * Creates a new the first time, if one doesn't exist.
   *
   * Then returns the same object each time after that.
   *
   * Initializes the globalPrintSettings from the default printer
   */

readonly attribute nsIPrintSettings globalPrintSettings;

pub unsafe fn GetCurrentPrintSettings(
&self,
aCurrentPrintSettings: *mut *const nsIPrintSettings
) -> nsresult

/**
   * Returns a pointer to the PrintSettings object that
   * that was passed into either "print" or "print preview"
   *
   * This enables any consumers of the interface to have access
   * to the "current" PrintSetting at later points in the execution
   */

readonly attribute nsIPrintSettings currentPrintSettings;

pub unsafe fn GetCurrentChildDOMWindow(
&self,
aCurrentChildDOMWindow: *mut *const mozIDOMWindowProxy
) -> nsresult

/**
   * Returns a pointer to the current child DOMWindow
   * that is being print previewed. (FrameSet Frames)
   *
   * Returns null if parent document is not a frameset or the entire FrameSet
   * document is being print previewed
   *
   * This enables any consumers of the interface to have access
   * to the "current" child DOMWindow at later points in the execution
   */

readonly attribute mozIDOMWindowProxy currentChildDOMWindow;

pub unsafe fn GetDoingPrint(&self, aDoingPrint: *mut bool) -> nsresult

/**
   * Returns whether it is in Print mode
   */

readonly attribute boolean doingPrint;

pub unsafe fn GetDoingPrintPreview(
&self,
aDoingPrintPreview: *mut bool
) -> nsresult

/**
   * Returns whether it is in Print Preview mode
   */

readonly attribute boolean doingPrintPreview;

pub unsafe fn GetIsFramesetDocument(
&self,
aIsFramesetDocument: *mut bool
) -> nsresult

/**
   * This returns whether the current document is a frameset document
   */

readonly attribute boolean isFramesetDocument;

pub unsafe fn GetIsFramesetFrameSelected(
&self,
aIsFramesetFrameSelected: *mut bool
) -> nsresult

/**
   * This returns whether the current document is a frameset document
   */

readonly attribute boolean isFramesetFrameSelected;

pub unsafe fn GetIsIFrameSelected(
&self,
aIsIFrameSelected: *mut bool
) -> nsresult

/**
   * This returns whether there is an IFrame selected
   */

readonly attribute boolean isIFrameSelected;

pub unsafe fn GetIsRangeSelection(
&self,
aIsRangeSelection: *mut bool
) -> nsresult

/**
   * This returns whether there is a "range" selection
   */

readonly attribute boolean isRangeSelection;

pub unsafe fn GetPrintPreviewNumPages(
&self,
aPrintPreviewNumPages: *mut int32_t
) -> nsresult

/**
   * This returns the total number of pages for the Print Preview
   */

readonly attribute long printPreviewNumPages;

pub unsafe fn Print(
&self,
aThePrintSettings: *const nsIPrintSettings,
aWPListener: *const nsIWebProgressListener
) -> nsresult

/**
   * Print the specified DOM window
   *
   * @param aThePrintSettings - Printer Settings for the print job, if aThePrintSettings is null
   *                            then the global PS will be used.
   * @param aWPListener - is updated during the print
   * @return void
   */

void print (in nsIPrintSettings aThePrintSettings, in nsIWebProgressListener aWPListener);

pub unsafe fn PrintPreview(
&self,
aThePrintSettings: *const nsIPrintSettings,
aChildDOMWin: *const mozIDOMWindowProxy,
aWPListener: *const nsIWebProgressListener
) -> nsresult

/**
   * Print Preview the specified DOM window
   *
   * @param aThePrintSettings - Printer Settings for the print preview, if aThePrintSettings is null
   *                            then the global PS will be used.
   * @param aChildDOMWin - DOM Window to be print previewed.
   * @param aWPListener - is updated during the printpreview
   * @return void
   */

void printPreview (in nsIPrintSettings aThePrintSettings, in mozIDOMWindowProxy aChildDOMWin, in nsIWebProgressListener aWPListener);

pub unsafe fn PrintPreviewNavigate(
&self,
aNavType: int16_t,
aPageNum: int32_t
) -> nsresult

/**
   * Print Preview - Navigates within the window
   *
   * @param aNavType - navigation enum
   * @param aPageNum - page num to navigate to when aNavType = ePrintPreviewGoToPageNum
   * @return void
   */

void printPreviewNavigate (in short aNavType, in long aPageNum);

pub unsafe fn Cancel(&self) -> nsresult

/**
   * Cancels the current print
   * @return void
   */

void cancel ();

pub unsafe fn EnumerateDocumentNames(
&self,
aCount: *mut uint32_t,
aResult: *mut *mut *const int16_t
) -> nsresult

/**
   * Returns an array of the names of all documents names (Title or URL)
   * and sub-documents. This will return a single item if the attr "isFramesetDocument" is false
   * and may return any number of items is "isFramesetDocument" is true
   *
   * @param  aCount - returns number of printers returned
   * @param  aResult - returns array of names
   * @return void
   */

void enumerateDocumentNames (out uint32_t aCount, [array, size_is (aCount), retval] out wstring aResult);

pub unsafe fn ExitPrintPreview(&self) -> nsresult

/**
   * This exists PrintPreview mode and returns browser window to galley mode
   * @return void
   */

void exitPrintPreview ();