Struct xpcom::interfaces::nsIWebContentConverterService

#[repr(C)]

pub struct nsIWebContentConverterService { /* fields omitted */ }

interface nsIWebContentConverterService : nsIWebContentHandlerRegistrar

Methods

impl nsIWebContentConverterService

pub fn coerce<T: nsIWebContentConverterServiceCoerce>(&self) -> &T

Cast this nsIWebContentConverterService to one of its base interfaces.

impl nsIWebContentConverterService

pub unsafe fn SetAutoHandler(
&self,
contentType: &nsAString,
handler: *const nsIWebContentHandlerInfo
) -> nsresult

/**
   * Specifies the handler to be used to automatically handle all links of a
   * certain content type from now on.
   * @param   contentType
   *          The content type to automatically load with the specified handler
   * @param   handler
   *          A web service handler. If this is null, no automatic action is
   *          performed and the user must choose.
   * @throws  NS_ERROR_NOT_AVAILABLE if the service refered to by |handler| is
   *          not already registered.
   */

void setAutoHandler (in AString contentType, in nsIWebContentHandlerInfo handler);

pub unsafe fn GetAutoHandler(
&self,
contentType: &nsAString,
_retval: *mut *const nsIWebContentHandlerInfo
) -> nsresult

/**
   * Gets the auto handler specified for a particular content type
   * @param   contentType
   *          The content type to look up an auto handler for.
   * @returns The web service handler that will automatically handle all
   *          documents of the specified type. null if there is no automatic
   *          handler. (Handlers may be registered, just none of them specified
   *          as "automatic").
   */

nsIWebContentHandlerInfo getAutoHandler (in AString contentType);

pub unsafe fn GetWebContentHandlerByURI(
&self,
contentType: &nsAString,
uri: &nsAString,
_retval: *mut *const nsIWebContentHandlerInfo
) -> nsresult

/**
   * Gets a web handler for the specified service URI
   * @param   contentType
   *          The content type of the service being located
   * @param   uri
   *          The service URI of the handler to locate.
   * @returns A web service handler that uses the specified uri.
   */

nsIWebContentHandlerInfo getWebContentHandlerByURI (in AString contentType, in AString uri);

pub unsafe fn LoadPreferredHandler(
&self,
request: *const nsIRequest
) -> nsresult

/**
   * Loads the preferred handler when content of a registered type is about
   * to be loaded.
   * @param   request
   *          The nsIRequest for the load of the content
   */

void loadPreferredHandler (in nsIRequest request);

pub unsafe fn RemoveProtocolHandler(
&self,
protocol: &nsAString,
uri: &nsAString
) -> nsresult

/**
   * Removes a registered protocol handler
   * @param   protocol
   *          The protocol scheme to remove a service handler for
   * @param   uri
   *          The uri of the service handler to remove
   */

void removeProtocolHandler (in AString protocol, in AString uri);

pub unsafe fn RemoveContentHandler(
&self,
contentType: &nsAString,
uri: &nsAString
) -> nsresult

/**
   * Removes a registered content handler
   * @param   contentType
   *          The content type to remove a service handler for
   * @param   uri
   *          The uri of the service handler to remove
   */

void removeContentHandler (in AString contentType, in AString uri);

pub unsafe fn GetContentHandlers(
&self,
contentType: &nsAString,
count: *mut uint32_t,
handlers: *mut *mut *const nsIWebContentHandlerInfo
) -> nsresult

/**
   * Gets the list of content handlers for a particular type.
   * @param   contentType
   *          The content type to get handlers for
   * @returns An array of nsIWebContentHandlerInfo objects
   */

void getContentHandlers (in AString contentType, [optional] out unsigned long count, [array, size_is (count), retval] out nsIWebContentHandlerInfo handlers);

pub unsafe fn ResetHandlersForType(&self, contentType: &nsAString) -> nsresult

/**
   * Resets the list of available content handlers to the default set from
   * the distribution.
   * @param   contentType
   *          The content type to reset handlers for
   */

void resetHandlersForType (in AString contentType);