Struct xpcom::interfaces::nsIWebHandlerApp

#[repr(C)]

pub struct nsIWebHandlerApp { /* fields omitted */ }

interface nsIWebHandlerApp : nsIHandlerApp

/**
 * nsIWebHandlerApp is a web-based handler, as speced by the WhatWG HTML5
 * draft.  Currently, only GET-based handlers are supported.  At some point,
 * we probably want to work with WhatWG to spec out and implement POST-based
 * handlers as well.
 */

Methods

impl nsIWebHandlerApp

pub fn coerce<T: nsIWebHandlerAppCoerce>(&self) -> &T

Cast this nsIWebHandlerApp to one of its base interfaces.

impl nsIWebHandlerApp

pub unsafe fn GetUriTemplate(&self, aUriTemplate: &mut nsACString) -> nsresult

/**
     * Template used to construct the URI to GET.  Template is expected to have
     * a %s in it, and the escaped URI to be handled is inserted in place of
     * that %s, as per the HTML5 spec.
     */

attribute AUTF8String uriTemplate;

pub unsafe fn SetUriTemplate(&self, aUriTemplate: &nsACString) -> nsresult

/**
     * Template used to construct the URI to GET.  Template is expected to have
     * a %s in it, and the escaped URI to be handled is inserted in place of
     * that %s, as per the HTML5 spec.
     */

attribute AUTF8String uriTemplate;