Struct xpcom::interfaces::nsIWebNavigationInfo

#[repr(C)]

pub struct nsIWebNavigationInfo { /* fields omitted */ }

interface nsIWebNavigationInfo : nsISupports

/**
 * The nsIWebNavigationInfo interface exposes a way to get information
 * on the capabilities of Gecko webnavigation objects.
 */

Methods

impl nsIWebNavigationInfo

pub fn coerce<T: nsIWebNavigationInfoCoerce>(&self) -> &T

Cast this nsIWebNavigationInfo to one of its base interfaces.

impl nsIWebNavigationInfo

pub const UNSUPPORTED: i64

UNSUPPORTED: i64 = 0

/**
   * Returned by isTypeSupported to indicate lack of support for a type.
   * @note this is guaranteed not to change, so that boolean tests can be done
   * on the return value if isTypeSupported to detect whether a type is
   * supported at all.
   */

pub const IMAGE: i64

IMAGE: i64 = 1

/**
   * Returned by isTypeSupported to indicate that a type is supported as an
   * image.
   */

pub const PLUGIN: i64

PLUGIN: i64 = 2

/**
   * Returned by isTypeSupported to indicate that a type is supported via an
   * NPAPI ("Netscape 4 API") plug-in.  This is not the value returned for
   * "XPCOM plug-ins".
   */

pub const OTHER: i64

OTHER: i64 = 32768

/**
   * @note Other return types may be added here in the future as they become
   * relevant.
   */
/**
   * Returned by isTypeSupported to indicate that a type is supported via some
   * other means.
   */

pub unsafe fn IsTypeSupported(
&self,
aType: &nsACString,
aWebNav: *const nsIWebNavigation,
_retval: *mut uint32_t
) -> nsresult

/**
   * Query whether aType is supported.
   * @param aType the MIME type in question.
   * @param aWebNav the nsIWebNavigation object for which the request
   *        is being made.  This is allowed to be null.  If it is non-null,
   *        the return value of this method may depend on the exact state of
   *        aWebNav and the values set through nsIWebBrowserSetup; otherwise
   *        the method will assume that the caller is interested in information
   *        about nsIWebNavigation objects in their default state.
   * @return an enum value indicating whether and how aType is supported.
   * @note This method may rescan plugins to ensure that they're properly
   *       registered for the types they support.
   */

unsigned long isTypeSupported (in ACString aType, in nsIWebNavigation aWebNav);