Struct xpcom::interfaces::nsIWebProgress

#[repr(C)]

pub struct nsIWebProgress { /* fields omitted */ }

interface nsIWebProgress : nsISupports

/**
 * The nsIWebProgress interface is used to add or remove nsIWebProgressListener
 * instances to observe the loading of asynchronous requests (usually in the
 * context of a DOM window).
 *
 * nsIWebProgress instances may be arranged in a parent-child configuration,
 * corresponding to the parent-child configuration of their respective DOM
 * windows.  However, in some cases a nsIWebProgress instance may not have an
 * associated DOM window.  The parent-child relationship of nsIWebProgress
 * instances is not made explicit by this interface, but the relationship may
 * exist in some implementations.
 *
 * A nsIWebProgressListener instance receives notifications for the
 * nsIWebProgress instance to which it added itself, and it may also receive
 * notifications from any nsIWebProgress instances that are children of that
 * nsIWebProgress instance.
 */

Methods

impl nsIWebProgress

pub fn coerce<T: nsIWebProgressCoerce>(&self) -> &T

Cast this nsIWebProgress to one of its base interfaces.

impl nsIWebProgress

pub const NOTIFY_STATE_REQUEST: i64

NOTIFY_STATE_REQUEST: i64 = 1

/**
   * The following flags may be combined to form the aNotifyMask parameter for
   * the addProgressListener method.  They limit the set of events that are
   * delivered to an nsIWebProgressListener instance.
   */
/**
   * These flags indicate the state transistions to observe, corresponding to
   * nsIWebProgressListener::onStateChange.
   *
   * NOTIFY_STATE_REQUEST
   *   Only receive the onStateChange event if the aStateFlags parameter
   *   includes nsIWebProgressListener::STATE_IS_REQUEST.
   *
   * NOTIFY_STATE_DOCUMENT
   *   Only receive the onStateChange event if the aStateFlags parameter
   *   includes nsIWebProgressListener::STATE_IS_DOCUMENT.
   *
   * NOTIFY_STATE_NETWORK
   *   Only receive the onStateChange event if the aStateFlags parameter
   *   includes nsIWebProgressListener::STATE_IS_NETWORK.
   *
   * NOTIFY_STATE_WINDOW
   *   Only receive the onStateChange event if the aStateFlags parameter
   *   includes nsIWebProgressListener::STATE_IS_WINDOW.
   *
   * NOTIFY_STATE_ALL
   *   Receive all onStateChange events.
   */

pub const NOTIFY_STATE_DOCUMENT: i64

NOTIFY_STATE_DOCUMENT: i64 = 2

pub const NOTIFY_STATE_NETWORK: i64

NOTIFY_STATE_NETWORK: i64 = 4

pub const NOTIFY_STATE_WINDOW: i64

NOTIFY_STATE_WINDOW: i64 = 8

pub const NOTIFY_STATE_ALL: i64

NOTIFY_STATE_ALL: i64 = 15

pub const NOTIFY_PROGRESS: i64

NOTIFY_PROGRESS: i64 = 16

/**
   * These flags indicate the other events to observe, corresponding to the
   * other four methods defined on nsIWebProgressListener.
   *
   * NOTIFY_PROGRESS
   *   Receive onProgressChange events.
   *
   * NOTIFY_STATUS
   *   Receive onStatusChange events.
   *
   * NOTIFY_SECURITY
   *   Receive onSecurityChange events.
   *
   * NOTIFY_LOCATION
   *   Receive onLocationChange events.
   *
   * NOTIFY_REFRESH
   *   Receive onRefreshAttempted events.
   *   This is defined on nsIWebProgressListener2.
   */

pub const NOTIFY_STATUS: i64

NOTIFY_STATUS: i64 = 32

pub const NOTIFY_SECURITY: i64

NOTIFY_SECURITY: i64 = 64

pub const NOTIFY_LOCATION: i64

NOTIFY_LOCATION: i64 = 128

pub const NOTIFY_REFRESH: i64

NOTIFY_REFRESH: i64 = 256

pub const NOTIFY_ALL: i64

NOTIFY_ALL: i64 = 511

/**
   * This flag enables all notifications.
   */

pub unsafe fn AddProgressListener(
&self,
aListener: *const nsIWebProgressListener,
aNotifyMask: uint32_t
) -> nsresult

/**
   * Registers a listener to receive web progress events.
   *
   * @param aListener
   *        The listener interface to be called when a progress event occurs.
   *        This object must also implement nsISupportsWeakReference.
   * @param aNotifyMask
   *        The types of notifications to receive.
   *
   * @throw NS_ERROR_INVALID_ARG
   *        Indicates that aListener was either null or that it does not
   *        support weak references.
   * @throw NS_ERROR_FAILURE
   *        Indicates that aListener was already registered.
   */

void addProgressListener (in nsIWebProgressListener aListener, in unsigned long aNotifyMask);

pub unsafe fn RemoveProgressListener(
&self,
aListener: *const nsIWebProgressListener
) -> nsresult

/**
   * Removes a previously registered listener of progress events.
   *
   * @param aListener
   *        The listener interface previously registered with a call to
   *        addProgressListener.
   *
   * @throw NS_ERROR_FAILURE
   *        Indicates that aListener was not registered.
   */

void removeProgressListener (in nsIWebProgressListener aListener);

pub unsafe fn GetDOMWindow(
&self,
aDOMWindow: *mut *const mozIDOMWindowProxy
) -> nsresult

/**
   * The DOM window associated with this nsIWebProgress instance.
   *
   * @throw NS_ERROR_FAILURE
   *        Indicates that there is no associated DOM window.
   */

readonly attribute mozIDOMWindowProxy DOMWindow;

pub unsafe fn GetDOMWindowID(&self, aDOMWindowID: *mut uint64_t) -> nsresult

readonly attribute uint64_t DOMWindowID;

pub unsafe fn GetInnerDOMWindowID(
&self,
aInnerDOMWindowID: *mut uint64_t
) -> nsresult

readonly attribute uint64_t innerDOMWindowID;

pub unsafe fn GetIsTopLevel(&self, aIsTopLevel: *mut bool) -> nsresult

/**
   * Indicates whether DOMWindow.top == DOMWindow.
   */

readonly attribute boolean isTopLevel;

pub unsafe fn GetIsLoadingDocument(
&self,
aIsLoadingDocument: *mut bool
) -> nsresult

/**
   * Indicates whether or not a document is currently being loaded
   * in the context of this nsIWebProgress instance.
   */

readonly attribute boolean isLoadingDocument;

pub unsafe fn GetLoadType(&self, aLoadType: *mut uint32_t) -> nsresult

/**
   * Contains a load type as specified by the load* constants in
   * nsIDocShellLoadInfo.idl.
   */

readonly attribute unsigned long loadType;

pub unsafe fn GetTarget(&self, aTarget: *mut *const nsIEventTarget) -> nsresult

/**
   * Main thread event target to which progress updates should be
   * dispatched. This typically will be a SchedulerEventTarget
   * corresponding to the tab requesting updates.
   */

attribute nsIEventTarget target;

pub unsafe fn SetTarget(&self, aTarget: *const nsIEventTarget) -> nsresult

/**
   * Main thread event target to which progress updates should be
   * dispatched. This typically will be a SchedulerEventTarget
   * corresponding to the tab requesting updates.
   */

attribute nsIEventTarget target;