Struct xpcom::interfaces::nsIWindowlessBrowser

#[repr(C)]

pub struct nsIWindowlessBrowser { /* fields omitted */ }

interface nsIWindowlessBrowser : nsIWebNavigation

/**
 * This interface represents a nsIWebBrowser instance with no associated OS
 * window. Its main function is to manage the lifetimes of those windows.
 * A strong reference to this object must be held until the window is
 * ready to be destroyed.
 */

Methods

impl nsIWindowlessBrowser

pub fn coerce<T: nsIWindowlessBrowserCoerce>(&self) -> &T

Cast this nsIWindowlessBrowser to one of its base interfaces.

impl nsIWindowlessBrowser

pub unsafe fn Close(&self) -> nsresult

/**
   * "Closes" the windowless browser and destroys its associated nsIWebBrowser
   * and docshell.
   *
   * This method *must* be called for every windowless browser before its last
   * reference is released.
   */

void close ();