Struct xpcom::interfaces::nsIWyciwygChannel

#[repr(C)]

pub struct nsIWyciwygChannel { /* fields omitted */ }

interface nsIWyciwygChannel : nsIChannel

/**
 * A channel to  manage all cache-related interactions for layout
 * when it is dealing with dynamic pages created through
 * document.write(). This interface provides methods that will
 * help layout save dynamic pages in cache for future retrievals.
 */

Methods

impl nsIWyciwygChannel

pub fn coerce<T: nsIWyciwygChannelCoerce>(&self) -> &T

Cast this nsIWyciwygChannel to one of its base interfaces.

impl nsIWyciwygChannel

pub unsafe fn WriteToCacheEntry(&self, aData: &nsAString) -> nsresult

/**
   * Append data to the cache entry; opens the cache entry if necessary.
   */

void writeToCacheEntry (in AString aData);

pub unsafe fn CloseCacheEntry(&self, reason: nsresult) -> nsresult

/**
   * Close the cache entry; subsequent writes have undefined behavior.
   */

void closeCacheEntry (in nsresult reason);

pub unsafe fn SetSecurityInfo(
&self,
aSecurityInfo: *const nsISupports
) -> nsresult

/**
   * Set the wyciwyg channels security info
   */

void setSecurityInfo (in nsISupports aSecurityInfo);

pub unsafe fn SetCharsetAndSource(
&self,
aSource: int32_t,
aCharset: &nsACString
) -> nsresult

/**
   * Store and read a charset and charset source on the wyciwyg channel.  These
   * are opaque values to the channel; consumers who set them should know what
   * they mean.
   */

void setCharsetAndSource (in long aSource, in ACString aCharset);

pub unsafe fn GetCharsetAndSource(
&self,
aSource: *mut int32_t,
_retval: &mut nsACString
) -> nsresult

/**
   * The return value is the charset.  Throws if either the charset or the
   * source cannot be retrieved.  This is guaranteed to return a nonzero source
   * and a nonempty charset if it does not throw.
   */

ACString getCharsetAndSource (out long aSource);