Struct xpcom::interfaces::nsIXPCComponents_Utils

#[repr(C)]

pub struct nsIXPCComponents_Utils { /* fields omitted */ }

interface nsIXPCComponents_Utils : nsISupports

/**
* interface of Components.utils
*/

Methods

impl nsIXPCComponents_Utils

pub fn coerce<T: nsIXPCComponents_UtilsCoerce>(&self) -> &T

Cast this nsIXPCComponents_Utils to one of its base interfaces.

impl nsIXPCComponents_Utils

pub unsafe fn GetSandbox(
&self,
aSandbox: *mut *const nsIXPCComponents_utils_Sandbox
) -> nsresult

readonly attribute nsIXPCComponents_utils_Sandbox Sandbox;

pub unsafe fn IsModuleLoaded(
&self,
aResourceURI: &nsACString,
_retval: *mut bool
) -> nsresult

/**
     * Returns true if the js file located at 'registryLocation' location has
     * been loaded previously via the import method above. Returns false
     * otherwise.
     *
     * @param resourceURI A resource:// URI string representing the location of
     *        the js file to be checked if it is already loaded or not.
     * @returns boolean, true if the js file has been loaded via import. false
     *          otherwise
     */

boolean isModuleLoaded (in AUTF8String aResourceURI);

pub unsafe fn Unload(&self, registryLocation: &nsACString) -> nsresult

void unload (in AUTF8String registryLocation);

pub unsafe fn ForceGC(&self) -> nsresult

void forceGC ();

pub unsafe fn ForceCC(
&self,
aListener: *const nsICycleCollectorListener
) -> nsresult

void forceCC ([optional] in nsICycleCollectorListener aListener);

pub unsafe fn FinishCC(&self) -> nsresult

void finishCC ();

pub unsafe fn CcSlice(&self, budget: int64_t) -> nsresult

void ccSlice (in long long budget);

pub unsafe fn GetMaxCCSliceTimeSinceClear(
&self,
_retval: *mut int32_t
) -> nsresult

long getMaxCCSliceTimeSinceClear ();

pub unsafe fn ClearMaxCCTime(&self) -> nsresult

void clearMaxCCTime ();

pub unsafe fn ForceShrinkingGC(&self) -> nsresult

void forceShrinkingGC ();

pub unsafe fn SchedulePreciseGC(
&self,
callback: *const ScheduledGCCallback
) -> nsresult

void schedulePreciseGC (in ScheduledGCCallback callback);

pub unsafe fn SchedulePreciseShrinkingGC(
&self,
callback: *const ScheduledGCCallback
) -> nsresult

void schedulePreciseShrinkingGC (in ScheduledGCCallback callback);

pub unsafe fn UnlinkGhostWindows(&self) -> nsresult

void unlinkGhostWindows ();

pub unsafe fn GetIsInAutomation(&self, aIsInAutomation: *mut bool) -> nsresult

readonly attribute boolean isInAutomation;

pub unsafe fn CrashIfNotInAutomation(&self) -> nsresult

void crashIfNotInAutomation ();

pub unsafe fn GetDOMClassInfo(
&self,
aClassName: &nsAString,
_retval: *mut *const nsIClassInfo
) -> nsresult

/**
     * Get a DOM classinfo for the given classname.  Only some class
     * names are supported.
     */

nsIClassInfo getDOMClassInfo (in AString aClassName);

pub unsafe fn GetWatchdogTimestamp(
&self,
aCategory: &nsAString,
_retval: *mut PRTime
) -> nsresult

/**
      * Retrieve the last time, in microseconds since epoch, that a given
      * watchdog-related event occured.
      *
      * Valid categories:
      *   "ContextStateChange"      - Context switching between active and inactive states
      *   "WatchdogWakeup"          - Watchdog waking up from sleeping
      *   "WatchdogHibernateStart"  - Watchdog begins hibernating
      *   "WatchdogHibernateStop"   - Watchdog stops hibernating
      */

PRTime getWatchdogTimestamp (in AString aCategory);

pub unsafe fn GetWebIDLCallerPrincipal(
&self,
_retval: *mut *const nsIPrincipal
) -> nsresult

nsIPrincipal getWebIDLCallerPrincipal ();

pub unsafe fn Now(&self, _retval: *mut c_double) -> nsresult

double now ();

pub unsafe fn ReadUTF8File(
&self,
file: *const nsIFile,
_retval: &mut nsACString
) -> nsresult

AUTF8String readUTF8File (in nsIFile file);

pub unsafe fn ReadUTF8URI(
&self,
url: *const nsIURI,
_retval: &mut nsACString
) -> nsresult

AUTF8String readUTF8URI (in nsIURI url);

pub unsafe fn BlockThreadedExecution(
&self,
aBlockedCallback: *const nsIBlockThreadedExecutionCallback
) -> nsresult

/**
     * If the main thread is using any kind of fancy cooperative
     * scheduling (e.g., Quantum DOM scheduling),
     * blockThreadedExecution disables it temporarily. The
     * aBlockedCallback is called when it has been completely disabled
     * and events are back to running sequentially on a single main
     * thread. Calling unblockThreadedExecution will re-enable thread
     * scheduling of the main thread. Multiple calls to
     * blockThreadedExecution will require the same number of calls to
     * unblockThreadedExecution in order to resume cooperative
     * scheduling.
     */

void blockThreadedExecution (in nsIBlockThreadedExecutionCallback aBlockedCallback);

pub unsafe fn UnblockThreadedExecution(&self) -> nsresult

void unblockThreadedExecution ();