Struct xpcom::interfaces::nsIXSLTProcessor

#[repr(C)]

pub struct nsIXSLTProcessor { /* fields omitted */ }

interface nsIXSLTProcessor : nsISupports

Methods

impl nsIXSLTProcessor

pub fn coerce<T: nsIXSLTProcessorCoerce>(&self) -> &T

Cast this nsIXSLTProcessor to one of its base interfaces.

impl nsIXSLTProcessor

pub unsafe fn ImportStylesheet(&self, style: *const nsIDOMNode) -> nsresult

/**
     * Import the stylesheet into this XSLTProcessor for transformations.
     *
     * @param style The root-node of a XSLT stylesheet. This can be either
     *              a document node or an element node. If a document node
     *              then the document can contain either a XSLT stylesheet
     *              or a LRE stylesheet.
     *              If the argument is an element node it must be the
     *              xsl:stylesheet (or xsl:transform) element of an XSLT
     *              stylesheet.
     */

void importStylesheet (in nsIDOMNode style);

pub unsafe fn TransformToFragment(
&self,
source: *const nsIDOMNode,
output: *const nsIDOMDocument,
_retval: *mut *const nsIDOMDocumentFragment
) -> nsresult

/**
     * Transforms the node source applying the stylesheet given by
     * the importStylesheet() function. The owner document of the output node
     * owns the returned document fragment.
     *
     * @param source The node to be transformed
     * @param output This document is used to generate the output
     * @return DocumentFragment The result of the transformation
     */

nsIDOMDocumentFragment transformToFragment (in nsIDOMNode source, in nsIDOMDocument output);

pub unsafe fn TransformToDocument(
&self,
source: *const nsIDOMNode,
_retval: *mut *const nsIDOMDocument
) -> nsresult

/**
     * Transforms the node source applying the stylesheet given by the
     * importStylesheet() function.
     *
     * @param source The node to be transformed
     * @return Document The result of the transformation
     */

nsIDOMDocument transformToDocument (in nsIDOMNode source);

pub unsafe fn SetParameter(
&self,
namespaceURI: &nsAString,
localName: &nsAString,
value: *const nsIVariant
) -> nsresult

/**
     * Sets a parameter to be used in subsequent transformations with this
     * nsIXSLTProcessor. If the parameter doesn't exist in the stylesheet the
     * parameter will be ignored.
     *
     * @param namespaceURI The namespaceURI of the XSLT parameter
     * @param localName    The local name of the XSLT parameter
     * @param value        The new value of the XSLT parameter
     *
     * @exception NS_ERROR_ILLEGAL_VALUE The datatype of value is
     *                                   not supported
     */

void setParameter (in DOMString namespaceURI, in DOMString localName, in nsIVariant value);

pub unsafe fn GetParameter(
&self,
namespaceURI: &nsAString,
localName: &nsAString,
_retval: *mut *const nsIVariant
) -> nsresult

/**
     * Gets a parameter if previously set by setParameter. Returns null
     * otherwise.
     *
     * @param namespaceURI The namespaceURI of the XSLT parameter
     * @param localName    The local name of the XSLT parameter
     * @return nsIVariant  The value of the XSLT parameter
     */

nsIVariant getParameter (in DOMString namespaceURI, in DOMString localName);

pub unsafe fn RemoveParameter(
&self,
namespaceURI: &nsAString,
localName: &nsAString
) -> nsresult

/**
     * Removes a parameter, if set. This will make the processor use the
     * default-value for the parameter as specified in the stylesheet.
     *
     * @param namespaceURI The namespaceURI of the XSLT parameter
     * @param localName    The local name of the XSLT parameter
     */

void removeParameter (in DOMString namespaceURI, in DOMString localName);

pub unsafe fn ClearParameters(&self) -> nsresult

/**
     * Removes all set parameters from this nsIXSLTProcessor. This will make
     * the processor use the default-value for all parameters as specified in
     * the stylesheet.
     */

void clearParameters ();

pub unsafe fn Reset(&self) -> nsresult

/**
     * Remove all parameters and stylesheets from this nsIXSLTProcessor.
     */

void reset ();