Struct xpcom::interfaces::nsIXULRuntime

#[repr(C)]

pub struct nsIXULRuntime { /* fields omitted */ }

interface nsIXULRuntime : nsISupports

/**
 * Provides information about the XUL runtime.
 * @status UNSTABLE - This interface is not frozen and will probably change in
 *                    future releases. If you need this functionality to be
 *                    stable/frozen, please contact Benjamin Smedberg.
 */

Methods

impl nsIXULRuntime

pub fn coerce<T: nsIXULRuntimeCoerce>(&self) -> &T

Cast this nsIXULRuntime to one of its base interfaces.

impl nsIXULRuntime

pub const PROCESS_TYPE_DEFAULT: i64

PROCESS_TYPE_DEFAULT: i64 = 0

/**
   * The legal values of processType.
   */

pub const PROCESS_TYPE_PLUGIN: i64

PROCESS_TYPE_PLUGIN: i64 = 1

pub const PROCESS_TYPE_CONTENT: i64

PROCESS_TYPE_CONTENT: i64 = 2

pub const PROCESS_TYPE_IPDLUNITTEST: i64

PROCESS_TYPE_IPDLUNITTEST: i64 = 3

pub const PROCESS_TYPE_GMPLUGIN: i64

PROCESS_TYPE_GMPLUGIN: i64 = 4

pub const PROCESS_TYPE_GPU: i64

PROCESS_TYPE_GPU: i64 = 5

pub const PROCESS_TYPE_PDFIUM: i64

PROCESS_TYPE_PDFIUM: i64 = 6

pub const E10S_MULTI_EXPERIMENT: i64

E10S_MULTI_EXPERIMENT: i64 = 1

/**
   * The current e10s-multi experiment number. Set dom.ipc.multiOptOut to (at
   * least) this to disable it until the next experiment.
   */

pub unsafe fn GetInSafeMode(&self, aInSafeMode: *mut bool) -> nsresult

/**
   * Whether the application was launched in safe mode.
   */

readonly attribute boolean inSafeMode;

pub unsafe fn GetLogConsoleErrors(
&self,
aLogConsoleErrors: *mut bool
) -> nsresult

/**
   * Whether to write console errors to a log file. If a component
   * encounters startup errors that might prevent the app from showing
   * proper UI, it should set this flag to "true".
   */

attribute boolean logConsoleErrors;

pub unsafe fn SetLogConsoleErrors(&self, aLogConsoleErrors: bool) -> nsresult

/**
   * Whether to write console errors to a log file. If a component
   * encounters startup errors that might prevent the app from showing
   * proper UI, it should set this flag to "true".
   */

attribute boolean logConsoleErrors;

pub unsafe fn GetOS(&self, aOS: &mut nsACString) -> nsresult

/**
   * A string tag identifying the current operating system. This is taken
   * from the OS_TARGET configure variable. It will always be available.
   */

readonly attribute AUTF8String OS;

pub unsafe fn GetXPCOMABI(&self, aXPCOMABI: &mut nsACString) -> nsresult

/**
   * A string tag identifying the binary ABI of the current processor and
   * compiler vtable. This is taken from the TARGET_XPCOM_ABI configure
   * variable. It may not be available on all platforms, especially
   * unusual processor or compiler combinations.
   *
   * The result takes the form <processor>-<compilerABI>, for example:
   *   x86-msvc
   *   ppc-gcc3
   *
   * This value should almost always be used in combination with "OS".
   *
   * @throw NS_ERROR_NOT_AVAILABLE if not available.
   */

readonly attribute AUTF8String XPCOMABI;

pub unsafe fn GetWidgetToolkit(
&self,
aWidgetToolkit: &mut nsACString
) -> nsresult

/**
   * A string tag identifying the target widget toolkit in use.
   * This is taken from the MOZ_WIDGET_TOOLKIT configure variable.
   */

readonly attribute AUTF8String widgetToolkit;

pub unsafe fn GetProcessType(&self, aProcessType: *mut uint32_t) -> nsresult

/**
   * The type of the caller's process.  Returns one of the values above.
   */

readonly attribute unsigned long processType;

pub unsafe fn GetProcessID(&self, aProcessID: *mut uint32_t) -> nsresult

/**
   * The system process ID of the caller's process.
   */

readonly attribute unsigned long processID;

pub unsafe fn GetUniqueProcessID(
&self,
aUniqueProcessID: *mut uint64_t
) -> nsresult

/**
   * A globally unique and non-recycled ID of the caller's process.
   */

readonly attribute uint64_t uniqueProcessID;

pub unsafe fn GetRemoteType(&self, aRemoteType: &mut nsAString) -> nsresult

/**
   * The type of remote content process we're running in.
   * null if we're in the parent/chrome process.
   */

readonly attribute DOMString remoteType;

pub unsafe fn GetBrowserTabsRemoteAutostart(
&self,
aBrowserTabsRemoteAutostart: *mut bool
) -> nsresult

/**
   * If true, browser tabs may be opened by default in a different process
   * from the main browser UI.
   */

readonly attribute boolean browserTabsRemoteAutostart;

pub unsafe fn GetMaxWebProcessCount(
&self,
aMaxWebProcessCount: *mut uint32_t
) -> nsresult

/**
   * Returns the number of content processes to use for normal web pages. If
   * this value is > 1, then e10s-multi should be considered to be "on".
   *
   * NB: If browserTabsRemoteAutostart is false, then this value has no
   * meaning and e10s should be considered to be "off"!
   */

readonly attribute uint32_t maxWebProcessCount;

pub unsafe fn GetAccessibilityEnabled(
&self,
aAccessibilityEnabled: *mut bool
) -> nsresult

/**
   * If true, the accessibility service is running.
   */

readonly attribute boolean accessibilityEnabled;

pub unsafe fn GetAccessibleHandlerUsed(
&self,
aAccessibleHandlerUsed: *mut bool
) -> nsresult

/**
   * If true, the AccessibleHandler dll is used.
   */

readonly attribute boolean accessibleHandlerUsed;

pub unsafe fn GetAccessibilityInstantiator(
&self,
aAccessibilityInstantiator: &mut nsAString
) -> nsresult

/**
   * Executable of Windows service that activated accessibility.
   */

readonly attribute DOMString accessibilityInstantiator;

pub unsafe fn GetShouldBlockIncompatJaws(
&self,
aShouldBlockIncompatJaws: *mut bool
) -> nsresult

/**
   * Temporary, do not use. Indicates if an incompat version of JAWS
   * screen reader software is loaded in our process space.
   */

readonly attribute boolean shouldBlockIncompatJaws;

pub unsafe fn GetIs64Bit(&self, aIs64Bit: *mut bool) -> nsresult

/**
   * Indicates whether the current Firefox build is 64-bit.
   */

readonly attribute boolean is64Bit;

pub unsafe fn InvalidateCachesOnRestart(&self) -> nsresult

/**
   * Signal the apprunner to invalidate caches on the next restart.
   * This will cause components to be autoregistered and all
   * fastload data to be re-created.
   */

void invalidateCachesOnRestart ();

pub unsafe fn EnsureContentProcess(&self) -> nsresult

/**
   * Starts a child process. This method is intented to pre-start a
   * content child process so that when it is actually needed, it is
   * ready to go.
   *
   * @throw NS_ERROR_NOT_AVAILABLE if not available.
   */

void ensureContentProcess ();

pub unsafe fn GetReplacedLockTime(
&self,
aReplacedLockTime: *mut PRTime
) -> nsresult

/**
   * Modification time of the profile lock before the profile was locked on
   * this startup. Used to know the last time the profile was used and not
   * closed cleanly. This is set to 0 if there was no existing profile lock.
   */

readonly attribute PRTime replacedLockTime;

pub unsafe fn GetIsReleaseOrBeta(&self, aIsReleaseOrBeta: *mut bool) -> nsresult

/**
   * True if this is RELEASE_OR_BETA.
   */

readonly attribute boolean isReleaseOrBeta;

pub unsafe fn GetIsOfficialBranding(
&self,
aIsOfficialBranding: *mut bool
) -> nsresult

/**
   * True if this build uses official branding (MOZ_OFFICIAL_BRANDING).
   */

readonly attribute boolean isOfficialBranding;

pub unsafe fn GetDefaultUpdateChannel(
&self,
aDefaultUpdateChannel: &mut nsACString
) -> nsresult

/**
   * The default update channel (MOZ_UPDATE_CHANNEL).
   */

readonly attribute AUTF8String defaultUpdateChannel;

pub unsafe fn GetDistributionID(
&self,
aDistributionID: &mut nsACString
) -> nsresult

/**
   * The distribution ID for this build (MOZ_DISTRIBUTION_ID).
   */

readonly attribute AUTF8String distributionID;

pub unsafe fn GetWindowsDLLBlocklistStatus(
&self,
aWindowsDLLBlocklistStatus: *mut bool
) -> nsresult

/**
   * True if Windows DLL blocklist initialized correctly. This is
   * primarily for automated testing purposes.
   */

readonly attribute boolean windowsDLLBlocklistStatus;