Struct xpcom::interfaces::nsIXULWindow

#[repr(C)]

pub struct nsIXULWindow { /* fields omitted */ }

interface nsIXULWindow : nsISupports

Methods

impl nsIXULWindow

pub fn coerce<T: nsIXULWindowCoerce>(&self) -> &T

Cast this nsIXULWindow to one of its base interfaces.

impl nsIXULWindow

pub const lowestZ: i64

lowestZ: i64 = 0

pub const loweredZ: i64

loweredZ: i64 = 4

pub const normalZ: i64

normalZ: i64 = 5

pub const raisedZ: i64

raisedZ: i64 = 6

pub const highestZ: i64

highestZ: i64 = 9

pub unsafe fn GetDocShell(&self, aDocShell: *mut *const nsIDocShell) -> nsresult

/**
   * The docshell owning the XUL for this window.
   */

readonly attribute nsIDocShell docShell;

pub unsafe fn GetIntrinsicallySized(
&self,
aIntrinsicallySized: *mut bool
) -> nsresult

/**
   * Indicates if this window is instrinsically sized.
   */

attribute boolean intrinsicallySized;

pub unsafe fn SetIntrinsicallySized(
&self,
aIntrinsicallySized: bool
) -> nsresult

/**
   * Indicates if this window is instrinsically sized.
   */

attribute boolean intrinsicallySized;

pub unsafe fn GetPrimaryContentShell(
&self,
aPrimaryContentShell: *mut *const nsIDocShellTreeItem
) -> nsresult

/**
   * The primary content shell.
   *
   * Note that this is a docshell tree item and therefore can not be assured of
   * what object it is. It could be an editor, a docshell, or a browser object.
   * Or down the road any other object that supports being a DocShellTreeItem
   * Query accordingly to determine the capabilities.
   */

readonly attribute nsIDocShellTreeItem primaryContentShell;

pub unsafe fn GetPrimaryTabParent(
&self,
aPrimaryTabParent: *mut *const nsITabParent
) -> nsresult

/**
   * In multiprocess case we may not have primaryContentShell but
   * primaryTabParent.
   */

readonly attribute nsITabParent primaryTabParent;

pub unsafe fn TabParentAdded(
&self,
aTab: *const nsITabParent,
aPrimary: bool
) -> nsresult

void tabParentAdded (in nsITabParent aTab, in boolean aPrimary);

pub unsafe fn TabParentRemoved(&self, aTab: *const nsITabParent) -> nsresult

void tabParentRemoved (in nsITabParent aTab);

pub unsafe fn AddChildWindow(&self, aChild: *const nsIXULWindow) -> nsresult

/**
   * Tell this window that it has picked up a child XUL window
   * @param aChild the child window being added
   */

void addChildWindow (in nsIXULWindow aChild);

pub unsafe fn RemoveChildWindow(&self, aChild: *const nsIXULWindow) -> nsresult

/**
   * Tell this window that it has lost a child XUL window
   * @param aChild the child window being removed
   */

void removeChildWindow (in nsIXULWindow aChild);

pub unsafe fn Center(
&self,
aRelative: *const nsIXULWindow,
aScreen: bool,
aAlert: bool
) -> nsresult

/**
   * Move the window to a centered position.
   * @param aRelative If not null, the window relative to which the window is
   *                  moved. See aScreen parameter for details.
   * @param aScreen   PR_TRUE to center the window relative to the screen
   *                  containing aRelative if aRelative is not null. If
   *                  aRelative is null then relative to the screen of the
   *                  opener window if it was initialized by passing it to
   *                  nsWebShellWindow::Initialize. Failing that relative to
   *                  the main screen.
   *                  PR_FALSE to center it relative to aRelative itself.
   * @param aAlert    PR_TRUE to move the window to an alert position,
   *                  generally centered horizontally and 1/3 down from the top.
   */

void center (in nsIXULWindow aRelative, in boolean aScreen, in boolean aAlert);

pub unsafe fn ShowModal(&self) -> nsresult

/**
   * Shows the window as a modal window. That is, ensures that it is visible
   * and runs a local event loop, exiting only once the window has been closed.
   */

void showModal ();

pub unsafe fn GetZLevel(&self, aZLevel: *mut uint32_t) -> nsresult

attribute unsigned long zLevel;

pub unsafe fn SetZLevel(&self, aZLevel: uint32_t) -> nsresult

attribute unsigned long zLevel;

pub unsafe fn GetChromeFlags(&self, aChromeFlags: *mut uint32_t) -> nsresult

attribute uint32_t chromeFlags;

pub unsafe fn SetChromeFlags(&self, aChromeFlags: uint32_t) -> nsresult

attribute uint32_t chromeFlags;

pub unsafe fn AssumeChromeFlagsAreFrozen(&self) -> nsresult

/**
   * Begin assuming |chromeFlags| don't change hereafter, and assert
   * if they do change.  The state change is one-way and idempotent.
   */

void assumeChromeFlagsAreFrozen ();

pub unsafe fn CreateNewWindow(
&self,
aChromeFlags: int32_t,
aOpeningTab: *const nsITabParent,
aOpener: *const mozIDOMWindowProxy,
aNextTabParentId: uint64_t,
_retval: *mut *const nsIXULWindow
) -> nsresult

/**
   * Create a new window.
   * @param aChromeFlags see nsIWebBrowserChrome
   * @param aOpeningTab the TabParent that requested this new window be opened.
   *                    Can be left null.
   * @param aOpener The window which is requesting that this new window be opened.
   * @param aNextTabParentId The integer ID of the next tab parent actor to use.
   *        0 means there is no next tab parent actor to use.
   * @return the newly minted window
   */

nsIXULWindow createNewWindow (in int32_t aChromeFlags, in nsITabParent aOpeningTab, in mozIDOMWindowProxy aOpener, in unsigned long long aNextTabParentId);

pub unsafe fn GetXULBrowserWindow(
&self,
aXULBrowserWindow: *mut *const nsIXULBrowserWindow
) -> nsresult

attribute nsIXULBrowserWindow XULBrowserWindow;

pub unsafe fn SetXULBrowserWindow(
&self,
aXULBrowserWindow: *const nsIXULBrowserWindow
) -> nsresult

attribute nsIXULBrowserWindow XULBrowserWindow;

pub unsafe fn ApplyChromeFlags(&self) -> nsresult

/**
   * Back-door method to force application of chrome flags at a particular
   * time.  Do NOT call this unless you know what you're doing!  In particular,
   * calling this when this XUL window doesn't yet have a document in its
   * docshell could cause problems.
   */

[noscript] void applyChromeFlags ();

pub unsafe fn SizeShellToWithLimit(
&self,
aDesiredWidth: int32_t,
aDesiredHeight: int32_t,
shellItemWidth: int32_t,
shellItemHeight: int32_t
) -> c_void

/**
   * Given the dimensions of some content area held within this
   * XUL window, and assuming that that content area will change
   * its dimensions in linear proportion to the dimensions of this
   * XUL window, changes the size of the XUL window so that the
   * content area reaches a particular size.
   *
   * We need to supply the content area dimensions because sometimes
   * the child's nsDocShellTreeOwner needs to propagate a SizeShellTo
   * call to the parent. But the shellItem argument of the call will
   * not be available on the parent side.
   *
   * Note: this is an internal method, other consumers should never call this.
   *
   * @param aDesiredWidth
   *        The desired width of the content area in device pixels.
   * @param aDesiredHeight
   *        The desired height of the content area in device pixels.
   * @param shellItemWidth
   *        The current width of the content area.
   * @param shellItemHeight
   *        The current height of the content area.
   */

[noscript,notxpcom] void sizeShellToWithLimit (in int32_t aDesiredWidth, in int32_t aDesiredHeight, in int32_t shellItemWidth, in int32_t shellItemHeight);

pub unsafe fn GetNextTabParentId(
&self,
aNextTabParentId: *mut uint64_t
) -> nsresult

/**
   * If the window was opened as a content window by script, this will return the
   * integer ID of the next TabParent actor to use.
   */

[noscript] readonly attribute unsigned long long nextTabParentId;