Struct xpcom::interfaces::nsIZipEntry

#[repr(C)]

pub struct nsIZipEntry { /* fields omitted */ }

interface nsIZipEntry : nsISupports

Methods

impl nsIZipEntry

pub fn coerce<T: nsIZipEntryCoerce>(&self) -> &T

Cast this nsIZipEntry to one of its base interfaces.

impl nsIZipEntry

pub unsafe fn GetCompression(&self, aCompression: *mut uint16_t) -> nsresult

/**
     * The type of compression used for the item.  The possible values and
     * their meanings are defined in the zip file specification at
     * http://www.pkware.com/business_and_developers/developer/appnote/
     */

readonly attribute unsigned short compression;

pub unsafe fn GetSize(&self, aSize: *mut uint32_t) -> nsresult

/**
     * The compressed size of the data in the item.
     */

readonly attribute unsigned long size;

pub unsafe fn GetRealSize(&self, aRealSize: *mut uint32_t) -> nsresult

/**
     * The uncompressed size of the data in the item.
     */

readonly attribute unsigned long realSize;

pub unsafe fn GetCRC32(&self, aCRC32: *mut uint32_t) -> nsresult

/**
     * The CRC-32 hash of the file in the entry.
     */

readonly attribute unsigned long CRC32;

pub unsafe fn GetIsDirectory(&self, aIsDirectory: *mut bool) -> nsresult

/**
     * True if the name of the entry ends with '/' and false otherwise.
     */

readonly attribute boolean isDirectory;

pub unsafe fn GetLastModifiedTime(
&self,
aLastModifiedTime: *mut PRTime
) -> nsresult

/**
     * The time at which this item was last modified.
     */

readonly attribute PRTime lastModifiedTime;

pub unsafe fn GetIsSynthetic(&self, aIsSynthetic: *mut bool) -> nsresult

/**
     * Use this attribute to determine whether this item is an actual zip entry
     * or is one synthesized for part of a real entry's path.  A synthesized
     * entry represents a directory within the zip file which has no
     * corresponding entry within the zip file.  For example, the entry for the
     * directory foo/ in a zip containing exactly one entry for foo/bar.txt
     * is synthetic.  If the zip file contains an actual entry for a directory,
     * this attribute will be false for the nsIZipEntry for that directory.
     * It is impossible for a file to be synthetic.
     */

readonly attribute boolean isSynthetic;

pub unsafe fn GetPermissions(&self, aPermissions: *mut uint32_t) -> nsresult

/**
     * The UNIX style file permissions of this item.
     */

readonly attribute unsigned long permissions;