Struct xpcom::interfaces::nsPIDNSService

#[repr(C)]

pub struct nsPIDNSService { /* fields omitted */ }

interface nsPIDNSService : nsIDNSService

/**
 * This is a private interface used by the internals of the networking library.
 * It will never be frozen.  Do not use it in external code.
 */

Methods

impl nsPIDNSService

pub fn coerce<T: nsPIDNSServiceCoerce>(&self) -> &T

Cast this nsPIDNSService to one of its base interfaces.

impl nsPIDNSService

pub unsafe fn Init(&self) -> nsresult

/**
     * called to initialize the DNS service.
     */

void init ();

pub unsafe fn Shutdown(&self) -> nsresult

/**
     * called to shutdown the DNS service.  any pending asynchronous
     * requests will be canceled, and the local cache of DNS records
     * will be cleared.  NOTE: the operating system may still have
     * its own cache of DNS records, which would be unaffected by
     * this method.
     */

void shutdown ();

pub unsafe fn GetPrefetchEnabled(&self, aPrefetchEnabled: *mut bool) -> nsresult

/**
     * Whether or not DNS prefetching (aka RESOLVE_SPECULATE) is enabled
     */

attribute boolean prefetchEnabled;

pub unsafe fn SetPrefetchEnabled(&self, aPrefetchEnabled: bool) -> nsresult

/**
     * Whether or not DNS prefetching (aka RESOLVE_SPECULATE) is enabled
     */

attribute boolean prefetchEnabled;