Struct xpcom::interfaces::nsPIPlacesDatabase

#[repr(C)]

pub struct nsPIPlacesDatabase { /* fields omitted */ }

interface nsPIPlacesDatabase : nsISupports

/**
 * This is a private interface used by Places components to get access to the
 * database.  If outside consumers wish to use this, they should only read from
 * the database so they do not break any internal invariants.
 */

Methods

impl nsPIPlacesDatabase

pub fn coerce<T: nsPIPlacesDatabaseCoerce>(&self) -> &T

Cast this nsPIPlacesDatabase to one of its base interfaces.

impl nsPIPlacesDatabase

pub unsafe fn GetDBConnection(
&self,
aDBConnection: *mut *const mozIStorageConnection
) -> nsresult

/**
   * The database connection used by Places.
   */

readonly attribute mozIStorageConnection DBConnection;

pub unsafe fn AsyncExecuteLegacyQueries(
&self,
aQueries: *mut *const nsINavHistoryQuery,
aQueryCount: uint32_t,
aOptions: *const nsINavHistoryQueryOptions,
aCallback: *const mozIStorageStatementCallback,
_retval: *mut *const mozIStoragePendingStatement
) -> nsresult

/**
   * Asynchronously executes the statement created from queries.
   *
   * @see nsINavHistoryService::executeQueries
   * @note THIS IS A TEMPORARY API.  Don't rely on it, since it will be replaced
   *       in future versions by a real async querying API.
   * @note Results obtained from this method differ from results obtained from
   *       executeQueries, because there is additional filtering and sorting
   *       done by the latter.  Thus you should use executeQueries, unless you
   *       are absolutely sure that the returned results are fine for
   *       your use-case.
   */

mozIStoragePendingStatement asyncExecuteLegacyQueries ([array, size_is (aQueryCount)] in nsINavHistoryQuery aQueries, in unsigned long aQueryCount, in nsINavHistoryQueryOptions aOptions, in mozIStorageStatementCallback aCallback);

pub unsafe fn GetShutdownClient(
&self,
aShutdownClient: *mut *const nsIAsyncShutdownClient
) -> nsresult

/**
   * Hook for clients who need to perform actions during/by the end of
   * the shutdown of the database.
   * May be null if it's too late to get one.
   */

readonly attribute nsIAsyncShutdownClient shutdownClient;

pub unsafe fn GetConnectionShutdownClient(
&self,
aConnectionShutdownClient: *mut *const nsIAsyncShutdownClient
) -> nsresult

/**
   * Hook for internal clients who need to perform actions just before the
   * connection gets closed.
   * May be null if it's too late to get one.
   */

readonly attribute nsIAsyncShutdownClient connectionShutdownClient;