Struct xpcom::interfaces::nsPIWindowWatcher

#[repr(C)]

pub struct nsPIWindowWatcher { /* fields omitted */ }

interface nsPIWindowWatcher : nsISupports

Methods

impl nsPIWindowWatcher

pub fn coerce<T: nsPIWindowWatcherCoerce>(&self) -> &T

Cast this nsPIWindowWatcher to one of its base interfaces.

impl nsPIWindowWatcher

pub unsafe fn AddWindow(
&self,
aWindow: *const mozIDOMWindowProxy,
aChrome: *const nsIWebBrowserChrome
) -> nsresult

/** A window has been created. Add it to our list.
      @param aWindow the window to add
      @param aChrome the corresponding chrome window. The DOM window
                     and chrome will be mapped together, and the corresponding
                     chrome can be retrieved using the (not private)
                     method getChromeForWindow. If null, any extant mapping
                     will be cleared.
  */

void addWindow (in mozIDOMWindowProxy aWindow, in nsIWebBrowserChrome aChrome);

pub unsafe fn RemoveWindow(
&self,
aWindow: *const mozIDOMWindowProxy
) -> nsresult

/** A window has been closed. Remove it from our list.
      @param aWindow the window to remove
  */

void removeWindow (in mozIDOMWindowProxy aWindow);

pub unsafe fn OpenWindow2(
&self,
aParent: *const mozIDOMWindowProxy,
aUrl: *const c_char,
aName: *const c_char,
aFeatures: *const c_char,
aCalledFromScript: bool,
aDialog: bool,
aNavigate: bool,
aArgs: *const nsISupports,
aIsPopupSpam: bool,
aForceNoOpener: bool,
aLoadInfo: *const nsIDocShellLoadInfo,
_retval: *mut *const mozIDOMWindowProxy
) -> nsresult

/** Like the public interface's open(), but can handle openDialog-style
      arguments and calls which shouldn't result in us navigating the window.

      @param aParent parent window, if any. Null if no parent.  If it is
             impossible to get to an nsIWebBrowserChrome from aParent, this
             method will effectively act as if aParent were null.
      @param aURL url to which to open the new window. Must already be
             escaped, if applicable. can be null.
      @param aName window name from JS window.open. can be null.  If a window
             with this name already exists, the openWindow call may just load
             aUrl in it (if aUrl is not null) and return it.
      @param aFeatures window features from JS window.open. can be null.
      @param aCalledFromScript true if we were called from script.
      @param aDialog use dialog defaults (see nsIDOMWindow::openDialog)
      @param aNavigate true if we should navigate the new window to the
             specified URL.
      @param aArgs Window argument
      @param aIsPopupSpam true if the window is a popup spam window; used for
                          popup blocker internals.
      @param aForceNoOpener If true, force noopener behavior.  This means not
                            looking for existing windows with the given name,
                            not setting an opener on the newly opened window,
                            and returning null from this method.
      @param aLoadInfo if aNavigate is true, this allows the caller to pass in
                       an nsIDocShellLoadInfo to use for the navigation.
                       Callers can pass in null if they want the windowwatcher
                       to just construct a loadinfo itself.  If aNavigate is
                       false, this argument is ignored.

      @return the new window

      @note This method may examine the JS context stack for purposes of
            determining the security context to use for the search for a given
            window named aName.
      @note This method should try to set the default charset for the new
            window to the default charset of the document in the calling window
            (which is determined based on the JS stack and the value of
            aParent).  This is not guaranteed, however.
  */

mozIDOMWindowProxy openWindow2 (in mozIDOMWindowProxy aParent, in string aUrl, in string aName, in string aFeatures, in boolean aCalledFromScript, in boolean aDialog, in boolean aNavigate, in nsISupports aArgs, in boolean aIsPopupSpam, in boolean aForceNoOpener, in nsIDocShellLoadInfo aLoadInfo);

pub unsafe fn OpenWindowWithTabParent(
&self,
aOpeningTab: *const nsITabParent,
aFeatures: &nsACString,
aCalledFromJS: bool,
aOpenerFullZoom: c_float,
aNextTabParentId: uint64_t,
aForceNoOpener: bool,
_retval: *mut *const nsITabParent
) -> nsresult

/**
   * Opens a new window so that the window that aOpeningTab belongs to
   * is set as the parent window. The newly opened window will also
   * inherit load context information from aOpeningTab.
   *
   * @param aOpeningTab
   *        The nsITabParent that is requesting the new window be opened.
   * @param aFeatures
   *        Window features if called with window.open or similar.
   * @param aCalledFromJS
   *        True if called via window.open or similar.
   * @param aOpenerFullZoom
   *        The current zoom multiplier for the opener tab. This is then
   *        applied to the newly opened window.
   * @param aNextTabParentId
   *        The integer ID for the next tab parent actor.
   *        0 means there is no next tab parent actor to use.
   * @param aForceNoOpener
   *        If true, then aOpeningTab will not be used to set the opener
   *        for the newly created window.
   *
   * @return the nsITabParent of the initial browser for the newly opened
   *         window.
   */

nsITabParent openWindowWithTabParent (in nsITabParent aOpeningTab, in ACString aFeatures, in boolean aCalledFromJS, in float aOpenerFullZoom, in unsigned long long aNextTabParentId, in boolean aForceNoOpener);

pub unsafe fn FindItemWithName(
&self,
aName: &nsAString,
aRequestor: *const nsIDocShellTreeItem,
aOriginalRequestor: *const nsIDocShellTreeItem,
_retval: *mut *const nsIDocShellTreeItem
) -> nsresult

/**
   * Find a named docshell tree item amongst all windows registered
   * with the window watcher.  This may be a subframe in some window,
   * for example.
   *
   * @param aName the name of the window.  Must not be null.
   * @param aRequestor the tree item immediately making the request.
   *        We should make sure to not recurse down into its findItemWithName
   *        method.
   * @param aOriginalRequestor the original treeitem that made the request.
   *        Used for security checks.
   * @return the tree item with aName as the name, or null if there
   *         isn't one.  "Special" names, like _self, _top, etc, will be
   *         treated specially only if aRequestor is null; in that case they
   *         will be resolved relative to the first window the windowwatcher
   *         knows about.
   * @see findItemWithName methods on nsIDocShellTreeItem and
   *      nsIDocShellTreeOwner
   */

nsIDocShellTreeItem findItemWithName (in AString aName, in nsIDocShellTreeItem aRequestor, in nsIDocShellTreeItem aOriginalRequestor);