Trait xpcom::XpCom

pub unsafe trait XpCom: RefCounted {
    const IID: nsIID;
    fn query_interface<T: XpCom>(&self) -> Option<RefPtr<T>> { ... }
}

A type which implements XpCom must follow the following rules:

It must be a legal XPCOM interface.

The result of a QueryInterface or similar call, passing IID, must return a valid reference to an object of the given type.

It must be valid to cast a &self reference to a &nsISupports reference.

Associated Constants

const IID: nsIID

Provided Methods

fn query_interface<T: XpCom>(&self) -> Option<RefPtr<T>>

Perform a QueryInterface call on this object, attempting to dynamically cast it to the requested interface type. Returns Some(RefPtr) if the cast succeeded, and None otherwise.